Autodocodec.Extended

Documentation

disjointEnumCodec :: forall enum context. Eq enum => NonEmpty (enum, Codec context enum enum) -> Codec context enum enum

A codec for an enum that can be written each with their own codec. Unlike enumCodec, disjointEnumCodec assumes that each provided codec is disjoint.

WARNING

If you don't provide a string for one of the type's constructors, the last codec in the list will be used instead.

disjointStringConstCodec :: forall constant. Eq constant => NonEmpty (constant, Text) -> JSONCodec constant

A codec for an enum that can be written as constant string values Unlike stringConstCodec, this function assumes that the provided values and strings are disjoint.

Example usage

>>> data Fruit = Apple | Orange deriving (Show, Eq)
>>> let c = disjointStringConstCodec [(Apple, "foo"), (Orange, "bar")]
>>> toJSONVia c Orange
String "bar"
>>> JSON.parseMaybe (parseJSONVia c) (String "foo") :: Maybe Fruit
Just Apple

WARNING

If you don't provide a string for one of the type's constructors, the last string in the list will be used instead:

>>> let c = disjointStringConstCodec [(Apple, "foo")]
>>> toJSONVia c Orange
String "foo"

class HasObjectCodec a where

Class for types that have a JSONObjectCodec, but not necessarily a JSONValueCodec. Used for providec codecs for sum type alternatives via TypeAlternative.

Methods

objectCodec :: JSONObjectCodec a

requiredTypeField :: Text -> ObjectCodec void ()

altCodec :: HasObjectCodec a => Text -> Text -> JSONCodec a

disjointMatchChoiceCodec

Arguments

:: Codec context input output
First codec

-> Codec context input' output
Second codec

-> (newInput -> Either input input')
Rendering chooser

-> Codec context newInput output

Disjoint version of matchChoiceCodec

data DisjunctCodec context newInput output where

Constructors

DisjunctCodec :: (newInput -> Maybe input) -> Codec context input output -> DisjunctCodec context newInput output

disjointMatchChoicesNECodec

Arguments

:: NonEmpty (DisjunctCodec context newInput output)
Codecs, each which their own rendering matcher

-> Codec context newInput output

A choice codec for a disjoint non-empty list of options Note that this list of options must be complete. There is a variant of newInput for which a DisjunctCodec is not provided then encoding may fail with a call to error (via fromJust)

data TypeAlternative a where

Data needed to generate a codec for one alternative of a sum type a. The existenstially quantified type b represents the object type contained within the alternative.

Constructors

TypeAlternative

Fields

:: HasObjectCodec b

=> Text

Name of the object type for the alternative

-> Text

Value to require in the "type" field of the object

-> Prism' a b

Prism to access values of the alternative

-> TypeAlternative a

sumTypeCodec :: NonEmpty (TypeAlternative a) -> JSONCodec a

A codec for a sum type. Note: the list of TypeAlternatives must cover all constructors of the sum type Otherwise encoding may fail with a call to error. This is not checked by the compiler. Example: > > data Field > = ColumnField (ValueWrapper "column" API.V0.ColumnName) > | RelationshipField RelField > deriving stock (Eq, Ord, Show, Generic, Data) > > $(makePrisms ''Field) > > instance HasCodec Field where > codec = > named Field $ > sumTypeCodec > [ TypeAlternative ColumnField "column" _ColumnField, > TypeAlternative RelationshipField "relationship" _RelationshipField > ]

newtype ValueWrapper (t :: Symbol) a

Constructors

ValueWrapper

Fields

getValue :: a

data ValueWrapper2 (t1 :: Symbol) a1 (t2 :: Symbol) a2

Constructors

ValueWrapper2

Fields

getValue1 :: a1

getValue2 :: a2

data ValueWrapper3 (t1 :: Symbol) a1 (t2 :: Symbol) a2 (t3 :: Symbol) a3

Constructors

ValueWrapper3

Fields

getValue1_ :: a1

getValue2_ :: a2

getValue3_ :: a3