Control.Monad.Memoize

Documentation

class Monad m => MonadMemoize (m :: Type -> Type) where

Methods

memoizeOn

Arguments

:: (Ord a, Typeable a, Typeable p)

=> Name
A unique name used to identify the function being memoized. There isn’t really any metaprogramming going on here, we just use a Template Haskell Name as a convenient source for a static, unique identifier.

-> a
The value to use as the memoization key. It’s the caller’s responsibility to ensure multiple calls to the same function don’t use the same key.

-> m p

-> m p

Memoizes a parser constructor function for the extent of a single schema construction process. This is mostly useful for recursive parsers; see Note [Tying the knot] for more details.

The generality of the type here allows us to use this with multiple concrete parser types:

memoizeOn :: (MonadMemoize m, MonadParse n) => Name -> a -> m (Parser n b) -> m (Parser n b)
memoizeOn :: (MonadMemoize m, MonadParse n) => Name -> a -> m (FieldParser n b) -> m (FieldParser n b)

memoize :: (MonadMemoize m, Ord a, Typeable a, Typeable p) => Name -> (a -> m p) -> a -> m p

A wrapper around memoizeOn that memoizes a function by using its argument as the key.

data MemoizeT (m :: Type -> Type) a

runMemoizeT :: Monad m => MemoizeT m a -> m a