Data.Aeson.Extended

Documentation

class ToJSONKeyValue a where

Methods

toJSONKeyValue :: a -> (Key, Value)

class FromJSONKeyValue a where

Methods

parseJSONKeyValue :: (Key, Value) -> Parser a

class FromJSONWithContext ctx a | a -> ctx where

Similar to FromJSON, except the parser can also source data with which to construct a from a context ctx.

This can be useful if the a value contains some data that is not from the current piece of JSON (the Value). For example, some data from higher up in the overall JSON graph, or from some system context.

Methods

parseJSONWithContext :: ctx -> Value -> Parser a

(.=?) :: (ToJSON v, KeyValue kv) => Key -> Maybe v -> Maybe kv infixr 8

An optional key-value pair for encoding a JSON object.

object $ ["foo" .= 0] <> catMaybes [ "bar" .=? Nothing, "baz" .=? 2 ]

mapWithJSONPath :: (a -> Parser b) -> [a] -> Parser [b]

Map a Parser over a list, keeping the JSONPath context

encodeToStrictText :: ToJSON a => a -> Text