Data.Aeson.Kriti.Functions

Description

Module of reusable functions for Kriti transforms.

NOTE: This defines an alternative runKritiWith that includes the basicFunctions by default. You should probably invoke Kriti through this module rather than directly in order to make updating the functions available only require touching this module.

TODO: This should be added to the documentation and referenced in (for-example) REST Connectors once the documentation refactor project is complete.

Documentation

runKriti :: Text -> [(Text, Value)] -> Either SerializedError Value

runKriti attaches the basicFunctions by default NOTE: The error type is SerializedError due to KritiError not currently being exported

runKritiWith :: Text -> [(Text, Value)] -> HashMap Text KritiFunc -> Either SerializedError Value

runKritiWith attaches the basicFunctions by default.

basicFunctions :: HashMap Text KritiFunc

Re-Export of the Kriti stdlib

environmentFunctions :: Environment -> HashMap Text KritiFunc

Functions that interact with environment variables

sessionFunctions :: Maybe SessionVariables -> HashMap Text KritiFunc

Functions that interact with HGE session during requests