Data.Aeson.Ordered

Description

A version of aeson that parses with key order preserved.

Copyright: (c) 2011-2016 Bryan O'Sullivan (c) 2011 MailRank, Inc.

Documentation

newtype Object

A JSON "object" (key/value map). This is where this type differs to the aeson package.

Constructors

Object_

Fields

unObject_ :: InsOrdHashMap Text Value

safeUnion :: Object -> Object -> Either String Object

Union the keys, ordered, in two maps, erroring on duplicates.

empty :: Object

Empty object.

insert :: (Int, Text) -> Value -> Object -> Object

Insert before the element at index i. Think of it in terms of splitAt, which is (take k, drop k). Deletes existing key, if any.

lookup :: Text -> Object -> Maybe Value

Lookup a key.

delete :: Text -> Object -> Object

Delete a key.

adjust :: (Value -> Value) -> Text -> Object -> Object

toList :: Object -> [(Text, Value)]

ToList a key.

fromList :: [(Text, Value)] -> Object

FromList a key.

type Array = Vector Value

A JSON "array" (sequence).

data Value

A JSON value represented as a Haskell value. Intentionally shadowing the Value from the aeson package.

Constructors

Object !Object

Array !Array

String !Text

Number !Scientific

Bool !Bool

Null

object :: [(Text, Value)] -> Value

Value pairs to Value

array :: [Value] -> Value

Value list to Value

toOrdered :: ToJSON a => a -> Value

Convert Aeson Value to Ordered Value

fromOrdered :: Value -> Value

Convert Ordered Value to Aeson Value

asObject :: IsString s => Value -> Either s Object

eitherDecode :: ByteString -> Either String Value

decode :: ByteString -> Maybe Value

arrayValues :: Parser Array

objectValues :: Parser (InsOrdHashMap Text Value)

value :: Parser Value

skipSpace :: Parser ()

The only valid whitespace in a JSON document is space, newline, carriage return, and tab.