Data.HashMap.Strict.InsOrd.Extended

Documentation

partition :: Hashable k => (v -> Bool) -> InsOrdHashMap k v -> (InsOrdHashMap k v, InsOrdHashMap k v)

alterF :: (Functor f, Hashable k) => (Maybe v -> f (Maybe v)) -> k -> InsOrdHashMap k v -> f (InsOrdHashMap k v)

Alter a hashmap using a function that can fail, in which case the entire operation fails. (Maybe a version with the key also being passed to the function could be useful.)

Orphan instances

Filterable (InsOrdHashMap k)

Instance details

Methods

mapMaybe :: (a -> Maybe b) -> InsOrdHashMap k a -> InsOrdHashMap k b

catMaybes :: InsOrdHashMap k (Maybe a) -> InsOrdHashMap k a

filter :: (a -> Bool) -> InsOrdHashMap k a -> InsOrdHashMap k a