Data.HashMap.Strict.Multi

Description

HashMap to multiple values.

Documentation

data MultiMap k v

Map from keys to sets of values

singleton :: Hashable k => k -> v -> MultiMap k v

Construct a MmultiMap with a single key, to which only one value is associated.

fromMap :: HashMap k (Set v) -> MultiMap k v

Construct a MultiMap with the supplied mappings.

toMap :: MultiMap k v -> HashMap k (Set v)

Convert a MultiMap to a HashMap.

fromList :: (Eq k, Hashable k, Ord v) => [(k, v)] -> MultiMap k v

Creates a MultiMap from an association list.

If the provided list constains duplicate mappings, the resulting MultiMap will store the set of all mapped values for each duplicate key.

toList :: MultiMap k v -> [(k, [v])]

Creates an association list from a MultiMap.

Each set of values associated with a given key is transformed back into a list.

lookup :: (Eq k, Hashable k) => k -> MultiMap k v -> Set v

Return the value to which the specified key is mapped, or Nothing if this map contains no mapping for the key.

insert :: (Eq k, Hashable k, Ord v) => k -> v -> MultiMap k v -> MultiMap k v

Associate the specified value with the specified key in this map.

If this map previously contained a mapping for the key, the new value is inserted in the set, and does not replace the previous mapping.

keys :: MultiMap k v -> [k]

Returns a list of this map's keys.

elems :: MultiMap k v -> [Set v]

Returns a list of this map's set of values.