Data.HashMap.Strict.NonEmpty

Description

Non-empty hash maps.

Documentation

newtype NEHashMap k v

A non-empty hashmap is a wrapper around a normal hashmap, that only provides a restricted set of functionalities. It doesn't provide a Monoid instance, nor an empty function.

Constructors

NEHashMap

Fields

unNEHashMap :: HashMap k v

singleton :: Hashable k => k -> v -> NEHashMap k v

Construct a non-empty map with a single element.

fromHashMap :: HashMap k v -> Maybe (NEHashMap k v)

Construct a non-empty map with the supplied mappings. Returns Nothing if the provided HashMap is empty.

fromList :: (Eq k, Hashable k) => [(k, v)] -> Maybe (NEHashMap k v)

Construct a non-empty map with the supplied mappings as follows:

if the provided list contains duplicate mappings, the later mappings take precedence;

if the provided list is empty, returns Nothing.

toHashMap :: NEHashMap k v -> HashMap k v

Convert a non-empty map to a HashMap.

lookup :: (Eq k, Hashable k) => k -> NEHashMap k v -> Maybe v

Return the value to which the specified key is mapped, or Nothing if this map contains no mapping for the key.

(!?) :: (Eq k, Hashable k) => NEHashMap k v -> k -> Maybe v

Return the value to which the specified key is mapped, or Nothing if this map contains no mapping for the key.

This is a flipped version of lookup.

keys :: NEHashMap k v -> [k]

Return a list of this map's keys.

unionWith :: (Eq k, Hashable k) => (v -> v -> v) -> NEHashMap k v -> NEHashMap k v -> NEHashMap k v

The union of two maps.

If a key occurs in both maps, the provided function (first argument) will be used to compute the result.

mapKeys :: (Eq k2, Hashable k2) => (k1 -> k2) -> NEHashMap k1 v -> NEHashMap k2 v

mapKeys f s is the map obtained by applying f to each key of s.

The size of the result may be smaller if f maps two or more distinct keys to the same new key. In this case there is no guarantee which of the associated values is chosen for the conflicting key.