Data.SqlCommenter

Description

Add metadata tags as comments to SQL queries.

Documentation

sqlCommenterGoogle :: QueryTagsAttributes -> QueryTagsComment

query-tags format as defined in the Spec https://google.github.io/sqlcommenter/spec/#sql-commenter

sqlCommenterStandard :: QueryTagsAttributes -> QueryTagsComment

Default 'Query Tags' format in the Hasura GraphQL Engine Creates simple 'key=value' pairs of query tags. No sorting, No URL encoding. If the format of query tags is not mentioned in the metadata then, query-tags are formatted using hte below format

generateCommentTags :: NonEmpty Attribute -> Text

Top-level algorithm to generate the string comment from list of Attributes Spec https://google.github.io/sqlcommenter/spec/#sql-commenter See Note [Ambiguous SQLCommenterGoogle Specification]