Data.Text.Casing

Description

This module is inspired by casing package. Instead of String this package uses Data.Text.Text

PascalCase - no spacing between words, first letter in word is uppercase, all others are lowercase.

camelCase - like PascalCase, but the very first letter is lowercase.

snake_Case - underscores delimit words, case is unrestricted.

Documentation

type NameWithOrigin = (Name, NameOrigin)

type NameSuffixWithOrigin = (NameSuffix, NameOrigin)

data GQLNameIdentifier

An opaque type, representing a parsed identifier with prefix and suffixes.

Constructors

GQLNameIdentifier

Fields

namePrefix :: NameWithOrigin

nameSuffixes :: [NameSuffixWithOrigin]

data NameOrigin

Represents the origin of a name entity.

CustomName represents a custom user provided name

AutogeneratedName represents a name which is generated by Hasura

For a custom table name foo, the select by pk field name elements:

foo is a CustomName

by is an AutogeneratedName

pk is an AutogeneratedName

However, for a table name foo_bar, the select by pk field name elements foo, bar, by and pk are all AutogeneratedName

Constructors

CustomName

AutogeneratedName

fromTupleWith :: NameOrigin -> (Name, [NameSuffix]) -> GQLNameIdentifier

fromNameWith :: NameOrigin -> Name -> GQLNameIdentifier

fromAutogeneratedName :: Name -> GQLNameIdentifier

fromCustomName :: Name -> GQLNameIdentifier

fromAutogeneratedTuple :: (Name, [NameSuffix]) -> GQLNameIdentifier

fromCustomTuple :: (Name, [NameSuffix]) -> GQLNameIdentifier

fromNonEmptyList :: NonEmpty NameWithOrigin -> GQLNameIdentifier

transformNameWith :: (Text -> Text) -> Name -> Name

transforms a graphql name with a transforming function

Note: This will return the graphql name without transformation if the transformed name is not a valid GraphQL identifier

transformNameAndOriginWith :: (Text -> Text) -> NameWithOrigin -> NameWithOrigin

same as transformNameWith but will not transform if the name is a custom name

transformGQLSuffixWith :: (Text -> Text) -> NameSuffix -> NameSuffix

similar to transformNameWith but transforms NameSuffix instead of Name

transformGQLSuffixAndOriginWith :: (Text -> Text) -> NameSuffixWithOrigin -> NameSuffixWithOrigin

This is essentially same as second transformGQLSuffixWith

transformGQLIdentifierWith :: (Text -> Text) -> GQLNameIdentifier -> GQLNameIdentifier

transforms a GQLNameIdentifier; this will apply the transformations on prefix as well as suffixes

identifierToList :: GQLNameIdentifier -> [Text]

converts identifiers to Text (i.e. unNames and unNameSuffixs identifiers)

toSnakeT :: [Text] -> Text

To snake_case for Data.Text

>>> toSnakeT ["my","random","text","list"]
"my_random_text_list"

toPascalT :: [Text] -> Text

To PascalCase for Data.Text

>>> toPascalT ["my","random","text","list"]
"MyRandomTextList"

toCamelT :: [Text] -> Text

To camelCase for Data.Text

>>> toCamelT ["my","random","text","list"]
"myRandomTextList"

toSnakeG :: GQLNameIdentifier -> Name

To snake_case for GQLNameIdentifier

toPascalG :: GQLNameIdentifier -> Name

To PascalCase for GQLNameIdentifier

toCamelG :: GQLNameIdentifier -> Name

To camelCase for GQLNameIdentifier

transformPrefixAndSuffixAndConcat :: GQLNameIdentifier -> (Text -> Text) -> (Text -> Text) -> Name

Transforms GQLNameIdentifier and returns a G.Name

fromSnake :: Text -> [Text]

Convert from snake_cased

>>> fromSnake "_hello_world_foo"
["_hello","world","foo"]

snakeToPascal :: Text -> Text

Directly convert to PascalCase through fromSnake

snakeToCamel :: Text -> Text

Directly convert to camelCase through fromSnake

lowerFirstChar :: Text -> Text

An internal helper function to lowercase the first character

upperFirstChar :: Text -> Text

An internal helper function to uppercase the first character