Data.Time.Clock.Units

Description

Types for time intervals of various units. Each newtype wraps DiffTime, but they have different Num instances. The intent is to use the record selectors to write literals with particular units, like this:

>>> milliseconds 500
0.5s
>>> hours 3
10800s
>>> minutes 1.5 + seconds 30
120s

You can also go the other way using the constructors rather than the selectors:

>>> toRational $ Minutes (seconds 17)
17 % 60
>>> realToFrac (Days (hours 12)) :: Double
0.5

NOTE: the Real and Fractional instances just essentially add or strip the unit label (as above), so you can't use realToFrac to convert between the units types here. Instead try convertDuration which is less of a foot-gun.

The Read instances for these types mirror the behavior of the RealFrac instance wrt numeric literals for convenient serialization (e.g. when working with env vars):

>>> read "1.2" :: Milliseconds
Milliseconds {milliseconds = 0.0012s}

Generally, if you need to pass around a duration between functions you should use DiffTime directly. However if storing a duration in a type that will be serialized, e.g. one having a ToJSON instance, it is better to use one of these explicit wrapper types so that it's obvious what units will be used.

Documentation

newtype Seconds

Constructors

Seconds

Fields

seconds :: DiffTime

newtype Days

Constructors

Days

Fields

days :: DiffTime

newtype Hours

Constructors

Hours

Fields

hours :: DiffTime

newtype Minutes

Constructors

Minutes

Fields

minutes :: DiffTime

newtype Milliseconds

Constructors

Milliseconds

Fields

milliseconds :: DiffTime

newtype Microseconds

Constructors

Microseconds

Fields

microseconds :: DiffTime

newtype Nanoseconds

Constructors

Nanoseconds

Fields

nanoseconds :: DiffTime

newtype TimeUnit (picosPerUnit :: Nat)

Constructors

TimeUnit DiffTime

type SecondsP n = n * 1000000000000

natNum :: forall n a. (KnownNat n, Num a) => a

class Duration d where

Duration types isomorphic to DiffTime, powering convertDuration.

Methods

fromDiffTime :: DiffTime -> d

toDiffTime :: d -> DiffTime

convertDuration :: (Duration x, Duration y) => x -> y

Safe conversion between duration units.

diffTimeToMicroSeconds :: DiffTime -> Integer