Data.Trie

Description

Prefix trees on arbitrary keys.

Documentation

data Trie k v

Data structure for storing a value v keyed on a sequence of ks

Constructors

Trie

Fields

trieMap :: HashMap k (Trie k v)

trieData :: Maybe v

empty :: Trie k v

Construct an empty trie.

singleton :: Hashable k => [k] -> v -> Trie k v

Creates a trie from a path and a value

>>> singleton ["a", "b"] 5
Trie (fromList [("a", Trie (fromList [("b", Trie (fromList []) (Just 5))]) Nothing)]) Nothing

>>> singleton [] 5
Trie (fromList []) (Just 5)

lookup :: (Eq k, Hashable k) => [k] -> Trie k v -> Maybe v

Find a value at the given path, if any.

insert :: (Eq k, Hashable k) => [k] -> v -> Trie k v -> Trie k v

Insert the given value at the given path.

If there's already a value at the given path, it is replaced.

insertWith :: (Eq k, Hashable k) => (v -> v -> v) -> [k] -> v -> Trie k v -> Trie k v

Insert the value at the given path.

If there's already a value at the given path, the old value is replaced by the result of applying the given function to the new and old value.

elems :: Trie k v -> [v]

Extract all values of the trie, discarding any path information.