Database.MSSQL.Transaction

Documentation

newtype TxET e m a

The transaction command to run, parameterised over: e - the exception type (usually MSSQLTxError) m - some Monad, (usually some MonadIO) a - the successful result type

Constructors

TxET

Fields

txHandler :: ReaderT Connection (ExceptT e m) a

data MSSQLTxError

Error type generally used in TxET.

Constructors

MSSQLQueryError !Query !ODBCException

MSSQLConnError !ODBCException

MSSQLInternal !Text

type TxE e a = TxET e IO a

type TxT m a = TxET MSSQLTxError m a

The transaction command to run, returning an MSSQLTxError or the result.

runTx :: (MonadIO m, MonadBaseControl IO m) => TxIsolation -> TxT m a -> MSSQLPool -> ExceptT MSSQLTxError m a

Run a command on the given connection wrapped in a transaction.

See runTxE if you need to map the error type as well.

runTxE :: (MonadIO m, MonadBaseControl IO m) => (MSSQLTxError -> e) -> TxIsolation -> TxET e m a -> MSSQLPool -> ExceptT e m a

Run a command on the given connection wrapped in a transaction.

unitQuery :: MonadIO m => Query -> TxT m ()

Useful for building transactions which return no data.

insertId :: TxT m ()
insertId = unitQuery "INSERT INTO some_table VALUES (1, "hello")"

See unitQueryE if you need to map the error type as well.

unitQueryE :: MonadIO m => (MSSQLTxError -> e) -> Query -> TxET e m ()

Useful for building transactions which return no data.

singleRowQuery :: forall a m. (MonadIO m, FromRow a) => Query -> TxT m a

Useful for building query transactions which return a single one row.

returnOne :: TxT m Int
returnOne = singleRowQuery "SELECT 1"

See singleRowQueryE if you need to map the error type as well.

singleRowQueryE :: forall m a e. (MonadIO m, FromRow a) => (MSSQLTxError -> e) -> Query -> TxET e m a

Useful for building query transactions which return a single one row.

forJsonQueryE :: forall m e. MonadIO m => (MSSQLTxError -> e) -> Query -> TxET e m Text

MSSQL splits up results that have a SELECT .. FOR JSON at the top-level into multiple rows with a single column, see https://docs.microsoft.com/en-us/sql/relational-databases/json/format-query-results-as-json-with-for-json-sql-server?view=sql-server-ver15#output-of-the-for-json-clause

This function simply concatenates each single-column row into one long Text string.

multiRowQuery :: forall a m. (MonadIO m, FromRow a) => Query -> TxT m [a]

Useful for building query transactions which return multiple rows.

selectIds :: TxT m [Int]
selectIds = multiRowQuery "SELECT id FROM author"

See multiRowQueryE if you need to map the error type as well.

multiRowQueryE :: forall m a e. (MonadIO m, FromRow a) => (MSSQLTxError -> e) -> Query -> TxET e m [a]

Useful for building query transactions which return multiple rows.

buildGenericQueryTxE

Arguments

:: MonadIO m

=> (MSSQLTxError -> e)
map MSSQLTxError to some other type

-> query
query to run

-> (query -> Query)
how to map a query to a Query

-> (Connection -> query -> IO a)
run the query on a provided Connection

-> TxET e m a

Build a generic transaction out of an IO action.

withTxET :: Monad m => (e1 -> e2) -> TxET e1 m a -> TxET e2 m a

Map the error type for a TxET.

data TxIsolation

https://learn.microsoft.com/en-us/sql/t-sql/statements/set-transaction-isolation-level-transact-sql

Constructors

ReadUncommitted

ReadCommitted

RepeatableRead

Snapshot

Serializable