GHC.Generics.Extended

Documentation

constrName :: (HasConstructor (Rep a), Generic a) => a -> String

Return the constructor name of some data with a Generic instance, as a string. Useful when constructing error or trace messages, or debugging.