Hasura.App.State

Documentation

data AppEnv

Represents the Read-Only Hasura State, these fields are immutable and the state cannot be changed during runtime.

Constructors

AppEnv

Fields

appEnvPort :: Port

appEnvHost :: HostPreference

appEnvMetadataDbPool :: PGPool

appEnvManager :: Manager

appEnvLoggers :: Loggers

appEnvMetadataVersionRef :: TMVar MetadataResourceVersion

appEnvInstanceId :: InstanceId

appEnvEnableMaintenanceMode :: MaintenanceMode ()

appEnvLoggingSettings :: LoggingSettings

appEnvEventingMode :: EventingMode

appEnvEnableReadOnlyMode :: ReadOnlyMode

appEnvServerMetrics :: ServerMetrics

appEnvShutdownLatch :: ShutdownLatch

appEnvMetaVersionRef :: TMVar MetadataResourceVersion

appEnvPrometheusMetrics :: PrometheusMetrics

appEnvTraceSamplingPolicy :: SamplingPolicy

appEnvSubscriptionState :: SubscriptionsState

appEnvLockedEventsCtx :: LockedEventsCtx

appEnvConnParams :: ConnParams

appEnvTxIso :: TxIsolation

appEnvConsoleAssetsDir :: Maybe Text

appEnvConsoleSentryDsn :: Maybe Text

appEnvConnectionOptions :: ConnectionOptions

appEnvWebSocketKeepAlive :: KeepAliveDelay

appEnvWebSocketConnectionInitTimeout :: WSConnectionInitTimeout

appEnvGracefulShutdownTimeout :: Refined NonNegative Seconds

appEnvCheckFeatureFlag :: FeatureFlag -> IO Bool

data AppContext

Represents the Dynamic Hasura State, these field are mutable and can be changed during runtime.

Constructors

AppContext

Fields

acCacheRef :: SchemaCacheRef

acAuthMode :: AuthMode

acSQLGenCtx :: SQLGenCtx

acEnabledAPIs :: HashSet API

acEnableAllowlist :: AllowListStatus

acResponseInternalErrorsConfig :: ResponseInternalErrorsConfig

acEnvironment :: Environment

acRemoteSchemaPermsCtx :: RemoteSchemaPermissions

acFunctionPermsCtx :: InferFunctionPermissions

acExperimentalFeatures :: HashSet ExperimentalFeature

acDefaultNamingConvention :: NamingCase

acMetadataDefaults :: MetadataDefaults

acLiveQueryOptions :: LiveQueriesOptions

acStreamQueryOptions :: StreamQueriesOptions

acCorsConfig :: CorsConfig

acConsoleStatus :: ConsoleStatus

acEnableTelemetry :: TelemetryStatus

acEventsHttpPoolSize :: Refined Positive Int

acEventsFetchInterval :: Refined NonNegative Milliseconds

acEventsFetchBatchSize :: Refined NonNegative Int

acAsyncActionsFetchInterval :: OptionalInterval

acSchemaPollInterval :: OptionalInterval

data Loggers

Collection of the LoggerCtx, the regular Logger and the PGLogger

Constructors

Loggers

Fields

_lsLoggerCtx :: LoggerCtx Hasura

_lsLogger :: Logger Hasura

_lsPgLogger :: PGLogger