Hasura.Backends.BigQuery.DDL.ComputedField

Description

BigQuery DDL ComputedField

Implementation to build ComputedFieldInfo for a BigQuery table from metadata

Documentation

data ComputedFieldError

Errors that occur when validating and building computed fields.

Constructors

CFENoArgumentType FunctionArgName

CFENotTableValuedFunction

CFENoInputArguments

CFENoArgumentName

CFEInvalidArgumentName FunctionArgName

CFEInvalidColumnName TableName ColumnName

CFEReturnTableNotTracked TableName

CFENeedReturnTableName

CFENotRelevantReturnTable TableName

CFEReturnTableSchemaError ReturnTableSchemaError

data ReturnTableSchemaError

Errors that occur when validating returning table schema fields

Constructors

RTSENoFieldName

RTSENoType Text

RTSENotValidGraphQLName Text

showError :: FunctionName -> ComputedFieldError -> Text

Generate read-able error message

buildComputedFieldInfo :: forall m. MonadError QErr m => HashSet TableName -> TableName -> HashSet ColumnName -> ComputedFieldName -> ComputedFieldDefinition -> RestRoutine -> Comment -> m (ComputedFieldInfo 'BigQuery)

Validate computed field metadata and build field information