Hasura.Backends.BigQuery.Execute

Description

Execute a Select query against the BigQuery REST API.

Documentation

data RecordSet

A set of records produced by the database. These are joined together. There are all sorts of optimizations possible here, from using a matrix/flat vector, unboxed sums for Value, etc. Presently we choose a naive implementation in the interest of getting other work done.

Constructors

RecordSet

Fields

rows :: !(Vector (InsOrdHashMap FieldNameText OutputValue))

wantedFields :: !(Maybe [Text])

newtype FieldNameText

As opposed to BigQuery.FieldName which is a qualified name, this is just the unqualified text name itself.

Constructors

FieldNameText Text

data OutputValue

Constructors

DecimalOutputValue !Decimal

BigDecimalOutputValue !BigDecimal

IntegerOutputValue !Int64

FloatOutputValue !Float64

GeographyOutputValue !Geography

TextOutputValue !Text

TimestampOutputValue !Timestamp

DateOutputValue !Date

TimeOutputValue !Time

DatetimeOutputValue !Datetime

BytesOutputValue !Base64

BoolOutputValue !Bool

ArrayOutputValue !(Vector OutputValue)

RecordOutputValue !(InsOrdHashMap FieldNameText OutputValue)

NullOutputValue

data ExecuteReader

Constructors

ExecuteReader

Fields

sourceConfig :: !BigQuerySourceConfig

data ExecuteProblem

Constructors

GetJobDecodeProblem String

CreateQueryJobDecodeProblem String

InsertDatasetDecodeProblem String

ExecuteRunBigQueryProblem BigQueryProblem

RESTRequestNonOK Status Value

data ShowDetails

We use this to hide certain details from the front-end, while allowing them in tests. We have not actually decided whether showing the details is insecure, but until we decide otherwise, it's probably best to err on the side of caution.

Constructors

HideDetails

InsecurelyShowDetails

executeProblemMessage :: ShowDetails -> ExecuteProblem -> Text

newtype Execute a

Execute monad; as queries are performed, the record sets are stored in the map.

Constructors

Execute

Fields

unExecute :: ReaderT ExecuteReader (ExceptT ExecuteProblem IO) a

data BigQueryType

Big query parameters must be accompanied by an explicit type signature.

Constructors

DECIMAL

INTEGER

FLOAT

BYTES

STRING

BOOL

ARRAY BigQueryType

GEOGRAPHY

DATE

TIMESTAMP

DATETIME

TIME

BIGDECIMAL

data BigQuery

Constructors

BigQuery

Fields

query :: !Text

parameters :: !(InsOrdHashMap ParameterName Parameter)

data Parameter

Constructors

Parameter

Fields

typ :: !BigQueryType

value :: !Value

newtype ParameterName

Constructors

ParameterName Text

data BigQueryField

Constructors

BigQueryField

Fields

name :: !FieldNameText

typ :: !BigQueryFieldType

mode :: !Mode

data BigQueryFieldType

Constructors

FieldSTRING

FieldBYTES

FieldINTEGER

FieldFLOAT

FieldBOOL

FieldTIMESTAMP

FieldDATE

FieldTIME

FieldDATETIME

FieldGEOGRAPHY

FieldDECIMAL

FieldBIGDECIMAL

FieldSTRUCT (Vector BigQueryField)

data Mode

Constructors

Nullable

NotNullable

Repeated

data IsNullable

Constructors

IsNullable

IsRequired

streamDelaySeconds :: DiffTime

Delay between attempts to get job results if the job is incomplete.

bigQueryProjectUrl :: Text -> String

runExecute :: MonadIO m => BigQuerySourceConfig -> Execute RecordSet -> m (Either ExecuteProblem RecordSet)

executeSelect :: Select -> Execute RecordSet

getFinalRecordSet :: RecordSet -> Execute RecordSet

This is needed to strip out unneeded fields (join keys) in the final query. This is a relic of the data loader approach. A later improvement would be to update the FromIr code to explicitly reselect the query. But the purpose of this commit is to drop the dataloader code and not modify the from IR code which is more delicate.

selectToBigQuery :: Select -> BigQuery

valueType :: Value -> BigQueryType

Make a BigQuery type for the given value.

valueToBigQueryJson :: Value -> Value

Make a JSON representation of the type of the given value.

streamBigQuery :: MonadIO m => BigQueryConnection -> BigQuery -> m (Either ExecuteProblem RecordSet)

TODO: WARNING: This function hasn't been tested on Big Data(tm), and therefore I was unable to get BigQuery to produce paginated results that would contain the $sel:pageToken:JobResults field in the JSON response. Until that test has been done, we should consider this a preliminary implementation.

executeBigQuery :: MonadIO m => BigQueryConnection -> BigQuery -> m (Either ExecuteProblem ())

Execute a query without expecting any output (e.g. CREATE TABLE or INSERT)

data JobResults

Constructors

JobResults

Fields

pageToken :: Maybe Text

recordSet :: RecordSet

data JobResultsResponse

Constructors

JobIncomplete

JobComplete JobResults

data Fetch

Constructors

Fetch

Fields

pageToken :: Maybe Text

getJobResults :: MonadIO m => BigQueryConnection -> Job -> Fetch -> m (Either ExecuteProblem JobResultsResponse)

Get results of a job.

data Job

Constructors

Job

Fields

state :: !Text

jobId :: !Text

location :: !Text

jsonRequestHeader :: Request -> Request

Make a Request return JSON

createQueryJob :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> BigQuery -> m Job

Create a job asynchronously.

data Dataset

Constructors

Dataset

Fields

datasetId :: Text

deleteDataset :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Text -> m ()

Delete a dataset

runBigQueryExcept :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Request -> m (Response ByteString)

Run request and map errors into ExecuteProblem

insertDataset :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Text -> m Dataset

Insert a new dataset

parseAsJsonOrText :: ByteString -> Value

Parse given @ByteString as JSON value. If not a valid JSON, encode to plain text.

parseRecordSetPayload :: Object -> Parser RecordSet

parseRow :: Vector BigQueryField -> Value -> Parser (InsOrdHashMap FieldNameText OutputValue)

parseBigQueryRow :: Vector BigQueryField -> Value -> Parser OutputValue

Parse a row, which at the top-level of the "rows" output has no {"v":..} wrapper. But when appearing nestedly, does have the wrapper. See parseBigQueryValue.

parseBigQueryValue :: IsNullable -> BigQueryFieldType -> Value -> Parser OutputValue

parseTimestamp :: Value -> Parser Timestamp

Parse upstream timestamp value in epoch milliseconds and convert it to calendar date time format https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#timestamp_type

parseBigQueryField :: BigQueryField -> Value -> Parser (FieldNameText, OutputValue)

has_v :: IsNullable -> (Value -> Parser OutputValue) -> Value -> Parser OutputValue

has_v_generic :: (Value -> Parser a) -> Value -> Parser a