Hasura.Backends.BigQuery.Parser.Scalars

Description

This module defines the scalars we use specific to the BigQuery schema.

An idiosyncracy of BigQuery is that numbers serialized via JSON uses string literals instead of number literals, because BigQuery handles wider-bit numbers than JSON/JavaScript does.

Therefore, the BigQuery Backend uses bespoke parsers for numeric scalar input, which accept string literals as well as number literals, such that we preserve symmetry with with output formats.

Documentation

bqInt64 :: forall origin m. MonadParse m => Parser origin 'Both m Int64

bqFloat64 :: forall origin m. MonadParse m => Parser origin 'Both m Float64

bqBigDecimal :: MonadParse m => Parser origin 'Both m BigDecimal

bqDecimal :: MonadParse m => Parser origin 'Both m Decimal

decimal :: MonadParse f => Name -> InputValue Variable -> f Scientific

mkScalar :: MonadParse m => Name -> Description -> (InputValue Variable -> m a) -> Parser origin 'Both m a

typeNamed :: Name -> Maybe Description -> Type origin 'Both

stringNotationError :: MonadParse m => Name -> Text -> m a