Hasura.Backends.BigQuery.Source

Documentation

data PKey

Constructors

PKey

Fields

unPKey :: PrivateKey

originalBS :: Text

newtype GoogleAccessToken

Constructors

GoogleAccessToken Text

data TokenResp

Constructors

TokenResp

Fields

_trAccessToken :: !GoogleAccessToken

_trExpiresAt :: !Integer

data ServiceAccount

Constructors

ServiceAccount

Fields

_saClientEmail :: !Text

_saPrivateKey :: !PKey

_saProjectId :: !Text

data ConfigurationJSON a

Constructors

FromEnvJSON Text

FromYamlJSON a

data ConfigurationInputs

Configuration inputs when they are a YAML array or an Env var whos value is a comma-separated string

Constructors

FromYamls ![Text]

FromEnvs !Text

data ConfigurationInput

Configuration input when the YAML value as well as the Env var have singlular values

Constructors

FromYaml !Text

FromEnv !Text

data BigQueryConnSourceConfig

Constructors

BigQueryConnSourceConfig

Fields

_cscServiceAccount :: ConfigurationJSON ServiceAccount

_cscDatasets :: ConfigurationInputs

_cscProjectId :: ConfigurationInput

_cscGlobalSelectLimit :: Maybe ConfigurationInput

_cscRetryBaseDelay :: Maybe ConfigurationInput

_cscRetryLimit :: Maybe ConfigurationInput

data RetryOptions

Constructors

RetryOptions

Fields

_retryBaseDelay :: Microseconds

_retryNumRetries :: Int

data BigQueryConnection

Constructors

BigQueryConnection

Fields

_bqServiceAccount :: ServiceAccount

_bqProjectId :: Text

_bqRetryOptions :: Maybe RetryOptions

_bqAccessTokenMVar :: MVar (Maybe TokenResp)

data BigQuerySourceConfig

Constructors

BigQuerySourceConfig

Fields

_scConnection :: BigQueryConnection

_scDatasets :: [Text]

_scGlobalSelectLimit :: Int64

Orphan instances

Generic PrivateKey

Instance details

Associated Types

type Rep PrivateKey :: Type -> Type

Methods

from :: PrivateKey -> Rep PrivateKey x

to :: Rep PrivateKey x -> PrivateKey

Generic PublicKey

Instance details

Associated Types

type Rep PublicKey :: Type -> Type

Methods

from :: PublicKey -> Rep PublicKey x

to :: Rep PublicKey x -> PublicKey

Hashable PrivateKey

Instance details

Methods

hashWithSalt :: Int -> PrivateKey -> Int

hash :: PrivateKey -> Int

Hashable PublicKey

Instance details

Methods

hashWithSalt :: Int -> PublicKey -> Int

hash :: PublicKey -> Int

ToJSON PrivateKey

Instance details

Methods

toJSON :: PrivateKey -> Value

toEncoding :: PrivateKey -> Encoding

toJSONList :: [PrivateKey] -> Value

toEncodingList :: [PrivateKey] -> Encoding

ToJSON PublicKey

Instance details

Methods

toJSON :: PublicKey -> Value

toEncoding :: PublicKey -> Encoding

toJSONList :: [PublicKey] -> Value

toEncodingList :: [PublicKey] -> Encoding