Hasura.Backends.BigQuery.ToQuery

Description

Convert the simple BigQuery AST to an SQL query, ready to be passed to the odbc package's query/exec functions.

Documentation

data Printer

Constructors

SeqPrinter [Printer]

SepByPrinter Printer [Printer]

NewlinePrinter

UnsafeTextPrinter Text

IndentPrinter Int Printer

ValuePrinter TypedValue

fromExpression :: Expression -> Printer

fromSelect :: Select -> Printer

fromReselect :: Reselect -> Printer

toBuilderFlat :: Printer -> Builder

toBuilderPretty :: Printer -> Builder

toTextPretty :: Printer -> Text

toTextFlat :: Printer -> Text

renderBuilderFlat :: Printer -> (Builder, InsOrdHashMap Int TypedValue)

Produces a query with holes, and a mapping for each

renderBuilderPretty :: Printer -> (Builder, InsOrdHashMap Int TypedValue)

Produces a query with holes, and a mapping for each

paramName :: Int -> Builder

Orphan instances

ToJSON Expression

Instance details

Methods

toJSON :: Expression -> Value

toEncoding :: Expression -> Encoding

toJSONList :: [Expression] -> Value

toEncodingList :: [Expression] -> Encoding