Hasura.Backends.BigQuery.Types

Description

Types for BigQuery

Documentation

data Select

Constructors

Select

Fields

selectWith :: Maybe With

selectTop :: Top

selectAsStruct :: AsStruct

selectProjections :: NonEmpty Projection

selectFrom :: From

selectJoins :: [Join]

selectWhere :: Where

selectOrderBy :: Maybe (NonEmpty OrderBy)

selectOffset :: Maybe Expression

selectGroupBy :: [FieldName]

selectFinalWantedFields :: Maybe [Text]

selectCardinality :: Cardinality

data PartitionableSelect

Helper type allowing addition of extra fields used in PARTITION BY.

The main purpose of this type is sumulation of DISTINCT ON implemented in Hasura.Backends.BigQuery.FromIr.simulateDistinctOn

Constructors

PartitionableSelect

Fields

pselectFinalize :: Maybe [FieldName] -> Select

pselectFrom :: From

simpleSelect :: Select -> PartitionableSelect

noExtraPartitionFields :: PartitionableSelect -> Select

withExtraPartitionFields :: PartitionableSelect -> [FieldName] -> Select

data ArrayAgg

Constructors

ArrayAgg

Fields

arrayAggProjections :: NonEmpty Projection

arrayAggOrderBy :: Maybe (NonEmpty OrderBy)

arrayAggTop :: Top

data Reselect

Constructors

Reselect

Fields

reselectProjections :: NonEmpty Projection

reselectWhere :: Where

data OrderBy

Constructors

OrderBy

Fields

orderByExpression :: Expression

orderByFieldName :: FieldName

orderByOrder :: Order

orderByNullsOrder :: NullsOrder

data Order

Constructors

AscOrder

DescOrder

data NullsOrder

Constructors

NullsFirst

NullsLast

NullsAnyOrder

data FieldOrigin

Constructors

NoOrigin

AggregateOrigin [Aliased Aggregate]

aggregateProjectionsFieldOrigin :: Projection -> FieldOrigin

data Projection

Constructors

ExpressionProjection (Aliased Expression)

FieldNameProjection (Aliased FieldName)

AggregateProjections (Aliased (NonEmpty (Aliased Aggregate)))

AggregateProjection (Aliased Aggregate)

StarProjection

ArrayAggProjection (Aliased ArrayAgg)

EntityProjection (Aliased [(FieldName, FieldOrigin)])

ArrayEntityProjection EntityAlias (Aliased [FieldName])

WindowProjection (Aliased WindowFunction)

data WindowFunction

Constructors

RowNumberOverPartitionBy (NonEmpty FieldName) (Maybe (NonEmpty OrderBy))
ROW_NUMBER() OVER(PARTITION BY field)

data JoinType

Constructors

LeftOuter

Inner

data Join

Constructors

Join

Fields

joinSource :: JoinSource

joinAlias :: EntityAlias

joinOn :: [(FieldName, FieldName)]

joinProvenance :: JoinProvenance

joinFieldName :: Text

joinExtractPath :: Maybe Text

joinRightTable :: EntityAlias

joinType :: JoinType

data JoinProvenance

Constructors

OrderByJoinProvenance

ObjectJoinProvenance [Text]

ArrayAggregateJoinProvenance [(Text, FieldOrigin)]

ArrayJoinProvenance [Text]

data JoinSource

Constructors

JoinSelect Select

newtype Where

Constructors

Where [Expression]

data Cardinality

Constructors

Many

One

data AsStruct

Constructors

NoAsStruct

AsStruct

newtype With

A Common Table Expression clause.

Constructors

With (NonEmpty (Aliased (InterpolatedQuery Expression)))

data Top

Constructors

NoTop

Top Int64

data Expression

Constructors

ValueExpression TypedValue

InExpression Expression TypedValue

AndExpression [Expression]

OrExpression [Expression]

NotExpression Expression

ExistsExpression Select

SelectExpression Select

IsNullExpression Expression

IsNotNullExpression Expression

ColumnExpression FieldName

EqualExpression Expression Expression

NotEqualExpression Expression Expression

JsonQueryExpression Expression
This one acts like a "cast to JSON" and makes SQL Server behave like it knows your field is JSON and not double-encode it.

ToStringExpression Expression

JsonValueExpression Expression JsonPath
This is for getting actual atomic values out of a JSON string.

OpExpression Op Expression Expression

ListExpression [Expression]

CastExpression Expression ScalarType

FunctionExpression FunctionName [Expression]

ConditionalProjection Expression FieldName

FunctionNamedArgument Text Expression
A function input argument expression with argument name argument_name => argument_value

data JsonPath

Constructors

RootPath

FieldPath JsonPath Text

IndexPath JsonPath Integer

data Aggregate

Constructors

CountAggregate (Countable Expression)

OpAggregates Text (NonEmpty (Text, Expression))

OpAggregate Text Expression

TextAggregate Text

newtype CountType field

Constructors

CountType

Fields

getCountType :: Countable (ColumnName, AnnRedactionExp 'BigQuery field)

data Countable fieldname

Constructors

StarCountable

NonNullFieldCountable (NonEmpty fieldname)

DistinctCountable (NonEmpty fieldname)

data From

Constructors

FromQualifiedTable (Aliased TableName)

FromSelect (Aliased Select)

FromSelectJson (Aliased SelectJson)

FromFunction (Aliased SelectFromFunction)

FromNativeQuery (Aliased NativeQueryName)

data SelectJson

Constructors

SelectJson

Fields

selectJsonBody :: Expression

selectJsonFields :: [(ColumnName, ScalarType)]

data SelectFromFunction

Constructors

SelectFromFunction

Fields

sffFunctionName :: FunctionName

sffArguments :: [Expression]

data Aliased a

Constructors

Aliased

Fields

aliasedThing :: a

aliasedAlias :: Text

data TableName

Constructors

TableName

Fields

tableName :: Text

tableNameSchema :: Text

data FieldName

Constructors

FieldName

Fields

fieldName :: Text

fieldNameEntity :: Text

newtype ColumnName

Constructors

ColumnName Text

newtype EntityAlias

Constructors

EntityAlias

Fields

entityAliasText :: Text

columnToFieldName :: EntityAlias -> ColumnName -> FieldName

data Op

Constructors

LessOp

LessOrEqualOp

MoreOp

MoreOrEqualOp

InOp

NotInOp

LikeOp

NotLikeOp

data Value

Source for this represenation type:

https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.Value

BigQuery results come in via the REST API as one of these simply types.

TODO: This omits StructValue -- do we need it?

Constructors

NullValue

IntegerValue Int64
64-bit https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#integer_type

DecimalValue Decimal
Fixed precision https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#decimal_types

BigDecimalValue BigDecimal
Fixed precision https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#decimal_types

FloatValue Float64
Floating point https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#floating_point_types

GeographyValue Geography

StringValue Text

BytesValue Base64

BoolValue Bool

ArrayValue (Vector Value)

TimestampValue Timestamp

DateValue Date

TimeValue Time

JsonValue Value

DatetimeValue Datetime

newtype Timestamp

BigQuery's conception of a timestamp.

Constructors

Timestamp Text

newtype Date

BigQuery's conception of a date.

Constructors

Date Text

newtype Time

BigQuery's conception of a time.

Constructors

Time Text

newtype Datetime

BigQuery's conception of a datetime.

Constructors

Datetime Text

newtype Int64

BigQuery's conception of an INTEGER/INT64 (they are the same).

Constructors

Int64 Text

data TypedValue

Constructors

TypedValue

Fields

tvType :: ScalarType

tvValue :: Value

intToInt64 :: Int64 -> Int64

int64Expr :: Int64 -> Expression

newtype Decimal

BigQuery's conception of a fixed precision decimal.

Constructors

Decimal Text

scientificToText :: Scientific -> Text

Convert Scientific to Text

newtype BigDecimal

BigQuery's conception of a "big" fixed precision decimal.

Constructors

BigDecimal Text

doubleToBigDecimal :: Double -> BigDecimal

data Float64

BigQuery's conception of a fixed precision decimal.

doubleToFloat64 :: Double -> Float64

data Base64

A base-64 encoded binary string.

newtype Geography

Constructors

Geography Text

data ScalarType

Constructors

StringScalarType

BytesScalarType

IntegerScalarType

FloatScalarType

BoolScalarType

TimestampScalarType

DateScalarType

TimeScalarType

DatetimeScalarType

GeographyScalarType

DecimalScalarType

BigDecimalScalarType

JsonScalarType

StructScalarType

data BooleanOperators a

Constructors

ASTContains a

ASTEquals a

ASTTouches a

ASTWithin a

ASTIntersects a

ASTDWithin (DWithinGeogOp a)

data FunctionName

Constructors

FunctionName

Fields

functionName :: Text

functionNameSchema :: Maybe Text

System functions like "unnest" don't have schema/dataset

data ComputedFieldDefinition

The metadata required to define a computed field for a BigQuery table

Constructors

ComputedFieldDefinition

Fields

_bqcfdFunction :: FunctionName

Name of the user defined routine

_bqcfdReturnTable :: Maybe TableName

Name of the table which the function returns. If not provided the return table schema is inferred from the routine API metadata.

_bqcfdArgumentMapping :: HashMap FunctionArgName ColumnName

A mapping context to determine argument value from table column

data ArgumentExp v

A argument expression for SQL functions

Constructors

AEInput v
Value coming from user's input through GraphQL query

AETableColumn ColumnName
For computed fields, value of column from the table

type ComputedFieldImplicitArguments = HashMap FunctionArgName ColumnName

data ComputedFieldReturn

Returning type of the function underlying a computed field

Constructors

ReturnExistingTable TableName
Returns existing table, needs to be present in the metadata

ReturnTableSchema [(ColumnName, Name, ScalarType)]
An arbitrary table schema specified by column name and type pairs

data FunctionArgument

Function input argument specification

Constructors

FunctionArgument

Fields

_faName :: FunctionArgName

Argument name of a table valued function is required Ref: https://cloud.google.com/bigquery/docs/reference/standard-sql/data-definition-language#create_table_function_statement

_faType :: ScalarType

The data type of the argument

parseScalarValue :: ScalarType -> Value -> Either QErr Value

isComparableType :: ScalarType -> Bool

isNumType :: ScalarType -> Bool

getGQLTableName :: TableName -> Either QErr Name

projectionAlias :: Projection -> Maybe Text

data Job

Constructors

Job

Fields

state :: Text

jobId :: Text

location :: Text

data ExecutionStatistics

Constructors

ExecutionStatistics

Fields

_esJob :: Job