Hasura.Backends.DataConnector.Adapter.Types

Documentation

data ConnSourceConfig

Constructors

ConnSourceConfig

Fields

value :: Config

An arbitrary JSON payload to be passed to the agent in a header. HGE validates this against the OpenAPI Spec provided by the agent.

template :: Maybe Text

Kriti Template for transforming the supplied Config value.

timeout :: Maybe SourceTimeout

Timeout setting for HTTP requests to the agent. -- TODO: verify with lyndon

data SourceTimeout

sourceTimeoutMicroseconds :: SourceTimeout -> Int

data SourceConfig

Constructors

SourceConfig

Fields

_scEndpoint :: BaseUrl

_scConfig :: Config

_scTemplate :: Maybe Text

_scCapabilities :: Capabilities

_scManager :: Manager

_scTimeoutMicroseconds :: Maybe Int

_scDataConnectorName :: DataConnectorName

_scEnvironment :: Environment

data FunctionReturnType

This represents what information can be known about the return type of a user-defined function. For now, either the return type will be the name of a table that exists in the schema, or Unknown - implying that this information can be derived from another source, or if there is no other source, then it is an error. In future, this type may be extended with additional constructors including scalar and row types from the Logical Models feature.

Note: This is very similar to ComputedFieldReturnType defined above. The two types may be unified in future.

Constructors

FunctionReturnsTable TableName

FunctionReturnsUnknown

data DataConnectorOptions

Constructors

DataConnectorOptions

Fields

_dcoUri :: BaseUrl

_dcoDisplayName :: Maybe Text

data DataConnectorInfo

Constructors

DataConnectorInfo

Fields

_dciOptions :: DataConnectorOptions

_dciCapabilities :: Capabilities

_dciConfigSchemaResponse :: ConfigSchemaResponse

_dciDisplayName :: Maybe Text

_dciReleaseName :: Maybe Text

newtype TableName

The fully qualified name of a table. The last element in the list is the table name and all other elements represent namespacing of the table name. For example, for a database that has schemas, the name would be '[schema,name]'

Constructors

TableName

Fields

unTableName :: NonEmpty Text

newtype ConstraintName

Constructors

ConstraintName

Fields

unConstraintName :: Text

newtype ColumnName

Constructors

ColumnName

Fields

unColumnName :: Text

newtype FunctionName

Constructors

FunctionName

Fields

unFunctionName :: NonEmpty Text

data ArgumentExp a

Constructors

AEInput a
Table row accessor AETableRow | -- | Hardcoded reference to hdb_catalog.hdb_action_log.response_payload AEActionResponsePayload | -- | JSON/JSONB hasura session variable object AESession a |

data CountAggregate v

Constructors

StarCount

ColumnCount (ColumnName, AnnRedactionExp 'DataConnector v)

ColumnDistinctCount (ColumnName, AnnRedactionExp 'DataConnector v)

data Literal

Constructors

ValueLiteral ScalarType Value

ArrayLiteral ScalarType [Value]

data OrderDirection

Constructors

Ascending

Descending

newtype ScalarType

Constructors

ScalarType

Fields

unScalarType :: Text

fromGQLType :: Name -> ScalarType

data ExtraTableMetadata

This type captures backend-specific "extra" information about tables and is used on types like DBTableMetadata

Constructors

ExtraTableMetadata

Fields

_etmTableType :: TableType

_etmExtraColumnMetadata :: HashMap ColumnName ExtraColumnMetadata

data ExtraColumnMetadata

Constructors

ExtraColumnMetadata

Fields

_ecmValueGenerated :: Maybe ColumnValueGenerationStrategy

scTimeoutMicroseconds :: Lens' SourceConfig (Maybe Int)

scTemplate :: Lens' SourceConfig (Maybe Text)

scManager :: Lens' SourceConfig Manager

scEnvironment :: Lens' SourceConfig Environment

scEndpoint :: Lens' SourceConfig BaseUrl

scDataConnectorName :: Lens' SourceConfig DataConnectorName

scConfig :: Lens' SourceConfig Config

scCapabilities :: Lens' SourceConfig Capabilities