Hasura.Backends.DataConnector.Adapter.Types.Mutations

Documentation

data BackendInsert v

The Data Connector-specific data of an Insert expression. Currently, we don't have any.

This is parameterised over v which enables different phases of IR transformation to maintain the overall structure while enriching/transforming the data at the leaves.

Constructors

BackendInsert

data DataConnectorUpdateVariant v

Constructors

SingleBatch (UpdateBatch 'DataConnector UpdateOperator v)

MultipleBatches [UpdateBatch 'DataConnector UpdateOperator v]

data UpdateOperator v

The operators that are used to mutate specific columns on a table

Constructors

UpdateSet v

UpdateCustomOperator UpdateColumnOperatorName v