Hasura.Backends.DataConnector.Agent.Client

Documentation

client :: forall m. (MonadIO m, MonadError QErr m) => Manager -> BaseUrl -> IO (Routes (AsClientT m))

Create a record of client functions (see Routes) from a BaseUrl. This function takes care to add trace headers, and to propagate useful errors back to the client for debugging purposes.