Hasura.Backends.DataConnector.IR.Expression

Documentation

data Expression

A "concrete" expression type for datasource-agnostic queries (as opposed to our existing polymorphic intermediate representation).

This type should be seen as an intermediate phase of the processing pipeline which provides a high-level interface that the GraphQL Engine can use to inspect, manipulate, optimize, etc. before sending off to an agent that will be responsible for performing query generation/execution.

This type should ascribe clear semantics to its sub-expressions; when this is not possible, it should clearly defer to the semantics of some reference datasource with clearer documentation.

e.g. https://www.postgresql.org/docs/13/sql-expressions.html

Constructors

And [Expression]
A logical AND fold.

cf. https://www.postgresql.org/docs/13/functions-logical.html

Or [Expression]
A logical OR fold.

cf. https://www.postgresql.org/docs/13/functions-logical.html

Not Expression
A logical NOT function.

cf. https://www.postgresql.org/docs/13/functions-logical.html

ApplyBinaryComparisonOperator BinaryComparisonOperator ComparisonColumn ComparisonValue
Apply a BinaryComparisonOperator that compares a column to a ComparisonValue; the result of this application will return "true" or "false" depending on the BinaryComparisonOperator that's being applied.

ApplyBinaryArrayComparisonOperator BinaryArrayComparisonOperator ComparisonColumn [Value]
Apply a BinaryArrayComparisonOperator that evaluates a column with the BinaryArrayComparisonOperator against an array of ComparisonValues. The result of this application will return "true" or "false" depending on the BinaryArrayComparisonOperator that's being applied.

ApplyUnaryComparisonOperator UnaryComparisonOperator ComparisonColumn
Apply a UnaryComparisonOperator that evaluates a column with the UnaryComparisonOperator; the result of this application will return "true" or "false" depending on the UnaryComparisonOperator that's being applied.

data BinaryComparisonOperator

Operators which are typically applied to two Expressions (via the ApplyOperator sub-Expression) to perform a boolean comparison.

cf. https://www.postgresql.org/docs/13/functions-comparison.html

XXX(jkachmar): Comparison operations are tricky business!

We should define the semantics of these comparisons in a way that is clear and carefully considered.

Constructors

LessThan

LessThanOrEqual

GreaterThan

GreaterThanOrEqual

Equal

CustomBinaryComparisonOperator Text

data UnaryComparisonOperator

Constructors

IsNull

CustomUnaryComparisonOperator Text

data BinaryArrayComparisonOperator

Constructors

In

CustomBinaryArrayComparisonOperator Text

data ComparisonColumn

Constructors

ComparisonColumn

Fields

_ccPath :: [RelationshipName]

_ccName :: Name

data ComparisonValue

Constructors

AnotherColumn ComparisonColumn

ScalarValue Value