Hasura.Backends.DataConnector.IR.Name

Documentation

newtype Name ty

A tagged, opaque wrapper around Text that provides a number of derived derived instances (primarily as required by the Backend typeclass).

This wrapper is indexed by NameType so that different "names" can be represented as semantically distinct types without the boilerplate of actually defining these wrappers separately.

Constructors

Name

Fields

unName :: Text

data NameType

The "type" of "name" that the Name type is meant to provide a textual representation for.

In other words: an enumeration of all the types for which Name acts as a shared abstraction.

Constructors

Column

Function

Table