Hasura.Backends.DataConnector.IR.Query

Documentation

data QueryRequest

An abstract request to retrieve structured data from some source.

Constructors

QueryRequest

Fields

_qrTable :: Name

_qrTableRelationships :: TableRelationships

_qrQuery :: Query

data Query

The details of a query against a table

Constructors

Query

Fields

_qFields :: HashMap FieldName Field

_qAggregates :: HashMap FieldName Aggregate

_qLimit :: Maybe Int

Optionally limit to N results.

_qOffset :: Maybe Int

Optionally offset from the Nth result.

_qWhere :: Maybe Expression

Optionally constrain the results to satisfy some predicate.

_qOrderBy :: [OrderBy]

Optionally order the results by the value of one or more fields.

memptyToNothing :: (Monoid m, Eq m) => m -> Maybe m

data Field

The specific fields that are targeted by a Query.

A field conceptually falls under one of the two following categories: 1. a "column" within the data store that the query is being issued against 2. a "relationship", which indicates that the field is the result of another query that must be executed on its own NOTE: The ToJSON instance is only intended for logging purposes.

Constructors

ColumnField Name

RelField RelationshipField

data RelationshipField

A relationship consists of the following components: - a sub-query, from the perspective that a relationship field will occur within a broader Query - a join condition relating the data returned by the sub-query with that of the broader Query

cf. https://en.wikipedia.org/wiki/Join_(SQL) https://www.postgresql.org/docs/13/tutorial-join.html https://www.postgresql.org/docs/13/queries-table-expressions.html#QUERIES-FROM

NOTE: The ToJSON instance is only intended for logging purposes.

Constructors

RelationshipField

Fields

_rfRelationship :: RelationshipName

_rfQuery :: Query