Hasura.Backends.DataConnector.Plan

Documentation

data QueryPlan

Constructors

QueryPlan

Fields

_qpRequest :: QueryRequest

_qpResponseReshaper :: forall m. MonadError QErr m => QueryResponse -> m Encoding

data ResponseError

Error type for the postProcessor continuation. Failure can occur if the Agent returns bad data.

Constructors

RequiredFieldMissing

UnexpectedFields

ExpectedObject

ExpectedArray

renderQuery :: QueryRequest -> Text

Extract the Q from a Plan and render it as Text.

NOTE: This is for logging and debug purposes only.

mkPlan :: forall m. MonadError QErr m => SessionVariables -> SourceConfig -> QueryDB 'DataConnector Void (UnpreparedValue 'DataConnector) -> m QueryPlan

Map a 'QueryDB DataConnector term into a Plan

queryHasRelations :: QueryRequest -> Bool

Validate if a Q contains any relationships.

data Cardinality

Constructors

Single

Many

reshapeResponseToQueryShape :: MonadError QErr m => QueryDB 'DataConnector Void (UnpreparedValue 'DataConnector) -> QueryResponse -> m Encoding

reshapeToAnnSimpleSelect :: MonadError QErr m => Cardinality -> AnnSimpleSelectG 'DataConnector Void (UnpreparedValue 'DataConnector) -> QueryResponse -> m Encoding

reshapeSimpleSelectRows :: MonadError QErr m => Cardinality -> AnnFieldsG 'DataConnector Void (UnpreparedValue 'DataConnector) -> QueryResponse -> m Encoding

reshapeFields :: MonadError QErr m => AnnFieldsG 'DataConnector Void (UnpreparedValue 'DataConnector) -> KeyMap FieldValue -> m Encoding

encodeAssocListAsObject :: [(Text, Encoding)] -> Encoding

reshapeField :: MonadError QErr m => AnnFieldG 'DataConnector Void (UnpreparedValue 'DataConnector) -> m FieldValue -> m Encoding