Hasura.Backends.DataWrapper.Agent.Schema

Description

TODO(Solomon): Add Haddocks

Documentation

data Routes mode

Constructors

Routes

Fields

_schema :: mode :- ("schema" :> Get '[JSON] SchemaResponse)

'GET /schema'

_query :: mode :- ("query" :> (ReqBody '[JSON] Query :> Post '[JSON] QueryResponse))

'POST /query'

data SchemaResponse

The Schema Response provides the schemas for tracked tables and Capabilities supported by the service.

Constructors

SchemaResponse

Fields

srCapabilities :: Capabilities

srTables :: [TableInfo]

data Capabilities

The Capabilities describes the _capabilities_ of the service. Specifically, the service is capable of serving queries which involve relationships.

Constructors

Capabilities

Fields

dcRelationships :: Bool

newtype QueryResponse

The resolved query response provided by the 'POST /query' endpoint encoded as Value.

Constructors

QueryResponse

Fields

getQueryResponse :: [Value]