Hasura.Backends.DataWrapper.IR.OrderBy

Documentation

data OrderBy

Indicates a particular sort order that should be applied based on some Name returned within a data source query.

TODO: We should use a sum type like Query.Field here so that we can handle order by constraints on object/array relationships as well.

cf. https://www.postgresql.org/docs/13/queries-order.html

Constructors

OrderBy

Fields

column :: Name

ordering :: OrderType

data OrderType

Names may be sorted in either ascending or descending order.

cf. https://www.postgresql.org/docs/13/queries-order.html

Constructors

Ascending

Descending