Hasura.Backends.DataWrapper.IR.Query

Documentation

data Query

An abstract request to retrieve structured data from some source.

Constructors

Query

Fields

fields :: HashMap Text Field

from :: Name

Reference to the table these fields are in.

limit :: Maybe Int

Optionally limit to N results.

offset :: Maybe Int

Optionally offset from the Nth result.

where_ :: Maybe Expression

Optionally constrain the results to satisfy some predicate.

orderBy :: [OrderBy]

Optionally order the results by the value of one or more fields.

data Field

The specific fields that are targeted by a Query.

A field conceptually falls under one of the two following categories: 1. a "column" within the data store that the query is being issued against 2. a "relationship", which indicates that the field is the result of another query that must be executed on its own

Constructors

Column ColumnContents

Relationship RelationshipContents

newtype ColumnContents

Constructors

ColumnContents

Fields

column :: Name

data RelationshipContents

A relationship consists of the following components: - a sub-query, from the perspective that a relationship field will occur within a broader Query - a join condition relating the data returned by the sub-query with that of the broader Query

cf. https://en.wikipedia.org/wiki/Join_(SQL) https://www.postgresql.org/docs/13/tutorial-join.html https://www.postgresql.org/docs/13/queries-table-expressions.html#QUERIES-FROM

Constructors

RelationshipContents

Fields

joinCondition :: HashMap PrimaryKey ForeignKey

query :: Query

newtype PrimaryKey

Constructors

PrimaryKey Name

newtype ForeignKey

Constructors

ForeignKey Name