Hasura.Backends.DataWrapper.Plan

Documentation

data Plan

A Plan consists of an Q describing the query to be performed by the Agent and a continuation for post processing the response. See the postProcessResponseRow haddock for more information on why we need a post-processor.

Constructors

Plan

Fields

query :: Query

postProcessor :: QueryResponse -> Either ResponseError QueryResponse

data ResponseError

Error type for the postProcessor continuation. Failure can occur if the Agent returns bad data.

Constructors

RequiredFieldMissing

UnexpectedFields

ExpectedObject

ExpectedArray

UnexpectedResponseCardinality

renderPlan :: Plan -> Text

Extract the Q from a Plan and render it as Text.

NOTE: This is for logging and debug purposes only.

mkPlan :: forall m. MonadError QErr m => SessionVariables -> SourceConfig -> QueryDB 'DataWrapper Void (UnpreparedValue 'DataWrapper) -> m Plan

Map a 'QueryDB DataWrapper term into a Plan

postProcessResponseRow :: Capabilities -> Query -> Object -> Either ResponseError Object

We need to modify any JSON substructures which appear as a result of fetching object relationships, to peel off the outer array sent by the backend.

This function takes a response object, and the Plan used to fetch it, and modifies any such arrays accordingly.

parseObject :: Value -> Either ResponseError Object

applyCardinalityToResponse :: Cardinality -> [Object] -> Either ResponseError Value

If a fk-to-pk lookup comes from an object relationship then we expect 0 or 1 items in the response and we should return it as an object. if it's an array, we have to send back all of the results

queryHasRelations :: Query -> Bool

Validate if a Q contains any relationships.