Hasura.Backends.DataWrapper.Schema.Column

Documentation

data Column

A schematic representation which captures common attributes associated with a piece of data that is stored in a given backend.

These attributes ascribe meaningful semantics to the data that they are associated with.

cf. https://en.wikipedia.org/wiki/Column_(database) https://www.postgresql.org/docs/13/ddl-basics.html

XXX: Instead of an isNullable flag, should we instead add a Nullable data constructor to Type?

Constructors

Column

Fields

name :: Name

type_ :: Type

isNullable :: Bool

description :: Maybe Text