Hasura.Backends.MSSQL.DDL.Source

Description

MSSQL Source

Implements the Source related methods of the BackendMetadata type class for the MSSQL backend, which provides an interface for identifying the MSSQL database instance (source) and manipulate it.

The actual instance is defined in Hasura.Backends.MSSQL.Instances.Metadata.

Documentation

resolveSourceConfig :: (MonadIO m, MonadResolveSource m) => SourceName -> MSSQLConnConfiguration -> BackendSourceKind 'MSSQL -> BackendConfig 'MSSQL -> Environment -> m (Either QErr MSSQLSourceConfig)

resolveDatabaseMetadata :: (MonadIO m, MonadBaseControl IO m) => MSSQLSourceConfig -> SourceTypeCustomization -> m (Either QErr (ResolvedSource 'MSSQL))

postDropSourceHook :: (MonadIO m, MonadBaseControl IO m) => MSSQLSourceConfig -> m ()

doesSchemaExist :: MonadMSSQLTx m => SchemaName -> m Bool

doesTableExist :: MonadMSSQLTx m => TableName -> m Bool

initCatalogForSource :: MonadMSSQLTx m => m RecreateEventTriggers

Initialise catalog tables for a source, including those required by the event delivery subsystem.

dropSourceCatalog :: MonadMSSQLTx m => m ()

migrateSourceCatalog :: MonadMSSQLTx m => m RecreateEventTriggers

migrateSourceCatalogFrom :: MonadMSSQLTx m => Int -> m RecreateEventTriggers

sourceMigrations :: [(Int, TxE QErr ())]