Hasura.Backends.MSSQL.FromIr

Description

Translate from the DML to the TSql dialect.

We use StateT (newtype FromIr) as the base monad for all operations, since state is used to mangle names such that the scope of identifiers in the IR is preserved in the resulting TSQL.

For the MSSQL backend, a supported subset of the constructs that make up its TSQL dialect are represented in the form of data-types in the Hasura.Backends.MSSQL.Types module. In this module, we translate from RQL to those TSQL types. And in ToQuery we renderserializeprint the TSQL types to query-strings that are suitable to be executed on the actual MSSQL database.

In places where a series of transations are scoped under a context, we use ReaderT. For example, such translations as pertaining to a table with an alias, will require the alias for their translation operations, like qualified equality checks under where clauses, etc., perhaps below multiple layers of nested function calls.

Documentation

data Error

Most of these errors should be checked for legitimacy.

Constructors

UnsupportedOpExpG (OpExpG 'MSSQL Expression)

FunctionNotSupported

NodesUnsupportedForNow

ConnectionsNotSupported

newtype FromIr a

The base monad used throughout this module for all conversion functions.

It's a Validate, so it'll continue going when it encounters errors to accumulate as many as possible.

It also contains a mapping from entity prefixes to counters. So if my prefix is "table" then there'll be a counter that lets me generate table1, table2, etc. Same for any other prefix needed (e.g. names for joins).

A ReaderT is used around this in most of the module too, for setting the current entity that a given field name refers to. See fromColumn.

Constructors

FromIr

Fields

unFromIr :: StateT (Map Text Int) (Validate (NonEmpty Error)) a

runFromIr :: FromIr a -> Validate (NonEmpty Error) a

mkSQLSelect :: JsonAggSelect -> AnnSelectG 'MSSQL (AnnFieldG 'MSSQL Void) Expression -> FromIr Select

fromRootField :: QueryDB 'MSSQL Void Expression -> FromIr Select

Convert from the IR database query into a select.

fromSelectRows :: AnnSelectG 'MSSQL (AnnFieldG 'MSSQL Void) Expression -> FromIr Select

Toproot-level Select. All descendentsub-translations are collected to produce a root TSQL.Select.

mkNodesSelect :: Args -> Where -> Expression -> Top -> From -> [(Int, (FieldName, [FieldSource]))] -> [(Int, Projection)]

mkAggregateSelect :: Args -> Where -> Expression -> From -> [(Int, (FieldName, [Projection]))] -> [(Int, Projection)]

fromSelectAggregate :: Maybe (EntityAlias, HashMap ColumnName ColumnName) -> AnnSelectG 'MSSQL (TableAggregateFieldG 'MSSQL Void) Expression -> FromIr Select

data Args

Constructors

Args

Fields

argsWhere :: Where

argsOrderBy :: Maybe (NonEmpty OrderBy)

argsJoins :: [Join]

argsTop :: Top

argsOffset :: Maybe Expression

argsDistinct :: Proxy (Maybe (NonEmpty FieldName))

argsExistingJoins :: Map TableName EntityAlias

data UnfurledJoin

Constructors

UnfurledJoin

Fields

unfurledJoin :: Join

unfurledObjectTableAlias :: Maybe (TableName, EntityAlias)

Recorded if we joined onto an object relation.

fromSelectArgsG :: SelectArgsG 'MSSQL Expression -> ReaderT EntityAlias FromIr Args

fromAnnotatedOrderByItemG :: AnnotatedOrderByItemG 'MSSQL Expression -> WriterT (Seq UnfurledJoin) (ReaderT EntityAlias FromIr) OrderBy

Produce a valid ORDER BY construct, telling about any joins needed on the side.

unfurlAnnotatedOrderByElement :: AnnotatedOrderByElement 'MSSQL Expression -> WriterT (Seq UnfurledJoin) (ReaderT EntityAlias FromIr) (FieldName, Maybe ScalarType)

Unfurl the nested set of object relations (tell'd in the writer) that are terminated by field name (IR.AOCColumn and IR.AOCArrayAggregation).

tableNameText :: TableName -> Text

fromQualifiedTable :: TableName -> FromIr From

This is really the start where you query the base table, everything else is joins attached to it.

fromTableName :: TableName -> FromIr EntityAlias

fromAnnBoolExpFld :: AnnBoolExpFld 'MSSQL Expression -> ReaderT EntityAlias FromIr Expression

Translate an AnnBoolExpFld within an EntityAlias context referring to the table the AnnBoolExpFld field belongs to.

This is mutually recursive with fromGBoolExp, mirroring the mutually recursive structure between AnnBoolExpFld and 'AnnBoolExp b a' (alias of 'GBoolExp b (AnnBoolExpFld b a)').

fromColumnInfoForBoolExp :: ColumnInfo 'MSSQL -> ReaderT EntityAlias FromIr Expression

For boolean operators, various comparison operators used need special handling to ensure that SQL Server won't outright reject the comparison. See also shouldCastToVarcharMax.

shouldCastToVarcharMax :: ColumnType 'MSSQL -> Bool

There's a problem of comparing text fields with =, <, etc. that SQL Server completely refuses to do so. So one way to workaround this restriction is to automatically cast such text fields to varchar(max).

fromColumnInfo :: ColumnInfo 'MSSQL -> ReaderT EntityAlias FromIr FieldName

data FieldSource

Constructors

ExpressionFieldSource (Aliased Expression)

JoinFieldSource JsonCardinality (Aliased Join)

fromTableExpFieldG :: (Int, (FieldName, TableAggregateFieldG 'MSSQL Void Expression)) -> Maybe (ReaderT EntityAlias FromIr (Int, Projection))

Get FieldSource from a TAFExp type table aggregate field

fromTableAggFieldG :: (Int, (FieldName, TableAggregateFieldG 'MSSQL Void Expression)) -> Maybe (Int, (FieldName, [Projection]))

fromTableNodesFieldG :: Map TableName EntityAlias -> StringifyNumbers -> (Int, (FieldName, TableAggregateFieldG 'MSSQL Void Expression)) -> Maybe (ReaderT EntityAlias FromIr (Int, (FieldName, [FieldSource])))

fromAggregateField :: Text -> AggregateField 'MSSQL -> Projection

fromAnnFieldsG :: Map TableName EntityAlias -> StringifyNumbers -> (FieldName, AnnFieldG 'MSSQL Void Expression) -> ReaderT EntityAlias FromIr FieldSource

The main sources of fields, either constants, fields or via joins.

fromAnnColumnField :: StringifyNumbers -> AnnColumnField 'MSSQL Expression -> ReaderT EntityAlias FromIr Expression

Here is where we project a field as a column expression. If number stringification is on, then we wrap it in a ToStringExpression so that it's casted when being projected.

fromColumn :: ColumnName -> ReaderT EntityAlias FromIr FieldName

This is where a field name "foo" is resolved to a fully qualified field name [table].[foo]. The table name comes from EntityAlias in the ReaderT.

fieldSourceProjections :: FieldSource -> Projection

joinAliasToField :: JoinAlias -> FieldName

fieldSourceJoin :: FieldSource -> Maybe Join

fromObjectRelationSelectG :: Map TableName EntityAlias -> ObjectRelationSelectG 'MSSQL Void Expression -> ReaderT EntityAlias FromIr Join

lookupTableFrom :: Map TableName EntityAlias -> TableName -> FromIr (Either EntityAlias From)

fromArraySelectG :: ArraySelectG 'MSSQL Void Expression -> ReaderT EntityAlias FromIr Join

fromArrayAggregateSelectG :: AnnRelationSelectG 'MSSQL (AnnAggregateSelectG 'MSSQL Void Expression) -> ReaderT EntityAlias FromIr Join

fromArrayRelationSelectG :: ArrayRelationSelectG 'MSSQL Void Expression -> ReaderT EntityAlias FromIr Join

fromRelName :: RelName -> FromIr Text

fromMapping :: From -> HashMap ColumnName ColumnName -> ReaderT EntityAlias FromIr [Expression]

The context given by the reader is of the previous/parent "remote" table. The WHERE that we're generating goes in the child, "local" query. The From passed in as argument is the local table.

We should hope to see e.g. "post.category = category.id" for a local table of post and a remote table of category.

The left/right columns in HashMap ColumnName ColumnName corresponds to the left/right of select ... join .... Therefore left=remote, right=local in this context.

selectFromMapping :: Select -> HashMap ColumnName ColumnName -> ReaderT EntityAlias FromIr [Expression]

fromOpExpG :: Expression -> OpExpG 'MSSQL Expression -> FromIr Expression

nullableBoolEquality :: Expression -> Expression -> Expression

nullableBoolInequality :: Expression -> Expression -> Expression

fromGBoolExp :: GBoolExp 'MSSQL (AnnBoolExpFld 'MSSQL Expression) -> ReaderT EntityAlias FromIr Expression

Translate a GBoolExp of a AnnBoolExpFld, within an EntityAlias context.

It is mutually recursive with fromAnnBoolExpFld and fromGExists.

fromInsert :: AnnInsert 'MSSQL Void Expression -> Insert

normalizeInsertRows :: [[(Column 'MSSQL, Expression)]] -> [[(Column 'MSSQL, Expression)]]

Normalize a row by adding missing columns with DEFAULT value and sort by column name to make sure all rows are consistent in column values and order.

Example: A table "author" is defined as:

CREATE TABLE author ([id] INTEGER NOT NULL PRIMARY KEY, name TEXT NOT NULL, age INTEGER)

Consider the following mutation:

mutation {
  insert_author(
    objects: [{id: 1, name: "Foo", age: 21}, {id: 2, name: "Bar"}]
  ){
    affected_rows
  }
}

We consider DEFAULT value for age column which is missing in second insert row.

The corresponding INSERT statement looks like:

INSERT INTO author (id, name, age)
  OUTPUT INSERTED.id
  VALUES (1, 'Foo', 21), (2, 'Bar', DEFAULT)

toMerge :: TableName -> [AnnotatedInsertRow 'MSSQL Expression] -> [ColumnInfo 'MSSQL] -> IfMatched Expression -> FromIr Merge

Construct a MERGE statement from AnnInsert information. A MERGE statement is responsible for actually inserting and/or updating the data in the table.

toInsertValuesIntoTempTable :: TempTableName -> AnnInsert 'MSSQL Void Expression -> InsertValuesIntoTempTable

As part of an INSERT/UPSERT process, insert VALUES into a temporary table. The content of the temporary table will later be inserted into the original table using a MERGE statement.

We insert the values into a temporary table first in order to replace the missing fields with DEFAULT in normalizeInsertRows, and we can't do that in a MERGE statement directly.

fromDelete :: AnnDel 'MSSQL -> FromIr Delete

Convert IR AST representing delete into MSSQL AST representing a delete statement

fromUpdate :: AnnotatedUpdate 'MSSQL -> FromIr Update

Convert IR AST representing update into MSSQL AST representing an update statement

toSelectIntoTempTable :: TempTableName -> TableName -> [ColumnInfo 'MSSQL] -> SITTConstraints -> SelectIntoTempTable

Create a temporary table with the same schema as the given table.

columnInfoToUnifiedColumn :: ColumnInfo 'MSSQL -> UnifiedColumn

Extracts the type and column name of a ColumnInfo

trueExpression :: Expression

safeJsonQueryExpression :: JsonCardinality -> Expression -> Expression

A version of JSON_QUERY(..) that returns a proper json literal, rather than SQL null, which does not compose properly with FOR JSON clauses.

nullExpression :: Expression

emptyArrayExpression :: Expression

jsonFieldName :: Text

aggFieldName :: Text

aggSubselectName :: Text

existsFieldName :: Text

data NameTemplate

Constructors

ArrayRelationTemplate Text

ArrayAggregateTemplate Text

ObjectRelationTemplate Text

TableTemplate Text

ForOrderAlias Text

generateAlias :: NameTemplate -> FromIr Text

Generate an alias for a given entity to remove ambiguity and naming conflicts between scopes at the TSQL level. Keeps track of the increments for the alias index in the StateT

fromAlias :: From -> EntityAlias

columnNameToFieldName :: ColumnName -> EntityAlias -> FieldName