Hasura.Backends.MSSQL.FromIr.Expression

Description

This module translates the IR of boolean expressions into TSQL boolean expressions.

Boolean expressions typically arise from permissions and where-clause filters.

Documentation

fromGBoolExp :: GBoolExp 'MSSQL (AnnBoolExpFld 'MSSQL Expression) -> ReaderT EntityAlias FromIr Expression

Translate boolean expressions into TSQL Expressions.

The AnnBoolExpFld references fields and columns. The entity (e.g. table) that binds these columns is supplied in the `ReaderT EntityAlias` environment, such that the columns can be referred to unambiguously.