Hasura.Backends.MSSQL.FromIr.SelectIntoTempTable

Description

This module contains supporting definitions for building temporary tables based off of the schema of other tables. This is used in mutations to capture the data of rows that are affected.

Documentation

toSelectIntoTempTable :: TempTableName -> TableName -> [ColumnInfo 'MSSQL] -> SITTConstraints -> SelectIntoTempTable

Create a temporary table with the same schema as the given table.