Hasura.Backends.MSSQL.SQL.Error

Description

Functions and datatypes for interpreting MSSQL database errors.

Documentation

data ErrorClass

The top-level error class. Errors in MSSQL are divided into different classes, which are further subdivided into individual error subclasses. It is useful to determine the class of database exception and handle it appropriately.

Constructors

DataException (ErrorSubclass DataExceptionSubclass)

IntegrityConstraintViolation

SyntaxErrorOrAccessViolation (ErrorSubclass SyntaxErrorOrAccessViolationSubclass)

data ErrorSubclass a

Constructors

NoSubclass
represents non-specific 000 subclass code

Subclass a
represents known, more specific sub class

data DataExceptionSubclass

Constructors

StringDataRightTruncated

NumericValueOutOfRange

InvalidDatetimeFormat

DatetimeFieldOverflow

IntervalFieldOverflow

InvalidEscapeCharacter

InvalidEscapeSequence

data SyntaxErrorOrAccessViolationSubclass

Constructors

TableOrViewAlreadyExists

TableOrViewNotFound

IndexAlreadyExists

IndexNotFound

ColumnAlreadyExists

ColumnNotFound

errorClassCode :: ErrorClass -> Code

Assign each error class' subclasses an appropriate API error code

parseErrorClass :: String -> Maybe ErrorClass

Parsing error class and subclass information from a SQLSTATE code. SQLSTATE provides detailed information about the cause of a warning or error. A SQLSTATE consists of 5 chars. They are divided into two parts: the first and second chars contain a class and the following three a subclass.

defaultMSSQLTxErrorHandler :: MSSQLTxError -> QErr

A default transaction error handler where all errors are unexpected.

mutationMSSQLTxErrorHandler :: MSSQLTxError -> QErr

A transaction error handler to be used in constructing mutation transactions, i.e INSERT, UPDATE and DELETE. We expect data exception and integrity constraint violation.

mkMSSQLTxErrorHandler :: (ErrorClass -> Bool) -> MSSQLTxError -> QErr

Constructs a transaction error handler given a predicate that determines which error classes (and subclasses) are expected and should be reported to the user. All other errors are considered internal errors. Example:- Consider a insert mutation where we insert some data into columns of a table. Except for the basic data type, such as Boolean, String, Float, Int etc. we cannot invalidate data any further, such as validating timestamp string format. In this case, a @DataException is expected from the database and it is handled and thrown with proper error message.

odbcExceptionToJSONValue :: ODBCException -> Value

The ODBCException type has no ToJSON instance. This is an attempt to convert the odbc exception to a JSON @Value