Hasura.Backends.MSSQL.Schema.IfMatched

Description

MSSQL Schema IfMatched

This module contains the building blocks for parsing if_matched clauses (represented as IfMatched), which in the MSSQL backend are used to implement upsert functionality.

These are used by backendInsertParser to construct a mssql-specific schema parser for insert (and upsert) mutations.

Documentation

ifMatchedFieldParser :: forall r m n. MonadBuildSchema 'MSSQL r m n => SourceName -> TableInfo 'MSSQL -> Maybe (SelPermInfo 'MSSQL) -> Maybe (UpdPermInfo 'MSSQL) -> m (InputFieldsParser n (Maybe (IfMatched (UnpreparedValue 'MSSQL))))

Field-parser for:

if_matched: tablename_if_matched

input tablename_if_matched {
  match_columns: [tablename_select_column!]
  update_columns: [tablename_update_columns!]
  where: tablename_bool_exp
}

Note that the types ordinarily produced by this parser are only created if the active role has both select and update permissions to the table tablename defined and these grant non-empty column permissions.

ifMatchedObjectParser :: forall r m n. MonadBuildSchema 'MSSQL r m n => SourceName -> TableInfo 'MSSQL -> Maybe (SelPermInfo 'MSSQL) -> Maybe (UpdPermInfo 'MSSQL) -> m (Maybe (Parser 'Input n (IfMatched (UnpreparedValue 'MSSQL))))

Parse a tablename_if_matched object.

tableInsertMatchColumnsEnum :: forall m n r. (MonadSchema n m, MonadRole r m, MonadTableInfo r m, Has MkTypename r) => SourceName -> TableInfo 'MSSQL -> SelPermInfo 'MSSQL -> m (Maybe (Parser 'Both n (Column 'MSSQL)))

Table insert_match_columns enum

Parser for an enum type that matches the columns that can be used for insert match_columns for a given table. Maps to the insert_match_columns object.

Return Nothing if there's no column the current user has "select" permissions for.

isMatchColumnValid :: ColumnInfo 'MSSQL -> Bool

Check whether a column can be used for match_columns.