Hasura.Backends.MSSQL.Types.Insert

Description

MSSQL Types Insert

Types for MSSQL Insert IR.

Documentation

data BackendInsert v

Defines the part in insert mutation that is unique for MSSQL the if_matched clause.

Constructors

BackendInsert

Fields

_biIfMatched :: Maybe (IfMatched v)

if_matched can be omitted (and in that case will be Nothing). If omitted, we only insert new rows (without upserting).

data IfMatched v

The IR data representing an if_matched clause, which handles upserts.

Constructors

IfMatched

Fields

_imMatchColumns :: [Column 'MSSQL]

Columns to compare when checking if there's a match

_imUpdateColumns :: [Column 'MSSQL]

Columns to update when there's a match

_imConditions :: AnnBoolExp 'MSSQL v

A condition for updating columns in case of a match

_imColumnPresets :: HashMap ColumnName v

Default values (presets) for some columns