Hasura.Backends.MSSQL.Types.Internal

Description

MSSQL Types Internal

Types for Transact-SQL aka T-SQL; the language of SQL Server.

In this module we define various MS SQL Server specific data types used for T-SQL generation.

These types are also used as underlying types in the Backend 'MSSQL instance which is defined in Hasura.Backends.MSSQL.Instances.Types.

We convert RQL IR ASTs to types defined here in the Hasura.Backends.MSSQL.FromIr module, and we implement pretty-printing for these types in the Hasura.Backends.MSSQL.ToQuery module.

NOTE: Various type class instances (including simple once such as Eq and Show) are implemented in the Hasura.Backends.MSSQL.Types.Instances module.

Documentation

type Column (b :: BackendType) = ColumnName

type ColumnType (b :: BackendType) = ScalarType

type Value = Value

data UnifiedColumn

Constructors

UnifiedColumn

Fields

name :: Text

type' :: ScalarType

data UnifiedTableName

Constructors

UnifiedTableName

Fields

schema :: Text

name :: Text

data UnifiedObjectRelationship

Constructors

UnifiedObjectRelationship

Fields

using :: UnifiedUsing

name :: Text

data UnifiedArrayRelationship

Constructors

UnifiedArrayRelationship

Fields

using :: UnifiedUsing

name :: Text

newtype UnifiedUsing

Constructors

UnifiedUsing

Fields

foreign_key_constraint_on :: UnifiedOn

data UnifiedOn

Constructors

UnifiedOn

Fields

table :: UnifiedTableName

column :: Text

data BooleanOperators a

Constructors

ASTContains a

ASTCrosses a

ASTEquals a

ASTIntersects a

ASTOverlaps a

ASTTouches a

ASTWithin a

data Select

Constructors

Select

Fields

selectWith :: Maybe With

selectTop :: Top

selectProjections :: [Projection]

selectFrom :: Maybe From

selectJoins :: [Join]

selectWhere :: Where

selectFor :: For

selectOrderBy :: Maybe (NonEmpty OrderBy)

selectOffset :: Maybe Expression

emptySelect :: Select

newtype OutputColumn

Constructors

OutputColumn

Fields

unOutputColumn :: ColumnName

data Inserted

Constructors

Inserted

data Deleted

Constructors

Deleted

data Output t

Constructors

Output

Fields

outputType :: t

outputColumns :: [OutputColumn]

type InsertOutput = Output Inserted

newtype Values

Constructors

Values [Expression]

data Insert

Constructors

Insert

Fields

insertTable :: TableName

insertColumns :: [ColumnName]

insertOutput :: InsertOutput

insertTempTable :: TempTable

insertValues :: [Values]

data SetValue

Constructors

SetON

SetOFF

data SetIdentityInsert

Constructors

SetIdentityInsert

Fields

setTable :: SomeTableName

setValue :: !SetValue

type DeleteOutput = Output Deleted

data Delete

Constructors

Delete

Fields

deleteTable :: Aliased TableName

deleteOutput :: DeleteOutput

deleteTempTable :: TempTable

deleteWhere :: Where

data Merge

MERGE statement. Used for upserts and is responsible for actually inserting or updating the data in the table.

Constructors

Merge

Fields

mergeTargetTable :: TableName

mergeUsing :: MergeUsing

mergeOn :: MergeOn

mergeWhenMatched :: MergeWhenMatched

mergeWhenNotMatched :: MergeWhenNotMatched

mergeInsertOutput :: InsertOutput

mergeOutputTempTable :: TempTable

data MergeUsing

The USING section of a MERGE statement. Specifies the temp table schema where the input values are.

Constructors

MergeUsing

Fields

mergeUsingTempTable :: TempTableName

mergeUsingColumns :: [ColumnName]

data MergeOn

The ON section of a MERGE statement. Which columns to match on?

Constructors

MergeOn

Fields

mergeOnColumns :: [ColumnName]

data MergeWhenMatched

The WHEN MATCHED section of a MERGE statement. Which columns to update when match_columns match (including presets), and on which condition to actually update the values.

Constructors

MergeWhenMatched

Fields

mwmUpdateColumns :: [ColumnName]

mwmCondition :: Expression

mwmUpdatePreset :: HashMap ColumnName Expression

newtype MergeWhenNotMatched

The WHEN MATCHED section of a MERGE statement. Which columns to insert?

Constructors

MergeWhenNotMatched

Fields

mergeWhenNotMatchedInsertColumns :: [ColumnName]

data SelectIntoTempTable

SELECT INTO temporary table statement without values. Used to create a temporary table with the same schema as an existing table.

Constructors

SelectIntoTempTable

Fields

sittTempTableName :: TempTableName

sittColumns :: [UnifiedColumn]

sittFromTableName :: TableName

sittConstraints :: SITTConstraints

data SITTConstraints

When creating a temporary table from an existing table schema, what should we do with the constraints (such as IDENTITY?)

Constructors

KeepConstraints

RemoveConstraints

data InsertValuesIntoTempTable

Simple insert into a temporary table.

Constructors

InsertValuesIntoTempTable

Fields

ivittTempTableName :: TempTableName

ivittColumns :: [ColumnName]

ivittValues :: [Values]

newtype TempTableName

A temporary table name is prepended by a hash-sign

Constructors

TempTableName Text

tempTableNameInserted :: TempTableName

tempTableNameValues :: TempTableName

tempTableNameDeleted :: TempTableName

tempTableNameUpdated :: TempTableName

data SomeTableName

A name of a regular table or temporary table

Constructors

RegularTableName TableName

TemporaryTableName TempTableName

data TempTable

Constructors

TempTable

Fields

ttName :: TempTableName

ttColumns :: [ColumnName]

data Reselect

Constructors

Reselect

Fields

reselectProjections :: [Projection]

reselectFor :: For

reselectWhere :: Where

data OrderBy

Constructors

OrderBy

Fields

orderByFieldName :: FieldName

orderByOrder :: Order

orderByNullsOrder :: NullsOrder

orderByType :: Maybe ScalarType

data Order

Constructors

AscOrder

DescOrder

data NullsOrder

Constructors

NullsFirst

NullsLast

NullsAnyOrder

data For

Constructors

JsonFor ForJson

NoFor

data ForJson

Constructors

ForJson

Fields

jsonCardinality :: JsonCardinality

jsonRoot :: Root

data Root

Constructors

NoRoot

Root Text

data JsonCardinality

Constructors

JsonArray

JsonSingleton

data Projection

Constructors

ExpressionProjection (Aliased Expression)

FieldNameProjection (Aliased FieldName)

AggregateProjection (Aliased Aggregate)

StarProjection

data Join

Constructors

Join

Fields

joinSource :: JoinSource

joinJoinAlias :: JoinAlias

data JoinSource

Constructors

JoinSelect Select

JoinReselect Reselect

data JoinAlias

Constructors

JoinAlias

Fields

joinAliasEntity :: Text

joinAliasField :: Maybe Text

newtype Where

Constructors

Where [Expression]

newtype With

Constructors

With (NonEmpty (Aliased Select))

data Top

Constructors

NoTop

Top Int

data Expression

Constructors

ValueExpression Value

AndExpression [Expression]

OrExpression [Expression]

NotExpression Expression

ExistsExpression Select

SelectExpression Select

IsNullExpression Expression

IsNotNullExpression Expression

ColumnExpression FieldName

JsonQueryExpression Expression
This one acts like a "cast to JSON" and makes SQL Server behave like it knows your field is JSON and not double-encode it.

ToStringExpression Expression

MethodApplicationExpression Expression MethodApplicationExpression

FunctionApplicationExpression FunctionApplicationExpression

JsonValueExpression Expression JsonPath
This is for getting actual atomic values out of a JSON string.

OpExpression Op Expression Expression

ListExpression [Expression]

STOpExpression SpatialOp Expression Expression

CastExpression Expression ScalarType DataLength

ConditionalExpression Expression Expression Expression
"CASE WHEN (expression) THEN (expression) ELSE (expression) END"

DefaultExpression
The DEFAULT value. TODO: Make this as a part of @Value.

data DataLength

Data type describing the length of a datatype. Used in CastExpressions.

Constructors

DataLengthUnspecified

DataLengthInt Int

DataLengthMax

data FunctionApplicationExpression

SQL functions application: some_function(e1, e2, ..).

Constructors

FunExpISNULL Expression Expression

data MethodApplicationExpression

Object expression method application: (expression).text(e1, e2, ..)

Constructors

MethExpSTAsText

data JsonPath

Constructors

RootPath

FieldPath JsonPath Text

IndexPath JsonPath Integer

data Aggregate

Constructors

CountAggregate (Countable FieldName)

OpAggregate Text [Expression]

TextAggregate Text

data Countable name

Constructors

StarCountable

NonNullFieldCountable name

DistinctCountable name

data From

Constructors

FromQualifiedTable (Aliased TableName)

FromOpenJson (Aliased OpenJson)

FromSelect (Aliased Select)

FromIdentifier Text

FromTempTable (Aliased TempTableName)

data OpenJson

Constructors

OpenJson

Fields

openJsonExpression :: Expression

openJsonWith :: Maybe (NonEmpty JsonFieldSpec)

data JsonFieldSpec

Constructors

IntField Text (Maybe JsonPath)

JsonField Text (Maybe JsonPath)

StringField Text (Maybe JsonPath)

UuidField Text (Maybe JsonPath)

data Aliased a

Constructors

Aliased

Fields

aliasedThing :: a

aliasedAlias :: Text

newtype SchemaName

Constructors

SchemaName

Fields

schemaNameParts :: [Text]

data TableName

Constructors

TableName

Fields

tableName :: Text

tableSchema :: Text

data FieldName

Constructors

FieldName

Fields

fieldName :: Text

fieldNameEntity :: Text

data Comment

Constructors

DueToPermission

RequestedSingleObject

newtype EntityAlias

Constructors

EntityAlias

Fields

entityAliasText :: Text

data Op

Constructors

LT

LTE

GT

GTE

IN

LIKE

NLIKE

NIN

EQ'

NEQ'

data SpatialOp

Supported operations for spatial data types

Constructors

STEquals

STContains

STCrosses

STIntersects

STOverlaps

STWithin

STTouches

newtype ColumnName

Column name of some database table -- this differs to FieldName that is used for referring to things within a query.

Constructors

ColumnName

Fields

columnNameText :: Text

data ScalarType

Derived from the odbc package.

Constructors

CharType

NumericType

DecimalType

IntegerType

SmallintType

FloatType

RealType

DateType

Ss_time2Type

VarcharType

WcharType

WvarcharType

WtextType

TimestampType

TextType

BinaryType

VarbinaryType

BigintType

TinyintType

BitType

GuidType

GeographyType

GeometryType

UnknownType Text

scalarTypeDBName :: DataLength -> ScalarType -> Text

fromDataLength :: DataLength -> Text

mkMSSQLScalarTypeName :: MonadError QErr m => ScalarType -> m Name

parseScalarValue :: ScalarType -> Value -> Either QErr Value

isComparableType :: ScalarType -> Bool

isNumType :: ScalarType -> Bool

getGQLTableName :: TableName -> Either QErr Name

snakeCaseTableName :: TableName -> Text

stringTypes :: [ScalarType]

geoTypes :: [ScalarType]