Hasura.Backends.MSSQL.Types.Update

Description

MSSQL Types Update

This module defines the Update-related IR types specific to MSSQL.

Documentation

data BackendUpdate v

The MSSQL-specific data of an Update expression.

This is parameterised over v which enables different phases of IR transformation to maintain the overall structure while enriching/transforming the data at the leaves.

Constructors

BackendUpdate

Fields

updateOperations :: HashMap ColumnName (UpdateOperator v)

The update operations to perform on each column.

This HashMap will also contain the update operators for the presets. They are populated as part of the schema building in msBuildTableUpdateMutationFields in the call to buildUpdateOperators.

data UpdateOperator v

The various update operators supported by MSSQL, i.e. the _set, _inc operators that appear in the schema.

TODO: Docs for MSSQL! See Update Mutations User docs

Constructors

UpdateSet v

UpdateInc v

type UpdateSet = HashMap ColumnName (UpdateOperator Expression)

type UpdateOutput = Output Inserted

data Update

UPDATE [table_alias] SET [table_alias].column = value OUTPUT INSERTED.column INTO #updated FROM [table_name] AS [table_alias] WHERE filter-expression

Constructors

Update

Fields

updateTable :: Aliased TableName

updateSet :: UpdateSet

updateOutput :: UpdateOutput

updateTempTable :: TempTable

updateWhere :: Where