Hasura.Backends.MySQL.DataLoader.Execute

Description

Execute the plan given from .Plan.

Documentation

data RecordSet

A set of records produced by the database. These are joined together. There are all sorts of optimizations possible here, from using a matrix/flat vector, unboxed sums for Value, etc. Presently we choose a naive implementation in the interest of getting other work done.

Constructors

RecordSet

Fields

origin :: Maybe PlannedAction

rows :: Vector (InsOrdHashMap FieldName OutputValue)

wantedFields :: Maybe [Text]

data ExecuteProblem

Any problem encountered while executing the plan.

Constructors

GetJobDecodeProblem String

CreateQueryJobDecodeProblem String

JoinProblem ExecuteProblem

UnsupportedJoinBug JoinType

MissingRecordSetBug Ref

BrokenJoinInvariant [FieldName]

data OutputValue

A value outputted by this execute module in a record set.

Constructors

ArrayOutputValue (Vector OutputValue)

RecordOutputValue (InsOrdHashMap FieldName OutputValue)

ScalarOutputValue Value

NullOutputValue

runExecute :: MonadIO m => SourceConfig -> HeadAndTail -> Execute a -> m (Either ExecuteProblem RecordSet)

Using the config, run the execute action. Finally, resolve the head-and-tail to a record set.

execute :: Forest PlannedAction -> Execute ()

Execute the forest of actions.

leftObjectJoin :: Maybe [Text] -> Text -> [(FieldName, FieldName)] -> RecordSet -> RecordSet -> Either ExecuteProblem RecordSet

Inefficient but clean left object join.

joinObjectRows :: Maybe [Text] -> Text -> InsOrdHashMap FieldName OutputValue -> Vector (InsOrdHashMap FieldName OutputValue) -> Either ExecuteProblem (InsOrdHashMap FieldName OutputValue)

Join a row with another as an object join.

If rightRow is not a single row, we throw BrokenJoinInvariant.