Hasura.Backends.MySQL.DataLoader.Plan

Description

Make a plan for the data loader to execute (.Execute).

It will produce a graph of actions, to be executed by .Execute.

Documentation

data Ref

A reference to a result of loading a recordset from the database.

Constructors

Ref

Fields

idx :: !Int

This index will be generated by the planner.

text :: !Text

A display name. The idx gives us uniqueness.

data Select

A almost-the-same version of Select from Types.Internal, except with some fields used for planning and executing.

Constructors

Select

Fields

selectAggUnwrap :: !(Maybe Text)

selectFrom :: !From

selectGroupBy :: ![FieldName]

selectHaskellJoins :: ![Join]

selectOrderBy :: !(Maybe (NonEmpty OrderBy))

selectProjections :: ![Projection]

selectRelationship :: !(Maybe Relationship)

selectWhere :: !Where

selectWantedFields :: !(Maybe [Text])

selectSqlOffset :: !(Maybe Int)

selectSqlTop :: !Top

data Join

An join action.

Constructors

Join

Fields

leftRecordSet :: Ref

Join this side...

rightRecordSet :: Ref

with this side.

joinRhsTop :: !Top

Join only the top N results. It's important that we do this IN HASKELL, therefore this is not part of the generated SQL.

joinRhsOffset :: !(Maybe Int)

Offset applied to the right-hand-side table.

joinType :: !JoinType

Type of relational join to do.

joinFieldName :: !Text

Field name to return the join result as; e.g. "albums" for an artist with an array relation of albums.

wantedFields :: !(Maybe [Text])

The SQL queries may achieve the data using joining fields, but those fields aren't supposed to be returned back to the user. To avoid that, we explicitly specify which fields are wanted from this join. E.g. "title" and "year", but not artist_id which was used to Haskell-join the row with an album_artist_id, or whatever.

data Action

An action that the executor will perform. Either pull data from the database directly via a select, or join two other actions' record sets together.

Constructors

SelectAction Select

JoinAction Join

data PlannedAction

An action planned, with a unique reference. I.e. the action performed yields a result stored at reference ref.

Constructors

PlannedAction

Fields

ref :: Ref

action :: Action

data Relationship

A relationship lets the executor insert on-the-fly WHERE fkey1=fkey2 for relationships. These can only be inserted on-the-fly and aren't known at the time of planning, because the keys come from the left-hand-side table for a join.

Constructors

Relationship

Fields

leftRecordSet :: Ref

rightTable :: EntityAlias

joinType :: JoinType

newtype FieldName

Just a wrapper to clarify some types. It's different from the MySQL.FieldName because it doesn't care about schemas: schemas aren't returned in recordsets from the database.

Constructors

FieldName Text

data HeadAndTail

The reason for this is subtle. Read this documentation. For each join on a select (see above, there is a list), we split that out into three jobs:

One job for the left hand side (i.e. the select).

One job for the right hand side (i.e. the join).

One job to join them (And in the darkness bind them...)

This is performed as a fold, like: foldM planJoin head joins. A nice linked-list or tree-like structure arises. The planner code produces a graph out of this; so it's possible that some parallelism can be achieved by running multiple jobs at once.

The "head" is the first, original select. The "tail" is the (indirectly) linked list of joins. That list may also be empty. In that case, the tail is simply the same as the head.

If the tail is different to the head, then we choose the tail, as it represents the joined up version of both. If they're the same, we take whichever.

Constructors

HeadAndTail

Fields

head :: Ref

tail :: Ref

data PlanState

We're simply accumulating a set of actions with this. The counter lets us generate unique refs.

Constructors

PlanState

Fields

actions :: !(Seq PlannedAction)

counter :: !Int

newtype Plan a

Simple monad to collect actions.

Constructors

Plan

Fields

unPlan :: State PlanState a

toFieldName :: FieldName -> FieldName

Note that we're intentionally discarding the table qualification.

joinAliasName :: EntityAlias -> Text

selectFromName :: From -> Text

Used for display purposes, not semantic content.

runPlan :: Plan r -> (r, [PlannedAction])

planSelectHeadAndTail :: Maybe Relationship -> Maybe Text -> Select -> Plan HeadAndTail

See the documentation for HeadAndTail.

planJoin :: Ref -> Join -> Plan Ref

Given a left-hand-side table and a join spec, produce a single reference that refers to the composition of the two.

tell :: PlannedAction -> Plan ()

Write the planned action to the state, like a writer's tell.

generate :: Text -> Plan Ref

Generate a unique reference with a label for debugging.

actionsForest :: (Graph -> Graph) -> [PlannedAction] -> Forest PlannedAction

Graph the set of planned actions ready for execution in the correct order.

selectQuery :: Select -> Select

Used by the executor to produce a plain old select that can be sent to the MySQL server.

fromSelect :: Maybe Relationship -> Maybe Text -> Select -> Select

From a plain select, and possibly a parent/left-hand-side relationship, produce a select that is useful for execution.