Hasura.Backends.MySQL.FromIr

Description

Translate from the DML to the MySQL dialect.

Documentation

data Error

Most of these errors should be checked for legitimacy.

Constructors

UnsupportedOpExpG (OpExpG 'MySQL Expression)

IdentifierNotSupported

FunctionNotSupported

NodesUnsupportedForNow

ConnectionsNotSupported

data FromIr a

The base monad used throughout this module for all conversion functions.

It's a Validate, so it'll continue going when it encounters errors to accumulate as many as possible.

It also contains a mapping from entity prefixes to counters. So if my prefix is "table" then there'll be a counter that lets me generate table1, table2, etc. Same for any other prefix needed (e.g. names for joins).

A ReaderT is used around this in most of the module too, for setting the current entity that a given field name refers to. See fromColumn.

runFromIr :: FromIr a -> Validate (NonEmpty Error) a

fromSelectRows :: AnnSelectG 'MySQL (AnnFieldG 'MySQL Void) Expression -> FromIr Select

mkSQLSelect :: JsonAggSelect -> AnnSelectG 'MySQL (AnnFieldG 'MySQL Void) Expression -> FromIr Select

fromRootField :: QueryDB 'MySQL Void Expression -> FromIr Select

Convert from the IR database query into a select.