Hasura.Backends.MySQL.Plan

Description

Planning MySQL queries and subscriptions.

Documentation

queryToActionForest :: MonadError QErr m => UserInfo -> QueryDB 'MySQL Void (UnpreparedValue 'MySQL) -> m (HeadAndTail, Forest PlannedAction)

Plan the query and then produce a forest of actions for the executor.

planQuery :: MonadError QErr m => SessionVariables -> QueryDB 'MySQL Void (UnpreparedValue 'MySQL) -> m Select

prepareValueQuery :: MonadError QErr m => SessionVariables -> UnpreparedValue 'MySQL -> m Expression

Prepare a value without any query planning; we just execute the query with the values embedded.