Hasura.Backends.MySQL.ToQuery

Description

Convert the simple AST to an SQL query, ready to be passed to the mysql package's query/exec functions.

Documentation

newtype Query

Constructors

Query

Fields

unQuery :: ByteString

data Printer

Constructors

SeqPrinter [Printer]

SepByPrinter Printer [Printer]

NewlinePrinter

QueryPrinter Query

IndentPrinter Int Printer

(<+>) :: Printer -> Printer -> Printer

fromExpression :: Expression -> Printer

fromScalarType :: ScalarValue -> Query

fromOp :: Op -> Printer

fromFieldName :: FieldName -> Printer

fromSelect :: Select -> Printer

fromOffsetAndLimit :: Top -> Maybe Int -> Printer

fromOrderBys :: Maybe (NonEmpty OrderBy) -> Printer

fromOrderBy :: OrderBy -> [Printer]

fromOrder :: Order -> Printer

fromNullsOrder :: FieldName -> NullsOrder -> Printer

fromProjection :: Projection -> Printer

fromAggregate :: Aggregate -> Printer

fromCountable :: Countable FieldName -> Printer

fromWhere :: Where -> Printer

fromFrom :: From -> Printer

parens :: Printer -> Printer

fromTableName :: TableName -> Printer

fromAliased :: Aliased Printer -> Printer

fromNameText :: Text -> Printer

truePrinter :: Printer

falsePrinter :: Printer

toQueryFlat :: Printer -> Query

toQueryPretty :: Printer -> Query

renderBuilderPretty :: Printer -> (Builder, InsOrdHashMap Int ScalarValue)

Produces a query with holes, and a mapping for each

runBuilderPretty :: Printer -> State (InsOrdHashMap ScalarValue Int) Builder

fromFieldOrigin :: FieldOrigin -> Printer

fromNullAggregate :: Aggregate -> Printer