Hasura.Backends.Postgres.Connection

Description

Postgres Connetion

This module handles the connection against a Postgres server. It provides an assortment of features, such as:

MonadTx, a class which abstracts the QErr in TxE via MonadError

various run combinators for executing TxE into ExceptT or MonadIO

dealing with trace contexts (withTraceContext and an orphan MonadTrace instance for TxET)

connection pool settings (PostgresPoolSettings, setPostgresPoolSettings, etc)

other settings, including source connection info and read replicas, etc. (PostgresConnConfiguration)

setting or getting miscellaneous properties (setHeadersTx, sessionInfoExp, doesTableExist)

Documentation

class MonadError QErr m => MonadTx m where

Methods

liftTx :: TxE QErr a -> m a

runTx :: (MonadIO m, MonadBaseControl IO m) => PGExecCtx -> TxAccess -> TxET QErr m a -> ExceptT QErr m a

Executes the given query in a transaction of the specified mode, within the provided PGExecCtx.

runTxWithCtx :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadTrace m, UserInfoM m) => PGExecCtx -> TxAccess -> TxET QErr m a -> m a

runQueryTx :: (MonadIO m, MonadError QErr m) => PGExecCtx -> TxET QErr IO a -> m a

This runs the given set of statements (Tx) without wrapping them in BEGIN and COMMIT. This should only be used for running a single statement query!

setHeadersTx :: MonadIO m => SessionVariables -> TxET QErr m ()

sessionInfoJsonExp :: SessionVariables -> SQLExp

withUserInfo :: MonadIO m => UserInfo -> TxET QErr m a -> TxET QErr m a

setTraceContextInTx :: MonadIO m => TraceContext -> TxET QErr m ()

withTraceContext :: MonadIO m => TraceContext -> TxET QErr m a -> TxET QErr m a

Inject the trace context as a transaction-local variable, so that it can be picked up by any triggers (including event triggers).

checkDbConnection :: MonadIO m => PGPool -> m Bool

doesSchemaExist :: MonadTx m => SchemaName -> m Bool

doesTableExist :: MonadTx m => SchemaName -> TableName -> m Bool

isExtensionAvailable :: MonadTx m => Text -> m Bool

enablePgcryptoExtension :: forall m. MonadTx m => m ()

dropHdbCatalogSchema :: MonadTx m => m ()

data PostgresPoolSettings

Constructors

PostgresPoolSettings

Fields

_ppsMaxConnections :: !(Maybe Int)

_ppsIdleTimeout :: !(Maybe Int)

_ppsRetries :: !(Maybe Int)

_ppsPoolTimeout :: !(Maybe NominalDiffTime)

_ppsConnectionLifetime :: !(Maybe NominalDiffTime)

data DefaultPostgresPoolSettings

Constructors

DefaultPostgresPoolSettings

Fields

_dppsMaxConnections :: !Int

_dppsIdleTimeout :: !Int

_dppsRetries :: !Int

_dppsConnectionLifetime :: !(Maybe NominalDiffTime)

defaultPostgresPoolSettings :: DefaultPostgresPoolSettings

setPostgresPoolSettings :: PostgresPoolSettings

getDefaultPGPoolSettingIfNotExists :: Maybe PostgresPoolSettings -> DefaultPostgresPoolSettings -> (Int, Int, Int)

data SSLMode

Constructors

Disable

Allow

Prefer

Require

VerifyCA

VerifyFull

data CertVar

Constructors

CertVar String

CertLiteral String

newtype CertData

Constructors

CertData

Fields

unCert :: Text

data PGClientCerts p a

Constructors

PGClientCerts

Fields

pgcSslCert :: a

pgcSslKey :: a

pgcSslRootCert :: a

pgcSslMode :: SSLMode

pgcSslPassword :: Maybe p

data PostgresSourceConnInfo

Constructors

PostgresSourceConnInfo

Fields

_psciDatabaseUrl :: !UrlConf

_psciPoolSettings :: !(Maybe PostgresPoolSettings)

_psciUsePreparedStatements :: !Bool

_psciIsolationLevel :: !TxIsolation

_psciSslConfiguration :: !(Maybe (PGClientCerts CertVar CertVar))

psciUsePreparedStatements :: Lens' PostgresSourceConnInfo Bool

psciSslConfiguration :: Lens' PostgresSourceConnInfo (Maybe (PGClientCerts CertVar CertVar))

psciPoolSettings :: Lens' PostgresSourceConnInfo (Maybe PostgresPoolSettings)

psciIsolationLevel :: Lens' PostgresSourceConnInfo TxIsolation

psciDatabaseUrl :: Lens' PostgresSourceConnInfo UrlConf

data PostgresConnConfiguration

Constructors

PostgresConnConfiguration

Fields

_pccConnectionInfo :: !PostgresSourceConnInfo

_pccReadReplicas :: !(Maybe (NonEmpty PostgresSourceConnInfo))

pccReadReplicas :: Lens' PostgresConnConfiguration (Maybe (NonEmpty PostgresSourceConnInfo))

pccConnectionInfo :: Lens' PostgresConnConfiguration PostgresSourceConnInfo

Orphan instances

Generic TxIsolation

Instance details

Associated Types

type Rep TxIsolation :: Type -> Type

Methods

from :: TxIsolation -> Rep TxIsolation x

to :: Rep TxIsolation x -> TxIsolation

NFData TxIsolation

Instance details

Methods

rnf :: TxIsolation -> ()

Hashable TxIsolation

Instance details

Methods

hashWithSalt :: Int -> TxIsolation -> Int

hash :: TxIsolation -> Int

FromJSON TxIsolation

Instance details

Methods

parseJSON :: Value -> Parser TxIsolation

parseJSONList :: Value -> Parser [TxIsolation]

ToJSON TxIsolation

Instance details

Methods

toJSON :: TxIsolation -> Value

toEncoding :: TxIsolation -> Encoding

toJSONList :: [TxIsolation] -> Value

toEncodingList :: [TxIsolation] -> Encoding

Cacheable TxIsolation

Instance details

Methods

unchanged :: Accesses -> TxIsolation -> TxIsolation -> Bool

MonadTrace m => MonadTrace (TxET e m)

Instance details

Methods

trace :: Text -> TxET e m a -> TxET e m a

currentContext :: TxET e m TraceContext

currentReporter :: TxET e m Reporter

attachMetadata :: TracingMetadata -> TxET e m ()