Hasura.Backends.Postgres.Connection.MonadTx

Description

Postgres Connection MonadTx

This module contains MonadTx and related combinators.

MonadTx, a class which abstracts the QErr in TxE via MonadError.

The combinators are used for running, tracing, or otherwise perform database related tasks. Please consult the individual documentation for more information.

Documentation

class MonadError QErr m => MonadTx m where

Methods

liftTx :: TxE QErr a -> m a

runTx :: (MonadIO m, MonadBaseControl IO m) => PGExecCtx -> TxAccess -> TxET QErr m a -> ExceptT QErr m a

Executes the given query in a transaction of the specified mode, within the provided PGExecCtx.

runTxWithCtx :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadTrace m, UserInfoM m) => PGExecCtx -> TxAccess -> TxET QErr m a -> m a

runQueryTx :: (MonadIO m, MonadError QErr m) => PGExecCtx -> TxET QErr IO a -> m a

This runs the given set of statements (Tx) without wrapping them in BEGIN and COMMIT. This should only be used for running a single statement query!

setHeadersTx :: MonadIO m => SessionVariables -> TxET QErr m ()

sessionInfoJsonExp :: SessionVariables -> SQLExp

withUserInfo :: MonadIO m => UserInfo -> TxET QErr m a -> TxET QErr m a

setTraceContextInTx :: MonadIO m => TraceContext -> TxET QErr m ()

withTraceContext :: MonadIO m => TraceContext -> TxET QErr m a -> TxET QErr m a

Inject the trace context as a transaction-local variable, so that it can be picked up by any triggers (including event triggers).

checkDbConnection :: MonadTx m => m ()

doesSchemaExist :: MonadTx m => SchemaName -> m Bool

doesTableExist :: MonadTx m => SchemaName -> TableName -> m Bool

isExtensionAvailable :: MonadTx m => Text -> m Bool

enablePgcryptoExtension :: forall m. MonadTx m => ExtensionsSchema -> m ()

dropHdbCatalogSchema :: MonadTx m => m ()

Orphan instances

MonadTrace m => MonadTrace (TxET e m)

Instance details

Methods

trace :: Text -> TxET e m a -> TxET e m a

currentContext :: TxET e m TraceContext

currentReporter :: TxET e m Reporter

attachMetadata :: TracingMetadata -> TxET e m ()