Hasura.Backends.Postgres.Connection.Settings

Description

Postgres Connection Settings

This module contains types and combinators related to postgres connection, pool, and replica related settings.

Documentation

data PostgresPoolSettings

Constructors

PostgresPoolSettings

Fields

_ppsMaxConnections :: Maybe Int

_ppsTotalMaxConnections :: Maybe Int

_ppsIdleTimeout :: Maybe Int

_ppsRetries :: Maybe Int

_ppsPoolTimeout :: Maybe NominalDiffTime

_ppsConnectionLifetime :: Maybe NominalDiffTime

data DefaultPostgresPoolSettings

Constructors

DefaultPostgresPoolSettings

Fields

_dppsMaxConnections :: Int

_dppsIdleTimeout :: Int

_dppsRetries :: Int

_dppsConnectionLifetime :: Maybe NominalDiffTime

defaultPostgresPoolSettings :: DefaultPostgresPoolSettings

setPostgresPoolSettings :: PostgresPoolSettings

getDefaultPGPoolSettingIfNotExists :: Maybe PostgresPoolSettings -> DefaultPostgresPoolSettings -> (Int, Int, Int)

data SSLMode

Constructors

Disable

Allow

Prefer

Require

VerifyCA

VerifyFull

newtype CertVar

Constructors

CertVar String

newtype CertData

Constructors

CertData

Fields

unCert :: Text

data PGClientCerts p a

Constructors

PGClientCerts

Fields

pgcSslCert :: Maybe a

pgcSslKey :: Maybe a

pgcSslRootCert :: Maybe a

pgcSslMode :: SSLMode

pgcSslPassword :: Maybe p

data PostgresSourceConnInfo

Constructors

PostgresSourceConnInfo

Fields

_psciDatabaseUrl :: UrlConf

_psciPoolSettings :: Maybe PostgresPoolSettings

_psciUsePreparedStatements :: Bool

_psciIsolationLevel :: TxIsolation

_psciSslConfiguration :: Maybe (PGClientCerts CertVar CertVar)

defaultPostgresExtensionsSchema :: ExtensionsSchema

data KritiTemplate

`kriti-lang` template.

Constructors

KritiTemplate

Fields

_ktSource :: Text

Raw kriti template

_ktParsedAST :: ValueExt

Parsed kriti template

data ConnectionTemplate

Connection template for the dynamic DB connection.

Constructors

ConnectionTemplate

Fields

_ctVersion :: Int

Version for the connection template. Please read more about this in the dynamic DB connection RFC (Metadata API > Versioning).

_ctTemplate :: KritiTemplate

`kriti-lang` template for the dynamic DB connection.

newtype PostgresConnectionSetMemberName

Name of the member of a connection set.

Constructors

PostgresConnectionSetMemberName

Fields

getPostgresConnectionSetMemberName :: NonEmptyText

data PostgresConnectionSetMember

Constructors

PostgresConnectionSetMember

Fields

_pscmName :: PostgresConnectionSetMemberName

_pscmConnectionInfo :: PostgresSourceConnInfo

newtype PostgresConnectionSet

HashMap of the connection set. This is used for the dynamic DB connection feature.

Constructors

PostgresConnectionSet

Fields

getPostgresConnectionSet :: NEHashMap PostgresConnectionSetMemberName PostgresConnectionSetMember

data PostgresConnConfiguration

Constructors

PostgresConnConfiguration

Fields

_pccConnectionInfo :: PostgresSourceConnInfo

_pccReadReplicas :: Maybe (NonEmpty PostgresSourceConnInfo)

_pccExtensionsSchema :: ExtensionsSchema

_pccConnectionTemplate :: Maybe ConnectionTemplate

_pccConnectionSet :: Maybe PostgresConnectionSet

Orphan instances

FromJSON TxIsolation

Instance details

Methods

parseJSON :: Value -> Parser TxIsolation

parseJSONList :: Value -> Parser [TxIsolation]

ToJSON TxIsolation

Instance details

Methods

toJSON :: TxIsolation -> Value

toEncoding :: TxIsolation -> Encoding

toJSONList :: [TxIsolation] -> Value

toEncodingList :: [TxIsolation] -> Encoding

HasCodec TxIsolation

Instance details

Methods

codec :: JSONCodec TxIsolation

listCodecForStringCompatibility :: JSONCodec [TxIsolation]

Generic TxIsolation

Instance details

Associated Types

type Rep TxIsolation :: Type -> Type

Methods

from :: TxIsolation -> Rep TxIsolation x

to :: Rep TxIsolation x -> TxIsolation

NFData TxIsolation

Instance details

Methods

rnf :: TxIsolation -> ()

Hashable TxIsolation

Instance details

Methods

hashWithSalt :: Int -> TxIsolation -> Int

hash :: TxIsolation -> Int