Hasura.Backends.Postgres.DDL.EventTrigger

Documentation

data OpVar

Constructors

OLD

NEW

fetchUndeliveredEvents :: (MonadIO m, MonadError QErr m) => SourceConfig ('Postgres pgKind) -> SourceName -> MaintenanceMode -> FetchBatchSize -> m [Event ('Postgres pgKind)]

setRetry :: (MonadIO m, MonadError QErr m) => SourceConfig ('Postgres pgKind) -> Event ('Postgres pgKind) -> UTCTime -> Maybe MaintenanceModeVersion -> m ()

insertManualEvent :: (MonadIO m, MonadError QErr m) => SourceConfig ('Postgres pgKind) -> TableName ('Postgres pgKind) -> TriggerName -> Value -> UserInfo -> TraceContext -> m EventId

getMaintenanceModeVersion :: (MonadIO m, MonadError QErr m) => SourceConfig ('Postgres pgKind) -> m MaintenanceModeVersion

recordSuccess :: MonadIO m => SourceConfig ('Postgres pgKind) -> Event ('Postgres pgKind) -> Invocation 'EventType -> Maybe MaintenanceModeVersion -> m (Either QErr ())

recordError :: MonadIO m => SourceConfig ('Postgres pgKind) -> Event ('Postgres pgKind) -> Invocation 'EventType -> ProcessEventError -> Maybe MaintenanceModeVersion -> m (Either QErr ())

recordError' :: MonadIO m => SourceConfig ('Postgres pgKind) -> Event ('Postgres pgKind) -> Maybe (Invocation 'EventType) -> ProcessEventError -> Maybe MaintenanceModeVersion -> m (Either QErr ())

redeliverEvent :: (MonadIO m, MonadError QErr m) => SourceConfig ('Postgres pgKind) -> EventId -> m ()

dropTriggerAndArchiveEvents :: (MonadIO m, MonadError QErr m) => SourceConfig ('Postgres pgKind) -> TriggerName -> m ()

createTableEventTrigger :: (Backend ('Postgres pgKind), MonadIO m, MonadBaseControl IO m) => ServerConfigCtx -> PGSourceConfig -> QualifiedTable -> [ColumnInfo ('Postgres pgKind)] -> TriggerName -> TriggerOpsDef ('Postgres pgKind) -> m (Either QErr ())

updateColumnInEventTrigger :: QualifiedTable -> PGCol -> PGCol -> QualifiedTable -> EventTriggerConf ('Postgres pgKind) -> EventTriggerConf ('Postgres pgKind)

unlockEventsInSource :: MonadIO m => SourceConfig ('Postgres pgKind) -> Set EventId -> m (Either QErr Int)

insertInvocation :: Invocation 'EventType -> TxE QErr ()

insertPGManualEvent :: QualifiedTable -> TriggerName -> Value -> TxE QErr EventId

archiveEvents :: TriggerName -> TxE QErr ()

getMaintenanceModeVersionTx :: TxE QErr MaintenanceModeVersion

fetchEvents :: SourceName -> FetchBatchSize -> TxE QErr [Event ('Postgres pgKind)]

Lock and return events not yet being processed or completed, up to some limit. Process events approximately in created_at order, but we make no ordering guarentees; events can and will race. Nevertheless we want to ensure newer change events don't starve older ones.

fetchEventsMaintenanceMode :: SourceName -> FetchBatchSize -> MaintenanceModeVersion -> TxE QErr [Event ('Postgres pgKind)]

setSuccessTx :: Event ('Postgres pgKind) -> Maybe MaintenanceModeVersion -> TxE QErr ()

setErrorTx :: Event ('Postgres pgKind) -> Maybe MaintenanceModeVersion -> TxE QErr ()

setRetryTx :: Event ('Postgres pgKind) -> UTCTime -> Maybe MaintenanceModeVersion -> TxE QErr ()

dropTriggerQ :: TriggerName -> TxE QErr ()

checkEvent :: EventId -> TxE QErr ()

markForDelivery :: EventId -> TxE QErr ()

redeliverEventTx :: EventId -> TxE QErr ()

unlockEventsTx :: [EventId] -> TxE QErr Int

unlockEvents takes an array of EventId and unlocks them. This function is called when a graceful shutdown is initiated.

pgIdenTrigger :: Ops -> TriggerName -> Text

pgIdenTrigger is a method used to construct the name of the pg function used for event triggers which are present in the hdb_catalog schema.

mkTriggerQ :: forall pgKind m. (Backend ('Postgres pgKind), MonadTx m, MonadReader ServerConfigCtx m) => TriggerName -> QualifiedTable -> [ColumnInfo ('Postgres pgKind)] -> Ops -> SubscribeOpSpec ('Postgres pgKind) -> m ()

Define the pgSQL trigger functions on database events.

mkAllTriggersQ :: forall pgKind m. (Backend ('Postgres pgKind), MonadTx m, MonadReader ServerConfigCtx m) => TriggerName -> QualifiedTable -> [ColumnInfo ('Postgres pgKind)] -> TriggerOpsDef ('Postgres pgKind) -> m ()