Hasura.Backends.Postgres.Execute.LiveQuery

Description

Postgres Execute LiveQuery

Multiplex is an optimization which allows us to group similar queries into a single query, and routing the response rows afterwards. See https://hasura.io/docs/latest/graphql/core/databases/postgres/subscriptions/execution-and-performance.html for more details

See Execute.

Documentation

data QueryParametersInfo (b :: BackendType)

Internal: Used to collect information about various parameters of a subscription field's AST as we resolve them to SQL expressions.

Constructors

QueryParametersInfo

Fields

_qpiReusableVariableValues :: !(HashMap Name (ColumnValue b))

_qpiSyntheticVariableValues :: !(Seq (ColumnValue b))

_qpiReferencedSessionVariables :: !(HashSet SessionVariable)

The session variables that are referenced in the query root fld's AST. This information is used to determine a cohort's required session variables

qpiSyntheticVariableValues :: forall b. Lens' (QueryParametersInfo b) (Seq (ColumnValue b))

qpiReusableVariableValues :: forall b. Lens' (QueryParametersInfo b) (HashMap Name (ColumnValue b))

qpiReferencedSessionVariables :: forall b. Lens' (QueryParametersInfo b) (HashSet SessionVariable)

validateVariables :: forall pgKind f m. (Traversable f, MonadError QErr m, MonadIO m) => PGExecCtx -> f (ColumnValue ('Postgres pgKind)) -> m (ValidatedVariables f)

Checks if the provided arguments are valid values for their corresponding types. | Generates SQL of the format "select v1::t1, v2::t2 ..."

newtype MultiplexedQuery

Constructors

MultiplexedQuery

Fields

unMultiplexedQuery :: Query

toSQLFromItem :: (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => Alias -> QueryDB ('Postgres pgKind) Void SQLExp -> FromItem

mkMultiplexedQuery :: (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => InsOrdHashMap Name (QueryDB ('Postgres pgKind) Void SQLExp) -> MultiplexedQuery

resolveMultiplexedValue :: (MonadState (QueryParametersInfo ('Postgres pgKind)) m, MonadError QErr m) => SessionVariables -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

Resolves an UnresolvedVal by converting UVPG values to SQL expressions that refer to the result_vars input object, collecting information about various parameters of the query along the way.

executeMultiplexedQuery :: MonadTx m => MultiplexedQuery -> [(CohortId, CohortVariables)] -> m [(CohortId, ByteString)]

executeQuery :: (MonadTx m, FromRow a) => Query -> [(CohortId, CohortVariables)] -> m [a]

Internal; used by both executeMultiplexedQuery and pgDBLiveQueryExplain.