Hasura.Backends.Postgres.Execute.Mutation

Description

Postgres Execute Mutation

Generic combinators for translating and excecuting IR mutation statements. Used by the specific mutation modules, e.g. Insert.

See Execute.

Documentation

data MutateResp (b :: BackendType) a

Constructors

MutateResp

Fields

_mrAffectedRows :: Int

_mrReturningColumns :: [ColumnValues b a]

execUpdateQuery :: forall pgKind m. (MonadTx m, Backend ('Postgres pgKind), PostgresTranslateSelect pgKind, MonadReader QueryTagsComment m, MonadIO m) => StringifyNumbers -> Maybe NamingCase -> UserInfo -> (AnnotatedUpdate ('Postgres pgKind), Seq PrepArg) -> m EncJSON

execDeleteQuery :: forall pgKind m. (MonadTx m, MonadIO m, Backend ('Postgres pgKind), PostgresTranslateSelect pgKind, MonadReader QueryTagsComment m) => StringifyNumbers -> Maybe NamingCase -> UserInfo -> (AnnDel ('Postgres pgKind), Seq PrepArg) -> m EncJSON

execInsertQuery :: (MonadTx m, Backend ('Postgres pgKind), PostgresTranslateSelect pgKind, MonadReader QueryTagsComment m, MonadIO m) => StringifyNumbers -> Maybe NamingCase -> UserInfo -> (InsertQueryP1 ('Postgres pgKind), Seq PrepArg) -> m EncJSON

executeMutationOutputQuery

Arguments

:: forall pgKind m. (MonadTx m, Backend ('Postgres pgKind), PostgresTranslateSelect pgKind, MonadReader QueryTagsComment m, MonadIO m)

=> UserInfo

-> QualifiedTable

-> [ColumnInfo ('Postgres pgKind)]

-> Maybe Int

-> MutationCTE

-> MutationOutput ('Postgres pgKind)

-> StringifyNumbers

-> Maybe NamingCase

-> [PrepArg]
Prepared params

-> m EncJSON

mutateAndFetchCols :: forall pgKind. (Backend ('Postgres pgKind), PostgresTranslateSelect pgKind) => UserInfo -> QualifiedTable -> [ColumnInfo ('Postgres pgKind)] -> (MutationCTE, Seq PrepArg) -> StringifyNumbers -> Maybe NamingCase -> TxE QErr (MutateResp ('Postgres pgKind) TxtEncodedVal)

type ValidateInputPayloadVersion = Int

validateInputPayloadVersion :: ValidateInputPayloadVersion

newtype ValidateInputErrorResponse

Constructors

ValidateInputErrorResponse

Fields

_vierMessage :: Text

data HttpHandlerLog

Constructors

HttpHandlerLog

Fields

_hhlUrl :: Text

_hhlRequest :: Value

_hhlRequestHeaders :: [HeaderConf]

_hhlResponse :: Value

_hhlResponseStatus :: Int

data ValidateInsertInputLog

Constructors

VIILHttpHandler HttpHandlerLog

type InsertValidationPayloadMap pgKind = InsOrdHashMap (TableName ('Postgres pgKind)) ([AnnotatedInsertRow ('Postgres pgKind) (UnpreparedValue ('Postgres pgKind))], ValidateInput ResolvedWebhook)

Map of table name and the value that is being inserted for that table This map is helpful for collecting all the insert mutation arguments for the nested tables and then sending them all at onve to the input validation webhook.

validateUpdateMutation :: forall pgKind m. (MonadError QErr m, MonadIO m, MonadTrace m) => Environment -> Manager -> Logger Hasura -> UserInfo -> ResolvedWebhook -> [HeaderConf] -> Timeout -> Bool -> [Header] -> AnnotatedUpdateG ('Postgres pgKind) Void (UnpreparedValue ('Postgres pgKind)) -> Maybe (HashMap Name (Value Variable)) -> m ()

validateDeleteMutation :: forall m pgKind. (MonadError QErr m, MonadIO m, MonadTrace m) => Environment -> Manager -> Logger Hasura -> UserInfo -> ResolvedWebhook -> [HeaderConf] -> Timeout -> Bool -> [Header] -> AnnDelG ('Postgres pgKind) Void (UnpreparedValue ('Postgres pgKind)) -> Maybe (HashMap Name (Value Variable)) -> m ()

validateMutation :: forall m. (MonadError QErr m, MonadIO m, MonadTrace m) => Environment -> Manager -> Logger Hasura -> UserInfo -> ResolvedWebhook -> [HeaderConf] -> Timeout -> Bool -> [Header] -> Value -> m ()