Hasura.Backends.Postgres.Execute.Subscription

Description

Postgres Execute subscription

Multiplex is an optimization which allows us to group similar queries into a single query, and routing the response rows afterwards. See https://hasura.io/docs/latest/graphql/core/databases/postgres/subscriptions/execution-and-performance.html for more details

See Execute.

Documentation

newtype MultiplexedQuery

Constructors

MultiplexedQuery

Fields

unMultiplexedQuery :: Query

data QueryParametersInfo (b :: BackendType)

Internal: Used to collect information about various parameters of a subscription field's AST as we resolve them to SQL expressions.

Constructors

QueryParametersInfo

Fields

_qpiReusableVariableValues :: HashMap Name (ColumnValue b)

_qpiSyntheticVariableValues :: Seq (ColumnValue b)

_qpiReferencedSessionVariables :: HashSet SessionVariable

The session variables that are referenced in the query root fld's AST. This information is used to determine a cohort's required session variables

mkMultiplexedQuery :: forall (pgKind :: PostgresKind) m. (Backend ('Postgres pgKind), PostgresTranslateSelect pgKind, MonadIO m, MonadError QErr m) => UserInfo -> InsOrdHashMap Name (QueryDB ('Postgres pgKind) Void SQLExp) -> m MultiplexedQuery

mkStreamingMultiplexedQuery :: forall (pgKind :: PostgresKind) m. (Backend ('Postgres pgKind), PostgresTranslateSelect pgKind, MonadIO m, MonadError QErr m) => UserInfo -> (Name, QueryDB ('Postgres pgKind) Void SQLExp) -> m MultiplexedQuery

resolveMultiplexedValue :: forall (pgKind :: PostgresKind) m. (MonadState (QueryParametersInfo ('Postgres pgKind)) m, MonadError QErr m) => SessionVariables -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

Resolves an UnresolvedVal by converting UVPG values to SQL expressions that refer to the result_vars input object, collecting information about various parameters of the query along the way.

validateVariablesTx :: forall (pgKind :: PostgresKind) f m. (Traversable f, MonadTx m, MonadIO m) => f (ColumnValue ('Postgres pgKind)) -> m (ValidatedVariables f)

Checks if the provided arguments are valid values for their corresponding types. | Generates SQL of the format "select v1::t1, v2::t2 ..."

executeMultiplexedQuery :: MonadTx m => MultiplexedQuery -> [(CohortId, CohortVariables)] -> m [(CohortId, ByteString)]

executeStreamingMultiplexedQuery :: MonadTx m => MultiplexedQuery -> [(CohortId, CohortVariables)] -> m [(CohortId, ByteString, ViaJSON CursorVariableValues)]

executeQuery :: (MonadTx m, FromRes a) => Query -> [(CohortId, CohortVariables)] -> m a

Internal; used by both executeMultiplexedQuery, executeStreamingMultiplexedQuery and pgDBSubscriptionExplain.

data SubscriptionType

Constructors

Streaming

LiveQuery