Hasura.Backends.Postgres.Instances.Types

Description

Postgres Instances Types

Defines a Backend type class instance for Postgres.

Documentation

class (Representable (PgExtraTableMetadata pgKind), ToJSON (PgExtraTableMetadata pgKind), FromJSON (PgExtraTableMetadata pgKind)) => PostgresBackend (pgKind :: PostgresKind)

This class is an implementation detail of Backend. Some types of Backend differ across different Postgres "kinds". This class alllows each "kind" to specify its own specific implementation. All common code is directly part of the Backend instance.

Note: Users shouldn't ever put this as a constraint. Use `Backend ('Postgres pgKind)` instead.

Associated Types

type PgExtraTableMetadata pgKind :: Type

Orphan instances

(HasTag ('Postgres pgKind), Typeable ('Postgres pgKind), PostgresBackend pgKind, FromJSON (BackendSourceKind ('Postgres pgKind))) => Backend ('Postgres pgKind)

Instance details

Associated Types

type BackendConfig ('Postgres pgKind)

type SourceConnConfiguration ('Postgres pgKind)

type SourceConfig ('Postgres pgKind)

type TableName ('Postgres pgKind)

type FunctionName ('Postgres pgKind)

type RawFunctionInfo ('Postgres pgKind)

type ConstraintName ('Postgres pgKind)

type BasicOrderType ('Postgres pgKind)

type NullsOrderType ('Postgres pgKind)

type CountType ('Postgres pgKind)

type Column ('Postgres pgKind)

type ScalarValue ('Postgres pgKind)

type ScalarType ('Postgres pgKind)

type BooleanOperators ('Postgres pgKind) :: Type -> Type

type SQLExpression ('Postgres pgKind)

type ComputedFieldDefinition ('Postgres pgKind)

type ScalarSelectionArguments ('Postgres pgKind)

type ExtraTableMetadata ('Postgres pgKind)

type FunctionArgument ('Postgres pgKind)

type FunctionArgumentExp ('Postgres pgKind) :: Type -> Type

type ComputedFieldImplicitArguments ('Postgres pgKind)

type ComputedFieldReturn ('Postgres pgKind)

type BackendUpdate ('Postgres pgKind) :: Type -> Type

type BackendInsert ('Postgres pgKind) :: Type -> Type

type XComputedField ('Postgres pgKind)

type XRelay ('Postgres pgKind)

type XNodesAgg ('Postgres pgKind)

type XNestedInserts ('Postgres pgKind)

type XStreamingSubscription ('Postgres pgKind)

Methods

isComparableType :: ScalarType ('Postgres pgKind) -> Bool

isNumType :: ScalarType ('Postgres pgKind) -> Bool

textToScalarValue :: Maybe Text -> ScalarValue ('Postgres pgKind)

parseScalarValue :: ScalarType ('Postgres pgKind) -> Value -> Either QErr (ScalarValue ('Postgres pgKind))

scalarValueToJSON :: ScalarValue ('Postgres pgKind) -> Value

functionToTable :: FunctionName ('Postgres pgKind) -> TableName ('Postgres pgKind)

tableToFunction :: TableName ('Postgres pgKind) -> FunctionName ('Postgres pgKind)

computedFieldFunction :: ComputedFieldDefinition ('Postgres pgKind) -> FunctionName ('Postgres pgKind)

computedFieldReturnType :: ComputedFieldReturn ('Postgres pgKind) -> ComputedFieldReturnType ('Postgres pgKind)

fromComputedFieldImplicitArguments :: v -> ComputedFieldImplicitArguments ('Postgres pgKind) -> [FunctionArgumentExp ('Postgres pgKind) v]

tableGraphQLName :: TableName ('Postgres pgKind) -> Either QErr Name

functionGraphQLName :: FunctionName ('Postgres pgKind) -> Either QErr Name

scalarTypeGraphQLName :: ScalarType ('Postgres pgKind) -> Either QErr Name

snakeCaseTableName :: TableName ('Postgres pgKind) -> Text