Hasura.Backends.Postgres.SQL.DML

Description

Postgres SQL DML

Provide types and combinators for defining Postgres SQL queries and mutations.

Documentation

data Select

An select statement that does not require mutation CTEs.

See SelectWithG or SelectWithG for select statements with mutations as CTEs.

Constructors

Select [(TableAlias, Select)] (Maybe DistinctExpr) [Extractor] (Maybe FromExp) (Maybe WhereFrag) (Maybe GroupByExp) (Maybe HavingExp) (Maybe OrderByExp) (Maybe LimitExp) (Maybe OffsetExp)

mkSelect :: Select

An empty select statement.

dummySelectList :: [Extractor]

A dummy select list to avoid an empty select list, which doesn't work for cockroach db. This is just the value 1 without an alias.

newtype LimitExp

Constructors

LimitExp SQLExp

newtype OffsetExp

Constructors

OffsetExp SQLExp

newtype OrderByExp

Constructors

OrderByExp (NonEmpty OrderByItem)

data OrderByItem

Constructors

OrderByItem SQLExp (Maybe OrderType) (Maybe NullsOrder)

data OrderType

Order by ascending or descending

Constructors

OTAsc

OTDesc

data NullsOrder

Constructors

NullsFirst

NullsLast

newtype GroupByExp

Constructors

GroupByExp [SQLExp]

newtype FromExp

Constructors

FromExp [FromItem]

mkIdenFromExp :: TableIdentifier -> FromExp

mkSimpleFromExp :: QualifiedTable -> FromExp

mkSelFromExp :: Bool -> Select -> TableName -> FromItem

mkRowExp :: [Extractor] -> SQLExp

newtype HavingExp

Constructors

HavingExp BoolExp

newtype WhereFrag

Constructors

WhereFrag BoolExp

mkSIdenExp :: IsIdentifier a => a -> SQLExp

mkQIdenExp :: (IsIdentifier a, IsIdentifier b) => a -> b -> SQLExp

data Qual

Constructors

QualifiedIdentifier Identifier (Maybe TypeAnn)

QualTable QualifiedTable

QualVar Text

mkQual :: QualifiedTable -> Qual

mkQIdentifier :: (IsIdentifier a, IsIdentifier b) => a -> b -> QIdentifier

mkQIdentifierTable :: IsIdentifier a => QualifiedTable -> a -> QIdentifier

mkIdentifierSQLExp :: forall a. IsIdentifier a => Qual -> a -> SQLExp

data QIdentifier

Constructors

QIdentifier Qual Identifier

data ColumnOp

Constructors

ColumnOp

Fields

_colOp :: SQLOp

_colExp :: SQLExp

newtype SQLOp

Constructors

SQLOp Text

incOp :: SQLOp

mulOp :: SQLOp

jsonbPathOp :: SQLOp

jsonbConcatOp :: SQLOp

jsonbDeleteOp :: SQLOp

jsonbDeleteAtPathOp :: SQLOp

newtype TypeAnn

Constructors

TypeAnn Text

mkTypeAnn :: CollectableType PGScalarType -> TypeAnn

intTypeAnn :: TypeAnn

numericTypeAnn :: TypeAnn

textTypeAnn :: TypeAnn

textArrTypeAnn :: TypeAnn

jsonTypeAnn :: TypeAnn

jsonbTypeAnn :: TypeAnn

boolTypeAnn :: TypeAnn

data CountType

Constructors

CTStar

CTSimple [PGCol]

CTDistinct [PGCol]

newtype TupleExp

Constructors

TupleExp [SQLExp]

data SQLExp

Constructors

SEPrep Int

SENull

SELit Text

SEUnsafe Text

SESelect Select

SEStar (Maybe Qual)
all fields (*) or all fields from relation (iden.*)

SEIdentifier Identifier
A column name

SERowIdentifier Identifier
SEIdentifier and SERowIdentifier are distinguished for easier rewrite rules

SEQIdentifier QIdentifier
A qualified column name

SEFnApp Text [SQLExp] (Maybe OrderByExp)
this is used to apply a sql function to an expression. The Text is the function name

SEOpApp SQLOp [SQLExp]

SETyAnn SQLExp TypeAnn

SECond BoolExp SQLExp SQLExp

SEBool BoolExp

SEExcluded Identifier

SEArray [SQLExp]

SEArrayIndex SQLExp SQLExp

SETuple TupleExp

SECount CountType

SENamedArg Identifier SQLExp

SEFunction FunctionExp

newtype ColumnAlias

Represents an alias assignment for a column

Constructors

ColumnAlias

Fields

getColumnAlias :: Identifier

toColumnAlias :: IsIdentifier a => a -> ColumnAlias

newtype TableAlias

Represents an alias assignment for a table, relation or row

Constructors

TableAlias

Fields

getTableAlias :: Identifier

mkTableAlias :: Text -> TableAlias

Create a table alias.

tableAliasToIdentifier :: TableAlias -> TableIdentifier

Create a table identifier from a table alias.

toTableAlias :: IsIdentifier a => a -> TableAlias

countStar :: SQLExp

intToSQLExp :: Int -> SQLExp

int64ToSQLExp :: Int64 -> SQLExp

data Extractor

Extractor can be used to apply Postgres alias to a column

Constructors

Extractor SQLExp (Maybe ColumnAlias)

mkSQLOpExp :: SQLOp -> SQLExp -> SQLExp -> SQLExp

columnDefaultValue :: SQLExp

handleIfNull :: SQLExp -> SQLExp -> SQLExp

applyJsonBuildObj :: [SQLExp] -> SQLExp

applyJsonBuildArray :: [SQLExp] -> SQLExp

applyRowToJson :: [Extractor] -> SQLExp

applyUppercase :: SQLExp -> SQLExp

mkExtr :: IsIdentifier a => a -> Extractor

data DistinctExpr

Constructors

DistinctSimple

DistinctOn [SQLExp]

data FunctionArgs

Constructors

FunctionArgs [SQLExp] (HashMap Text SQLExp)

data FunctionDefinitionListItem

Constructors

FunctionDefinitionListItem

Fields

_dliColumn :: ColumnAlias

_dliType :: PGScalarType

data FunctionAlias

We can alias the result of a function call that returns a SETOF RECORD by naming the result relation, and the columns and their types. For example:

SELECT * FROM
function_returns_record(arg1, arg2 ...) AS relation_name(column_1 column_1_type, column_2 column_2_type, ...)

Note: a function that returns a table (instead of a record) cannot name the types as seen in the above example.

Constructors

FunctionAlias TableAlias (Maybe [FunctionDefinitionListItem])

mkFunctionAlias :: QualifiedObject FunctionName -> Maybe [(ColumnAlias, PGScalarType)] -> FunctionAlias

Construct a function alias which represents the "relation signature" for the function invocation, Using the function name as the relation name, and the columns as the relation schema.

data FunctionExp

A function call

Constructors

FunctionExp QualifiedFunction FunctionArgs (Maybe FunctionAlias)

data FromItem

See from_item in https://www.postgresql.org/docs/current/sql-select.html

Constructors

FISimple QualifiedTable (Maybe TableAlias)
A simple table

FIIdentifier TableIdentifier
An identifier (from CTEs)

FIFunc FunctionExp
A function call (that should return a relation (SETOF) and not a scalar)

FIUnnest [SQLExp] TableAlias [ColumnAlias]
unnest converts (an) array(s) to a relation.

We have: * The unnest function arguments * The relation alias * A list of column aliases

See unnest in https://www.postgresql.org/docs/current/functions-array.html.

FISelect Lateral Select TableAlias

FISelectWith Lateral (SelectWithG Select) TableAlias

FIValues ValuesExp TableAlias (Maybe [ColumnAlias])

FIJoin JoinExpr

mkSelFromItem :: Select -> TableAlias -> FromItem

mkSelectWithFromItem :: SelectWithG Select -> TableAlias -> FromItem

mkLateralFromItem :: Select -> TableAlias -> FromItem

newtype Lateral

Constructors

Lateral Bool

data JoinExpr

Constructors

JoinExpr FromItem JoinType FromItem JoinCond

data JoinType

Constructors

Inner

LeftOuter

data JoinCond

Constructors

JoinOn BoolExp

JoinUsing [Identifier]

data BoolExp

Constructors

BELit Bool

BEBin BinOp BoolExp BoolExp

BENot BoolExp

BECompare CompareOp SQLExp SQLExp

BECompareAny CompareOp SQLExp SQLExp

BENull SQLExp

BENotNull SQLExp

BEExists Select

BEIN SQLExp [SQLExp]

BEExp SQLExp

simplifyBoolExp :: BoolExp -> BoolExp

mkExists :: FromItem -> BoolExp -> BoolExp

data BinOp

Constructors

AndOp

OrOp

data CompareOp

Constructors

SEQ

SGT

SLT

SNE

SGTE

SLTE

SLIKE

SNLIKE

SILIKE

SNILIKE

SSIMILAR

SNSIMILAR

SREGEX

SIREGEX

SNREGEX

SNIREGEX

SContains

SContainedIn

SHasKey

SHasKeysAny

SHasKeysAll

SMatchesFulltext

data SQLDelete

Constructors

SQLDelete QualifiedTable (Maybe UsingExp) (Maybe WhereFrag) (Maybe RetExp)

data SQLUpdate

Constructors

SQLUpdate

Fields

upTable :: QualifiedTable

upSet :: SetExp

upFrom :: Maybe FromExp

upWhere :: Maybe WhereFrag

upRet :: Maybe RetExp

newtype SetExp

Constructors

SetExp [SetExpItem]

newtype SetExpItem

Constructors

SetExpItem (PGCol, SQLExp)

buildUpsertSetExp :: [PGCol] -> HashMap PGCol SQLExp -> SetExp

newtype RetExp

Constructors

RetExp [Extractor]

selectStar :: Extractor

selectStar' :: Qual -> Extractor

returningStar :: RetExp

data SQLConflictTarget

Constructors

SQLColumn [PGCol]

SQLConstraint ConstraintName

data SQLConflict

Constructors

DoNothing (Maybe SQLConflictTarget)

Update SQLConflictTarget SetExp (Maybe WhereFrag)

newtype ValuesExp

Constructors

ValuesExp [TupleExp]

data SQLInsert

Constructors

SQLInsert

Fields

siTable :: QualifiedTable

siCols :: [PGCol]

siValues :: ValuesExp

siConflict :: Maybe SQLConflict

siRet :: Maybe RetExp

data TopLevelCTE

Top-level Common Table Expression statement.

A top level CTE can be a query or a mutation statement.

Postgres supports mutations only in top-level CTEs. See https://www.postgresql.org/docs/current/queries-with.html#QUERIES-WITH-MODIFYING

Constructors

CTESelect Select

CTEInsert SQLInsert

CTEUpdate SQLUpdate

CTEDelete SQLDelete

data SelectWithG statement

A SELECT statement with Common Table Expressions. https://www.postgresql.org/docs/current/queries-with.html

These CTEs are determined by the statement parameter. Currently they are either TopLevelCTE, which allow for a query or mutation statement, or Select, which only allow for querying results.

The distinction is required because Postgres only supports mutations in CTEs at the top level. See https://www.postgresql.org/docs/current/queries-with.html#QUERIES-WITH-MODIFYING

Constructors

SelectWith [(TableAlias, statement)] Select

type SelectWith = SelectWithG TopLevelCTE

A top-level select with CTEs.