Hasura.Backends.Postgres.SQL.Error

Description

Functions and datatypes for interpreting Postgres errors.

Documentation

data PGErrorType

The top-level error code type. Errors in Postgres are divided into different classes, which are further subdivided into individual error codes. Even if a particular status code is not known to the application, it’s possible to determine its class and handle it appropriately.

Constructors

PGDataException !(Maybe (PGErrorCode PGDataException))

PGIntegrityConstraintViolation !(Maybe (PGErrorCode PGIntegrityConstraintViolation))

PGSyntaxErrorOrAccessRuleViolation !(Maybe (PGErrorCode PGSyntaxErrorOrAccessRuleViolation))

data PGErrorCode a

Constructors

PGErrorGeneric
represents errors that have the non-specific 000 status code

PGErrorSpecific !a
represents errors with a known, more specific status code

data PGDataException

Constructors

PGInvalidDatetimeFormat

PGInvalidParameterValue

PGInvalidEscapeSequence

PGInvalidTextRepresentation

data PGIntegrityConstraintViolation

Constructors

PGRestrictViolation

PGNotNullViolation

PGForeignKeyViolation

PGUniqueViolation

PGCheckViolation

PGExclusionViolation

data PGSyntaxErrorOrAccessRuleViolation

Constructors

PGUndefinedObject

PGInvalidColumnReference

_PGSyntaxErrorOrAccessRuleViolation :: Prism' PGErrorType (Maybe (PGErrorCode PGSyntaxErrorOrAccessRuleViolation))

_PGIntegrityConstraintViolation :: Prism' PGErrorType (Maybe (PGErrorCode PGIntegrityConstraintViolation))

_PGDataException :: Prism' PGErrorType (Maybe (PGErrorCode PGDataException))

_PGErrorSpecific :: forall a a. Prism (PGErrorCode a) (PGErrorCode a) a a

_PGErrorGeneric :: forall a. Prism' (PGErrorCode a) ()

pgErrorType :: PGStmtErrDetail -> Maybe PGErrorType