Hasura.Backends.Postgres.SQL.IdentifierUniqueness

Description

Postgres SQL IdentifierUniqueness

Bypass the Postgres limitation of truncating identifiers to 63 characters long by prepending unique numbers.

See Postgres docs: https://www.postgresql.org/docs/current/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS

Also see Note [Postgres identifier length limitations]

API

prefixNumToAliases :: Select -> Select

Prefix an Int to all aliases to preserve the uniqueness of identifiers. See Note [Postgres identifier length limitations] above.

prefixNumToAliasesSelectWith :: SelectWithG Select -> SelectWithG Select

Data types

runUniq :: Uniq a -> a

emptyUniqState :: UniqState

data UniqState

Constructors

UniqState

Fields

_uqFreshId :: Int

Used to generate fresh prefix ids

_uqIdentifierMap :: HashMap Identifier Int

Mapping from an identifier to their prefix id

type Uniq = State UniqState

Utilities

mkPrefixedName :: Identifier -> Int -> Identifier

attach a prefix id to an identifier

addAlias :: TableAlias -> Uniq TableAlias

Return a prefixed alias by: 1. Generating a fresh prefix id for the alias' identifier 2. Adding the identifier + id to the identifier mapping

getIdentifier :: Identifier -> Uniq Identifier

Search for the identifier in the identifier mapping and return a prefixed identifier if found, or the original identifier if not found in the identifier map.

restoringIdentifierMap :: Uniq a -> Uniq a

Run an action that might change the identifier mapping and discard the changes made to the identifier map.

Algorithm

uSelectWith :: SelectWithG Select -> Uniq (SelectWithG Select)

We run the algorithm on each CTE separately and discard the identifier mapping, then we run the algorithm on the main select and return that result (with the identifier mapping generated from that).

uSelect :: Select -> Uniq Select

We go in order of each component in the select, starting with the from and CTE clauses (as those introduce new names). We return a transformed Select (with the identifier map in the Uniq state).

We transform the table names but not column names.

uFromExp :: FromExp -> Uniq FromExp

Transform every from_item. Potentially introduces a new alias.

uFromItem :: FromItem -> Uniq FromItem

Transform a single from_item. Potentially introduces a new alias.

uFunctionExp :: FunctionExp -> Uniq FunctionExp

Transform a function call expression.

uFunctionArgs :: FunctionArgs -> Uniq FunctionArgs

Transform function call arguments.

uFunctionAlias :: FunctionAlias -> Uniq FunctionAlias

Transform a function call alias.

uJoinExp :: JoinExpr -> Uniq JoinExpr

Transform join expressions. Potentially introduces a new alias.

uJoinCond :: JoinCond -> Uniq JoinCond

Transform Join condition. ON join condition might contain references to table names and aliases.

uBoolExp :: BoolExp -> Uniq BoolExp

Transform boolean expression.

The boolean expression structure does not contain a table name currently, So we look for SQLExps and transform those, as those may contain table names and aliases.

We discard new aliases that might be generated here because we don't use them outside of the boolean expression.

uSqlExp :: SQLExp -> Uniq SQLExp

Transform a SQL expression. We look for table names and aliases and rename them if needed. SQL expressions do not introduce new table aliases, so we discard the new aliases that might be generated here.

uOrderBy :: OrderByExp -> Uniq OrderByExp

Transform order by clauses. Since order by does not introduce new aliases we can discard the new names that might be added, this is already done by uSqlExp though.