Hasura.Backends.Postgres.SQL.RenameIdentifiers

Description

Postgres SQL Rename Identifiers

Prefix table names with underscores to avoid issues where column names and tables conflict. This can happen because we give columns and tables the name root for some reason, and that can trip up row_to_json. See https://github.com/PostgREST/postgrest/issues/993#issuecomment-340377813. An alternative solution would be to not create a TableAlias with the name root, but that seemed a bit complicated for me to do at the time.

Bypass the Postgres limitation of truncating identifiers to 63 characters long by prepending they identifier's md5 hash when they are longer than 63 characters.

We do both operations in the same traversal for performance reasons, but a simpler implementation of (1) would be transformBi prefixHash from the uniplate or the generic-plate package.

See Postgres docs: https://www.postgresql.org/docs/current/sql-syntax-lexical.html#SQL-SYNTAX-IDENTIFIERS

Exported API

renameIdentifiers :: Select -> Select

Prefix table names with undescores and rename long identifiers.

renameIdentifiersSelectWith :: SelectWithG Select -> SelectWithG Select

prefix table names with undescores and rename long identifiers.

renameIdentifiersSelectWithTopLevelCTE :: SelectWithG TopLevelCTE -> SelectWithG TopLevelCTE

prefix table names with undescores and rename long identifiers.