Hasura.Backends.Postgres.SQL.Types

Description

Postgres SQL Types

Combinators and helpers for dealing with Postgres types such as strings, numerals, geography and geometry, etc.

Documentation

newtype Identifier

Constructors

Identifier

Fields

getIdenTxt :: Text

class IsIdentifier a where

Methods

toIdentifier :: a -> Identifier

newtype TableIdentifier

The type of identifiers representing tabular values. While we are transitioning away from Identifier we provisionally export the value constructor.

Constructors

TableIdentifier

Fields

unTableIdentifier :: Text

tableIdentifierToIdentifier :: TableIdentifier -> Identifier

Temporary conversion function, to be removed once Identifier has been entirely split into TableIdentifier and ColumnIdentifier.

identifierToTableIdentifier :: Identifier -> TableIdentifier

Temporary conversion function, to be removed once Identifier has been entirely split into TableIdentifier and ColumnIdentifier.

newtype ColumnIdentifier

The type of identifiers representing scalar values

Constructors

ColumnIdentifier

Fields

unColumnIdentifier :: Text

pgFmtIdentifier :: Text -> Text

pgFmtLit :: Text -> Text

newtype TableName

Constructors

TableName

Fields

getTableTxt :: Text

isView :: TableType -> Bool

newtype ConstraintName

Constructors

ConstraintName

Fields

getConstraintTxt :: Text

newtype FunctionName

Constructors

FunctionName

Fields

getFunctionTxt :: Text

newtype SchemaName

Constructors

SchemaName

Fields

getSchemaTxt :: Text

publicSchema :: SchemaName

hdbCatalogSchema :: SchemaName

data QualifiedObject a

Constructors

QualifiedObject

Fields

qSchema :: SchemaName

qName :: a

qualifiedObjectToText :: ToTxt a => QualifiedObject a -> Text

snakeCaseQualifiedObject :: ToTxt a => QualifiedObject a -> Text

getIdentifierQualifiedObject :: ToTxt a => QualifiedObject a -> Either QErr GQLNameIdentifier

namingConventionSupport :: SupportedNamingCase

qualifiedObjectToName :: (ToTxt a, MonadError QErr m) => QualifiedObject a -> m Name

type QualifiedTable = QualifiedObject TableName

Represents a database table qualified with the schema name.

type QualifiedFunction = QualifiedObject FunctionName

newtype PGDescription

Constructors

PGDescription

Fields

getPGDescription :: Text

data PGCol

unsafePGCol :: Text -> PGCol

showPGCols :: (Foldable t, Functor t) => t PGCol -> Text

data PGScalarType

Constructors

PGSmallInt

PGInteger

PGBigInt

PGSerial

PGBigSerial

PGFloat

PGDouble

PGNumeric

PGMoney

PGBoolean

PGChar

PGVarchar

PGText

PGCitext

PGDate

PGTimeStamp

PGTimeStampTZ

PGTimeTZ

PGJSON

PGJSONB

PGGeometry

PGGeography

PGRaster

PGUUID

PGLtree

PGLquery

PGLtxtquery

PGArray PGScalarType

PGUnknown Text

PGCompositeScalar Text

PGEnumScalar Text

pgScalarTypeToText :: PGScalarType -> Text

textToPGScalarType :: Text -> PGScalarType

pgScalarTranslations :: [(Text, PGScalarType)]

isNumType :: PGScalarType -> Bool

stringTypes :: [PGScalarType]

isStringType :: PGScalarType -> Bool

isJSONType :: PGScalarType -> Bool

isComparableType :: PGScalarType -> Bool

isBigNum :: PGScalarType -> Bool

geoTypes :: [PGScalarType]

isGeoType :: PGScalarType -> Bool

data PGTypeKind

Constructors

PGKindBase

PGKindComposite

PGKindDomain

PGKindEnum

PGKindRange

PGKindPseudo

PGKindUnknown Text

data QualifiedPGType

Constructors

QualifiedPGType

Fields

_qptSchema :: SchemaName

_qptName :: PGScalarType

_qptType :: PGTypeKind

isBaseType :: QualifiedPGType -> Bool

typeToTable :: QualifiedPGType -> QualifiedTable

mkFunctionArgScalarType :: QualifiedPGType -> PGScalarType

data PGRawFunctionInfo

Metadata describing SQL functions at the DB level, i.e. below the GraphQL layer.

Constructors

PGRawFunctionInfo

Fields

rfiOid :: OID

rfiHasVariadic :: Bool

rfiFunctionType :: FunctionVolatility

rfiReturnTypeSchema :: SchemaName

rfiReturnTypeName :: PGScalarType

rfiReturnTypeType :: PGTypeKind

rfiReturnsSet :: Bool

rfiInputArgTypes :: [QualifiedPGType]

rfiInputArgNames :: [FunctionArgName]

rfiDefaultArgs :: Int

rfiReturnsTable :: Bool

rfiDescription :: Maybe PGDescription

mkScalarTypeName :: MonadError QErr m => PGScalarType -> m Name

pgTypeOid :: PGScalarType -> Oid

data PGExtraTableMetadata

Constructors

PGExtraTableMetadata

Fields

_petmTableType :: SourceTableType