Hasura.Backends.Postgres.Translate.BoolExp

Description

Postgres Translate BoolExp

Convert IR boolean expressions to Postgres-specific SQL expressions.

Documentation

annBoolExp :: (QErrM m, TableCoreInfoRM b m, BackendMetadata b) => BoolExpRHSParser b m v -> TableName b -> FieldInfoMap (FieldInfo b) -> GBoolExp b ColExp -> m (AnnBoolExp b v)

toSQLBoolExp

Arguments

:: forall pgKind. Backend ('Postgres pgKind)

=> Qual
The name of the tabular value in query scope that the boolean expression applies to

-> AnnBoolExpSQL ('Postgres pgKind)
The boolean expression to translate

-> BoolExp

Translate an IR boolean expression to an SQL boolean expression. References to columns etc are relative to the given rootReference.