Hasura.Backends.Postgres.Translate.Select

Description

Postgres Translate Select

This module is a translation layer between IR and postgres-specific select queries.

There are three main types of selects (as distinguished from the IR):

"simple" selects

aggregate selects

connection selects (used for relay)

Most exports from this module showcase this distinction. The "interesting" parts of the call tree of these functions is similar:

selectQuerySQL -> mkSQLSelect -> processAnnSimpleSelect -> processSelectParams/processAnnFields

selectAggregateQuerySQL -> mkAggregateSelect -> processAnnAggregateSelect -> processSelectParams/processAnnFields

connetionSelectQuerySQL -> mkConnectionSelect -> processConnectionSelection -> processSelectParams

Random thoughts that might help when diving deeper in this module:

Extractors are a pair of an SQL expression and an alias; they get translated like "[SELECT ...] expr as alias"

a SelectSource consists of a prefix, a source, a boolean conditional expression, and info on whether sorting or slicing is done (needed to handle the LIMIT optimisation)

For details on creating the selection tree for relationships via MonadWriter JoinTree, see withWriteJoinTree