Hasura.Backends.Postgres.Translate.Select.Aux1

Description

Stuff gutted from Translate.Select.

Documentation

mkFirstElementExp :: SQLExp -> SQLExp

First element extractor expression from given record set For example:- To get first "id" column from given row set, the function generates the SQL expression AS `(array_agg("id"))[1]`

mkLastElementExp :: SQLExp -> SQLExp

Last element extractor expression from given record set. For example:- To get first "id" column from given row set, the function generates the SQL expression AS `(array_agg("id"))[array_length(array_agg("id"), 1)]`

cursorIdentifier :: Identifier

startCursorIdentifier :: Identifier

endCursorIdentifier :: Identifier

hasPreviousPageIdentifier :: Identifier

hasNextPageIdentifier :: Identifier

pageInfoSelectAliasIdentifier :: Identifier

cursorsSelectAliasIdentifier :: Identifier

encodeBase64 :: SQLExp -> SQLExp

fromTableRowArgs :: Identifier -> FunctionArgsExpTableRow SQLExp -> FunctionArgs

selectFromToFromItem :: Identifier -> SelectFrom ('Postgres pgKind) -> FromItem

functionToIdentifier :: QualifiedFunction -> Identifier

Converts a function name to an Identifier.

If the schema name is public, it will just use its name, otherwise it will prefix it by the schema name.

withJsonBuildObj :: FieldName -> [SQLExp] -> (Alias, SQLExp)

withForceAggregation :: TypeAnn -> SQLExp -> SQLExp

Forces aggregation