Hasura.Backends.Postgres.Translate.Select.Internal.Extractor

Description

Extractors are a pair of an SQL expression and an alias; they get translated like "[SELECT ...] expr as alias"

Documentation

aggregateFieldsToExtractorExps :: forall pgKind m. (Backend ('Postgres pgKind), MonadIO m, MonadError QErr m) => TableIdentifier -> UserInfo -> AggregateFields ('Postgres pgKind) SQLExp -> m [(ColumnAlias, SQLExp)]

Creates node extractors for all of the columns and computed fields used in aggregated fields. The ColumnAliases for all the extractors are namespaced aliases using the 'contextualize*` functions so that none of the extractors names will conflict with one another (for example, if a column name is the same as a field name (eg nodes))

mkAggregateOrderByExtractorAndFields :: forall pgKind. Backend ('Postgres pgKind) => TableIdentifier -> AnnotatedAggregateOrderBy ('Postgres pgKind) SQLExp -> (Extractor, AggregateFields ('Postgres pgKind) SQLExp)

mkRawComputedFieldExpression :: forall pgKind m. (Backend ('Postgres pgKind), MonadIO m, MonadError QErr m) => TableIdentifier -> UserInfo -> ComputedFieldScalarSelect ('Postgres pgKind) SQLExp -> m SQLExp

withJsonAggExtr :: PermissionLimitSubQuery -> Maybe OrderByExp -> ColumnAlias -> SQLExp

asSingleRowExtr :: ColumnAlias -> SQLExp

asJsonAggExtr :: JsonAggSelect -> ColumnAlias -> PermissionLimitSubQuery -> Maybe OrderByExp -> Extractor

withColumnOp :: Maybe ColumnOp -> SQLExp -> SQLExp