Hasura.Backends.Postgres.Translate.Select.Internal.Helpers

Description

Stuff gutted from Translate.Select.

Documentation

mkFirstElementExp :: SQLExp -> SQLExp

First element extractor expression from given record set For example:- To get first "id" column from given row set, the function generates the SQL expression AS `(array_agg("id"))[1]`

mkLastElementExp :: SQLExp -> SQLExp

Last element extractor expression from given record set. For example:- To get first "id" column from given row set, the function generates the SQL expression AS `(array_agg("id"))[array_length(array_agg("id"), 1)]`

cursorIdentifier :: Identifier

startCursorIdentifier :: Identifier

endCursorIdentifier :: Identifier

hasPreviousPageIdentifier :: Identifier

hasNextPageIdentifier :: Identifier

pageInfoSelectAlias :: TableAlias

pageInfoSelectAliasIdentifier :: TableIdentifier

cursorsSelectAlias :: TableAlias

cursorsSelectAliasIdentifier :: TableIdentifier

encodeBase64 :: SQLExp -> SQLExp

fromTableRowArgs :: TableIdentifier -> FunctionArgsExpG (ArgumentExp SQLExp) -> FunctionArgs

selectFromToFromItem :: TableIdentifier -> SelectFrom ('Postgres pgKind) -> FromItem

nativeQueryNameToAlias :: NativeQueryName -> TableAlias

Given a NativeQueryName, what should we call the CTE generated for it?

functionToIdentifier :: QualifiedFunction -> Identifier

Converts a function name to an Identifier.

If the schema name is public, it will just use its name, otherwise it will prefix it by the schema name.

withJsonBuildObj :: FieldName -> [SQLExp] -> (ColumnAlias, SQLExp)

withForceAggregation :: TypeAnn -> SQLExp -> SQLExp

Forces aggregation

selectToSelectWith :: Writer CustomSQLCTEs Select -> SelectWith

unwrap any emitted TopLevelCTEs for custom sql from the Writer and combine them with a Select to create a SelectWith

customSQLToTopLevelCTEs :: CustomSQLCTEs -> [(TableAlias, TopLevelCTE)]

convert map of CustomSQL CTEs into named TopLevelCTEs

customSQLToInnerCTEs :: CustomSQLCTEs -> [(TableAlias, InnerCTE)]

convert map of CustomSQL CTEs into named InnerCTEs

toQuery :: SelectWithG TopLevelCTE -> Query