Hasura.Backends.Postgres.Translate.Select.Streaming

Description

This module defines the top-level translation functions pertaining to streaming selects into Postgres AST.

Streaming subscriptions are subscriptions based on a user-provided cursor column. Unlike live queries, streaming subscriptions can be used to only get the part that has changed in the query's response, although this will be dependent on the user's choice of the cursor column. The streaming starts from the initial value provided by the user.

Documentation

selectStreamQuerySQL :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => AnnSimpleStreamSelect ('Postgres pgKind) -> Query

mkStreamSQLSelect :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => AnnSimpleStreamSelect ('Postgres pgKind) -> Select