Hasura.Backends.Postgres.Translate.Types

Description

Postgres Translate Types

Intermediary / helper types used for translating IR to Postgres SQL.

Documentation

data SourcePrefixes

Constructors

SourcePrefixes

Fields

_pfThis :: Identifier

Current source prefix

_pfBase :: Identifier

Base table source row identifier to generate the table's column identifiers for computed field function input parameters

data SelectSlicing

Select portion of rows generated by the query using limit and offset

Constructors

SelectSlicing

Fields

_ssLimit :: Maybe Int

_ssOffset :: Maybe Int64

data DistinctAndOrderByExpr

Constructors

ASorting

Fields

_sortAtNode :: (OrderByExp, Maybe DistinctExpr)

_sortAtBase :: Maybe (OrderByExp, Maybe DistinctExpr)

data SelectSorting

Sorting with -- Note [Optimizing queries using limit/offset])

Constructors

NoSorting (Maybe DistinctExpr)

Sorting DistinctAndOrderByExpr

data SortingAndSlicing

Constructors

SortingAndSlicing

Fields

_sasSorting :: SelectSorting

_sasSlicing :: SelectSlicing

data SelectSource

Constructors

SelectSource

Fields

_ssPrefix :: Identifier

_ssFrom :: FromItem

_ssWhere :: BoolExp

_ssSortingAndSlicing :: SortingAndSlicing

noSortingAndSlicing :: SortingAndSlicing

noSlicing :: SelectSlicing

orderByForJsonAgg :: SelectSource -> Maybe OrderByExp

data ApplySortingAndSlicing

Constructors

ApplySortingAndSlicing

Fields

_applyAtBase :: (Maybe OrderByExp, SelectSlicing, Maybe DistinctExpr)

_applyAtNode :: (Maybe OrderByExp, SelectSlicing, Maybe DistinctExpr)

applySortingAndSlicing :: SortingAndSlicing -> ApplySortingAndSlicing

data SelectNode

Constructors

SelectNode

Fields

_snExtractors :: HashMap ColumnAlias SQLExp

_snJoinTree :: JoinTree

data ObjectSelectSource

Constructors

ObjectSelectSource

Fields

_ossPrefix :: Identifier

_ossFrom :: FromItem

_ossWhere :: BoolExp

objectSelectSourceToSelectSource :: ObjectSelectSource -> SelectSource

data ObjectRelationSource

Constructors

ObjectRelationSource

Fields

_orsRelationshipName :: RelName

_orsRelationMapping :: HashMap PGCol PGCol

_orsSelectSource :: ObjectSelectSource

data ArrayRelationSource

Constructors

ArrayRelationSource

Fields

_arsAlias :: TableAlias

_arsRelationMapping :: HashMap PGCol PGCol

_arsSelectSource :: SelectSource

data MultiRowSelectNode

Constructors

MultiRowSelectNode

Fields

_mrsnTopExtractors :: [Extractor]

_mrsnSelectNode :: SelectNode

data ComputedFieldTableSetSource

Constructors

ComputedFieldTableSetSource

Fields

_cftssFieldName :: FieldName

_cftssSelectSource :: SelectSource

data ArrayConnectionSource

Constructors

ArrayConnectionSource

Fields

_acsAlias :: TableAlias

_acsRelationMapping :: HashMap PGCol PGCol

_acsSplitFilter :: Maybe BoolExp

_acsSlice :: Maybe ConnectionSlice

_acsSource :: SelectSource

data JoinTree

Constructors

JoinTree

Fields

_jtObjectRelations :: HashMap ObjectRelationSource SelectNode

_jtArrayRelations :: HashMap ArrayRelationSource MultiRowSelectNode

_jtArrayConnections :: HashMap ArrayConnectionSource MultiRowSelectNode

_jtComputedFieldTableSets :: HashMap ComputedFieldTableSetSource MultiRowSelectNode

data PermissionLimitSubQuery

Constructors

PLSQRequired Int
Permission limit

PLSQNotRequired

type SimilarArrayFields = HashMap FieldName [FieldName]