Hasura.Backends.Postgres.Types.Aggregates

Documentation

newtype CountAggregate (pgKind :: PostgresKind) v

This newtype allows us to reuse CountType for the CountType type family We reuse the overall structure, but our column type is a PGCol column name, plus the column redaction expression used by inherited roles. See [SQL generation for inherited roles] for more information about column redaction

Constructors

CountAggregate

Fields

getCountType :: CountType (PGCol, AnnRedactionExp ('Postgres pgKind) v)