Hasura.Backends.Postgres.Types.ComputedField

Description

Postgres types related to computed fields

Documentation

data ComputedFieldDefinition

Constructors

ComputedFieldDefinition

Fields

_cfdFunction :: QualifiedFunction

_cfdTableArgument :: Maybe FunctionArgName

_cfdSessionArgument :: Maybe FunctionArgName

data FunctionTableArgument

The function table argument is either the very first argument or the named argument with an index. The index is 0 if the named argument is the first.

Constructors

FTAFirst

FTANamed

Fields

FunctionArgName

argument name

Int

argument index

data FunctionSessionArgument

The session argument, which passes Hasura session variables to a SQL function as a JSON object.

Constructors

FunctionSessionArgument

Fields

FunctionArgName

The argument name

Int

The ordinal position in the function input parameters

data ComputedFieldImplicitArguments

Constructors

ComputedFieldImplicitArguments

Fields

_cffaTableArgument :: FunctionTableArgument

_cffaSessionArgument :: Maybe FunctionSessionArgument

fromComputedFieldImplicitArguments :: v -> ComputedFieldImplicitArguments -> [ArgumentExp v]

data ComputedFieldReturn

Constructors

CFRScalar PGScalarType

CFRSetofTable QualifiedTable

_CFRSetofTable :: Prism' ComputedFieldReturn QualifiedTable

_CFRScalar :: Prism' ComputedFieldReturn PGScalarType