Hasura.Backends.Postgres.Types.Insert

Description

Postgres Types Insert

This module defines the Insert-related IR types specific to Postgres.

Documentation

newtype BackendInsert pgKind v

The PostgreSQL-specific data of an Insert expression.

This is parameterised over v which enables different phases of IR transformation to maintain the overall structure while enriching/transforming the data at the leaves.

Constructors

BackendInsert

Fields

_biConflictClause :: Maybe (OnConflictClause ('Postgres pgKind) v)