Hasura.Backends.Postgres.Types.Update

Description

Postgres Types Update

This module defines the Update-related IR types specific to Postgres.

Documentation

data UpdateOpExpression v

The various update operators supported by PostgreSQL, i.e. the _set, _inc operators that appear in the schema.

See Update Mutations User docs

Constructors

UpdateSet v

UpdateInc v

UpdateAppend v

UpdatePrepend v

UpdateDeleteKey v

UpdateDeleteElem v

UpdateDeleteAtPath [v]

data PgUpdateVariant (pgKind :: PostgresKind) v

The different variants of updates that the Postgres backend supports.

This is parameterised over v which enables different phases of IR transformation to maintain the overall structure while enriching/transforming the data at the leaves.

Constructors

SingleBatch (UpdateBatch ('Postgres pgKind) UpdateOpExpression v)

MultipleBatches [UpdateBatch ('Postgres pgKind) UpdateOpExpression v]

updateVariantIsEmpty :: forall (b :: PostgresKind) v. PgUpdateVariant b v -> Bool

Are we updating anything?