Hasura.Base.Error

Documentation

data Code

Constructors

AccessDenied

ActionWebhookCode !Text

AlreadyExists

AlreadyTracked

AlreadyUntracked

BadRequest

BigQueryError

Busy

CoercionError

Conflict

ConstraintError

ConstraintViolation

CustomCode !Text
Custom code for extending this sum-type easily

CyclicDependency

DataException

DependencyError

InvalidConfiguration

InvalidHeaders

InvalidJSON

InvalidParams

JWTInvalid

JWTInvalidClaims

JWTRoleClaimMissing

MSSQLError

MethodNotAllowed

NotExists

NotFound

NotSupported

ParseFailed

PermissionDenied

PermissionError

PostgresError

PostgresMaxConnectionsError

RemoteSchemaConflicts

RemoteSchemaError

StartFailed
Websockets

Unexpected

UnexpectedPayload

ValidationFailed

data QErr

Constructors

QErr

Fields

qePath :: !JSONPath

qeStatus :: !Status

qeError :: !Text

qeCode :: !Code

qeInternal :: !(Maybe QErrExtra)

data QErrExtra

Extra context for a QErr, which can either be information from an internal error (e.g. from Postgres, or from a network operation timing out), or context provided when an external service or operation fails, for instance, a webhook error response may provide additional context in the extensions key.

Constructors

ExtraExtensions !Value

ExtraInternal !Value

noInternalQErrEnc :: QErr -> Value

encodeGQLErr :: Bool -> QErr -> Value

encodeQErr :: Bool -> QErr -> Value

encodeJSONPath :: JSONPath -> String

err400 :: Code -> Text -> QErr

err404 :: Code -> Text -> QErr

err405 :: Code -> Text -> QErr

err401 :: Code -> Text -> QErr

err409 :: Code -> Text -> QErr

err429 :: Code -> Text -> QErr

err500 :: Code -> Text -> QErr

type QErrM m = MonadError QErr m

throw400 :: QErrM m => Code -> Text -> m a

throw404 :: QErrM m => Text -> m a

throw405 :: QErrM m => Text -> m a

MethodNotAllowed

throw401 :: QErrM m => Text -> m a

AccessDenied

throw409 :: QErrM m => Text -> m a

Conflict

throw429 :: QErrM m => Code -> Text -> m a

throw500 :: QErrM m => Text -> m a

internalError :: Text -> QErr

throw500WithDetail :: QErrM m => Text -> Value -> m a

modifyQErr :: QErrM m => (QErr -> QErr) -> m a -> m a

modifyErr :: QErrM m => (Text -> Text) -> m a -> m a

modifyErrA :: ArrowError QErr arr => arr (e, s) a -> arr (e, (Text -> Text, s)) a

liftTxtMod :: (Text -> Text) -> QErr -> QErr

modifyErrAndSet500 :: QErrM m => (Text -> Text) -> m a -> m a

liftTxtMod500 :: (Text -> Text) -> QErr -> QErr

withPathE :: ArrowError QErr arr => arr (e, s) a -> arr (e, (JSONPathElement, s)) a

withPathKA :: ArrowError QErr arr => arr (e, s) a -> arr (e, (Text, s)) a

withPathK :: QErrM m => Text -> m a -> m a

withPathIA :: ArrowError QErr arr => arr (e, s) a -> arr (e, (Int, s)) a

withPathI :: QErrM m => Int -> m a -> m a

indexedFoldlA' :: (ArrowChoice arr, ArrowError QErr arr, Foldable t) => arr (e, (b, (a, s))) b -> arr (e, (b, (t a, s))) b

indexedFoldM :: (QErrM m, Foldable t) => (b -> a -> m b) -> b -> t a -> m b

indexedTraverseA_ :: (ArrowChoice arr, ArrowError QErr arr, Foldable t) => arr (e, (a, s)) b -> arr (e, (t a, s)) ()

indexedMapM_ :: (QErrM m, Foldable t) => (a -> m b) -> t a -> m ()

indexedForM_ :: (QErrM m, Foldable t) => t a -> (a -> m b) -> m ()

indexedTraverseA :: (ArrowChoice arr, ArrowError QErr arr) => arr (e, (a, s)) b -> arr (e, ([a], s)) [b]

indexedMapM :: QErrM m => (a -> m b) -> [a] -> m [b]

indexedForM :: QErrM m => [a] -> (a -> m b) -> m [b]

liftIResult :: QErrM m => IResult a -> m a

iResultToMaybe :: IResult a -> Maybe a

formatMsg :: String -> String

runAesonParser :: QErrM m => (v -> Parser a) -> v -> m a

decodeValue :: (FromJSON a, QErrM m) => Value -> m a