Hasura.Base.ErrorMessage

Description

Error messages

This module defines a type for user facing error messages.

To construct a value of this type, use toErrorMessage or the IsString interface, the type class ToErrorValue defined in the Hasura.Base.ToErrorValue module, or use the utility functions defined in the Hasura.Base.ErrorValue module.

ErrorMessages can also be composed using the Semigroup interface.

To fail with an error message in monadic code, use the failWithMessage function.

Documentation

newtype ErrorMessage

ErrorMessage wraps a Text value such that it's easy to build up, but difficult to break apart or extract the underlying text value.

Constructors

ErrorMessage

Fields

fromErrorMessage :: Text

A temporary extractor which will go away once ErrorMessage is pervasive.

toErrorMessage :: Text -> ErrorMessage

A smart constructor for ErrorMessage so that it cannot be deconstructed.

failWithMessage :: MonadFail m => ErrorMessage -> m ()

Fails with the given message.

type ErrorMessageOr = Either ErrorMessage

A simple alias to prevent duplication.