Hasura.Cache.Bounded

Description

graphql-engine used to have a query plan cache, which cached the execution plan for a given graphql query (sans JSON variable values). After the PDV refactor (see hasura/graphql-engine#4111), this query plan cache was not needed anymore. For backwards compatibility reasons, we still need to parse the configuration options from the CLI, although the CLI option gets ignored.

Eventually, we can decide to stop parsing the CLI option --query-plan-cache-size, at which point this module can be removed.

Documentation

newtype CacheSize

Constructors

CacheSize

Fields

unCacheSize :: Word16

parseCacheSize :: String -> Either String CacheSize