Hasura.CustomReturnType.API

Description

Define and handle v1/metadata API operations to track, untrack, and get custom return types.

Documentation

data TrackCustomReturnType (b :: BackendType)

Default implementation of the track_custom_return_type request payload.

Constructors

TrackCustomReturnType

Fields

tctSource :: SourceName

tctName :: CustomReturnTypeName

tctDescription :: Maybe Text

tctFields :: InsOrdHashMap (Column b) (NullableScalarType b)

data GetCustomReturnType (b :: BackendType)

API payload for the get_custom_return_type endpoint.

Constructors

GetCustomReturnType

Fields

gcrtSource :: SourceName

runGetCustomReturnType :: forall b m. (BackendMetadata b, MetadataM m, HasFeatureFlagChecker m, MonadError QErr m) => GetCustomReturnType b -> m EncJSON

Handler for the get_custom_return_type endpoint.

getCustomTypes :: forall b. Backend b => SourceName -> Traversal' Metadata (CustomReturnTypes b)

runTrackCustomReturnType :: forall b m. (BackendMetadata b, CacheRWM m, MetadataM m, MonadError QErr m, HasFeatureFlagChecker m) => TrackCustomReturnType b -> m EncJSON

Handler for the track_custom_return_type endpoint. The type 'TrackCustomReturnType b' (appearing here in wrapped as 'BackendTrackCustomReturnType b' for AnyBackend compatibility) is defined in 'class CustomReturnTypeMetadata'.

data UntrackCustomReturnType (b :: BackendType)

API payload for the untrack_custom_return_type endpoint.

Constructors

UntrackCustomReturnType

Fields

utctSource :: SourceName

utctName :: CustomReturnTypeName

runUntrackCustomReturnType :: forall b m. (BackendMetadata b, MonadError QErr m, CacheRWM m, MetadataM m) => UntrackCustomReturnType b -> m EncJSON

Handler for the untrack_custom_return_type endpoint.

data CreateCustomReturnTypePermission a (b :: BackendType)

A permission for logical models is tied to a specific root field name and source. This wrapper adds both of those things to the JSON object that describes the permission.

Constructors

CreateCustomReturnTypePermission

Fields

ccrtpSource :: SourceName

ccrtpName :: CustomReturnTypeName

ccrtpInfo :: PermDef b a

runCreateSelectCustomReturnTypePermission :: forall b m. (Backend b, CacheRWM m, MetadataM m, MonadError QErr m, HasFeatureFlagChecker m) => CreateCustomReturnTypePermission SelPerm b -> m EncJSON

data DropCustomReturnTypePermission (b :: BackendType)

To drop a permission, we need to know the source and root field name of the logical model, as well as the role whose permission we want to drop.

Constructors

DropCustomReturnTypePermission

Fields

dcrtpSource :: SourceName

dcrtpName :: CustomReturnTypeName

dcrtpRole :: RoleName

runDropSelectCustomReturnTypePermission :: forall b m. (Backend b, CacheRWM m, MetadataM m, MonadError QErr m, HasFeatureFlagChecker m) => DropCustomReturnTypePermission b -> m EncJSON

dropCustomReturnTypeInMetadata :: forall b. BackendMetadata b => SourceName -> CustomReturnTypeName -> MetadataModifier

TODO: should this cascade and also delete associated permissions?