Hasura.CustomReturnType.Metadata

Documentation

data CustomReturnTypeMetadata (b :: BackendType)

Description of a custom return type for use in metadata (before schema cache)

Constructors

CustomReturnTypeMetadata

Fields

_crtmName :: CustomReturnTypeName

_crtmFields :: InsOrdHashMap (Column b) (NullableScalarType b)

_crtmDescription :: Maybe Text

_crtmSelectPermissions :: InsOrdHashMap RoleName (SelPermDef b)

crtmSelectPermissions :: forall b. Lens' (CustomReturnTypeMetadata b) (InsOrdHashMap RoleName (SelPermDef b))

crtmName :: forall b. Lens' (CustomReturnTypeMetadata b) CustomReturnTypeName

crtmFields :: forall b. Lens' (CustomReturnTypeMetadata b) (InsOrdHashMap (Column b) (NullableScalarType b))

crtmDescription :: forall b. Lens' (CustomReturnTypeMetadata b) (Maybe Text)

data WithCustomReturnType a

A wrapper to tie something to a particular logical model. Specifically, it assumes the underlying _wlmInfo is represented as an object, and adds two keys to that object: source and root_field_name.

Constructors

WithCustomReturnType

Fields

_wcrtSource :: SourceName

_wcrtName :: CustomReturnTypeName

_wcrtInfo :: a