Hasura.Eventing.Common

Documentation

data LockedEventsCtx

Constructors

LockedEventsCtx

Fields

leCronEvents :: TVar (Set CronEventId)

leOneOffEvents :: TVar (Set OneOffScheduledEventId)

leEvents :: TVar (HashMap SourceName (Set EventId))

leActionEvents :: TVar (Set LockedActionEventId)

saveLockedEvents :: MonadIO m => [EventId] -> TVar (Set EventId) -> m ()

After the events are fetched from the DB, we store the locked events in a hash set(order doesn't matter and look ups are faster) in the event engine context

removeEventFromLockedEvents :: MonadIO m => EventId -> TVar (Set EventId) -> m ()

Remove an event from the LockedEventsCtx after it has been processed

generateScheduleTimes :: UTCTime -> Int -> CronSchedule -> [UTCTime]

Generates next n events starting from according to CronSchedule

cleanupSchedulesToBeGenerated :: Int

number of cleanup schedules to be generated in one iteration

deleteEventTriggerLogsInBatchesWith :: (MonadIO m, MonadError QErr m) => IO (Maybe (TriggerLogCleanupConfig, EventTriggerCleanupStatus)) -> TriggerLogCleanupConfig -> (TriggerLogCleanupConfig -> IO (Either QErr DeletedEventLogStats)) -> m DeletedEventLogStats

createStatsLogger :: forall m stats. (MonadIO m, ToEngineLog stats Hasura, Monoid stats) => Logger Hasura -> m (Trigger stats stats)

A logger useful for accumulating stats, evaluated in forever running loops, over a period of time and log them only once. Use @logStats to record statistics for logging.

closeStatsLogger :: MonadIO m => EngineLogType Hasura -> Logger Hasura -> Trigger stats stats -> m ()

Safely close the statistics logger. When occurred, exception is logged.

logStats :: MonadIO m => Trigger stats stats -> stats -> m ()

This won't log the given stats immediately. The stats are accumulated over the specific timeframe and logged only once. See @createStatsLogger for more details.

Orphan instances

ToEngineLog (OpException, EngineLogType Hasura) Hasura

Instance details

Methods

toEngineLog :: (OpException, EngineLogType Hasura) -> (LogLevel, EngineLogType Hasura, Value)