Hasura.Eventing.HTTP

Description

Hasura.Eventing.HTTP

This module is an utility module providing HTTP utilities for Hasura.Eventing.EventTriggers and Hasura.Eventing.ScheduledTriggers.

The event triggers and scheduled triggers share the event delivery mechanism using the tryWebhook function defined in this module.

Documentation

data ExtraLogContext

Constructors

ExtraLogContext

Fields

elEventId :: !EventId

elEventName :: !(Maybe TriggerName)

data HTTPResp (a :: TriggerTypes)

Constructors

HTTPResp

Fields

hrsStatus :: !Int

hrsHeaders :: ![HeaderConf]

hrsBody :: !SerializableBlob

hrsSize :: !Int64

data HTTPErr (a :: TriggerTypes)

Constructors

HClient !HttpException

HStatus !(HTTPResp a)

HOther !String

httpExceptionErrorEncoding :: HttpException -> ByteString

data RequestDetails

Constructors

RequestDetails

Fields

_rdOriginalRequest :: Request

_rdOriginalSize :: Int64

_rdTransformedRequest :: Maybe Request

_rdTransformedSize :: Maybe Int64

_rdReqTransformCtx :: Maybe RequestContext

_rdSessionVars :: Maybe SessionVariables

extractRequest :: RequestDetails -> Request

isNetworkError :: HTTPErr a -> Bool

isNetworkErrorHC :: HttpException -> Bool

logHTTPForET :: (MonadReader r m, Has (Logger Hasura) r, MonadIO m) => Either (HTTPErr 'EventType) (HTTPResp 'EventType) -> ExtraLogContext -> RequestDetails -> Text -> [HeaderConf] -> TriggersErrorLogLevelStatus -> m ()

logHTTPForST :: (MonadReader r m, Has (Logger Hasura) r, MonadIO m) => Either (HTTPErr 'ScheduledType) (HTTPResp 'ScheduledType) -> ExtraLogContext -> RequestDetails -> Text -> [HeaderConf] -> TriggersErrorLogLevelStatus -> m ()

runHTTP :: MonadIO m => Manager -> Request -> m (Either (HTTPErr a) (HTTPResp a))

data TransformableRequestError a

Constructors

HTTPError Value (HTTPErr a)

TransformationError Value TransformErrorBundle

mkRequest

Arguments

:: MonadError (TransformableRequestError a) m

=> [Header]

-> ResponseTimeout

-> ByteString
the request body. It is passed as a Bytestring because we need to log the request size. As the logging happens outside the function, we pass it the final request body, instead of Value

-> Maybe RequestTransform

-> ResolvedWebhook

-> m RequestDetails

invokeRequest :: (MonadReader r m, MonadError (TransformableRequestError a) m, Has Manager r, Has (Logger Hasura) r, MonadIO m, MonadTrace m) => RequestDetails -> Maybe ResponseTransform -> Maybe SessionVariables -> (Either (HTTPErr a) (HTTPResp a) -> RequestDetails -> m ()) -> m (HTTPResp a)

mkResp :: Int -> SerializableBlob -> [HeaderConf] -> Response a

mkClientErr :: SerializableBlob -> Response a

mkWebhookReq :: Value -> [HeaderConf] -> InvocationVersion -> WebhookRequest

mkInvocationResp :: Maybe Int -> SerializableBlob -> [HeaderConf] -> Response a

isClientError :: Int -> Bool

prepareHeaders :: [EventHeaderInfo] -> ([Header], [HeaderConf])

Encodes given request headers along with our defaultHeaders and returns them along with the re-decoded set of headers (for logging purposes).

getRetryAfterHeaderFromHTTPErr :: HTTPErr a -> Maybe Text

getRetryAfterHeaderFromResp :: HTTPResp a -> Maybe Text

parseRetryHeaderValue :: Text -> Maybe Int