Hasura.Function.Common

Documentation

getFunctionArgsGQLName

Arguments

:: Name
The GQL version of the DB name of the function

-> FunctionConfig b

-> (Name -> Name)
Custom function for setting naming case

-> Name

Apply function name customization to function arguments, as detailed in 'rfcs/function-root-field-customisation.md'. We want the different variations of a function (i.e. basic, aggregate) to share the same type name for their arguments.

getFunctionGQLName

Arguments

:: Name

-> FunctionConfig b

-> (Name -> Name)
Custom function for setting naming case

-> Name

Apply function name customization to the basic function variation, as detailed in 'rfcs/function-root-field-customisation.md'.

getFunctionAggregateGQLName

Arguments

:: Name

-> FunctionConfig b

-> (Name -> Name)
Custom function for setting naming case

-> Name

Apply function name customization to the aggregate function variation, as detailed in 'rfcs/function-root-field-customisation.md'.

getInputArgs :: FunctionInfo b -> Seq (FunctionArgument b)