Hasura.GraphQL.Analyse

Description

Tools to analyze the structure of a GraphQL request.

Documentation

data Structure

Overall structure of a given query. We extract the tree of fields in the output, and the graph of input variables.

Constructors

Structure

Fields

_stSelection :: HashMap Name FieldInfo

_stVariables :: HashMap Name VariableInfo

data FieldInfo

Information about the type of an output field; whether the base type is an object or a scalar, we store the correspoding GType to keep track of the modifiers applied to it (list or non-nullability).

Constructors

FieldObjectInfo GType ObjectInfo

FieldScalarInfo GType ScalarInfo

FieldEnumInfo GType EnumInfo

data ScalarInfo

Constructors

ScalarInfo

Fields

_siTypeDefinition :: ScalarTypeDefinition

data EnumInfo

Constructors

EnumInfo

Fields

_eiTypeDefinition :: EnumTypeDefinition

data ObjectInfo

Constructors

ObjectInfo

Fields

_oiTypeDefinition :: ObjectTypeDefinition InputValueDefinition

_oiSelection :: HashMap Name FieldInfo

data VariableInfo

Information about a single variable of the query.

Constructors

VariableInfo

Fields

_viType :: GType

_viTypeInfo :: InputFieldInfo

_viDefaultValue :: Maybe (Value Void)

data InputFieldInfo

Information about the type of an input field; whether the base type is an object or a scalar, we store the correspoding GType to keep track of the modifiers applied to it (list or non-nullability).

Constructors

InputFieldScalarInfo ScalarInfo

InputFieldEnumInfo EnumInfo

InputFieldObjectInfo InputObjectInfo

data InputObjectInfo

Constructors

InputObjectInfo

Fields

_ioiTypeDefinition :: InputObjectTypeDefinition InputValueDefinition

_ioiFields :: ~(HashMap Name (GType, InputFieldInfo))

lazy for knot-tying, as we build a graph

diagnoseGraphQLQuery :: SchemaIntrospection -> TypedOperationDefinition NoFragments Name -> Maybe [Text]

Given the schema's definition, and a query, validate that the query is consistent. We do this by running the analysis, but discarding the result: we do not care about the structure, only about the validity of the query.

Returns Nothing if the query is valid, or a list of messages otherwise.

analyzeGraphQLQuery :: SchemaIntrospection -> TypedOperationDefinition NoFragments Name -> (Maybe Structure, [Text])

Given the schema's definition, and a query, run the analysis.

We process all possible fields, and return a partially filled structure if necessary. Given the following query:

query {
  foo {
    bar
  }
  does_not_exist {
    ghsdflgh
  }
}

We would return a structure containing:

foo: {
  bar: {
  }
}

AND an error about "does_not_exist" not existing.

In some cases, however, we might not be able to produce a structure at all, in which case we return Nothing. This either indicates that something was fundamentally wrong with the structure of the query (such as not finding an object at the top level), or that a recoverable error was not caught properly (see withCatchAndRecord).

analyzeObjectSelectionSet :: ObjectTypeDefinition InputValueDefinition -> SelectionSet NoFragments Name -> Analysis (HashMap Name FieldInfo)

Analyze the fields of an object selection set against its definition, and emit the corresponding Selection. We ignore the fields that fail, and we continue accumulating the others.

analyzeField :: GType -> TypeDefinition [Name] InputValueDefinition -> Field NoFragments Name -> Analysis (Maybe FieldInfo)

Analyze a given field, and attempt to build a corresponding FieldInfo.

analyzeVariables :: [VariableDefinition] -> Analysis (HashMap Name VariableInfo)

Analyzes the variables in the given query. This builds the graph of input types associated with the variable. This process is, like any GraphQL schema operation, inherently self-recursive, and we use CircularT (a lesser SchemaT) to tie the knot.

analyzeInputField :: Name -> TypeDefinition [Name] InputValueDefinition -> CircularT Name InputFieldInfo Analysis InputFieldInfo

Builds an InputFieldInfo for a given typename.

This function is "memoized" using withCircular to prevent processing the same type more than once in case the input types are self-recursive.

newtype Analysis a

The monad in which we run our analysis.

Has three capabilities: - reader carries the current path, and the full schema for lookups - writer logs all errors we have caught - except allows for short-circuiting errors

Constructors

Analysis (ExceptT AnalysisError (ReaderT (Path, SchemaIntrospection) (Writer [AnalysisError])) a)

runAnalysis :: SchemaIntrospection -> Analysis a -> (Maybe a, [Text])

lookupType :: MonadReader (Path, SchemaIntrospection) m => Name -> m (Maybe (TypeDefinition [Name] InputValueDefinition))

Look up a type in the schema.

withField :: MonadReader (Path, SchemaIntrospection) m => Name -> m a -> m a

Add the current field to the error path.

throwDiagnosis :: (MonadReader (Path, SchemaIntrospection) m, MonadError AnalysisError m) => Diagnosis -> m a

Throws an AnalysisError by combining the given diagnosis with the current path. This interrupts the computation in the given branch, and must be caught for the analysis to resume.

withCatchAndRecord :: (MonadReader (Path, SchemaIntrospection) m, MonadWriter [AnalysisError] m, MonadError AnalysisError m) => m a -> m (Maybe a)

Runs the given computation. if it fails, cacthes the error, records it in the monad, and return Nothing. This allows for a clean recovery.

data AnalysisError

Constructors

AnalysisError

Fields

_aePath :: Path

_aeDiagnosis :: Diagnosis

type Path = Seq Text

data Diagnosis

Constructors

RootTypeNotAnObject

TypeNotFound Name

EnumSelectionSet Name

ScalarSelectionSet Name

InputObjectInOutput Name

UnionInInput Name

ObjectInInput Name

InterfaceInInput Name

ObjectFieldNotFound Name Name

ObjectMissingSelectionSet Name

MismatchedFields Name GType GType

render :: AnalysisError -> Text

queryRootName :: Name

mutationRootName :: Name

subscriptionRootName :: Name

typenameField :: FieldDefinition InputValueDefinition

schemaField :: FieldDefinition InputValueDefinition

typeField :: FieldDefinition InputValueDefinition

mkReservedField :: Name -> Name -> FieldDefinition InputValueDefinition