Hasura.GraphQL.ApolloFederation

Description

Tools for generating fields for Apollo federation

Documentation

data ApolloFederationParserFunction n

Internal parser function for entities field

Constructors

ApolloFederationParserFunction

Fields

aafuGetRootField :: ApolloFederationAnyType -> n (QueryRootField UnpreparedValue)

data ApolloFederationAnyType

Haskell representation of _Any scalar

Constructors

ApolloFederationAnyType

Fields

afTypename :: Name

afPKValues :: Object

anyParser :: Parser origin 'Both Parse ApolloFederationAnyType

Parser for _Any scalar

convertToApolloFedParserFunc :: (Monad n, MonadParse n, Backend b) => SourceInfo b -> TableInfo b -> TablePermG b (UnpreparedValue b) -> StringifyNumbers -> Maybe NamingCase -> NESeq (ColumnInfo b) -> Parser 'Output n (AnnotatedFields b) -> Parser 'Output n (ApolloFederationParserFunction n)

modifyApolloFedParserFunc :: (MonadParse n, Backend b) => SourceInfo b -> TableInfo b -> TablePermG b (UnpreparedValue b) -> StringifyNumbers -> Maybe NamingCase -> NESeq (ColumnInfo b) -> AnnotatedFields b -> ApolloFederationParserFunction n

mkServiceField :: FieldParser Parse (SchemaIntrospection -> QueryRootField UnpreparedValue)

Creates _service FieldParser using the schema introspection. This will allow us to process the following query:

query {
  _service {
    sdl
  }
}

apolloRootFields :: HashSet ExperimentalFeature -> [(Name, Parser 'Output Parse (ApolloFederationParserFunction Parse))] -> [FieldParser Parse (SchemaIntrospection -> QueryRootField UnpreparedValue)]

generateSDL :: SchemaIntrospection -> Text

Generate sdl from the schema introspection

filterAndWrapTypeSystemDefinition :: TypeDefinition [Name] InputValueDefinition -> Maybe TypeSystemDefinition

Filter out schema components from sdl which are not required by apollo federation and wraps it in TypeSystemDefinition

mkEntityUnionFieldParser :: [(Name, Parser 'Output Parse (ApolloFederationParserFunction Parse))] -> FieldParser Parse (QueryRootField UnpreparedValue)

Creates _entities FieldParser using Parsers for Entity union, schema introspection and a list of all query FieldParser. This will allow us to process the following query:

query ($representations: [_Any!]!) {
  _entities(representations: $representations) {
    ... on SomeType {
      foo
      bar
    }
  }
}

concatQueryRootFields :: [QueryRootField UnpreparedValue] -> QueryRootField UnpreparedValue

concatenates multiple fields