Hasura.GraphQL.Execute.LiveQuery.State

Description

Top-level management of live query poller threads. The implementation of the polling itself is in Hasura.GraphQL.Execute.LiveQuery.Poll. See Hasura.GraphQL.Execute.LiveQuery for high-level details.

Documentation

data LiveQueriesState

The top-level datatype that holds the state for all active live queries.

NOTE!: This must be kept consistent with a websocket connection's OperationMap, in onClose and onStart.

Constructors

LiveQueriesState

Fields

_lqsOptions :: !LiveQueriesOptions

_lqsLiveQueryMap :: !PollerMap

_lqsPostPollHook :: !LiveQueryPostPollHook

A hook function which is run after each fetch cycle

_lqsAsyncActions :: !AsyncActionSubscriptionState

initLiveQueriesState :: LiveQueriesOptions -> LiveQueryPostPollHook -> IO LiveQueriesState

dumpLiveQueriesState :: Bool -> LiveQueriesState -> IO Value

data LiveQueryId

Constructors

LiveQueryId

Fields

_lqiPoller :: !PollerKey

_lqiCohort :: !CohortKey

_lqiSubscriber :: !SubscriberId

addLiveQuery

Arguments

:: forall b. BackendTransport b

=> Logger Hasura

-> ServerMetrics

-> SubscriberMetadata

-> LiveQueriesState

-> SourceName

-> ParameterizedQueryHash

-> Maybe OperationName
operation name of the query

-> RequestId

-> LiveQueryPlan b (MultiplexedQuery b)

-> OnChange
the action to be executed when result changes

-> IO LiveQueryId

removeLiveQuery :: Logger Hasura -> ServerMetrics -> LiveQueriesState -> LiveQueryId -> IO ()

data LiveAsyncActionQueryOnSource

An async action query whose relationships are refered to table in a source. We need to generate an SQL statement with the action response and execute it in the source database so as to fetch response joined with relationship rows. For more details see Note [Resolving async action query]

Constructors

LiveAsyncActionQueryOnSource

Fields

_laaqpCurrentLqId :: !LiveQueryId

_laaqpPrevActionLogMap :: !ActionLogResponseMap

_laaqpRestartLq :: !(LiveQueryId -> ActionLogResponseMap -> IO (Maybe LiveQueryId))

An IO action to restart the live query poller with updated action log responses fetched from metadata storage Restarting a live query re-generates the SQL statement with new action log responses to send latest action response to the client.

data LiveAsyncActionQueryWithNoRelationships

Constructors

LiveAsyncActionQueryWithNoRelationships

Fields

_laaqwnrSendResponse :: !(ActionLogResponseMap -> IO ())

An IO action to send response to the websocket client

_laaqwnrSendCompleted :: !(IO ())

An IO action to send "completed" message to the websocket client

data LiveAsyncActionQuery

Constructors

LAAQNoRelationships !LiveAsyncActionQueryWithNoRelationships

LAAQOnSourceDB !LiveAsyncActionQueryOnSource

data AsyncActionQueryLive

Constructors

AsyncActionQueryLive

Fields

_aaqlActionIds :: !(NonEmpty ActionId)

_aaqlOnException :: !(QErr -> IO ())

An IO action to send error message (in case of any exception) to the websocket client

_aaqlLiveExecution :: !LiveAsyncActionQuery

type AsyncActionSubscriptionState = TMap OperationId AsyncActionQueryLive

A share-able state map which stores an async action live query with it's subscription operation id

addAsyncActionLiveQuery :: AsyncActionSubscriptionState -> OperationId -> NonEmpty ActionId -> (QErr -> IO ()) -> LiveAsyncActionQuery -> IO ()

removeAsyncActionLiveQuery :: AsyncActionSubscriptionState -> OperationId -> IO ()