Hasura.GraphQL.Execute.LiveQuery.TMap

Documentation

newtype TMap k v

A coarse-grained transactional map implemented by simply wrapping a HashMap in a TVar. Compared to StmContainers.Map, this provides much faster iteration over the elements at the cost of significantly increased contention on writes.

Constructors

TMap

Fields

unTMap :: TVar (HashMap k v)

new :: STM (TMap k v)

reset :: TMap k v -> STM ()

null :: TMap k v -> STM Bool

lookup :: (Eq k, Hashable k) => k -> TMap k v -> STM (Maybe v)

insert :: (Eq k, Hashable k) => v -> k -> TMap k v -> STM ()

delete :: (Eq k, Hashable k) => k -> TMap k v -> STM ()

toList :: TMap k v -> STM [(k, v)]