Hasura.GraphQL.Execute.RemoteJoin.Collect

Documentation

newtype Collector a

A writer monad used to collect together all remote joins appearing in some data structure.

In the functions below, the withField function is used to track the context of the path from the root of the current selection set.

It is important that we work bottom-up, and do not collect duplicate field names at any level, because the Semigroup instance for RemoteJoins does not allow for these duplicates.

Constructors

Collector

Fields

runCollector :: (a, Maybe RemoteJoins)

collect :: NEHashMap FieldName RemoteJoin -> Collector ()

Collect some remote joins appearing at the given field names in the current context.

withField :: FieldName -> Collector a -> Collector a

Keep track of the given field name in the current path from the root of the selection set.

getRemoteJoins :: Backend b => QueryDB b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (QueryDB b Void (UnpreparedValue b), Maybe RemoteJoins)

Collects remote joins from the AST and also adds the necessary join fields

getRemoteJoinsSelect :: Backend b => AnnSimpleSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (AnnSimpleSelectG b Void (UnpreparedValue b), Maybe RemoteJoins)

Traverse through AnnSimpleSel and collect remote join fields (if any).

getRemoteJoinsAggregateSelect :: Backend b => AnnAggregateSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (AnnAggregateSelectG b Void (UnpreparedValue b), Maybe RemoteJoins)

Traverse through @AnnAggregateSelect and collect remote join fields (if any).

getRemoteJoinsConnectionSelect :: Backend b => ConnectionSelect b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (ConnectionSelect b Void (UnpreparedValue b), Maybe RemoteJoins)

Traverse through @ConnectionSelect and collect remote join fields (if any).

getRemoteJoinsMutationOutput :: Backend b => MutationOutputG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (MutationOutputG b Void (UnpreparedValue b), Maybe RemoteJoins)

Traverse through MutationOutput and collect remote join fields (if any)

getRemoteJoinsActionFields :: Backend b => ActionFieldsG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (ActionFieldsG b Void (UnpreparedValue b), Maybe RemoteJoins)

getRemoteJoinsMutationDB :: Backend b => MutationDB b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (MutationDB b Void (UnpreparedValue b), Maybe RemoteJoins)

getRemoteJoinsSyncAction :: Backend b => AnnActionExecution b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (AnnActionExecution b Void (UnpreparedValue b), Maybe RemoteJoins)

getRemoteJoinsActionQuery :: Backend b => ActionQuery b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (ActionQuery b Void (UnpreparedValue b), Maybe RemoteJoins)

getRemoteJoinsActionMutation :: Backend b => ActionMutation b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (ActionMutation b Void (UnpreparedValue b), Maybe RemoteJoins)

transformSelect :: Backend b => AnnSimpleSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> Collector (AnnSimpleSelectG b Void (UnpreparedValue b))

transformAggregateSelect :: Backend b => AnnAggregateSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> Collector (AnnAggregateSelectG b Void (UnpreparedValue b))

transformConnectionSelect :: forall b. Backend b => ConnectionSelect b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> Collector (ConnectionSelect b Void (UnpreparedValue b))

transformObjectSelect :: Backend b => AnnObjectSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> Collector (AnnObjectSelectG b Void (UnpreparedValue b))

createRemoteJoin :: HashMap FieldName JoinColumnAlias -> RemoteRelationshipField UnpreparedValue -> RemoteJoin

Converts a remote relationship field into a RemoteJoin that the execution engine understands

transformAnnFields :: forall src. Backend src => AnnFieldsG src (RemoteRelationshipField UnpreparedValue) (UnpreparedValue src) -> Collector (AnnFieldsG src Void (UnpreparedValue src))

transformAnnRelation :: (a -> Collector b) -> AnnRelationSelectG src a -> Collector (AnnRelationSelectG src b)

Annotate an element a remote source join from _rssJoinMapping so that a remote join can be constructed.

transformActionFields :: forall src. Backend src => ActionFieldsG src (RemoteRelationshipField UnpreparedValue) (UnpreparedValue src) -> Collector (ActionFieldsG src Void (UnpreparedValue src))

getJoinColumnAlias :: (Eq field, Hashable field) => FieldName -> field -> HashMap field FieldName -> [FieldName] -> JoinColumnAlias

getRemoteJoinsSourceRelation :: Backend b => SourceRelationshipSelection b (RemoteRelationshipField UnpreparedValue) UnpreparedValue -> (SourceRelationshipSelection b Void UnpreparedValue, Maybe RemoteJoins)