Hasura.GraphQL.Execute.RemoteJoin.RemoteSchema

Documentation

newtype ResponsePath

Used to extract the value from a remote schema response.

For example: if a remote relationship is defined to retrieve data from some deeply nested field, this is the path towards that deeply nested field.

Constructors

ResponsePath (NonEmpty Name)

type Alias = Name

The name that we generate when performing a remote join, which shall always be the first field in a ResponsePath.

parseGraphQLName :: MonadError QErr m => Text -> m Name

data RemoteSchemaCall

Intermediate type containing all of the information required to perform a remote schema call.

See buildRemoteSchemaCall for details.

Constructors

RemoteSchemaCall

Fields

_rscInfo :: !RemoteSchemaInfo

_rscCustomizer :: !ResultCustomizer

_rscGQLRequest :: !GQLReqOutgoing

_rscResponsePaths :: !(IntMap ResponsePath)

buildRemoteSchemaCall :: MonadError QErr m => UserInfo -> RemoteSchemaJoin -> IntMap JoinArgument -> m (Maybe RemoteSchemaCall)

Constructs an outgoing response from the remote relationships definition (i.e. RemoteSchemaJoin) and the arguments collected from the database's response.

NOTE: We need to pass along some additional information with the raw outgoing GraphQL request, hence the RemoteSchemaCall type.

buildJoinIndex :: (Monad m, MonadError QErr m) => Object -> IntMap ResponsePath -> m JoinIndex

Construct a JoinIndex from the remote source's Value response.

If the response does not have value at any of the provided ResponsePaths, throw a generic QErr.

NOTE(jkachmar): If we switch to an Applicative validator, we can collect more than one missing ResponsePaths (rather than short-circuiting on the first missing value).

getRemoteSchemaResponse :: (MonadError QErr m, MonadIO m, MonadTrace m) => Environment -> Manager -> [Header] -> UserInfo -> RemoteSchemaCall -> m Object

extractAtPath :: forall m. MonadError QErr m => Value -> ResponsePath -> m Value

Attempt to extract a deeply nested value from a remote source's Value response, according to the JSON path provided by ResponsePath.

ordJSONValueToGValue :: MonadError QErr n => Value -> n (Value Void)

convertFieldWithVariablesToName :: Field NoFragments Variable -> Field NoFragments Name

inputValueToJSON :: InputValue Void -> Value

collectVariablesFromValue :: Value Variable -> HashMap VariableDefinition Value

TODO: Documentation.

collectVariablesFromField :: Field NoFragments Variable -> HashMap VariableDefinition Value

TODO: Documentation.

fieldsToRequest :: NonEmpty (Field NoFragments Variable) -> GQLReqOutgoing

TODO: Documentation.

Extension of the documentation required for collectVariablesFromField and collectVariablesFromValue.

fieldCallsToField

Arguments

:: forall m. MonadError QErr m

=> HashMap Name (InputValue RemoteSchemaVariable)
user input arguments to the remote join field

-> HashMap Name (Value Void)
Contains the values of the variables that have been defined in the remote join definition

-> SelectionSet NoFragments RemoteSchemaVariable
Inserted at leaf of nested FieldCalls

-> Alias
Top-level name to set for this Field

-> NonEmpty FieldCall

-> m (Field NoFragments RemoteSchemaVariable)

Fold nested FieldCalls into a bare Field, inserting the passed selection set at the leaf of the tree we construct.

mergeValue :: Value RemoteSchemaVariable -> Value RemoteSchemaVariable -> Value RemoteSchemaVariable

createArguments :: MonadError QErr m => HashMap Name (Value Void) -> RemoteArguments -> m (HashMap Name (Value Void))

Create an argument map using the inputs taken from the hasura database.

substituteVariables

Arguments

:: HashMap Name (Value Void)
Values of the variables to substitute.

-> HashMap Name (Value Name)
Template which contains the variables.

-> Validation [Text] (HashMap Name (Value Void))

Substitute values in the argument list.