Hasura.GraphQL.Execute.Subscription.Poll.Common

Description

Multiplexed subscription poller threads; see Hasura.GraphQL.Execute.Subscription for details.

Documentation

newtype SubscriberId

Constructors

SubscriberId

Fields

unSubscriberId :: UUID

newSubscriberId :: IO SubscriberId

newtype SubscriberMetadata

Allows a user of the live query subsystem (currently websocket transport) to attach arbitrary metadata about a subscriber. This information is available as part of Subscriber in CohortExecutionDetails and can be logged by customizing in pollerlog

Constructors

SubscriberMetadata

Fields

unSubscriberMetadata :: Value

mkSubscriberMetadata :: WSId -> OperationId -> Maybe OperationName -> RequestId -> SubscriberMetadata

data Subscriber

Constructors

Subscriber

Fields

_sId :: !SubscriberId

_sMetadata :: !SubscriberMetadata

_sRequestId :: !RequestId

_sOperationName :: !(Maybe OperationName)

_sOnChangeCallback :: !OnChange

data SubscriptionMetadata

Subscription onChange metadata, used for adding more extra analytics data

Constructors

SubscriptionMetadata

Fields

_sqmExecutionTime :: !DiffTime

data SubscriptionResponse

Constructors

SubscriptionResponse

Fields

_lqrPayload :: !ByteString

_lqrExecutionTime :: !DiffTime

type SubscriptionGQResponse = GQResult SubscriptionResponse

type OnChange = SubscriptionGQResponse -> IO ()

type SubscriberMap = TMap SubscriberId Subscriber

data Cohort streamCursorVars

A batched group of Subscribers who are not only listening to the same query but also have identical session and query variables. Each result pushed to a Cohort is forwarded along to each of its Subscribers.

In SQL, each Cohort corresponds to a single row in the laterally-joined _subs table (and therefore a single row in the query result).

See also CohortMap.

Constructors

Cohort

Fields

_cCohortId :: !CohortId

a unique identifier used to identify the cohort in the generated query

_cPreviousResponse :: !(TVar (Maybe ResponseHash))

a hash of the previous query result, if any, used to determine if we need to push an updated result to the subscribers or not

_cExistingSubscribers :: !SubscriberMap

the subscribers we’ve already pushed a result to; we push new results to them iff the response changes

_cNewSubscribers :: !SubscriberMap

subscribers we haven’t yet pushed any results to; we push results to them regardless if the result changed, then merge them in the map of existing subscribers

_cStreamCursorVariables :: !streamCursorVars

a mutable type which holds the latest value of the subscription stream cursor. In case of live query subscription, this field is ignored by setting streamCursorVars to ()

newtype BatchId

The BatchId is a number based ID to uniquely identify a batch in a single poll and it's used to identify the batch to which a cohort belongs to.

Constructors

BatchId

Fields

_unBatchId :: Int

newtype ResponseHash

A hash used to determine if the result changed without having to keep the entire result in memory. Using a cryptographic hash ensures that a hash collision is almost impossible: with 256 bits, even if a subscription changes once per second for an entire year, the probability of a hash collision is ~4.294417×10-63. See Note [Blake2b faster than SHA-256].

Constructors

ResponseHash

Fields

unResponseHash :: Digest Blake2b_256

mkRespHash :: ByteString -> ResponseHash

type CohortKey = CohortVariables

A key we use to determine if two Subscribers belong in the same Cohort (assuming they already meet the criteria to be in the same Poller). Note the distinction between this and CohortId; the latter is a completely synthetic key used only to identify the cohort in the generated SQL query.

type CohortMap streamCursor = TMap CohortKey (Cohort streamCursor)

This has the invariant, maintained in removeLiveQuery, that it contains no Cohort with zero total (existing + new) subscribers.

dumpCohortMap :: CohortMap streamCursor -> IO Value

data CohortSnapshot

Constructors

CohortSnapshot

Fields

_csVariables :: !CohortVariables

_csPreviousResponse :: !(TVar (Maybe ResponseHash))

_csExistingSubscribers :: ![Subscriber]

_csNewSubscribers :: ![Subscriber]

data Poller streamCursor

A unique, multiplexed query. Each Poller has its own polling thread that periodically polls Postgres and pushes results to each of its listening Cohorts.

In SQL, an Poller corresponds to a single, multiplexed query, though in practice, Pollers with large numbers of Cohorts are batched into multiple concurrent queries for performance reasons.

Constructors

Poller

Fields

_pCohorts :: !(CohortMap streamCursor)

_pIOState :: !(TMVar PollerIOState)

This is in a separate TMVar because it’s important that we are able to construct Poller values in STM --- we need the insertion into the PollerMap to be atomic to ensure that we don’t accidentally create two for the same query due to a race. However, we can’t spawn the worker thread or create the metrics store in STM, so we insert it into the Poller only after we’re certain we won’t create any duplicates.

This var is "write once", moving monotonically from empty to full. TODO this could probably be tightened up to something like 'STM PollerIOState'

data PollerIOState

Constructors

PollerIOState

Fields

_pThread :: !Thread

a handle on the poller’s worker thread that can be used to stop it if all its cohorts stop listening

_pId :: !PollerId

data PollerKey

Constructors

PollerKey

Fields

_lgSource :: !SourceName

_lgRole :: !RoleName

_lgQuery :: !Text

type PollerMap streamCursor = Map PollerKey (Poller streamCursor)

dumpPollerMap :: Bool -> PollerMap streamCursor -> IO Value

newtype PollerId

An ID to track unique Pollers, so that we can gather metrics about each poller

Constructors

PollerId

Fields

unPollerId :: UUID

data SubscriberExecutionDetails

Constructors

SubscriberExecutionDetails

Fields

_sedSubscriberId :: !SubscriberId

_sedSubscriberMetadata :: !SubscriberMetadata

data CohortExecutionDetails

Execution information related to a cohort on a poll cycle

Constructors

CohortExecutionDetails

Fields

_cedCohortId :: !CohortId

_cedVariables :: !CohortVariables

_cedResponseSize :: !(Maybe Int)

Nothing in case of an error

_cedPushedTo :: ![SubscriberExecutionDetails]

The response on this cycle has been pushed to these above subscribers New subscribers (those which haven't been around during the previous poll cycle) will always be part of this

_cedIgnored :: ![SubscriberExecutionDetails]

The response on this cycle has *not* been pushed to these above subscribers. This would when the response hasn't changed from the previous polled cycle

_cedBatchId :: !BatchId

data BatchExecutionDetails

Execution information related to a single batched execution

Constructors

BatchExecutionDetails

Fields

_bedPgExecutionTime :: !DiffTime

postgres execution time of each batch

_bedPushTime :: !DiffTime

time to taken to push to all cohorts belonging to this batch

_bedBatchId :: !BatchId

id of the batch

_bedCohorts :: ![CohortExecutionDetails]

execution details of the cohorts belonging to this batch

_bedBatchResponseSizeBytes :: !(Maybe Int)

batchExecutionDetailMinimal :: BatchExecutionDetails -> Value

see Note [Minimal LiveQuery Poller Log]

data PollDetails

Constructors

PollDetails

Fields

_pdPollerId :: !PollerId

the unique ID (basically a thread that run as a Poller) for the Poller

_pdGeneratedSql :: !Text

the multiplexed SQL query to be run against the database with all the variables together

_pdSnapshotTime :: !DiffTime

the time taken to get a snapshot of cohorts from our SubscriptionsState data structure

_pdBatches :: ![BatchExecutionDetails]

list of execution batches and their details

_pdTotalTime :: !DiffTime

total time spent on a poll cycle

_pdLiveQueryOptions :: !SubscriptionsOptions

_pdSource :: !SourceName

_pdRole :: !RoleName

_pdParameterizedQueryHash :: !ParameterizedQueryHash

pollDetailMinimal :: PollDetails -> Value

see Note [Minimal LiveQuery Poller Log]

type SubscriptionPostPollHook = PollDetails -> IO ()

defaultSubscriptionPostPollHook :: Logger Hasura -> SubscriptionPostPollHook