Hasura.GraphQL.Execute.Subscription.Poll.StreamingQuery

Documentation

mergeOldAndNewCursorValues :: CursorVariableValues -> CursorVariableValues -> CursorVariableValues

pushResultToCohort

Arguments

:: GQResult ByteString

-> Maybe ResponseHash

-> SubscriptionMetadata

-> CursorVariableValues

-> Name
Root field name

-> (CohortSnapshot 'Streaming, Cohort 'Streaming)
subscribers to which data has been pushed, subscribers which already have this data (this information is exposed by metrics reporting)

-> IO ([SubscriberExecutionDetails], [SubscriberExecutionDetails])

pollStreamingQuery :: forall b. BackendTransport b => PollerId -> SubscriptionsOptions -> (SourceName, SourceConfig b) -> RoleName -> ParameterizedQueryHash -> MultiplexedQuery b -> CohortMap 'Streaming -> Name -> SubscriptionPostPollHook -> Maybe (IO ()) -> IO ()

A single iteration of the streaming query polling loop. Invocations on the same mutable objects may race.