Hasura.GraphQL.Logging.ExecutionLog

Description

This module holds functions and data types used for logging at the GraphQL layer. Unlike QueryLog, these are fired after queries are finished so could include things like execution time in future.

Documentation

data ExecutionLog

A GraphQL query, optionally generated SQL, and the request id makes up the | ExecutionLog

Constructors

ExecutionLog

Fields

_elRequestId :: !RequestId

_elStatistics :: !(Maybe (AnyBackend ExecutionStats))

newtype ExecutionStats b

ExecutionStatistics is a type family, which means we can't partially apply it (in AnyBackend, for example). To get round this, we have a newtype that really just wraps the type family.

Constructors

ExecutionStats (ExecutionStatistics b)

statsToAnyBackend :: forall b. HasTag b => ActionResult b -> (Maybe (AnyBackend ExecutionStats), EncJSON)

When we want to log anything from DBStepInfo, we first need to transform the backend-specific execution statistics into AnyBackend statistics. This is fine in practice because all we do with it is log it as JSON.

class Monad m => MonadExecutionLog m where

Methods

logExecutionLog :: Logger Hasura -> ExecutionLog -> m ()