Hasura.GraphQL.Logging.QueryLog

Description

This module holds functions and data types used for logging at the GraphQL layer. In contrast with, logging at the HTTP server layer.

Documentation

data QueryLog

A GraphQL query, optionally generated SQL, and the request id makes up the | QueryLog

Constructors

QueryLog

Fields

_qlQuery :: !GQLReqUnparsed

_qlGeneratedSql :: !(Maybe (RootFieldAlias, GeneratedQuery))

_qlRequestId :: !RequestId

_qlKind :: !QueryLogKind

data QueryLogKind

Constructors

QueryLogKindDatabase (Maybe BackendResolvedConnectionTemplate)

QueryLogKindAction

QueryLogKindRemoteSchema

QueryLogKindCached

QueryLogKindIntrospection

data GeneratedQuery

Constructors

GeneratedQuery

Fields

_gqQueryString :: Text

_gqPreparedArgs :: Value

class Monad m => MonadQueryLog m where

Methods

logQueryLog :: Logger Hasura -> QueryLog -> m ()