Hasura.GraphQL.Parser.Class

Description

Classes for monads used during schema construction and query parsing.

Documentation

class (Monad m, MonadParse n) => MonadSchema n m | m -> n where

A class that provides functionality used when building the GraphQL schema, i.e. constructing the Parser graph.

Methods

memoizeOn

Arguments

:: forall p a b. (Ord a, Typeable p, Typeable a, Typeable b)

=> Name
A unique name used to identify the function being memoized. There isn’t really any metaprogramming going on here, we just use a Template Haskell Name as a convenient source for a static, unique identifier.

-> a
The value to use as the memoization key. It’s the caller’s responsibility to ensure multiple calls to the same function don’t use the same key.

-> m (p n b)

-> m (p n b)

Memoizes a parser constructor function for the extent of a single schema construction process. This is mostly useful for recursive parsers; see Note [Tying the knot] for more details.

The generality of the type here allows us to use this with multiple concrete parser types:

memoizeOn :: MonadSchema n m => Name -> a -> m (Parser n b) -> m (Parser n b)
memoizeOn :: MonadSchema n m => Name -> a -> m (FieldParser n b) -> m (FieldParser n b)

memoize :: (MonadSchema n m, Ord a, Typeable a, Typeable b, Typeable p) => Name -> (a -> m (p n b)) -> a -> m (p n b)

A wrapper around memoizeOn that memoizes a function by using its argument as the key.