Hasura.GraphQL.Parser.Collect

Description

This module implements two parts of the GraphQL specification:

5.3.2 Field Selection Merging http://spec.graphql.org/June2018/#sec-Field-Selection-Merging

6.3.2 Field Collection http://spec.graphql.org/June2018/#sec-Field-Collection

These are described in completely different sections of the specification, but they’re actually highly related: both essentially normalize fields in a selection set.

Documentation

collectFields

Arguments

:: (MonadParse m, Foldable t)

=> t Name
The names of the object types and interface types the SelectionSet is selecting against.

-> SelectionSet NoFragments Variable

-> m (InsOrdHashMap Name (Field NoFragments Variable))

Collects the effective set of fields queried by a selection set by flattening fragments and merging duplicate fields.

flattenSelectionSet

Arguments

:: (MonadParse m, Foldable t)

=> t Name
The name of the object type the SelectionSet is selecting against.

-> SelectionSet NoFragments Variable

-> m [Field NoFragments Variable]

Flattens inline fragments in a selection set. For example,

{
  bar
  ... on Foo {
    baz
    qux
  }
}

is flattened to:

{
  bar
  baz
  qux
}

Nested fragments are similarly flattened, but only as is necessary: fragments inside subselection sets of individual fields are not flattened. For example,

{
  bar
  ... on Foo {
    baz {
      ... on Baz {
        foo
      }
    }
    qux
  }
}

is flattened to

{
  bar
  baz {
    ... on Baz {
      foo
    }
  }
  qux
}

leaving the innermost fragment on baz alone.

This function also applies @include and @skip directives, since they should be applied before fragments are flattened.

mergeFields :: (MonadParse m, Eq var) => [Field NoFragments var] -> m (InsOrdHashMap Name (Field NoFragments var))

Merges fields according to the rules in the GraphQL specification, specifically 5.3.2 Field Selection Merging http://spec.graphql.org/June2018/#sec-Field-Selection-Merging.