Hasura.GraphQL.Parser.Directives

Description

Definition of all supported GraphQL directives.

Documentation

directivesInfo :: forall m. MonadParse m => [DirectiveInfo]

Returns the schema information for all supported directives, for each of which a DirectiveDefinition will be inserted into the schema.

To add a new custom directives to the schema, add it to customDirectives. Inclusion directives are treated separately, as they are not concerned with altering the result of the parsers, but about whether a given field should be parsed at all (see collectFields for more information).

Alternatively, you could manually add a DirectiveInfo to this list; however, if you do so, and do not provide a corresponding definition in customDirectives, calls to parseDirectives will NOT attempt to parse that new directive: they will only validate that it is used correctly. You can use this option to implement directives that do not match our simple inclusion / custom dichotomy.

To create a directive definition, you need to create a Directive using mkDirective, and export from this module its corresponding DirectiveKey (see skip below for an example). Make sure the type in the key matches the type of the parser!

Directives may be "hidden", in which case they won't advertised in the schema, but silently accepted. This is un-advisable and should only be used when there's no other way around it.

allDirectives :: forall m. MonadParse m => [DirectiveInfo]

Not exported, only used internally; identical to directivesInfo, but also contains hidden directives.

inclusionDirectives :: forall m. MonadParse m => [Directive m]

customDirectives :: forall m. MonadParse m => [Directive m]

parseDirectives :: forall m. MonadParse m => [Directive m] -> DirectiveLocation -> [Directive Variable] -> m DirectiveMap

Parses directives, given a location. Ensures that all directives are known and match the location; subsequently builds a dependent map of the results, that can be then introspected with withDirective. The list of parsers that should be applied is given as an argument: if a valid directive is found, but for which no parser is provided, it will be ignored.

Example use:

dMap <- parseDirectives customDirectives (DLExecutable EDLQUERY) directives withDirective dMap cached $ onJust _ -> tagAsCached

withDirective :: DirectiveMap -> DirectiveKey a -> (Maybe a -> m b) -> m b

cachedDirective :: forall m. MonadParse m => Directive m

data CachedDirective

Constructors

CachedDirective

Fields

cdTtl :: Int

cdRefresh :: Bool

cached :: DirectiveKey CachedDirective

multipleRootFieldsDirective :: forall m. MonadParse m => Directive m

multipleRootFields :: DirectiveKey ()

skipDirective :: MonadParse m => Directive m

includeDirective :: MonadParse m => Directive m

skip :: DirectiveKey Bool

include :: DirectiveKey Bool

ifArgument :: MonadParse m => InputFieldsParser m Bool

data Directive m where

Constructors

Directive

Fields

:: forall m a. (MonadParse m, Typeable a)

=> { dDefinition :: DirectiveInfo

, dAdvertised :: Bool

, dParser :: Directive Variable -> m a

} -> Directive m

data DirectiveKey a where

Constructors

DirectiveKey :: Typeable a => Name -> DirectiveKey a

type DirectiveMap = DMap DirectiveKey Identity

mkDirective :: (MonadParse m, Typeable a) => Name -> Maybe Description -> Bool -> [DirectiveLocation] -> InputFieldsParser m a -> Directive m