Hasura.GraphQL.Parser.Internal.Convert

Description

This module defines all functions that convert between different representations of values in the schema; most commonly: GraphQL literals, JSON values, and InputValue, a type that provides an abstraction above both of those.

Documentation

valueToJSON :: MonadParse m => GType -> InputValue Variable -> m Value

jsonToGraphQL :: Value -> Either ErrorMessage (Value Void)