Hasura.GraphQL.Parser.Internal.Input

Description

Defines the Parser type and its primitive combinators.

Documentation

inputParserInput :: forall k. 'Input <: k => ParserInput k :~: InputValue Variable

pInputParser :: forall origin k m a. 'Input <: k => Parser origin k m a -> InputValue Variable -> m a

data InputFieldsParser origin m a

Parses some collection of input fields. Build an InputFieldsParser using field, fieldWithDefault, or fieldOptional, combine several together with the Applicative instance, and finish it off using object to turn it into a Parser.

Constructors

InputFieldsParser

Fields

ifDefinitions :: [Definition origin (InputFieldInfo origin)]

ifParser :: HashMap Name (InputValue Variable) -> m a

field :: (MonadParse m, 'Input <: k) => Name -> Maybe Description -> Parser origin k m a -> InputFieldsParser origin m a

Creates a parser for an input field. For nullable value parsers, if an input field is omitted, the provided parser will be called with a value of null.

fieldOptional :: (MonadParse m, 'Input <: k) => Name -> Maybe Description -> Parser origin k m a -> InputFieldsParser origin m (Maybe a)

Creates a parser for a nullable field with no default value. If the field is omitted, the provided parser will not be called. This allows a field to distinguish an omitted field from a field supplied with null (which is permitted by the GraphQL specification); see Note [The value of omitted fields] for more details.

If you want a field with a default value of null, use field with a nullable value parser, instead.

fieldWithDefault

Arguments

:: (MonadParse m, 'Input <: k)

=> Name

-> Maybe Description

-> Value Void
default value

-> Parser origin k m a

-> InputFieldsParser origin m a

Creates a parser for an input field with the given default value. The resulting field will always be optional, even if the underlying parser rejects null values. The underlying parser is always called.

enum :: MonadParse m => Name -> Maybe Description -> NonEmpty (Definition origin EnumValueInfo, a) -> Parser origin 'Both m a

object :: MonadParse m => Name -> Maybe Description -> InputFieldsParser origin m a -> Parser origin 'Input m a

list :: forall origin k m a. (MonadParse m, 'Input <: k) => Parser origin k m a -> Parser origin k m [a]