Hasura.GraphQL.Parser.Internal.Parser

Description

Defines the Parser type and its primitive combinators.

Documentation

bind :: Monad m => Parser origin k m a -> (a -> m b) -> Parser origin k m b infixl 1

bindFields :: Monad m => InputFieldsParser origin m a -> (a -> m b) -> InputFieldsParser origin m b infixl 1

data FieldParser origin m a

A parser for a single field in a selection set. Build a FieldParser with selection or subselection, and combine them together with selectionSet to obtain a Parser.

Constructors

FieldParser

Fields

fDefinition :: Definition origin (FieldInfo origin)

fParser :: Field NoFragments Variable -> m a

bindField :: Monad m => FieldParser origin m a -> (a -> m b) -> FieldParser origin m b infixl 1

data ParsedSelection a

A single parsed field in a selection set.

Constructors

SelectField a
An ordinary field.

SelectTypename Name
The magical __typename field, implicitly available on all objects part of GraphQL introspection http://spec.graphql.org/June2018/#sec-Type-Name-Introspection.

handleTypename :: (Name -> a) -> ParsedSelection a -> a

nullable :: forall origin k m a. (MonadParse m, 'Input <: k) => Parser origin k m a -> Parser origin k m (Maybe a)

nonNullableField :: forall m origin a. FieldParser origin m a -> FieldParser origin m a

Decorate a schema field as NON_NULL

nullableField :: forall m origin a. FieldParser origin m a -> FieldParser origin m a

Decorate a schema field as NULL

multipleField :: forall m origin a. FieldParser origin m a -> FieldParser origin m a

wrapFieldParser :: forall m origin a. GType -> FieldParser origin m a -> FieldParser origin m a

Decorate a schema field with reference to given @GType

nonNullableParser :: forall m origin a. Parser origin 'Output m a -> Parser origin 'Output m a

Decorate a schema output type as NON_NULL

nullableParser :: forall m origin a. Parser origin 'Output m a -> Parser origin 'Output m a

Make a schema output as nullable

multiple :: forall m origin a. Parser origin 'Output m a -> Parser origin 'Output m a

setDefinitionOrigin :: origin -> Definition origin a -> Definition origin a

Set the metadata origin of a Definition

setParserOrigin :: forall origin k m a. origin -> Parser origin k m a -> Parser origin k m a

Set the metadata origin of a Parser

setFieldParserOrigin :: forall m origin a. origin -> FieldParser origin m a -> FieldParser origin m a

Set the metadata origin of a FieldParser

setInputFieldsParserOrigin :: forall m origin a. origin -> InputFieldsParser origin m a -> InputFieldsParser origin m a

Set the metadata origin of the arguments in a InputFieldsParser

setDefinitionDirectives :: [Directive Void] -> Definition origin a -> Definition origin a

Set the directives of a Definition

setParserDirectives :: forall origin k m a. [Directive Void] -> Parser origin k m a -> Parser origin k m a

Set the directives of a Parser

setFieldParserDirectives :: forall m origin a. [Directive Void] -> FieldParser origin m a -> FieldParser origin m a

Set the directives of a FieldParser

setInputFieldsParserDirectives :: forall m origin a. [Directive Void] -> InputFieldsParser origin m a -> InputFieldsParser origin m a

Set the directives of the arguments in a InputFieldsParser

selectionSet :: MonadParse m => Name -> Maybe Description -> [FieldParser origin m a] -> Parser origin 'Output m (InsOrdHashMap Name (ParsedSelection a))

A variant of selectionSetObject which doesn't implement any interfaces

safeSelectionSet :: forall n m origin a. (MonadError ErrorMessage n, MonadParse m, Eq origin, Hashable origin, ToErrorValue origin) => Name -> Maybe Description -> [FieldParser origin m a] -> n (Parser origin 'Output m (InsOrdHashMap Name (ParsedSelection a)))

selectionSetObject

Arguments

:: MonadParse m

=> Name

-> Maybe Description

-> [FieldParser origin m a]
Fields of this object, including any fields that are required from the interfaces that it implements. Note that we can't derive those fields from the list of interfaces (next argument), because the types of the fields of the object are only required to be *subtypes* of the types of the fields of the interfaces it implements.

-> [Parser origin 'Output m b]
Interfaces implemented by this object; see Note [The interfaces story] in Hasura.GraphQL.Parser.Schema.

-> Parser origin 'Output m (InsOrdHashMap Name (ParsedSelection a))

selectionSetInterface

Arguments

:: (MonadParse n, Traversable t)

=> Name

-> Maybe Description

-> [FieldParser origin n a]
Fields defined in this interface

-> t (Parser origin 'Output n b)
Parsers for the object types that implement this interface; see Note [The interfaces story] in Hasura.GraphQL.Parser.Schema for details.

-> Parser origin 'Output n (t b)

selectionSetUnion

Arguments

:: (MonadParse n, Traversable t)

=> Name

-> Maybe Description

-> t (Parser origin 'Output n b)
The member object types.

-> Parser origin 'Output n (t b)

selection

Arguments

:: forall m origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Both m b
type of the result

-> FieldParser origin m a

Builds a FieldParser for a field that does not take a subselection set, i.e. a field that returns a scalar or enum. The field’s type is taken from the provided Parser, but the Parser is not otherwise used.

See also Note [The delicate balance of GraphQL kinds] in Hasura.GraphQL.Parser.Schema.

rawSelection

Arguments

:: forall m origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Both m b
type of the result

-> FieldParser origin m (Maybe Name, HashMap Name (Value Variable), a)
alias provided (if any), and the arguments

subselection

Arguments

:: forall m origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Output m b
parser for the subselection set

-> FieldParser origin m (a, b)

Builds a FieldParser for a field that takes a subselection set, i.e. a field that returns an object.

For example, subselection name _ args fields produces schema:

name (args) { fields }

See also Note [The delicate balance of GraphQL kinds] in Hasura.GraphQL.Parser.Schema.

rawSubselection

Arguments

:: forall m origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Output m b
parser for the subselection set

-> FieldParser origin m (Maybe Name, HashMap Name (Value Variable), a, b)

selection_

Arguments

:: MonadParse m

=> Name

-> Maybe Description

-> Parser origin 'Both m a
type of the result

-> FieldParser origin m ()

A shorthand for a selection that takes no arguments.

subselection_

Arguments

:: MonadParse m

=> Name

-> Maybe Description

-> Parser origin 'Output m a
parser for the subselection set

-> FieldParser origin m a

A shorthand for a subselection that takes no arguments.