Hasura.GraphQL.Parser.Internal.Scalars

Description

This module defines all backend-agnostic scalars we use throughout the schema. This includes GraphQL scalars, and several other custom ones.

Documentation

boolean :: MonadParse m => Parser 'Both m Bool

int :: MonadParse m => Parser 'Both m Int32

float :: MonadParse m => Parser 'Both m Double

string :: MonadParse m => Parser 'Both m Text

identifier :: MonadParse m => Parser 'Both m Text

As an input type, any string or integer input value should be coerced to ID as Text https://spec.graphql.org/June2018/#sec-ID

uuid :: MonadParse m => Parser 'Both m UUID

json :: MonadParse m => Parser 'Both m Value

jsonb :: MonadParse m => Parser 'Both m Value

nonNegativeInt :: MonadParse m => Parser 'Both m Int32

Additional validation on integers. We do keep the same type name in the schema for backwards compatibility. TODO: when we can do a breaking change, we can rename the type to NonNegativeInt.

bigInt :: MonadParse m => Parser 'Both m Int64

GraphQL ints are 32-bit integers; but in some places we want to accept bigger ints. To do so, we declare a cusom scalar that can represent 64-bit ints, which accepts both int literals and string literals. We do keep the same type name in the schema for backwards compatibility. TODO: when we can do a breaking change, we can rename the type to BigInt.

unsafeRawScalar :: MonadParse n => Name -> Maybe Description -> Parser 'Both n (InputValue Variable)

Explicitly define any desired scalar type.

This was considered unsafe because, unlike all other scalar definitions, it doesn't enforce that we properly peel variables, and let the caller decide how they want to deal with potential variables. This allows the caller to bypass type-checking and (deprecated) non-reusability semantics (see comment about re-usability in TypeChecking).

In practice, this function isn't very dangerous, and preferable to an explicit use of the Parser constructor.

jsonScalar :: MonadParse m => Name -> Maybe Description -> Parser 'Both m Value

Creates a parser that transforms its input into a JSON value. valueToJSON does properly unpack variables.

mkScalar :: MonadParse m => Name -> Maybe Description -> (InputValue Variable -> m a) -> Parser 'Both m a

convertWith :: MonadParse m => (a -> Parser b) -> a -> m b