Hasura.GraphQL.Parser.Internal.TypeChecking

Description

Here are all functions related to type-checking within the schema. This includes functions that deal with variables, and functions that craft error messages.

Documentation

peelVariable :: MonadParse m => GType -> InputValue Variable -> m (InputValue Variable)

Peeling a variable.

Whenever we need to inspect a value, for instance when parsing the value of an input argument, there are two main cases: either we have a GraphQL literal (be it explicitly given, or a default argument), or we have a variable:

where(limit: 42) where(limit: $limit)

In the case of a variable, we need to "open" it and inspect its content. This forces us to do several things:

Type-checking We know what the variable was declared as, we know what type we expect at our current location in the schema: we can check that they match.

Update re-usability In the past, we used to cache the generated execution plan for a given query. To do so properly, we had to identify queries that couldn't be cached; and any root field that uses the content of a variable (outside of a column's value) couldn't be cached: the same graphql expression would lead to a different execution plan if the value of the variable were to change. We no longer cache execution plans; but we might do it again in the future, which is why we haven't removed some of the code that deals with re-usability.

peelVariableWith :: MonadParse m => Bool -> GType -> InputValue Variable -> m (InputValue Variable)

typeCheck :: MonadParse m => Bool -> GType -> Variable -> m ()

Type-checking.

Checks whether the type of a variable is compatible with the type at the location at which it is used. This is an implementation of the function described in 5.8.5 of the spec: http://spec.graphql.org/June2018/#sec-All-Variable-Usages-are-Allowed

No input type coercion is allowed between variables: coercion rules only allow when translating a value from a literal. It is therefore not allowed to use an Int variable at a Float location, despite the fact that it is legal to use an Int literal at a Float location.

Furthermore, it's also worth noting that there's one tricky case where we might allow a nullable variable at a non-nullable location: when either side has a non-null default value. That's because GraphQL conflates nullability and optionality. See also Note [When are fields optional?].

isVariableUsageAllowed

Arguments

:: Bool
does the location have a default value

-> GType
the location type

-> Variable
the variable

-> Bool

typeMismatch :: (HasName n, MonadParse m) => n -> Text -> InputValue Variable -> m a

describeValue :: InputValue Variable -> Text

describeValueWith :: (var -> Text) -> InputValue var -> Text