Hasura.GraphQL.Parser.Internal.Types

Description

Defines the Parser type and its primitive combinators.

Documentation

data Parser k m a

A Parser that corresponds to a type in the GraphQL schema. A Parser is really two things at once:

As its name implies, a Parser can be used to parse GraphQL queries (via runParser).

Less obviously, a Parser represents a slice of the GraphQL schema, since every Parser corresponds to a particular GraphQL type, and information about that type can be recovered (via parserType).

A natural way to view this is that Parsers support a sort of dynamic reflection: in addition to running a Parser on an input query, you can ask it to tell you about what type of input it expects. Importantly, you can do this even if you don’t have a query to parse; this is necessary to implement GraphQL introspection, which provides precisely this sort of reflection on types.

Another way of viewing a Parser is a little more quantum: just as light “sometimes behaves like a particle and sometimes behaves like a wave,” a Parser “sometimes behaves like a query parser and sometimes behaves like a type.” In this way, you can think of a function that produces a Parser as simultaneously both a function that constructs a GraphQL schema and a function that parses a GraphQL query. Parser constructors therefore interleave two concerns: information about a type definition (like the type’s name and description) and information about how to parse a query on that type.

Notably, these two concerns happen at totally different phases in the program: GraphQL schema construction happens when graphql-engine first starts up, before it receives any GraphQL queries at all. But query parsing obviously can’t happen until there is actually a query to parse. For that reason, it’s useful to take care to distinguish which effects are happening at which phase during Parser construction, since otherwise you may get mixed up!

For some more information about how to interpret the meaning of a Parser, see Note [The meaning of Parser 'Output].

Constructors

Parser

Fields

pType :: ~(Type k)

Lazy for knot-tying reasons; see Note [Tying the knot] in Hasura.GraphQL.Parser.Class.

pParser :: ParserInput k -> m a

type family ParserInput k where ...

Equations

ParserInput 'Both = InputValue Variable

ParserInput 'Input = InputValue Variable

ParserInput 'Output = SelectionSet NoFragments Variable

parserType :: Parser k m a -> Type k

runParser :: Parser k m a -> ParserInput k -> m a