Hasura.GraphQL.Parser.Monad

Description

Monad transformers for GraphQL schema construction and query parsing.

Documentation

newtype SchemaT n m a

Constructors

SchemaT

Fields

unSchemaT :: StateT (DMap ParserId (ParserById n)) m a

runSchemaT :: forall m n a. Monad m => SchemaT n m a -> m a

data ParserId (t :: ((Type -> Type) -> Type -> Type, Type)) where

A key used to distinguish calls to memoized functions. The Name distinguishes calls to completely different parsers, and the a value records the arguments.

Constructors

ParserId :: (Ord a, Typeable p, Typeable a, Typeable b) => Name -> a -> ParserId '(p, b)

data family ParserById (m :: Type -> Type) (a :: ((Type -> Type) -> Type -> Type, Type))

A newtype wrapper around a Parser that rearranges the type parameters so that it can be indexed by a ParserId in a DMap.

This is really just a single newtype, but it’s implemented as a data family because GHC doesn’t allow ordinary datatype declarations to pattern-match on type parameters, and we want to match on the tuple.

newtype ParseT m a

Constructors

ParseT

Fields

unParseT :: ReaderT JSONPath (ValidateT (NESeq ParseError) m) a

runParseT :: Functor m => ParseT m a -> m (Either (NESeq ParseError) a)

data ParseError

Constructors

ParseError

Fields

pePath :: JSONPath

peMessage :: Text

peCode :: Code

reportParseErrors :: MonadError QErr m => NESeq ParseError -> m a