Hasura.GraphQL.Parser.Names

Documentation

class HasName a where

Methods

getName :: a -> Name

newtype MkTypename

Type name customization

Constructors

MkTypename

Fields

runMkTypename :: Name -> Name

withTypenameCustomization :: forall m r a. (MonadReader r m, Has MkTypename r) => MkTypename -> m a -> m a

Inject a new MkTypename customization function into the environment. This can be used by schema-building code (with MonadBuildSchema constraint) to ensure the correct type name customizations are applied.

mkTypename :: (MonadReader r m, Has MkTypename r) => Name -> m Name

Apply the type name customization function from the current environment.