Hasura.GraphQL.Parser.Schema

Description

Types for representing a GraphQL schema.

Documentation

class HasName a where

Methods

getName :: a -> Name

newtype MkTypename

Type name customization

Constructors

MkTypename

Fields

runMkTypename :: Name -> Name

withTypenameCustomization :: forall m r a. (MonadReader r m, Has MkTypename r) => MkTypename -> m a -> m a

Inject a new MkTypename customization function into the environment. This can be used by schema-building code (with MonadBuildSchema constraint) to ensure the correct type name customizations are applied.

mkTypename :: (MonadReader r m, Has MkTypename r) => Name -> m Name

Apply the type name customization function from the current environment.

data Kind

GraphQL types are divided into two classes: input types and output types. The GraphQL spec does not use the word “kind” to describe these classes, but it’s an apt term.

Some GraphQL types can be used at either kind, so we also include the Both kind, the superkind of both Input and Output. The <: class provides kind subsumption constraints.

For more details, see http://spec.graphql.org/June2018/#sec-Input-and-Output-Types.

Constructors

Both
see Note [The 'Both kind]

Input

Output

data k1 :<: k2 where

Evidence for <:.

Constructors

KRefl :: k :<: k

KBoth :: k :<: 'Both

class k1 <: k2 where

Kind subsumption. The GraphQL kind hierarchy is extremely simple:

    Both
    /  \
Input  Output

Various functions in this module use <: to allow Both to be used in places where Input or Output would otherwise be expected.

Methods

subKind :: k1 :<: k2

data Nullability

Constructors

Nullable

NonNullable

isNullable :: Nullability -> Bool

data Type k

Constructors

TNamed Nullability (Definition (TypeInfo k))

TList Nullability (Type k)

eqType :: Type k1 -> Type k2 -> Bool

Like ==, but can compare Types of different kinds.

typeNullability :: Type k -> Nullability

nullableType :: Type k -> Type k

nonNullableType :: Type k -> Type k

toGraphQLType :: Type k -> GType

data InputObjectInfo

Constructors

InputObjectInfo ~[Definition InputFieldInfo]

data ObjectInfo

Constructors

ObjectInfo

Fields

oiFields :: ~[Definition FieldInfo]

The fields that this object has. This consists of the fields of the interfaces that it implements, as well as any additional fields.

oiImplements :: ~[Definition InterfaceInfo]

The interfaces that this object implements (inheriting all their fields). See Note [The interfaces story] for more details.

data InterfaceInfo

Type information for a GraphQL interface; see Note [The interfaces story] for more details.

Note: in the current working draft of the GraphQL specification (> June 2018), interfaces may implement other interfaces, but we currently don't support this.

Constructors

InterfaceInfo

Fields

iiFields :: ~[Definition FieldInfo]

Fields declared by this interface. Every object implementing this interface must include those fields.

iiPossibleTypes :: ~[Definition ObjectInfo]

Objects that implement this interface. See Note [The interfaces story] for why we include that information here.

data UnionInfo

Constructors

UnionInfo

Fields

uiPossibleTypes :: ~[Definition ObjectInfo]

The member object types of this union.

data TypeInfo k where

Constructors

TIScalar :: TypeInfo 'Both

TIEnum :: NonEmpty (Definition EnumValueInfo) -> TypeInfo 'Both

TIInputObject :: InputObjectInfo -> TypeInfo 'Input

TIObject :: ObjectInfo -> TypeInfo 'Output

TIInterface :: InterfaceInfo -> TypeInfo 'Output

TIUnion :: UnionInfo -> TypeInfo 'Output

eqTypeInfo :: TypeInfo k1 -> TypeInfo k2 -> Bool

Like ==, but can compare TypeInfos of different kinds.

getTypeInfo :: Type k -> Definition (TypeInfo k)

getObjectInfo :: Type k -> Maybe (Definition ObjectInfo)

getInterfaceInfo :: Type k -> Maybe (Definition InterfaceInfo)

data SomeTypeInfo

Constructors

forall k. SomeTypeInfo (TypeInfo k)

data Definition a

Constructors

Definition

Fields

dName :: Name

dDescription :: Maybe Description

dInfo :: ~a

Lazy to allow mutually-recursive type definitions.

data EnumValueInfo

Enum values have no extra information except for the information common to all definitions, so this is just a placeholder for use as Definition EnumValueInfo.

Constructors

EnumValueInfo

data InputFieldInfo

Constructors

forall k.'Input <: k => InputFieldInfo (Type k) (Maybe (Value Void))
An input field with a type and possibly a default value. If a default value is provided, it should be a valid value for the type.

Note that a default value of VNull is subtly different from having no default value at all. If no default value is provided (i.e. Nothing), the GraphQL specification allows distinguishing provided null values from values left completely absent; see Note [The value of omitted fields] in Hasura.GraphQL.Parser.Internal.Parser.

data FieldInfo

Constructors

forall k.'Output <: k => FieldInfo

Fields

fArguments :: [Definition InputFieldInfo]

fType :: Type k

data InputValue v

See Note [Parsing variable values].

Constructors

GraphQLValue (Value v)

JSONValue Value

data Variable

Constructors

Variable

Fields

vInfo :: VariableInfo

vType :: GType

vValue :: InputValue Void

Note: if the variable was null or was not provided and the field has a non-null default value, this field contains the default value, not VNull.

data VariableInfo

Constructors

VIRequired Name

VIOptional Name (Value Void)
Unlike fields (see InputFieldInfo), nullable variables with no default value are indistinguishable from variables with a default value of null, so we don’t distinguish those cases here.

data DirectiveInfo

This type represents the directives information to be served over GraphQL introspection

Constructors

DirectiveInfo

Fields

diName :: !Name

diDescription :: !(Maybe Description)

diArguments :: ![Definition InputFieldInfo]

diLocations :: ![DirectiveLocation]

data Schema

This type contains all the information needed to efficiently serve GraphQL introspection queries. It corresponds to the GraphQL __Schema type defined in 4.5 Schema Introspection http://spec.graphql.org/June2018/#sec-Introspection. See also Note [Basics of introspection schema generation].

Constructors

Schema

Fields

sDescription :: Maybe Description

sTypes :: HashMap Name (Definition SomeTypeInfo)

sQueryType :: Type 'Output

sMutationType :: Maybe (Type 'Output)

sSubscriptionType :: Maybe (Type 'Output)

sDirectives :: [DirectiveInfo]

data TypeDefinitionsWrapper where

Constructors

TypeDefinitionsWrapper :: HasTypeDefinitions a => a -> TypeDefinitionsWrapper

collectTypeDefinitions :: HasTypeDefinitions a => a -> Either ConflictingDefinitions (HashMap Name (Definition SomeTypeInfo))

Recursively collects all type definitions accessible from the given value.

newtype TypeOriginStack

Constructors

TypeOriginStack [Name]

typeOriginRecurse :: Name -> TypeOriginStack -> TypeOriginStack

typeRootRecurse :: Name -> TypeOriginStack -> TypeOriginStack

data ConflictingDefinitions

Constructors

ConflictingDefinitions (Definition SomeTypeInfo, TypeOriginStack) (Definition SomeTypeInfo, NonEmpty TypeOriginStack)
Type collection has found at least two types with the same name.

newtype TypeAccumulation a

Although the majority of graphql-engine is written in terms of abstract mtl-style effect monads, we figured out that this particular codepath is quite hot, and that mtl has a measurable negative effect for accumulating types from the schema, both in profiling and in benchmarking. Using an explicit transformers-style effect stack seems to overall memory usage by about 3-7%.

Constructors

TypeAccumulation

Fields

runTypeAccumulation :: ReaderT TypeOriginStack (StateT (HashMap Name (Definition SomeTypeInfo, NonEmpty TypeOriginStack)) (ExceptT ConflictingDefinitions Identity)) a

class HasTypeDefinitions a where

Methods

accumulateTypeDefinitions :: a -> TypeAccumulation ()

Recursively accumulates all type definitions accessible from the given value. This is done statefully to avoid infinite loops arising from recursive type definitions; see Note [Tying the knot] in Hasura.GraphQL.Parser.Class.