Hasura.GraphQL.Parser.Schema.Convert

Description

Provides utilities to convert a runtime schema back to a static intrsospection schema.

Documentation

convertToSchemaIntrospection :: Schema -> SchemaIntrospection

Convert back from our "live" schema representation into a flat static set of definitions.

convertType :: SomeDefinitionTypeInfo -> TypeDefinition [Name] InputValueDefinition

convertEnumValue :: Definition EnumValueInfo -> EnumValueDefinition

convertInputField :: Definition InputFieldInfo -> InputValueDefinition

convertField :: Definition FieldInfo -> FieldDefinition InputValueDefinition

getDefinitionName :: Definition a -> Name

noDirectives :: [Directive Void]