Hasura.GraphQL.Parser.Variable

Documentation

data InputValue v

See Note [Parsing variable values].

Constructors

GraphQLValue (Value v)

JSONValue Value

data Variable

Constructors

Variable

Fields

vInfo :: VariableInfo

vType :: GType

vValue :: InputValue Void

Note: if the variable was null or was not provided and the field has a non-null default value, this field contains the default value, not VNull.

data VariableInfo

Constructors

VIRequired Name

VIOptional Name (Value Void)
Unlike fields (see InputFieldInfo), nullable variables with no default value are indistinguishable from variables with a default value of null, so we don’t distinguish those cases here.