Hasura.GraphQL.Schema

Documentation

buildGQLContext :: forall m. (MonadError QErr m, MonadIO m, HasServerConfigCtx m) => GraphQLQueryType -> SourceCache -> RemoteSchemaCache -> ActionCache -> AnnotatedCustomTypes -> m (HashMap RoleName (RoleContext GQLContext), GQLContext, Seq InconsistentMetadata)

customizeFields :: forall f n db remote action. (Functor f, MonadParse n) => SourceCustomization -> MkTypename -> f [FieldParser n (RootField db remote action Value)] -> f [FieldParser n (NamespacedField (RootField db remote action Value))]

buildRoleContext :: forall m. (MonadError QErr m, MonadIO m) => (SQLGenCtx, GraphQLQueryType, FunctionPermissionsCtx) -> SourceCache -> RemoteSchemaCache -> [ActionInfo] -> AnnotatedCustomTypes -> [(RemoteSchemaName, RemoteRelationshipQueryContext)] -> RoleName -> RemoteSchemaPermsCtx -> m (RoleContext GQLContext)

buildRelayRoleContext :: forall m. (MonadError QErr m, MonadIO m) => (SQLGenCtx, GraphQLQueryType, FunctionPermissionsCtx) -> SourceCache -> [ActionInfo] -> AnnotatedCustomTypes -> RoleName -> m (RoleContext GQLContext)

buildFullestDBSchema :: forall m. (MonadError QErr m, MonadIO m) => QueryContext -> SourceCache -> [ActionInfo] -> AnnotatedCustomTypes -> m (Parser 'Output (ParseT Identity) (RootFieldMap (QueryRootField UnpreparedValue)), Maybe (Parser 'Output (ParseT Identity) (RootFieldMap (MutationRootField UnpreparedValue))))

unauthenticatedContext :: forall m. (MonadError QErr m, MonadIO m) => [FieldParser (ParseT Identity) (NamespacedField RemoteField)] -> [FieldParser (ParseT Identity) (NamespacedField RemoteField)] -> RemoteSchemaPermsCtx -> m GQLContext

buildRoleBasedRemoteSchemaParser :: forall m. (MonadError QErr m, MonadIO m) => RoleName -> RemoteSchemaCache -> m [(RemoteSchemaName, RemoteRelationshipQueryContext)]

remoteSchemaFields :: forall m. MonadState (Seq InconsistentMetadata) m => [Name] -> [Name] -> HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject) -> m [(RemoteSchemaName, RemoteRelationshipQueryContext)]

buildQueryFields :: forall b r m n. MonadBuildSchema b r m n => SourceName -> SourceConfig b -> TableCache b -> FunctionCache b -> Maybe QueryTagsConfig -> m [FieldParser n (QueryRootField UnpreparedValue)]

buildRelayQueryFields :: forall b r m n. MonadBuildSchema b r m n => SourceName -> SourceConfig b -> TableCache b -> FunctionCache b -> Maybe QueryTagsConfig -> m [FieldParser n (QueryRootField UnpreparedValue)]

buildMutationFields :: forall b r m n. MonadBuildSchema b r m n => Scenario -> SourceName -> SourceConfig b -> TableCache b -> FunctionCache b -> Maybe QueryTagsConfig -> m [FieldParser n (MutationRootField UnpreparedValue)]

buildQueryParser :: forall m n r. (MonadSchema n m, MonadTableInfo r m, MonadRole r m, Has QueryContext r, Has MkTypename r, Has MkRootFieldName r, Has CustomizeRemoteFieldName r) => [FieldParser n (NamespacedField (QueryRootField UnpreparedValue))] -> [FieldParser n (NamespacedField RemoteField)] -> [ActionInfo] -> AnnotatedCustomTypes -> Maybe (Parser 'Output n (RootFieldMap (MutationRootField UnpreparedValue))) -> Maybe (Parser 'Output n (RootFieldMap (QueryRootField UnpreparedValue))) -> m (Parser 'Output n (RootFieldMap (QueryRootField UnpreparedValue)))

Prepare the parser for query-type GraphQL requests, but with introspection for queries, mutations and subscriptions built in.

queryWithIntrospectionHelper :: forall n m. (MonadSchema n m, MonadError QErr m) => [FieldParser n (NamespacedField (QueryRootField UnpreparedValue))] -> Maybe (Parser 'Output n (RootFieldMap (MutationRootField UnpreparedValue))) -> Maybe (Parser 'Output n (RootFieldMap (QueryRootField UnpreparedValue))) -> m (Parser 'Output n (RootFieldMap (QueryRootField UnpreparedValue)))

queryRootFromFields :: forall n m. (MonadError QErr m, MonadParse n) => [FieldParser n (NamespacedField (QueryRootField UnpreparedValue))] -> m (Parser 'Output n (RootFieldMap (QueryRootField UnpreparedValue)))

collectTypes :: forall m a. (MonadError QErr m, HasTypeDefinitions a) => a -> m (HashMap Name (Definition SomeTypeInfo))

buildSubscriptionParser :: forall m n r. (MonadSchema n m, MonadTableInfo r m, MonadRole r m, Has QueryContext r, Has MkTypename r, Has MkRootFieldName r, Has CustomizeRemoteFieldName r) => [FieldParser n (NamespacedField (QueryRootField UnpreparedValue))] -> [ActionInfo] -> AnnotatedCustomTypes -> m (Maybe (Parser 'Output n (RootFieldMap (QueryRootField UnpreparedValue))))

Prepare the parser for subscriptions. Every postgres query field is exposed as a subscription along with fields to get the status of asynchronous actions.

buildMutationParser :: forall m n r. (MonadSchema n m, MonadTableInfo r m, MonadRole r m, Has QueryContext r, Has MkTypename r, Has MkRootFieldName r, Has CustomizeRemoteFieldName r) => [FieldParser n (NamespacedField RemoteField)] -> [ActionInfo] -> AnnotatedCustomTypes -> [FieldParser n (NamespacedField (MutationRootField UnpreparedValue))] -> m (Maybe (Parser 'Output n (RootFieldMap (MutationRootField UnpreparedValue))))

mkRootField :: forall b m n a db remote action raw. (HasTag b, Functor m, Functor n) => SourceName -> SourceConfig b -> Maybe QueryTagsConfig -> (a -> db b) -> m [FieldParser n a] -> m [FieldParser n (RootField db remote action raw)]

All the BackendSchema methods produce something of the form m [FieldParser n a], where a is something specific to what is being parsed by the given method.

In order to build the complete schema these must be homogenised and be annotated with query-tag data, which this function makes easy.

takeExposedAs :: FunctionExposedAs -> FunctionCache b -> FunctionCache b

subscriptionRoot :: Name

mutationRoot :: Name

queryRoot :: Name

finalizeParser :: Parser 'Output (ParseT Identity) a -> ParserFn a

type ConcreteSchemaT m a = SchemaT (ParseT Identity) (ReaderT (RoleName, SourceCache, QueryContext, MkTypename, MkRootFieldName, CustomizeRemoteFieldName) m) a

runMonadSchema :: forall m a. Monad m => RoleName -> QueryContext -> SourceCache -> ConcreteSchemaT m a -> m a

type RemoteSchemaCache = HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject)

buildBackendSource :: (forall b. BackendSchema b => SourceInfo b -> r) -> AnyBackend SourceInfo -> r

typenameToNamespacedRawRF :: ParsedSelection (NamespacedField (RootField db remote action Value)) -> NamespacedField (RootField db remote action Value)

typenameToRawRF :: ParsedSelection (RootField db remote action Value) -> RootField db remote action Value