Hasura.GraphQL.Schema.BoolExp

Documentation

class AggregationPredicatesSchema (b :: BackendType) where

Backends implement this type class to specify the schema of aggregation predicates.

The default implementation results in a parser that does not parse anything.

The scope of this class is local to the function boolExp. In particular, methods in `class BackendSchema` and `type MonadBuildSchema` should *NOT* include this class as a constraint.

Methods

aggregationPredicatesParser :: forall r m n. MonadBuildSourceSchema b r m n => TableInfo b -> SchemaT r m (Maybe (InputFieldsParser n [AggregationPredicates b (UnpreparedValue b)]))

logicalModelBoolExp :: forall b r m n. (MonadBuildSchema b r m n, AggregationPredicatesSchema b) => LogicalModelInfo b -> SchemaT r m (Parser 'Input n (AnnBoolExp b (UnpreparedValue b)))

input type_bool_exp {
  _or: [type_bool_exp!]
  _and: [type_bool_exp!]
  _not: type_bool_exp
  column: type_comparison_exp
  ...
}

| Boolean expression for logical models

tableBoolExp :: forall b r m n. (MonadBuildSchema b r m n, AggregationPredicatesSchema b) => TableInfo b -> SchemaT r m (Parser 'Input n (AnnBoolExp b (UnpreparedValue b)))

input type_bool_exp {
  _or: [type_bool_exp!]
  _and: [type_bool_exp!]
  _not: type_bool_exp
  column: type_comparison_exp
  ...
}

| Booleans expressions for tables

mkBoolOperator

Arguments

:: (MonadParse n, 'Input <: k)

=> NamingCase
Naming convention for the field

-> DangerouslyCollapseBooleans
shall this be collapsed to True when null is given?

-> GQLNameIdentifier
name of this operator

-> Maybe Description
optional description

-> Parser k n a
parser for the underlying value

-> InputFieldsParser n (Maybe a)

equalityOperators

Arguments

:: (MonadParse n, 'Input <: k)

=> NamingCase

-> DangerouslyCollapseBooleans
shall this be collapsed to True when null is given?

-> Parser k n (UnpreparedValue b)
parser for one column value

-> Parser k n (UnpreparedValue b)
parser for a list of column values

-> [InputFieldsParser n (Maybe (OpExpG b (UnpreparedValue b)))]

comparisonOperators

Arguments

:: (MonadParse n, 'Input <: k)

=> NamingCase

-> DangerouslyCollapseBooleans
shall this be collapsed to True when null is given?

-> Parser k n (UnpreparedValue b)
parser for one column value

-> [InputFieldsParser n (Maybe (OpExpG b (UnpreparedValue b)))]