Hasura.GraphQL.Schema.BoolExp.AggregationPredicates

Description

This module defines the schema aspect of the default implementation of aggregation predicates.

Documentation

defaultAggregationPredicatesParser :: forall b r m n. (MonadBuildSchema b r m n, AggregationPredicatesSchema b) => [FunctionSignature b] -> SourceInfo b -> TableInfo b -> SchemaT r m (Maybe (InputFieldsParser n [AggregationPredicatesImplementation b (UnpreparedValue b)]))

This function is meant to serve as the default schema for Aggregation Predicates represented in the IR by the type AggregationPredicates.

data FunctionSignature (b :: BackendType)

Constructors

FunctionSignature

Fields

fnName :: Text

fnGQLName :: Name

fnArguments :: ArgumentsSignature b

fnReturnType :: ScalarType b

data ArgumentsSignature (b :: BackendType)

Constructors

ArgumentsStar

SingleArgument (ScalarType b)

Arguments (NonEmpty (ArgumentSignature b))

data ArgumentSignature (b :: BackendType)

Constructors

ArgumentSignature

Fields

argType :: ScalarType b

argName :: Name