Hasura.GraphQL.Schema.Common

Documentation

data SchemaContext

Aggregation of contextual information required to build the schema.

Constructors

SchemaContext

Fields

scSchemaKind :: SchemaKind

the kind of schema being built

scRemoteRelationshipParserBuilder :: RemoteRelationshipParserBuilder

how to process remote relationships

scRole :: RoleName

the role for which the schema is being built

data SchemaKind

The kind of schema we're building, and its associated options.

Constructors

HasuraSchema

RelaySchema NodeInterfaceParserBuilder

isHasuraSchema :: SchemaKind -> Bool

type MonadBuildSchemaBase r m n = (MonadError QErr m, MonadMemoize m, MonadParse n, Has SchemaContext r)

The set of common constraints required to build the schema.

newtype RemoteRelationshipParserBuilder

How a remote relationship field should be processed when building a schema. Injecting this function from the top level avoids having to know how to do top-level dispatch from deep within the schema code.

Note: the inner function type uses an existential qualifier: it is expected that the given function will work for _any_ monad m that has the relevant constraints. This prevents us from passing a function that is specfic to the monad in which the schema construction will run, but avoids having to propagate type annotations to each call site.

Constructors

RemoteRelationshipParserBuilder (forall lhsJoinField r n m. MonadBuildSchemaBase r m n => RemoteFieldInfo lhsJoinField -> SchemaT r m (Maybe [FieldParser n (RemoteRelationshipField UnpreparedValue)]))

ignoreRemoteRelationship :: RemoteRelationshipParserBuilder

A RemoteRelationshipParserBuilder that ignores the field altogether, that can be used in tests or to build a source or remote schema in isolation.

newtype NodeInterfaceParserBuilder

How to build the Relay node.

Similarly to what we do for remote relationships, we pass in the context the builder function required to build the Node interface, in order to avoid the cross-sources cycles it creates otherwise.

Constructors

NodeInterfaceParserBuilder

Fields

runNodeBuilder :: forall r m n. MonadBuildSourceSchema r m n => SchemaT r m (Parser 'Output n NodeMap)

retrieve :: (MonadReader r m, Has a r) => (a -> b) -> m b

newtype SchemaT r m a

The monad in which the schema is built.

The implementation of SchemaT is intended to be opaque: running a computation in SchemaT is intended to be done via calls to runSourceSchema and runRemoteSchema, which also enforce what the r parameter should be in each case.

The reason why we want to enforce that the schema is built in a reader on top of an arbitrary base monad is for performance: see Note [SchemaT and stacking] for more information.

In the future, we might monomorphize this further to make MemoizeT explicit.

Constructors

SchemaT

Fields

runSchemaT :: ReaderT r m a

type MonadBuildSourceSchema r m n = (MonadBuildSchemaBase r m n, Has SchemaOptions r, Has MkTypename r, Has NamingCase r)

runSourceSchema :: SchemaContext -> SchemaOptions -> SchemaT (SchemaContext, SchemaOptions, MkTypename, NamingCase) m a -> m a

Runs a schema-building computation with all the context required to build a source.

type MonadBuildRemoteSchema r m n = (MonadBuildSchemaBase r m n, Has CustomizeRemoteFieldName r, Has MkTypename r)

runRemoteSchema :: SchemaContext -> SchemaT (SchemaContext, MkTypename, CustomizeRemoteFieldName) m a -> m a

Runs a schema-building computation with all the context required to build a remote schema.

type SelectExp b = AnnSimpleSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type StreamSelectExp b = AnnSimpleStreamSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AggSelectExp b = AnnAggregateSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type ConnectionSelectExp b = ConnectionSelect b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type SelectArgs b = SelectArgsG b (UnpreparedValue b)

type SelectStreamArgs b = SelectStreamArgsG b (UnpreparedValue b)

type TablePerms b = TablePermG b (UnpreparedValue b)

type AnnotatedFields b = AnnFieldsG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AnnotatedField b = AnnFieldG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type ConnectionFields b = ConnectionFields b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type EdgeFields b = EdgeFields b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AnnotatedActionFields = ActionFieldsG (RemoteRelationshipField UnpreparedValue)

type AnnotatedActionField = ActionFieldG (RemoteRelationshipField UnpreparedValue)

data RemoteSchemaParser n

Constructors

RemoteSchemaParser

Fields

piQuery :: [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

piMutation :: Maybe [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

piSubscription :: Maybe [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

getTableRoles :: BackendSourceInfo -> [RoleName]

askTableInfo :: forall b m. (Backend b, MonadError QErr m) => SourceInfo b -> TableName b -> m (TableInfo b)

Looks up table information for the given table name. This function should never fail, since the schema cache construction process is supposed to ensure all dependencies are resolved. TODO: deduplicate this with CacheRM.

data Scenario

Whether the request is sent with `x-hasura-use-backend-only-permissions` set to true.

Constructors

Backend

Frontend

textToName :: MonadError QErr m => Text -> m Name

partialSQLExpToUnpreparedValue :: PartialSQLExp b -> UnpreparedValue b

mapField :: Functor m => InputFieldsParser m (Maybe a) -> (a -> b) -> InputFieldsParser m (Maybe b)

parsedSelectionsToFields

Arguments

:: (Text -> a)
how to handle __typename fields

-> InsOrdHashMap Name (ParsedSelection a)

-> Fields a

numericAggOperators :: [Name]

comparisonAggOperators :: [Name]

mkDescriptionWith :: Maybe PGDescription -> Text -> Description

takeValidTables :: forall b. Backend b => TableCache b -> TableCache b

takeValidFunctions :: forall b. FunctionCache b -> FunctionCache b

requiredFieldParser :: (Functor n, Functor m) => (a -> b) -> m (FieldParser n a) -> m (Maybe (FieldParser n b))

optionalFieldParser :: (Functor n, Functor m) => (a -> b) -> m (Maybe (FieldParser n a)) -> m (Maybe (FieldParser n b))

mkEnumTypeName :: forall b m r. (Backend b, MonadReader r m, Has MkTypename r, MonadError QErr m, Has NamingCase r) => EnumReference b -> m Name

Builds the type name for referenced enum tables.

addEnumSuffix :: (MonadReader r m, Has MkTypename r) => GQLNameIdentifier -> Maybe Name -> NamingCase -> m Name

peelWithOrigin :: MonadParse m => Parser 'Both m a -> Parser 'Both m (ValueWithOrigin a)