Hasura.GraphQL.Schema.NamingCase

Documentation

isGraphqlCase :: NamingCase -> Bool

hasNamingConventionChanged :: (HashSet ExperimentalFeature, NamingCase) -> (HashSet ExperimentalFeature, NamingCase) -> Bool

Check if naming convention has changed The value of naming convention depends on whether the naming convention is enabled in experimental features and what the default naming convention (HASURA_GRAPHQL_DEFAULT_NAMING_CONVENTION) is hence use both these values to decide if naming convention has changed