Hasura.GraphQL.Schema.Node

Description

A collection of types and utilities around the Node GraphQL type exposed by the Relay API.

Documentation

data NodeId

Constructors

NodeIdV1 V1NodeId

NodeIdV2 (AnyBackend V2NodeId)

data V1NodeId

V1 format of a node.

This id does NOT uniquely identify the table properly, as it only knows the table's name, but doesn't store a source name.

Constructors

V1NodeId

Fields

_ni1Table :: QualifiedTable

_ni1Columns :: NESeq Value

data V2NodeId b

V2 format of a node.

Uniquely identifies a table with source name and table name, and uniquely identifies a row within that table with a list of primary key values.

Constructors

V2NodeId

Fields

_ni2Source :: SourceName

_ni2Table :: TableName b

_ni2Columns :: NESeq Value

parseNodeIdV1 :: [Value] -> Parser V1NodeId

data NodeIdVersion

Enum representing the supported versions of the API.

Constructors

NIVersion1

NIVersion2

nodeIdVersionInt :: NodeIdVersion -> Int

currentNodeIdVersion :: NodeIdVersion

type NodeMap = HashMap SourceName (AnyBackend TableMap)

data NodeInfo b

All the information required to craft a query to a row pointed to by a NodeId.

Constructors

NodeInfo

Fields

nvSourceConfig :: SourceConfig b

nvSelectPermissions :: SelPermInfo b

nvPrimaryKeys :: PrimaryKeyColumns b

nvAnnotatedFields :: AnnFieldsG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

newtype TableMap b

Constructors

TableMap (HashMap (TableName b) (NodeInfo b))

findNode :: forall b. Backend b => SourceName -> TableName b -> NodeMap -> Maybe (NodeInfo b)

Given a source name and table name, peform the double lookup within a NodeMap.