Hasura.GraphQL.Schema.Options

Description

Definitions for schema building configuration.

Documentation

data SchemaOptions

The record of options required to build the schema. This includes options that can be enabled with ExperimentalFeature flags.

Constructors

SchemaOptions

Fields

soStringifyNumbers :: StringifyNumbers

soDangerousBooleanCollapse :: DangerouslyCollapseBooleans

soInferFunctionPermissions :: InferFunctionPermissions

soOptimizePermissionFilters :: OptimizePermissionFilters

data StringifyNumbers

Should we represent numbers in our responses as numbers, or strings? Some backends can return numbers that exceed the bounds of JSON's own number type. In these cases, we can use stringified versions of these numbers to avoid this problem.

Constructors

StringifyNumbers

Don'tStringifyNumbers

data DangerouslyCollapseBooleans

Should Boolean fields be collapsed to True when a null value is given? This was the behaviour of Hasura V1, and is now discouraged.

Constructors

DangerouslyCollapseBooleans

Don'tDangerouslyCollapseBooleans

data InferFunctionPermissions

Should we infer function permissions? If this flag is set to InferFunctionPermissions, we may fail to build expression parsers in buildQueryAndSubscriptionFields for users with unrecognised roles.

Constructors

InferFunctionPermissions

Don'tInferFunctionPermissions

data RemoteSchemaPermissions

Should we enable remote schema permissions? If so, these permissions will influence buildGQLContext.

Constructors

EnableRemoteSchemaPermissions

DisableRemoteSchemaPermissions

data OptimizePermissionFilters

Should we attempt to deduplicate permission filters? This flag is used in relationshipField to determine whether certain row-level permission filters can be dropped in certain cases.

Constructors

OptimizePermissionFilters

Don'tOptimizePermissionFilters