Hasura.GraphQL.Schema.OrderBy

Documentation

tableOrderByExp :: forall (b :: BackendType) r (m :: Type -> Type) (n :: Type -> Type). MonadBuildSchema b r m n => TableInfo b -> SchemaT r m (Parser 'Input n [AnnotatedOrderByItemG b (UnpreparedValue b)])

Corresponds to an object type for an order by.

input table_order_by {
  col1: order_by
  col2: order_by
  .     .
  .     .
  coln: order_by
  obj-rel: <remote-table>_order_by
}

logicalModelOrderByExp :: forall (b :: BackendType) r (m :: Type -> Type) (n :: Type -> Type). MonadBuildSchema b r m n => LogicalModelInfo b -> SchemaT r m (Parser 'Input n [AnnotatedOrderByItemG b (UnpreparedValue b)])

Corresponds to an object type for an order by.

input table_order_by {
  col1: order_by
  col2: order_by
  .     .
  .     .
  coln: order_by
  obj-rel: <remote-table>_order_by
}

TODO: When there are no columns accessible to a role, the `table_order_by` will be an empty input object. In such a case, we can avoid exposing the order_by argument.