Hasura.GraphQL.Schema.RemoteRelationship

Documentation

remoteRelationshipField :: SourceCache -> RemoteSchemaMap -> RemoteRelationshipParserBuilder

Remote relationship field parsers

remoteRelationshipToSchemaField :: forall r m n lhsJoinField. MonadBuildSchemaBase r m n => RemoteSchemaMap -> HashMap FieldName lhsJoinField -> RemoteSchemaFieldInfo -> m (Maybe (FieldParser n (RemoteSchemaSelect (RemoteRelationshipField UnpreparedValue))))

Parser(s) for remote relationship fields to a remote schema

lookupNestedFieldType' :: (MonadSchema n m, MonadError QErr m) => Name -> RemoteSchemaIntrospection -> FieldCall -> m GType

lookupNestedFieldType :: (MonadSchema n m, MonadError QErr m) => Name -> RemoteSchemaIntrospection -> NonEmpty FieldCall -> m GType

remoteRelationshipToSourceField :: forall r m n tgt. (MonadBuildSchemaBase r m n, BackendSchema tgt) => SourceCache -> RemoteSourceFieldInfo tgt -> m [FieldParser n (RemoteSourceSelect (RemoteRelationshipField UnpreparedValue) UnpreparedValue tgt)]

Parser(s) for remote relationship fields to a database table. Note that when the target is a database table, an array relationship declaration would have the _aggregate field in addition to the array relationship field, hence [FieldParser ...] instead of FieldParser