Hasura.GraphQL.Schema.Update.Batch

Documentation

updateTable

Arguments

:: forall b r m n. (MonadBuildSchema b r m n, AggregationPredicatesSchema b, BackendTableSelectSchema b, BackendUpdateOperatorsSchema b)

=> (UpdateBatch b (UpdateOperators b) (UnpreparedValue b) -> UpdateVariant b (UnpreparedValue b))

-> Scenario

-> TableInfo b

-> GQLNameIdentifier
table field display name

-> SchemaT r m (Maybe (FieldParser n (AnnotatedUpdateG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b))))

Construct a root field, normally called update_tablename, that can be used to update rows in a DB table specified by filters. Only returns a parser if there are columns the user is allowed to update; otherwise returns Nothing.

updateTableMany

Arguments

:: forall b r m n. (MonadBuildSchema b r m n, AggregationPredicatesSchema b, BackendTableSelectSchema b, BackendUpdateOperatorsSchema b)

=> ([UpdateBatch b (UpdateOperators b) (UnpreparedValue b)] -> UpdateVariant b (UnpreparedValue b))

-> Scenario

-> TableInfo b

-> GQLNameIdentifier
table field display name

-> SchemaT r m (Maybe (FieldParser n (AnnotatedUpdateG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b))))

updateTableByPk

Arguments

:: forall b r m n. (MonadBuildSchema b r m n, BackendTableSelectSchema b, BackendUpdateOperatorsSchema b)

=> (UpdateBatch b (UpdateOperators b) (UnpreparedValue b) -> UpdateVariant b (UnpreparedValue b))

-> Scenario

-> TableInfo b

-> GQLNameIdentifier
table field display name

-> SchemaT r m (Maybe (FieldParser n (AnnotatedUpdateG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b))))

Construct a root field, normally called update_tablename_by_pk, that can be used to update a single in a DB table, specified by primary key. Only returns a parser if there are columns the user is allowed to update and if the user has select permissions on all primary keys; otherwise returns Nothing.