Hasura.GraphQL.Transport.Backend

Documentation

class BackendExecute b => BackendTransport (b :: BackendType) where

This typeclass enacapsulates how a given backend sends queries and mutations over the network. Each backend is currently responsible for both logging and tracing, for now.

Methods

runDBQuery :: forall m. (MonadIO m, MonadError QErr m, MonadQueryLog m, MonadTrace m) => RequestId -> GQLReqUnparsed -> RootFieldAlias -> UserInfo -> Logger Hasura -> SourceConfig b -> ExecutionMonad b EncJSON -> Maybe (PreparedQuery b) -> m (DiffTime, EncJSON)

runDBMutation :: forall m. (MonadIO m, MonadError QErr m, MonadQueryLog m, MonadTrace m) => RequestId -> GQLReqUnparsed -> RootFieldAlias -> UserInfo -> Logger Hasura -> SourceConfig b -> ExecutionMonad b EncJSON -> Maybe (PreparedQuery b) -> m (DiffTime, EncJSON)

runDBSubscription

Arguments

:: forall m. MonadIO m

=> SourceConfig b

-> MultiplexedQuery b

-> [(CohortId, CohortVariables)]
WARNING: Postgres-specific, ignored by other backends

-> m (DiffTime, Either QErr [(CohortId, ByteString)])

runDBStreamingSubscription

Arguments

:: forall m. MonadIO m

=> SourceConfig b

-> MultiplexedQuery b

-> [(CohortId, CohortVariables)]
WARNING: Postgres-specific, ignored by other backends

-> m (DiffTime, Either QErr [(CohortId, ByteString, CursorVariableValues)])

runDBQueryExplain :: forall m. (MonadIO m, MonadError QErr m) => DBStepInfo b -> m EncJSON