Hasura.GraphQL.Transport.HTTP.Protocol

Documentation

newtype GQLExecDoc

Constructors

GQLExecDoc

Fields

unGQLExecDoc :: [ExecutableDefinition Name]

newtype OperationName

Constructors

OperationName

Fields

_unOperationName :: Name

type VariableValues = HashMap Name Value

data GQLReq a

https://graphql.org/learn/serving-over-http/#post-request

See GQLReqParsed for invariants.

Constructors

GQLReq

Fields

_grOperationName :: !(Maybe OperationName)

_grQuery :: !a

_grVariables :: !(Maybe VariableValues)

data GQLBatchedReqs a

Batched queries are sent as a JSON array of GQLReq records. This newtype exists to support the unusual JSON encoding.

See https://github.com/hasura/graphql-engine/issues/1812.

Constructors

GQLSingleRequest a

GQLBatchedReqs [a]

newtype GQLQueryText

Constructors

GQLQueryText

Fields

_unGQLQueryText :: Text

type GQLReqUnparsed = GQLReq GQLQueryText

We've not yet parsed the graphql query string parameter of the POST.

type GQLReqParsed = GQLReq GQLExecDoc

Invariants:

when _grOperationName is Nothing, _grQuery contains exactly one ExecutableDefinitionOperation (and zero or more ExecutableDefinitionFragment)

when _grOperationName is present, there is a corresponding ExecutableDefinitionOperation in _grQuery

type ReqsText = GQLBatchedReqs (GQLReq GQLQueryText)

type GQLReqOutgoing = GQLReq SingleOperation

A simplified form of GQLReqParsed which is more ergonomic in particular for APIs that act as graphql clients (e.g. in remote relationship execution). This is a "desugared" request in which fragments have been inlined (see inlineSelectionSet), and the operation (_grOperationName) to be executed is the only payload (in contrast to a ExecutableDocument with possibly many named operations).

_grOperationName is essentially ignored here, but should correspond with _todName if present.

These could maybe benefit from an HKD refactoring.

type SingleOperation = TypedOperationDefinition NoFragments Name

A single graphql operation to be executed, with fragment definitions inlined. This is the simplified form of GQLExecDoc or ExecutableDocument:

renderGQLReqOutgoing :: GQLReqOutgoing -> GQLReqUnparsed

getSingleOperation :: MonadError QErr m => GQLReqParsed -> m SingleOperation

Obtain the actual single operation to be executed, from the possibly- multi-operation document, validating per the spec and inlining any fragment definitions (pre-defined parts of a graphql query) at fragment spreads (locations where fragments are "spliced"). See:

https://spec.graphql.org/June2018/#sec-Executable-Definitions and... https://graphql.org/learn/serving-over-http/

toParsed :: MonadError QErr m => GQLReqUnparsed -> m GQLReqParsed

encodeGQErr :: Bool -> QErr -> Value

type GQResult a = Either GQExecError a

newtype GQExecError

Constructors

GQExecError [Value]

type GQResponse = GQResult ByteString

isExecError :: GQResult a -> Bool

encodeGQResp :: GQResponse -> EncJSON

decodeGQResp :: EncJSON -> (Maybe GQResponse, EncJSON)

encodeHTTPResp :: GQResponse -> EncJSON