Hasura.GraphQL.Transport.WebSocket.Protocol

Description

This file contains types for both the websocket protocols (Apollo) and (graphql-ws) | See Apollo: https://github.com/apollographql/subscriptions-transport-ws/blob/master/PROTOCOL.md | See graphql-ws: https://github.com/enisdenjo/graphql-ws/blob/master/PROTOCOL.md

Documentation

data WSSubProtocol

Constructors

Apollo

GraphQLWS

showSubProtocol :: WSSubProtocol -> String

toWSSubProtocol :: String -> WSSubProtocol

newtype OperationId

Constructors

OperationId

Fields

unOperationId :: Text

unsafeMkOperationId :: Text -> OperationId

data ServerMsgType

Constructors

SMT_GQL_CONNECTION_KEEP_ALIVE

SMT_GQL_CONNECTION_ERROR

SMT_GQL_DATA

SMT_GQL_NEXT

SMT_GQL_PING

SMT_GQL_PONG

SMT_GQL_CONNECTION_ACK

SMT_GQL_ERROR

SMT_GQL_COMPLETE

data ConnParams

Constructors

ConnParams

Fields

_cpHeaders :: Maybe (HashMap Text Text)

data StartMsg

Constructors

StartMsg

Fields

_smId :: !OperationId

_smPayload :: !GQLReqUnparsed

data StopMsg

Constructors

StopMsg

Fields

_stId :: OperationId

data PingPongPayload

Constructors

PingPongPayload

Fields

_smMessage :: !(Maybe Text)

keepAliveMessage :: PingPongPayload

data SubscribeMsg

Constructors

SubscribeMsg

Fields

_subId :: !OperationId

_subPayload :: !GQLReqUnparsed

data ClientMsg

Constructors

CMConnInit !(Maybe ConnParams)

CMStart !StartMsg

CMStop !StopMsg

CMConnTerm

CMPing !(Maybe PingPongPayload)

CMPong !(Maybe PingPongPayload)

data DataMsg

Constructors

DataMsg

Fields

_dmId :: !OperationId

_dmPayload :: !GQResponse

data ErrorMsg

Constructors

ErrorMsg

Fields

_emId :: !OperationId

_emPayload :: !Value

newtype CompletionMsg

Constructors

CompletionMsg

Fields

unCompletionMsg :: OperationId

newtype ConnErrMsg

Constructors

ConnErrMsg

Fields

unConnErrMsg :: Text

data ServerErrorMsg

Constructors

ServerErrorMsg

Fields

unServerErrorMsg :: Text

data ServerMsg

Constructors

SMConnAck

SMConnKeepAlive

SMConnErr !ConnErrMsg

SMData !DataMsg

SMErr !ErrorMsg

SMComplete !CompletionMsg

SMNext !DataMsg

SMPing !(Maybe PingPongPayload)

SMPong !(Maybe PingPongPayload)

data ServerErrorCode

This is sent from the server to the client while closing the websocket on encountering an error.

Constructors

ProtocolError1002

GenericError4400 !String

Unauthorized4401

Forbidden4403

ConnectionInitTimeout4408

NonUniqueSubscription4409 !OperationId

TooManyRequests4429

encodeServerErrorMsg :: ServerErrorCode -> ByteString

encodeServerMsg :: ServerMsg -> ByteString

data WSConnInitTimerStatus

Constructors

Running

Done

type WSConnInitTimer = (TVar WSConnInitTimerStatus, TMVar ())

getWSTimerState :: WSConnInitTimer -> IO WSConnInitTimerStatus

getNewWSTimer :: Seconds -> IO WSConnInitTimer