Hasura.Incremental

Description

A simple implementation of incremental build rules, which can be used to avoid unnecessary recomputation on incrementally-changing input. See Rule for more details.

Documentation

newtype InvalidationKey

A simple helper type that can be used to implement explicit cache invalidation. Internally, each InvalidationKey is a counter; initialInvalidationKey starts the counter at 0 and invalidate increments it by 1. Two InvalidationKeys are equal iff they have the same internal count, so depending on an InvalidationKey provides a mechanism to force portions of the build process to be reexecuted by calling invalidate before running the build.

Constructors

InvalidationKey Int

initialInvalidationKey :: InvalidationKey

invalidate :: InvalidationKey -> InvalidationKey