Hasura.Incremental.Internal.Cache

Documentation

class ArrowKleisli m arr => ArrowCache m arr | arr -> m where

Methods

cache :: Cacheable a => arr a b -> arr a b

Adds equality-based caching to the given arrow. After each execution of the arrow, its input and result values are cached. On the next execution, the new input value is compared via == to the previous input value. If they are the same, the previous result is returned without re-executing the arrow. Otherwise, the old cached values are discarded, and the arrow is re-executed to produce a new set of cached values.

Indescriminate use of cache is likely to have little effect except to increase memory usage, since the input and result of each execution must be retained in memory. Avoid using cache around arrows with large input or output that is likely to change often unless profiling indicates it is computationally expensive enough to be worth the memory overhead.

Note that only direct inputs and outputs of the given arrow are cached. If an arrow provides access to values through a side-channel, they will not participate in caching.

newDependency :: arr a (Dependency a)

Creates a new Dependency, which allows fine-grained caching of composite values; see the documentation for Dependency for more details.

dependOn :: Cacheable a => arr (Dependency a) a

Extract the value from a Dependency, incurring a dependency on its entirety. To depend on only a portion of the value, use selectD or selectKeyD before passing it to dependOn.

bindDepend :: arr (DependT m a) a

Run a monadic sub-computation with the ability to access dependencies; see MonadDepend for more details.

class Monad m => MonadDepend m where

A restricted, monadic variant of ArrowCache that can only read dependencies, not create new ones or add local caching. This serves as a limited adapter between arrow and monadic code.

Methods

dependOnM :: Cacheable a => Dependency a -> m a

newtype DependT m a

Constructors

DependT

Fields

unDependT :: StateT Accesses m a