Hasura.Incremental.Internal.Dependency

Description

Supporting functionality for fine-grained dependency tracking.

Documentation

data Dependency a

A Dependency represents a value that a Rule can conditionally depend on. A Dependency is created using newDependency, and it can be “opened” again using dependOn. What makes a Dependency useful is the way it cooperates with cache---if a Dependency is passed to a cached rule, but that rule (or any of its sub-rules) never “opens” it using dependOn, then subsequent executions of the rule will ignore the Dependency when computing whether or not it is necessary to re-execute the rule.

The above functionality is useful on its own to express conditional dependencies, but even more useful is the ability to express partial dependencies. For example, if a Dependency contains a HashMap, a rule can choose to only depend on the value associated with a particular key by using selectKeyD (or the more general selectD). Only the parts that are actually used will be counted when computing whether a rule needs to be re-executed.

Constructors

Dependency (DependencyKey a) a

selectD :: Select a => Selector a b -> Dependency a -> Dependency b

Applies a Selector to select part of a Dependency.

selectKeyD :: (Select a, Selector a ~ ConstS k v) => k -> Dependency a -> Dependency v

Selects a single key from a dependency containing a map-like data structure.

selectMaybeD :: Monoid a => Dependency (Maybe a) -> Dependency a

data DependencyKey a where

Tracks whether a Dependency is a “root” dependency created by newDependency or a “child” dependency created from an existing dependency using selectD.

Constructors

DependencyRoot :: UniqueS a -> DependencyKey a

data Accesses

A mapping from root Dependency keys to the accesses made against those dependencies.

recordAccess :: DependencyKey a -> Access a -> Accesses -> Accesses

unchanged :: (Given Accesses => Eq a) => Accesses -> a -> a -> Bool

data Access a where

Records the accesses made within a single Dependency and its children. The Semigroup instance for Access computes a least upper bound:

AccessedAll serves as the top of the lattice and records the dependency’s entire value was accessed.

AccessedParts records a set of accesses for individual parts of a dependency.

Constructors

AccessedAll :: Eq a => Access a