Hasura.Incremental.Select

Documentation

class GCompare (Selector a) => Select a where

The Select class provides a way to access subparts of a product type using a reified Selector. A Selector a b is essentially a function from a to b, and indeed select converts a Selector to such a function. However, unlike functions, Selectors can be compared for equality using GEq and ordered using GCompare.

This is useful to implement dependency tracking, since it’s possible to track in a reified form exactly which parts of a data structure are used.

Instances of Select can be automatically derived for record types (just define an empty instance). The instance uses the magical HasField constraints, and Selectors for the type can be written using OverloadedLabels.

Minimal complete definition

Nothing

Associated Types

type Selector a :: Type -> Type

type Selector r = FieldS r

Methods

select :: Selector a b -> a -> b

default select :: Selector a ~ FieldS a => Selector a b -> a -> b

data ConstS k a b where

The constant selector, which is useful for representing selectors into data structures where all fields have the same type. Matching on a value of type ConstS k a b causes a and b to unify, effectively “pinning” b to a.

Constructors

ConstS :: !k -> ConstS k a a

selectKey :: (Select a, Selector a ~ ConstS k v) => k -> a -> v

data FieldS r a where

Constructors

FieldS :: (KnownSymbol s, HasField s r a) => !(Proxy s) -> FieldS r a

newtype UniqueS a

A UniqueS is, as the name implies, a globally-unique Selector, which can be created using newUniqueS. If a value of type UniqueS a is found to be equal (via geq) with another value of type UniqueS b, then a and b must be the same type. This effectively allows the creation of a dynamically-extensible sum type, where new constructors can be created at runtime using newUniqueS.

Constructors

UniqueS Unique

newUniqueS :: MonadUnique m => m (UniqueS a)

data DMapS k f a where

Constructors

DMapS :: !(k a) -> DMapS k f (Maybe (f a))