Hasura.Logging

Documentation

class (Eq (EngineLogType impl), Hashable (EngineLogType impl)) => EnabledLogTypes impl where

Typeclass representing any type which can be parsed into a list of enabled log types, and has a Set of default enabled log types, and can find out if a log type is enabled

Methods

parseEnabledLogTypes :: String -> Either String [EngineLogType impl]

defaultEnabledLogTypes :: HashSet (EngineLogType impl)

isLogTypeEnabled :: HashSet (EngineLogType impl) -> EngineLogType impl -> Bool

data family EngineLogType impl

A family of EngineLogType types

data Hasura

data InternalLogTypes

Constructors

ILTUnstructured
mostly for debug logs - see debugT, debugBS and debugLBS functions

ILTUnhandledInternalError

ILTEventTrigger

ILTEventTriggerProcess

ILTScheduledTrigger

ILTScheduledTriggerProcess

ILTCronEventGeneratorProcess

ILTWsServer
internal logs for the websocket server

ILTPgClient

ILTMetadata
log type for logging metadata related actions; currently used in logging inconsistent metadata

ILTTelemetry

ILTSchemaSync

ILTSourceCatalogMigration

ILTStoredIntrospection

ILTStoredIntrospectionStorage

defaultEnabledEngineLogTypes :: HashSet (EngineLogType Hasura)

isEngineLogTypeEnabled :: HashSet (EngineLogType Hasura) -> EngineLogType Hasura -> Bool

readLogTypes :: String -> Either String [EngineLogType Hasura]

userAllowedLogTypes :: [EngineLogType Hasura]

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

data EngineLog impl

Constructors

EngineLog

Fields

_elTimestamp :: !FormattedTime

_elLevel :: !LogLevel

_elType :: !(EngineLogType impl)

_elDetail :: !Value

class EnabledLogTypes impl => ToEngineLog a impl where

Typeclass representing any data type that can be converted to EngineLog for the purpose of logging

Methods

toEngineLog :: a -> (LogLevel, EngineLogType impl, Value)

data UnstructuredLog

Constructors

UnstructuredLog

Fields

_ulLevel :: !LogLevel

_ulPayload :: !SerializableBlob

debugT :: Text -> UnstructuredLog

debugBS :: ByteString -> UnstructuredLog

debugLBS :: ByteString -> UnstructuredLog

data LoggerCtx impl

Constructors

LoggerCtx

Fields

_lcLoggerSet :: !LoggerSet

_lcLogLevel :: !LogLevel

_lcTimeGetter :: !(IO FormattedTime)

_lcEnabledLogTypes :: !(HashSet (EngineLogType impl))

Unhandled Internal Errors

newtype UnhandledInternalErrorLog

We expect situations where there are code paths that should not occur and we throw an error on this code paths. If our assumptions are incorrect and infact these errors do occur, we want to log them.

Constructors

UnhandledInternalErrorLog ErrorCall

LoggerSettings

data LoggerSettings

Constructors

LoggerSettings

Fields

_lsCachedTimestamp :: !Bool

should current time be cached (refreshed every sec)

_lsTimeZone :: !(Maybe TimeZone)

_lsLevel :: !LogLevel

defaultLoggerSettings :: Bool -> LogLevel -> LoggerSettings

getFormattedTime :: Maybe TimeZone -> IO FormattedTime

mkLoggerCtx :: (MonadIO io, MonadBaseControl IO io) => LoggerSettings -> HashSet (EngineLogType impl) -> ManagedT io (LoggerCtx impl)

Creates a new LoggerCtx.

The underlying LoggerSet is bound to the ManagedT context: when it exits, the log will be flushed and cleared regardless of whether it was exited properly or not (ManagedT uses bracket underneath). This guarantees that the logs will always be flushed, even in case of error, avoiding a repeat of https://github.com/hasura/graphql-engine/issues/4772.

cleanLoggerCtx :: LoggerCtx a -> IO ()

newtype Logger impl

Constructors

Logger

Fields

unLogger :: forall a m. (ToEngineLog a impl, MonadIO m) => a -> m ()

mkLogger :: ToJSON (EngineLogType impl) => LoggerCtx impl -> Logger impl

nullLogger :: Logger Hasura

eventTriggerLogType :: EngineLogType Hasura

eventTriggerProcessLogType :: EngineLogType Hasura

scheduledTriggerLogType :: EngineLogType Hasura

scheduledTriggerProcessLogType :: EngineLogType Hasura

cronEventGeneratorProcessType :: EngineLogType Hasura

sourceCatalogMigrationLogType :: EngineLogType Hasura

data StoredIntrospectionLog

Emit when stored introspection is used

Constructors

StoredIntrospectionLog

Fields

silMessage :: Text

silSourceError :: QErr

upstream data source errors

data StoredIntrospectionStorageLog

Logs related to errors while interacting with the stored introspection storage

Constructors

StoredIntrospectionStorageLog

Fields

sislMessage :: Text

sislError :: QErr

createStatsLogger :: forall m stats impl. (MonadIO m, ToEngineLog stats impl, Monoid stats) => Logger impl -> m (Trigger stats stats)

A logger useful for accumulating and logging stats, in tight polling loops. It also debounces to not flood with excessive logs. Use @logStats to record statistics for logging.

closeStatsLogger :: (MonadIO m, EnabledLogTypes impl) => EngineLogType impl -> Logger impl -> Trigger stats stats -> m ()

Safely close the statistics logger. When occurred, exception is logged.

logStats :: MonadIO m => Trigger stats stats -> stats -> m ()

This won't log the given stats immediately. The stats are accumulated over the specific timeframe and logged only once. See @createStatsLogger for more details.