Hasura.LogicalModel.IR

Description

The RQL IR representation of an invocation of a logical model.

Documentation

data LogicalModel b field

The RQL IR representation of an invocation of a logical model.

Constructors

LogicalModel

Fields

lmRootFieldName :: LogicalModelName

The graphql name of the logical model.

lmInterpolatedQuery :: InterpolatedQuery field

The raw sql to use in the query

lmArgs :: HashMap LogicalModelArgumentName (ColumnValue b)

The arguments passed to the query, if any.

lmReturnType :: CustomReturnType b

The return type of the logical model