Hasura.LogicalModel.Metadata

Description

Metadata representation of a logical model in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

data InterpolatedItem variable

A component of an interpolated query

Constructors

IIText Text
normal text

IIVariable variable
a captured variable

newtype InterpolatedQuery variable

A list of logical model components representing a single logical model, separating the variables from the text.

Constructors

InterpolatedQuery

Fields

getInterpolatedQuery :: [InterpolatedItem variable]

newtype LogicalModelArgumentName

Constructors

LogicalModelArgumentName

Fields

getLogicalModelArgumentName :: Text

data LogicalModelMetadata (b :: BackendType)

The representation of logical models within the metadata structure.

Constructors

LogicalModelMetadata

Fields

_lmmRootFieldName :: LogicalModelName

_lmmCode :: InterpolatedQuery LogicalModelArgumentName

_lmmReturns :: CustomReturnType b

_lmmArguments :: HashMap LogicalModelArgumentName (ScalarType b)

_lmmSelectPermissions :: InsOrdHashMap RoleName (SelPermDef b)

_lmmDescription :: Maybe Text

data WithLogicalModel a

A wrapper to tie something to a particular logical model. Specifically, it assumes the underlying _wlmInfo is represented as an object, and adds two keys to that object: source and root_field_name.

Constructors

WithLogicalModel

Fields

_wlmSource :: SourceName

_wlmName :: LogicalModelName

_wlmInfo :: a

parseInterpolatedQuery :: Text -> Either Text (InterpolatedQuery LogicalModelArgumentName)

extract all of the `{{ variable }}` inside our query string

lmmSelectPermissions :: forall b. Lens' (LogicalModelMetadata b) (InsOrdHashMap RoleName (SelPermDef b))

lmmRootFieldName :: forall b. Lens' (LogicalModelMetadata b) LogicalModelName

lmmReturns :: forall b. Lens' (LogicalModelMetadata b) (CustomReturnType b)

lmmDescription :: forall b. Lens' (LogicalModelMetadata b) (Maybe Text)

lmmCode :: forall b. Lens' (LogicalModelMetadata b) (InterpolatedQuery LogicalModelArgumentName)

lmmArguments :: forall b. Lens' (LogicalModelMetadata b) (HashMap LogicalModelArgumentName (ScalarType b))