Hasura.LogicalModel.NullableScalarType

Description

A scalar type to be used for logical models and resolvers.

Documentation

data NullableScalarType b

A ScalarType that can be nullable with an optional description

Constructors

NullableScalarType

Fields

nstType :: ScalarType b

nstNullable :: Bool

nstDescription :: Maybe Text

nullableScalarTypeMapCodec :: forall b. Backend b => Codec Value (InsOrdHashMap (Column b) (NullableScalarType b)) (InsOrdHashMap (Column b) (NullableScalarType b))