Hasura.LogicalModel.Types

Description

A name for a logical model as it is recognized by the graphql schema.

Documentation

newtype LogicalModelName

Constructors

LogicalModelName

Fields

getLogicalModelName :: Name

data LogicalModelField b

either a scalar type or a reference to another logical model

Constructors

LogicalModelScalarField

Fields

lmfName :: Column b

lmfType :: ScalarType b

lmfNullable :: Bool

lmfDescription :: Maybe Text

LogicalModelArrayReference

Fields

lmfName :: Column b

lmfLogicalModel :: LogicalModelName

logicalModelFieldMapCodec :: forall b. Backend b => Codec Value (InsOrdHashMap (Column b) (LogicalModelField b)) (InsOrdHashMap (Column b) (LogicalModelField b))