Hasura.LogicalModelResolver.Metadata

Documentation

data LogicalModelIdentifier (b :: BackendType)

the name of a Logical Model, or an inline Logical Model

Constructors

LMILogicalModelName LogicalModelName

LMIInlineLogicalModel (InlineLogicalModelMetadata b)

data InlineLogicalModelMetadata (b :: BackendType)

Description of an inline logical model to use in metadata (before schema cache) this has no name - it is up to the resolving user (ie, the Native Query, etc) to give the generated type a name

Constructors

InlineLogicalModelMetadata

Fields

_ilmmFields :: InsOrdHashMap (Column b) (LogicalModelField b)

_ilmmSelectPermissions :: InsOrdHashMap RoleName (SelPermDef b)