Hasura.LogicalModelResolver.Schema

Description

Schema parsers for common functionality of logical model resolvers.

Documentation

argumentsSchema

Arguments

:: forall b r m n. MonadBuildSchema b r m n

=> Text
The resolver description, such as "Stored Procedure" or "Native Query".

-> Name
The resolver name.

-> HashMap ArgumentName (NullableScalarType b)
Arguments

-> MaybeT (SchemaT r m) (InputFieldsParser n (HashMap ArgumentName (ColumnValue b)))

Schema parser for native query or stored procedure arguments.