Hasura.Metadata.DTO.Utils

Description

Utility functions for use defining autodocodec codecs.

Documentation

boolConstCodec :: Eq a => a -> a -> JSONCodec a

Map a fixed set of two values to boolean values when serializing. The first argument is the value to map to True, the second is the value to map to False.

boundedEnumCodec :: forall enum. (Eq enum, Enum enum, Bounded enum) => (enum -> String) -> JSONCodec enum

A codec for a Bounded Enum that maps to literal strings using a provided function.

Example usage

>>> data Fruit = FruitApple | FruitOrange deriving (Show, Eq, Enum, Bounded)
>>> let c = boundedEnumCodec (snakeCase . drop 5)
>>> toJSONVia c Apple
String "apple"
>>> JSON.parseMaybe (parseJSONVia c) (String "orange") :: Maybe Fruit
Just Orange

versionField :: Integer -> ObjectCodec a Scientific

Defines a required object field named version that must have the given integer value. On serialization the field will have the given value automatically. On deserialization parsing will fail unless the field has the exact given value.

optionalVersionField :: Integer -> ObjectCodec a (Maybe Scientific)

Defines an optional object field named version that must have the given integer value if the field is present. On serialization the field will have the given value automatically. On deserialization parsing will fail unless the field has the exact given value, or is absent.

discriminatorField :: Text -> Text -> ObjectCodec a ()

Useful in an object codec for a field that indicates the type of the object within a union.

codecNamePrefix :: forall b. HasTag b => Text

Provides a title-cased name for a database kind, inferring the appropriate database kind from type context.

typeableName :: forall a. Typeable a => Text

Provides a string based on the given type to use to uniquely name instantiations of polymorphic codecs.

fromEnvCodec :: JSONCodec Text

Represents a text field wrapped in an object with a single property named from_env.

Objects of this form appear in many places in the Metadata API. If we reproduced this codec in each use case the OpenAPI document would have many identical object definitions. Using a shared codec allows a single shared reference.