Hasura.NativeQuery.API

Description

Define and handle v1/metadata API operations to track, untrack, and get native queries.

Documentation

data TrackNativeQuery (b :: BackendType)

Default implementation of the track_native_query request payload.

Constructors

TrackNativeQuery

Fields

tnqSource :: SourceName

tnqRootFieldName :: NativeQueryName

tnqCode :: Text

tnqArguments :: HashMap ArgumentName (NullableScalarType b)

tnqArrayRelationships :: InsOrdHashMap RelName (RelDef (RelManualConfig b))

tnqObjectRelationships :: InsOrdHashMap RelName (RelDef (RelManualConfig b))

tnqDescription :: Maybe Text

tnqReturns :: LogicalModelIdentifier b

data GetNativeQuery (b :: BackendType)

API payload for the get_native_query endpoint.

Constructors

GetNativeQuery

Fields

gnqSource :: SourceName

runGetNativeQuery :: forall b m. (BackendMetadata b, MetadataM m, MonadError QErr m) => GetNativeQuery b -> m EncJSON

Handler for the get_native_query endpoint.

execTrackNativeQuery :: forall b m. (BackendMetadata b, MonadError QErr m) => TrackNativeQuery b -> Metadata -> m (MetadataObjId, MetadataModifier)

Handler for the track_native_query endpoint. The type 'TrackNativeQuery b' (appearing here in wrapped as 'BackendTrackNativeQuery b' for AnyBackend compatibility) is defined in 'class NativeQueryMetadata'.

data UntrackNativeQuery (b :: BackendType)

API payload for the untrack_native_query endpoint.

Constructors

UntrackNativeQuery

Fields

utnqSource :: SourceName

utnqRootFieldName :: NativeQueryName

execUntrackNativeQuery :: forall b m. (BackendMetadata b, MonadError QErr m) => UntrackNativeQuery b -> Metadata -> m (MetadataObjId, MetadataModifier)

Handler for the untrack_native_query endpoint.

dropNativeQueryInMetadata :: forall b. BackendMetadata b => SourceName -> NativeQueryName -> MetadataModifier

assertNativeQueryExists :: forall b m. (Backend b, MonadError QErr m) => SourceName -> NativeQueryName -> Metadata -> m ()

Check whether a native query with the given root field name exists for the given source.