Hasura.NativeQuery.IR

Description

The RQL IR representation of an invocation of a native query.

Documentation

data NativeQuery b field

The RQL IR representation of an invocation of a native query.

Constructors

NativeQuery

Fields

nqRootFieldName :: NativeQueryName

The graphql name of the native query.

nqInterpolatedQuery :: InterpolatedQuery field

The raw sql to use in the query

nqLogicalModel :: LogicalModel b

The return type of the native query