Hasura.NativeQuery.InterpolatedQuery

Description

Parser and prettyprinter for native query code.

Documentation

data InterpolatedItem variable

A component of an interpolated query

Constructors

IIText Text
normal text

IIVariable variable
a captured variable

newtype InterpolatedQuery variable

A list of stored procedure components representing a single stored procedure, separating the variables from the text.

Constructors

InterpolatedQuery

Fields

getInterpolatedQuery :: [InterpolatedItem variable]

parseInterpolatedQuery :: Text -> Either Text (InterpolatedQuery ArgumentName)

extract all of the `{{ variable }}` inside our query string

getUniqueVariables :: Ord var => InterpolatedQuery var -> Set var

Get a set of all arguments used in an interpolated query.