Hasura.NativeQuery.Metadata

Description

This module houses the types and functions associated with the default implementation of the metadata of native queries.

Documentation

newtype NativeQueryArgumentName

Constructors

NativeQueryArgumentName

Fields

getNativeQueryArgumentName :: Text

data NativeQueryInfoImpl b

Provisional data type for modelling metadata

Constructors

NativeQueryInfoImpl

Fields

nqiiName :: NativeQueryName

nqiiCode :: Text

nqiiReturns :: TableName b

nqiiArgs :: HashMap NativeQueryArgumentName (ScalarType b)

nqiiComment :: Text