Hasura.NativeQuery.Types

Description

This module houses the types that are necessary to even talk about native queries abstract of a concrete implementation.

The default implementation is given in modules Metadata, and API, but backends are free to provide their own as needed.

Documentation

class (APIType (TrackNativeQuery b), APIType (NativeQueryInfo b), HasCodec (NativeQueryInfo b), Representable (NativeQueryInfo b)) => NativeQueryMetadata (b :: BackendType) where

This type class models the types and functions necessary to talk about Native Queries.

Uninstantiable defaults are given for types and methods.

Minimal complete definition

Nothing

Associated Types

type NativeQueryInfo b :: Type

The type of persisted metadata.

type NativeQueryInfo b = Void

type TrackNativeQuery b :: Type

The API payload of the track_native_query api endpoint.

type TrackNativeQuery b = Void

Methods

trackNativeQuerySource :: TrackNativeQuery b -> SourceName

Projection function identifying the name of the source a track_native_query request concerns.

default trackNativeQuerySource :: TrackNativeQuery b ~ Void => TrackNativeQuery b -> SourceName

nativeQueryInfoName :: NativeQueryInfo b -> NativeQueryName

Projection function giving the name of a native query.

default nativeQueryInfoName :: NativeQueryInfo b ~ Void => NativeQueryInfo b -> NativeQueryName

nativeQueryTrackToInfo :: SourceConnConfiguration b -> TrackNativeQuery b -> ExceptT NativeQueryError IO (NativeQueryInfo b)

Projection function, producing a 'NativeQueryInfo b' from a 'TrackNativeQuery b'.

default nativeQueryTrackToInfo :: TrackNativeQuery b ~ Void => SourceConnConfiguration b -> TrackNativeQuery b -> ExceptT NativeQueryError IO (NativeQueryInfo b)

validateNativeQueryAgainstSource :: (MonadIO m, MonadError NativeQueryError m) => SourceConnConfiguration b -> NativeQueryInfo b -> m ()

Validate the native query against the database.

default validateNativeQueryAgainstSource :: NativeQueryInfo b ~ Void => SourceConnConfiguration b -> NativeQueryInfo b -> m ()

newtype BackendTrackNativeQuery b

Our API endpoint solution wraps all request payload types in AnyBackend for its multi-backend support, but type families must be fully applied to all their arguments.

So in order to be usable as an API request payload data type, 'TrackNativeQuery b' needs to be wrapped in a newtype.

Constructors

BackendTrackNativeQuery

Fields

getBackendTrackNativeQuery :: TrackNativeQuery b

data NativeQueryError

Constructors

NativeQueryParseError Text

NativeQueryValidationError QErr

newtype NativeQueryName

Constructors

NativeQueryName

Fields

getNativeQueryName :: Name