Hasura.PingSources

Documentation

runPingSources :: Environment -> (String -> IO ()) -> IO SourcePingCache -> IO a

A forever running IO loop that performs regular pings for DBs that need it these are used to send a fingerprint to third parties that wish to attribute users to Hasura