Hasura.Prelude

Documentation

alphabet :: String

alphaNumerics :: String

onNothing :: Applicative m => Maybe a -> m a -> m a

onNothingM :: Monad m => m (Maybe a) -> m a -> m a

onJust :: Applicative m => Maybe a -> (a -> m ()) -> m ()

withJust :: Applicative m => Maybe a -> (a -> m (Maybe b)) -> m (Maybe b)

onLeft :: Applicative m => Either e a -> (e -> m a) -> m a

mapLeft :: (e1 -> e2) -> Either e1 a -> Either e2 a

whenMaybe :: Applicative m => Bool -> m a -> m (Maybe a)

choice :: Alternative f => [f a] -> f a

afold :: (Foldable t, Alternative f) => t a -> f a

bsToTxt :: ByteString -> Text

lbsToTxt :: ByteString -> Text

txtToBs :: Text -> ByteString

base64Decode :: Text -> ByteString

liftEitherM :: MonadError e m => m (Either e a) -> m a

spanMaybeM :: (Foldable f, Monad m) => (a -> m (Maybe b)) -> f a -> m ([b], [a])

findWithIndex :: (a -> Bool) -> [a] -> Maybe (a, Int)

mapFromL :: (Eq k, Hashable k) => (a -> k) -> [a] -> HashMap k a

oMapFromL :: (Eq k, Hashable k) => (a -> k) -> [a] -> InsOrdHashMap k a

withElapsedTime :: MonadIO m => m a -> m (DiffTime, a)

Time an IO action, returning the time with microsecond precision. The result of the input action will be evaluated to WHNF.

The result DiffTime is guarenteed to be >= 0.

startTimer :: (MonadIO m, MonadIO n) => m (n DiffTime)

Start timing and return an action to return the elapsed time since startTimer was called.

  timer <- startTimer
  someStuffToTime
  elapsed <- timer
  moreStuff
  elapsedBoth <- timer

hoistMaybe :: Applicative m => Maybe b -> MaybeT m b

Upgrade a Maybe to a MaybeT.

cf. http://hackage.haskell.org/package/errors-2.3.0/docs/src/Control.Error.Util.html#hoistMaybe

hoistEither :: Applicative m => Either e a -> ExceptT e m a

Upgrade an Either to an ExceptT.

cf. http://hackage.haskell.org/package/errors-2.3.0/docs/src/Control.Error.Util.html#hoistEither

tshow :: Show a => a -> Text

readJson :: FromJSON a => String -> Either String a

hasuraJSON :: Options

Customized Options which apply "snake case" to Generic or Template Haskell JSON derivations.

For example, a Haskell field fooBar would be deserialized fromto JSON as foo_bar.

fold' :: (Monoid m, Foldable t) => t m -> m

Given a structure with elements whose type is a Monoid, combine them via the monoid's (<>) operator.

This fold is right-associative and strict in the accumulator; it's defined as foldMap id, per the documentation in Data.Foldable.

ltrace :: Show a => String -> a -> a

Warning: ltrace left in code

Labeled, prettified traceShowId

ltraceM :: Applicative m => Show a => String -> a -> m ()

Warning: ltraceM left in code

Labeled, prettified traceShowM

hashNub :: (Hashable a, Eq a) => [a] -> [a]

Remove duplicates from a list. Like nub but runs in O(n * log_16(n)) time and requires Hashable and Eq instances. hashNub is faster than ordNub when there're not so many different values in the list.

>>> hashNub [1,3,2,9,4,1,5,7,3,3,1,2,5,4,3,2,1,0]
[0,1,2,3,4,5,7,9]

Orphan instances

Hashable a => Hashable (Seq a)

Instance details

Methods

hashWithSalt :: Int -> Seq a -> Int

hash :: Seq a -> Int