Hasura.QueryTags

Description

This module has the various metadata we want to attach to the generated/executed query

Documentation

data QueryTags

Query Tags are SQL comments which are made up of (key=value) pairs.

These are appended to the SQL statements generated by Hasura for GraphQL operations. This enables the ability to get some application context in the database logs and also use native database monitoring tools (e.g. pganalyze) for better performance analysis.

The application context(query tags) can be used to detect slow GQL operation and relate them back to the SQL that was generated.

For eg: SELECT name FROM child * request_id=487c2ed5-08a4-429a-b0e0-4666a82e3cc6, field_name=child, operation_name=GetChild *

For more usage information, refer Query Tags Docs

Constructors

QTQuery !QueryMetadata

QTMutation !MutationMetadata

QTLiveQuery !LivequeryMetadata

newtype QueryTagsComment

query-tags as SQL comment which is appended to the prepared SQL statement

Constructors

QueryTagsComment

Fields

_unQueryTagsComment :: Text

type Attribute = (Text, Text)

data QueryTagsAttributes

emptyQueryTagsAttributes :: QueryTagsAttributes

emptyQueryTagsComment :: QueryTagsComment

data QueryMetadata

Constructors

QueryMetadata (Maybe RequestId) (Maybe Name) RootFieldAlias ParameterizedQueryHash

data MutationMetadata

Constructors

MutationMetadata (Maybe RequestId) (Maybe Name) RootFieldAlias ParameterizedQueryHash

data LivequeryMetadata

Constructors

LivequeryMetadata RootFieldAlias ParameterizedQueryHash

encodeQueryTags :: QueryTags -> QueryTagsAttributes

class Monad m => MonadQueryTags m where

Minimal complete definition

Nothing

Methods

createQueryTags :: QueryTagsAttributes -> Maybe QueryTagsConfig -> Tagged m QueryTagsComment

Creates Query Tags. These are appended to the Generated SQL. Helps users to use native database monitoring tools to get some 'application-context'.

default createQueryTags :: forall t n. (m ~ t n, MonadQueryTags n) => QueryTagsAttributes -> Maybe QueryTagsConfig -> Tagged m QueryTagsComment