Hasura.QueryTags.Types

Documentation

data QueryTagsFormat

Constructors

Standard

SQLCommenter

data QueryTagsConfig

QueryTagsConfig is the configuration created by the users to control query tags

This config let's hasura know about the following: 1. In what format should the query tags be created 2. Should they be appended to the SQL 3. Should the request id be part of the query tags (which varies on each request and will cause prepared statements to be re-prepared every time)

FWIW, QueryTagsConfig are coupled along with the Source metadata. So you can also think QueryTagsConfig as the query tags configuration for each source.

The workflow is something like this:

The QueryTagsConfig for a source is created from the metadata we get from the user.

This configuration is packaged (for the lack of better word) along with SourceConfigWith

These query tags configuration are extracted from the SourceConfigWith in the mkDBQueryPlan, mkDBMutationPlan, mkDBSubscriptionPlan functions and are passed along to the point where the actual SQL generation takes place

Note that, it is important for QueryTagsConfig to be a part of SourceConfigWith because that's the only sane way (that we can think of) the `mkDB..Plan` functions can get the QueryTagsConfig.

Constructors

QueryTagsConfig

Fields

_qtcDisabled :: !Bool

_qtcFormat :: !QueryTagsFormat

_qtcOmitRequestId :: !Bool

defaultQueryTagsConfig :: QueryTagsConfig