Hasura.RQL.DDL.Action

Documentation

getActionInfo :: (QErrM m, CacheRM m) => ActionName -> m ActionInfo

data CreateAction

Constructors

CreateAction

Fields

_caName :: !ActionName

_caDefinition :: !ActionDefinitionInput

_caComment :: !(Maybe Text)

runCreateAction :: (QErrM m, CacheRWM m, MetadataM m) => CreateAction -> m EncJSON

resolveAction :: QErrM m => Environment -> AnnotatedCustomTypes -> ActionDefinitionInput -> DMap BackendTag ScalarSet -> m (ResolvedActionDefinition, AnnotatedObjectType)

Note [Postgres scalars in action input arguments] ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ It's very comfortable to be able to reference Postgres scalars in actions input arguments. For example, see the following action mutation:

extend type mutation_root { create_user ( name: String! created_at: timestamptz ): User }

The timestamptz is a Postgres scalar. We need to validate the presence of timestamptz type in the Postgres database. So, the resolveAction function takes all Postgres scalar types as one of the inputs and returns the set of referred scalars.

data UpdateAction

Constructors

UpdateAction

Fields

_uaName :: !ActionName

_uaDefinition :: !ActionDefinitionInput

_uaComment :: !(Maybe Text)

runUpdateAction :: forall m. (QErrM m, CacheRWM m, MetadataM m) => UpdateAction -> m EncJSON

newtype ClearActionData

Constructors

ClearActionData

Fields

unClearActionData :: Bool

shouldClearActionData :: ClearActionData -> Bool

defaultClearActionData :: ClearActionData

data DropAction

Constructors

DropAction

Fields

_daName :: !ActionName

_daClearData :: !(Maybe ClearActionData)

runDropAction :: (CacheRWM m, MetadataM m, MonadMetadataStorageQueryAPI m) => DropAction -> m EncJSON

dropActionInMetadata :: ActionName -> MetadataModifier

newtype ActionMetadataField

Constructors

ActionMetadataField

Fields

unActionMetadataField :: Text

doesActionPermissionExist :: Metadata -> ActionName -> RoleName -> Bool

data CreateActionPermission

Constructors

CreateActionPermission

Fields

_capAction :: !ActionName

_capRole :: !RoleName

_capDefinition :: !(Maybe Value)

_capComment :: !(Maybe Text)

runCreateActionPermission :: (QErrM m, CacheRWM m, MetadataM m) => CreateActionPermission -> m EncJSON

data DropActionPermission

Constructors

DropActionPermission

Fields

_dapAction :: !ActionName

_dapRole :: !RoleName

runDropActionPermission :: (QErrM m, CacheRWM m, MetadataM m) => DropActionPermission -> m EncJSON

dropActionPermissionInMetadata :: ActionName -> RoleName -> MetadataModifier