Hasura.RQL.DDL.EventTrigger

Documentation

data CreateEventTriggerQuery (b :: BackendType)

Constructors

CreateEventTriggerQuery

Fields

_cetqSource :: SourceName

_cetqName :: TriggerName

_cetqTable :: TableName b

_cetqInsert :: Maybe (SubscribeOpSpec b)

_cetqUpdate :: Maybe (SubscribeOpSpec b)

_cetqDelete :: Maybe (SubscribeOpSpec b)

_cetqEnableManual :: Maybe Bool

_cetqRetryConf :: Maybe RetryConf

_cetqWebhook :: Maybe InputWebhook

_cetqWebhookFromEnv :: Maybe Text

_cetqHeaders :: Maybe [HeaderConf]

_cetqReplace :: Bool

_cetqRequestTransform :: Maybe RequestTransform

_cetqResponseTrasnform :: Maybe MetadataResponseTransform

_cteqCleanupConfig :: Maybe AutoTriggerLogCleanupConfig

cteqCleanupConfig :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe AutoTriggerLogCleanupConfig)

cetqWebhookFromEnv :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe Text)

cetqWebhook :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe InputWebhook)

cetqUpdate :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe (SubscribeOpSpec b))

cetqTable :: forall b. Lens' (CreateEventTriggerQuery b) (TableName b)

cetqSource :: forall b. Lens' (CreateEventTriggerQuery b) SourceName

cetqRetryConf :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe RetryConf)

cetqResponseTrasnform :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe MetadataResponseTransform)

cetqRequestTransform :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe RequestTransform)

cetqReplace :: forall b. Lens' (CreateEventTriggerQuery b) Bool

cetqName :: forall b. Lens' (CreateEventTriggerQuery b) TriggerName

cetqInsert :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe (SubscribeOpSpec b))

cetqHeaders :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe [HeaderConf])

cetqEnableManual :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe Bool)

cetqDelete :: forall b. Lens' (CreateEventTriggerQuery b) (Maybe (SubscribeOpSpec b))

data DeleteEventTriggerQuery (b :: BackendType)

Constructors

DeleteEventTriggerQuery

Fields

_detqSource :: SourceName

_detqName :: TriggerName

data RedeliverEventQuery (b :: BackendType)

Constructors

RedeliverEventQuery

Fields

_rdeqEventId :: EventId

_rdeqSource :: SourceName

data InvokeEventTriggerQuery (b :: BackendType)

Constructors

InvokeEventTriggerQuery

Fields

_ietqName :: TriggerName

_ietqSource :: SourceName

_ietqPayload :: Value

class Monad m => MonadEventLogCleanup m where

This typeclass have the implementation logic for the event trigger log cleanup

Methods

runLogCleaner :: TriggerLogCleanupConfig -> m (Either QErr EncJSON)

generateCleanupSchedules :: AnyBackend SourceInfo -> TriggerName -> AutoTriggerLogCleanupConfig -> m (Either QErr ())

resolveEventTriggerQuery :: forall b m. (Backend b, UserInfoM m, QErrM m, CacheRM m) => CreateEventTriggerQuery b -> m (Bool, EventTriggerConf b)

droppedTriggerOps :: TriggerOpsDef b -> TriggerOpsDef b -> HashSet Ops

createEventTriggerQueryMetadata :: forall b m r. (BackendMetadata b, QErrM m, UserInfoM m, CacheRWM m, MetadataM m, BackendEventTrigger b, MonadIO m, MonadEventLogCleanup m, MonadReader r m, Has (Logger Hasura) r) => CreateEventTriggerQuery b -> m ()

runCreateEventTriggerQuery :: forall b m r. (BackendMetadata b, BackendEventTrigger b, QErrM m, UserInfoM m, CacheRWM m, MetadataM m, MonadIO m, MonadEventLogCleanup m, MonadReader r m, Has (Logger Hasura) r) => CreateEventTriggerQuery b -> m EncJSON

runDeleteEventTriggerQuery :: forall b m. (BackendEventTrigger b, MonadError QErr m, CacheRWM m, MonadIO m, MetadataM m) => DeleteEventTriggerQuery b -> m EncJSON

runRedeliverEvent :: forall b m. (BackendEventTrigger b, MonadIO m, CacheRM m, QErrM m, MetadataM m) => RedeliverEventQuery b -> m EncJSON

runInvokeEventTrigger :: forall b m. (MonadIO m, QErrM m, CacheRM m, MetadataM m, MonadTrace m, UserInfoM m, BackendEventTrigger b) => InvokeEventTriggerQuery b -> m EncJSON

askTabInfoFromTrigger :: (Backend b, QErrM m, CacheRM m) => SourceName -> TriggerName -> m (TableInfo b)

getTabInfoFromSchemaCache :: (Backend b, QErrM m) => SchemaCache -> SourceName -> TriggerName -> m (TableInfo b)

askEventTriggerInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TriggerName -> m (EventTriggerInfo b)

maxTriggerNameLength :: Int

getHeaderInfosFromConf :: QErrM m => Environment -> [HeaderConf] -> m [EventHeaderInfo]

getWebhookInfoFromConf :: QErrM m => Environment -> WebhookConf -> m WebhookConfInfo

buildEventTriggerInfo :: forall b m. (Backend b, QErrM m) => Environment -> SourceName -> TableName b -> EventTriggerConf b -> m (EventTriggerInfo b, [SchemaDependency])

getTrigDefDeps :: forall b. Backend b => SourceName -> TableName b -> TriggerOpsDef b -> [SchemaDependency]

getTriggersMap :: SourceMetadata b -> InsOrdHashMap TriggerName (EventTriggerConf b)

getTriggerNames :: SourceMetadata b -> HashSet TriggerName

getTableNameFromTrigger :: forall b m. (Backend b, QErrM m) => SchemaCache -> SourceName -> TriggerName -> m (TableName b)

runCleanupEventTriggerLog :: (MonadEventLogCleanup m, MonadError QErr m) => TriggerLogCleanupConfig -> m EncJSON

updateCleanupStatusInMetadata :: forall b m. (Backend b, QErrM m, CacheRWM m, MetadataM m) => AutoTriggerLogCleanupConfig -> EventTriggerCleanupStatus -> SourceName -> TableName b -> TriggerName -> m ()

Updates the cleanup switch in metadata given the source, table and trigger name The Bool value represents the status of the cleaner, whether to start or pause it

toggleEventTriggerCleanupAction :: forall m. (MonadIO m, QErrM m, CacheRWM m, MetadataM m) => TriggerLogCleanupToggleConfig -> EventTriggerCleanupStatus -> m EncJSON

Function to start/stop the cleanup action based on the event triggers supplied in TriggerLogCleanupToggleConfig conf

runEventTriggerResumeCleanup :: forall m. (MonadIO m, QErrM m, CacheRWM m, MetadataM m) => TriggerLogCleanupToggleConfig -> m EncJSON

runEventTriggerPauseCleanup :: (MonadError QErr m, CacheRWM m, MonadIO m, MetadataM m) => TriggerLogCleanupToggleConfig -> m EncJSON

getAllEventTriggersWithCleanupConfig :: TableInfo b -> [(TriggerName, AutoTriggerLogCleanupConfig)]

Collects and returns all the event triggers with cleanup config

hasCleanupCronScheduleUpdated :: Maybe AutoTriggerLogCleanupConfig -> Maybe AutoTriggerLogCleanupConfig -> Bool

getAllETWithCleanupConfigInTableMetadata :: TableMetadata b -> [(TriggerName, AutoTriggerLogCleanupConfig)]