Hasura.RQL.DDL.Metadata.Types

Description

Types related to metadata management API

Documentation

data ClearMetadata

Constructors

ClearMetadata

data ExportMetadata

Constructors

ExportMetadata

data ReloadSpec a

Constructors

RSReloadAll

RSReloadList !(HashSet a)

type ReloadRemoteSchemas = ReloadSpec RemoteSchemaName

type ReloadSources = ReloadSpec SourceName

reloadAllRemoteSchemas :: ReloadRemoteSchemas

reloadAllSources :: ReloadSources

data ReloadMetadata

Constructors

ReloadMetadata

Fields

_rmReloadRemoteSchemas :: ReloadRemoteSchemas

_rmReloadSources :: ReloadSources

_rmRecreateEventTriggers :: ReloadSources

Provides a way for the user to allow to explicitly recreate event triggers for some or all the sources. This is useful when a user may have fiddled with the SQL trigger in the source and they'd simply want the event trigger to be recreated without deleting and creating the event trigger. By default, no source's event triggers will be recreated.

data DumpInternalState

Constructors

DumpInternalState

data GetInconsistentMetadata

Constructors

GetInconsistentMetadata

data DropInconsistentMetadata

Constructors

DropInconsistentMetadata

data AllowInconsistentMetadata

Constructors

AllowInconsistentMetadata

NoAllowInconsistentMetadata

data ReplaceMetadataV1

Constructors

RMWithSources !Metadata

RMWithoutSources !MetadataNoSources

data ReplaceMetadataV2

Constructors

ReplaceMetadataV2

Fields

_rmv2AllowInconsistentMetadata :: !AllowInconsistentMetadata

_rmv2Metadata :: !ReplaceMetadataV1

data ReplaceMetadata

Constructors

RMReplaceMetadataV1 !ReplaceMetadataV1

RMReplaceMetadataV2 !ReplaceMetadataV2

data WebHookUrl

Constructors

EnvVar String

URL Text

data TestWebhookTransform

Constructors

TestWebhookTransform

Fields

_twtEnv :: Environment

_twtWebhookUrl :: WebHookUrl

_twtPayload :: Value

_twtTransformer :: MetadataTransform

_twtSessionVariables :: Maybe SessionVariables