Hasura.RQL.DDL.OpenTelemetry

Documentation

runSetOpenTelemetryConfig :: (MonadError QErr m, MetadataM m, CacheRWM m) => OpenTelemetryConfig -> m EncJSON

Set the OpenTelemetry configuration to the provided value.

runSetOpenTelemetryStatus :: (MonadError QErr m, MetadataM m, CacheRWM m) => OtelStatus -> m EncJSON

Set just the "status" field of the OpenTelemetry configuration.

parseOtelExporterConfig :: OtelStatus -> Environment -> OtelExporterConfig -> Either QErr (Maybe OtelExporterInfo)

Smart constructor for OtelExporterInfo.

Returns a Left qErr to signal a validation error. Returns Right Nothing to signal that the exporter should be disabled without raising an error.

Allows the trace endpoint to be unset if the entire OpenTelemetry system is disabled.

parseOtelBatchSpanProcessorConfig :: OtelBatchSpanProcessorConfig -> Either QErr OtelBatchSpanProcessorInfo