Hasura.RQL.DDL.Permission

Documentation

data CreatePerm (a :: BackendType -> Type) (b :: BackendType)

runCreatePerm :: forall m (b :: BackendType) (a :: BackendType -> Type). (UserInfoM m, CacheRWM m, MonadError QErr m, MetadataM m, BackendMetadata b) => CreatePerm a b -> m EncJSON

data PermDef (b :: BackendType) (perm :: BackendType -> Type)

Constructors

PermDef

Fields

_pdRole :: RoleName

_pdPermission :: PermDefPermission b perm

_pdComment :: Maybe Text

data InsPerm (b :: BackendType)

Constructors

InsPerm

Fields

ipCheck :: BoolExp b

ipSet :: Maybe (ColumnValues b Value)

ipColumns :: Maybe (PermColSpec b)

ipBackendOnly :: Bool

ipValidateInput :: Maybe (ValidateInput InputWebhook)

type InsPermDef (b :: BackendType) = PermDef b InsPerm

buildInsPermInfo :: forall (b :: BackendType) m r. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => Environment -> SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> InsPerm b -> m (WithDeps (InsPermInfo b))

data SelPerm (b :: BackendType)

Constructors

SelPerm

Fields

spColumns :: PermColSpec b

Allowed columns

spFilter :: BoolExp b

Filter expression

spLimit :: Maybe Int

Limit value

spAllowAggregations :: Bool

Allow aggregation

spComputedFields :: [ComputedFieldName]

Allowed computed fields which should not include the fields returning rows of existing table.

spAllowedQueryRootFields :: AllowedRootFields QueryRootFieldType

spAllowedSubscriptionRootFields :: AllowedRootFields SubscriptionRootFieldType

type SelPermDef (b :: BackendType) = PermDef b SelPerm

buildSelPermInfo :: forall (b :: BackendType) m r. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> RoleName -> SelPerm b -> m (WithDeps (SelPermInfo b))

data UpdPerm (b :: BackendType)

Constructors

UpdPerm

Fields

ucColumns :: PermColSpec b

ucSet :: Maybe (ColumnValues b Value)

ucFilter :: BoolExp b

ucCheck :: Maybe (BoolExp b)

Check expression, which must be true after update. This is optional because we don't want to break the v1 API but Nothing should be equivalent to the expression which always returns true.

ucBackendOnly :: Bool

ucValidateInput :: Maybe (ValidateInput InputWebhook)

type UpdPermDef (b :: BackendType) = PermDef b UpdPerm

buildUpdPermInfo :: forall (b :: BackendType) m r. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => Environment -> SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> UpdPerm b -> m (WithDeps (UpdPermInfo b))

data DelPerm (b :: BackendType)

Constructors

DelPerm

Fields

dcFilter :: BoolExp b

dcBackendOnly :: Bool

dcValidateInput :: Maybe (ValidateInput InputWebhook)

type DelPermDef (b :: BackendType) = PermDef b DelPerm

buildDelPermInfo :: forall (b :: BackendType) m r. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => Environment -> SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> DelPerm b -> m (WithDeps (DelPermInfo b))

data DropPerm (b :: BackendType)

runDropPerm :: forall (b :: BackendType) m. (UserInfoM m, CacheRWM m, MonadError QErr m, MetadataM m, BackendMetadata b) => PermType -> DropPerm b -> m EncJSON

dropPermissionInMetadata :: forall (b :: BackendType). RoleName -> PermType -> TableMetadata b -> TableMetadata b

data SetPermComment (b :: BackendType)

Constructors

SetPermComment

Fields

apSource :: SourceName

apTable :: TableName b

apRole :: RoleName

apPermission :: PermType

apComment :: Maybe Text

runSetPermComment :: forall (b :: BackendType) m. (QErrM m, CacheRWM m, MetadataM m, BackendMetadata b) => SetPermComment b -> m EncJSON

type family PermInfo (perm :: BackendType -> Type) :: BackendType -> Type where ...

Equations

PermInfo SelPerm = SelPermInfo

PermInfo InsPerm = InsPermInfo

PermInfo UpdPerm = UpdPermInfo

PermInfo DelPerm = DelPermInfo

buildPermInfo :: forall (b :: BackendType) m r (perm :: BackendType -> Type). (BackendMetadata b, QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => Environment -> SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> RoleName -> PermDefPermission b perm -> m (WithDeps (PermInfo perm b))

buildLogicalModelPermInfo :: forall (b :: BackendType) m r (perm :: BackendType -> Type). (BackendMetadata b, QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => SourceName -> LogicalModelLocation -> InsOrdHashMap (Column b) (LogicalModelField b) -> PermDefPermission b perm -> m (WithDeps (PermInfo perm b))

Given the logical model's definition and the permissions as defined in the logical model's metadata, try to construct the permission definition.

addPermissionToMetadata :: forall (b :: BackendType) (a :: BackendType -> Type). PermDef b a -> TableMetadata b -> TableMetadata b

annBoolExp :: forall m (b :: BackendType) v. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b) => BoolExpRHSParser b m v -> FieldInfoMap (FieldInfo b) -> FieldInfoMap (FieldInfo b) -> GBoolExp b ColExp -> m (AnnBoolExp b v)