Hasura.RQL.DDL.RemoteRelationship

Documentation

data CreateFromSourceRelationship (b :: BackendType)

Argument to the _create_remote_relationship and _update_remote_relationship families of metadata commands.

For historical reason, this type is also used to represent a db-to-rs schema in the metadata.

Constructors

CreateFromSourceRelationship

Fields

_crrSource :: SourceName

_crrTable :: TableName b

_crrName :: RelName

_crrDefinition :: RemoteRelationshipDefinition

runCreateRemoteRelationship :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => CreateFromSourceRelationship b -> m EncJSON

runUpdateRemoteRelationship :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => CreateFromSourceRelationship b -> m EncJSON

data DeleteFromSourceRelationship (b :: BackendType)

Argument to the _drop_remote_relationship family of metadata commands.

Constructors

DeleteFromSourceRelationship

Fields

_drrSource :: SourceName

_drrTable :: TableName b

_drrName :: RelName

runDeleteRemoteRelationship :: forall b m. (BackendMetadata b, MonadError QErr m, CacheRWM m, MetadataM m) => DeleteFromSourceRelationship b -> m EncJSON

data CreateRemoteSchemaRemoteRelationship

Constructors

CreateRemoteSchemaRemoteRelationship

Fields

_crsrrRemoteSchema :: RemoteSchemaName

_crsrrType :: Name

_crsrrName :: RelName

_crsrrDefinition :: RemoteRelationshipDefinition

runCreateRemoteSchemaRemoteRelationship :: forall m. (MonadError QErr m, CacheRWM m, MetadataM m) => CreateRemoteSchemaRemoteRelationship -> m EncJSON

runUpdateRemoteSchemaRemoteRelationship :: forall m. (MonadError QErr m, CacheRWM m, MetadataM m) => CreateRemoteSchemaRemoteRelationship -> m EncJSON

data DeleteRemoteSchemaRemoteRelationship

Argument to the _drop_remote_relationship family of metadata commands.

Constructors

DeleteRemoteSchemaRemoteRelationship

Fields

_drsrrRemoteSchema :: RemoteSchemaName

_drsrrTypeName :: Name

_drsrrName :: RelName

runDeleteRemoteSchemaRemoteRelationship :: forall m. (MonadError QErr m, CacheRWM m, MetadataM m) => DeleteRemoteSchemaRemoteRelationship -> m EncJSON

data PartiallyResolvedSource b

Internal intermediary step.

We build the output of sources in two steps: 1. we first resolve sources, and collect the core info of their tables 2. we then build the entire output from the collection of partially resolved sources

We need this split to be able to resolve cross-source relationships: to process one source's remote relationship, we need to know about the target source's tables core info.

This data structure is used as an argument to AnyBackend in the backend-agnostic intermediary collection, and used here to build remote field info.

Constructors

PartiallyResolvedSource

Fields

_prsSourceMetadata :: !(SourceMetadata b)

_resolvedSource :: !(ResolvedSource b)

_tableCoreInfoMap :: !(HashMap (TableName b) (TableCoreInfoG b (ColumnInfo b) (ColumnInfo b)))

buildRemoteFieldInfo

Arguments

:: QErrM m

=> LHSIdentifier
The entity on which the remote relationship is defined

-> HashMap FieldName lhsJoinField
join fields provided by the LHS entity

-> RemoteRelationship
definition of remote relationship

-> HashMap SourceName (AnyBackend PartiallyResolvedSource)
Required context to process cross boundary relationships

-> RemoteSchemaMap
Required context to process cross boundary relationships

-> m (RemoteFieldInfo lhsJoinField, [SchemaDependency])
returns 1. schema cache representation of the remote relationships 2. the dependencies on the RHS of the join. The dependencies on the LHS entities has to be handled by the calling function

Builds the schema cache representation of a remote relationship TODO: this is not actually called by the remote relationship DDL API and is only used as part of the schema cache process. Should this be moved elsewhere?

getRemoteSchemaEntityJoinColumns :: MonadError QErr m => RemoteSchemaName -> RemoteSchemaIntrospection -> Name -> m (HashMap FieldName Name)