Hasura.RQL.DDL.ScheduledTrigger

Documentation

populateInitialCronTriggerEvents :: (MonadIO m, MonadMetadataStorageQueryAPI m) => CronSchedule -> TriggerName -> m ()

runCreateCronTrigger :: (CacheRWM m, MonadIO m, MetadataM m, MonadMetadataStorageQueryAPI m) => CreateCronTrigger -> m EncJSON

runCreateCronTrigger will update a existing cron trigger when the replace value is set to true and when replace is false a new cron trigger will be created

resolveCronTrigger :: QErrM m => Environment -> CronTriggerMetadata -> m CronTriggerInfo

updateCronTrigger :: (CacheRWM m, MonadIO m, MetadataM m, MonadMetadataStorageQueryAPI m) => CronTriggerMetadata -> m EncJSON

runDeleteCronTrigger :: (CacheRWM m, MetadataM m, MonadMetadataStorageQueryAPI m) => ScheduledTriggerName -> m EncJSON

dropCronTriggerInMetadata :: TriggerName -> MetadataModifier

runCreateScheduledEvent :: MonadMetadataStorageQueryAPI m => CreateScheduledEvent -> m EncJSON

checkExists :: (CacheRM m, MonadError QErr m) => TriggerName -> m ()

runDeleteScheduledEvent :: MonadMetadataStorageQueryAPI m => DeleteScheduledEvent -> m EncJSON

runGetScheduledEvents :: (CacheRM m, MonadMetadataStorageQueryAPI m) => GetScheduledEvents -> m EncJSON

runGetEventInvocations :: (CacheRM m, MonadMetadataStorageQueryAPI m) => GetEventInvocations -> m EncJSON

runGetCronTriggers :: MetadataM m => m EncJSON

Metadata API handler to retrieve all the cron triggers from the metadata