Hasura.RQL.DDL.Schema.Cache.Common

Description

Types/functions shared between modules that implement Hasura.RQL.DDL.Schema.Cache. Other modules should not import this module directly.

Documentation

newtype BackendInvalidationKeysWrapper (b :: BackendType)

Constructors

BackendInvalidationKeysWrapper

Fields

unBackendInvalidationKeysWrapper :: BackendInvalidationKeys b

data InvalidationKeys

InvalidationKeys used to apply requested CacheInvalidations.

Constructors

InvalidationKeys

Fields

_ikMetadata :: InvalidationKey

_ikRemoteSchemas :: HashMap RemoteSchemaName InvalidationKey

_ikSources :: HashMap SourceName InvalidationKey

_ikBackends :: BackendMap BackendInvalidationKeysWrapper

ikSources :: Lens' InvalidationKeys (HashMap SourceName InvalidationKey)

ikRemoteSchemas :: Lens' InvalidationKeys (HashMap RemoteSchemaName InvalidationKey)

ikMetadata :: Lens' InvalidationKeys InvalidationKey

ikBackends :: Lens' InvalidationKeys (BackendMap BackendInvalidationKeysWrapper)

initialInvalidationKeys :: InvalidationKeys

invalidateKeys :: CacheInvalidations -> InvalidationKeys -> InvalidationKeys

data TableBuildInput b

Constructors

TableBuildInput (TableName b) Bool (TableConfig b) (Maybe ApolloFederationConfig) (Maybe LogicalModelName)

data NonColumnTableInputs b

Constructors

NonColumnTableInputs

Fields

_nctiTable :: TableName b

_nctiObjectRelationships :: [ObjRelDef b]

_nctiArrayRelationships :: [ArrRelDef b]

_nctiComputedFields :: [ComputedFieldMetadata b]

_nctiRemoteRelationships :: [RemoteRelationship]

data TablePermissionInputs b

Constructors

TablePermissionInputs

Fields

_tpiTable :: TableName b

_tpiInsert :: [InsPermDef b]

_tpiSelect :: [SelPermDef b]

_tpiUpdate :: [UpdPermDef b]

_tpiDelete :: [DelPermDef b]

mkTableInputs :: TableMetadata b -> (TableBuildInput b, NonColumnTableInputs b, TablePermissionInputs b)

data BuildOutputs

The direct output of buildSchemaCacheRule. Contains most of the things necessary to build a schema cache, but dependencies and inconsistent metadata objects are collected via a separate MonadWriter side channel.

See also Note [Avoiding GraphQL schema rebuilds when changing irrelevant Metadata]

Constructors

BuildOutputs

Fields

_boSources :: SourceCache

_boActions :: ActionCache

_boRemoteSchemas :: HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject)

We preserve the MetadataObject from the original catalog metadata in the output so we can reuse it later if we need to mark the remote schema inconsistent during GraphQL schema generation (because of field conflicts).

_boCustomTypes :: AnnotatedCustomTypes

_boRoles :: HashMap RoleName Role

_boBackendCache :: BackendCache

boSources :: Lens' BuildOutputs SourceCache

boRoles :: Lens' BuildOutputs (HashMap RoleName Role)

boRemoteSchemas :: Lens' BuildOutputs (HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject))

boCustomTypes :: Lens' BuildOutputs AnnotatedCustomTypes

boBackendCache :: Lens' BuildOutputs BackendCache

boActions :: Lens' BuildOutputs ActionCache

data CacheBuildParams

Parameters required for schema cache build

Constructors

CacheBuildParams Manager (SourceResolver ('Postgres 'Vanilla)) (SourceResolver 'MSSQL) CacheStaticConfig

data CacheBuild a

The monad in which @RebuildableSchemaCache is being run

runCacheBuild :: (MonadIO m, MonadError QErr m) => CacheBuildParams -> CacheBuild a -> m a

runCacheBuildM :: (MonadIO m, MonadError QErr m, MonadResolveSource m, ProvidesNetwork m, HasCacheStaticConfig m) => CacheBuild a -> m a

data SourcesIntrospectionStatus

The status of collection of stored introspections of remote schemas and data sources.

Constructors

SourcesIntrospectionChangedFull StoredIntrospection
A full introspection collection of all available remote schemas and data sources.

SourcesIntrospectionChangedPartial StoredIntrospection
A partial introspection collection. Does not include all configured remote schemas and data sources, because they were not available.

SourcesIntrospectionUnchanged
None of remote schemas or data sources introspection is refetched.

data RebuildableSchemaCache

Constructors

RebuildableSchemaCache SchemaCache InvalidationKeys (Rule (ReaderT BuildReason CacheBuild) (MetadataWithResourceVersion, CacheDynamicConfig, InvalidationKeys, Maybe StoredIntrospection) (SchemaCache, (SourcesIntrospectionStatus, SchemaRegistryAction)))

withRecordDependencies :: ArrowWriter (Seq CollectItem) arr => WriterA (Seq SchemaDependency) arr (e, s) a -> arr (e, (MetadataObject, (SchemaObjId, s))) a

buildInfoMap :: (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectItem) arr, Hashable k) => (a -> k) -> (a -> MetadataObject) -> ((e, a) `arr` Maybe b) -> (e, [a]) `arr` HashMap k b

Processes a list of catalog metadata into a map of processed information, marking any duplicate entries inconsistent.

buildInfoMapM :: (MonadWriter (Seq CollectItem) m, Hashable k) => (a -> k) -> (a -> MetadataObject) -> (a -> m (Maybe b)) -> [a] -> m (HashMap k b)

buildInfoMapPreservingMetadata :: (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectItem) arr, Hashable k) => (a -> k) -> (a -> MetadataObject) -> ((e, a) `arr` Maybe b) -> (e, [a]) `arr` HashMap k (b, MetadataObject)

Like buildInfoMap, but includes each processed info’s associated MetadataObject in the result. This is useful if the results will be further processed, and the MetadataObject is still needed to mark the object inconsistent.

buildInfoMapPreservingMetadataM :: (MonadWriter (Seq CollectItem) m, Hashable k) => (a -> k) -> (a -> MetadataObject) -> (a -> m (Maybe b)) -> [a] -> m (HashMap k (b, MetadataObject))

addTableContext :: Backend b => TableName b -> Text -> Text

addLogicalModelContext :: LogicalModelLocation -> Text -> Text