Hasura.RQL.DDL.Schema.Cache.Config

Documentation

data CacheStaticConfig

This type aggregates all of the "static" configuration of the cache build.

Static arguments are the ones that will not change during the execution of the engine. They are a subset of the environment of the engine (see AppEnv and Note [Hasura Application State] for more information).

While AppEnv has access to the union of *all* the static configuration of the engine, more specific parts of the code should avoid relying directly on it to avoid being tied to unrelated parts of the codebase. (See FIXME).

Constructors

CacheStaticConfig

Fields

_cscMaintenanceMode :: MaintenanceMode ()

_cscEventingMode :: EventingMode

_cscReadOnlyMode :: ReadOnlyMode

_cscAreNativeQueriesEnabled :: Bool

Native queries can be enabled or disabled on the fly via a feature flag, however we only recognise a change on a restart

_cscAreStoredProceduresEnabled :: Bool

Stored procedures can be enabled or disabled on the fly via a feature flag, however we only recognise a change on a restart

class Monad m => HasCacheStaticConfig m where

Methods

askCacheStaticConfig :: m CacheStaticConfig

data CacheDynamicConfig

This type aggregates all of the "dynamic" configuration of the cache build.

Dynamic arguments are the ones that might change during the execution of the engine. They are a subset of the AppContext (see Note [Hasura Application State] for more information).

While AppContext has access to the union of *all* the dynamic configuration of the engine, more specific parts of the code should avoid relying directly on it to avoid being tied to unrelated parts of the codebase. (See FIXME).

Constructors

CacheDynamicConfig

Fields

_cdcFunctionPermsCtx :: InferFunctionPermissions

_cdcRemoteSchemaPermsCtx :: RemoteSchemaPermissions

_cdcSQLGenCtx :: SQLGenCtx

_cdcExperimentalFeatures :: HashSet ExperimentalFeature

_cdcDefaultNamingConvention :: NamingCase

_cdcMetadataDefaults :: MetadataDefaults

_cdcApolloFederationStatus :: ApolloFederationStatus