Hasura.RQL.DDL.Schema.Cache.Dependencies

Documentation

resolveDependencies :: (ArrowKleisli m arr, QErrM m) => (BuildOutputs, [(MetadataObject, SchemaObjId, SchemaDependency)]) `arr` (BuildOutputs, [InconsistentMetadata], DepMap)

Processes collected CIDependency values into a DepMap, performing integrity checking to ensure the dependencies actually exist. If a dependency is missing, its transitive dependents are removed from the cache, and InconsistentMetadatas are returned.

performIteration :: QErrM m => Int -> BuildOutputs -> [InconsistentMetadata] -> HashMap SchemaObjId [(MetadataObject, SchemaDependency)] -> m (BuildOutputs, [InconsistentMetadata], DepMap)

pruneDanglingDependents :: BuildOutputs -> HashMap SchemaObjId [(MetadataObject, SchemaDependency)] -> ([InconsistentMetadata], HashMap SchemaObjId [(MetadataObject, SchemaDependency)])

deleteMetadataObject :: MetadataObjId -> BuildOutputs -> BuildOutputs