Hasura.RQL.DDL.Schema.Cache.Permission

Documentation

mkBooleanPermissionMap :: (RoleName -> a) -> HashMap RoleName a -> OrderedRoles -> HashMap RoleName a

data OrderedRoles

OrderedRoles is a data type to hold topologically sorted roles according to each role's parent roles, see orderRoles for more details.

orderRoles :: MonadError QErr m => [Role] -> m OrderedRoles

orderRoles is used to order the roles, in such a way that given a role R with n parent roles - PR1, PR2 .. PRn, then the orderRoles function will order the roles in such a way that all the parent roles precede the role R. Note that the order of the parent roles itself doesn't matter as long as they precede the roles on which they are dependent on.

For example, the orderRoles may return `[PR1, PR3, PR2, ... PRn, R]` or `[PR5, PR3, PR1 ... R]`, both of them are correct because all the parent roles precede the inherited role R, assuming the parent roles themselves don't have any parents for the sake of this example.

resolveCheckPermission :: forall m p. MonadWriter (Seq CollectedInfo) m => CheckPermission p -> RoleName -> InconsistentRoleEntity -> m (Maybe p)

resolveCheckPermission is a helper function which will convert the indermediate type CheckPermission to its original type. It will record any metadata inconsistencies, if exists.

buildTablePermissions :: forall b m arr. (ArrowChoice arr, ArrowDistribute arr, ArrowCache m arr, MonadError QErr m, ArrowWriter (Seq CollectedInfo) arr, BackendMetadata b, Cacheable (Proxy b), GetAggregationPredicatesDeps b) => (Proxy b, SourceName, Dependency (TableCoreCache b), FieldInfoMap (FieldInfo b), TablePermissionInputs b, OrderedRoles) `arr` RolePermInfoMap b

mkPermissionMetadataObject :: forall b a. BackendMetadata b => SourceName -> TableName b -> PermDef b a -> MetadataObject

mkRemoteSchemaPermissionMetadataObject :: AddRemoteSchemaPermission -> MetadataObject