Hasura.RQL.DDL.Schema.Function

Description

Documentation

newtype TrackFunction b

Constructors

TrackFunction

Fields

tfName :: FunctionName b

trackFunctionP1 :: forall b m. (CacheRM m, QErrM m, Backend b) => SourceName -> FunctionName b -> m ()

Track function, Phase 1: Validate function tracking operation. Fails if function is already being tracked, or if a table with the same name is being tracked.

trackFunctionP2 :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => SourceName -> FunctionName b -> FunctionConfig -> Maybe Text -> m EncJSON

handleMultipleFunctions :: forall b m a. (QErrM m, Backend b) => FunctionName b -> [a] -> m a

runTrackFunc :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => TrackFunction b -> m EncJSON

data TrackFunctionV2 (b :: BackendType)

JSON API payload for v2 of track_function:

https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-functions.html#track-function-v2

Constructors

TrackFunctionV2

Fields

_tfv2Source :: !SourceName

_tfv2Function :: !(FunctionName b)

_tfv2Configuration :: !FunctionConfig

_tfv2Comment :: !(Maybe Text)

runTrackFunctionV2 :: forall b m. (BackendMetadata b, QErrM m, CacheRWM m, MetadataM m) => TrackFunctionV2 b -> m EncJSON

data UnTrackFunction b

JSON API payload for untrack_function:

https://hasura.io/docs/latest/graphql/core/api-reference/schema-metadata-api/custom-functions.html#untrack-function

Constructors

UnTrackFunction

Fields

_utfFunction :: !(FunctionName b)

_utfSource :: !SourceName

runUntrackFunc :: forall b m. (CacheRWM m, MonadError QErr m, MetadataM m, BackendMetadata b) => UnTrackFunction b -> m EncJSON

data FunctionPermissionArgument b

Constructors

FunctionPermissionArgument

Fields

_afpFunction :: !(FunctionName b)

_afpSource :: !SourceName

_afpRole :: !RoleName

runCreateFunctionPermission :: forall b m. (CacheRWM m, MonadError QErr m, MetadataM m, BackendMetadata b) => FunctionPermissionArgument b -> m EncJSON

dropFunctionPermissionInMetadata :: forall b. BackendMetadata b => SourceName -> FunctionName b -> RoleName -> MetadataModifier

doesFunctionPermissionExist :: forall b. BackendMetadata b => Metadata -> SourceName -> FunctionName b -> RoleName -> Bool

runDropFunctionPermission :: forall m b. (CacheRWM m, MonadError QErr m, MetadataM m, BackendMetadata b) => FunctionPermissionArgument b -> m EncJSON

data SetFunctionCustomization b

Represents the payload of the API command pg_set_function_customization.

See the Hasura API reference for a detailed description.

Constructors

SetFunctionCustomization

Fields

_sfcSource :: SourceName

_sfcFunction :: FunctionName b

_sfcConfiguration :: FunctionConfig

runSetFunctionCustomization :: forall b m. (QErrM m, CacheRWM m, MetadataM m, Backend b, BackendMetadata b) => SetFunctionCustomization b -> m EncJSON

Changes the custom names of a function. Used in the API command pg_set_function_customization.