Hasura.RQL.DDL.Schema.Source

Documentation

data AddSource b

runAddSource :: forall m b. (MonadIO m, MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => Environment -> AddSource b -> m EncJSON

data RenameSource

runRenameSource :: forall m. (MonadError QErr m, CacheRWM m, MetadataM m) => RenameSource -> m EncJSON

data DropSource

Constructors

DropSource

Fields

_dsName :: SourceName

_dsCascade :: Bool

runDropSource :: forall m r. (MonadError QErr m, CacheRWM m, MonadIO m, MonadBaseControl IO m, MetadataM m, MonadReader r m, Has (Logger Hasura) r) => DropSource -> m EncJSON

runPostDropSourceHook :: forall m r b. (MonadError QErr m, MonadIO m, MonadBaseControl IO m, MonadReader r m, Has (Logger Hasura) r, BackendMetadata b) => SourceName -> SourceInfo b -> m ()

data UpdateSource b

runUpdateSource :: forall m b. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => UpdateSource b -> m EncJSON

newtype GetSourceTables (b :: BackendType)

Constructors

GetSourceTables

Fields

_gstSourceName :: SourceName

runGetSourceTrackables :: forall b m r. (BackendMetadata b, CacheRM m, MonadError QErr m, MetadataM m, MonadIO m, MonadBaseControl IO m, MonadReader r m, Has (Logger Hasura) r, ProvidesNetwork m) => GetSourceTrackables b -> m EncJSON

Fetch a list of tables for the request data source.

newtype GetSourceTrackables (b :: BackendType)

Constructors

GetSourceTrackables

Fields

_gstrSourceName :: SourceName

runGetSourceTables :: forall b m r. (BackendMetadata b, CacheRM m, MonadError QErr m, MetadataM m, MonadIO m, MonadBaseControl IO m, MonadReader r m, Has (Logger Hasura) r, ProvidesNetwork m) => GetSourceTables b -> m EncJSON

Fetch a list of tables for the request data source.

data GetTableInfo_

Constructors

GetTableInfo_

Fields

_gtiSourceName_ :: SourceName

_gtiTableName_ :: TableName 'DataConnector

runGetTableInfo_ :: (CacheRM m, MonadError QErr m, MetadataM m, MonadBaseControl IO m, MonadIO m) => GetTableInfo_ -> m EncJSON

Legacy data connector command. This doesn't use the DataConnector ScalarType to represent types.

data GetTableInfo (b :: BackendType)

Constructors

GetTableInfo

Fields

_gtiSourceName :: SourceName

_gtiTableName :: TableName b

runGetTableInfo :: forall b m. (BackendMetadata b, CacheRM m, MonadError QErr m, MetadataM m, MonadBaseControl IO m, MonadIO m) => GetTableInfo b -> m EncJSON

Get information about the given table.