Hasura.RQL.DDL.Schema.Table

Description

Documentation

data TrackTable b

Constructors

TrackTable

Fields

tSource :: !SourceName

tName :: !(TableName b)

tIsEnum :: !Bool

data SetTableIsEnum

Constructors

SetTableIsEnum

Fields

stieSource :: !SourceName

stieTable :: !QualifiedTable

stieIsEnum :: !Bool

data UntrackTable b

Constructors

UntrackTable

Fields

utSource :: !SourceName

utTable :: !(TableName b)

utCascade :: !Bool

isTableTracked :: forall b. Backend b => SourceInfo b -> TableName b -> Bool

trackExistingTableOrViewP1 :: forall b m. (QErrM m, CacheRWM m, Backend b, MetadataM m) => SourceName -> TableName b -> m ()

Track table/view, Phase 1: Validate table tracking operation. Fails if table is already being tracked, or if a function with the same name is being tracked.

checkConflictingNode :: forall m. MonadError QErr m => SchemaCache -> Text -> m ()

Check whether a given name would conflict with the current schema by doing an internal introspection

trackExistingTableOrViewP2 :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => SourceName -> TableName b -> Bool -> TableConfig b -> m EncJSON

runTrackTableQ :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => TrackTable b -> m EncJSON

data TrackTableV2 b

Constructors

TrackTableV2

Fields

ttv2Table :: !(TrackTable b)

ttv2Configuration :: !(TableConfig b)

runTrackTableV2Q :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => TrackTableV2 b -> m EncJSON

runSetExistingTableIsEnumQ :: (MonadError QErr m, CacheRWM m, MetadataM m) => SetTableIsEnum -> m EncJSON

data SetTableCustomization b

Constructors

SetTableCustomization

Fields

_stcSource :: !SourceName

_stcTable :: !(TableName b)

_stcConfiguration :: !(TableConfig b)

data SetTableCustomFields

Constructors

SetTableCustomFields

Fields

_stcfSource :: !SourceName

_stcfTable :: !QualifiedTable

_stcfCustomRootFields :: !TableCustomRootFields

_stcfCustomColumnNames :: !(CustomColumnNames ('Postgres 'Vanilla))

runSetTableCustomFieldsQV2 :: (QErrM m, CacheRWM m, MetadataM m) => SetTableCustomFields -> m EncJSON

runSetTableCustomization :: forall b m. (QErrM m, CacheRWM m, MetadataM m, Backend b, BackendMetadata b) => SetTableCustomization b -> m EncJSON

unTrackExistingTableOrViewP1 :: forall b m. (CacheRM m, QErrM m, Backend b) => UntrackTable b -> m ()

unTrackExistingTableOrViewP2 :: forall b m. (CacheRWM m, QErrM m, MetadataM m, BackendMetadata b) => UntrackTable b -> m EncJSON

runUntrackTableQ :: forall b m. (CacheRWM m, QErrM m, MetadataM m, BackendMetadata b) => UntrackTable b -> m EncJSON

buildTableCache :: forall arr m b. (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectedInfo) arr, ArrowCache m arr, MonadIO m, MonadBaseControl IO m, BackendMetadata b) => (SourceName, SourceConfig b, DBTablesMetadata b, [TableBuildInput b], Dependency InvalidationKey) `arr` HashMap (TableName b) (TableCoreInfoG b (ColumnInfo b) (ColumnInfo b))

Builds an initial TableCache ColumnInfo from catalog information. Does not fill in _tiRolePermInfoMap or _tiEventTriggerInfoMap at all, and _tiFieldInfoMap only contains columns, not relationships; those pieces of information are filled in by later stages.