Hasura.RQL.DDL.Webhook.Transform.Body

Body Transformations

data Body

HTTP message body being transformed.

Constructors

JSONBody (Maybe Value)

RawBody ByteString

data family TransformFn a

The associated type 'TransformFn a' is the defunctionalized version of some transformation that should be applied to a given request field.

In most cases it is some variation on a piece of template text describing the transformation.

data family TransformCtx a

data BodyTransformFn

The transformations which can be applied to an HTTP message body.

Constructors

Remove
Remove the HTTP message body.

ModifyAsJSON Template
Modify the JSON message body by applying a Template transformation.

ModifyAsFormURLEncoded (HashMap Text UnescapedTemplate)
Modify the JSON message body by applying UnescapedTemplate transformations to each field with a matching Text key.

foldFormEncoded :: HashMap Text ByteString -> ByteString

Fold a HashMap of header key/value pairs into an x-www-form-urlencoded message body.

validateBodyTransformFn :: TemplatingEngine -> BodyTransformFn -> Validation TransformErrorBundle ()

Validate that the provided BodyTransformFn is correct in the context of a particular TemplatingEngine.

This is a product of the fact that the correctness of a given transformation may be dependent on zero, one, or more of the templated transformations encoded within the given BodyTransformFn.

Orphan instances

Transform Body

Instance details

Methods

transform :: MonadError TransformErrorBundle m => TransformFn Body -> TransformCtx Body -> Body -> m Body

validate :: TemplatingEngine -> TransformFn Body -> Validation TransformErrorBundle ()