Hasura.RQL.DDL.Webhook.Transform.Class

Description

The Transform typeclass with various types and helper functions for evaluating transformations.

Documentation

class Transform a where

Transform describes how to reify a defunctionalized transformation for a particular request field.

Associated Types

data TransformFn a :: Type

The associated type 'TransformFn a' is the defunctionalized version of some transformation that should be applied to a given request field.

In most cases it is some variation on a piece of template text describing the transformation.

data TransformCtx a :: Type

Methods

transform :: MonadError TransformErrorBundle m => TransformFn a -> TransformCtx a -> a -> m a

transform is a function which takes TransformFn of a and reifies it into a function of the form:

  ReqTransformCtx -> a -> m a

validate :: TemplatingEngine -> TransformFn a -> Validation TransformErrorBundle ()

Validate a TransformFn of a.

newtype TransformErrorBundle

We use collect all transformation failures as a '[J.Value]'.

Constructors

TransformErrorBundle

Fields

tebMessages :: [Value]

throwErrorBundle :: MonadError TransformErrorBundle m => Text -> Maybe Value -> m a

A helper function for serializing transformation errors to JSON.

data TemplatingEngine

Available templating engines.

Constructors

Kriti

newtype Template

Textual transformation template.

Constructors

Template

Fields

unTemplate :: Text

newtype UnescapedTemplate

Validated textual transformation template /for string interpolation only/.

This is necessary due to Kriti not distinguishing between string literals and string templates.

Constructors

UnescapedTemplate

Fields

getUnescapedTemplate :: Text

wrapUnescapedTemplate :: UnescapedTemplate -> Template

Wrap an UnescapedTemplate with escaped double quotes.

encodeScalar :: MonadError TransformErrorBundle m => Value -> m ByteString

Encode a JSON Scalar Value as a ByteString. If a non-Scalar value is provided, will return a TrnasformErrorBundle