Hasura.RQL.DDL.Webhook.Transform.Headers

Documentation

newtype Headers

The actual header data we are transforming.

This newtype is necessary because otherwise we end up with an orphan instance.

Constructors

Headers [Header]

newtype HeadersTransformAction

The defunctionalized transformation on Headers

Constructors

ReplaceHeaders ReplaceHeaderFields

data ReplaceHeaderFields

The user can supply a set of header keys to be filtered from the request and a set of headers to be added to the request.

Constructors

ReplaceHeaderFields

Fields

rhf_addHeaders :: [(CI Text, UnescapedTemplate)]

rhf_removeHeaders :: [CI Text]

newtype HeaderKey

This newtype exists solely to anchor a FromJSON instance and is eliminated in the TransformHeaders FromJSON instance.

Constructors

HeaderKey

Fields

unHeaderKey :: CI Text