Hasura.RQL.DDL.Webhook.Transform.Response

Description

Types and subroutines related to constructing transformations on HTTP responses.

Documentation

data ResponseTransformCtx

Common context that is made available to all response transformations.

Constructors

ResponseTransformCtx

Fields

responseTransformBody :: Value

responseTransformReqCtx :: Value

responseSessionVariables :: Maybe SessionVariables

responseTransformEngine :: TemplatingEngine

responseStatusCode :: Int

runResponseTemplateTransform :: Template -> ResponseTransformCtx -> Either TransformErrorBundle Value

A helper function for executing transformations from a Template and a ResponseTransformCtx.

NOTE: This and all related funtions are hard-coded to Kriti at the moment. When we add additional template engines this function will need to take a TemplatingEngine parameter.

runUnescapedResponseTemplateTransform :: ResponseTransformCtx -> UnescapedTemplate -> Either TransformErrorBundle ByteString

Run an UnescapedTemplate with a ResponseTransformCtx.

runUnescapedResponseTemplateTransform' :: ResponseTransformCtx -> UnescapedTemplate -> Validation TransformErrorBundle ByteString

Run an UnescapedTemplate with a ResponseTransformCtx in Validation.