Hasura.RQL.DDL.WebhookTransforms

Documentation

data ReqTransformCtx

Constructors

ReqTransformCtx

Fields

tcUrl :: Maybe Value

tcBody :: Value

tcSessionVars :: Value

tcQueryParams :: Maybe Value

data RespTransformCtx

Constructors

RespTransformCtx

Fields

rtcBody :: Value

rtcReqCtx :: Value

data RequestTransform

A set of data transformation functions and substitutions generated from a MetadataRequestTransform. Nothing values mean use the original request value, unless otherwise indicated.

Constructors

RequestTransform

Fields

reqTransformRequestMethod :: Maybe (ReqTransformCtx -> Either TransformErrorBundle RequestMethod)

Change the request method to one provided here. Nothing means POST.

reqTransformRequestURL :: Maybe (ReqTransformCtx -> Either TransformErrorBundle Text)

A function which contructs a new URL given the request context.

reqTransformBody :: Maybe (RemoveOrTransform (ReqTransformCtx -> Either TransformErrorBundle Value))

A function for transforming the request body.

reqTransformQueryParams :: Maybe (ReqTransformCtx -> Either TransformErrorBundle Query)

A function which contructs new query parameters given the request context.

reqTransformRequestHeaders :: Maybe (ReqTransformCtx -> Either TransformErrorBundle ([Header] -> [Header]))

A function which contructs new Headers given the request context.

newtype ResponseTransform

A set of data transformation functions generated from a MetadataResponseTransform. Nothing means use the original response value.

Constructors

ResponseTransform

Fields

respTransformBody :: Maybe (RemoveOrTransform (RespTransformCtx -> Either TransformErrorBundle Value))

data MetadataRequestTransform

A de/serializable request transformation template which can be stored in the metadata associated with an actionevent triggeretc. and used to produce a RequestTransform.

NOTE: This data type is _only_ intended to be parsed from and stored as user metadata; no direct logical transformations should be done upon it.

NOTE: Users should convert this to RequestTransform as close as possible to the call site that performs a transformed HTTP request, as RequestTransform has a representation that makes it more difficult to deserialize for debugging.

Nothing values mean use the original request value, unless otherwise indicated.

Constructors

MetadataRequestTransform

Fields

mtVersion :: Version

The Schema Version

mtRequestMethod :: Maybe StringTemplateText

Change the request method to one provided here. Nothing means POST.

mtRequestURL :: Maybe StringTemplateText

Template script for transforming the URL.

mtBodyTransform :: Maybe (RemoveOrTransform TemplateText)

Template script for transforming the request body.

mtQueryParams :: Maybe [(StringTemplateText, Maybe StringTemplateText)]

A list of template scripts for constructing new Query Params.

mtRequestHeaders :: Maybe TransformHeaders

A list of template scripts for constructing headers.

mtTemplatingEngine :: TemplatingEngine

The template engine to use for transformations. Default: Kriti

data MetadataResponseTransform

Constructors

MetadataResponseTransform

Fields

mrtVersion :: Version

The Schema Version

mrtBodyTransform :: Maybe (RemoveOrTransform TemplateText)

Template script for transforming the response body.

mrtTemplatingEngine :: TemplatingEngine

The template engine to use for transformations. Default: Kriti

data RemoveOrTransform a

Constructors

Remove

Transform a

newtype RequestMethod

Constructors

RequestMethod (CI Text)

data TemplatingEngine

Available Template Languages

Constructors

Kriti

renderTemplatingEngine :: TemplatingEngine -> Text

data Version

Constructors

V1

V2

newtype TemplateText

Unparsed Kriti templating code

Constructors

TemplateText

Fields

unTemplateText :: Text

newtype StringTemplateText

Unparsed Kriti templating code for string interpolations only. NOTE: WE are highly coupled to Kriti at this point and will likely need to rethink Kriti string interpolation when we start adding other template engines.

Constructors

StringTemplateText

Fields

unStringTemplateText :: Text

newtype HeaderKey

This newtype exists solely to anchor a FromJSON instance and is eliminated in the TransformHeaders FromJSON instance.

Constructors

HeaderKey

Fields

unHeaderKey :: CI Text

data TransformHeaders

Constructors

TransformHeaders

Fields

addHeaders :: [(CI Text, StringTemplateText)]

removeHeaders :: [CI Text]

newtype TransformErrorBundle

Constructors

TransformErrorBundle

Fields

teMessages :: [Value]

mkRequestTransform :: MetadataRequestTransform -> RequestTransform

Construct a RequestTransform from its metadata representation.

mkResponseTransform :: MetadataResponseTransform -> ResponseTransform

transformST :: TemplatingEngine -> ReqTransformCtx -> StringTemplateText -> Either TransformErrorBundle ByteString

Transform a String Template

mkMethodTransform :: TemplatingEngine -> StringTemplateText -> ReqTransformCtx -> Either TransformErrorBundle RequestMethod

mkQueryParamsTransform :: TemplatingEngine -> [(StringTemplateText, Maybe StringTemplateText)] -> ReqTransformCtx -> Either TransformErrorBundle Query

mkHeaderTransform :: TemplatingEngine -> TransformHeaders -> ReqTransformCtx -> Either TransformErrorBundle ([Header] -> [Header])

Given a TransformHeaders and the ReqTransformCtx, Construct a function to transform the existing headers.

mkUrlTransform :: TemplatingEngine -> StringTemplateText -> ReqTransformCtx -> Either TransformErrorBundle Text

mkReqTemplateTransform :: TemplatingEngine -> TemplateText -> ReqTransformCtx -> Either TransformErrorBundle Value

Construct a Template Transformation function for Requests

mkRespTemplateTransform :: TemplatingEngine -> TemplateText -> RespTransformCtx -> Either TransformErrorBundle Value

Construct a Template Transformation function for Responses

buildReqTransformCtx :: Text -> Maybe SessionVariables -> Request -> ReqTransformCtx

buildRespTransformCtx :: ReqTransformCtx -> ByteString -> RespTransformCtx

applyRequestTransform :: (Request -> ReqTransformCtx) -> RequestTransform -> Request -> Either TransformErrorBundle Request

Transform an Request with a RequestTransform.

Note: we pass in the request url explicitly for use in the ReqTransformCtx. We do this so that we can ensure that the url is syntactically identical to what the use submits. If we use the parsed request from the Request term then it is possible that the url is semantically equivalent but syntactically different. An example of this is the presence or lack of a trailing slash on the URL path. This important when performing string interpolation on the request url.

Because of this requirement, we must ensure the sequence of transformation applications. If a query param transformation occured before the url transformation then the query param transformation would be wiped out. applyRequestTransform :: T.Text -> RequestTransform -> HTTP.Request -> Maybe SessionVariables -> Either TransformErrorBundle HTTP.Request applyRequestTransform reqUrl RequestTransform {..} reqData sessionVars =

applyResponseTransform :: ResponseTransform -> RespTransformCtx -> Either TransformErrorBundle ByteString

At the moment we only transform the body of Responses. 'http-client' does not export the constructors for Response. If we want to transform then we will need additional apply functions.

valueToString :: MonadError TransformErrorBundle m => Value -> m ByteString

throwErrorBundle :: MonadError TransformErrorBundle m => Text -> Maybe Value -> m a

(.=?) :: ToJSON v => k -> Maybe v -> Maybe (k, Value) infixr 8

wrapTemplate :: StringTemplateText -> StringTemplateText

Wrap a template with escaped double quotes.