Hasura.RQL.IR.Action

Documentation

data ActionFieldG (r :: Type)

Internal representation for a selection of fields on the result of an action. Type parameter r will be either r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

Constructors

ACFScalar Name
Scalar value. G.Name is the original field name from the object type.

ACFRemote (ActionRemoteRelationshipSelect r)
Remote relationship

ACFExpression Text
Constant text value (used for __typename fields)

ACFNestedObject Name (ActionFieldsG r)
Nested object. G.Name is the original field name from the object type.

type ActionFieldsG r = Fields (ActionFieldG r)

type ActionFields = ActionFieldsG Void

data ActionRemoteRelationshipSelect r

Constructors

ActionRemoteRelationshipSelect

Fields

_arrsLHSJoinFields :: HashMap FieldName Name

The fields on the table that are required for the join condition of the remote relationship

_arrsRelationship :: r

The field that captures the relationship r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

_ACFNestedObject :: forall r. Prism' (ActionFieldG r) (Name, ActionFieldsG r)

_ACFExpression :: forall r. Prism' (ActionFieldG r) Text

_ACFRemote :: forall r. Prism' (ActionFieldG r) (ActionRemoteRelationshipSelect r)

_ACFScalar :: forall r. Prism' (ActionFieldG r) Name

data AnnActionExecution (r :: Type)

Constructors

AnnActionExecution

Fields

_aaeName :: ActionName

_aaeOutputType :: GraphQLType

output type

_aaeFields :: ActionFieldsG r

output selection

_aaePayload :: Value

jsonified input arguments

_aaeOutputFields :: ActionOutputFields

to validate the response fields from webhook

_aaeWebhook :: EnvRecord ResolvedWebhook

_aaeHeaders :: [HeaderConf]

_aaeForwardClientHeaders :: Bool

_aaeTimeOut :: Timeout

_aaeRequestTransform :: Maybe RequestTransform

_aaeResponseTransform :: Maybe MetadataResponseTransform

type ActionOutputFields = HashMap Name GType

getActionOutputFields :: AnnotatedOutputType -> ActionOutputFields

data AnnActionMutationAsync

Constructors

AnnActionMutationAsync

Fields

_aamaName :: ActionName

_aamaForwardClientHeaders :: Bool

_aamaPayload :: Value

jsonified input arguments

data AsyncActionQueryFieldG (r :: Type)

Constructors

AsyncTypename Text

AsyncOutput (ActionFieldsG r)

AsyncId

AsyncCreatedAt

AsyncErrors

type AsyncActionQueryFieldsG r = Fields (AsyncActionQueryFieldG r)

data AnnActionAsyncQuery (b :: BackendType) (r :: Type)

Constructors

AnnActionAsyncQuery

Fields

_aaaqName :: ActionName

_aaaqActionId :: ActionId

_aaaqOutputType :: GraphQLType

_aaaqFields :: AsyncActionQueryFieldsG r

_aaaqDefinitionList :: [(Column b, ScalarType b)]

_aaaqStringifyNum :: StringifyNumbers

_aaaqForwardClientHeaders :: Bool

_aaaqSource :: ActionSourceInfo b

data ActionSourceInfo b

Constructors

ASINoSource
No relationships defined on the action output object

ASISource SourceName (SourceConfig b)
All relationships refer to tables in one source

aaaqStringifyNum :: forall b r. Lens' (AnnActionAsyncQuery b r) StringifyNumbers

aaaqSource :: forall b r. Lens' (AnnActionAsyncQuery b r) (ActionSourceInfo b)

aaaqOutputType :: forall b r. Lens' (AnnActionAsyncQuery b r) GraphQLType

aaaqName :: forall b r. Lens' (AnnActionAsyncQuery b r) ActionName

aaaqForwardClientHeaders :: forall b r. Lens' (AnnActionAsyncQuery b r) Bool

aaaqFields :: forall b r r. Lens (AnnActionAsyncQuery b r) (AnnActionAsyncQuery b r) (AsyncActionQueryFieldsG r) (AsyncActionQueryFieldsG r)

aaaqDefinitionList :: forall b r. Lens' (AnnActionAsyncQuery b r) [(Column b, ScalarType b)]

aaaqActionId :: forall b r. Lens' (AnnActionAsyncQuery b r) ActionId

aaeWebhook :: forall r. Lens' (AnnActionExecution r) (EnvRecord ResolvedWebhook)

aaeTimeOut :: forall r. Lens' (AnnActionExecution r) Timeout

aaeResponseTransform :: forall r. Lens' (AnnActionExecution r) (Maybe MetadataResponseTransform)

aaeRequestTransform :: forall r. Lens' (AnnActionExecution r) (Maybe RequestTransform)

aaePayload :: forall r. Lens' (AnnActionExecution r) Value

aaeOutputType :: forall r. Lens' (AnnActionExecution r) GraphQLType

aaeOutputFields :: forall r. Lens' (AnnActionExecution r) ActionOutputFields

aaeName :: forall r. Lens' (AnnActionExecution r) ActionName

aaeHeaders :: forall r. Lens' (AnnActionExecution r) [HeaderConf]

aaeForwardClientHeaders :: forall r. Lens' (AnnActionExecution r) Bool

aaeFields :: forall r r. Lens (AnnActionExecution r) (AnnActionExecution r) (ActionFieldsG r) (ActionFieldsG r)

_AsyncErrors :: forall r. Prism' (AsyncActionQueryFieldG r) ()

_AsyncCreatedAt :: forall r. Prism' (AsyncActionQueryFieldG r) ()

_AsyncId :: forall r. Prism' (AsyncActionQueryFieldG r) ()

_AsyncOutput :: forall r r. Prism (AsyncActionQueryFieldG r) (AsyncActionQueryFieldG r) (ActionFieldsG r) (ActionFieldsG r)

_AsyncTypename :: forall r. Prism' (AsyncActionQueryFieldG r) Text