Hasura.RQL.IR.BoolExp

Description

Boolean Expressions

This module defines the IR representation of boolean expressions used in _where clauses in GraphQL queries, permissions, and so on.

The types in this module define a generic structure with "holes" to be filled by each backend. Specifically, holes will include things like types for table names, and backend field types.

Documentation

data GBoolExp (backend :: BackendType) field

This type represents a boolean expression tree. It is parametric over the actual implementation of the actual boolean term values. It nonetheless leaks some information: "exists" is only used in permissions, to add conditions based on another table.

The backend parameter is used to find the backend-specific type for table names in the BoolExists constructor.

The field type represent the type of database-specific field types.

Constructors

BoolAnd [GBoolExp backend field]

BoolOr [GBoolExp backend field]

BoolNot (GBoolExp backend field)

BoolExists (GExists backend field)
Represents a condition on an aribtrary table. since the backend and field are the same, the table must be of the same database type.

BoolField field
A column field

gBoolExpTrue :: GBoolExp backend field

A default representation for a true boolean value.

data GExists (backend :: BackendType) field

Represents a condition on an aribtrary table. Used as part of our permissions boolean expressions. See our documentation for more information: https://hasura.io/docs/latest/graphql/core/auth/authorization/permission-rules.html#using-unrelated-tables-views

Constructors

GExists

Fields

_geTable :: TableName backend

_geWhere :: GBoolExp backend field

geWhere :: forall backend field field. Lens (GExists backend field) (GExists backend field) (GBoolExp backend field) (GBoolExp backend field)

geTable :: forall backend field. Lens' (GExists backend field) (TableName backend)

data ColExp

We don't allow conditions across relationships in permissions: the type we use as the terms in GBoolExp is this one, ColExp, which only contains a FieldName and a JSON Value.

Constructors

ColExp

Fields

ceCol :: FieldName

ceVal :: Value

newtype BoolExp (b :: BackendType)

This BoolExp type is a simple alias for the boolean expressions used in permissions, that uses ColExp as the term in GBoolExp.

Constructors

BoolExp

Fields

unBoolExp :: GBoolExp b ColExp

_BoolExists :: forall backend field. Prism' (GBoolExp backend field) (GExists backend field)

data PartialSQLExp (backend :: BackendType)

Permissions get translated into boolean expressions that are threaded throuhgout the parsers. For the leaf values of those permissions, we use this type, which references but doesn't inline the session variables.

Constructors

PSESessVar (SessionVarType backend) SessionVariable

PSESession

PSESQLExp (SQLExpression backend)

isStaticValue :: PartialSQLExp backend -> Bool

hasStaticExp :: Backend b => OpExpG b (PartialSQLExp b) -> Bool

type CastExp backend field = HashMap (ScalarType backend) [OpExpG backend field]

Operand for cast operator

data OpExpG (backend :: BackendType) field

This type represents the boolean operators that can be applied on values of a column. This type only contains the common core, that we expect to be ultimately entirely supported in most if not all backends. Backends can extend this with the BooleanOperators type in Backend.

Constructors

ACast (CastExp backend field)

AEQ Bool field

ANE Bool field

AIN field

ANIN field

AGT field

ALT field

AGTE field

ALTE field

ALIKE field

ANLIKE field

CEQ (RootOrCurrentColumn backend)

CNE (RootOrCurrentColumn backend)

CGT (RootOrCurrentColumn backend)

CLT (RootOrCurrentColumn backend)

CGTE (RootOrCurrentColumn backend)

CLTE (RootOrCurrentColumn backend)

ANISNULL

ANISNOTNULL

ABackendSpecific (BooleanOperators backend field)

data RootOrCurrentColumn b

Constructors

RootOrCurrentColumn RootOrCurrent (Column b)

data RootOrCurrent

The arguments of column-operators may refer to either the so-called 'root tabular value' or 'current tabular value'.

Constructors

IsRoot

IsCurrent

opExpDepCol :: OpExpG backend field -> Maybe (RootOrCurrentColumn backend)

data ComputedFieldBoolExp (backend :: BackendType) scalar

This type is used to represent the kinds of boolean expression used for compouted fields based on the return type of the SQL function.

Constructors

CFBEScalar [OpExpG backend scalar]
SQL function returning a scalar

CFBETable (TableName backend) (AnnBoolExp backend scalar)
SQL function returning SET OF table

data AnnComputedFieldBoolExp (backend :: BackendType) scalar

Using a computed field in boolean expression. Example: A computed field "full_name" ("first_name" || "last_name") is defined to the "user" table. Boolean expression to filter whose "full_name" is LIKE "%bob%" query { user(where: {full_name: {_like: "%bob%"}}){ id first_name last_name full_name } } Limitation: We only support computed fields in boolean expressions when they are functions with no input arguments, because it is complex to generate schema for where clauses for functions that have input arguments.

Constructors

AnnComputedFieldBoolExp

Fields

_acfbXFieldInfo :: XComputedField backend

_acfbName :: ComputedFieldName

_acfbFunction :: FunctionName backend

_acfbFunctionArgsExp :: FunctionArgsExp backend scalar

_acfbBoolExp :: ComputedFieldBoolExp backend scalar

data AnnBoolExpFld (backend :: BackendType) leaf

This type is used for boolean terms in GBoolExp in the schema; there are four kinds boolean terms: - operators on a column of the current table, using the OpExpG kind of operators - arbitrary expressions on columns of tables in relationships (in the same source) - A computed field of the current table - aggregation operations on array relationships on the current tables.

This type is parameterized over the type of leaf values, the values on which we operate.

Constructors

AVColumn (ColumnInfo backend) [OpExpG backend leaf]

AVRelationship (RelInfo backend) (AnnBoolExp backend leaf)

AVComputedField (AnnComputedFieldBoolExp backend leaf)

AVAggregationPredicates (AggregationPredicates backend leaf)

type AnnBoolExp backend scalar = GBoolExp backend (AnnBoolExpFld backend scalar)

A simple alias for the kind of boolean expressions used in the schema, that ties together GBoolExp, OpExpG, and AnnBoolExpFld.

type AnnBoolExpFldSQL backend = AnnBoolExpFld backend (SQLExpression backend)

type AnnBoolExpSQL backend = AnnBoolExp backend (SQLExpression backend)

type AnnBoolExpPartialSQL backend = AnnBoolExp backend (PartialSQLExp backend)

annBoolExpTrue :: AnnBoolExp backend scalar

andAnnBoolExps :: AnnBoolExp backend scalar -> AnnBoolExp backend scalar -> AnnBoolExp backend scalar

data DWithinGeomOp field

Operand for STDWithin opoerator

Constructors

DWithinGeomOp

Fields

dwgeomDistance :: field

dwgeomFrom :: field

data DWithinGeogOp field

Operand for STDWithin opoerator

Constructors

DWithinGeogOp

Fields

dwgeogDistance :: field

dwgeogFrom :: field

dwgeogUseSpheroid :: field

data STIntersectsNbandGeommin field

Operand for STIntersect

Constructors

STIntersectsNbandGeommin

Fields

singNband :: field

singGeommin :: field

data STIntersectsGeomminNband field

Operand for STIntersect

Constructors

STIntersectsGeomminNband

Fields

signGeommin :: field

signNband :: Maybe field

newtype AnnColumnCaseBoolExpField (backend :: BackendType) field

This is a simple newtype over AnnBoolExpFld. At time of writing, I do not know why we want this, and why it exists. It might be a relic of a needed differentiation, now lost? TODO: can this be removed?

Constructors

AnnColumnCaseBoolExpField

Fields

_accColCaseBoolExpField :: AnnBoolExpFld backend field

type AnnColumnCaseBoolExp b a = GBoolExp b (AnnColumnCaseBoolExpField b a)

Similar to AnnBoolExp, this type alias ties together GBoolExp, OpExpG, and AnnColumnCaseBoolExpFld.

type AnnColumnCaseBoolExpPartialSQL b = AnnColumnCaseBoolExp b (PartialSQLExp b)

type PreSetColsG b v = HashMap (Column b) v

type PreSetColsPartial b = HashMap (Column b) (PartialSQLExp b)