Hasura.RQL.IR.Root

Documentation

data SourceConfigWith (db :: BackendType -> Type) (b :: BackendType)

Constructors

SourceConfigWith (SourceConfig b) (Maybe QueryTagsConfig) (db b)

data RootField (db :: BackendType -> Type) remote action raw where

Constructors

RFDB :: SourceName -> AnyBackend (SourceConfigWith db) -> RootField db remote action raw

RFRemote :: remote -> RootField db remote action raw

RFAction :: action -> RootField db remote action raw

RFRaw :: raw -> RootField db remote action raw

data MutationDB (b :: BackendType) (r :: Type) v

Constructors

MDBInsert (AnnotatedInsert b r v)

MDBUpdate (AnnotatedUpdateG b r v)

MDBDelete (AnnDelG b r v)

MDBFunction JsonAggSelect (AnnSimpleSelectG b r v)
This represents a VOLATILE function, and is AnnSimpleSelG for easy re-use of non-VOLATILE function tracking code.

data ActionQuery (r :: Type)

Constructors

AQQuery !(AnnActionExecution r)

AQAsync !(AnnActionAsyncQuery ('Postgres 'Vanilla) r)

data ActionMutation (r :: Type)

Constructors

AMSync !(AnnActionExecution r)

AMAsync !AnnActionMutationAsync

newtype QueryDBRoot r v b

Constructors

QDBR (QueryDB b r (v b))

newtype MutationDBRoot r v b

Constructors

MDBR (MutationDB b r (v b))

data RemoteRelationshipField vf

IR of a remote relationship. A remote relationship currently can be to either a remote schema or a database's table. See RemoteSourceSelect for explanation on vf.

Constructors

RemoteSchemaField (RemoteSchemaSelect (RemoteRelationshipField vf))

RemoteSourceField (AnyBackend (RemoteSourceSelect (RemoteRelationshipField vf) vf))
AnyBackend is used here to capture a relationship to an arbitrary target

type QueryActionRoot v = ActionQuery (RemoteRelationshipField v)

Represents a query root field to an action

type MutationActionRoot v = ActionMutation (RemoteRelationshipField v)

Represents a mutation root field to an action

type QueryRootField v = RootField (QueryDBRoot (RemoteRelationshipField v) v) (RemoteSchemaRootField (RemoteRelationshipField v) RemoteSchemaVariable) (QueryActionRoot v) Value

type MutationRootField v = RootField (MutationDBRoot (RemoteRelationshipField v) v) (RemoteSchemaRootField (RemoteRelationshipField v) RemoteSchemaVariable) (MutationActionRoot v) Value

type SubscriptionRootField v = RootField (QueryDBRoot (RemoteRelationshipField v) v) (RemoteSchemaRootField (RemoteRelationshipField v) RemoteSchemaVariable) (QueryActionRoot v) Value