Hasura.RQL.IR.Select

Description

This modules defines the tree of Select types: how we represent a query internally, from its top level QueryDB down to each individual field. Most of those types have three type arguments:

b: BackendType The backend that is targeted by that specific select (Postgres Vanilla, MSSQL...); we use the type families in the Backend class to decide how different parts of the IR are represented in different backends.

v: Type The type of the leaf values in our AST; used almost exclusively for column values, over which queries can be parameterized. The output of the parser phase will use UnpreparedValue b for the leaves, and most backends will then transform the AST to interpret those values and consequently change v to be SQLExpression b

r: BackendType -> Type Joins across backends mean that the aforementioned b parameter won't be the same throughout the entire tree; at some point we will have an AnyBackend used to encapsulate a branch that uses a different b. We still want, however, to be able to parameterize the values of the leaves in that separate branch, and that's what the r parameter is for. We also use UnpreparedValue here during the parsing phase, meaning all leaf values will be UnpreparedValue b for their respective backend b, and most backends will then transform their AST, cutting all such remote branches, and therefore using Const Void for r.

Documentation

data QueryDB (b :: BackendType) (r :: Type) v

Constructors

QDBMultipleRows (AnnSimpleSelectG b r v)

QDBSingleRow (AnnSimpleSelectG b r v)

QDBAggregation (AnnAggregateSelectG b r v)

QDBConnection (ConnectionSelect b r v)

QDBStreamMultipleRows (AnnSimpleStreamSelectG b r v)

type AnnSimpleSelectG b r v = AnnSelectG b (AnnFieldG b r) v

type AnnAggregateSelectG b r v = AnnSelectG b (TableAggregateFieldG b r) v

type AnnSimpleStreamSelectG b r v = AnnSelectStreamG b (AnnFieldG b r) v

type AnnSimpleSelect b = AnnSimpleSelectG b Void (SQLExpression b)

type AnnAggregateSelect b = AnnAggregateSelectG b Void (SQLExpression b)

type AnnSimpleStreamSelect b = AnnSimpleStreamSelectG b Void (SQLExpression b)

data ConnectionSelect (b :: BackendType) (r :: Type) v

Constructors

ConnectionSelect

Fields

_csXRelay :: XRelay b

_csPrimaryKeyColumns :: PrimaryKeyColumns b

_csSplit :: Maybe (NonEmpty (ConnectionSplit b v))

_csSlice :: Maybe ConnectionSlice

_csSelect :: AnnSelectG b (ConnectionField b r) v

data ConnectionSplit (b :: BackendType) v

Constructors

ConnectionSplit

Fields

_csKind :: ConnectionSplitKind

_csValue :: v

_csOrderBy :: OrderByItemG b (AnnotatedOrderByElement b v)

data ConnectionSlice

Constructors

SliceFirst Int

SliceLast Int

data ConnectionSplitKind

Constructors

CSKBefore

CSKAfter

data RemoteRelationshipSelect b r

captures a remote relationship's selection and the necessary context

Constructors

RemoteRelationshipSelect

Fields

_rrsLHSJoinFields :: HashMap FieldName (DBJoinField b)

The fields on the table that are required for the join condition of the remote relationship

_rrsRelationship :: r

The field that captures the relationship r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

data AnnFieldG (b :: BackendType) (r :: Type) v

Constructors

AFColumn (AnnColumnField b v)

AFObjectRelation (ObjectRelationSelectG b r v)

AFArrayRelation (ArraySelectG b r v)

AFComputedField (XComputedField b) ComputedFieldName (ComputedFieldSelect b r v)

AFRemote (RemoteRelationshipSelect b r)
A remote relationship field

AFNodeId (XRelay b) SourceName (TableName b) (PrimaryKeyColumns b)

AFExpression Text

AFNestedObject (AnnNestedObjectSelectG b r v)
Nested object.

AFNestedArray (XNestedObjects b) (AnnNestedArraySelectG b r v)
Nested array

type AnnField b = AnnFieldG b Void (SQLExpression b)

type AnnFields b = AnnFieldsG b Void (SQLExpression b)

mkAnnColumnField :: Column backend -> ColumnType backend -> AnnRedactionExp backend v -> Maybe (ScalarSelectionArguments backend) -> AnnFieldG backend r v

mkAnnColumnFieldAsText :: ColumnInfo backend -> AnnFieldG backend r v

traverseSourceRelationshipSelection :: (Applicative f, Backend backend) => (vf backend -> f (vg backend)) -> SourceRelationshipSelection backend r vf -> f (SourceRelationshipSelection backend r vg)

data TableAggregateFieldG (b :: BackendType) (r :: Type) v

Constructors

TAFAgg (AggregateFields b v)

TAFNodes (XNodesAgg b) (AnnFieldsG b r v)

TAFGroupBy (XGroupBy b) (GroupByG b r v)

TAFExp Text

data AggregateField (b :: BackendType) v

Constructors

AFCount (CountType b v)

AFOp (AggregateOp b v)

AFExp Text

data AggregateOp (b :: BackendType) v

Constructors

AggregateOp

Fields

_aoOp :: Text

_aoFields :: SelectionFields b v

data GroupByG (b :: BackendType) r v

Constructors

GroupByG

Fields

_gbgKeys :: [GroupKeyField b]

_gbgFields :: Fields (GroupByField b r v)

data GroupByField (b :: BackendType) r v

Constructors

GBFGroupKey (Fields (GroupKeyField b))

GBFAggregate (AggregateFields b v)

GBFNodes (AnnFieldsG b r v)

GBFExp Text

data GroupKeyField (b :: BackendType)

Constructors

GKFColumn (Column b)

GKFExp Text

data SelectionField (b :: BackendType) v

Types of fields that can be selected in a user query.

Constructors

SFCol

Fields

(Column b)

(ColumnType b)

(AnnRedactionExp b v)

This type is used to determine whether the column should be redacted before being aggregated

SFComputedField ComputedFieldName (ComputedFieldScalarSelect b v)

SFExp Text

type TableAggregateField b = TableAggregateFieldG b Void (SQLExpression b)

type TableAggregateFields b = TableAggregateFieldsG b Void (SQLExpression b)

type TableAggregateFieldsG b r v = Fields (TableAggregateFieldG b r v)

type SelectionFields b v = Fields (SelectionField b v)

type AggregateFields b v = Fields (AggregateField b v)

type AnnFieldsG b r v = Fields (AnnFieldG b r v)

data ConnectionField (b :: BackendType) (r :: Type) v

Constructors

ConnectionTypename Text

ConnectionPageInfo PageInfoFields

ConnectionEdges (EdgeFields b r v)

data PageInfoField

Constructors

PageInfoTypename Text

PageInfoHasNextPage

PageInfoHasPreviousPage

PageInfoStartCursor

PageInfoEndCursor

data EdgeField (b :: BackendType) (r :: Type) v

Constructors

EdgeTypename Text

EdgeCursor

EdgeNode (AnnFieldsG b r v)

type ConnectionFields b r v = Fields (ConnectionField b r v)

type PageInfoFields = Fields PageInfoField

type EdgeFields b r v = Fields (EdgeField b r v)

data AnnColumnField (b :: BackendType) v

Constructors

AnnColumnField

Fields

_acfColumn :: Column b

_acfType :: ColumnType b

_acfAsText :: Bool

If this field is True, columns are explicitly casted to text when fetched, which avoids an issue that occurs because we don’t currently have proper support for array types. See https://github.com/hasura/graphql-engine/pull/3198 for more details.

_acfArguments :: Maybe (ScalarSelectionArguments b)

Arguments of this column's selection. See ScalarSelectionArguments

_acfRedactionExpression :: AnnRedactionExp b v

This type is used to determine whether the column should be redacted

data ComputedFieldScalarSelect (b :: BackendType) v

Constructors

ComputedFieldScalarSelect

Fields

_cfssFunction :: FunctionName b

_cfssArguments :: FunctionArgsExp b v

_cfssType :: ScalarType b

_cfssScalarArguments :: Maybe (ScalarSelectionArguments b)

_cfssRedactionExpression :: AnnRedactionExp b v

This type is used to determine whether the computed field should be redacted

data ComputedFieldSelect (b :: BackendType) (r :: Type) v

Constructors

CFSScalar

Fields

(ComputedFieldScalarSelect b v)

Type containing info about the computed field

CFSTable JsonAggSelect (AnnSimpleSelectG b r v)

type ArrayRelationSelectG b r v = AnnRelationSelectG b (AnnSimpleSelectG b r v)

type ArrayAggregateSelectG b r v = AnnRelationSelectG b (AnnAggregateSelectG b r v)

type ArrayConnectionSelect b r v = AnnRelationSelectG b (ConnectionSelect b r v)

type ArrayAggregateSelect b = ArrayAggregateSelectG b Void (SQLExpression b)

data AnnObjectSelectG (b :: BackendType) (r :: Type) v

Constructors

AnnObjectSelectG

Fields

_aosFields :: AnnFieldsG b r v

_aosTarget :: SelectFromG b v

_aosTargetFilter :: AnnBoolExp b v

type AnnObjectSelect b r = AnnObjectSelectG b r (SQLExpression b)

type ObjectRelationSelectG b r v = AnnRelationSelectG b (AnnObjectSelectG b r v)

type ObjectRelationSelect b = ObjectRelationSelectG b Void (SQLExpression b)

data ArraySelectG (b :: BackendType) (r :: Type) v

Constructors

ASSimple (ArrayRelationSelectG b r v)

ASAggregate (ArrayAggregateSelectG b r v)

ASConnection (ArrayConnectionSelect b r v)

type ArraySelect b = ArraySelectG b Void (SQLExpression b)

type ArraySelectFieldsG b r v = Fields (ArraySelectG b r v)

data SourceRelationshipSelection (b :: BackendType) (r :: Type) (vf :: BackendType -> Type)

Captures the selection set of a remote source relationship.

Constructors

SourceRelationshipObject (AnnObjectSelectG b r (vf b))

SourceRelationshipArray (AnnSimpleSelectG b r (vf b))

SourceRelationshipArrayAggregate (AnnAggregateSelectG b r (vf b))

data RemoteSourceSelect (r :: Type) (vf :: BackendType -> Type) (tgt :: BackendType)

A relationship to a remote source. vf (could use a better name) is analogous to v in other IR types such as AnnFieldG. vf's kind is (BackendType -> Type) instead of v's Type so that v of AnnFieldG can be specific to the backend that it captures (b of an AnnFieldG changes as we walk down the IR branches which capture relationships to other databases)

Constructors

RemoteSourceSelect

Fields

_rssName :: SourceName

_rssConfig :: SourceConfig tgt

_rssSelection :: SourceRelationshipSelection tgt r vf

_rssJoinMapping :: HashMap FieldName (ScalarType tgt, Column tgt)

Additional information about the source's join columns: (ScalarType tgt) so that the remote can interpret the join values coming from src (Column tgt) so that an appropriate join condition / IN clause can be built by the remote

_rssStringifyNums :: StringifyNumbers

data AnnNestedObjectSelectG (b :: BackendType) (r :: Type) v

Constructors

AnnNestedObjectSelectG

Fields

_anosSupportsNestedObjects :: XNestedObjects b

_anosColumn :: Column b

_anosFields :: AnnFieldsG b r v

type AnnNestedObjectSelect b r = AnnNestedObjectSelectG b r (SQLExpression b)

data AnnNestedArraySelectG (b :: BackendType) (r :: Type) v

Constructors

ANASSimple (AnnFieldG b r v)

ANASAggregate (AnnAggregateSelectG b r v)

type AnnNestedArraySelect b r = AnnNestedArraySelectG b r (SQLExpression b)

insertFunctionArg :: FunctionArgName -> Int -> a -> FunctionArgsExpG a -> FunctionArgsExpG a

If argument positional index is less than or equal to length of positional arguments then insert the value in positional arguments else insert the value with argument name in named arguments

data CountDistinct

The "distinct" input field inside "count" aggregate field

count ( distinct: Boolean ): Int

Constructors

SelectCountDistinct

SelectCountNonDistinct