Hasura.RQL.IR.Select.AnnSelectG

Description

More leaves from Select

Documentation

data AnnSelectG (b :: BackendType) (f :: Type -> Type) (v :: Type)

Constructors

AnnSelectG

Fields

_asnFields :: Fields (f v)

_asnFrom :: SelectFromG b v

_asnPerm :: TablePermG b v

_asnArgs :: SelectArgsG b v

_asnStrfyNum :: StringifyNumbers

_asnNamingConvention :: Maybe NamingCase

data AnnSelectStreamG (b :: BackendType) (f :: Type -> Type) (v :: Type)

IR type representing nodes for streaming subscriptions

Constructors

AnnSelectStreamG

Fields

_assnXStreamingSubscription :: XStreamingSubscription b

type to indicate if streaming subscription has been enabled in the BackendType. This type helps avoiding missing case match patterns for backends where it's disabled.

_assnFields :: Fields (f v)

output selection fields

_assnFrom :: SelectFromG b v

table information to select from

_assnPerm :: TablePermG b v

select permissions

_assnArgs :: SelectStreamArgsG b v

streaming arguments

_assnStrfyNum :: StringifyNumbers

bifoldMapAnnSelectG :: (Backend b, Bifoldable (f b), Monoid m) => (r -> m) -> (v -> m) -> AnnSelectG b (f b r) v -> m

We can't write a Bifoldable instance for AnnSelectG because the types don't line up. Instead, we provide this function which can be used to help define Bifoldable instances of other types containing AnnSelectG values.

bifoldMapAnnSelectStreamG :: (Backend b, Bifoldable (f b), Monoid m) => (r -> m) -> (v -> m) -> AnnSelectStreamG b (f b r) v -> m