Hasura.RQL.IR.Select.Args

Description

Removed from Select to speed up compilation

Documentation

data SelectStreamArgsG (b :: BackendType) v

Constructors

SelectStreamArgsG

Fields

_ssaWhere :: Maybe (AnnBoolExp b v)

optional filter to filter the stream results

_ssaBatchSize :: Int

maximum number of rows to be returned in a single fetch

_ssaCursorArg :: StreamCursorItem b

info related to the cursor column, a single item data type currently because only single column cursors are supported

type SelectStreamArgs b = SelectStreamArgsG b (SQLExpression b)

data SelectArgsG (b :: BackendType) v

Constructors

SelectArgs

Fields

_saWhere :: Maybe (AnnBoolExp b v)

_saOrderBy :: Maybe (NonEmpty (AnnotatedOrderByItemG b v))

_saLimit :: Maybe Int

_saOffset :: Maybe Int64

_saDistinct :: Maybe (NonEmpty (AnnDistinctColumn b v))

type SelectArgs b = SelectArgsG b (SQLExpression b)

noSelectArgs :: SelectArgsG backend v

data AnnDistinctColumn b v

Constructors

AnnDistinctColumn

Fields

_adcColumn :: Column b

_adcCaseBoolExpression :: Maybe (AnnColumnCaseBoolExp b v)

This type is used to determine whether the column should be nullified. When the value is Nothing, the column value will be outputted as computed and when the value is `Just c`, the column will be outputted when c evaluates to true and null when c evaluates to false.

data StreamCursorItem (b :: BackendType)

Cursor for streaming subscription

Constructors

StreamCursorItem

Fields

_sciOrdering :: CursorOrdering

Specifies how the cursor item should be ordered

_sciColInfo :: ColumnInfo b

Column info of the cursor item

_sciInitialValue :: ColumnValue b

Initial value of the cursor item from where the streaming should start