Hasura.RQL.IR.Select.Lenses

Documentation

asnArgs :: forall (b :: BackendType) (f1 :: Type -> Type) v f2. Functor f2 => (SelectArgsG b v -> f2 (SelectArgsG b v)) -> AnnSelectG b f1 v -> f2 (AnnSelectG b f1 v)

asnFields :: forall (b :: BackendType) f1 v f2 f3. Functor f3 => (Fields (f1 v) -> f3 (Fields (f2 v))) -> AnnSelectG b f1 v -> f3 (AnnSelectG b f2 v)

asnFrom :: forall (b :: BackendType) (f1 :: Type -> Type) v f2. Functor f2 => (SelectFromG b v -> f2 (SelectFromG b v)) -> AnnSelectG b f1 v -> f2 (AnnSelectG b f1 v)

asnNamingConvention :: forall (b :: BackendType) (f1 :: Type -> Type) v f2. Functor f2 => (Maybe NamingCase -> f2 (Maybe NamingCase)) -> AnnSelectG b f1 v -> f2 (AnnSelectG b f1 v)

asnPerm :: forall (b :: BackendType) (f1 :: Type -> Type) v f2. Functor f2 => (TablePermG b v -> f2 (TablePermG b v)) -> AnnSelectG b f1 v -> f2 (AnnSelectG b f1 v)

asnStrfyNum :: forall (b :: BackendType) (f1 :: Type -> Type) v f2. Functor f2 => (StringifyNumbers -> f2 StringifyNumbers) -> AnnSelectG b f1 v -> f2 (AnnSelectG b f1 v)

saDistinct :: forall (b :: BackendType) v f. Functor f => (Maybe (NonEmpty (AnnDistinctColumn b v)) -> f (Maybe (NonEmpty (AnnDistinctColumn b v)))) -> SelectArgsG b v -> f (SelectArgsG b v)

saLimit :: forall (b :: BackendType) v f. Functor f => (Maybe Int -> f (Maybe Int)) -> SelectArgsG b v -> f (SelectArgsG b v)

saOffset :: forall (b :: BackendType) v f. Functor f => (Maybe Int64 -> f (Maybe Int64)) -> SelectArgsG b v -> f (SelectArgsG b v)

saOrderBy :: forall (b :: BackendType) v f. Functor f => (Maybe (NonEmpty (AnnotatedOrderByItemG b v)) -> f (Maybe (NonEmpty (AnnotatedOrderByItemG b v)))) -> SelectArgsG b v -> f (SelectArgsG b v)

saWhere :: forall (b :: BackendType) v f. Functor f => (Maybe (AnnBoolExp b v) -> f (Maybe (AnnBoolExp b v))) -> SelectArgsG b v -> f (SelectArgsG b v)

aarAnnSelect :: forall (b :: BackendType) a1 a2 f. Functor f => (a1 -> f a2) -> AnnRelationSelectG b a1 -> f (AnnRelationSelectG b a2)

aarColumnMapping :: forall (b1 :: BackendType) a (b2 :: BackendType) f. Functor f => (HashMap (ColumnPath b1) (ColumnPath b1) -> f (HashMap (ColumnPath b2) (ColumnPath b2))) -> AnnRelationSelectG b1 a -> f (AnnRelationSelectG b2 a)

aarRelationshipName :: forall (b :: BackendType) a f. Functor f => (RelName -> f RelName) -> AnnRelationSelectG b a -> f (AnnRelationSelectG b a)

aarNullable :: forall (b :: BackendType) a f. Functor f => (Nullable -> f Nullable) -> AnnRelationSelectG b a -> f (AnnRelationSelectG b a)

anosSupportsNestedObjects :: forall (b :: BackendType) r v f. Functor f => (XNestedObjects b -> f (XNestedObjects b)) -> AnnNestedObjectSelectG b r v -> f (AnnNestedObjectSelectG b r v)

anosColumn :: forall (b :: BackendType) r v f. Functor f => (Column b -> f (Column b)) -> AnnNestedObjectSelectG b r v -> f (AnnNestedObjectSelectG b r v)

anosFields :: forall (b :: BackendType) r1 v1 r2 v2 f. Functor f => (AnnFieldsG b r1 v1 -> f (AnnFieldsG b r2 v2)) -> AnnNestedObjectSelectG b r1 v1 -> f (AnnNestedObjectSelectG b r2 v2)

aosFields :: forall (b :: BackendType) r1 v r2 f. Functor f => (AnnFieldsG b r1 v -> f (AnnFieldsG b r2 v)) -> AnnObjectSelectG b r1 v -> f (AnnObjectSelectG b r2 v)

aosTarget :: forall (b :: BackendType) r v f. Functor f => (SelectFromG b v -> f (SelectFromG b v)) -> AnnObjectSelectG b r v -> f (AnnObjectSelectG b r v)

aosTargetFilter :: forall (b :: BackendType) r v f. Functor f => (AnnBoolExp b v -> f (AnnBoolExp b v)) -> AnnObjectSelectG b r v -> f (AnnObjectSelectG b r v)

csXRelay :: forall (b :: BackendType) r v f. Functor f => (XRelay b -> f (XRelay b)) -> ConnectionSelect b r v -> f (ConnectionSelect b r v)

csPrimaryKeyColumns :: forall (b :: BackendType) r v f. Functor f => (PrimaryKeyColumns b -> f (PrimaryKeyColumns b)) -> ConnectionSelect b r v -> f (ConnectionSelect b r v)

csSplit :: forall (b :: BackendType) r v f. Functor f => (Maybe (NonEmpty (ConnectionSplit b v)) -> f (Maybe (NonEmpty (ConnectionSplit b v)))) -> ConnectionSelect b r v -> f (ConnectionSelect b r v)

csSlice :: forall (b :: BackendType) r v f. Functor f => (Maybe ConnectionSlice -> f (Maybe ConnectionSlice)) -> ConnectionSelect b r v -> f (ConnectionSelect b r v)

csSelect :: forall (b :: BackendType) r1 v r2 f. Functor f => (AnnSelectG b (ConnectionField b r1) v -> f (AnnSelectG b (ConnectionField b r2) v)) -> ConnectionSelect b r1 v -> f (ConnectionSelect b r2 v)

gbgFields :: forall (b :: BackendType) r1 v1 r2 v2 f. Functor f => (Fields (GroupByField b r1 v1) -> f (Fields (GroupByField b r2 v2))) -> GroupByG b r1 v1 -> f (GroupByG b r2 v2)

gbgKeys :: forall (b :: BackendType) r v f. Functor f => ([GroupKeyField b] -> f [GroupKeyField b]) -> GroupByG b r v -> f (GroupByG b r v)

insertFunctionArg :: FunctionArgName -> Int -> a -> FunctionArgsExpG a -> FunctionArgsExpG a

If argument positional index is less than or equal to length of positional arguments then insert the value in positional arguments else insert the value with argument name in named arguments

mkAnnColumnField :: forall (backend :: BackendType) v r. Column backend -> ColumnType backend -> AnnRedactionExp backend v -> Maybe (ScalarSelectionArguments backend) -> AnnFieldG backend r v

mkAnnColumnFieldAsText :: forall (backend :: BackendType) r v. ColumnInfo backend -> AnnFieldG backend r v

traverseSourceRelationshipSelection :: forall f (backend :: BackendType) vf vg r. (Applicative f, Backend backend) => (vf backend -> f (vg backend)) -> SourceRelationshipSelection backend r vf -> f (SourceRelationshipSelection backend r vg)

_AOCColumn :: forall (b :: BackendType) v p f. (Choice p, Applicative f) => p (ColumnInfo b, AnnRedactionExp b v) (f (ColumnInfo b, AnnRedactionExp b v)) -> p (AnnotatedOrderByElement b v) (f (AnnotatedOrderByElement b v))

_AOCObjectRelation :: forall (b :: BackendType) v p f. (Choice p, Applicative f) => p (RelInfo b, AnnBoolExp b v, AnnotatedOrderByElement b v) (f (RelInfo b, AnnBoolExp b v, AnnotatedOrderByElement b v)) -> p (AnnotatedOrderByElement b v) (f (AnnotatedOrderByElement b v))

_AOCArrayAggregation :: forall (b :: BackendType) v p f. (Choice p, Applicative f) => p (RelInfo b, AnnBoolExp b v, AnnotatedAggregateOrderBy b v) (f (RelInfo b, AnnBoolExp b v, AnnotatedAggregateOrderBy b v)) -> p (AnnotatedOrderByElement b v) (f (AnnotatedOrderByElement b v))

_AOCComputedField :: forall (b :: BackendType) v p f. (Choice p, Applicative f) => p (ComputedFieldOrderBy b v) (f (ComputedFieldOrderBy b v)) -> p (AnnotatedOrderByElement b v) (f (AnnotatedOrderByElement b v))

_AFArrayRelation :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (ArraySelectG b r v) (f (ArraySelectG b r v)) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_AFColumn :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (AnnColumnField b v) (f (AnnColumnField b v)) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_AFComputedField :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (XComputedField b, ComputedFieldName, ComputedFieldSelect b r v) (f (XComputedField b, ComputedFieldName, ComputedFieldSelect b r v)) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_AFExpression :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p Text (f Text) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_AFNodeId :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (XRelay b, SourceName, TableName b, PrimaryKeyColumns b) (f (XRelay b, SourceName, TableName b, PrimaryKeyColumns b)) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_AFObjectRelation :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (ObjectRelationSelectG b r v) (f (ObjectRelationSelectG b r v)) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_AFRemote :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (RemoteRelationshipSelect b r) (f (RemoteRelationshipSelect b r)) -> p (AnnFieldG b r v) (f (AnnFieldG b r v))

_TAFAgg :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (AggregateFields b v) (f (AggregateFields b v)) -> p (TableAggregateFieldG b r v) (f (TableAggregateFieldG b r v))

_TAFNodes :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (XNodesAgg b, AnnFieldsG b r v) (f (XNodesAgg b, AnnFieldsG b r v)) -> p (TableAggregateFieldG b r v) (f (TableAggregateFieldG b r v))

_TAFExp :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p Text (f Text) -> p (TableAggregateFieldG b r v) (f (TableAggregateFieldG b r v))

_GBFGroupKey :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (Fields (GroupKeyField b)) (f (Fields (GroupKeyField b))) -> p (GroupByField b r v) (f (GroupByField b r v))

_GBFAggregate :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p (AggregateFields b v) (f (AggregateFields b v)) -> p (GroupByField b r v) (f (GroupByField b r v))

_GBFNodes :: forall (b :: BackendType) r1 v r2 p f. (Choice p, Applicative f) => p (AnnFieldsG b r1 v) (f (AnnFieldsG b r2 v)) -> p (GroupByField b r1 v) (f (GroupByField b r2 v))

_GBFExp :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p Text (f Text) -> p (GroupByField b r v) (f (GroupByField b r v))

_ConnectionTypename :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p Text (f Text) -> p (ConnectionField b r v) (f (ConnectionField b r v))

_ConnectionPageInfo :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p PageInfoFields (f PageInfoFields) -> p (ConnectionField b r v) (f (ConnectionField b r v))

_ConnectionEdges :: forall (b1 :: BackendType) r1 v1 (b2 :: BackendType) r2 v2 p f. (Choice p, Applicative f) => p (EdgeFields b1 r1 v1) (f (EdgeFields b2 r2 v2)) -> p (ConnectionField b1 r1 v1) (f (ConnectionField b2 r2 v2))

_EdgeTypename :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p Text (f Text) -> p (EdgeField b r v) (f (EdgeField b r v))

_EdgeCursor :: forall (b :: BackendType) r v p f. (Choice p, Applicative f) => p () (f ()) -> p (EdgeField b r v) (f (EdgeField b r v))

_EdgeNode :: forall (b1 :: BackendType) r1 v1 (b2 :: BackendType) r2 v2 p f. (Choice p, Applicative f) => p (AnnFieldsG b1 r1 v1) (f (AnnFieldsG b2 r2 v2)) -> p (EdgeField b1 r1 v1) (f (EdgeField b2 r2 v2))