Hasura.RQL.IR.Select.OrderBy

Documentation

data AnnotatedOrderByElement (b :: BackendType) v

Constructors

AOCColumn

Fields

(ColumnInfo b)

(AnnRedactionExp b v)

This type is used to determine whether the column should be redacted before being ordered over

AOCNestedObject (NestedObjectInfo b) (AnnotatedOrderByElement b v)

AOCObjectRelation

Fields

(RelInfo b)

(AnnBoolExp b v)

Permission filter of the remote table to which the relationship is defined

(AnnotatedOrderByElement b v)

AOCArrayAggregation

Fields

(RelInfo b)

(AnnBoolExp b v)

Permission filter of the remote table to which the relationship is defined

(AnnotatedAggregateOrderBy b v)

AOCComputedField (ComputedFieldOrderBy b v)

data AnnotatedAggregateOrderBy (b :: BackendType) v

Constructors

AAOCount

AAOOp (AggregateOrderByColumn b v)
Order by an aggregate function applied to a column

data AggregateOrderByColumn b v

Constructors

AggregateOrderByColumn

Fields

_aobcAggregateFunctionName :: Text

_aobcAggregateFunctionReturnType :: ColumnType b

_aobcColumn :: ColumnInfo b

_aobcRedactionExpression :: AnnRedactionExp b v

This type is used to determine whether the column should be redacted before being aggregated and then ordered over

type AnnotatedOrderByItemG b v = OrderByItemG b (AnnotatedOrderByElement b v)

type AnnotatedOrderByItem b = AnnotatedOrderByItemG b (SQLExpression b)

data ComputedFieldOrderByElement (b :: BackendType) v

The order by element for a computed field based on its return type

Constructors

CFOBEScalar
Sort by the scalar computed field

Fields

(ScalarType b)

(AnnRedactionExp b v)

This type is used to determine whether the computed field should be redacted before being ordered over

CFOBETableAggregation

Fields

(TableName b)

(AnnBoolExp b v)

Permission filter of the retuning table

(AnnotatedAggregateOrderBy b v)

Sort by aggregation fields of table rows returned by computed field

data ComputedFieldOrderBy (b :: BackendType) v

Constructors

ComputedFieldOrderBy

Fields

_cfobXField :: XComputedField b

_cfobName :: ComputedFieldName

_cfobFunction :: FunctionName b

_cfobFunctionArgsExp :: FunctionArgsExp b v

_cfobOrderByElement :: ComputedFieldOrderByElement b v