Hasura.RQL.IR.Update.Batch

Description

Contains types that can be used by backends to structure updates to batches of rows in a table

Documentation

data UpdateBatch (b :: BackendType) (updateOperators :: Type -> Type) v

Represents a set of update operations (_ubOperations) applied to a batch of rows selected from a table by filtering it with a boolean expression (_ubWhere).

This type may be used by specific backends as a part their UpdateVariant. See PgUpdateVariant for an example.

The actual operators used to affect changes against columns in _ubOperations are abstract here and are specified by the specific backends based on what they actually support

Constructors

UpdateBatch

Fields

_ubOperations :: HashMap (Column b) (updateOperators v)

_ubWhere :: AnnBoolExp b v

updateBatchIsEmpty :: forall (b :: BackendType) (updateOperators :: Type -> Type) v. UpdateBatch b updateOperators v -> Bool

Are we actually updating anything in the batch?