Hasura.RQL.Types.Allowlist

Documentation

newtype DropCollectionFromAllowlist

Constructors

DropCollectionFromAllowlist

Fields

_dcfaCollection :: CollectionName

data AllowlistScope

Constructors

AllowlistScopeGlobal

AllowlistScopeRoles !(NonEmpty RoleName)

data AllowlistEntry

Constructors

AllowlistEntry

Fields

aeCollection :: !CollectionName

aeScope :: !AllowlistScope

newtype UpdateScopeOfCollectionInAllowlist

Wrap AllowlistEntry with a FromJSON instance that requires scope to be set.

Constructors

UpdateScopeOfCollectionInAllowlist AllowlistEntry

type MetadataAllowlist = InsOrdHashMap CollectionName AllowlistEntry

metadataAllowlistInsert :: AllowlistEntry -> MetadataAllowlist -> Either Text MetadataAllowlist

metadataAllowlistUpdateScope :: AllowlistEntry -> MetadataAllowlist -> Either Text MetadataAllowlist

metadataAllowlistAllCollections :: MetadataAllowlist -> [CollectionName]

Produce a list of all collections in the allowlist. This is used in runDropCollection to function to ensure that we don't delete any collections which are referred to in the allowlist.

newtype NormalizedQuery

A query stripped of typenames. A query is allowed if it occurs in an allowed query collection after normalization.

Compare docsgraphqlcoredeploymentallow-list.rst.

Constructors

NormalizedQuery

Fields

unNormalizedQuery :: ExecutableDocument Name

normalizeQuery :: ExecutableDocument Name -> NormalizedQuery

Normalize query for comparison by stripping type names.

data InlinedAllowlist

InlinedAllowlist is the data type with which the allowlist is represented in the schema cache, it contains a global and a per role allowlist and when allowlist is enabled in the graphql-engine, the incoming query for a non-admin role should either be in the global allowlist or in the given role's role based allowlist.

Essentially, it's a memoization of allowlistAllowsQuery implemented in terms of MetadataAllowlist.

Constructors

InlinedAllowlist

Fields

iaGlobal :: HashSet NormalizedQuery

iaPerRole :: HashMap RoleName (HashSet NormalizedQuery)

inlineAllowlist :: QueryCollections -> MetadataAllowlist -> InlinedAllowlist

data AllowlistMode

The mode in which the allowlist functions. In global mode, collections with non-global scope are ignored.

Constructors

AllowlistModeGlobalOnly

AllowlistModeFull

allowlistAllowsQuery :: InlinedAllowlist -> AllowlistMode -> RoleName -> ExecutableDocument Name -> Bool