Hasura.RQL.Types.ApiLimit

Documentation

data ApiLimit

Constructors

ApiLimit

Fields

_alRateLimit :: Maybe RateLimit

_alDepthLimit :: Maybe DepthLimit

_alNodeLimit :: Maybe NodeLimit

_alTimeLimit :: Maybe TimeLimit

_alBatchLimit :: Maybe BatchLimit

_alDisabled :: Bool

emptyApiLimit :: ApiLimit

data Limit a

Constructors

Limit

Fields

_lGlobal :: a

_lPerRole :: Map RoleName a

type RateLimit = Limit RateLimitConfig

type DepthLimit = Limit MaxDepth

type NodeLimit = Limit MaxNodes

type TimeLimit = Limit MaxTime

type BatchLimit = Limit MaxBatchSize

data RateLimitConfig

Constructors

RateLimitConfig

Fields

_rlcMaxReqsPerMin :: Int

_rlcUniqueParams :: Maybe UniqueParamConfig

data UniqueParamConfig

The unique key using which an authenticated client can be identified

Constructors

UPCSessionVar [Text]
it can be a list of session variable (like session var in UserInfo)

UPCIpAddress
or it can be an IP address

newtype MaxDepth

Constructors

MaxDepth

Fields

unMaxDepth :: Int

newtype MaxNodes

Constructors

MaxNodes

Fields

unMaxNodes :: Int

newtype MaxTime

Constructors

MaxTime

Fields

unMaxTime :: Seconds

newtype MaxBatchSize

Constructors

MaxBatchSize

Fields

unMaxBatchSize :: Int