Hasura.RQL.Types.BackendTag

Documentation

data BackendTag (b :: BackendType) where

A singleton-like GADT that associates a tag to each backend.

Constructors

PostgresVanillaTag :: BackendTag ('Postgres 'Vanilla)

PostgresCitusTag :: BackendTag ('Postgres 'Citus)

PostgresCockroachTag :: BackendTag ('Postgres 'Cockroach)

MSSQLTag :: BackendTag 'MSSQL

BigQueryTag :: BackendTag 'BigQuery

DataConnectorTag :: BackendTag 'DataConnector

class HasTag (b :: BackendType) where

This class describes how to get a tag for a given type. We use it in AnyBackend: `case backendTag @b of`...

Methods

backendTag :: BackendTag b

reify :: BackendTag b -> BackendType

How to convert back from a tag to a runtime value.

backendPrefix :: forall b. HasTag b => Text

Provides a title-cased name for a database kind, inferring the appropriate database kind from type context.