Hasura.RQL.Types.Common

Documentation

newtype RelName

Constructors

RelName

Fields

getRelTxt :: NonEmptyText

relNameToTxt :: RelName -> Text

fromRemoteRelationship :: RelName -> FieldName

data RelType

Constructors

ObjRel

ArrRel

relTypeToTxt :: RelType -> Text

data JsonAggSelect

Constructors

JASMultipleRows

JASSingleObject

data InsertOrder

Constructors

BeforeParent

AfterParent

newtype OID

Postgres OIDs. https://www.postgresql.org/docs/12/datatype-oid.html

Constructors

OID

Fields

unOID :: Int

newtype FieldName

Constructors

FieldName

Fields

getFieldNameTxt :: Text

type Fields a = [(FieldName, a)]

class ToAesonPairs a where

Methods

toAesonPairs :: KeyValue v => a -> [v]

data SourceName

Constructors

SNDefault

SNName NonEmptyText

sourceNameToText :: SourceName -> Text

defaultSource :: SourceName

data InpValInfo

Constructors

InpValInfo

Fields

_iviDesc :: Maybe Description

_iviName :: Name

_iviDefVal :: Maybe (Value Void)

_iviType :: GType

newtype SystemDefined

Constructors

SystemDefined

Fields

unSystemDefined :: Bool

isSystemDefined :: SystemDefined -> Bool

data SQLGenCtx

Constructors

SQLGenCtx

Fields

stringifyNum :: StringifyNumbers

dangerousBooleanCollapse :: DangerouslyCollapseBooleans

optimizePermissionFilters :: OptimizePermissionFilters

bigqueryStringNumericInput :: BigQueryStringNumericInput

successMsg :: EncJSON

failureMsg :: EncJSON

newtype ResolvedWebhook

Constructors

ResolvedWebhook

Fields

unResolvedWebhook :: Text

newtype InputWebhook

Constructors

InputWebhook

Fields

unInputWebhook :: URLTemplate

newtype ResolveWebhookError

Constructors

ResolveWebhookError

Fields

unResolveWebhookError :: Text

resolveWebhook :: QErrM m => Environment -> InputWebhook -> m ResolvedWebhook

resolveWebhookEither :: Environment -> InputWebhook -> Either ResolveWebhookError ResolvedWebhook

newtype Timeout

Constructors

Timeout

Fields

unTimeout :: Int

defaultActionTimeoutSecs :: Timeout

data PGConnectionParams

See API reference here: https://hasura.io/docs/latest/graphql/core/api-reference/syntax-defs.html#pgconnectionparameters

Constructors

PGConnectionParams

Fields

_pgcpHost :: Text

_pgcpUsername :: Text

_pgcpPassword :: Maybe Text

_pgcpPort :: Int

_pgcpDatabase :: Text

data UrlConf

Constructors

UrlValue InputWebhook
the database connection string

UrlFromEnv Text
the name of environment variable containing the connection string

UrlFromParams PGConnectionParams
the minimum required `connection parameters` to construct a valid connection string

getConnOptionsFromConnParams :: PGConnectionParams -> ConnOptions

getPGConnectionStringFromParams :: PGConnectionParams -> String

Construct a Postgres connection URI as a String from PGConnectionParams.

NOTE: This function takes care to properly escape all URI components, as Postgres requires that a connection URI is percent-encoded if it includes symbols with "special meaning".

See the libpq documentation for details: https://www.postgresql.org/docs/13/libpq-connect.html#id-1.7.3.8.3.6

resolveUrlConf :: MonadError QErr m => Environment -> UrlConf -> m Text

getEnv :: QErrM m => Environment -> Text -> m Text

getEnvEither :: Environment -> Text -> Either Text Text

data MetricsConfig

Various user-controlled configuration for metrics used by Pro

Constructors

MetricsConfig

Fields

_mcAnalyzeQueryVariables :: Bool

should the query-variables be logged and analyzed for metrics

_mcAnalyzeResponseBody :: Bool

should the response-body be analyzed for empty and null responses

emptyMetricsConfig :: MetricsConfig

data Comment

Constructors

Automatic
Automatically generate a comment (derive it from DB comments, or a sensible default describing the source of the data)

Explicit (Maybe NonEmptyText)
The user's explicitly provided comment, or explicitly no comment (ie. leave it blank, do not autogenerate one)

commentToMaybeText :: Comment -> Maybe Text

commentFromMaybeText :: Maybe Text -> Comment

data EnvRecord a

We use the following type, after we resolve the env var. | This will store both the env var name and the resolved value.

Constructors

EnvRecord

Fields

_envVarName :: Text

_envVarValue :: a

data ApolloFederationVersion

Constructors

V1

data ApolloFederationConfig

Constructors

ApolloFederationConfig

Fields

enable :: ApolloFederationVersion

isApolloFedV1enabled :: Maybe ApolloFederationConfig -> Bool

data TriggerOnReplication

Type to indicate if the SQL trigger should be enabled when data is inserted into a table through replication.

Constructors

TOREnableTrigger

TORDisableTrigger

data RemoteRelationshipG definition

Metadata representation of a generic remote relationship, regardless of the source: all sources use this same agnostic definition. The internal definition field is where we differentiate between different targets.

TODO: This needs to be moved to an appropriate module, maybe something like Hasura.RemoteRelationships.Metadata.

Constructors

RemoteRelationship

Fields

_rrName :: RelName

_rrDefinition :: definition

rrName :: Lens (RemoteRelationshipG def) (RemoteRelationshipG def) RelName RelName

rrDefinition :: Lens (RemoteRelationshipG def) (RemoteRelationshipG def') def def'

remoteRelationshipCodec :: forall definition. Typeable definition => JSONCodec definition -> JSONCodec (RemoteRelationshipG definition)