Hasura.RQL.Types.ComputedField

Description

Documentation

data FunctionTrackedAs (b :: BackendType)

Constructors

FTAComputedField ComputedFieldName SourceName (TableName b)

FTACustomFunction CustomFunctionNames

data CustomFunctionNames

The function name and input arguments name for the "args" field parser.

function_name(args: args_name)

Constructors

CustomFunctionNames

Fields

cfnFunctionName :: Name

cfnArgsName :: Name

data ComputedFieldFunction (b :: BackendType)

Constructors

ComputedFieldFunction

Fields

_cffName :: FunctionName b

_cffInputArgs :: Seq (FunctionArgument b)

_cffComputedFieldImplicitArgs :: ComputedFieldImplicitArguments b

_cffDescription :: Maybe PGDescription

data ComputedFieldInfo (b :: BackendType)

Constructors

ComputedFieldInfo

Fields

_cfiXComputedFieldInfo :: XComputedField b

_cfiName :: ComputedFieldName

_cfiFunction :: ComputedFieldFunction b

_cfiReturnType :: ComputedFieldReturn b

_cfiDescription :: Maybe Text

onlyNumComputedFields :: forall b. Backend b => [ComputedFieldInfo b] -> [ComputedFieldInfo b]

Return all the computed fields in the given list that have numeric types.

isNumComputedField :: forall b. Backend b => ComputedFieldInfo b -> Bool

Check whether a computed field has a numeric type.

onlyComparableComputedFields :: forall b. Backend b => [ComputedFieldInfo b] -> [ComputedFieldInfo b]

Return all the computed fields in the given list that have numeric types.

isComparableComputedField :: forall b. Backend b => ComputedFieldInfo b -> Bool

Check whether a computed field has a numeric type.

cfiXComputedFieldInfo :: forall b. Lens' (ComputedFieldInfo b) (XComputedField b)

cfiReturnType :: forall b. Lens' (ComputedFieldInfo b) (ComputedFieldReturn b)

cfiName :: forall b. Lens' (ComputedFieldInfo b) ComputedFieldName

cfiFunction :: forall b. Lens' (ComputedFieldInfo b) (ComputedFieldFunction b)

cfiDescription :: forall b. Lens' (ComputedFieldInfo b) (Maybe Text)

removeComputedFieldsReturningExistingTable :: forall backend. Backend backend => [ComputedFieldInfo backend] -> [ComputedFieldInfo backend]

Filter computed fields not returning rows of existing table