Hasura.RQL.Types.CustomTypes

Documentation

newtype GraphQLType

A wrapper around GType which allows us to define custom JSON instances.

TODO: this name is ambiguous, and conflicts with Hasura.RQL.DDL.RemoteSchema.Permission.GraphQLType; it should perhaps be renamed, made internal to this module, or removed altogether?

Constructors

GraphQLType

Fields

unGraphQLType :: GType

isListType :: GraphQLType -> Bool

isNullableType :: GraphQLType -> Bool

isInBuiltScalar :: Text -> Bool

data CustomTypes

A set of custom GraphQL types, sorted by "kind".

Constructors

CustomTypes

Fields

_ctInputObjects :: [InputObjectTypeDefinition]

_ctObjects :: [ObjectTypeDefinition]

_ctScalars :: [ScalarTypeDefinition]

_ctEnums :: [EnumTypeDefinition]

emptyCustomTypes :: CustomTypes

data InputObjectTypeDefinition

Constructors

InputObjectTypeDefinition

Fields

_iotdName :: InputObjectTypeName

_iotdDescription :: Maybe Description

_iotdFields :: NonEmpty InputObjectFieldDefinition

newtype InputObjectTypeName

Constructors

InputObjectTypeName

Fields

unInputObjectTypeName :: Name

data InputObjectFieldDefinition

Constructors

InputObjectFieldDefinition

Fields

_iofdName :: InputObjectFieldName

_iofdDescription :: Maybe Description

_iofdType :: GraphQLType

newtype InputObjectFieldName

Constructors

InputObjectFieldName

Fields

unInputObjectFieldName :: Name

data ObjectTypeDefinition

Constructors

ObjectTypeDefinition

Fields

_otdName :: ObjectTypeName

_otdDescription :: Maybe Description

_otdFields :: NonEmpty (ObjectFieldDefinition GraphQLType)

_otdRelationships :: [TypeRelationshipDefinition]

newtype ObjectTypeName

Constructors

ObjectTypeName

Fields

unObjectTypeName :: Name

data ObjectFieldDefinition field

Constructors

ObjectFieldDefinition

Fields

_ofdName :: ObjectFieldName

_ofdArguments :: Maybe Value

_ofdDescription :: Maybe Description

_ofdType :: field

newtype ObjectFieldName

Constructors

ObjectFieldName

Fields

unObjectFieldName :: Name

data ScalarTypeDefinition

Constructors

ScalarTypeDefinition

Fields

_stdName :: Name

_stdDescription :: Maybe Description

defaultGraphQLScalars :: HashMap Name ScalarTypeDefinition

data EnumTypeDefinition

Constructors

EnumTypeDefinition

Fields

_etdName :: EnumTypeName

_etdDescription :: Maybe Description

_etdValues :: NonEmpty EnumValueDefinition

newtype EnumTypeName

Constructors

EnumTypeName

Fields

unEnumTypeName :: Name

data EnumValueDefinition

Constructors

EnumValueDefinition

Fields

_evdValue :: EnumValue

_evdDescription :: Maybe Description

_evdIsDeprecated :: Maybe Bool

data TypeRelationshipDefinition

Constructors

TypeRelationshipDefinition

Fields

_trdName :: RelationshipName

_trdType :: RelType

_trdSource :: SourceName

_trdRemoteTable :: QualifiedTable

_trdFieldMapping :: HashMap ObjectFieldName PGCol

newtype RelationshipName

TODO: deduplicate this in favour of RelName

Constructors

RelationshipName

Fields

unRelationshipName :: Name

data AnnotatedCustomTypes

While we do not persist resolved types in the schema cache, they are used when building the cache to validate and resolve actions and their relationships.

Constructors

AnnotatedCustomTypes

Fields

_actInputTypes :: HashMap Name AnnotatedInputType

_actObjectTypes :: HashMap Name AnnotatedObjectType

data AnnotatedInputType

Constructors

NOCTScalar AnnotatedScalarType

NOCTEnum EnumTypeDefinition

NOCTInputObject InputObjectTypeDefinition

data AnnotatedScalarType

Constructors

ASTCustom ScalarTypeDefinition

ASTReusedScalar Name (AnyBackend ScalarWrapper)

newtype ScalarWrapper b

Constructors

ScalarWrapper

Fields

unwrapScalar :: ScalarType b

data AnnotatedOutputType

Constructors

AOTObject AnnotatedObjectType

AOTScalar AnnotatedScalarType

data AnnotatedObjectType

Constructors

AnnotatedObjectType

Fields

_aotName :: ObjectTypeName

_aotDescription :: Maybe Description

_aotFields :: NonEmpty (ObjectFieldDefinition (GType, AnnotatedObjectFieldType))

_aotRelationships :: [AnnotatedTypeRelationship]

data AnnotatedObjectFieldType

Constructors

AOFTScalar AnnotatedScalarType

AOFTEnum EnumTypeDefinition

AOFTObject Name

data AnnotatedTypeRelationship

Constructors

AnnotatedTypeRelationship

Fields

_atrName :: RelationshipName

_atrType :: RelType

_atrSource :: SourceName

_atrSourceConfig :: SourceConfig ('Postgres 'Vanilla)

_atrTableInfo :: TableInfo ('Postgres 'Vanilla)

_atrFieldMapping :: HashMap ObjectFieldName (ColumnInfo ('Postgres 'Vanilla))

trdType :: Lens' TypeRelationshipDefinition RelType

trdSource :: Lens' TypeRelationshipDefinition SourceName

trdRemoteTable :: Lens' TypeRelationshipDefinition QualifiedTable

trdName :: Lens' TypeRelationshipDefinition RelationshipName

trdFieldMapping :: Lens' TypeRelationshipDefinition (HashMap ObjectFieldName PGCol)