Hasura.RQL.Types.EventTrigger

Documentation

newtype TriggerName

Unique name for event trigger.

Constructors

TriggerName

Fields

unTriggerName :: NonEmptyText

triggerNameToTxt :: TriggerName -> Text

isIllegalTriggerName :: TriggerName -> Bool

data Ops

Constructors

INSERT

UPDATE

DELETE

MANUAL

data SubscribeColumns (b :: BackendType)

Constructors

SubCStar

SubCArray [Column b]

data SubscribeOpSpec (b :: BackendType)

Constructors

SubscribeOpSpec

Fields

sosColumns :: SubscribeColumns b

Columns of the table that user can subscribe to listen for changes.

sosPayload :: Maybe (SubscribeColumns b)

Columns that the event trigger payload should consists. If set, only those columns will be visible in the payload. By default, the payload consists of all the columns of the table.

defaultTimeoutSeconds :: Int

defaultRetryConf :: RetryConf

data RetryConf

Constructors

RetryConf

Fields

rcNumRetries :: Int

rcIntervalSec :: Int

rcTimeoutSec :: Maybe Int

data EventHeaderInfo

Constructors

EventHeaderInfo

Fields

ehiHeaderConf :: HeaderConf

ehiCachedValue :: Text

data WebhookConf

Constructors

WCValue InputWebhook

WCEnv Text

data WebhookConfInfo

Constructors

WebhookConfInfo

Fields

wciWebhookConf :: WebhookConf

wciCachedValue :: EnvRecord ResolvedWebhook

data TriggerOpsDef (b :: BackendType)

The table operations on which the event trigger will be invoked.

Constructors

TriggerOpsDef

Fields

tdInsert :: Maybe (SubscribeOpSpec b)

tdUpdate :: Maybe (SubscribeOpSpec b)

tdDelete :: Maybe (SubscribeOpSpec b)

tdEnableManual :: Maybe Bool

data EventTriggerCleanupStatus

Constructors

ETCSPaused

ETCSUnpaused

data AutoTriggerLogCleanupConfig

Automatic event trigger log cleanup configuration

Constructors

AutoTriggerLogCleanupConfig

Fields

_atlccSchedule :: CronSchedule

cron schedule for the automatic cleanup

_atlccBatchSize :: Int

maximum number of events to be deleted in a single cleanup action

_atlccClearOlderThan :: Int

retention period (in hours) for the event trigger logs

_atlccTimeout :: Int

SQL query timeout (in seconds)

_atlccCleanInvocationLogs :: Bool

should we clean the invocation logs as well

_atlccPaused :: EventTriggerCleanupStatus

is the cleanup action paused

data TriggerLogCleanupConfig

Manual event trigger log cleanup configuration

Constructors

TriggerLogCleanupConfig

Fields

tlccEventTriggerName :: TriggerName

name of the event trigger

tlccSourceName :: SourceName

source of the event trigger

tlccBatchSize :: Int

batch size of for the cleanup action

tlccClearOlderThan :: Int

retention period (in hours) for the event trigger logs

tlccTimeout :: Int

SQL query timeout (in seconds)

tlccCleanInvocationLogs :: Bool

should we clean the invocation logs as well

data EventTriggerQualifier

Constructors

EventTriggerQualifier

Fields

_etqSourceName :: SourceName

_etqEventTriggers :: NonEmpty TriggerName

data TriggerLogCleanupSources

Constructors

TriggerAllSource

TriggerSource (NonEmpty SourceName)

data TriggerLogCleanupToggleConfig

Constructors

TriggerLogCleanupSources TriggerLogCleanupSources

TriggerQualifier (NonEmpty EventTriggerQualifier)

data EventTriggerConf (b :: BackendType)

Constructors

EventTriggerConf

Fields

etcName :: TriggerName

etcDefinition :: TriggerOpsDef b

etcWebhook :: Maybe InputWebhook

etcWebhookFromEnv :: Maybe Text

etcRetryConf :: RetryConf

etcHeaders :: Maybe [HeaderConf]

etcRequestTransform :: Maybe RequestTransform

etcResponseTransform :: Maybe MetadataResponseTransform

etcCleanupConfig :: Maybe AutoTriggerLogCleanupConfig

etcTriggerOnReplication :: TriggerOnReplication

updateCleanupConfig :: Maybe AutoTriggerLogCleanupConfig -> EventTriggerConf b -> EventTriggerConf b

data RecreateEventTriggers

Constructors

RETRecreate

RETDoNothing

data TriggerMetadata

Constructors

TriggerMetadata

Fields

tmName :: TriggerName

data Event (b :: BackendType)

Change data for a particular row

https://docs.hasura.io/1.0/graphql/manual/event-triggers/payload.html

Constructors

Event

Fields

eId :: EventId

eSource :: SourceName

eTable :: TableName b

eTrigger :: TriggerMetadata

eEvent :: Value

eTries :: Int

eCreatedAt :: LocalTime

eRetryAt :: Maybe UTCTime

eCreatedAtUTC :: UTCTime

The values eCreatedAtUTC and eRetryAtUTC are only used for calculating the event_processing_time metric.

eRetryAtUTC :: Maybe UTCTime

data EventWithSource (b :: BackendType)

The event payload processed by processEvent

Constructors

EventWithSource

Fields

_ewsEvent :: Event b

_ewsSourceConfig :: SourceConfig b

_ewsSourceName :: SourceName

_ewsFetchTime :: UTCTime

The UTCTime represents the time when the event was fetched from DB. ^ Used to calculate Event Lock time

data ProcessEventError

Constructors

PESetRetry UTCTime

PESetError

data EventTriggerInfo (b :: BackendType)

Constructors

EventTriggerInfo

Fields

etiName :: TriggerName

etiOpsDef :: TriggerOpsDef b

etiRetryConf :: RetryConf

etiWebhookInfo :: WebhookConfInfo

The HTTP(s) URL which will be called with the event payload on configured operation. Must be a POST handler. This URL can be entered manually or can be picked up from an environment variable (the environment variable needs to be set before using it for this configuration).

etiHeaders :: [EventHeaderInfo]

Custom headers can be added to an event trigger. Each webhook request will have these headers added.

etiRequestTransform :: Maybe RequestTransform

etiResponseTransform :: Maybe MetadataResponseTransform

etiCleanupConfig :: Maybe AutoTriggerLogCleanupConfig

etiTriggerOnReplication :: TriggerOnReplication

type EventTriggerInfoMap b = HashMap TriggerName (EventTriggerInfo b)

newtype FetchBatchSize

Constructors

FetchBatchSize

Fields

_unFetchBatchSize :: Int

data DeletedEventLogStats

Statistics of deleted event logs and invocation logs

Constructors

DeletedEventLogStats

Fields

deletedEventLogs :: Int

deletedInvocationLogs :: Int

data EventLogStatus

Constructors

Processed

Pending

All

data GetEventLogs (b :: BackendType)

Constructors

GetEventLogs

Fields

_gelName :: TriggerName

_gelSourceName :: SourceName

_gelLimit :: Int

_gelOffset :: Int

_gelStatus :: EventLogStatus

data EventLog

Constructors

EventLog

Fields

elId :: EventId

elSchemaName :: Text

elTableName :: Text

elTriggerName :: TriggerName

elPayload :: Value

elDelivered :: Bool

elError :: Bool

elTries :: Int

elCreatedAt :: UTCTime

elLocked :: Maybe UTCTime

elNextRetryAt :: Maybe UTCTime

elArchived :: Bool

data GetEventInvocations (b :: BackendType)

Constructors

GetEventInvocations

Fields

_geiName :: TriggerName

_geiSourceName :: SourceName

_geiLimit :: Int

_geiOffset :: Int

data EventInvocationLog

Constructors

EventInvocationLog

Fields

eilId :: Text

eilTriggerName :: TriggerName

eilEventId :: EventId

eilHttpStatus :: Maybe Int

eilRequest :: Value

eilResponse :: Value

eilCreatedAt :: UTCTime

data GetEventById (b :: BackendType)

Constructors

GetEventById

Fields

_gebiSourceName :: SourceName

_gebiEventId :: EventId

_gebiInvocationLogLimit :: Int

_gebiInvocationLogOffset :: Int

data EventLogWithInvocations

Constructors

EventLogWithInvocations

Fields

elwiEvent :: Maybe EventLog

elwiInvocations :: [EventInvocationLog]