Hasura.RQL.Types.Eventing.Backend

Documentation

class Backend b => BackendEventTrigger (b :: BackendType) where

The BackendEventTrigger type class contains functions which interact with the source database to perform event trigger related operations like fetching pending events from the database or inserting a new invocation log after processing an event.

Methods

insertManualEvent :: (MonadIO m, MonadError QErr m) => SourceConfig b -> TableName b -> TriggerName -> Value -> UserInfo -> TraceContext -> m EventId

insertManualEvent inserts the specified event in the event log table, note that this method should also set the trace context and session variables in the source database context (if available)

fetchUndeliveredEvents :: (MonadIO m, MonadError QErr m) => SourceConfig b -> SourceName -> MaintenanceMode -> FetchBatchSize -> m [Event b]

fetchUndeliveredEvents fetches the undelivered events from the source and locks the events which are fetched for processing. The locking is done so that when there are multiple instances of graphql-engine connected to the same source they don't end up processing the same events concurrently. Also, it's crucial that the SQL query used to fetch events in this function uses something like Postgres's `FOR UPDATE SKIP LOCKED` mechanism so that it skips past the events which are locked by the database and pick newer undelivered events to achieve maximum throughput.

The locking mechanism for event triggers is timestamp based i.e. when an event is fetched from the database, the locked column will contain the timestamp of when it was fetched from the database. Undelivered events will have NULL value as their locked column value. The idea behind having a timestamp based locking mechanism is that if the graphql-engine is shutdown abruptly with events being fetched by the events processor, it will be locked and after the shutdown it will remain locked. Now with a timestamp based lock, when the graphql-engine is started again it will also fetch events which have a locked value of older than 30 mins along with the undelivered events. So, this way no events remain in a locked state.

setRetry :: (MonadIO m, MonadError QErr m) => SourceConfig b -> Event b -> UTCTime -> Maybe MaintenanceModeVersion -> m ()

Ad-hoc function to set a retry for an undelivered event

getMaintenanceModeVersion :: (MonadIO m, MonadError QErr m) => SourceConfig b -> m MaintenanceModeVersion

getMaintenanceModeVersion gets the source catalog version from the source

recordSuccess :: MonadIO m => SourceConfig b -> Event b -> Invocation 'EventType -> Maybe MaintenanceModeVersion -> m (Either QErr ())

recordSuccess records a successful event invocation, it does a couple of things,

Insert the invocation in the invocation logs table

Mark the event as delivered in the event_log table

recordError :: MonadIO m => SourceConfig b -> Event b -> Invocation 'EventType -> ProcessEventError -> Maybe MaintenanceModeVersion -> m (Either QErr ())

recordError records an erronous event invocation, it does a couple of things,

Insert the invocation in the invocation logs table

Depending on the value of ProcessEventError, it will either, - Set a retry for the given event - Mark the event as error

recordError' :: MonadIO m => SourceConfig b -> Event b -> Maybe (Invocation 'EventType) -> ProcessEventError -> Maybe MaintenanceModeVersion -> m (Either QErr ())

recordError' records an erronous event invocation, it does a couple of things,

If present, insert the invocation in the invocation logs table

Depending on the value of ProcessEventError, it will either, - Set a retry for the given event - Mark the event as error

dropTriggerAndArchiveEvents :: (MonadIO m, MonadError QErr m) => SourceConfig b -> TriggerName -> m ()

dropTriggerAndArchiveEvents drops the database trigger and marks all the events related to the event trigger as archived. See Note [Cleanup for dropped triggers]

redeliverEvent :: (MonadIO m, MonadError QErr m) => SourceConfig b -> EventId -> m ()

unlockEventsInSource :: MonadIO m => SourceConfig b -> Set EventId -> m (Either QErr Int)

unlockEventsInSource unlocks the cached locked events captured by the events when a graceful shutdown is initiated, so that when the graphql-engine is started up again these events can be fetched to process them immediately.

createTableEventTrigger :: (MonadBaseControl IO m, MonadIO m) => ServerConfigCtx -> SourceConfig b -> TableName b -> [ColumnInfo b] -> TriggerName -> TriggerOpsDef b -> m (Either QErr ())