Hasura.RQL.Types.Function

Documentation

data FunctionVolatility

https://www.postgresql.org/docs/current/xfunc-volatility.html

Constructors

FTVOLATILE

FTIMMUTABLE

FTSTABLE

funcTypToTxt :: FunctionVolatility -> Text

newtype FunctionArgName

Constructors

FunctionArgName

Fields

getFuncArgNameTxt :: Text

newtype HasDefault

Constructors

HasDefault

Fields

unHasDefault :: Bool

data FunctionArg (b :: BackendType)

Constructors

FunctionArg

Fields

faName :: !(Maybe FunctionArgName)

faType :: !(FunctionArgType b)

faHasDefault :: !HasDefault

data InputArgument a

Constructors

IAUserProvided !a

IASessionVariables !FunctionArgName

_IASessionVariables :: forall a. Prism' (InputArgument a) FunctionArgName

_IAUserProvided :: forall a a. Prism (InputArgument a) (InputArgument a) a a

type FunctionInputArgument b = InputArgument (FunctionArg b)

data FunctionExposedAs

Indicates whether the user requested the corresponding function to be tracked as a mutation or a query/subscription, in track_function.

Constructors

FEAQuery

FEAMutation

newtype FunctionPermissionInfo

Constructors

FunctionPermissionInfo

Fields

_fpmRole :: RoleName

fpmRole :: Iso' FunctionPermissionInfo RoleName

type FunctionPermissionsMap = HashMap RoleName FunctionPermissionInfo

data FunctionCustomRootFields

Custom root fields for functions. When set, will be the names exposed to the user in the schema.

See rfcs/function-root-field-customisation.md for more information.

Constructors

FunctionCustomRootFields

Fields

_fcrfFunction :: Maybe Name

_fcrfFunctionAggregate :: Maybe Name

emptyFunctionCustomRootFields :: FunctionCustomRootFields

A function custom root fields without custom names set. This is the default.

data FunctionInfo (b :: BackendType)

Tracked SQL function metadata. See buildFunctionInfo.

Constructors

FunctionInfo

Fields

_fiSQLName :: FunctionName b

_fiGQLName :: Name

_fiGQLArgsName :: Name

_fiGQLAggregateName :: Name

_fiSystemDefined :: SystemDefined

_fiVolatility :: FunctionVolatility

_fiExposedAs :: FunctionExposedAs

In which part of the schema should this function be exposed?

See mkFunctionInfo and _fcExposedAs.

_fiInputArgs :: Seq (FunctionInputArgument b)

_fiReturnType :: TableName b

NOTE: when a table is created, a new composite type of the same name is automatically created; so strictly speaking this field means "the function returns the composite type corresponding to this table".

_fiDescription :: Maybe Text

this field represents the description of the function as present on the database

_fiPermissions :: FunctionPermissionsMap

Roles to which the function is accessible

_fiJsonAggSelect :: JsonAggSelect

_fiComment :: Maybe Text

fiVolatility :: forall b. Lens' (FunctionInfo b) FunctionVolatility

fiSystemDefined :: forall b. Lens' (FunctionInfo b) SystemDefined

fiSQLName :: forall b. Lens' (FunctionInfo b) (FunctionName b)

fiReturnType :: forall b. Lens' (FunctionInfo b) (TableName b)

fiPermissions :: forall b. Lens' (FunctionInfo b) FunctionPermissionsMap

fiJsonAggSelect :: forall b. Lens' (FunctionInfo b) JsonAggSelect

fiInputArgs :: forall b. Lens' (FunctionInfo b) (Seq (FunctionInputArgument b))

fiGQLName :: forall b. Lens' (FunctionInfo b) Name

fiGQLArgsName :: forall b. Lens' (FunctionInfo b) Name

fiGQLAggregateName :: forall b. Lens' (FunctionInfo b) Name

fiExposedAs :: forall b. Lens' (FunctionInfo b) FunctionExposedAs

fiDescription :: forall b. Lens' (FunctionInfo b) (Maybe Text)

fiComment :: forall b. Lens' (FunctionInfo b) (Maybe Text)

getFunctionArgsGQLName

Arguments

:: Name
The GQL version of the DB name of the function

-> FunctionConfig

-> Name

Apply function name customization to function arguments, as detailed in 'rfcs/function-root-field-customisation.md'. We want the different variations of a function (i.e. basic, aggregate) to share the same type name for their arguments.

getFunctionGQLName :: Name -> FunctionConfig -> Name

Apply function name customization to the basic function variation, as detailed in 'rfcs/function-root-field-customisation.md'.

getFunctionAggregateGQLName :: Name -> FunctionConfig -> Name

Apply function name customization to the aggregate function variation, as detailed in 'rfcs/function-root-field-customisation.md'.

getInputArgs :: FunctionInfo b -> Seq (FunctionArg b)

type FunctionCache b = HashMap (FunctionName b) (FunctionInfo b)

data FunctionConfig

Tracked function configuration, and payload of the pg_track_function and pg_set_function_customization API calls.

Constructors

FunctionConfig

Fields

_fcSessionArgument :: Maybe FunctionArgName

_fcExposedAs :: Maybe FunctionExposedAs

In which top-level field should we expose this function?

The user might omit this, in which case we'll infer the location from the SQL functions volatility. See mkFunctionInfo or the track_function API docs for details of validation, etc.

_fcCustomRootFields :: FunctionCustomRootFields

_fcCustomName :: Maybe Name

emptyFunctionConfig :: FunctionConfig

The default function config; v1 of the API implies this.

type DBFunctionsMetadata b = HashMap (FunctionName b) [RawFunctionInfo b]

data FunctionPermissionsCtx

Constructors

FunctionPermissionsInferred

FunctionPermissionsManual