Hasura.RQL.Types.HealthCheckImplementation

Description

Provides a data type that holds all of the required implementation details for a backend that supports health checks.

Documentation

data HealthCheckImplementation healthCheckTest

Constructors

HealthCheckImplementation

Fields

_hciDefaultTest :: healthCheckTest

_hciTestCodec :: JSONCodec healthCheckTest