Hasura.RQL.Types.Metadata

Documentation

data MetadataVersion

Versioning the Metadata JSON structure to track backwards incompatible changes. This value is included in the metadata JSON object at top level version key. Always metadata is emitted in the latest version via export metadata API (runExportMetadata handler). Adding a new value constructor to @MetadataVersion type bumps the metadata version.

NOTE: When metadata version is bumped: 1. The Hasura CLI and Console actively use export metadata API to read metadata. Hence, it is necessary to update CLI and Console to read latest metadata. All changes SHOULD be released hand in hand (preferebly in one pull request) 2. There might be other third party services (developed by Hasura users) which use the export metadata API. Apart from changelog, we need to establish the metadata version update by bumping up the minor version of the GraphQL Engine.

Constructors

MVVersion1

MVVersion2

MVVersion3

currentMetadataVersion :: MetadataVersion

data Metadata

A complete GraphQL Engine metadata representation to be stored, exported/replaced via metadata queries.

Constructors

Metadata

Fields

_metaSources :: !Sources

_metaRemoteSchemas :: !RemoteSchemas

_metaQueryCollections :: !QueryCollections

_metaAllowlist :: !MetadataAllowlist

_metaCustomTypes :: !CustomTypes

_metaActions :: !Actions

_metaCronTriggers :: !CronTriggers

_metaRestEndpoints :: !Endpoints

_metaApiLimits :: !ApiLimit

_metaMetricsConfig :: !MetricsConfig

_metaInheritedRoles :: !InheritedRoles

_metaSetGraphqlIntrospectionOptions :: !SetGraphqlIntrospectionOptions

_metaNetwork :: !Network

_metaBackendConfigs :: BackendMap BackendConfigWrapper

metaSources :: Lens' Metadata Sources

metaSetGraphqlIntrospectionOptions :: Lens' Metadata SetGraphqlIntrospectionOptions

metaRestEndpoints :: Lens' Metadata Endpoints

metaRemoteSchemas :: Lens' Metadata RemoteSchemas

metaQueryCollections :: Lens' Metadata QueryCollections

metaNetwork :: Lens' Metadata Network

metaMetricsConfig :: Lens' Metadata MetricsConfig

metaInheritedRoles :: Lens' Metadata InheritedRoles

metaCustomTypes :: Lens' Metadata CustomTypes

metaCronTriggers :: Lens' Metadata CronTriggers

metaBackendConfigs :: Lens' Metadata (BackendMap BackendConfigWrapper)

metaApiLimits :: Lens' Metadata ApiLimit

metaAllowlist :: Lens' Metadata MetadataAllowlist

metaActions :: Lens' Metadata Actions

emptyMetadata :: Metadata

tableMetadataSetter :: Backend b => SourceName -> TableName b -> ASetter' Metadata (TableMetadata b)

functionMetadataSetter :: Backend b => SourceName -> FunctionName b -> ASetter' Metadata (FunctionMetadata b)

A lens setter for the metadata of a specific function as identified by the source name and function name.

class Monad m => MetadataM m where

A simple monad class which enables fetching and setting @Metadata in the state.

Methods

getMetadata :: m Metadata

putMetadata :: Metadata -> m ()

data MetadataNoSources

Constructors

MetadataNoSources

Fields

_mnsTables :: !(Tables ('Postgres 'Vanilla))

_mnsFunctions :: !(Functions ('Postgres 'Vanilla))

_mnsRemoteSchemas :: !RemoteSchemas

_mnsQueryCollections :: !QueryCollections

_mnsAllowlist :: !MetadataAllowlist

_mnsCustomTypes :: !CustomTypes

_mnsActions :: !Actions

_mnsCronTriggers :: !CronTriggers

newtype MetadataModifier

Constructors

MetadataModifier

Fields

runMetadataModifier :: Metadata -> Metadata

dropTableInMetadata :: forall b. Backend b => SourceName -> TableName b -> MetadataModifier

dropRelationshipInMetadata :: RelName -> TableMetadata b -> TableMetadata b

dropPermissionInMetadata :: RoleName -> PermType -> TableMetadata b -> TableMetadata b

dropComputedFieldInMetadata :: ComputedFieldName -> TableMetadata b -> TableMetadata b

dropEventTriggerInMetadata :: TriggerName -> TableMetadata b -> TableMetadata b

dropRemoteRelationshipInMetadata :: RelName -> TableMetadata b -> TableMetadata b

dropFunctionInMetadata :: forall b. Backend b => SourceName -> FunctionName b -> MetadataModifier

dropRemoteSchemaInMetadata :: RemoteSchemaName -> MetadataModifier

dropRemoteSchemaPermissionInMetadata :: RemoteSchemaName -> RoleName -> MetadataModifier

dropRemoteSchemaRemoteRelationshipInMetadata :: RemoteSchemaName -> Name -> RelName -> MetadataModifier

metadataToOrdJSON :: Metadata -> Value

Encode Metadata to JSON with deterministic ordering (e.g. "version" being at the top). The CLI system stores metadata in files and has option to show changes in git diff style. The diff shouldn't appear different for no metadata changes. So, the ordering of object keys and array elements should remain consistent across versions of graphql-engine if possible.

Note: While modifying any part of the code below, make sure the encoded JSON of each type is parsable via its FromJSON instance.

TODO: we can use 'aeson-pretty' to serialize in a consistent way, and to specify a (partial) order of keys, while getting the benefits of auto-generated To/FromJSON instances. `FromJSON TableMetadata` complicates this though...

See: https://github.com/hasura/graphql-engine/issues/6348