Hasura.RQL.Types.Metadata.Backend

Documentation

class (Backend b, Eq (BooleanOperators b (PartialSQLExp b)), Hashable (BooleanOperators b (PartialSQLExp b))) => BackendMetadata (b :: BackendType) where

Minimal complete definition

buildComputedFieldInfo, fetchAndValidateEnumValues, resolveSourceConfig, resolveDatabaseMetadata, parseBoolExpOperations, buildFunctionInfo, updateColumnInEventTrigger, parseCollectableType, postDropSourceHook

Methods

buildComputedFieldInfo :: MonadError QErr m => HashSet (TableName b) -> TableName b -> ComputedFieldName -> ComputedFieldDefinition b -> RawFunctionInfo b -> Comment -> m (ComputedFieldInfo b)

fetchAndValidateEnumValues :: (MonadIO m, MonadBaseControl IO m) => SourceConfig b -> TableName b -> Maybe (PrimaryKey b (RawColumnInfo b)) -> [RawColumnInfo b] -> m (Either QErr EnumValues)

resolveSourceConfig :: (MonadIO m, MonadResolveSource m) => SourceName -> SourceConnConfiguration b -> Environment -> m (Either QErr (SourceConfig b))

Function that resolves the connection related source configuration, and creates a connection pool (and other related parameters) in the process

resolveDatabaseMetadata :: (MonadIO m, MonadBaseControl IO m, MonadResolveSource m) => SourceConfig b -> SourceTypeCustomization -> m (Either QErr (ResolvedSource b))

Function that introspects a database for tables, columns, functions etc.

parseBoolExpOperations :: (MonadError QErr m, TableCoreInfoRM b m) => ValueParser b m v -> TableName b -> FieldInfoMap (FieldInfo b) -> ColumnReference b -> Value -> m [OpExpG b v]

buildFunctionInfo

Arguments

:: MonadError QErr m

=> SourceName

-> FunctionName b

-> SystemDefined

-> FunctionConfig

-> FunctionPermissionsMap

-> RawFunctionInfo b

-> Maybe Text
the function comment

-> m (FunctionInfo b, SchemaDependency)

updateColumnInEventTrigger :: TableName b -> Column b -> Column b -> TableName b -> EventTriggerConf b -> EventTriggerConf b

parseCollectableType :: MonadError QErr m => CollectableType (ColumnType b) -> Value -> m (PartialSQLExp b)

postDropSourceHook :: (MonadError QErr m, MonadIO m, MonadBaseControl IO m) => SourceConfig b -> m ()

validateRelationship :: MonadError QErr m => TableCache b -> TableName b -> Either (ObjRelDef b) (ArrRelDef b) -> m ()

default validateRelationship :: MonadError QErr m => TableCache b -> TableName b -> Either (ObjRelDef b) (ArrRelDef b) -> m ()