Hasura.RQL.Types.OpenTelemetry

User-facing configuration (metadata)

data OpenTelemetryConfig

Metadata configuration for all OpenTelemetry-related features

Constructors

OpenTelemetryConfig

Fields

_ocStatus :: OtelStatus

_ocEnabledDataTypes :: Set OtelDataType

_ocExporterOtlp :: OtelExporterConfig

_ocBatchSpanProcessor :: OtelBatchSpanProcessorConfig

emptyOpenTelemetryConfig :: OpenTelemetryConfig

data OpenTelemetryConfigSubobject

Subsets of the fields of OpenTelemetryConfig, serving as metadata object names for MetadataObjId.

Constructors

OtelSubobjectAll
The entire OpenTelemetry configuration

OtelSubobjectExporterOtlp

OtelSubobjectBatchSpanProcessor

data OtelStatus

Should the OpenTelemetry exporter be enabled?

Constructors

OtelEnabled

OtelDisabled

data OtelDataType

Constructors

OtelTraces

data OtelExporterConfig

https://opentelemetry.io/docs/reference/specification/protocol/exporter/

Constructors

OtelExporterConfig

Fields

_oecTracesEndpoint :: Maybe Text

Target URL to which the exporter is going to send traces. No default.

_oecProtocol :: OtlpProtocol

The transport protocol

_oecHeaders :: [HeaderConf]

Key-value pairs to be used as headers to send with an export request.

_oecResourceAttributes :: [NameValue]

Attributes to send as the resource attributes of an export request. We currently only support string-valued attributes.

defaultOtelExporterConfig :: OtelExporterConfig

data OtlpProtocol

Possible protocol to use with OTLP. Currently, only http/protobuf is supported.

Constructors

OtlpProtocolHttpProtobuf

data NameValue

Constructors

NameValue

Fields

nv_name :: Text

nv_value :: Text

newtype OtelBatchSpanProcessorConfig

Constructors

OtelBatchSpanProcessorConfig

Fields

_obspcMaxExportBatchSize :: Int

The maximum batch size of every export. It must be smaller or equal to maxQueueSize (not yet configurable). Default 512.

defaultOtelBatchSpanProcessorConfig :: OtelBatchSpanProcessorConfig

ocStatus :: Lens' OpenTelemetryConfig OtelStatus

ocExporterOtlp :: Lens' OpenTelemetryConfig OtelExporterConfig

ocEnabledDataTypes :: Lens' OpenTelemetryConfig (Set OtelDataType)

ocBatchSpanProcessor :: Lens' OpenTelemetryConfig OtelBatchSpanProcessorConfig

Parsed configuration (schema cache)

data OpenTelemetryInfo

Schema cache configuration for all OpenTelemetry-related features

Constructors

OpenTelemetryInfo

Fields

_otiExporterOtlp :: Maybe OtelExporterInfo

_otiBatchSpanProcessor :: Maybe OtelBatchSpanProcessorInfo

A value of Nothing indicates that the export of trace data is disabled.

emptyOpenTelemetryInfo :: OpenTelemetryInfo

data OtelExporterInfo

Constructors

OtelExporterInfo

Fields

_oteleiTracesBaseRequest :: Request

HTTP Request containing (1) the target URL to which the exporter is going to send spans, and (2) the user-specified request headers.

_oteleiResourceAttributes :: Map Text Text

Attributes to send as the resource attributes of an export request. We currently only support string-valued attributes.

Using Data.Map.Strict over Data.Hashmap.Strict because currently the only operations on data are (1) folding and (2) union with a small map of default attributes, and Map should be is faster than HashMap for the latter.

getOtelExporterTracesBaseRequest :: OtelExporterInfo -> Request

getOtelExporterResourceAttributes :: OtelExporterInfo -> Map Text Text

data OtelBatchSpanProcessorInfo

Constructors

OtelBatchSpanProcessorInfo

Fields

_obspiMaxExportBatchSize :: Int

The maximum batch size of every export. It must be smaller or equal to maxQueueSize. Default 512.

_obspiMaxQueueSize :: Int

The maximum span queue size. After the size is reached spans are dropped. Default 2048.

getMaxExportBatchSize :: OtelBatchSpanProcessorInfo -> Int

getMaxQueueSize :: OtelBatchSpanProcessorInfo -> Int

defaultOtelBatchSpanProcessorInfo :: OtelBatchSpanProcessorInfo

Defaults taken from https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/sdk.md#batching-processor

otiExporterOtlp :: Lens' OpenTelemetryInfo (Maybe OtelExporterInfo)

otiBatchSpanProcessor :: Lens' OpenTelemetryInfo (Maybe OtelBatchSpanProcessorInfo)