Hasura.RQL.Types.Permission

Documentation

data PermType

Constructors

PTInsert

PTSelect

PTUpdate

PTDelete

permTypeToCode :: PermType -> Text

data PermColSpec b

Constructors

PCStar

PCCols [Column b]

data PermDef (b :: BackendType) (perm :: BackendType -> Type)

Constructors

PermDef

Fields

_pdRole :: RoleName

_pdPermission :: PermDefPermission b perm

_pdComment :: Maybe Text

data PermDefPermission (b :: BackendType) (perm :: BackendType -> Type) where

The permission data as it appears in a PermDef. Since this type is a GADT it facilitates that values which are polymorphic may re-discover its specific type of permission by case analysis.

The fact that permission types are tracked in types are more accidental than intentional and something we want to move away from, see https://github.com/hasura/graphql-engine-mono/issues/4076.

Constructors

SelPerm' :: SelPerm b -> PermDefPermission b SelPerm

InsPerm' :: InsPerm b -> PermDefPermission b InsPerm

UpdPerm' :: UpdPerm b -> PermDefPermission b UpdPerm

DelPerm' :: DelPerm b -> PermDefPermission b DelPerm

unPermDefPermission :: PermDefPermission b perm -> perm b

reflectPermDefPermission :: PermDefPermission b a -> PermType

data QueryRootFieldType

Constructors

QRFTSelect

QRFTSelectByPk

QRFTSelectAggregate

data SubscriptionRootFieldType

Constructors

SRFTSelect

SRFTSelectByPk

SRFTSelectAggregate

SRFTSelectStream

data InsPerm (b :: BackendType)

Constructors

InsPerm

Fields

ipCheck :: BoolExp b

ipSet :: Maybe (ColumnValues b Value)

ipColumns :: Maybe (PermColSpec b)

ipBackendOnly :: Bool

ipValidateInput :: Maybe (ValidateInput InputWebhook)

type InsPermDef b = PermDef b InsPerm

data AllowedRootFields rootFieldType

Constructors

ARFAllowAllRootFields

ARFAllowConfiguredRootFields (HashSet rootFieldType)

isRootFieldAllowed :: Eq rootField => rootField -> AllowedRootFields rootField -> Bool

data ValidateInputHttpDefinition webhook

Constructors

ValidateInputHttpDefinition

Fields

_vihdUrl :: webhook

_vihdHeaders :: [HeaderConf]

_vihdTimeout :: Timeout

_vihdForwardClientHeaders :: Bool

data ValidateInput webhook

Constructors

VIHttp (ValidateInputHttpDefinition webhook)

data SelPerm (b :: BackendType)

Constructors

SelPerm

Fields

spColumns :: PermColSpec b

Allowed columns

spFilter :: BoolExp b

Filter expression

spLimit :: Maybe Int

Limit value

spAllowAggregations :: Bool

Allow aggregation

spComputedFields :: [ComputedFieldName]

Allowed computed fields which should not include the fields returning rows of existing table.

spAllowedQueryRootFields :: AllowedRootFields QueryRootFieldType

spAllowedSubscriptionRootFields :: AllowedRootFields SubscriptionRootFieldType

type SelPermDef b = PermDef b SelPerm

data DelPerm (b :: BackendType)

Constructors

DelPerm

Fields

dcFilter :: BoolExp b

dcBackendOnly :: Bool

type DelPermDef b = PermDef b DelPerm

data UpdPerm (b :: BackendType)

Constructors

UpdPerm

Fields

ucColumns :: PermColSpec b

ucSet :: Maybe (ColumnValues b Value)

ucFilter :: BoolExp b

ucCheck :: Maybe (BoolExp b)

Check expression, which must be true after update. This is optional because we don't want to break the v1 API but Nothing should be equivalent to the expression which always returns true.

ucBackendOnly :: Bool

type UpdPermDef b = PermDef b UpdPerm

pdRole :: forall b perm. Lens' (PermDef b perm) RoleName

pdPermission :: forall b perm b perm. Lens (PermDef b perm) (PermDef b perm) (PermDefPermission b perm) (PermDefPermission b perm)

pdComment :: forall b perm. Lens' (PermDef b perm) (Maybe Text)