Hasura.RQL.Types.Relationships.Remote

Documentation

data RemoteRelationship

Metadata representation of a generic remote relationship, regardless of the source: all sources use this same agnostic definition. The internal definition field is where we differentiate between different targets.

Constructors

RemoteRelationship

Fields

_rrName :: RelName

_rrDefinition :: RemoteRelationshipDefinition

data RRFormat

Represents the format of the metadata a remote relationship was read from and must be written back as. We don't have a good way of doing metadata versioning yet, and we therefore use this to keep track of the format used.

Constructors

RRFOldDBToRemoteSchema
The remote relationship was parsed from the old format, that was only used only for db-to-rs schemas.

RRFUnifiedFormat
The remote relationship was parsed from the new unified format.

data RemoteRelationshipDefinition

Metadata representation of the internal definition of a remote relationship.

Constructors

RelationshipToSource ToSourceRelationshipDef
Remote relationship targetting a source.

RelationshipToSchema RRFormat ToSchemaRelationshipDef
Remote relationship targetting a remote schema.

data RRParseMode

Whether to accept legacy fields when parsing RemoteRelationshipDefinition

Constructors

RRPLegacy
Only allow legacy fields when parsing RemoteRelationshipDefinition

RRPLenient
Allow legacy fields when parsing RemoteRelationshipDefinition

RRPStrict
Reject legacy fields when parsing RemoteRelationshipDefinition

parseRemoteRelationshipDefinition :: RRParseMode -> Value -> Parser RemoteRelationshipDefinition

Parse RemoteRelationshipDefinition letting the caller decide how lenient to be.

This is necessary because RemoteRelationshipDefinition is parsed in different contexts. In RemoteRelationship, the RemoteRelationshipDefinition is always parsed out from a top-level @"definition" field. Thus, a legacy payload looks like this:

{
  "name": "thing",
  "definition": {
    "remote_schema": "stuff",
    "hasura_fields": ...
    "remote_field": ...
  }
}

and a new payload looks like this:

{
  "name": "thing",
  "definition": {
    "to_remote_schema": {
      "schema": "stuff",
      "lhs_fields": ...
      "remote_field": ...
    }
  }
}

In contrast, CreateFromSourceRelationship does not have a top- level "definition" in its legacy format. Instead, the legacy fields themselves are top-level:

{
  "remote_schema": "stuff",
  "hasura_fields": ...
  "remote_field": ...
}

Furthermore, the presence of a "definition" field is used to detect that the new payload is being used:

{
  "definition": {
    "to_remote_schema": {
      "schema": "stuff",
      "lhs_fields": ...
      "remote_field": ...
    }
  }
}

In this latter case, we should not allow "remote_schema" to appear under "definition".

data RemoteFieldInfo lhsJoinField

Resolved remote relationship, as stored in the schema cache.

Constructors

RemoteFieldInfo

Fields

_rfiLHS :: HashMap FieldName lhsJoinField

_rfiRHS :: RemoteFieldInfoRHS

data RemoteFieldInfoRHS

Resolved remote relationship's RHS

Constructors

RFISchema !RemoteSchemaFieldInfo

RFISource !(AnyBackend RemoteSourceFieldInfo)

data DBJoinField (b :: BackendType)

Information about the field on the LHS of a join against a remote schema.

Constructors

JoinColumn !(Column b) !(ColumnType b)

JoinComputedField !(ScalarComputedField b)

data ScalarComputedField (b :: BackendType)

Information about a computed field appearing on the LHS of a remote join. FIXME: why do we need all of this?

Constructors

ScalarComputedField

Fields

_scfXField :: !(XComputedField b)

_scfName :: !ComputedFieldName

_scfFunction :: !(FunctionName b)

_scfTableArgument :: !FunctionTableArgument

_scfSessionArgument :: !(Maybe FunctionSessionArgument)

_scfType :: !(ScalarType b)

rrName :: Lens' RemoteRelationship RelName

rrDefinition :: Lens' RemoteRelationship RemoteRelationshipDefinition

_RelationshipToSchema :: Prism' RemoteRelationshipDefinition (RRFormat, ToSchemaRelationshipDef)

_RelationshipToSource :: Prism' RemoteRelationshipDefinition ToSourceRelationshipDef