Hasura.RQL.Types.Relationships.ToSchema

Documentation

data ToSchemaRelationshipDef

Metadata representation of a relationship to a remote schema.

FIXME: move this to Hasura/Metadata

Constructors

ToSchemaRelationshipDef

Fields

_trrdRemoteSchema :: RemoteSchemaName

Identifier for this mapping.

_trrdLhsFields :: HashSet FieldName

The lhs fields that must be forwarded to the remote schema.

_trrdRemoteField :: RemoteFields

newtype RemoteFields

Targeted field in a remote schema relationship. TODO: explain about subfields and why this is a container

Constructors

RemoteFields

Fields

unRemoteFields :: NonEmpty FieldCall

data FieldCall

Associates a field name with the arguments it will be passed in the query.

https://graphql.github.io/graphql-spec/June2018/#sec-Language.Arguments

Constructors

FieldCall

Fields

fcName :: Name

fcArguments :: RemoteArguments

newtype RemoteArguments

Arguments to a remote GraphQL fields, represented as a mapping from name to GraphQL Value. Said values can be variable names, in which case they'll be referring to values we're closed over. TODO: expand on this

Constructors

RemoteArguments

Fields

getRemoteArguments :: HashMap Name (Value Name)

newtype LHSIdentifier

Constructors

LHSIdentifier

Fields

getLHSIdentifier :: Text

tableNameToLHSIdentifier :: Backend b => TableName b -> LHSIdentifier

remoteSchemaToLHSIdentifier :: RemoteSchemaName -> LHSIdentifier

lhsIdentifierToGraphQLName :: LHSIdentifier -> Maybe Name

Generates a valid graphql name from an arbitrary LHS identifier. This is done by replacing all unrecognized characters by '_'. This function still returns a Maybe value, in cases we can't adjust the raw text (such as the case of empty identifiers).

data RemoteSchemaFieldInfo

Schema cache information for a table field targeting a remote schema.

Constructors

RemoteSchemaFieldInfo

Fields

_rrfiName :: RelName

Field name to which we'll map the remote in hasura; this becomes part of the hasura schema.

_rrfiParamMap :: HashMap Name RemoteSchemaInputValueDefinition

Input arguments to the remote field info; The _rfiParamMap will only include the arguments to the remote field that is being joined. The names of the arguments here are modified, it will be in the format of Field Name_remote_rel_table schema_table namerelationship name

_rrfiRemoteFields :: RemoteFields

_rrfiRemoteSchema :: RemoteSchemaInfo

_rrfiInputValueDefinitions :: [TypeDefinition [Name] RemoteSchemaInputValueDefinition]

The new input value definitions created for this remote field

_rrfiRemoteSchemaName :: RemoteSchemaName

Name of the remote schema, that's used for joining

_rrfiLHSIdentifier :: LHSIdentifier

TODO: this one should be gone when validateRemoteRelationship function is cleaned up

graphQLValueToJSON :: Value Void -> Value

trrdRemoteSchema :: Lens' ToSchemaRelationshipDef RemoteSchemaName

trrdRemoteField :: Lens' ToSchemaRelationshipDef RemoteFields

trrdLhsFields :: Lens' ToSchemaRelationshipDef (HashSet FieldName)