Hasura.RQL.Types.RemoteSchema

Documentation

type UrlFromEnv = Text

newtype RemoteSchemaName

Remote schema identifier.

NOTE: no validation on the character set is done here; it's likely there is a bug (FIXME) where this interacts with remote relationships and some name mangling needs to happen.

Constructors

RemoteSchemaName

Fields

unRemoteSchemaName :: NonEmptyText

data RemoteTypeCustomization

Constructors

RemoteTypeCustomization

Fields

_rtcPrefix :: Maybe Name

_rtcSuffix :: Maybe Name

_rtcMapping :: HashMap Name Name

data RemoteFieldCustomization

Constructors

RemoteFieldCustomization

Fields

_rfcParentType :: Name

_rfcPrefix :: Maybe Name

_rfcSuffix :: Maybe Name

_rfcMapping :: HashMap Name Name

data RemoteSchemaCustomization

Constructors

RemoteSchemaCustomization

Fields

_rscRootFieldsNamespace :: Maybe Name

_rscTypeNames :: Maybe RemoteTypeCustomization

_rscFieldNames :: Maybe [RemoteFieldCustomization]

data ValidatedRemoteSchemaDef

RemoteSchemaDef after validation and baking-in of defaults in validateRemoteSchemaDef.

Constructors

ValidatedRemoteSchemaDef

Fields

_vrsdUrl :: EnvRecord URI

_vrsdHeaders :: [HeaderConf]

_vrsdFwdClientHeaders :: Bool

_vrsdTimeoutSeconds :: Int

_vrsdCustomization :: Maybe RemoteSchemaCustomization

See _rsdCustomization.

data RemoteSchemaCustomizer

Constructors

RemoteSchemaCustomizer

Fields

_rscNamespaceFieldName :: Maybe Name

_rscCustomizeTypeName :: HashMap Name Name

type name -> type name

_rscCustomizeFieldName :: HashMap Name (HashMap Name Name)

type name -> field name -> field name

identityCustomizer :: RemoteSchemaCustomizer

remoteSchemaCustomizeTypeName :: RemoteSchemaCustomizer -> MkTypename

remoteSchemaCustomizeFieldName :: RemoteSchemaCustomizer -> CustomizeRemoteFieldName

hasTypeOrFieldCustomizations :: RemoteSchemaCustomizer -> Bool

data RemoteSchemaInfo

RemoteSchemaDef after the RemoteSchemaCustomizer has been generated by fetchRemoteSchema

Constructors

RemoteSchemaInfo

Fields

rsDef :: ValidatedRemoteSchemaDef

rsCustomizer :: RemoteSchemaCustomizer

data RemoteSchemaDef

Unvalidated remote schema config, from the user's API request

Constructors

RemoteSchemaDef

Fields

_rsdUrl :: Maybe InputWebhook

_rsdUrlFromEnv :: Maybe UrlFromEnv

_rsdHeaders :: Maybe [HeaderConf]

_rsdForwardClientHeaders :: Bool

_rsdTimeoutSeconds :: Maybe Int

_rsdCustomization :: Maybe RemoteSchemaCustomization

data AddRemoteSchemaQuery

The payload for add_remote_schema, and a component of Metadata.

Constructors

AddRemoteSchemaQuery

Fields

_arsqName :: RemoteSchemaName

An internal identifier for this remote schema.

_arsqDefinition :: RemoteSchemaDef

_arsqComment :: Maybe Text

An opaque description or comment. We might display this in the UI, for instance.

newtype RemoteSchemaNameQuery

Constructors

RemoteSchemaNameQuery

Fields

_rsnqName :: RemoteSchemaName

getUrlFromEnv :: (MonadIO m, MonadError QErr m) => Environment -> Text -> m (EnvRecord URI)

validateRemoteSchemaCustomization :: MonadError QErr m => Maybe RemoteSchemaCustomization -> m ()

validateRemoteSchemaDef :: (MonadError QErr m, MonadIO m) => Environment -> RemoteSchemaDef -> m ValidatedRemoteSchemaDef

newtype RemoteSchemaPermissionDefinition

Constructors

RemoteSchemaPermissionDefinition

Fields

_rspdSchema :: SchemaDocument

data AddRemoteSchemaPermission

Constructors

AddRemoteSchemaPermission

Fields

_arspRemoteSchema :: RemoteSchemaName

_arspRole :: RoleName

_arspDefinition :: RemoteSchemaPermissionDefinition

_arspComment :: Maybe Text

data DropRemoteSchemaPermissions

Constructors

DropRemoteSchemaPermissions

Fields

_drspRemoteSchema :: RemoteSchemaName

_drspRole :: RoleName

data SessionArgumentPresetInfo

See resolveRemoteVariable function. This data type is used for validation of the session variable value

Constructors

SessionArgumentPresetScalar

SessionArgumentPresetEnum (HashSet EnumValue)

data RemoteSchemaVariable

Details required to resolve a "session variable preset" variable.

See Notes [Remote Schema Argument Presets] and [Remote Schema Permissions Architecture] for additional information.

Constructors

SessionPresetVariable SessionVariable Name SessionArgumentPresetInfo

QueryVariable Variable

RemoteJSONValue GType Value

data RemoteSchemaInputValueDefinition

Extends InputValueDefinition with an optional preset argument.

See Note [Remote Schema Argument Presets] for additional information.

Constructors

RemoteSchemaInputValueDefinition

Fields

_rsitdDefinition :: InputValueDefinition

_rsitdPresetArgument :: Maybe (Value RemoteSchemaVariable)

newtype RemoteSchemaIntrospection

Constructors

RemoteSchemaIntrospection (HashMap Name (TypeDefinition [Name] RemoteSchemaInputValueDefinition))

getTypeName :: TypeDefinition possibleTypes inputType -> Name

Extracts the name of a given type from its definition. TODO: move this to Language.GraphQL.Draft.Syntax.

lookupType :: RemoteSchemaIntrospection -> Name -> Maybe (TypeDefinition [Name] RemoteSchemaInputValueDefinition)

lookupObject :: RemoteSchemaIntrospection -> Name -> Maybe (ObjectTypeDefinition RemoteSchemaInputValueDefinition)

lookupInterface :: RemoteSchemaIntrospection -> Name -> Maybe (InterfaceTypeDefinition [Name] RemoteSchemaInputValueDefinition)

lookupScalar :: RemoteSchemaIntrospection -> Name -> Maybe ScalarTypeDefinition

lookupUnion :: RemoteSchemaIntrospection -> Name -> Maybe UnionTypeDefinition

lookupEnum :: RemoteSchemaIntrospection -> Name -> Maybe EnumTypeDefinition

lookupInputObject :: RemoteSchemaIntrospection -> Name -> Maybe (InputObjectTypeDefinition RemoteSchemaInputValueDefinition)