Hasura.RQL.Types.ResultCustomization

Documentation

newtype AliasMapping

Mapping that can be provided to a ResultCustomizer to map top-level field aliases that were not available at field parse time. E.g. for aliases created in the remote server query for remote joins.

Constructors

AliasMapping

Fields

unAliasMapping :: Name -> Name

singletonAliasMapping :: Name -> Name -> AliasMapping

AliasMapping that maps a single field name to an alias

newtype ResultCustomizer

Function to modify JSON values returned from the remote server e.g. to map values of __typename fields to customized type names. The customizer uses Maybe to allow short-circuiting subtrees where no customizations are needed.

Constructors

ResultCustomizer

Fields

unResultCustomizer :: AliasMapping -> Value -> Value

applyResultCustomizer :: ResultCustomizer -> Value -> Value

Apply a ResultCustomizer to a JSON value

applyAliasMapping :: AliasMapping -> ResultCustomizer -> ResultCustomizer

Apply an AliasMapping to a ResultCustomizer.

modifyFieldByName :: Name -> ResultCustomizer -> ResultCustomizer

Take a ResultCustomizer for a JSON subtree, and a fieldName, and produce a ResultCustomizer for a parent object or array of objects that applies the subtree customizer to the subtree at the given fieldName.

customizeTypeNameString :: HashMap Name Name -> ResultCustomizer

Create a RemoteResultCustomizer that applies the typeNameMap to a JSON string value, e.g. for use in customizing a __typename field value.