Hasura.RQL.Types.Roles

Documentation

data RoleName

roleNameToTxt :: RoleName -> Text

mkRoleName :: Text -> Maybe RoleName

mkRoleNameSafe :: NonEmptyText -> RoleName

adminRoleName :: RoleName

newtype ParentRoles

Constructors

ParentRoles

Fields

_unParentRoles :: HashSet RoleName

data Role

The Role type represents a role by containing its name and the names of its parent roles. This type is used externally in the add_inherited_role metadata API and is also used internally in the permission building part of the schema cache building process

Constructors

Role

Fields

_rRoleName :: RoleName

_rParentRoles :: ParentRoles

set of the parent role names, in case of non-inherited roles it will be an empty set

type InheritedRole = Role

newtype DropInheritedRole

Constructors

DropInheritedRole

Fields

_ddrRoleName :: RoleName