Hasura.RQL.Types.Roles.Internal

Documentation

data CheckPermission permissionType

CheckPermission is a type which can be used to combine multiple permissions when the permission type implements the OnlyRelevantEq instance

Constructors

CPUndefined

CPInconsistent

CPDefined permissionType

data CombineRolePermInfo (b :: BackendType)

CombineRolePermInfo acts as an intermediate type to be able to combine multiple role permissions into one, using the Monoid instance. Multiple role permissions are combined for inherited role permissions where this is used.

Constructors

CombineRolePermInfo

Fields

crpiInsPerm :: CheckPermission (InsPermInfo b)

crpiSelPerm :: Maybe (CombinedSelPermInfo b)

crpiUpdPerm :: CheckPermission (UpdPermInfo b)

crpiDelPerm :: CheckPermission (DelPermInfo b)

rolePermInfoToCombineRolePermInfo :: RolePermInfo b -> CombineRolePermInfo b

maybeToCheckPermission :: Maybe a -> CheckPermission a