Hasura.RQL.Types.ScheduledTrigger

Description

These are types for Scheduled Trigger definition; see Hasura.Eventing.ScheduledTrigger

Documentation

type CronEventId = EventId

type OneOffScheduledEventId = EventId

type ScheduledEventId = EventId

type InvocationId = Text

data STRetryConf

Constructors

STRetryConf

Fields

strcNumRetries :: !Int

strcRetryIntervalSeconds :: !NonNegativeDiffTime

strcTimeoutSeconds :: !NonNegativeDiffTime

strcToleranceSeconds :: !NonNegativeDiffTime

The tolerance configuration is used to determine whether a scheduled event is not too old to process. The age of the scheduled event is the difference between the current timestamp and the scheduled event's timestamp, if the age is than the tolerance then the scheduled event is marked as dead.

defaultSTRetryConf :: STRetryConf

data CronTriggerMetadata

Constructors

CronTriggerMetadata

Fields

ctName :: !TriggerName

ctWebhook :: !InputWebhook

ctSchedule :: !CronSchedule

ctPayload :: !(Maybe Value)

ctRetryConf :: !STRetryConf

ctHeaders :: ![HeaderConf]

ctIncludeInMetadata :: !Bool

ctComment :: !(Maybe Text)

ctRequestTransform :: !(Maybe RequestTransform)

ctResponseTransform :: !(Maybe MetadataResponseTransform)

data CreateCronTrigger

Constructors

CreateCronTrigger

Fields

_cctName :: !TriggerName

_cctWebhook :: !InputWebhook

_cctCronSchedule :: !CronSchedule

_cctPayload :: !(Maybe Value)

_cctRetryConf :: !STRetryConf

_cctHeaders :: ![HeaderConf]

_cctIncludeInMetadata :: !Bool

_cctComment :: !(Maybe Text)

_cctReplace :: !Bool

_cctRequestTransform :: !(Maybe RequestTransform)

_cctResponseTransform :: !(Maybe MetadataResponseTransform)

cctWebhook :: Lens' CreateCronTrigger InputWebhook

cctRetryConf :: Lens' CreateCronTrigger STRetryConf

cctResponseTransform :: Lens' CreateCronTrigger (Maybe MetadataResponseTransform)

cctRequestTransform :: Lens' CreateCronTrigger (Maybe RequestTransform)

cctReplace :: Lens' CreateCronTrigger Bool

cctPayload :: Lens' CreateCronTrigger (Maybe Value)

cctName :: Lens' CreateCronTrigger TriggerName

cctIncludeInMetadata :: Lens' CreateCronTrigger Bool

cctHeaders :: Lens' CreateCronTrigger [HeaderConf]

cctCronSchedule :: Lens' CreateCronTrigger CronSchedule

cctComment :: Lens' CreateCronTrigger (Maybe Text)

newtype ScheduledTriggerName

Constructors

ScheduledTriggerName

Fields

unName :: TriggerName

formatTime' :: UTCTime -> Text

data CreateScheduledEvent

Constructors

CreateScheduledEvent

Fields

cseWebhook :: !InputWebhook

cseScheduleAt :: !UTCTime

The timestamp should be in the 8601 https://en.wikipedia.org/wiki/ISO_8601 format (which is what aeson expects by default for UTCTime).

csePayload :: !(Maybe Value)

cseHeaders :: ![HeaderConf]

cseRetryConf :: !STRetryConf

cseComment :: !(Maybe Text)

cseRequestTransform :: !(Maybe RequestTransform)

cseResponseTransform :: !(Maybe MetadataResponseTransform)

data ScheduledEventType

The ScheduledEventType data type is needed to differentiate between a CronScheduledEvent and OneOffScheduledEvent scheduled event because they both have different configurations and they live in different tables.

Constructors

Cron
A Cron scheduled event has a template defined which will contain the webhook, header configuration, retry configuration and a payload. Every cron event created uses the above mentioned configurations defined in the template. The configuration defined with the cron trigger is cached and hence it's not fetched along the cron scheduled events.

OneOff
A One-off scheduled event doesn't have any template defined so all the configuration is fetched along the scheduled events.

data ScheduledEventInvocation

Constructors

ScheduledEventInvocation

Fields

_seiId :: !InvocationId

_seiEventId :: !EventId

_seiStatus :: !(Maybe Int)

_seiRequest :: !(Maybe Value)

_seiResponse :: !(Maybe Value)

_seiCreatedAt :: !UTCTime

data ScheduledEvent

Constructors

SEOneOff

SECron !TriggerName

parseScheduledEvent :: Object -> Parser ScheduledEvent

scheduledEventToPairs :: ScheduledEvent -> [Pair]

data CronEventSeed

Constructors

CronEventSeed

Fields

cesName :: !TriggerName

cesScheduledTime :: !UTCTime

type OneOffEvent = CreateScheduledEvent

data ScheduledEventStatus

Constructors

SESScheduled

SESLocked

SESDelivered

SESError

SESDead

scheduledEventStatusToText :: ScheduledEventStatus -> Text

textToScheduledEventStatus :: Text -> Maybe ScheduledEventStatus

data OneOffScheduledEvent

Constructors

OneOffScheduledEvent

Fields

_ooseId :: !OneOffScheduledEventId

_ooseWebhookConf :: !InputWebhook

_ooseScheduledTime :: !UTCTime

_ooseRetryConf :: !STRetryConf

_oosePayload :: !(Maybe Value)

_ooseHeaderConf :: ![HeaderConf]

_ooseStatus :: !Text

_ooseTries :: !Int

_ooseCreatedAt :: !UTCTime

_ooseNextRetryAt :: !(Maybe UTCTime)

_ooseComment :: !(Maybe Text)

_ooseRequestTransform :: !(Maybe RequestTransform)

_ooseResponseTransform :: !(Maybe MetadataResponseTransform)

data CronEvent

Constructors

CronEvent

Fields

_ceId :: !CronEventId

_ceTriggerName :: !TriggerName

_ceScheduledTime :: !UTCTime

We expect this to always be at second zero, since cron events have minute resolution. Note that a OneOffScheduledEvent has full timestamp precision.

_ceStatus :: !Text

_ceTries :: !Int

_ceCreatedAt :: !UTCTime

it is the time at which the cron event generator created the event

_ceNextRetryAt :: !(Maybe UTCTime)

data ScheduledEventPagination

Constructors

ScheduledEventPagination

Fields

_sepLimit :: !(Maybe Int)

_sepOffset :: !(Maybe Int)

parseScheduledEventPagination :: Object -> Parser ScheduledEventPagination

scheduledEventPaginationToPairs :: ScheduledEventPagination -> [Pair]

data GetScheduledEvents

Query type to fetch all one-off/cron scheduled events

Constructors

GetScheduledEvents

Fields

_gseScheduledEvent :: !ScheduledEvent

_gsePagination :: !ScheduledEventPagination

_gseStatus :: ![ScheduledEventStatus]

data WithTotalCount a

Constructors

WithTotalCount

Fields

_wtcCount :: !Int

_wtcData :: !a

data DeleteScheduledEvent

Query type to delete cron/one-off events.

Constructors

DeleteScheduledEvent

Fields

_dseType :: !ScheduledEventType

_dseEventId :: !ScheduledEventId

data GetInvocationsBy

Constructors

GIBEventId !EventId !ScheduledEventType

GIBEvent !ScheduledEvent

data GetEventInvocations

Constructors

GetEventInvocations

Fields

_geiInvocationsBy :: !GetInvocationsBy

_geiPagination :: !ScheduledEventPagination

data ClearCronEvents

Constructors

SingleCronTrigger !TriggerName
Used to delete the cron events only of the specified cron trigger

MetadataCronTriggers ![TriggerName]
Used to delete all the cron events of the cron triggers with `include_in_metadata: true` It is used in the case of the replace_metadata API