Hasura.RQL.Types.Schema.Options

Description

Definitions for schema building configuration.

Documentation

data SchemaOptions

The record of options required to build the schema. This includes options that can be enabled with ExperimentalFeature flags.

Constructors

SchemaOptions

Fields

soStringifyNumbers :: StringifyNumbers

soDangerousBooleanCollapse :: DangerouslyCollapseBooleans

soRemoteNullForwardingPolicy :: RemoteNullForwardingPolicy

soInferFunctionPermissions :: InferFunctionPermissions

soOptimizePermissionFilters :: OptimizePermissionFilters

soIncludeUpdateManyFields :: IncludeUpdateManyFields

soIncludeAggregationPredicates :: IncludeAggregationPredicates

soIncludeStreamFields :: IncludeStreamFields

soBigQueryStringNumericInput :: BigQueryStringNumericInput

soIncludeGroupByAggregateFields :: IncludeGroupByAggregateFields

soPostgresArrays :: UsePostgresArrays

data StringifyNumbers

Should we represent numbers in our responses as numbers, or strings? Some backends can return numbers that exceed the bounds of JSON's own number type. In these cases, we can use stringified versions of these numbers to avoid this problem.

Constructors

StringifyNumbers

Don'tStringifyNumbers

data DangerouslyCollapseBooleans

Should Boolean fields be collapsed to True when a null value is given? This was the behaviour of Hasura V1, and is now discouraged.

Constructors

DangerouslyCollapseBooleans

Don'tDangerouslyCollapseBooleans

data BackwardsCompatibleNullInNonNullableVariables

Should null values allowed for variables whose type is declared as non-nullable in a GraphQL query. Until version 2.34.0 (except 2.11.11), null value is validated for non-nullable variables. It is being fixed, to keep backwards compatible this option is introduced.

Constructors

AllowNullInNonNullableVariables

Don'tAllowNullInNonNullableVariables

data RemoteNullForwardingPolicy

Constructors

RemoteForwardAccurately

RemoteOnlyForwardNonNull

data InferFunctionPermissions

Should we infer function permissions? If this flag is set to InferFunctionPermissions, we may fail to build expression parsers in buildQueryAndSubscriptionFields for users with unrecognised roles.

Constructors

InferFunctionPermissions

Don'tInferFunctionPermissions

data RemoteSchemaPermissions

Should we enable remote schema permissions? If so, these permissions will influence buildGQLContext.

Constructors

EnableRemoteSchemaPermissions

DisableRemoteSchemaPermissions

data OptimizePermissionFilters

Should we attempt to deduplicate permission filters? This flag is used in relationshipField to determine whether certain row-level permission filters can be dropped in certain cases.

Constructors

OptimizePermissionFilters

Don'tOptimizePermissionFilters

data IncludeAggregationPredicates

Should we include aggregation functions in where clauses? Because this has the potential to cause naming conflicts in graphql schema types, this flag allows users to toggle the feature off if it an upgrade breaks their setup.

Constructors

IncludeAggregationPredicates

Don'tIncludeAggregationPredicates

data IncludeStreamFields

Should we include TABLE_stream fields in schemas This is a toggle so that users can opt-in, and so that we can rename any tables that this may conflict with if needed

Constructors

IncludeStreamFields

Don'tIncludeStreamFields

data IncludeUpdateManyFields

Should we include TABLE_updates fields in schemas This is a toggle so that users can opt-in, and so that we can rename any tables that this may conflict with if needed

Constructors

IncludeUpdateManyFields

Don'tIncludeUpdateManyFields

data BigQueryStringNumericInput

Should we enable string-accepting scalar parsers for BigQuery sources

Constructors

EnableBigQueryStringNumericInput

DisableBigQueryStringNumericInput

data IncludeGroupByAggregateFields

Constructors

IncludeGroupByAggregateFields

ExcludeGroupByAggregateFields

data UsePostgresArrays

if we use Postgres arrays then an array of text becomes `[String!]`, however we want users to have the option to make it output _text like it did before for compatibility.

Constructors

UsePostgresArrays

DontUsePostgresArrays