Hasura.RQL.Types.SchemaCache.Build

Description

Types and functions used in the process of building the schema cache from metadata information stored in the hdb_catalog schema in Postgres.

Inconsistencies

recordInconsistencyM :: MonadWriter (Seq (Either InconsistentMetadata md)) m => Maybe Value -> MetadataObject -> Text -> m ()

recordInconsistenciesM :: MonadWriter (Seq (Either InconsistentMetadata md)) m => [MetadataObject] -> Text -> m ()

Dependencies

data MetadataDependency

Constructors

MetadataDependency

Fields

MetadataObject

for error reporting on missing dependencies

SchemaObjId

SchemaDependency

recordDependencies :: ArrowWriter (Seq (Either im MetadataDependency)) arr => (MetadataObject, SchemaObjId, Seq SchemaDependency) `arr` ()

recordDependenciesM :: MonadWriter (Seq (Either im MetadataDependency)) m => MetadataObject -> SchemaObjId -> Seq SchemaDependency -> m ()

Helpers

withRecordInconsistencyM :: MonadWriter (Seq (Either InconsistentMetadata md)) m => MetadataObject -> ExceptT QErr m a -> m (Maybe a)

Monadic version of withRecordInconsistency

withRecordInconsistency :: (ArrowChoice arr, ArrowWriter (Seq (Either InconsistentMetadata md)) arr) => ErrorA QErr arr (e, s) a -> arr (e, (MetadataObject, s)) (Maybe a)

Record any errors resulting from a computation as inconsistencies

withRecordInconsistencies :: (ArrowChoice arr, ArrowWriter (Seq (Either InconsistentMetadata md)) arr) => ErrorA QErr arr (e, s) a -> arr (e, ([MetadataObject], s)) (Maybe a)

class CacheRM m => CacheRWM m where

Methods

buildSchemaCacheWithOptions :: BuildReason -> CacheInvalidations -> Metadata -> m ()

setMetadataResourceVersionInSchemaCache :: MetadataResourceVersion -> m ()

data BuildReason

Constructors

CatalogUpdate (Maybe (HashSet SourceName))
The build was triggered by an update this instance made to the catalog (in the currently-active transaction), so information in Postgres that needs to be kept in sync with the catalog (i.e. table event triggers in hdb_catalog schema) should be updated.

CatalogSync
The build was triggered by a notification that some other currently-running Hasura instance updated the catalog. Since that instance already updated table event triggers in hdb_catalog, this build should be read-only.

data CacheInvalidations

Constructors

CacheInvalidations

Fields

ciMetadata :: Bool

Force reloading of all database information, including information not technically stored in metadata (currently just enum values). Set by the reload_metadata API.

ciRemoteSchemas :: HashSet RemoteSchemaName

Force refetching of the given remote schemas, even if their definition has not changed. Set by the reload_remote_schema API.

ciSources :: HashSet SourceName

Force re-establishing connections of the given data sources, even if their configuration has not changed. Set by the pg_reload_source API.

ciDataConnectors :: HashSet DataConnectorName

Force re-fetching of DataConnectorInfo from the named data connectors.

newtype MetadataT m a

Constructors

MetadataT

Fields

unMetadataT :: StateT Metadata m a

runMetadataT :: Metadata -> MetadataDefaults -> MetadataT m a -> m (a, Metadata)

runMetadataT puts a stateful metadata in scope. MetadataDefaults is provided so that it can be considered from the --metadataDefaults arguments.

buildSchemaCacheWithInvalidations :: (MetadataM m, CacheRWM m) => CacheInvalidations -> MetadataModifier -> m ()

buildSchemaCache :: (MetadataM m, CacheRWM m) => MetadataModifier -> m ()

buildSchemaCacheFor :: (QErrM m, CacheRWM m, MetadataM m) => MetadataObjId -> MetadataModifier -> m ()

Rebuilds the schema cache after modifying metadata. If an object with the given object id became newly inconsistent, raises an error about it specifically. Otherwise, raises a generic metadata inconsistency error.

throwOnInconsistencies :: (QErrM m, CacheRWM m) => m ()

Requests the schema cache, and fails if there is any inconsistent metadata.

withNewInconsistentObjsCheck :: (QErrM m, CacheRM m) => m a -> m a

Executes the given action, and if any new InconsistentMetadatas are added to the schema cache as a result of its execution, raises an error.

getInconsistentQueryCollections :: SchemaIntrospection -> QueryCollections -> ((CollectionName, ListedQuery) -> MetadataObject) -> EndpointTrie GQLQueryWithText -> [NormalizedQuery] -> [InconsistentMetadata]

getInconsistentQueryCollections is a helper function that runs the static analysis over the saved queries and reports any inconsistenties with the current schema.