Hasura.RQL.Types.SchemaCacheTypes

Documentation

data TableObjId (b :: BackendType)

Constructors

TOCol (Column b)

TORel RelName

TOComputedField ComputedFieldName

TORemoteRel RelName

TOForeignKey (ConstraintName b)

TOPerm RoleName PermType

TOTrigger TriggerName

data LogicalModelObjId (b :: BackendType)

Identifiers for components of logical models within the metadata. These are used to track dependencies within the resolved schema (see SourceInfo).

Constructors

LMOPerm RoleName PermType

LMOCol (Column b)

LMOReferencedLogicalModel LogicalModelName

data NativeQueryObjId (b :: BackendType)

Identifier for component of Native Queries within the metadata. These are used to track dependencies between items in the resolved schema. For instance, we use NQOCol along with TOCol from TableObjId to ensure that the two columns that join an array relationship actually exist.

Constructors

NQOCol (Column b)

newtype StoredProcedureObjId (b :: BackendType)

Identifier for component of Stored Procedures within the metadata. These are used to track dependencies between items in the resolved schema. For instance, we use SPOCol along with TOCol from TableObjId to ensure that the two columns that join an array relationship actually exist.

Constructors

SPOCol (Column b)

data SourceObjId (b :: BackendType)

Constructors

SOITable (TableName b)

SOITableObj (TableName b) (TableObjId b)

SOIFunction (FunctionName b)

SOINativeQuery NativeQueryName

SOINativeQueryObj NativeQueryName (NativeQueryObjId b)

SOIStoredProcedure (FunctionName b)

SOIStoredProcedureObj (FunctionName b) (StoredProcedureObjId b)

SOILogicalModel LogicalModelName

SOILogicalModelObj LogicalModelLocation (LogicalModelObjId b)

data SchemaObjId

Constructors

SOSource SourceName

SOSourceObj SourceName (AnyBackend SourceObjId)

SORemoteSchema RemoteSchemaName

SORemoteSchemaPermission RemoteSchemaName RoleName

SORemoteSchemaRemoteRelationship RemoteSchemaName Name RelName
A remote relationship on a remote schema type, identified by 1. remote schema name 2. remote schema type on which the relationship is defined 3. name of the relationship

SORole RoleName

reportSchemaObj :: SchemaObjId -> Text

reportSchemaObjs :: [SchemaObjId] -> Text

data DependencyReason

Constructors

DRTable

DRColumn

DRRemoteTable

DRLeftColumn

DRRightColumn

DRUsingColumn

DRFkey

DRRemoteFkey

DRUntyped

DROnType

DRSessionVariable

DRPayload

DRParent

DRRemoteSchema

DRRemoteRelationship

DRParentRole

DRLogicalModel

DRReferencedLogicalModel

reasonToTxt :: DependencyReason -> Text

data SchemaDependency

Constructors

SchemaDependency

Fields

sdObjId :: SchemaObjId

sdReason :: DependencyReason

reportDependentObjectsExist :: MonadError QErr m => [SchemaObjId] -> m ()

purgeSourceAndSchemaDependencies :: MonadError QErr m => SchemaObjId -> WriterT MetadataModifier m ()

purgeDependentObject :: forall b m. (MonadError QErr m, Backend b) => SourceName -> SourceObjId b -> m MetadataModifier

class Backend b => GetAggregationPredicatesDeps b where

Type class to collect schema dependencies from backend-specific aggregation predicates.

Minimal complete definition

Nothing

Methods

getAggregationPredicateDeps :: AggregationPredicates b (PartialSQLExp b) -> BoolExpM b [SchemaDependency]

default getAggregationPredicateDeps :: AggregationPredicates b ~ Const Void => AggregationPredicates b (PartialSQLExp b) -> BoolExpM b [SchemaDependency]

newtype BoolExpM b a

The monad for doing schema dependency discovery for boolean expressions. maintains the table context of the expressions being translated.

Constructors

BoolExpM

Fields

unBoolExpM :: Reader (BoolExpCtx b) a

data BoolExpCtx b

The table type context of schema dependency discovery. Boolean expressions may refer to a so-called 'root table' (identified by a $-sign in the expression input syntax) or the current table.

Constructors

BoolExpCtx

Fields

source :: SourceName

currTable :: TableName b

Reference to the current table type.

rootTable :: TableName b

Reference to the root table type.

runBoolExpM :: BoolExpCtx b -> BoolExpM b a -> a