Hasura.RQL.Types.Session

Documentation

newtype SessionVariable

Constructors

SessionVariable

Fields

unSessionVariable :: CI Text

type SessionVariableValue = Text

sessionVariablePrefix :: Text

isSessionVariable :: Text -> Bool

parseSessionVariable :: Text -> Parser SessionVariable

sessionVariableToText :: SessionVariable -> Text

mkSessionVariable :: Text -> SessionVariable

newtype SessionVariables

Constructors

SessionVariables

Fields

unSessionVariables :: HashMap SessionVariable SessionVariableValue

mkSessionVariablesText :: HashMap Text Text -> SessionVariables

data UserAdminSecret

Represent the admin secret state; whether the secret is sent in the request or if actually authorization is not configured.

Constructors

UAdminSecretSent

UAdminSecretNotSent

UAuthNotSet

data BackendOnlyFieldAccess

Represents the 'X-Hasura-Use-Backend-Only-Permissions' session variable and request made with 'X-Hasura-Admin-Secret' if any auth configured. For more details see Note [Backend only permissions]

Constructors

BOFAAllowed

BOFADisallowed

data UserInfo

Constructors

UserInfo

Fields

_uiRole :: !RoleName

_uiSession :: !SessionVariables

_uiBackendOnlyFieldAccess :: !BackendOnlyFieldAccess

class Monad m => UserInfoM m where

Methods

askUserInfo :: m UserInfo

data ExtraUserInfo

extra information used to identify a Hasura User

Constructors

ExtraUserInfo

Fields

_euiUserId :: Maybe Text

data UserRoleBuild

Represents how to build a role from the session variables

Constructors

URBFromSessionVariables
Look for `x-hasura-role` session variable value and absence will raise an exception

URBFromSessionVariablesFallback !RoleName
Look for `x-hasura-role` session variable value, if absent fall back to given role

URBPreDetermined !RoleName
Use only the pre-determined role