Hasura.RQL.Types.Source

Documentation

data SourceInfo b

Constructors

SourceInfo

Fields

_siName :: SourceName

_siSourceKind :: BackendSourceKind b

_siTables :: TableCache b

_siFunctions :: FunctionCache b

_siNativeQueries :: NativeQueryCache b

_siStoredProcedures :: StoredProcedureCache b

_siLogicalModels :: LogicalModelCache b

_siConfiguration :: ~(SourceConfig b)

_siQueryTagsConfig :: Maybe QueryTagsConfig

_siCustomization :: ResolvedSourceCustomization

_siDbObjectsIntrospection :: DBObjectsIntrospection b

type BackendSourceInfo = AnyBackend SourceInfo

type SourceCache = HashMap SourceName BackendSourceInfo

unsafeSourceInfo :: forall b. HasTag b => BackendSourceInfo -> Maybe (SourceInfo b)

unsafeSourceName :: BackendSourceInfo -> SourceName

unsafeSourceTables :: forall b. HasTag b => BackendSourceInfo -> Maybe (TableCache b)

unsafeSourceFunctions :: forall b. HasTag b => BackendSourceInfo -> Maybe (FunctionCache b)

unsafeSourceConfiguration :: forall b. HasTag b => BackendSourceInfo -> Maybe (SourceConfig b)

unsafeSourceLogicalModels :: forall b. HasTag b => BackendSourceInfo -> Maybe (LogicalModelCache b)

data DBObjectsIntrospection b

Contains metadata (introspection) from the database, used to build the schema cache. This type only contains results of introspecting DB objects, i.e. the DB types specified by tables, functions, and scalars. Notably, it does not include the additional introspection that takes place on Postgres, namely reading the contents of tables used as Enum Values -- see fetchAndValidateEnumValues.

Constructors

DBObjectsIntrospection

Fields

_rsTables :: DBTablesMetadata b

_rsFunctions :: DBFunctionsMetadata b

_rsScalars :: ScalarMap b

_rsLogicalModels :: LogicalModels b

newtype ScalarMap b

A map from GraphQL name to equivalent scalar type for a given backend.

Constructors

ScalarMap (HashMap Name (ScalarType b))

type SourceResolver b = Environment -> SourceName -> SourceConnConfiguration b -> IO (Either QErr (SourceConfig b))

FIXME: this should be either in BackendMetadata, or into a new dedicated BackendResolve, instead of listing backends explicitly. It could also be moved to the app level.

class Monad m => MonadResolveSource m where

Methods

getPGSourceResolver :: m (SourceResolver ('Postgres 'Vanilla))

getMSSQLSourceResolver :: m (SourceResolver 'MSSQL)

data MaintenanceModeVersion

Constructors

PreviousMMVersion
should correspond to the source catalog version from which the user is migrating from

CurrentMMVersion
should correspond to the latest source catalog version

data SourceHealthCheckInfo b

Constructors

SourceHealthCheckInfo

Fields

_shciName :: SourceName

_shciConnection :: SourceConnConfiguration b

_shciHealthCheck :: HealthCheckConfig b

type BackendSourceHealthCheckInfo = AnyBackend SourceHealthCheckInfo

type SourceHealthCheckCache = HashMap SourceName BackendSourceHealthCheckInfo

data SourcePingInfo b

Constructors

SourcePingInfo

Fields

_spiName :: SourceName

_spiConnection :: SourceConnConfiguration b

type BackendSourcePingInfo = AnyBackend SourcePingInfo

type SourcePingCache = HashMap SourceName BackendSourcePingInfo

siTables :: forall b. Lens' (SourceInfo b) (TableCache b)

siStoredProcedures :: forall b. Lens' (SourceInfo b) (StoredProcedureCache b)

siSourceKind :: forall b. Lens' (SourceInfo b) (BackendSourceKind b)

siQueryTagsConfig :: forall b. Lens' (SourceInfo b) (Maybe QueryTagsConfig)

siNativeQueries :: forall b. Lens' (SourceInfo b) (NativeQueryCache b)

siName :: forall b. Lens' (SourceInfo b) SourceName

siLogicalModels :: forall b. Lens' (SourceInfo b) (LogicalModelCache b)

siFunctions :: forall b. Lens' (SourceInfo b) (FunctionCache b)

siDbObjectsIntrospection :: forall b. Lens' (SourceInfo b) (DBObjectsIntrospection b)

siCustomization :: forall b. Lens' (SourceInfo b) ResolvedSourceCustomization

siConfiguration :: forall b. Lens' (SourceInfo b) (SourceConfig b)