Hasura.RQL.Types.SourceCustomization

Documentation

data RootFieldsCustomization

Constructors

RootFieldsCustomization

Fields

_rootfcNamespace :: Maybe Name

_rootfcPrefix :: Maybe Name

_rootfcSuffix :: Maybe Name

emptyRootFieldsCustomization :: RootFieldsCustomization

data SourceTypeCustomization

Constructors

SourceTypeCustomization

Fields

_stcPrefix :: Maybe Name

_stcSuffix :: Maybe Name

emptySourceTypeCustomization :: SourceTypeCustomization

mkCustomizedTypename :: Maybe SourceTypeCustomization -> NamingCase -> MkTypename

mkCustomizedFieldName :: Maybe RootFieldsCustomization -> NamingCase -> MkRootFieldName

applyTypeCust :: Maybe SourceTypeCustomization -> NamingCase -> Name -> Name

apply prefix and suffix to type name according to the source customization

applyFieldCust :: Maybe RootFieldsCustomization -> NamingCase -> Name -> Name

apply prefix and suffix to field name according to the source customization

applyTypeNameCaseCust :: NamingCase -> Name -> Name

apply naming convention to type name

applyFieldNameCaseCust :: NamingCase -> Name -> Name

apply naming convention to field name

getTextFieldName :: NamingCase -> GQLNameIdentifier -> Text

returns field name according to the naming conventions as Text

applyTypeNameCaseIdentifier :: NamingCase -> GQLNameIdentifier -> Name

applies naming convention and returns field name

Note: This can't possibly fail as GQLNameIdentifier contains already validated identifiers

getTextTypeName :: NamingCase -> GQLNameIdentifier -> Text

returns type name according to the naming conventions as Text

applyFieldNameCaseIdentifier :: NamingCase -> GQLNameIdentifier -> Name

applies naming convention and returns type name

Note: This can't possibly fail as GQLNameIdentifier contains already validated identifiers

applyEnumValueCase :: NamingCase -> Name -> Name

applyPrefixSuffix :: Maybe Name -> Maybe Name -> NamingCase -> Bool -> Name -> Name

appendprepend the suffixprefix in the graphql name

concatPrefixSuffix :: NamingCase -> Bool -> NonEmpty (Name, NameOrigin) -> Name

data SourceCustomization

Source customization information as it appears in the metadata.

Constructors

SourceCustomization

Fields

_scRootFields :: Maybe RootFieldsCustomization

_scTypeNames :: Maybe SourceTypeCustomization

_scNamingConvention :: Maybe NamingCase

emptySourceCustomization :: SourceCustomization

getSourceTypeCustomization :: SourceCustomization -> SourceTypeCustomization

getNamingConvention :: SourceCustomization -> Maybe NamingCase -> NamingCase

data ResolvedSourceCustomization

Source customization as it appears in the SchemaCache.

Constructors

ResolvedSourceCustomization

Fields

_rscRootFields :: MkRootFieldName

_rscTypeNames :: MkTypename

_rscNamingConvention :: NamingCase

_rscRootNamespace :: Maybe Name

mkResolvedSourceCustomization :: SourceCustomization -> NamingCase -> ResolvedSourceCustomization

newtype MkRootFieldName

Function to apply root field name customizations.

Constructors

MkRootFieldName

Fields

runMkRootFieldName :: Name -> Name

withSourceCustomization :: forall m r a. (MonadReader r m, Has MkTypename r, Has NamingCase r) => ResolvedSourceCustomization -> m a -> m a

Inject NamingCase, typename and root field name customizations from SourceCustomization into the environment.

getNamingCase :: forall m. MonadError QErr m => SourceCustomization -> SupportedNamingCase -> Maybe NamingCase -> m NamingCase

withNamingCaseCustomization :: forall m r a. (MonadReader r m, Has NamingCase r) => NamingCase -> m a -> m a

newtype CustomizeRemoteFieldName

Constructors

CustomizeRemoteFieldName

Fields

runCustomizeRemoteFieldName :: Name -> Name -> Name

withRemoteFieldNameCustomization :: forall m r a. (MonadReader r m, Has CustomizeRemoteFieldName r) => CustomizeRemoteFieldName -> m a -> m a

mkSelectField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectAggregateField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkInsertField :: GQLNameIdentifier -> GQLNameIdentifier

mkInsertOneField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateManyField :: GQLNameIdentifier -> GQLNameIdentifier

mkDeleteField :: GQLNameIdentifier -> GQLNameIdentifier

mkDeleteByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkRelayConnectionField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectStreamField :: GQLNameIdentifier -> GQLNameIdentifier

mkMultiRowUpdateTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkOnConflictTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableConstraintTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateFieldTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkFunctionArgsTypeName :: Name -> GQLNameIdentifier -> GQLNameIdentifier

mkTableBoolExpTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableInsertInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableObjRelInsertInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableArrRelInsertInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableMutationResponseTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableOrderByTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateOrderByTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggOperatorTypeName :: GQLNameIdentifier -> Name -> GQLNameIdentifier

mkTableSelectColumnTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableUpdateColumnTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableOperatorInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier -> GQLNameIdentifier

mkTablePkColumnsInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkEnumTableTypeName :: GQLNameIdentifier -> Maybe Name -> GQLNameIdentifier