Hasura.RQL.Types.Table

Documentation

data CustomRootField

Constructors

CustomRootField

Fields

_crfName :: Maybe Name

_crfComment :: Comment

defaultCustomRootField :: CustomRootField

data TableCustomRootFields

Constructors

TableCustomRootFields

Fields

_tcrfSelect :: CustomRootField

_tcrfSelectByPk :: CustomRootField

_tcrfSelectAggregate :: CustomRootField

_tcrfSelectStream :: CustomRootField

_tcrfInsert :: CustomRootField

_tcrfInsertOne :: CustomRootField

_tcrfUpdate :: CustomRootField

_tcrfUpdateByPk :: CustomRootField

_tcrfUpdateMany :: CustomRootField

_tcrfDelete :: CustomRootField

_tcrfDeleteByPk :: CustomRootField

emptyCustomRootFields :: TableCustomRootFields

getAllCustomRootFields :: TableCustomRootFields -> [CustomRootField]

data FieldInfo (b :: BackendType)

Constructors

FIColumn (ColumnInfo b)

FIRelationship (RelInfo b)

FIComputedField (ComputedFieldInfo b)

FIRemoteRelationship (RemoteFieldInfo (DBJoinField b))

_FIRemoteRelationship :: forall b. Prism' (FieldInfo b) (RemoteFieldInfo (DBJoinField b))

_FIComputedField :: forall b. Prism' (FieldInfo b) (ComputedFieldInfo b)

_FIRelationship :: forall b. Prism' (FieldInfo b) (RelInfo b)

_FIColumn :: forall b. Prism' (FieldInfo b) (ColumnInfo b)

type FieldInfoMap = HashMap FieldName

fieldInfoName :: forall b. Backend b => FieldInfo b -> FieldName

fieldInfoGraphQLName :: FieldInfo b -> Maybe Name

getRemoteFieldInfoName :: RemoteFieldInfo lhsJoinField -> RelName

fieldInfoGraphQLNames :: FieldInfo b -> [Name]

Returns all the field names created for the given field. Columns, object relationships, and computed fields only ever produce a single field, but array relationships also contain an _aggregate field.

getCols :: FieldInfoMap (FieldInfo backend) -> [ColumnInfo backend]

sortCols :: [ColumnInfo backend] -> [ColumnInfo backend]

Sort columns based on their ordinal position

getRels :: FieldInfoMap (FieldInfo backend) -> [RelInfo backend]

getComputedFieldInfos :: FieldInfoMap (FieldInfo backend) -> [ComputedFieldInfo backend]

data InsPermInfo (b :: BackendType)

Constructors

InsPermInfo

Fields

ipiCols :: HashSet (Column b)

ipiCheck :: AnnBoolExpPartialSQL b

ipiSet :: PreSetColsPartial b

ipiBackendOnly :: Bool

ipiRequiredHeaders :: HashSet Text

data CombinedSelPermInfo (b :: BackendType)

This type is only used as an intermediate type to combine more than one select permissions for inherited roles.

Constructors

CombinedSelPermInfo

Fields

cspiCols :: [HashMap (Column b) (Maybe (AnnColumnCaseBoolExpPartialSQL b))]

cspiComputedFields :: [HashMap ComputedFieldName (Maybe (AnnColumnCaseBoolExpPartialSQL b))]

cspiFilter :: [AnnBoolExpPartialSQL b]

cspiLimit :: Maybe (Max Int)

cspiAllowAgg :: Any

cspiRequiredHeaders :: HashSet Text

cspiAllowedQueryRootFieldTypes :: AllowedRootFields QueryRootFieldType

cspiAllowedSubscriptionRootFieldTypes :: AllowedRootFields SubscriptionRootFieldType

combinedSelPermInfoToSelPermInfo :: Backend b => Int -> CombinedSelPermInfo b -> SelPermInfo b

data SelPermInfo (b :: BackendType)

Constructors

SelPermInfo

Fields

spiCols :: HashMap (Column b) (Maybe (AnnColumnCaseBoolExpPartialSQL b))

HashMap of accessible columns to the role, the Column may be mapped to an AnnColumnCaseBoolExpPartialSQL, which happens only in the case of an inherited role, for a non-inherited role, it will be Nothing. The above bool exp will determine if the column should be nullified in a row, when there aren't requisite permissions.

spiComputedFields :: HashMap ComputedFieldName (Maybe (AnnColumnCaseBoolExpPartialSQL b))

HashMap of accessible computed fields to the role, mapped to AnnColumnCaseBoolExpPartialSQL, simililar to spiCols. These computed fields do not return rows of existing table.

spiFilter :: AnnBoolExpPartialSQL b

spiLimit :: Maybe Int

spiAllowAgg :: Bool

spiRequiredHeaders :: HashSet Text

spiAllowedQueryRootFields :: AllowedRootFields QueryRootFieldType

allowed root field types to be exposed in the query_root

spiAllowedSubscriptionRootFields :: AllowedRootFields SubscriptionRootFieldType

allowed root field types to be exposed in the subscription_root

data UpdPermInfo (b :: BackendType)

Constructors

UpdPermInfo

Fields

upiCols :: HashSet (Column b)

upiTable :: TableName b

upiFilter :: AnnBoolExpPartialSQL b

upiCheck :: Maybe (AnnBoolExpPartialSQL b)

upiSet :: PreSetColsPartial b

upiBackendOnly :: Bool

upiRequiredHeaders :: HashSet Text

data DelPermInfo (b :: BackendType)

Constructors

DelPermInfo

Fields

dpiTable :: TableName b

dpiFilter :: AnnBoolExpPartialSQL b

dpiBackendOnly :: !Bool

dpiRequiredHeaders :: HashSet Text

data RolePermInfo (b :: BackendType)

Constructors

RolePermInfo

Fields

_permIns :: Maybe (InsPermInfo b)

_permSel :: Maybe (SelPermInfo b)

_permUpd :: Maybe (UpdPermInfo b)

_permDel :: Maybe (DelPermInfo b)

permUpd :: forall b. Lens' (RolePermInfo b) (Maybe (UpdPermInfo b))

permSel :: forall b. Lens' (RolePermInfo b) (Maybe (SelPermInfo b))

permIns :: forall b. Lens' (RolePermInfo b) (Maybe (InsPermInfo b))

permDel :: forall b. Lens' (RolePermInfo b) (Maybe (DelPermInfo b))

type RolePermInfoMap b = HashMap RoleName (RolePermInfo b)

hasuraJSON ''TableConstraint)

data ViewInfo

Constructors

ViewInfo

Fields

viIsUpdatable :: Bool

viIsDeletable :: Bool

viIsInsertable :: Bool

isMutable :: (ViewInfo -> Bool) -> Maybe ViewInfo -> Bool

data ColumnConfig

Constructors

ColumnConfig

Fields

_ccfgCustomName :: Maybe Name

_ccfgComment :: Comment

data TableConfig b

Constructors

TableConfig

Fields

_tcCustomRootFields :: TableCustomRootFields

_tcColumnConfig :: HashMap (Column b) ColumnConfig

_tcCustomName :: Maybe Name

_tcComment :: Comment

tcCustomRootFields :: forall b. Lens' (TableConfig b) TableCustomRootFields

tcCustomName :: forall b. Lens' (TableConfig b) (Maybe Name)

tcComment :: forall b. Lens' (TableConfig b) Comment

tcColumnConfig :: forall b b. Lens (TableConfig b) (TableConfig b) (HashMap (Column b) ColumnConfig) (HashMap (Column b) ColumnConfig)

emptyTableConfig :: TableConfig b

data Constraint (b :: BackendType)

Constructors

Constraint

Fields

_cName :: ConstraintName b

_cOid :: OID

data PrimaryKey (b :: BackendType) a

Constructors

PrimaryKey

Fields

_pkConstraint :: Constraint b

_pkColumns :: NESeq a

pkConstraint :: forall b a b. Lens (PrimaryKey b a) (PrimaryKey b a) (Constraint b) (Constraint b)

pkColumns :: forall b a a. Lens (PrimaryKey b a) (PrimaryKey b a) (NESeq a) (NESeq a)

data UniqueConstraint (b :: BackendType)

Data type modelling uniqueness constraints. Occasionally this will include primary keys, although those are tracked separately in TableCoreInfoG.

For more information about unique constraints, visit the postgresql documentation: https://www.postgresql.org/docs/current/ddl-constraints.html#DDL-CONSTRAINTS-UNIQUE-CONSTRAINTS.

Constructors

UniqueConstraint

Fields

_ucConstraint :: Constraint b

The postgresql name and object id of a unique constraint.

_ucColumns :: HashSet (Column b)

The set of columns which should be unique for this particular constraint. Used for permissions calculation.

data ForeignKey (b :: BackendType)

Constructors

ForeignKey

Fields

_fkConstraint :: Constraint b

_fkForeignTable :: TableName b

_fkColumnMapping :: NEHashMap (Column b) (Column b)

data TableCoreInfoG (b :: BackendType) field primaryKeyColumn

The field and primaryKeyColumn type parameters vary as the schema cache is built and more information is accumulated. See also TableCoreInfo.

Constructors

TableCoreInfo

Fields

_tciName :: TableName b

_tciDescription :: Maybe PGDescription

_tciFieldInfoMap :: FieldInfoMap field

_tciPrimaryKey :: Maybe (PrimaryKey b primaryKeyColumn)

_tciUniqueConstraints :: HashSet (UniqueConstraint b)

Does not include the primary key; use tciUniqueOrPrimaryKeyConstraints if you need both.

_tciForeignKeys :: HashSet (ForeignKey b)

_tciViewInfo :: Maybe ViewInfo

_tciEnumValues :: Maybe EnumValues

_tciCustomConfig :: TableConfig b

_tciExtraTableMetadata :: ExtraTableMetadata b

tciViewInfo :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (Maybe ViewInfo)

tciUniqueConstraints :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (HashSet (UniqueConstraint b))

tciPrimaryKey :: forall b field primaryKeyColumn primaryKeyColumn. Lens (TableCoreInfoG b field primaryKeyColumn) (TableCoreInfoG b field primaryKeyColumn) (Maybe (PrimaryKey b primaryKeyColumn)) (Maybe (PrimaryKey b primaryKeyColumn))

tciName :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (TableName b)

tciForeignKeys :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (HashSet (ForeignKey b))

tciFieldInfoMap :: forall b field primaryKeyColumn field. Lens (TableCoreInfoG b field primaryKeyColumn) (TableCoreInfoG b field primaryKeyColumn) (FieldInfoMap field) (FieldInfoMap field)

tciExtraTableMetadata :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (ExtraTableMetadata b)

tciEnumValues :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (Maybe EnumValues)

tciDescription :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (Maybe PGDescription)

tciCustomConfig :: forall b field primaryKeyColumn. Lens' (TableCoreInfoG b field primaryKeyColumn) (TableConfig b)

type TableCoreInfo b = TableCoreInfoG b (FieldInfo b) (ColumnInfo b)

Fully-processed table info that includes non-column fields.

tciUniqueOrPrimaryKeyConstraints :: forall b f. (Eq (Column b), Hashable (Column b)) => TableCoreInfoG b f (ColumnInfo b) -> Maybe (NonEmpty (UniqueConstraint b))

data TableInfo (b :: BackendType)

Constructors

TableInfo

Fields

_tiCoreInfo :: TableCoreInfo b

_tiRolePermInfoMap :: RolePermInfoMap b

_tiEventTriggerInfoMap :: EventTriggerInfoMap b

_tiAdminRolePermInfo :: RolePermInfo b

tiRolePermInfoMap :: forall b. Lens' (TableInfo b) (RolePermInfoMap b)

tiEventTriggerInfoMap :: forall b. Lens' (TableInfo b) (EventTriggerInfoMap b)

tiCoreInfo :: forall b. Lens' (TableInfo b) (TableCoreInfo b)

tiAdminRolePermInfo :: forall b. Lens' (TableInfo b) (RolePermInfo b)

tiName :: Lens' (TableInfo b) (TableName b)

tableInfoName :: TableInfo b -> TableName b

getRolePermInfo :: RoleName -> TableInfo b -> RolePermInfo b

type TableCoreCache b = HashMap (TableName b) (TableCoreInfo b)

type TableCache b = HashMap (TableName b) (TableInfo b)

newtype ForeignKeyMetadata (b :: BackendType)

Metadata of a Postgres foreign key constraint which is being extracted from database via 'src-rsr/pg_table_metadata.sql'

Constructors

ForeignKeyMetadata

Fields

unForeignKeyMetadata :: ForeignKey b

data DBTableMetadata (b :: BackendType)

Metadata of any Backend table which is being extracted from source database

Constructors

DBTableMetadata

Fields

_ptmiOid :: OID

_ptmiColumns :: [RawColumnInfo b]

_ptmiPrimaryKey :: Maybe (PrimaryKey b (Column b))

_ptmiUniqueConstraints :: HashSet (UniqueConstraint b)

Does not include the primary key

_ptmiForeignKeys :: HashSet (ForeignKeyMetadata b)

_ptmiViewInfo :: Maybe ViewInfo

_ptmiDescription :: Maybe PGDescription

_ptmiExtraTableMetadata :: ExtraTableMetadata b

type DBTablesMetadata b = HashMap (TableName b) (DBTableMetadata b)

getFieldInfoM :: TableInfo b -> FieldName -> Maybe (FieldInfo b)

getColumnInfoM :: TableInfo b -> FieldName -> Maybe (ColumnInfo b)

askFieldInfo :: MonadError QErr m => FieldInfoMap fieldInfo -> FieldName -> m fieldInfo

askColumnType :: (MonadError QErr m, Backend backend) => FieldInfoMap (FieldInfo backend) -> Column backend -> Text -> m (ColumnType backend)

askColInfo :: forall m backend. (MonadError QErr m, Backend backend) => FieldInfoMap (FieldInfo backend) -> Column backend -> Text -> m (ColumnInfo backend)

askComputedFieldInfo :: MonadError QErr m => FieldInfoMap (FieldInfo backend) -> ComputedFieldName -> m (ComputedFieldInfo backend)

assertColumnExists :: forall backend m. (MonadError QErr m, Backend backend) => FieldInfoMap (FieldInfo backend) -> Text -> Column backend -> m ()

askRelType :: MonadError QErr m => FieldInfoMap (FieldInfo backend) -> RelName -> Text -> m (RelInfo backend)

askRemoteRel :: MonadError QErr m => FieldInfoMap (FieldInfo backend) -> RelName -> m (RemoteFieldInfo (DBJoinField backend))

mkAdminRolePermInfo :: Backend b => TableCoreInfo b -> RolePermInfo b