Hasura.RQL.Types.Webhook.Transform.Class

Documentation

newtype Template

Textual transformation template.

Constructors

Template

Fields

unTemplate :: Text

newtype UnescapedTemplate

Validated textual transformation template /for string interpolation only/.

This is necessary due to Kriti not distinguishing between string literals and string templates.

Constructors

UnescapedTemplate

Fields

getUnescapedTemplate :: Text

data TemplatingEngine

Available templating engines.

Constructors

Kriti

data family TransformFn a :: Type

The associated type 'TransformFn a' is the defunctionalized version of some transformation that should be applied to a given request field.

In most cases it is some variation on a piece of template text describing the transformation.

data family TransformCtx a :: Type

newtype TransformErrorBundle

We use collect all transformation failures as a '[J.Value]'.

Constructors

TransformErrorBundle

Fields

tebMessages :: [Value]