Hasura.RemoteSchema.Metadata.Base

Documentation

newtype RemoteSchemaName

Remote schema identifier.

NOTE: no validation on the character set is done here; it's likely there is a bug (FIXME) where this interacts with remote relationships and some name mangling needs to happen.

Constructors

RemoteSchemaName

Fields

unRemoteSchemaName :: NonEmptyText