Hasura.RemoteSchema.Metadata.Core

Documentation

data RemoteSchemaDef

Unvalidated remote schema config, from the user's API request

Constructors

RemoteSchemaDef

Fields

_rsdUrl :: Maybe InputWebhook

_rsdUrlFromEnv :: Maybe UrlFromEnv

_rsdHeaders :: Maybe [HeaderConf]

_rsdForwardClientHeaders :: Bool

_rsdTimeoutSeconds :: Maybe Int

_rsdCustomization :: Maybe RemoteSchemaCustomization

newtype RemoteSchemaName

Remote schema identifier.

NOTE: no validation on the character set is done here; it's likely there is a bug (FIXME) where this interacts with remote relationships and some name mangling needs to happen.

Constructors

RemoteSchemaName

Fields

unRemoteSchemaName :: NonEmptyText

type UrlFromEnv = Text

getUrlFromEnv :: MonadError QErr m => Environment -> Text -> m (EnvRecord URI)

data RemoteSchemaMetadataG r

Constructors

RemoteSchemaMetadata

Fields

_rsmName :: RemoteSchemaName

_rsmDefinition :: RemoteSchemaDef

_rsmComment :: Maybe Text

_rsmPermissions :: [RemoteSchemaPermissionMetadata]

_rsmRemoteRelationships :: SchemaRemoteRelationships r

rsmComment :: forall r f. Functor f => (Maybe Text -> f (Maybe Text)) -> RemoteSchemaMetadataG r -> f (RemoteSchemaMetadataG r)

rsmDefinition :: forall r f. Functor f => (RemoteSchemaDef -> f RemoteSchemaDef) -> RemoteSchemaMetadataG r -> f (RemoteSchemaMetadataG r)

rsmName :: forall r f. Functor f => (RemoteSchemaName -> f RemoteSchemaName) -> RemoteSchemaMetadataG r -> f (RemoteSchemaMetadataG r)

rsmPermissions :: forall r f. Functor f => ([RemoteSchemaPermissionMetadata] -> f [RemoteSchemaPermissionMetadata]) -> RemoteSchemaMetadataG r -> f (RemoteSchemaMetadataG r)

rsmRemoteRelationships :: forall r1 r2 f. Functor f => (SchemaRemoteRelationships r1 -> f (SchemaRemoteRelationships r2)) -> RemoteSchemaMetadataG r1 -> f (RemoteSchemaMetadataG r2)