Hasura.RemoteSchema.Metadata.RemoteRelationship

Documentation

data ToSchemaRelationshipDef

Metadata representation of a relationship to a remote schema.

Constructors

ToSchemaRelationshipDef

Fields

_trrdRemoteSchema :: RemoteSchemaName

Identifier for this mapping.

_trrdLhsFields :: HashSet FieldName

The lhs fields that must be forwarded to the remote schema.

_trrdRemoteField :: RemoteFields

newtype RemoteFields

Targeted field in a remote schema relationship. TODO: explain about subfields and why this is a container

Constructors

RemoteFields

Fields

unRemoteFields :: NonEmpty FieldCall

data FieldCall

Associates a field name with the arguments it will be passed in the query.

https://graphql.github.io/graphql-spec/June2018/#sec-Language.Arguments

Constructors

FieldCall

Fields

fcName :: Name

fcArguments :: RemoteArguments

newtype RemoteArguments

Arguments to a remote GraphQL fields, represented as a mapping from name to GraphQL Value. Said values can be variable names, in which case they'll be referring to values we're closed over. TODO: expand on this

Constructors

RemoteArguments

Fields

getRemoteArguments :: HashMap Name (Value Name)

data RemoteSchemaTypeRelationships r

Constructors

RemoteSchemaTypeRelationships

Fields

_rstrsName :: Name

_rstrsRelationships :: RemoteRelationships r

type SchemaRemoteRelationships r = InsOrdHashMap Name (RemoteSchemaTypeRelationships r)

rstrsRelationships :: forall r r. Lens (RemoteSchemaTypeRelationships r) (RemoteSchemaTypeRelationships r) (RemoteRelationships r) (RemoteRelationships r)

rstrsName :: forall r. Lens' (RemoteSchemaTypeRelationships r) Name

trrdRemoteSchema :: Lens' ToSchemaRelationshipDef RemoteSchemaName

trrdRemoteField :: Lens' ToSchemaRelationshipDef RemoteFields

trrdLhsFields :: Lens' ToSchemaRelationshipDef (HashSet FieldName)