Hasura.RemoteSchema.SchemaCache.Types

Documentation

type RemoteSchemaRelationshipsG remoteFieldInfo = InsOrdHashMap Name (InsOrdHashMap RelName remoteFieldInfo)

data IntrospectionResult

Constructors

IntrospectionResult

Fields

irDoc :: RemoteSchemaIntrospection

irQueryRoot :: Name

irMutationRoot :: Maybe Name

irSubscriptionRoot :: Maybe Name

data RemoteSchemaCtxG remoteFieldInfo

The resolved information of a remote schema. It is parameterized by remoteFieldInfo so as to work on an arbitrary 'remote relationship' TODO: Get rid of this G suffix using pattern synonyms or qualified usage

Constructors

RemoteSchemaCtx

Fields

_rscName :: RemoteSchemaName

_rscIntroOriginal :: IntrospectionResult

Original remote schema without customizations

_rscInfo :: RemoteSchemaInfo

_rscRawIntrospectionResult :: ByteString

The raw response from the introspection query against the remote server. We store this so we can efficiently service introspect_remote_schema.

_rscPermissions :: HashMap RoleName IntrospectionResult

_rscRemoteRelationships :: RemoteSchemaRelationshipsG remoteFieldInfo

data PartiallyResolvedRemoteRelationship remoteRelationshipDefinition

Resolved information of a remote relationship with the local information that we have. Currently this is only the typename on which the relationship is defined. TODO: also add the available join fields on the type

Constructors

PartiallyResolvedRemoteRelationship

Fields

_prrrTypeName :: Name

_prrrDefinition :: RemoteRelationshipG remoteRelationshipDefinition

type PartiallyResolvedRemoteSchemaCtxG remoteRelationshipDefinition = RemoteSchemaCtxG (PartiallyResolvedRemoteRelationship remoteRelationshipDefinition)

We can't go from RemoteSchemaMetadata to RemoteSchemaCtx in a single phase because we don't have information to resolve remote relationships. So we annotate remote relationships with as much information as we know about them which would be further resolved in a later stage.

data ValidatedRemoteSchemaDef

RemoteSchemaDef after validation and baking-in of defaults in validateRemoteSchemaDef.

Constructors

ValidatedRemoteSchemaDef

Fields

_vrsdUrl :: EnvRecord URI

_vrsdHeaders :: [HeaderConf]

_vrsdFwdClientHeaders :: Bool

_vrsdTimeoutSeconds :: Int

_vrsdCustomization :: Maybe RemoteSchemaCustomization

See _rsdCustomization.

data RemoteSchemaCustomizer

Constructors

RemoteSchemaCustomizer

Fields

_rscNamespaceFieldName :: Maybe Name

_rscCustomizeTypeName :: HashMap Name Name

type name -> type name

_rscCustomizeFieldName :: HashMap Name (HashMap Name Name)

type name -> field name -> field name

identityCustomizer :: RemoteSchemaCustomizer

remoteSchemaCustomizeTypeName :: RemoteSchemaCustomizer -> MkTypename

newtype CustomizeRemoteFieldName

Constructors

CustomizeRemoteFieldName

Fields

runCustomizeRemoteFieldName :: Name -> Name -> Name

withRemoteFieldNameCustomization :: forall m r a. (MonadReader r m, Has CustomizeRemoteFieldName r) => CustomizeRemoteFieldName -> m a -> m a

remoteSchemaCustomizeFieldName :: RemoteSchemaCustomizer -> CustomizeRemoteFieldName

hasTypeOrFieldCustomizations :: RemoteSchemaCustomizer -> Bool

data RemoteSchemaInfo

RemoteSchemaDef after the RemoteSchemaCustomizer has been generated by fetchRemoteSchema

Constructors

RemoteSchemaInfo

Fields

rsDef :: ValidatedRemoteSchemaDef

rsCustomizer :: RemoteSchemaCustomizer

validateRemoteSchemaCustomization :: MonadError QErr m => Maybe RemoteSchemaCustomization -> m ()

validateRemoteSchemaDef :: MonadError QErr m => Environment -> RemoteSchemaDef -> m ValidatedRemoteSchemaDef

data SessionArgumentPresetInfo

See resolveRemoteVariable function. This data type is used for validation of the session variable value

Constructors

SessionArgumentPresetScalar

SessionArgumentPresetEnum (HashSet EnumValue)

data RemoteSchemaVariable

Details required to resolve a "session variable preset" variable.

See Notes [Remote Schema Argument Presets] and [Remote Schema Permissions Architecture] for additional information.

Constructors

SessionPresetVariable SessionVariable Name SessionArgumentPresetInfo

QueryVariable Variable

RemoteJSONValue GType Value

data RemoteSchemaInputValueDefinition

Extends InputValueDefinition with an optional preset argument.

See Note [Remote Schema Argument Presets] for additional information.

Constructors

RemoteSchemaInputValueDefinition

Fields

_rsitdDefinition :: InputValueDefinition

_rsitdPresetArgument :: Maybe (Value RemoteSchemaVariable)

newtype RemoteSchemaIntrospection

Constructors

RemoteSchemaIntrospection (HashMap Name (TypeDefinition [Name] RemoteSchemaInputValueDefinition))

getTypeName :: TypeDefinition possibleTypes inputType -> Name

Extracts the name of a given type from its definition. TODO: move this to Language.GraphQL.Draft.Syntax.

lookupType :: RemoteSchemaIntrospection -> Name -> Maybe (TypeDefinition [Name] RemoteSchemaInputValueDefinition)

lookupObject :: RemoteSchemaIntrospection -> Name -> Maybe (ObjectTypeDefinition RemoteSchemaInputValueDefinition)

lookupInterface :: RemoteSchemaIntrospection -> Name -> Maybe (InterfaceTypeDefinition [Name] RemoteSchemaInputValueDefinition)

lookupScalar :: RemoteSchemaIntrospection -> Name -> Maybe ScalarTypeDefinition

lookupUnion :: RemoteSchemaIntrospection -> Name -> Maybe UnionTypeDefinition

lookupEnum :: RemoteSchemaIntrospection -> Name -> Maybe EnumTypeDefinition

lookupInputObject :: RemoteSchemaIntrospection -> Name -> Maybe (InputObjectTypeDefinition RemoteSchemaInputValueDefinition)

newtype LHSIdentifier

Constructors

LHSIdentifier

Fields

getLHSIdentifier :: Text

remoteSchemaToLHSIdentifier :: RemoteSchemaName -> LHSIdentifier

lhsIdentifierToGraphQLName :: LHSIdentifier -> Maybe Name

Generates a valid graphql name from an arbitrary LHS identifier. This is done by replacing all unrecognized characters by '_'. This function still returns a Maybe value, in cases we can't adjust the raw text (such as the case of empty identifiers).

data RemoteSchemaFieldInfo

Schema cache information for a table field targeting a remote schema.

Constructors

RemoteSchemaFieldInfo

Fields

_rrfiName :: RelName

Field name to which we'll map the remote in hasura; this becomes part of the hasura schema.

_rrfiParamMap :: HashMap Name RemoteSchemaInputValueDefinition

Input arguments to the remote field info; The _rfiParamMap will only include the arguments to the remote field that is being joined. The names of the arguments here are modified, it will be in the format of Field Name_remote_rel_table schema_table namerelationship name

_rrfiRemoteFields :: RemoteFields

_rrfiRemoteSchema :: RemoteSchemaInfo

_rrfiInputValueDefinitions :: [TypeDefinition [Name] RemoteSchemaInputValueDefinition]

The new input value definitions created for this remote field

_rrfiRemoteSchemaName :: RemoteSchemaName

Name of the remote schema, that's used for joining

_rrfiLHSIdentifier :: LHSIdentifier

TODO: this one should be gone when validateRemoteRelationship function is cleaned up

graphQLValueToJSON :: Value Void -> Value

rscRemoteRelationships :: forall remoteFieldInfo remoteFieldInfo. Lens (RemoteSchemaCtxG remoteFieldInfo) (RemoteSchemaCtxG remoteFieldInfo) (RemoteSchemaRelationshipsG remoteFieldInfo) (RemoteSchemaRelationshipsG remoteFieldInfo)

rscRawIntrospectionResult :: forall remoteFieldInfo. Lens' (RemoteSchemaCtxG remoteFieldInfo) ByteString

rscPermissions :: forall remoteFieldInfo. Lens' (RemoteSchemaCtxG remoteFieldInfo) (HashMap RoleName IntrospectionResult)

rscName :: forall remoteFieldInfo. Lens' (RemoteSchemaCtxG remoteFieldInfo) RemoteSchemaName

rscIntroOriginal :: forall remoteFieldInfo. Lens' (RemoteSchemaCtxG remoteFieldInfo) IntrospectionResult

rscInfo :: forall remoteFieldInfo. Lens' (RemoteSchemaCtxG remoteFieldInfo) RemoteSchemaInfo