Hasura.SQL.TH

Description

This module defines all basic Template Haskell functions we use in the rest of this folder, to generate code that deals with all possible known backends.

Those are all "normal" Haskell functions in the Q monad: they deal with values that represent Haskell code. Those functions are used, in other modules, within Template Haskell splices.

For more information about Template Haskell: * this tutorial is an incredible resource: https://markkarpov.com/tutorial/th.html * the definition of the AST is fairly readable: * Exp: https://hackage.haskell.org/package/template-haskell-2.16.0.0/docs/Language-Haskell-TH.html#t:Exp * Dec: https://hackage.haskell.org/package/template-haskell-2.16.0.0/docs/Language-Haskell-TH.html#t:Dec

Documentation

type BackendConstructor = NonEmpty Name

backendConstructors :: Q [BackendConstructor]

Inspects the BackendType to produce a list of its constructors in the Q monad. Each constructor is represented as a list of names, to include the arguments, if any. This assumes that the arguments themselves don't have arguments.

forEachBackend :: (BackendConstructor -> Q a) -> Q [a]

Associates a value in the Q monad to each backend Name.

getBackendValue :: BackendConstructor -> Exp

Associates to a backend the promoted type level value. getBackendValue [Postgres, Vanilla] = [| Postgres Vanilla |]

getBackendTypeValue :: BackendConstructor -> Type

Associates to a backend the promoted type level value. getBackendValue [Postgres, Vanilla] = [t| ('Postgres 'Vanilla) |]

getBackendTagName :: BackendConstructor -> Name

Associates to a backend the Name of its corresponding tag.

getBackendValueName :: BackendConstructor -> Name

Associates to a backend the Name of its corresponding AnyBackend constructor.

backendList :: (BackendConstructor -> Q Exp) -> Q Exp

Creates a list of values by associating an expression to each backend.

backendCase

Arguments

:: Q Exp
the expresion on which we do a case switch

-> (BackendConstructor -> Q Pat)
the match pattern for a given backend

-> (BackendConstructor -> Q Exp)
the match body for a given backend

-> Maybe (Q Exp)
the default case, if any

-> Q Exp

Creates a case expression with a match for each backend. It is not possible do directly expand a Q [Match], which is a body of a case, hence the need to instead generate the full Q Exp.

backendData

Arguments

:: Name
the name of the type

-> [TyVarBndr]
type variables of the type if any

-> (BackendConstructor -> Q Con)
the constructor for a given backend

-> [Name]
classes to derive using the stock strategy

-> Q [Dec]

Creates a data type in which there's one constructor per backend. While this only returns one declaration, it nonetheless returns a [Dec] as it's what the $() splice interpolation syntax expects.

mkDispatch

Arguments

:: Name
the name of the function to dispatch

-> Name
the name of the AnyBackend value

-> Q Exp

Generates a case expression that applies a function f to each possible value of an AnyBackend e:

case e of FooValue x -> f x BarValue x -> f x

This function needs to be in a separate file from AnyBackend, so that it can be used in splices of another module.