Hasura.SQL.Tag

Documentation

data BackendTag (b :: BackendType) where

A singleton-like GADT that associates a tag to each backend. It is generated with Template Haskell for each Backend. Its declaration results in the following type:

data BackendTag (b :: BackendType) where PostgresVanillaTag :: BackendTag ('Postgres 'Vanilla) PostgresCitusTag :: BackendTag ('Postgres 'Citus) MSSQLTag :: BackendTag 'MSSQL ...

Constructors

PostgresVanillaTag :: BackendTag ('Postgres 'Vanilla)

PostgresCitusTag :: BackendTag ('Postgres 'Citus)

MSSQLTag :: BackendTag 'MSSQL

BigQueryTag :: BackendTag 'BigQuery

MySQLTag :: BackendTag 'MySQL

DataWrapperTag :: BackendTag 'DataWrapper

class HasTag (b :: BackendType) where

This class describes how to get a tag for a given type. We use it in AnyBackend: `case backendTag @b of`...

Methods

backendTag :: BackendTag b

reify :: BackendTag b -> BackendType

How to convert back from a tag to a runtime value. This function is generated with Template Haskell for each Backend. The case switch looks like this:

PostgresVanillaTag -> Postgres Vanilla PostgresCitusTag -> Postgres Citus MSSQLTag -> MSSQL ...