Hasura.SQL.Types

Documentation

class ToSQL a where

Methods

toSQL :: a -> Builder

toSQLTxt :: ToSQL a => a -> Text

data CollectableType a

The type of all Postgres types (i.e. scalars and arrays). This type is parameterized so that we can have both CollectableType PGScalarType and CollectableType PGColumnType, for when we care about the distinction made by PGColumnType. If we ever change PGColumnType to handle arrays, not just scalars, then the parameterization can go away.

TODO (from master): This is incorrect modeling, as PGScalarType will capture anything (under PGUnknown). This should be fixed when support for all types is merged.

Constructors

CollectableTypeScalar !a

CollectableTypeArray !a

newtype ExtensionsSchema

The name of the schema in which the graphql-engine will install database extensions.

Constructors

ExtensionsSchema

Fields

_unExtensionsSchema :: Text