Hasura.Server.API.Backend

Description

BackendAPI

This module defines the BackendAPI class, alongside a few helpers. Its goal is to delegate to backends the responsibility of creating the parsers for the metadata API. Each backend is expected to provide a list of CommandParser, which in turn is a simple function from command name and command arguments to a corresponding parser. Command parsers can easily be created using the commandParser function.

Furthermore, for each set of related features, such as table tracking commands, or permission commands, a helper function is provided, that allows a backend to write its instance by simply listing the set of features it supports.

Documentation

class BackendAPI (b :: BackendType) where

Methods

metadataV1CommandParsers :: [CommandParser b]

commandParser

Arguments

:: FromJSON a

=> Text
expected command name

-> (a -> RQLMetadataV1)
corresponding parser

-> CommandParser b

sourceCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

tableCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

tablePermissionsCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

functionCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

functionPermissionsCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

relationshipCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

remoteRelationshipCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

eventTriggerCommands :: forall (b :: BackendType). Backend b => [CommandParser b]

computedFieldCommands :: forall (b :: BackendType). Backend b => [CommandParser b]