Hasura.Server.App

Documentation

data SchemaCacheRef

Constructors

SchemaCacheRef

Fields

_scrLock :: MVar ()

The idea behind explicit locking here is to

Allow maximum throughput for serving requests (v1graphql) (as each request reads the current schemacache)

We don't want to process more than one request at any point of time which would modify the schema cache as such queries are expensive.

Another option is to consider removing this lock in place of `_scrCache :: MVar ...` if it's okay or in fact correct to block during schema update in e.g. _wseGCtxMap. Vamshi says: It is theoretically possible to have a situation (in between building new schemacache and before writing it to the IORef) where we serve a request with a stale schemacache but I guess it is an okay trade-off to pay for a higher throughput (I remember doing a bunch of benchmarks to test this hypothesis).

_scrCache :: IORef (RebuildableSchemaCache, SchemaCacheVer)

data ServerCtx

Constructors

ServerCtx

Fields

scLogger :: !(Logger Hasura)

scCacheRef :: !SchemaCacheRef

scAuthMode :: !AuthMode

scManager :: !Manager

scSQLGenCtx :: !SQLGenCtx

scEnabledAPIs :: !(HashSet API)

scInstanceId :: !InstanceId

scLQState :: !LiveQueriesState

scEnableAllowlist :: !Bool

scEkgStore :: !(Store EmptyMetrics)

scResponseInternalErrorsConfig :: !ResponseInternalErrorsConfig

scEnvironment :: !Environment

scRemoteSchemaPermsCtx :: !RemoteSchemaPermsCtx

scFunctionPermsCtx :: !FunctionPermissionsCtx

scEnableMaintenanceMode :: !MaintenanceMode

scExperimentalFeatures :: !(HashSet ExperimentalFeature)

scEnabledLogTypes :: !(HashSet (EngineLogType Hasura))

this is only required for the short-term fix in https://github.com/hasura/graphql-engine-mono/issues/1770

scEventingMode :: !EventingMode

scEnableReadOnlyMode :: !ReadOnlyMode

data HandlerCtx

Constructors

HandlerCtx

Fields

hcServerCtx :: !ServerCtx

hcUser :: !UserInfo

hcReqHeaders :: ![Header]

hcRequestId :: !RequestId

hcSourceIpAddress :: !IpAddress

type Handler m = ReaderT HandlerCtx (MetadataStorageT m)

data APIResp

Constructors

JSONResp !(HttpResponse EncJSON)

RawResp !(HttpResponse ByteString)

data APIHandler m a where

API request handlers for different endpoints

Constructors

AHGet :: !(Handler m (HttpLogMetadata m, APIResp)) -> APIHandler m void
A simple GET request

AHPost :: !(a -> Handler m (HttpLogMetadata m, APIResp)) -> APIHandler m a
A simple POST request that expects a request body from which an a can be extracted

AHGraphQLRequest :: !(ReqsText -> Handler m (HttpLogMetadata m, APIResp)) -> APIHandler m ReqsText
A general GraphQL request (query or mutation) for which the content of the query is made available to the handler for authentication. This is a more specific version of the AHPost constructor.

boolToText :: Bool -> Text

isAdminSecretSet :: AuthMode -> Text

getSCFromRef :: MonadIO m => SchemaCacheRef -> m SchemaCache

logInconsObjs :: Logger Hasura -> [InconsistentMetadata] -> IO ()

withSCUpdate :: (MonadIO m, MonadBaseControl IO m) => SchemaCacheRef -> Logger Hasura -> Maybe (TVar Bool) -> m (a, RebuildableSchemaCache) -> m a

mkGetHandler :: Handler m (HttpLogMetadata m, APIResp) -> APIHandler m ()

mkPostHandler :: (a -> Handler m (HttpLogMetadata m, APIResp)) -> APIHandler m a

mkGQLRequestHandler :: (ReqsText -> Handler m (HttpLogMetadata m, APIResp)) -> APIHandler m ReqsText

mkAPIRespHandler :: Functor m => (a -> Handler m (HttpResponse EncJSON)) -> a -> Handler m APIResp

mkGQLAPIRespHandler :: Functor m => (a -> Handler m (b, HttpResponse EncJSON)) -> a -> Handler m (b, APIResp)

isMetadataEnabled :: ServerCtx -> Bool

isGraphQLEnabled :: ServerCtx -> Bool

isPGDumpEnabled :: ServerCtx -> Bool

isConfigEnabled :: ServerCtx -> Bool

isDeveloperAPIEnabled :: ServerCtx -> Bool

parseBody :: (FromJSON a, MonadError QErr m) => ByteString -> m (Value, a)

onlyAdmin :: (MonadError QErr m, MonadReader HandlerCtx m) => m ()

setHeader :: MonadIO m => Header -> ActionT m ()

class Monad m => MonadMetadataApiAuthorization m where

Typeclass representing the metadata API authorization effect

Methods

authorizeV1QueryApi :: RQLQuery -> HandlerCtx -> m (Either QErr ())

authorizeV1MetadataApi :: RQLMetadata -> HandlerCtx -> m (Either QErr ())

authorizeV2QueryApi :: RQLQuery -> HandlerCtx -> m (Either QErr ())

class Monad m => MonadConfigApiHandler m where

The config API (v1alpha1config) handler

Methods

runConfigApiHandler

Arguments

:: ServerCtx

-> Maybe Text
console assets directory

-> SpockCtxT () m ()

mapActionT :: (Monad m, Monad n) => (m (StT (ActionCtxT ()) a) -> n (StT (ActionCtxT ()) a)) -> ActionT m a -> ActionT n a

mkSpockAction

Arguments

:: (MonadIO m, MonadBaseControl IO m, FromJSON a, UserAuthentication (TraceT m), HttpLog m, HasReporter m, HasResourceLimits m)

=> ServerCtx

-> (Bool -> QErr -> Value)
QErr JSON encoder function

-> (QErr -> QErr)
QErr modifier

-> APIHandler (TraceT m) a

-> ActionT m ()

v1QueryHandler :: (MonadIO m, MonadBaseControl IO m, MonadMetadataApiAuthorization m, MonadTrace m, MonadReader HandlerCtx m, MonadMetadataStorage m, MonadResolveSource m, MonadQueryTags m) => RQLQuery -> m (HttpResponse EncJSON)

v1MetadataHandler :: (MonadIO m, MonadBaseControl IO m, MonadReader HandlerCtx m, MonadTrace m, MonadMetadataStorage m, MonadResolveSource m, MonadMetadataApiAuthorization m) => RQLMetadata -> m (HttpResponse EncJSON)

v2QueryHandler :: (MonadIO m, MonadBaseControl IO m, MonadMetadataApiAuthorization m, MonadTrace m, MonadReader HandlerCtx m, MonadMetadataStorage m, MonadResolveSource m, MonadQueryTags m) => RQLQuery -> m (HttpResponse EncJSON)

v1Alpha1GQHandler :: (MonadIO m, MonadBaseControl IO m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadTrace m, MonadExecuteQuery m, MonadError QErr m, MonadReader HandlerCtx m, HttpLog m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m, HasResourceLimits m) => GraphQLQueryType -> GQLBatchedReqs (GQLReq GQLQueryText) -> m (HttpLogMetadata m, HttpResponse EncJSON)

mkExecutionContext :: (MonadIO m, MonadReader HandlerCtx m) => m ExecutionCtx

v1GQHandler :: (MonadIO m, MonadBaseControl IO m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadTrace m, MonadExecuteQuery m, HttpLog m, MonadError QErr m, MonadReader HandlerCtx m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m, HasResourceLimits m) => GQLBatchedReqs (GQLReq GQLQueryText) -> m (HttpLogMetadata m, HttpResponse EncJSON)

v1GQRelayHandler :: (MonadIO m, MonadBaseControl IO m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadTrace m, HttpLog m, MonadExecuteQuery m, MonadError QErr m, MonadReader HandlerCtx m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m, HasResourceLimits m) => GQLBatchedReqs (GQLReq GQLQueryText) -> m (HttpLogMetadata m, HttpResponse EncJSON)

gqlExplainHandler :: forall m. (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadReader HandlerCtx m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m) => GQLExplain -> m (HttpResponse EncJSON)

v1Alpha1PGDumpHandler :: (MonadIO m, MonadError QErr m, MonadReader HandlerCtx m) => PGDumpReqBody -> m APIResp

consoleAssetsHandler :: (MonadIO m, HttpLog m) => Logger Hasura -> HashSet (EngineLogType Hasura) -> Text -> FilePath -> ActionT m ()

class Monad m => ConsoleRenderer m where

Methods

renderConsole :: Text -> AuthMode -> Bool -> Maybe Text -> m (Either String Text)

renderHtmlTemplate :: Template -> Value -> Either String Text

configApiGetHandler :: forall m. (MonadIO m, MonadBaseControl IO m, UserAuthentication (TraceT m), HttpLog m, HasReporter m, HasResourceLimits m) => ServerCtx -> Maybe Text -> SpockCtxT () m ()

Default implementation of the MonadConfigApiHandler

data HasuraApp

Constructors

HasuraApp

Fields

_hapApplication :: !Application

_hapSchemaRef :: !SchemaCacheRef

_hapAsyncActionSubscriptionState :: !AsyncActionSubscriptionState

_hapShutdownWsServer :: !(IO ())

mkWaiApp

Arguments

:: forall m. (MonadIO m, MonadStateless IO m, Forall (Pure m), ConsoleRenderer m, HttpLog m, UserAuthentication (TraceT m), MonadMetadataApiAuthorization m, MonadGQLExecutionCheck m, MonadConfigApiHandler m, MonadQueryLog m, MonadWSLog m, HasReporter m, MonadExecuteQuery m, HasResourceLimits m, MonadMetadataStorage (MetadataStorageT m), MonadResolveSource m, MonadQueryTags m)

=> (ServerCtx -> SpockT m ())

-> Environment
Set of environment variables for reference in UIs

-> Logger Hasura
a Hasura specific logger

-> SQLGenCtx

-> Bool
is AllowList enabled - TODO: change this boolean to sumtype

-> Manager
HTTP manager so that we can re-use sessions

-> AuthMode
AuthMode in which the application should operate in

-> CorsConfig

-> Bool
is console enabled - TODO: better type

-> Maybe Text
filepath to the console static assets directory - TODO: better type

-> Bool
is telemetry enabled

-> InstanceId
each application, when run, gets an InstanceId. this is used at various places including schema syncing and telemetry

-> HashSet API
set of the enabled APIs

-> LiveQueriesOptions

-> ResponseInternalErrorsConfig

-> Maybe LiveQueryPostPollHook

-> SchemaCacheRef

-> Store EmptyMetrics

-> ServerMetrics

-> RemoteSchemaPermsCtx

-> FunctionPermissionsCtx

-> ConnectionOptions

-> KeepAliveDelay

-> MaintenanceMode

-> EventingMode

-> ReadOnlyMode

-> HashSet ExperimentalFeature
Set of the enabled experimental features

-> HashSet (EngineLogType Hasura)

-> WSConnectionInitTimeout

-> m HasuraApp

initialiseCache :: MonadIO m => RebuildableSchemaCache -> m SchemaCacheRef

httpApp :: forall m. (MonadIO m, MonadBaseControl IO m, ConsoleRenderer m, HttpLog m, UserAuthentication (TraceT m), MonadMetadataApiAuthorization m, MonadGQLExecutionCheck m, MonadConfigApiHandler m, MonadQueryLog m, HasReporter m, MonadExecuteQuery m, MonadMetadataStorage (MetadataStorageT m), HasResourceLimits m, MonadResolveSource m, MonadQueryTags m) => (ServerCtx -> SpockT m ()) -> CorsConfig -> ServerCtx -> Bool -> Maybe Text -> Bool -> SpockT m ()

raiseGenericApiError :: (MonadIO m, HttpLog m) => Logger Hasura -> HashSet (EngineLogType Hasura) -> [Header] -> QErr -> ActionT m ()