Hasura.Server.AppStateRef

Documentation

data AppStateRef impl

A mutable reference to a AppState, plus

a write lock,

update version tracking, and

a gauge metric that tracks the metadata version of the SchemaCache.

Constructors

AppStateRef

Fields

_scrLock :: MVar ()

The idea behind explicit locking here is to

Allow maximum throughput for serving requests (v1graphql) (as each request reads the current schemacache)

We don't want to process more than one request at any point of time which would modify the schema cache as such queries are expensive.

Another option is to consider removing this lock in place of `_scrCache :: MVar ...` if it's okay or in fact correct to block during schema update in e.g. _wseGCtxMap. Vamshi says: It is theoretically possible to have a situation (in between building new schemacache and before writing it to the IORef) where we serve a request with a stale schemacache but I guess it is an okay trade-off to pay for a higher throughput (I remember doing a bunch of benchmarks to test this hypothesis).

_scrCache :: AppState impl

_scrMetadataVersionGauge :: Gauge

The gauge metric that tracks the current metadata version.

Invariant: This gauge must be updated via updateMetadataVersionGauge whenever the _scrCache IORef is updated.

data AppState impl

A mutable reference to '(RebuildableSchemaCache, SchemaCacheVer)' and RebuildableAppContext

Constructors

AppState

Fields

asSchemaCache :: IORef (RebuildableSchemaCache, SchemaCacheVer)

asAppCtx :: IORef (RebuildableAppContext impl)

initialiseAppStateRef :: MonadIO m => ServerMetrics -> RebuildableSchemaCache -> RebuildableAppContext impl -> m (AppStateRef impl)

Build a new AppStateRef

withSchemaCacheUpdate :: (MonadIO m, MonadBaseControl IO m) => AppStateRef impl -> Logger Hasura -> Maybe (TVar Bool) -> m (a, RebuildableSchemaCache) -> m a

Set the AppStateRef to the RebuildableSchemaCache produced by the given action.

An internal lock ensures that at most one update to the AppStateRef may proceed at a time.

readSchemaCacheRef :: AppStateRef impl -> IO (RebuildableSchemaCache, SchemaCacheVer)

Read the contents of the AppStateRef to get the latest RebuildableSchemaCache and SchemaCacheVer

readAppContextRef :: AppStateRef impl -> IO (RebuildableAppContext impl)

Read the contents of the AppStateRef to get the latest RebuildableAppContext

getSchemaCache :: AppStateRef impl -> IO SchemaCache

Utility function. Read the latest SchemaCache from the AppStateRef.

getSchemaCache == fmap (lastBuiltSchemaCache . fst) . readAppStateRef

getSchemaCacheWithVersion :: AppStateRef impl -> IO (SchemaCache, SchemaCacheVer)

getSchemaCacheRef :: AppStateRef impl -> IORef (RebuildableSchemaCache, SchemaCacheVer)

getAppContext :: AppStateRef impl -> IO AppContext

Utility function. Read the latest AppContext from the AppStateRef.

logInconsistentMetadata :: Logger Hasura -> [InconsistentMetadata] -> IO ()

Utility function