Hasura.Server.Auth.WebHook

Documentation

data AuthHookType

Constructors

AHTGet

AHTPost

data AuthHook

Constructors

AuthHook

Fields

ahUrl :: Text

ahType :: AuthHookType

ahSendRequestBody :: Bool

Whether to send the request body to the auth hook

userInfoFromAuthHook :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m) => Logger Hasura -> Manager -> AuthHook -> [Header] -> Maybe ReqsText -> m (UserInfo, Maybe UTCTime, [Header])

Makes an authentication request to the given AuthHook and returns UserInfo parsed from the response, plus an expiration time if one was returned. Optionally passes a batch of raw GraphQL requests for finer-grained auth. (#2666)