Hasura.Server.Compression

Documentation

data CompressionType

Compressed encodings which hasura supports

Constructors

CTGZip

type EncodingType = Maybe CompressionType

Accept-Encoding directives (from client) which hasura supports. Nothing indicates identity (no compression)

identityEncoding :: EncodingType

compressionTypeToTxt :: CompressionType -> Text

contentEncodingHeader :: CompressionType -> Header

A map from Accept-Encoding directives to corresponding Content-Encoding headers (from server). NOTE: identity is not a valid directive for this header.

compressResponse

Arguments

:: RequestHeaders

-> ByteString

-> (ByteString, EncodingType)
The response body (possibly compressed), and the encoding chosen

Maybe compress the response body, based on the client's Accept-Encoding and our own judgement.

compressFast :: CompressionType -> ByteString -> ByteString

Compress the bytestring preferring speed over compression ratio

getAcceptedEncodings :: RequestHeaders -> Set EncodingType

Which encodings can the client accept? The empty set returned here is an error condition and the server tecnically ought to return a 406.

https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Accept-Encoding