Hasura.Server.Init.Config

Description

Types and classes related to configuration when the server is initialised

Documentation

data Option def

The collected default value, env var, and help message for an option. If there should be no default value then use 'Option ()'.

Constructors

Option

Fields

_default :: def

_envVar :: String

_helpMessage :: String

optionPP :: Option a -> (String, String)

Helper function for pretty printing Option a.

data HGEOptionsRaw impl

Raw HGE Options from the arg parser and the env.

Constructors

HGEOptionsRaw

Fields

_horDatabaseUrl :: PostgresConnInfo (Maybe PostgresConnInfoRaw)

_horMetadataDbUrl :: Maybe String

_horCommand :: HGECommand impl

horDatabaseUrl :: Lens' (HGEOptionsRaw impl) (PostgresConnInfo (Maybe PostgresConnInfoRaw))

horMetadataDbUrl :: Lens' (HGEOptionsRaw impl) (Maybe String)

horCommand :: Lens' (HGEOptionsRaw impl) (HGECommand impl)

data HGEOptions impl

The final processed HGE options.

Constructors

HGEOptions

Fields

_hoDatabaseUrl :: PostgresConnInfo (Maybe UrlConf)

_hoMetadataDbUrl :: Maybe String

_hoCommand :: HGECommand impl

hoCommand :: Lens' (HGEOptions impl) (HGECommand impl)

data PostgresConnInfo a

Postgres connection info tupled with a retry count.

In practice, the a here is one of the following: 1. 'Maybe PostgresConnInfoRaw' 2. 'Maybe UrlConf' 3. 'Maybe Text' 4. 'Maybe DatabaseUrl' where DatabaseUrl is an alias for Text

If it contains a 'Maybe PostgresConnInfoRaw' then you have not yet processed your arg parser results.

Constructors

PostgresConnInfo

Fields

_pciDatabaseConn :: a

_pciRetries :: Maybe Int

pciDatabaseConn :: Lens' (PostgresConnInfo a) a

pciRetries :: Lens' (PostgresConnInfo a) (Maybe Int)

data PostgresConnInfoRaw

Postgres Connection info in the form of a templated URI string or structured data.

Constructors

PGConnDatabaseUrl Template

PGConnDetails PostgresConnDetailsRaw

mkUrlConnInfo :: String -> PostgresConnInfoRaw

_PGConnDatabaseUrl :: Prism' PostgresConnInfoRaw Template

_PGConnDetails :: Prism' PostgresConnInfoRaw PostgresConnDetailsRaw

rawConnDetailsToUrl :: PostgresConnDetailsRaw -> Template

data PostgresConnDetailsRaw

Structured Postgres connection information as provided by the arg parser or env vars.

Constructors

PostgresConnDetailsRaw

Fields

connHost :: String

connPort :: Int

connUser :: String

connPassword :: String

connDatabase :: String

connOptions :: Maybe String

rawConnDetailsToUrlText :: PostgresConnDetailsRaw -> Text

data HGECommand a

The HGE Arg parser Command choices.

This is polymorphic so that we can pack either ServeOptionsRaw or ProServeOptionsRaw in it.

Constructors

HCServe a

HCExport

HCClean

HCVersion

HCDowngrade !DowngradeOptions

_HCServe :: Prism' (HGECommand a) a

data ServeOptionsRaw impl

The Serve Command options accumulated from the Arg and Env parsers.

NOTE: A Nothing value indicates the absence of a particular flag. Hence types such as 'Maybe (HashSet X)' or 'Maybe Bool'.

Constructors

ServeOptionsRaw

Fields

rsoPort :: Maybe Port

rsoHost :: Maybe HostPreference

rsoConnParams :: ConnParamsRaw

rsoTxIso :: Maybe TxIsolation

rsoAdminSecret :: Maybe AdminSecretHash

rsoAuthHook :: AuthHookRaw

rsoJwtSecret :: Maybe JWTConfig

rsoUnAuthRole :: Maybe RoleName

rsoCorsConfig :: Maybe CorsConfig

rsoConsoleStatus :: ConsoleStatus

rsoConsoleAssetsDir :: Maybe Text

rsoConsoleSentryDsn :: Maybe Text

rsoEnableTelemetry :: Maybe TelemetryStatus

rsoWsReadCookie :: WsReadCookieStatus

rsoStringifyNum :: StringifyNumbers

rsoDangerousBooleanCollapse :: Maybe DangerouslyCollapseBooleans

rsoRemoteNullForwardingPolicy :: Maybe RemoteNullForwardingPolicy

rsoEnabledAPIs :: Maybe (HashSet API)

rsoMxRefetchInt :: Maybe RefetchInterval

rsoMxBatchSize :: Maybe BatchSize

rsoStreamingMxRefetchInt :: Maybe RefetchInterval

rsoStreamingMxBatchSize :: Maybe BatchSize

rsoEnableAllowList :: AllowListStatus

rsoEnabledLogTypes :: Maybe (HashSet (EngineLogType impl))

rsoLogLevel :: Maybe LogLevel

rsoDevMode :: DevModeStatus

rsoAdminInternalErrors :: Maybe AdminInternalErrorsStatus

rsoEventsHttpPoolSize :: Maybe (Refined Positive Int)

rsoEventsFetchInterval :: Maybe (Refined NonNegative Milliseconds)

rsoAsyncActionsFetchInterval :: Maybe OptionalInterval

rsoEnableRemoteSchemaPermissions :: RemoteSchemaPermissions

rsoWebSocketCompression :: CompressionOptions

rsoWebSocketKeepAlive :: Maybe KeepAliveDelay

rsoInferFunctionPermissions :: Maybe InferFunctionPermissions

rsoEnableMaintenanceMode :: MaintenanceMode ()

rsoSchemaPollInterval :: Maybe OptionalInterval

rsoExperimentalFeatures :: Maybe (HashSet ExperimentalFeature)

See Note '$experimentalFeatures' at bottom of module

rsoEventsFetchBatchSize :: Maybe (Refined NonNegative Int)

rsoGracefulShutdownTimeout :: Maybe (Refined NonNegative Seconds)

rsoWebSocketConnectionInitTimeout :: Maybe WSConnectionInitTimeout

rsoEnableMetadataQueryLoggingEnv :: MetadataQueryLoggingMode

rsoDefaultNamingConvention :: Maybe NamingCase

stores global default naming convention

rsoExtensionsSchema :: Maybe ExtensionsSchema

rsoMetadataDefaults :: Maybe MetadataDefaults

rsoApolloFederationStatus :: Maybe ApolloFederationStatus

rsoCloseWebsocketsOnMetadataChangeStatus :: Maybe CloseWebsocketsOnMetadataChangeStatus

rsoMaxTotalHeaderLength :: Maybe Int

rsoTriggersErrorLogLevelStatus :: Maybe TriggersErrorLogLevelStatus

data ConsoleStatus

Whether or not to serve Console assets.

Constructors

ConsoleEnabled

ConsoleDisabled

isConsoleEnabled :: ConsoleStatus -> Bool

data AdminInternalErrorsStatus

Whether or not internal errors will be sent in response to admin.

Constructors

AdminInternalErrorsEnabled

AdminInternalErrorsDisabled

isAdminInternalErrorsEnabled :: AdminInternalErrorsStatus -> Bool

isWebSocketCompressionEnabled :: CompressionOptions -> Bool

data AllowListStatus

A representation of whether or not to enable the GraphQL Query AllowList.

See: https://hasura.io/docs/latest/security/allow-list/#enable-allow-list

Constructors

AllowListEnabled

AllowListDisabled

isAllowListEnabled :: AllowListStatus -> Bool

data DevModeStatus

A representation of whether or not to enable Hasura Dev Mode.

See: https://hasura.io/docs/latest/deployment/graphql-engine-flags/config-examples/#dev-mode

Constructors

DevModeEnabled

DevModeDisabled

isDevModeEnabled :: DevModeStatus -> Bool

data TelemetryStatus

A representation of whether or not to enable telemetry that is isomorphic to Bool.

Constructors

TelemetryEnabled

TelemetryDisabled

isTelemetryEnabled :: TelemetryStatus -> Bool

data WsReadCookieStatus

A representation of whether or not to read the websocket cookie on initial handshake that is isomorphic to Bool. See wsReadCookieOption for more details.

Constructors

WsReadCookieEnabled

WsReadCookieDisabled

isWsReadCookieEnabled :: WsReadCookieStatus -> Bool

data Port

An Int representing a Port number in the range 0 to 65536.

mkPort :: Int -> Maybe Port

unsafePort :: Int -> Port

data API

Constructors

METADATA

GRAPHQL

PGDUMP

DEVELOPER

CONFIG

METRICS

data AuthHookRaw

Constructors

AuthHookRaw

Fields

ahrUrl :: Maybe Text

ahrType :: Maybe AuthHookType

ahrSendRequestBody :: Maybe Bool

data OptionalInterval

Sleep time interval for recurring activities such as (@asyncActionsProcessor) Presently msToOptionalInterval interprets `0` as Skip.

Constructors

Skip
No polling

Interval (Refined NonNegative Milliseconds)
Interval time

msToOptionalInterval :: Refined NonNegative Milliseconds -> OptionalInterval

data ConnParamsRaw

The Raw configuration data from the Arg and Env parsers needed to construct a ConnParams

Constructors

ConnParamsRaw

Fields

rcpStripes :: Maybe (Refined NonNegative Int)

rcpConns :: Maybe (Refined NonNegative Int)

rcpIdleTime :: Maybe (Refined NonNegative Int)

rcpConnLifetime :: Maybe (Refined NonNegative NominalDiffTime)

Time from connection creation after which to destroy a connection and choose a different/new one.

rcpAllowPrepare :: Maybe Bool

rcpPoolTimeout :: Maybe (Refined NonNegative NominalDiffTime)

See HASURA_GRAPHQL_PG_POOL_TIMEOUT

newtype KeepAliveDelay

Constructors

KeepAliveDelay

Fields

unKeepAliveDelay :: Refined NonNegative Seconds

newtype WSConnectionInitTimeout

The timeout duration in Seconds for a WebSocket connection.

Constructors

WSConnectionInitTimeout

Fields

unWSConnectionInitTimeout :: Refined NonNegative Seconds

data ServeOptions impl

The final Serve Command options accummulated from the Arg Parser and the Environment, fully processed and ready to apply when running the server.

Constructors

ServeOptions

Fields

soPort :: Port

soHost :: HostPreference

soConnParams :: ConnParams

soTxIso :: TxIsolation

soAdminSecret :: HashSet AdminSecretHash

soAuthHook :: Maybe AuthHook

soJwtSecret :: [JWTConfig]

soUnAuthRole :: Maybe RoleName

soCorsConfig :: CorsConfig

soConsoleStatus :: ConsoleStatus

soConsoleAssetsDir :: Maybe Text

soConsoleSentryDsn :: Maybe Text

soEnableTelemetry :: TelemetryStatus

soStringifyNum :: StringifyNumbers

soDangerousBooleanCollapse :: DangerouslyCollapseBooleans

soRemoteNullForwardingPolicy :: RemoteNullForwardingPolicy

soEnabledAPIs :: HashSet API

soLiveQueryOpts :: LiveQueriesOptions

soStreamingQueryOpts :: StreamQueriesOptions

soEnableAllowList :: AllowListStatus

soEnabledLogTypes :: HashSet (EngineLogType impl)

soLogLevel :: LogLevel

soEventsHttpPoolSize :: Refined Positive Int

soEventsFetchInterval :: Refined NonNegative Milliseconds

soAsyncActionsFetchInterval :: OptionalInterval

soEnableRemoteSchemaPermissions :: RemoteSchemaPermissions

soConnectionOptions :: ConnectionOptions

soWebSocketKeepAlive :: KeepAliveDelay

soInferFunctionPermissions :: InferFunctionPermissions

soEnableMaintenanceMode :: MaintenanceMode ()

soSchemaPollInterval :: OptionalInterval

soExperimentalFeatures :: HashSet ExperimentalFeature

See note '$experimentalFeatures'

soEventsFetchBatchSize :: Refined NonNegative Int

soDevMode :: DevModeStatus

soAdminInternalErrors :: AdminInternalErrorsStatus

soGracefulShutdownTimeout :: Refined NonNegative Seconds

soWebSocketConnectionInitTimeout :: WSConnectionInitTimeout

soEventingMode :: EventingMode

soReadOnlyMode :: ReadOnlyMode

See note '$readOnlyMode'

soEnableMetadataQueryLogging :: MetadataQueryLoggingMode

soDefaultNamingConvention :: NamingCase

soExtensionsSchema :: ExtensionsSchema

soMetadataDefaults :: MetadataDefaults

soApolloFederationStatus :: ApolloFederationStatus

soCloseWebsocketsOnMetadataChangeStatus :: CloseWebsocketsOnMetadataChangeStatus

soMaxTotalHeaderLength :: Int

soTriggersErrorLogLevelStatus :: TriggersErrorLogLevelStatus

data ResponseInternalErrorsConfig

ResponseInternalErrorsConfig represents the encoding of the internal errors in the response to the client.

For more details, see this github comment: https://github.com/hasura/graphql-engine/issues/4031#issuecomment-609747705

Constructors

InternalErrorsAllRequests

InternalErrorsAdminOnly

InternalErrorsDisabled

shouldIncludeInternal :: RoleName -> ResponseInternalErrorsConfig -> Bool

data DowngradeOptions

The Downgrade Command options. These are only sourced from the Arg Parser and are used directly in Migrate.

Constructors

DowngradeOptions

Fields

dgoTargetVersion :: Text

dgoDryRun :: Bool

Note Experimental features:

The graphql-engine accepts a list of experimental features that can be enabled at the startup. Experimental features are a way to introduce new, but not stable features to our users in a manner in which they have the choice to enable or disable a certain feature(s).

The objective of an experimental feature should be that when the feature is disabled, the graphql-engine should work the same way as it worked before adding the said feature.

The experimental feature's flag is `--experimental-features` and the corresponding environment variable is HASURA_GRAPHQL_EXPERIMENTAL_FEATURES which expects a comma-seperated value.

When an experimental feature is stable enough i.e. it's stable through multiple non-beta releases then we make the feature not experimental i.e. it will always be enabled. Note that when we do this we still have to support parsing of the experimental feature because users of the previous version will have it enabled and when they upgrade an error should not be thrown at the startup. For example:

The inherited roles was an experimental feature when introduced and it was enabled by setting `--experimental-features` to inherited_roles and then it was decided to make the inherited roles a stable feature, so it was removed as an experimental feature but the code was modified such that `--experimental-features inherited_roles` to not throw an error.

Note ReadOnly Mode:

This mode starts the server in a (database) read-only mode. That is, only read-only queries are allowed on users' database sources, and write queries throw a runtime error. The use-case is for failsafe operations. Metadata APIs are also disabled.

Following is the precise behaviour - 1. For any GraphQL API (relayhasura; httpwebsocket) - disable execution of mutations 2. Metadata API is disabled 3. v2query API - insert, delete, update, run_sql are disabled 4. v1query API - insert, delete, update, run_sql are disabled 5. No source catalog migrations are run 6. During build schema cache phase, building event triggers are disabled (as they create corresponding database triggers)