Hasura.Server.Init.Env

Documentation

considerEnv :: (Monad m, FromEnv a) => String -> WithEnvT m (Maybe a)

Lookup a key in the application environment then parse the value with a FromEnv instance'

considerEnvs :: (Monad m, FromEnv a) => [String] -> WithEnvT m (Maybe a)

Lookup a list of keys with considerEnv and return the first value to parse successfully.

withOptions :: (Monad m, FromEnv option) => Maybe option -> [Option ()] -> WithEnvT m (Maybe option)

Lookup a list of keys with withOption and return the first value to parse successfully.

withOption :: (Monad m, FromEnv option) => Maybe option -> Option () -> WithEnvT m (Maybe option)

Given the parse result for an option and the 'Option def' record for that option, query the environment, and then merge the results from the parser and environment.

withOptionDefault :: (Monad m, FromEnv option) => Maybe option -> Option option -> WithEnvT m option

Given the parse result for an option and the 'Option def' record for that option, query the environment, and then merge the results from the parser, environment, and the default.

withOptionSwitch :: Monad m => Bool -> Option Bool -> WithEnvT m Bool

Switches in 'optparse-applicative' have different semantics then ordinary flags. They are always optional and produce a False when absent rather then a Nothing.

In HGE we give Env Vars a higher precedence then an absent Switch but the ordinary withEnv operation expects a Nothing for an absent arg parser result.

This function executes with 'withOption Nothing' when the Switch is absent, otherwise it returns True.

An alternative solution would be to make Switches return 'Maybe _', where '_' is an option specific sum type. This would allow us to use withOptionDefault directly. Additionally, all fields of ServeOptionsRaw would become Maybe or First values which would allow us to write a 'Monoid ServeOptionsRaw' instance for combing different option sources.

class FromEnv a where

A Read style parser used for consuming envvars and building ReadM parsers for 'optparse-applicative'.

Methods

fromEnv :: String -> Either String a

type WithEnv = WithEnvT Identity

newtype WithEnvT m a

Constructors

WithEnvT

Fields

unWithEnvT :: ReaderT [(String, String)] (ExceptT String m) a

runWithEnv :: [(String, String)] -> WithEnv a -> Either String a

Given an environment run a WithEnv action producing either a parse error or an a.

runWithEnvT :: [(String, String)] -> WithEnvT m a -> m (Either String a)