Hasura.Server.Init.FeatureFlag

Description

Feature Flags are temporary toggles.

Documentation

newtype FeatureFlag

Constructors

FeatureFlag

Fields

ffIdentifier :: Text

ceCheckFeatureFlag :: Environment -> CheckFeatureFlag

In OSS we _may_ look for a environment variable or fall back to the default value.

data CheckFeatureFlag

Constructors

CheckFeatureFlag

Fields

runCheckFeatureFlag :: FeatureFlag -> IO Bool

Action that samples the value of a feature flag. Different products will want to do different things. For example, the Cloud product will want to use LaunchDarkly whereas the OSS and non-cloud EE products will want to sample environment variables.

listKnownFeatureFlags :: [(FeatureFlag, Text)]

A registry of flags that are known by the system. This is only used to inform of feature flag values via the 'v1alphaconfig' endpoint. Ideally, the console should have a dedicated endpoint to sample feature flags so we don't _have_ to centralise that knowledge here.

class Monad m => HasFeatureFlagChecker m where

Methods

checkFlag :: FeatureFlag -> m Bool