Hasura.Server.Limits

Documentation

data ResourceLimits

Resource limits, represented by a function which modifies IO actions to enforce those limits by throwing errors using MonadError in the case where they are exceeded.

Constructors

ResourceLimits

Fields

runResourceLimits :: forall m a. (MonadBaseControl IO m, MonadError QErr m) => m a -> m a

class Monad m => HasResourceLimits m where

Monads which support resource (memory, CPU time, etc.) limiting

Minimal complete definition

Nothing

Methods

askHTTPHandlerLimit :: m ResourceLimits

default askHTTPHandlerLimit :: (m ~ t n, MonadTrans t, HasResourceLimits n) => m ResourceLimits

askGraphqlOperationLimit :: RequestId -> UserInfo -> ApiLimit -> m ResourceLimits

default askGraphqlOperationLimit :: (m ~ t n, MonadTrans t, HasResourceLimits n) => RequestId -> UserInfo -> ApiLimit -> m ResourceLimits