Hasura.Server.Logging

Description

This is taken from wai-logger and customised for our use

Documentation

data StartupLog

Constructors

StartupLog

Fields

slLogLevel :: !LogLevel

slKind :: !Text

slInfo :: !Value

data PGLog

Constructors

PGLog

Fields

plLogLevel :: !LogLevel

plMessage :: !Text

data MetadataLog

Constructors

MetadataLog

Fields

mlLogLevel :: !LogLevel

mlMessage :: !Text

mlInfo :: !Value

mkInconsMetadataLog :: [InconsistentMetadata] -> MetadataLog

data WebHookLog

Constructors

WebHookLog

Fields

whlLogLevel :: !LogLevel

whlStatusCode :: !(Maybe Status)

whlUrl :: !Text

whlMethod :: !StdMethod

whlError :: !(Maybe HttpException)

whlResponse :: !(Maybe Text)

whlMessage :: !(Maybe Text)

data GQLQueryOperationSuccessLog

GQLQueryOperationSuccessLog captures all the data required to construct an HTTP success log.

Constructors

GQLQueryOperationSuccessLog

Fields

gqolQuery :: !GQLReqUnparsed

gqolQueryExecutionTime :: !DiffTime

gqolResponseSize :: !Int64

gqolRequestSize :: !Int64

gqolParameterizedQueryHash :: !ParameterizedQueryHash

data GQLQueryOperationErrorLog

GQLQueryOperationErrorLog captures the request along with the error message

Constructors

GQLQueryOperationErrorLog

Fields

gqelQuery :: !GQLReqUnparsed

gqelError :: !QErr

data GQLBatchQueryOperationLog

Constructors

GQLQueryOperationSuccess !GQLQueryOperationSuccessLog

GQLQueryOperationError !GQLQueryOperationErrorLog

data RequestMode

whether a request is executed in batched mode or not

Constructors

RequestModeBatched
this request is batched

RequestModeSingle
this is a single request

RequestModeNonBatchable
this request is of a kind for which batching is not done or does not make sense

RequestModeError
the execution of this request failed

data CommonHttpLogMetadata

Constructors

CommonHttpLogMetadata

Fields

_chlmRequestMode :: !RequestMode

_chlmBatchOperationLog :: !(Maybe (GQLBatchedReqs GQLBatchQueryOperationLog))

type HttpLogMetadata m = (CommonHttpLogMetadata, ExtraHttpLogMetadata m)

The http-log metadata attached to HTTP requests running in the monad m, split into a common portion that is present regardless of m, and a monad-specific one defined in the HttpLog instance.

This allows us to not have to duplicate the code that generates the common part of the metadata across OSS and Pro, so that instances only have to implement the part of it unique to them.

buildHttpLogMetadata :: forall m. HttpLog m => ParameterizedQueryHashList -> RequestMode -> Maybe (GQLBatchedReqs GQLBatchQueryOperationLog) -> HttpLogMetadata m

emptyHttpLogMetadata :: forall m. HttpLog m => HttpLogMetadata m

synonym for clarity, writing `emptyHttpLogMetadata m` instead of `def (HttpLogMetadata m)`

data MetadataQueryLoggingMode

Constructors

MetadataQueryLoggingEnabled

MetadataQueryLoggingDisabled

data LoggingSettings

Setting used to control the information in logs

Constructors

LoggingSettings

Fields

_lsEnabledLogTypes :: HashSet (EngineLogType Hasura)

this is only required for the short-term fix in https://github.com/hasura/graphql-engine-mono/issues/1770 See Note [Disable query printing when query-log is disabled]

_lsMetadataQueryLoggingMode :: MetadataQueryLoggingMode

class Monad m => HttpLog m where

Associated Types

type ExtraHttpLogMetadata m

Extra http-log metadata that we attach when operating in m.

Methods

emptyExtraHttpLogMetadata :: ExtraHttpLogMetadata m

buildExtraHttpLogMetadata :: ParameterizedQueryHashList -> ExtraHttpLogMetadata m

logHttpError

Arguments

:: Logger Hasura
the logger

-> LoggingSettings
setting used to control the information in logs

-> Maybe UserInfo
user info may or may not be present (error can happen during user resolution)

-> RequestId
request id of the request

-> Request
the Wai.Request object

-> (ByteString, Maybe Value)
the request body and parsed request

-> QErr
the error

-> [Header]
list of request headers

-> m ()

logHttpSuccess

Arguments

:: Logger Hasura
the logger

-> LoggingSettings
setting used to control the information in logs

-> Maybe UserInfo
user info may or may not be present (error can happen during user resolution)

-> RequestId
request id of the request

-> Request
the Wai.Request object

-> (ByteString, Maybe Value)
the request body and parsed request

-> ByteString
the response bytes

-> ByteString
the compressed response bytes ^ TODO (from master): make the above two type represented

-> Maybe (DiffTime, DiffTime)
IO/network wait time and service time (respectively) for this request, if available.

-> Maybe CompressionType
possible compression type

-> [Header]
list of request headers

-> HttpLogMetadata m

-> m ()

data HttpInfoLog

Log information about the HTTP request

Constructors

HttpInfoLog

Fields

hlStatus :: !Status

hlMethod :: !Text

hlSource :: !IpAddress

hlPath :: !Text

hlHttpVersion :: !HttpVersion

hlCompression :: !(Maybe CompressionType)

hlHeaders :: ![Header]

all the request headers

data OperationLog

Information about a GraphQL/Hasura metadata operation over HTTP

Constructors

OperationLog

Fields

olRequestId :: !RequestId

olUserVars :: !(Maybe SessionVariables)

olResponseSize :: !(Maybe Int64)

olRequestReadTime :: !(Maybe Seconds)

Request IO wait time, i.e. time spent reading the full request from the socket.

olQueryExecutionTime :: !(Maybe Seconds)

Service time, not including request IO wait time.

olQuery :: !(Maybe Value)

olRawQuery :: !(Maybe Text)

olError :: !(Maybe QErr)

olRequestMode :: !RequestMode

data BatchOperationSuccessLog

BatchOperationSuccessLog contains the information required for a single successful operation in a batch request for OSS. This type is a subset of the GQLQueryOperationSuccessLog

Constructors

BatchOperationSuccessLog

Fields

_bolQuery :: !(Maybe Value)

_bolResponseSize :: !Int64

_bolQueryExecutionTime :: !Seconds

data BatchOperationErrorLog

BatchOperationSuccessLog contains the information required for a single erroneous operation in a batch request for OSS. This type is a subset of the GQLQueryOperationErrorLog

Constructors

BatchOperationErrorLog

Fields

_belQuery :: !(Maybe Value)

_belError :: !QErr

data BatchOperationLog

Constructors

BatchOperationSuccess !BatchOperationSuccessLog

BatchOperationError !BatchOperationErrorLog

data HttpLogContext

Constructors

HttpLogContext

Fields

hlcHttpInfo :: !HttpInfoLog

hlcOperation :: !OperationLog

hlcRequestId :: !RequestId

hlcBatchedOperations :: !(Maybe (NonEmpty BatchOperationLog))

isQueryIncludedInLogs :: Text -> LoggingSettings -> Bool

Check if the query field should be included in the http-log

mkHttpAccessLogContext

Arguments

:: Maybe UserInfo
Maybe because it may not have been resolved

-> LoggingSettings

-> RequestId

-> Request

-> (ByteString, Maybe Value)

-> ByteString

-> Maybe (DiffTime, DiffTime)

-> Maybe CompressionType

-> [Header]

-> RequestMode

-> Maybe (GQLBatchedReqs GQLBatchQueryOperationLog)

-> HttpLogContext

mkHttpErrorLogContext

Arguments

:: Maybe UserInfo
Maybe because it may not have been resolved

-> LoggingSettings

-> RequestId

-> Request

-> (ByteString, Maybe Value)

-> QErr

-> Maybe (DiffTime, DiffTime)

-> Maybe CompressionType

-> [Header]

-> HttpLogContext

data HttpLogLine

Constructors

HttpLogLine

Fields

_hlLogLevel :: !LogLevel

_hlLogLine :: !HttpLogContext

mkHttpLog :: HttpLogContext -> HttpLogLine

logDeprecatedEnvVars :: Logger Hasura -> Environment -> SourceCache -> IO ()

Log warning messages for deprecated environment variables