Hasura.Server.Metrics

Documentation

data ServerMetricsSpec :: Symbol -> MetricType -> Type -> Type where

A specification of the metrics tracked by the server.

The use of the "unit" type () for the "tag structure" type parameter of a metric indicates that we prohibit that metric from being annotated with tags.

Constructors

WarpThreads :: ServerMetricsSpec "warp_threads" 'GaugeType ()
Current Number of active Warp threads

WebsocketConnections :: ServerMetricsSpec "websocket_connections" 'GaugeType ()
Current number of active websocket connections

ActiveSubscriptions :: ServerMetricsSpec "active_subscriptions" 'GaugeType ()
Current number of active subscriptions

NumEventsFetchedPerBatch :: ServerMetricsSpec "events_fetched_per_batch" 'DistributionType ()
Total Number of events fetched from last 'Event Trigger Fetch'

NumEventHTTPWorkers :: ServerMetricsSpec "num_event_trigger_http_workers" 'GaugeType ()
Current number of Event trigger's HTTP workers in process

EventQueueTime :: ServerMetricsSpec "event_queue_time" 'DistributionType ()
Time (in seconds) between the 'Event Trigger Fetch' from DB and the processing of the event

SchemaCacheMetadataResourceVersion :: ServerMetricsSpec "schema_cache_metadata_resource_version" 'GaugeType ()
The current schema cache metadata resource version

ActiveLiveQueries :: ServerMetricsSpec "active_livequeries" 'GaugeType ()
Current number active live queries

ActiveStreaming :: ServerMetricsSpec "active_streaming_subscriptions" 'GaugeType ()
Current number of streaming subscriptions

data ServerMetrics

Mutable references for the server metrics. See ServerMetricsSpec for a description of each metric.

Constructors

ServerMetrics

Fields

smWarpThreads :: !Gauge

smWebsocketConnections :: !Gauge

smActiveSubscriptions :: !Gauge

smNumEventsFetchedPerBatch :: !Distribution

smNumEventHTTPWorkers :: !Gauge

smEventQueueTime :: !Distribution

smSchemaCacheMetadataResourceVersion :: !Gauge

smActiveLiveQueries :: !Gauge

smActiveStreamingSubscriptions :: !Gauge

createServerMetrics :: Store ServerMetricsSpec -> IO ServerMetrics