Hasura.Server.Migrate.LatestVersion

Description

A module that defines the current catalog version and nothing else. This is necessary to circumvent the unfortunate “GHC stage restriction,” which prevents us from using a binding in a compile-time splice unless it is defined in a different module. The actual migration code is in Hasura.Server.Migrate.

Documentation

latestCatalogVersion :: MetadataCatalogVersion

The current catalog schema version. We store this in a file because we want to append the current verson to the catalog_versions file when tagging a new release, in tag-release.sh.

latestCatalogVersionString :: Text