Hasura.Server.Migrate.Version

Documentation

data MetadataCatalogVersion

Represents the catalog version. This is stored in the database and then compared with the latest version on startup.

Constructors

MetadataCatalogVersion Int
A typical catalog version.

MetadataCatalogVersion08
Maintained for compatibility with catalog version 0.8.

newtype SourceCatalogVersion (backend :: BackendType)

This is the source catalog version, used when deciding whether to (re-)create event triggers.

Constructors

SourceCatalogVersion Int