Hasura.Server.ResourceChecker

Documentation

data ComputeResourcesResponse

The response data of cpu and memory resources

Constructors

ComputeResourcesResponse

Fields

_rcrCpu :: Maybe Int

_rcrMemory :: Maybe Int64

_rcrErrorCode :: Maybe ResourceCheckerError

data ResourceCheckerError

Constructors

CGroupUnavailable

CpuInconclusive

MemoryInconclusive

CpuMemoryInconclusive

RCInternalError String

getServerResources :: MonadIO m => m ComputeResourcesResponse

Determine allocated cpu and memory resources of the host server or Container Runtime. because HGE mainly runs in the container runtime we need to determine the max cpu and memory limit constraints that are managed by cgroups or fallback to physical cpu and memory information of the server https://hasurahq.atlassian.net/browse/INFRA-772

Those information are stored in many files of cgroup folders, the logic is simply to read them and parse number values

In cgroup v1 systems there are several ways in which the amount of allocated cpu resources could be presented. We first try reading requests (quota & period); if that fails, we fallback to reading limits (shares); if that fails, we fallback to reading the physical cpu count, which should always succeed.

getServerResources_ :: MonadIO m => FilePath -> m ComputeResourcesResponse

getPhysicalCpuResource :: MonadIO m => m Int

Compute the cpu share allocations from the number of physical CPU cores

getMaxPhysicalMemory :: MonadIO m => m (Maybe Int64)

Compute the max physical memory size of the server

getPhysicalResources :: MonadIO m => Maybe ResourceCheckerError -> m ComputeResourcesResponse

getCGroupV1Resources :: MonadIO m => FilePath -> m ComputeResourcesResponse

Determine cpu and memory resource allocations if the OCI Container Runtime supports cgroup v1

getCGroupV2Resources :: MonadIO m => FilePath -> m ComputeResourcesResponse

Determine cpu and memory resource allocations if the OCI Container Runtime supports cgroup v2