Hasura.Server.Rest

Documentation

parseVariableNames :: EndpointMetadata GQLQueryWithText -> [Text]

alignVars :: [VariableDefinition] -> [(Text, Either Text Value)] -> HashMap Name (These VariableDefinition (Either Text Value))

resolveVar :: Name -> These VariableDefinition (Either Text Value) -> Either Text (Maybe Value)

mkPassthroughRequest :: EndpointMetadata GQLQueryWithText -> VariableValues -> GQLReq GQLQueryText

data RestRequest method

Constructors

RestRequest

Fields

reqPath :: Text

Remainder of the url path after `api/rest`

reqMethod :: method

reqArgs :: [(Text, Either Text Value)]

URL Query/Request Body Arguments

runCustomEndpoint :: forall m. (MonadIO m, MonadError QErr m, MonadTrace m, MonadBaseControl IO m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadExecuteQuery m, MonadMetadataStorage (MetadataStorageT m), HttpLog m, MonadQueryTags m, HasResourceLimits m) => Environment -> ExecutionCtx -> RequestId -> UserInfo -> [Header] -> IpAddress -> RestRequest EndpointMethod -> EndpointTrie GQLQueryWithText -> m (HttpLogMetadata m, HttpResponse EncJSON)

Implements all the custom endpoints by looking up the path/methods in the endpoint trie and delegating to the graphql handler.