Hasura.Server.SchemaCacheRef

Documentation

data SchemaCacheRef

A mutable reference to a RebuildableSchemaCache, plus

a write lock,

update version tracking, and

a gauge metric that tracks the metadata version of the SchemaCache.

initialiseSchemaCacheRef :: MonadIO m => ServerMetrics -> RebuildableSchemaCache -> m SchemaCacheRef

Build a new SchemaCacheRef

withSchemaCacheUpdate :: (MonadIO m, MonadBaseControl IO m) => SchemaCacheRef -> Logger Hasura -> Maybe (TVar Bool) -> m (a, RebuildableSchemaCache) -> m a

Set the SchemaCacheRef to the RebuildableSchemaCache produced by the given action.

An internal lock ensures that at most one update to the SchemaCacheRef may proceed at a time.

readSchemaCacheRef :: SchemaCacheRef -> IO (RebuildableSchemaCache, SchemaCacheVer)

Read the contents of the SchemaCacheRef

getSchemaCache :: SchemaCacheRef -> IO SchemaCache

Utility function. Read the latest SchemaCache from the SchemaCacheRef.

getSchemaCache == fmap (lastBuiltSchemaCache . fst) . readSchemaCacheRef

logInconsistentMetadata :: Logger Hasura -> [InconsistentMetadata] -> IO ()

Utility function