Hasura.Server.SchemaUpdate

Documentation

startSchemaSyncListenerThread :: ForkableMonadIO m => Logger Hasura -> PGPool -> InstanceId -> Refined NonNegative Milliseconds -> TMVar MetadataResourceVersion -> ManagedT m Thread

An async thread which listen to Postgres notify to enable schema syncing See Note [Schema Cache Sync]

startSchemaSyncProcessorThread :: (ForkableMonadIO m, HasAppEnv m, HasCacheStaticConfig m, MonadMetadataStorage m, MonadResolveSource m, ProvidesNetwork m) => AppStateRef impl -> TVar Bool -> ManagedT m Thread

An async thread which processes the schema sync events See Note [Schema Cache Sync]