Hasura.Server.Telemetry.Counters

Description

Counters used in telemetry collection. Additional counters can be added here.and serviced in Hasura.Server.Telemetry.

Documentation

data RequestDimensions

The properties that characterize this request. The dimensions over which we collect metrics for each serviced request.

Constructors

RequestDimensions

Fields

telemQueryType :: !QueryType

telemLocality :: !Locality

telemTransport :: !Transport

data RequestTimings

Accumulated time metrics.

Constructors

RequestTimings

Fields

telemTimeIO :: !Seconds

Time spent waiting on PG/remote http calls

telemTimeTot :: !Seconds

Total service time for request (including $sel:telemTimeIO:RequestTimings)

data RequestTimingsCount

RequestTimings along with the count

Constructors

RequestTimingsCount

Fields

telemTimeIO :: !Seconds

telemTimeTot :: !Seconds

telemCount :: !Word

The number of requests that have contributed to the accumulated timings above. So e.g. telemTimeTot / count would give the mean service time.

requestCounters :: IORef (HashMap (RequestDimensions, RunningTimeBucket) RequestTimingsCount)

Internal. Counts and durations across many RequestDimensions.

NOTE: We use the global mutable variable pattern for metric collection counters for convenience at collection site (don't wear hairshirts that discourage useful reporting).

approxStartTime :: POSIXTime

Internal. Since these metrics are accumulated while graphql-engine is running and sent periodically, we need to include a tag that is unique for each start of hge. This lets us e.g. query for just the latest uploaded sample for each start of hge.

We use time rather than a UUID since having this be monotonic increasing is convenient.

data QueryType

Was this request a mutation (involved DB writes)?

Constructors

Mutation

Query

data Locality

Was this a PG local query, or did it involve remote execution?

Constructors

Empty
No data was fetched

Local
local DB data

Remote
remote schema

Heterogeneous
mixed

data Transport

Was this a query over http or websockets?

Constructors

HTTP

WebSocket

newtype RunningTimeBucket

The timings and counts here were from requests with total time longer than $sel:bucketGreaterThan:RunningTimeBucket (but less than any larger bucket cutoff times).

Constructors

RunningTimeBucket

Fields

bucketGreaterThan :: Seconds

totalTimeBuckets :: [RunningTimeBucket]

recordTimingMetric :: MonadIO m => RequestDimensions -> RequestTimings -> m ()

Save a timing metric sample in our in-memory store. These will be accumulated and uploaded periodically in Hasura.Server.Telemetry.

data ServiceTimingMetrics

The final shape of this part of our metrics data JSON. This should allow reasonably efficient querying using GIN indexes and JSONB containment operations (which treat arrays as sets).

Constructors

ServiceTimingMetrics

Fields

collectionTag :: Int

This is set to a new unique value when the counters reset (e.g. because of a restart)

serviceTimingMetrics :: [ServiceTimingMetric]

data ServiceTimingMetric

Constructors

ServiceTimingMetric

Fields

dimensions :: RequestDimensions

bucket :: RunningTimeBucket

metrics :: RequestTimingsCount

dumpServiceTimingMetrics :: MonadIO m => m ServiceTimingMetrics