Hasura.Server.Telemetry.Types

Description

Telemetry types

Define anonymized metrics regarding usage of various features of Hasura.

Metrics

data RelationshipMetric

Constructors

RelationshipMetric

Fields

_rmManual :: Int

_rmAuto :: Int

data PermissionMetric

Constructors

PermissionMetric

Fields

_pmSelect :: Int

_pmInsert :: Int

_pmUpdate :: Int

_pmDelete :: Int

_pmRoles :: Int

data ActionMetric

Constructors

ActionMetric

Fields

_amSynchronous :: Int

_amAsynchronous :: Int

_amQueryActions :: Int

_amTypeRelationships :: Int

_amCustomTypes :: Int

data NativeQueriesMetrics

Constructors

NativeQueriesMetrics

Fields

_nqmWithParameters :: Int

_nqmWithoutParameters :: Int

data StoredProceduresMetrics

Constructors

StoredProceduresMetrics

Fields

_spmWithParameters :: Int

_spmWithoutParameters :: Int

newtype LogicalModelsMetrics

Constructors

LogicalModelsMetrics

Fields

_lmmCount :: Int

data Metrics

Constructors

Metrics

Fields

_mtTables :: Int

_mtViews :: Int

_mtEnumTables :: Int

_mtRelationships :: RelationshipMetric

_mtPermissions :: PermissionMetric

_mtEventTriggers :: Int

_mtFunctions :: Int

_mtRemoteSchemas :: Maybe Int

_mtServiceTimings :: Maybe ServiceTimingMetrics

_mtActions :: Maybe ActionMetric

_mtNativeQueries :: NativeQueriesMetrics

_mtStoredProcedures :: StoredProceduresMetrics

_mtLogicalModels :: LogicalModelsMetrics

data SourceMetadata

Constructors

SourceMetadata

Fields

_smDbUid :: Maybe DbUid

_smBackendType :: BackendType

_smDbKind :: Text

_smDbVersion :: Maybe DbVersion

data HasuraTelemetry

Constructors

HasuraTelemetry

Fields

_htMetadataDbUid :: MetadataDbId

_htInstanceUid :: InstanceId

_htHasuraVersion :: Version

_htCi :: Maybe CI

_htSourceMetadata :: SourceMetadata

_htMetrics :: Metrics

_htExperimentalFeatures :: HashSet ExperimentalFeature

data TelemetryPayload

Constructors

TelemetryPayload

Fields

_tpTopic :: Topic

_tpData :: HasuraTelemetry

newtype Topic

The telemetry table to which we'll add telemetry.

Constructors

Topic

Fields

getTopic :: Text

Counters

Local metric recording

data RequestDimensions

The properties that characterize this request. The dimensions over which we collect metrics for each serviced request.

Constructors

RequestDimensions

Fields

telemQueryType :: !QueryType

telemLocality :: !Locality

telemTransport :: !Transport

data RequestTimings

Accumulated time metrics.

Constructors

RequestTimings

Fields

telemTimeIO :: !Seconds

Time spent waiting on PG/remote http calls

telemTimeTot :: !Seconds

Total service time for request (including telemTimeIO)

Dimensions

data QueryType

Was this request a mutation (involved DB writes)?

Constructors

Mutation

Query

data Locality

Was this a PG local query, or did it involve remote execution?

Constructors

Empty
No data was fetched

Local
local DB data

Remote
remote schema

Heterogeneous
mixed

data Transport

Was this a query over http or websockets?

Constructors

HTTP

WebSocket

Metric upload

data ServiceTimingMetrics

The final shape of this part of our metrics data JSON. This should allow reasonably efficient querying using GIN indexes and JSONB containment operations (which treat arrays as sets).

Constructors

ServiceTimingMetrics

Fields

collectionTag :: Int

This is set to a new unique value when the counters reset (e.g. because of a restart)

serviceTimingMetrics :: [ServiceTimingMetric]

data ServiceTimingMetric

Constructors

ServiceTimingMetric

Fields

dimensions :: RequestDimensions

bucket :: RunningTimeBucket

metrics :: RequestTimingsCount

newtype RunningTimeBucket

The timings and counts here were from requests with total time longer than bucketGreaterThan (but less than any larger bucket cutoff times).

Constructors

RunningTimeBucket

Fields

bucketGreaterThan :: Seconds

data RequestTimingsCount

RequestTimings along with the count

Constructors

RequestTimingsCount

Fields

telemTimeIO :: !Seconds

telemTimeTot :: !Seconds

telemCount :: !Word

The number of requests that have contributed to the accumulated timings above. So e.g. telemTimeTot / count would give the mean service time.