Hasura.Server.Utils

Documentation

jsonHeader :: Header

sqlHeader :: Header

gzipHeader :: Header

userRoleHeader :: IsString a => a

deprecatedAccessKeyHeader :: IsString a => a

adminSecretHeader :: IsString a => a

userIdHeader :: IsString a => a

requestIdHeader :: IsString a => a

contentLengthHeader :: IsString a => a

useBackendOnlyPermissionsHeader :: IsString a => a

getRequestHeader :: HeaderName -> [Header] -> Maybe ByteString

parseStringAsBool :: String -> Either String Bool

quoteRegex :: CompOption -> ExecOption -> String -> Code Q Regex

Quotes a regex using Template Haskell so syntax errors can be reported at compile-time.

fmapL :: (a -> a') -> Either a b -> Either a' b

generateFingerprint :: IO Text

httpExceptToJSON :: HttpException -> Value

commonClientHeadersIgnored :: IsString a => [a]

sessionVariablePrefix :: Text

isReqUserId :: Text -> Bool

mkClientHeadersForward :: [Header] -> [Header]

mkSetCookieHeaders :: Response a -> ResponseHeaders

filterHeaders :: HashSet HeaderName -> [Header] -> [Header]

data APIVersion

The version integer

Constructors

VIVersion1

VIVersion2

englishList :: Text -> NonEmpty Text -> Text

makeReasonMessage :: [a] -> (a -> Text) -> Text

executeJSONPath :: JSONPath -> Value -> IResult Value

cryptoHash :: ToJSON a => a -> ByteString

readIsoLevel :: String -> Either String TxIsolation

parseConnLifeTime :: Maybe NominalDiffTime -> Maybe NominalDiffTime

newtype EnvVarsMovedToMetadata

The environment variables that were moved to metadata. These environment variables are available if a v1 hasura project is run an v2 hasura server. These environment variables are marked as deprecated only when the v1 hasura project is migrated to v2 project.

Constructors

EnvVarsMovedToMetadata

Fields

unEnvVarsMovedToMetadata :: [String]

newtype DeprecatedEnvVars

These env vars are completely deprecated

Constructors

DeprecatedEnvVars

Fields

unDeprecatedEnvVars :: [String]

envVarsMovedToMetadata :: EnvVarsMovedToMetadata

deprecatedEnvVars :: DeprecatedEnvVars

redactSensitiveHeader :: Header -> Header