Hasura.Services.Network

Description

Network service provider.

This module defines a Service (see Note [Services]) that provides access to the network; for now, that only means providing a HTTP Manager. This is consequentlt a simple analogue to `(MonadReader r m, Has Manager r)`, but could be updated to either encompass other network utilities, or to provide a more restricted interface if deemed useful. Alternatively this could be removed altogether if all network calls were to be hidden behind more specific services.

Documentation

class Monad m => ProvidesNetwork m where

Methods

askHTTPManager :: m Manager