Hasura.Session

Documentation

newtype RoleName

Constructors

RoleName

Fields

getRoleTxt :: NonEmptyText

roleNameToTxt :: RoleName -> Text

mkRoleName :: Text -> Maybe RoleName

adminRoleName :: RoleName

isAdmin :: RoleName -> Bool

newtype SessionVariable

Constructors

SessionVariable

Fields

unSessionVariable :: CI Text

sessionVariableToGraphQLName :: SessionVariable -> Maybe Name

Converts a SessionVariable value to a GraphQL name. This will fail if the session variable contains characters that are not valid for a graphql names. It is the caller's responsibility to decide what to do in such a case.

parseSessionVariable :: Text -> Parser SessionVariable

sessionVariableToText :: SessionVariable -> Text

mkSessionVariable :: Text -> SessionVariable

type SessionVariableValue = Text

newtype SessionVariables

Constructors

SessionVariables

Fields

unSessionVariables :: HashMap SessionVariable SessionVariableValue

filterSessionVariables :: (SessionVariable -> SessionVariableValue -> Bool) -> SessionVariables -> SessionVariables

mkSessionVariablesText :: HashMap Text Text -> SessionVariables

mkSessionVariablesHeaders :: [Header] -> SessionVariables

sessionVariablesToHeaders :: SessionVariables -> [Header]

getSessionVariables :: SessionVariables -> [Text]

getSessionVariablesSet :: SessionVariables -> HashSet SessionVariable

getSessionVariableValue :: SessionVariable -> SessionVariables -> Maybe SessionVariableValue

data UserAdminSecret

Represent the admin secret state; whether the secret is sent in the request or if actually authorization is not configured.

Constructors

UAdminSecretSent

UAdminSecretNotSent

UAuthNotSet

data BackendOnlyFieldAccess

Represents the 'X-Hasura-Use-Backend-Only-Permissions' session variable and request made with 'X-Hasura-Admin-Secret' if any auth configured. For more details see Note [Backend only permissions]

Constructors

BOFAAllowed

BOFADisallowed

data UserInfo

Constructors

UserInfo

Fields

_uiRole :: !RoleName

_uiSession :: !SessionVariables

_uiBackendOnlyFieldAccess :: !BackendOnlyFieldAccess

class Monad m => UserInfoM m where

Methods

askUserInfo :: m UserInfo

askCurRole :: UserInfoM m => m RoleName

data UserRoleBuild

Represents how to build a role from the session variables

Constructors

URBFromSessionVariables
Look for `x-hasura-role` session variable value and absence will raise an exception

URBFromSessionVariablesFallback !RoleName
Look for `x-hasura-role` session variable value, if absent fall back to given role

URBPreDetermined !RoleName
Use only the pre-determined role

mkUserInfo :: forall m. MonadError QErr m => UserRoleBuild -> UserAdminSecret -> SessionVariables -> m UserInfo

Build UserInfo from SessionVariables

maybeRoleFromSessionVariables :: SessionVariables -> Maybe RoleName

adminUserInfo :: UserInfo