Hasura.ShutdownLatch

Documentation

data ShutdownLatch

A latch for the graceful shutdown of a server process.

newShutdownLatch :: IO ShutdownLatch

waitForShutdown :: ShutdownLatch -> IO ()

Block the current thread, waiting on the latch.

shutdownGracefully :: ShutdownLatch -> IO ()

Initiate a graceful shutdown of the server associated with the provided latch.

shuttingDown :: ShutdownLatch -> IO Bool

Returns True if the latch is set for shutdown and vice-versa