Hasura.StoredProcedure.IR

Description

The RQL IR representation of an invocation of a stored procedure.

Documentation

data StoredProcedure b field

The RQL IR representation of an invocation of a stored procedure.

Constructors

StoredProcedure

Fields

spStoredProcedure :: FunctionName b

The SQL name of the stored procedure.

spGraphqlName :: Name

The graphql name of the stored procedure.

spArgs :: HashMap ArgumentName (ScalarType b, field)

The arguments passed to the query, if any.

spLogicalModel :: LogicalModel b

The return type of the stored procedure