Hasura.StoredProcedure.Metadata

Description

Metadata representation of a stored procedure in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

data StoredProcedureMetadata (b :: BackendType)

The representation of native queries within the metadata structure.

Constructors

StoredProcedureMetadata

Fields

_spmRootFieldName :: StoredProcedureName

_spmCode :: InterpolatedQuery ArgumentName

_spmReturns :: LogicalModelName

_spmArguments :: HashMap ArgumentName (NullableScalarType b)

_spmArrayRelationships :: Relationships (RelDef (RelManualConfig b))

_spmDescription :: Maybe Text

spmRootFieldName :: forall b. Lens' (StoredProcedureMetadata b) StoredProcedureName

spmReturns :: forall b. Lens' (StoredProcedureMetadata b) LogicalModelName

spmDescription :: forall b. Lens' (StoredProcedureMetadata b) (Maybe Text)

spmCode :: forall b. Lens' (StoredProcedureMetadata b) (InterpolatedQuery ArgumentName)

spmArrayRelationships :: forall b. Lens' (StoredProcedureMetadata b) (Relationships (RelDef (RelManualConfig b)))

spmArguments :: forall b. Lens' (StoredProcedureMetadata b) (HashMap ArgumentName (NullableScalarType b))