Hasura.StoredProcedure.Types

Description

Types for stored procedures.

Documentation

data StoredProcedureConfig

Tracked stored procedure configuration, and payload of the 'pg_track_stored procedure'.

Constructors

StoredProcedureConfig

Fields

_spcExposedAs :: StoredProcedureExposedAs

In which top-level field should we expose this stored procedure?

_spcCustomName :: Maybe Name

data StoredProcedureExposedAs

Indicates whether the user requested the corresponding stored procedure to be tracked as a mutation or a query, in track_stored_procedure. currently only query is supported.

Constructors

SPEAQuery