Hasura.Table.API

Description

Documentation

data TrackTable b

Constructors

TrackTable

Fields

tSource :: SourceName

tName :: TableName b

tIsEnum :: Bool

tApolloFedConfig :: Maybe ApolloFederationConfig

data SetTableIsEnum b

Constructors

SetTableIsEnum

Fields

stieSource :: SourceName

stieTable :: TableName b

stieIsEnum :: Bool

data UntrackTable b

Constructors

UntrackTable

Fields

utSource :: SourceName

utTable :: TableName b

utCascade :: Bool

checkConflictingNode :: forall m. MonadError QErr m => SchemaCache -> Text -> m ()

Check whether a given name would conflict with the current schema by doing an internal introspection

runTrackTableQ :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => TrackTable b -> m EncJSON

data TrackTableV2 b

Constructors

TrackTableV2

Fields

ttv2Table :: TrackTable b

ttv2Configuration :: TableConfig b

runTrackTableV2Q :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => TrackTableV2 b -> m EncJSON

data TrackTables b

Constructors

TrackTables

Fields

_ttv2Tables :: [TrackTableV2 b]

_ttv2AllowWarnings :: AllowWarnings

runTrackTablesQ :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => TrackTables b -> m EncJSON

data UntrackTables b

Constructors

UntrackTables

Fields

_utTables :: [UntrackTable b]

_utAllowWarnings :: AllowWarnings

runUntrackTablesQ :: forall b m. (CacheRWM m, QErrM m, MetadataM m, BackendMetadata b, BackendEventTrigger b, MonadIO m) => UntrackTables b -> m EncJSON

runSetExistingTableIsEnumQ :: forall b m. (MonadError QErr m, CacheRWM m, MetadataM m, BackendMetadata b) => SetTableIsEnum b -> m EncJSON

data SetTableCustomization b

Constructors

SetTableCustomization

Fields

_stcSource :: SourceName

_stcTable :: TableName b

_stcConfiguration :: TableConfig b

data SetTableCustomFields

Constructors

SetTableCustomFields

Fields

_stcfSource :: SourceName

_stcfTable :: QualifiedTable

_stcfCustomRootFields :: TableCustomRootFields

_stcfCustomColumnNames :: HashMap (Column ('Postgres 'Vanilla)) Name

runSetTableCustomFieldsQV2 :: (QErrM m, CacheRWM m, MetadataM m) => SetTableCustomFields -> m EncJSON

runSetTableCustomization :: forall b m. (QErrM m, CacheRWM m, MetadataM m, Backend b) => SetTableCustomization b -> m EncJSON

runUntrackTableQ :: forall b m. (CacheRWM m, QErrM m, MetadataM m, BackendMetadata b, BackendEventTrigger b, MonadIO m) => UntrackTable b -> m EncJSON

buildTableCache :: forall arr m b. (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectItem) arr, ArrowCache m arr, MonadIO m, MonadBaseControl IO m, BackendMetadata b) => (SourceName, SourceConfig b, DBTablesMetadata b, [TableBuildInput b], Dependency InvalidationKey, NamingCase) `arr` HashMap (TableName b) (TableCoreInfoG b (StructuredColumnInfo b) (ColumnInfo b))

Builds an initial table cache. Does not fill in permissions or event triggers, and the returned FieldInfoMaps only contain columns, not relationships; those pieces of information are filled in later.

data SetApolloFederationConfig b

Constructors

SetApolloFederationConfig

Fields

_safcSource :: SourceName

_safcTable :: TableName b

_safcApolloFederationConfig :: Maybe ApolloFederationConfig

Apollo Federation config for the table, setting Nothing would disable Apollo Federation support on the table.

runSetApolloFederationConfig :: forall b m. (QErrM m, CacheRWM m, MetadataM m, Backend b) => SetApolloFederationConfig b -> m EncJSON