Hasura.Tracing

Documentation

type TracingMetadata = [(Text, Text)]

Any additional human-readable key-value pairs relevant to the execution of a block of code.

newtype Reporter

Constructors

Reporter

Fields

runReporter :: forall io a. MonadIO io => TraceContext -> Text -> io (a, TracingMetadata) -> io a

noReporter :: Reporter

class Monad m => HasReporter m where

A type class for monads which support some way to report execution traces.

See instance Tracing.HasReporter (AppM impl) in HasuraPro.App.

Minimal complete definition

Nothing

Methods

askReporter :: m Reporter

Get the current tracer

default askReporter :: m Reporter

data TraceContext

A trace context records the current active trace, the active span within that trace, and the span's parent, unless the current span is the root.

Constructors

TraceContext

Fields

tcCurrentTrace :: !TraceId

TODO what is this exactly? The topmost span id?

tcCurrentSpan :: !SpanId

tcCurrentParent :: !(Maybe SpanId)

newtype TraceT m a

The TraceT monad transformer adds the ability to keep track of the current trace context.

Constructors

TraceT

Fields

unTraceT :: ReaderT (TraceContext, Reporter) (WriterT TracingMetadata m) a

runTraceT :: (HasReporter m, MonadIO m) => Text -> TraceT m a -> m a

Run an action in the TraceT monad transformer. runTraceT delimits a new trace with its root span, and the arguments specify a name and metadata for that span.

runTraceTWith :: MonadIO m => TraceContext -> Reporter -> Text -> TraceT m a -> m a

runTraceTInContext :: (MonadIO m, HasReporter m) => TraceContext -> Text -> TraceT m a -> m a

Run an action in the TraceT monad transformer in an existing context.

runTraceTWithReporter :: MonadIO m => Reporter -> Text -> TraceT m a -> m a

Run an action in the TraceT monad transformer in an existing context.

class Monad m => MonadTrace m where

Monads which support tracing. TraceT is the standard example.

Methods

trace :: Text -> m a -> m a

Trace the execution of a block of code, attaching a human-readable name.

currentContext :: m TraceContext

Ask for the current tracing context, so that we can provide it to any downstream services, e.g. in HTTP headers.

currentReporter :: m Reporter

Ask for the current tracing reporter

attachMetadata :: TracingMetadata -> m ()

Log some metadata to be attached to the current span

interpTraceT :: MonadTrace n => (m (a, TracingMetadata) -> n (b, TracingMetadata)) -> TraceT m a -> n b

Reinterpret a TraceT action in another MonadTrace. This can be useful when you need to reorganize a monad transformer stack, for example, to embed an action in some monadic computation, while preserving tracing metadata and context.

For example, we use this function in various places in BackendExecute, where we receive an action to execute in some concrete monad transformer stack. See the various implementations of runQuery for examples. Ideally, the input computation's type would be sufficiently polymorphic that we would not need to reorder monads inthe transformer stack. However, the monad transformer stacks must be concrete, because their types are defined by an associated type family ExecutionMonad. Hence, we need to use this function to peel off the outermost TraceT constructor, and embed the computation in some other MonadTrace.

A second example is related to caching. The cacheLookup function returns an action in a concrete transformer stack, again because we are constrained by the usage of a type class. We need to reinterpret the TraceT component of this concrete stack in some other abstract monad transformer stack, using this function.

Laws:

interpTraceT id (hoist f (TraceT x)) = interpTraceT f (TraceT x)

injectB3HttpContext :: TraceContext -> [Header]

Inject the trace context as a set of HTTP headers.

extractB3HttpContext :: [Header] -> IO (Maybe TraceContext)

Extract the trace and parent span headers from a HTTP request and create a new TraceContext. The new context will contain a fresh span ID, and the provided span ID will be assigned as the immediate parent span.

injectEventContext :: TraceContext -> Value

Inject the trace context as a JSON value, appropriate for storing in (e.g.) an event trigger payload.

extractEventContext :: Value -> IO (Maybe TraceContext)

Extract a trace context from an event trigger payload.

tracedHttpRequest

Arguments

:: MonadTrace m

=> Request
http request that needs to be made

-> (Request -> m a)
a function that takes the traced request and executes it

-> m a

Perform HTTP request which supports Trace headers using a HTTP.Request value

TODO REFACTOR: - inline performRequest so that we can be sure a trace is always logged - Inline try here since we always use that at call sites