Hasura.Tracing.Context

Documentation

type TraceMetadata = [(Text, Text)]

Any additional human-readable key-value pairs relevant to the execution of a span.

When the Open Telemetry exporter is in use these become attributes. Where possible and appropriate, consider using key names from the documented OT semantic conventions here: https://opentelemetry.io/docs/reference/specification/trace/semantic_conventions/ This can serve to document the metadata, even for users not using open telemetry.

We may make this type more closely align with the OT data model in the future (e.g. supporting int, etc)

data TraceContext

A trace context records the current active trace, the active span within that trace, and the span's parent, unless the current span is the root. This is like a call stack.

Constructors

TraceContext

Fields

tcCurrentTrace :: TraceId

tcCurrentSpan :: SpanId

tcCurrentParent :: Maybe SpanId

tcSamplingState :: SamplingState