Hasura.Tracing.Propagator

Documentation

data Propagator inboundCarrier outboundCarrier

A carrier is the medium used by Propagators to read values from and write values to. Each specific Propagator type defines its expected carrier type, such as a string map or a byte array.

Constructors

Propagator

Fields

extractor :: inboundCarrier -> SpanId -> Maybe TraceContext

injector :: TraceContext -> outboundCarrier -> outboundCarrier

type HttpPropagator = Propagator RequestHeaders ResponseHeaders

extract

Arguments

:: MonadIO m

=> Propagator i o

-> i
The carrier that holds the propagation fields. For example, an incoming message or HTTP request.

-> m TraceContext
a new Context derived from the Context passed as argument, containing the extracted value, which can be a SpanContext, Baggage or another cross-cutting concern context.

Extracts the value from an incoming request. For example, from the headers of an HTTP request.

If a value can not be parsed from the carrier, for a cross-cutting concern, the implementation MUST NOT throw an exception and MUST NOT store a new value in the Context, in order to preserve any previously existing valid value.

inject

Arguments

:: Propagator i o

-> TraceContext

-> o
The carrier that holds the propagation fields. For example, an outgoing message or HTTP request.

-> o

Injects the value into a carrier. For example, into the headers of an HTTP request.