Hasura.Tracing.Propagator.W3CTraceContext

Description

This module provides support for tracing context propagation in accordance with the W3C tracing context propagation specifications: https://www.w3.org/TR/trace-context/

Documentation

w3cTraceContextPropagator :: Propagator RequestHeaders ResponseHeaders

Propagate trace context information via headers using the w3c specification format