Hasura.Tracing.TraceId

Documentation

data TraceId

128-bit trace identifiers.

TraceIds are guaranteed to have at least one non-zero bit.

randomTraceId :: MonadIO m => m TraceId

traceIdFromBytes :: ByteString -> Maybe TraceId

Create a TraceId from a ByteString.

Fails if the ByteString is not exactly 16 bytes long, or if it contains only zero bytes.

traceIdToBytes :: TraceId -> ByteString

Convert a TraceId to a ByteString of 16 bytes.

traceIdFromHex :: ByteString -> Maybe TraceId

Parse a TraceId from the standard ASCII-encoded hex string format.

Fails if the ByteString is not exactly 32 characters long, or if it contains only zero characters.

traceIdToHex :: TraceId -> ByteString

Serialize a TraceId to the standard ASCII-encoded hex representation.

data SpanId

64-bit span identifiers

SpanIds are guaranteed to have at least one non-zero bit.

randomSpanId :: MonadIO m => m SpanId

spanIdFromBytes :: ByteString -> Maybe SpanId

Create a SpanId from a ByteString.

Fails if the ByteString is not exactly 8 bytes long, or if it contains only zero bytes.

spanIdToBytes :: SpanId -> ByteString

Convert a SpanId to a ByteString of 8 bytes.

spanIdFromHex :: ByteString -> Maybe SpanId

Parse a SpanId from the standard ASCII-encoded hex string format.

Fails if the ByteString is not exactly 16 characters long, or if it contains only zero characters.

spanIdToHex :: SpanId -> ByteString

Serialize a SpanId to the standard ASCII-encoded hex representation.