Net.IPv4

Documentation

data IPv4Range

The length should be between 0 and 32. These bounds are inclusive. This expectation is not in any way enforced by this library because it does not cause errors. A mask length greater than 32 will be treated as if it were 32.

newtype IPv4

A 32-bit Internet Protocol version 4 address. To use this with the network library, it is necessary to use Network.Socket.htonl to convert the underlying Word32 from host byte order to network byte order.

Constructors

IPv4

Fields

getIPv4 :: Word32

decodeRange :: Text -> Maybe IPv4Range

Decode an IPv4Range from Text.

>>> IPv4.decodeRange "172.16.0.0/12"
Just (IPv4Range {ipv4RangeBase = ipv4 172 16 0 0, ipv4RangeLength = 12})
>>> IPv4.decodeRange "192.168.25.254/16"
Just (IPv4Range {ipv4RangeBase = ipv4 192 168 0 0, ipv4RangeLength = 16})

parser :: Parser IPv4

Parse an IPv4 address using a Parser.

>>> AT.parseOnly IPv4.parser "192.168.2.47"
Right (ipv4 192 168 2 47)

>>> AT.parseOnly IPv4.parser "192.168.2.470"
Left "Failed reading: All octets in an IPv4 address must be between 0 and 255"

member :: IPv4 -> IPv4Range -> Bool

This is provided to mirror the interface provided by Data.Set. It behaves just like contains but with flipped arguments.

IPv4.member ip r == IPv4.contains r ip

fromTupleOctets :: (Word8, Word8, Word8, Word8) -> IPv4

An uncurried variant of fromOctets.