Net.IPv6

Documentation

decodeRange :: Text -> Maybe IPv6Range

Decode an IPv6Range from Text.

>>> addr = IPv6.ipv6 0xDEAD 0xBEEF 0x3240 0xA426 0xBA68 0x1CD0 0x4263 0x109B
>>> fmap IPv6.encodeRange $ IPv6.decodeRange (Text.pack "dead:beef:3240:a426:ba68:1cd0:4263:109b/28")
Just "dead:bee0::/28"

data IPv6Range

An IPv6Range. It is made up of the first IPv6 in the range and its length.

fromTupleWord32s :: (Word32, Word32, Word32, Word32) -> IPv6

Uncurried variant of fromWord32s.

member :: IPv6 -> IPv6Range -> Bool

This is provided to mirror the interface provided by Data.Set. It behaves just like contains but with flipped arguments.

IPv6.member ip r == IPv6.contains r ip