Network.HTTP.Client.CreateManager

Documentation

mkHttpManager :: IO [TlsAllow] -> Blocklist -> IO Manager

This mkHttpManager function takes a mechanism for finding the current allowlist, | Thus allowing it to be coupled from any ref type such as SchemaCacheRef. | A mechanism to block IPs (both IPv4 and IPv6) has also been added to it.