Network.HTTP.Client.DynamicTlsPermissions

Documentation

newtype TlsServiceDefinitionError

Constructors

TlsServiceDefinitionError

Fields

tlsServiceDefinitionError :: String

errorE :: String -> c

mkHttpManager :: IO [TlsAllow] -> IO Manager

This mkHttpManager function takes a mechanism for finding the current allowlist, Thus allowing it to be coupled from any ref type such as SchemaCacheRef.