Network.HTTP.Client.Restricted

Description

Restricted ManagerSettings for https://haskell-lang.org/library/http-client - - Portions from http-client-tls Copyright (c) 2013 Michael Snoyman - Portions from http-client-restricted Copyright 2018 Joey Hess id@joeyh.name - - License: MIT

Documentation

data Decision

Constructors

Allow

Deny

type Restriction = AddrInfo -> Decision

data ConnectionRestricted

Blocked requests raise this exception, wrapped as InternalException.

Constructors

ConnectionRestricted

Fields

crHostName :: String

crAddress :: AddrInfo

mkRestrictedManagerSettings :: Restriction -> Maybe ConnectionContext -> Maybe TLSSettings -> ManagerSettings

Makes a TLS-capable ManagerSettings with a Restriction applied to it.

The Restriction will be checked each time a Request is made, and for each redirect followed.

Aside from checking the Restriction, it should behave the same as mkManagerSettingsContext from http-client-tls.

 main = do
 	manager <- newManager $ mkRestrictedManagerSettings myRestriction Nothing Nothing
	request <- parseRequest "http://httpbin.org/get"
 	response <- httpLbs request manager
 	print $ responseBody response

See mkManagerSettingsContext for why it can be useful to provide a ConnectionContext.

Note that SOCKS is not supported.