Autodocodec.Extended

Documentation

optionalFieldOrIncludedNull

Arguments

:: HasCodec output

=> Text
Key

-> Text
Documentation

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNull in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

optionalFieldOrIncludedNull'

Arguments

:: HasCodec output

=> Text
Key

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNull' in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

optionalFieldOrIncludedNullWith

Arguments

:: Text
Key

-> JSONCodec output
Codec for the value

-> Text
Documentation

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNullWith in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

optionalFieldOrIncludedNullWith'

Arguments

:: Text
Key

-> JSONCodec output
Codec for the value

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNullWith' in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

orIncludedNullHelper :: ObjectCodec (Maybe (Maybe input)) (Maybe (Maybe output)) -> ObjectCodec (Maybe input) (Maybe output)