Autodocodec.Extended

Documentation

caseInsensitiveHashMapCodec :: forall k a. (FoldCase k, Hashable k, FromJSONKey k, ToJSONKey k) => JSONCodec a -> JSONCodec (HashMap (CI k) a)

Like hashMapCodec, but with case-insensitive keys.

caseInsensitiveTextCodec :: forall a. (FoldCase a, HasCodec a) => JSONCodec (CI a)

Codec for case-insensitive strings / text. The underlying value may be Text or another type that implements FoldCase and HasCodec.

graphQLEnumValueCodec :: JSONCodec EnumValue

graphQLExecutableDocumentCodec :: JSONCodec (ExecutableDocument Name)

graphQLFieldNameCodec :: JSONCodec Name

Codec for a GraphQL field name

graphQLFieldDescriptionCodec :: JSONCodec Description

graphQLValueCodec :: forall var. Typeable var => JSONCodec var -> JSONCodec (Value var)

graphQLSchemaDocumentCodec :: JSONCodec SchemaDocument

hashSetCodec :: (Hashable a, HasCodec a) => JSONCodec (HashSet a)

Serializes a hash set by converting it to a list. This matches the FromJSON and ToJSON instances in aeson.

hashSetCodecWith :: Hashable a => JSONCodec a -> JSONCodec (HashSet a)

Serializes a hash set by converting it to a list. This matches the FromJSON and ToJSON instances in aeson. This version accepts a codec for individual set values as an argument.

integerCodec :: JSONCodec Integer

Codec for integer with a generous bounds check that matches the behavior of aeson integer deserialization.

optionalFieldOrIncludedNull

Arguments

:: HasCodec output

=> Text
Key

-> Text
Documentation

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNull in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

optionalFieldOrIncludedNull'

Arguments

:: HasCodec output

=> Text
Key

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNull' in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

optionalFieldOrIncludedNullWith

Arguments

:: Text
Key

-> JSONCodec output
Codec for the value

-> Text
Documentation

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNullWith in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

optionalFieldOrIncludedNullWith'

Arguments

:: Text
Key

-> JSONCodec output
Codec for the value

-> ObjectCodec (Maybe output) (Maybe output)

An optional field that might be null where a Nothing value should be represented as null on serialization instead of omitting the field.

This differs from Autodocodec's stock optionalFieldOrNullWith' in that that function omits the field during serialization if the Haskell value is Nothing. This version includes the field with a serialized value of null.

refinedCodec :: (HasCodec a, Predicate p a) => JSONCodec (Refined p a)

Codec for values wrapped with a type-level predicate using the refined package.

This version assumes that the underlying value type implements HasCodec.

refinedCodecWith :: Predicate p a => JSONCodec a -> JSONCodec (Refined p a)

Codec for values wrapped with a type-level predicate using the refined package.

This version requires a codec to be provided for the underlying value type.