Data.HashMap.Strict.Extended

Documentation

catMaybes :: HashMap k (Maybe v) -> HashMap k v

fromListOn :: (Eq k, Hashable k) => (v -> k) -> [v] -> HashMap k v

unionsAll :: (Eq k, Hashable k, Foldable t) => t (HashMap k v) -> HashMap k (NonEmpty v)

Like unions, but keeping all elements in the result.

groupOn :: (Eq k, Hashable k, Foldable t) => (v -> k) -> t v -> HashMap k [v]

Given a Foldable sequence of values and a function that extracts a key from each value, returns a HashMap that maps each key to a list of all values in the sequence for which the given function produced it.

>>> groupOn (take 1) ["foo", "bar", "baz"]
fromList [("f", ["foo"]), ("b", ["bar", "baz"])]

groupOnNE :: (Eq k, Hashable k, Foldable t) => (v -> k) -> t v -> HashMap k (NonEmpty v)

differenceOn :: (Eq k, Hashable k, Foldable t) => (v -> k) -> t v -> t v -> HashMap k v

lpadZip :: (Eq k, Hashable k) => HashMap k a -> HashMap k b -> HashMap k (Maybe a, b)

Analogous to lpadZip, but on HashMaps instead of lists.

mapKeys :: (Ord k2, Hashable k2) => (k1 -> k2) -> HashMap k1 a -> HashMap k2 a

mapKeys f s is the map obtained by applying f to each key of s.

The size of the result may be smaller if f maps two or more distinct keys to the same new key. In this case the value at the greatest of the original keys is retained.

mapKeys (+ 1) (fromList [(5,"a"), (3,"b")])                        == fromList [(4, "b"), (6, "a")]
mapKeys (\ _ -> 1) (fromList [(1,"b"), (2,"a"), (3,"d"), (4,"c")]) == singleton 1 "c"
mapKeys (\ _ -> 3) (fromList [(1,"b"), (2,"a"), (3,"d"), (4,"c")]) == singleton 3 "c"

copied from https://hackage.haskell.org/package/containers-0.6.4.1/docs/src/Data.Map.Internal.html#mapKeys

unionsWith :: (Foldable f, Hashable k, Ord k) => (a -> a -> a) -> f (HashMap k a) -> HashMap k a

The union of a list of maps, with a combining operation: (unionsWith f == foldl (unionWith f) empty).

unionsWith (++) [(fromList [(5, "a"), (3, "b")]), (fromList [(5, "A"), (7, "C")]), (fromList [(5, "A3"), (3, "B3")])]
    == fromList [(3, "bB3"), (5, "aAA3"), (7, "C")]

copied from https://hackage.haskell.org/package/containers-0.6.4.1/docs/src/Data.Map.Internal.html#unionsWith

isInverseOf :: (Eq k, Hashable k, Eq v, Hashable v) => HashMap k v -> HashMap v k -> Bool

Determines whether the left-hand-side and the right-hand-side are inverses of each other.

More specifically, for two maps A and B, isInverseOf is satisfied when both of the following are true: 1. ∀ key ∈ A. A[key] ∈ B ∧ B[A[key]] == key 2. ∀ key ∈ B. B[key] ∈ A ∧ A[B[key]] == key