Data.HashMap.Strict.InsOrd.Extended

Documentation

catMaybes :: InsOrdHashMap k (Maybe v) -> InsOrdHashMap k v

partition :: (Eq k, Hashable k) => (v -> Bool) -> InsOrdHashMap k v -> (InsOrdHashMap k v, InsOrdHashMap k v)

alterF :: (Functor f, Eq k, Hashable k) => (Maybe v -> f (Maybe v)) -> k -> InsOrdHashMap k v -> f (InsOrdHashMap k v)

Alter a hashmap using a function that can fail, in which case the entire operation fails. (Maybe a version with the key also being passed to the function could be useful.)