Data.Parser.CacheControl

Description

Functions related to parsing Cache-Control header as defined in https://tools.ietf.org/html/rfc7234#section-5.2

To get max-age/s-maxage from Cache-Control header, use parseMaxAge. If you need to check other directives use parseCacheControl.

Rules which starts with obs- is not required to implement because they are maked as "obsolete" as per https://tools.ietf.org/html/rfc7230#section-1.2

Documentation

type CacheControl = [CacheControlDirective]

data CacheControlDirective

Constructors

CCDOnlyToken !Text

CCDTokenWithVal !Text !Text

parseMaxAge :: Integral a => Text -> Either String a

Tries to parse the max-age or s-maxage present in the value of Cache-Control header

findMaxAge :: Integral a => CacheControl -> Either String (Maybe a)

noCacheExists :: CacheControl -> Bool

Checks if the no-cache directive is present

noStoreExists :: CacheControl -> Bool

Checks if the no-store directive is present

mustRevalidateExists :: CacheControl -> Bool

Checks if the must-revalidate directive is present

findCCDOnlyToken :: (Text -> Bool) -> CacheControl -> Maybe Text

findCCDTokenWithVal :: (Text -> Bool) -> CacheControl -> Maybe (Text, Text)

parseCacheControl :: Text -> Either String CacheControl

Parses a Cache-Control header and returns a list of directives

cacheControlParser :: Parser CacheControl

optionalWhitespaceParser :: Parser (Maybe Char)

cacheDirectiveParser :: Parser CacheControlDirective

tokenParser :: Parser Text

tcharParser :: Parser Char

dquoteParser :: Parser Char

vcharParser :: Parser Char

quotedStringParser :: Parser Text

quotedPairParser :: Parser Char

qdTextParser :: Parser Char