Database.MSSQL.Transaction

Documentation

newtype TxET e m a

The transaction command to run, parameterised over: e - the exception type (usually MSSQLTxError) m - some Monad, (usually some MonadIO) a - the successful result type

Constructors

TxET

Fields

txHandler :: ReaderT Connection (ExceptT e m) a

data MSSQLTxError

Error type generally used in TxET.

Constructors

MSSQLQueryError !Query !ODBCException

MSSQLConnError !ODBCException

MSSQLInternal !Text

type TxT m a = TxET MSSQLTxError m a

The transaction command to run, returning an MSSQLTxError or the result.

runTx :: (MonadIO m, MonadBaseControl IO m) => TxT m a -> MSSQLPool -> ExceptT MSSQLTxError m a

Run a command on the given connection wrapped in a transaction.

See runTxE if you need to map the error type as well.

runTxE :: (MonadIO m, MonadBaseControl IO m) => (MSSQLTxError -> e) -> TxET e m a -> MSSQLPool -> ExceptT e m a

Run a command on the given connection wrapped in a transaction.

unitQuery :: MonadIO m => Query -> TxT m ()

Useful for building transactions which return no data.

insertId :: TxT m ()
insertId = unitQuery "INSERT INTO some_table VALUES (1, "hello")"

See unitQueryE if you need to map the error type as well.

unitQueryE :: MonadIO m => (MSSQLTxError -> e) -> Query -> TxET e m ()

Useful for building transactions which return no data.

singleRowQuery :: forall a m. (MonadIO m, FromRow a) => Query -> TxT m a

Useful for building query transactions which return a single one row.

returnOne :: TxT m Int
returnOne = singleRowQuery "SELECT 1"

See singleRowQueryE if you need to map the error type as well.

singleRowQueryE :: forall m a e. (MonadIO m, FromRow a) => (MSSQLTxError -> e) -> Query -> TxET e m a

Useful for building query transactions which return a single one row.

multiRowQuery :: forall a m. (MonadIO m, FromRow a) => Query -> TxT m [a]

Useful for building query transactions which return multiple rows.

selectIds :: TxT m [Int]
selectIds = multiRowQuery "SELECT id FROM author"

See multiRowQueryE if you need to map the error type as well.

multiRowQueryE :: forall m a e. (MonadIO m, FromRow a) => (MSSQLTxError -> e) -> Query -> TxET e m [a]

Useful for building query transactions which return multiple rows.

buildGenericQueryTxE

Arguments

:: MonadIO m

=> (MSSQLTxError -> e)
map MSSQLTxError to some other type

-> query
query to run

-> (query -> Query)
how to map a query to a Query

-> (Connection -> query -> IO a)
run the query on a provided Connection

-> TxET e m a

Build a generic transaction out of an IO action.

withTxET :: Monad m => (e1 -> e2) -> TxET e1 m a -> TxET e2 m a

Map the error type for a TxET.

newtype MSSQLResult

A successful result from a query is a list of rows where each row contains list of column values

Constructors

MSSQLResult [[Value]]

rawQueryE

Arguments

:: MonadIO m

=> (MSSQLTxError -> e)
Error modifier

-> (MSSQLResult -> Either String a)
Result modifier with a failure

-> Query
Query to run

-> TxET e m a

Packs a query, along with result and error converters into a TxET.

Used by unitQueryE, singleRowQueryE, and multiRowQueryE.

execQuery :: forall m a query. MonadIO m => query -> (query -> Query) -> (query -> IO a) -> ExceptT MSSQLTxError m a

Combinator for abstracting over the query type and ensuring we catch exceptions.

Used by buildGenericQueryTxE.

execTx :: Connection -> TxET e m a -> ExceptT e m a

Run a TxET with the given connection.

Used by runTxE and asTransaction.

data TransactionState

The transaction state of the current connection

Constructors

TSActive
Has an active transaction.

TSNoActive
Has no active transaction.

TSUncommittable
An error occurred that caused the transaction to be uncommittable. We cannot commit or rollback to a savepoint; we can only do a full rollback of the transaction.

asTransaction :: forall e a m. MonadIO m => (MSSQLTxError -> e) -> (Connection -> ExceptT e m a) -> Connection -> ExceptT e m a

Wraps an action in a transaction. Rolls back on errors.

beginTx :: MonadIO m => TxT m ()

commitTx :: MonadIO m => TxT m ()

rollbackTx :: MonadIO m => TxT m ()

getTransactionState :: MonadIO m => TxT m TransactionState

Get the @TransactionState of current connection For more details, refer to https://docs.microsoft.com/en-us/sql/t-sql/functions/xact-state-transact-sql?view=sql-server-ver15