Hasura.App

Documentation

data ExitCode

Constructors

InvalidEnvironmentVariableOptionsError

InvalidDatabaseConnectionParamsError

AuthConfigurationError

EventSubSystemError

DatabaseMigrationError

SchemaCacheInitError
used by MT because it initialises the schema cache only these are used in app/Main.hs:

MetadataExportError

MetadataCleanError

ExecuteProcessError

DowngradeProcessError

data ExitException

Constructors

ExitException

Fields

eeCode :: !ExitCode

eeMessage :: !ByteString

throwErrExit :: MonadIO m => forall a. ExitCode -> String -> m a

throwErrJExit :: (ToJSON a, MonadIO m) => forall b. ExitCode -> a -> m b

parseHGECommand :: EnabledLogTypes impl => Parser (RawHGECommand impl)

parseArgs :: EnabledLogTypes impl => IO (HGEOptions impl)

printJSON :: (ToJSON a, MonadIO m) => a -> m ()

printYaml :: (ToJSON a, MonadIO m) => a -> m ()

mkPGLogger :: Logger Hasura -> PGLogger

data GlobalCtx

Context required for all graphql-engine CLI commands

Constructors

GlobalCtx

Fields

_gcMetadataDbConnInfo :: !ConnInfo

_gcDefaultPostgresConnInfo :: !(Maybe (UrlConf, ConnInfo), Maybe Int)

-database-url option, @UrlConf is required to construct default source configuration and optional retries

readTlsAllowlist :: SchemaCacheRef -> IO [TlsAllow]

initGlobalCtx

Arguments

:: MonadIO m

=> Environment

-> Maybe String
the metadata DB URL

-> PostgresConnInfo (Maybe UrlConf)
the user's DB URL

-> m GlobalCtx

data ServeCtx

Context required for the serve CLI command.

Constructors

ServeCtx

Fields

_scHttpManager :: !Manager

_scInstanceId :: !InstanceId

_scLoggers :: !Loggers

_scEnabledLogTypes :: !(HashSet (EngineLogType Hasura))

_scMetadataDbPool :: !PGPool

_scShutdownLatch :: !ShutdownLatch

_scSchemaCache :: !RebuildableSchemaCache

_scSchemaCacheRef :: !SchemaCacheRef

_scMetaVersionRef :: !(TMVar MetadataResourceVersion)

data Loggers

Collection of the LoggerCtx, the regular Logger and the PGLogger TODO (from master): better naming?

Constructors

Loggers

Fields

_lsLoggerCtx :: !(LoggerCtx Hasura)

_lsLogger :: !(Logger Hasura)

_lsPgLogger :: !PGLogger

newtype PGMetadataStorageAppT m a

An application with Postgres database as a metadata storage

Constructors

PGMetadataStorageAppT

Fields

runPGMetadataStorageAppT :: (PGPool, PGLogger) -> m a

resolvePostgresConnInfo :: MonadIO m => Environment -> UrlConf -> Maybe Int -> m ConnInfo

initialiseServeCtx :: (ForkableMonadIO m, MonadCatch m) => Environment -> GlobalCtx -> ServeOptions Hasura -> ServerMetrics -> ManagedT m ServeCtx

Initializes or migrates the catalog and returns the context required to start the server.

mkLoggers :: (MonadIO m, MonadBaseControl IO m) => HashSet (EngineLogType Hasura) -> LogLevel -> ManagedT m Loggers

migrateCatalogSchema :: (MonadIO m, MonadBaseControl IO m) => Environment -> Logger Hasura -> PGPool -> Maybe (SourceConnConfiguration ('Postgres 'Vanilla)) -> Manager -> ServerConfigCtx -> SourceResolver ('Postgres 'Vanilla) -> SourceResolver 'MSSQL -> m RebuildableSchemaCache

helper function to initialize or migrate the hdb_catalog schema (used by pro as well)

newtype ShutdownLatch

A latch for the graceful shutdown of a server process.

Constructors

ShutdownLatch

Fields

unShutdownLatch :: MVar ()

data ShutdownAction

Event triggers live in the user's DB and other events (cron, one-off and async actions) live in the metadata DB, so we need a way to differentiate the type of shutdown action

Constructors

EventTriggerShutdownAction (IO ())

MetadataDBShutdownAction (MetadataStorageT IO ())

newShutdownLatch :: IO ShutdownLatch

waitForShutdown :: ShutdownLatch -> IO ()

Block the current thread, waiting on the latch.

shutdownGracefully :: ShutdownLatch -> IO ()

Initiate a graceful shutdown of the server associated with the provided latch.

flushLogger :: MonadIO m => LoggerCtx impl -> m ()

If an exception is encountered , flush the log buffer and rethrow If we do not flush the log buffer on exception, then log lines may be missed See: https://github.com/hasura/graphql-engine/issues/4772

runHGEServer

Arguments

:: forall m impl. (MonadIO m, MonadMask m, MonadStateless IO m, Forall (Pure m), UserAuthentication (TraceT m), HttpLog m, ConsoleRenderer m, MonadMetadataApiAuthorization m, MonadGQLExecutionCheck m, MonadConfigApiHandler m, MonadQueryLog m, MonadWSLog m, MonadExecuteQuery m, HasReporter m, HasResourceLimits m, MonadMetadataStorage (MetadataStorageT m), MonadResolveSource m, MonadQueryTags m)

=> (ServerCtx -> SpockT m ())

-> Environment

-> ServeOptions impl

-> ServeCtx

-> UTCTime
start time

-> Maybe SubscriptionPostPollHook

-> ServerMetrics

-> Store EmptyMetrics

-> ManagedT m ()

This function acts as the entrypoint for the graphql-engine webserver.

Note: at the exit of this function, or in case of a graceful server shutdown (SIGTERM, or more generally, whenever the shutdown latch is set), we need to make absolutely sure that we clean up any resources which were allocated during server setup. In the case of a multitenant process, failure to do so can lead to resource leaks.

To track these resources, we use the ManagedT monad, and attach finalizers at the same point in the code where we allocate resources. If you fork a new long-lived thread, or create a connection pool, or allocate any other long-lived resource, make sure to pair the allocator with its finalizer. There are plenty of examples throughout the code. For example, see forkManagedT.

Note also: the order in which the finalizers run can be important. Specifically, we want the finalizers for the logger threads to run last, so that we retain as many "thread stopping" log messages as possible. The order in which the finalizers is run is determined by the order in which they are introduced in the code.

mkHGEServer

Arguments

:: forall m impl. (MonadIO m, MonadMask m, MonadStateless IO m, Forall (Pure m), UserAuthentication (TraceT m), HttpLog m, ConsoleRenderer m, MonadMetadataApiAuthorization m, MonadGQLExecutionCheck m, MonadConfigApiHandler m, MonadQueryLog m, MonadWSLog m, MonadExecuteQuery m, HasReporter m, HasResourceLimits m, MonadMetadataStorage (MetadataStorageT m), MonadResolveSource m, MonadQueryTags m)

=> (ServerCtx -> SpockT m ())

-> Environment

-> ServeOptions impl

-> ServeCtx

-> UTCTime
start time

-> Maybe SubscriptionPostPollHook

-> ServerMetrics

-> Store EmptyMetrics

-> ManagedT m Application

Part of a factorization of runHGEServer to expose the constructed WAI application for testing purposes. See runHGEServer for documentation.

accessDeniedErrMsg :: Text

runInSeparateTx :: MonadIO m => TxE QErr a -> MetadataStorageT (PGMetadataStorageAppT m) a

notifySchemaCacheSyncTx :: MetadataResourceVersion -> InstanceId -> CacheInvalidations -> TxE QErr ()

getCatalogStateTx :: TxE QErr CatalogState

setCatalogStateTx :: CatalogStateType -> Value -> TxE QErr ()

mkConsoleHTML :: Text -> AuthMode -> Bool -> Maybe Text -> Either String Text

telemetryNotice :: String

mkPgSourceResolver :: PGLogger -> SourceResolver ('Postgres 'Vanilla)

mkMSSQLSourceResolver :: SourceResolver 'MSSQL