Hasura.Authentication.Session

Documentation

adminSecretHeader :: SessionVariable

userIdHeader :: SessionVariable

userRoleHeader :: SessionVariable

useBackendOnlyPermissionsHeader :: SessionVariable

deprecatedAccessKeyHeader :: SessionVariable

data SessionVariable

type SessionVariableValue = Text

parseSessionVariable :: Text -> Parser SessionVariable

class ToSessionVariable a where

Methods

isSessionVariable :: a -> Bool

Tests the input to find out whether it is a valid session variable.

mkSessionVariable :: a -> Maybe SessionVariable

Makes a session variable from the input. Returns Nothing on an invalid input.

fromSessionVariable :: SessionVariable -> a

Converts back to the input (lossily).

unsafeMkSessionVariable :: ToSessionVariable a => a -> SessionVariable

Unsafely makes a session variable. Errors if the input is invalid.

sessionVariableToHeader :: SessionVariable -> SessionVariableValue -> Header

data SessionVariables

singleSessionVariable :: SessionVariable -> SessionVariableValue -> SessionVariables

sessionVariablesFromMap :: HashMap Text Text -> SessionVariables

sessionVariablesWith :: SessionVariable -> SessionVariableValue -> SessionVariables -> SessionVariables

sessionVariablesWithout :: SessionVariable -> SessionVariables -> SessionVariables

sessionVariableToGraphQLName :: SessionVariable -> Maybe Name

Converts a SessionVariable value to a GraphQL name. This will fail if the session variable contains characters that are not valid for a graphql names. It is the caller's responsibility to decide what to do in such a case.

filterSessionVariables :: (SessionVariable -> SessionVariableValue -> Bool) -> SessionVariables -> SessionVariables

mkSessionVariablesHeaders :: [Header] -> SessionVariables

sessionVariablesToHeaders :: SessionVariables -> [Header]

getSessionVariables :: SessionVariables -> [Text]

getSessionVariablesSet :: SessionVariables -> HashSet SessionVariable

getSessionVariableValue :: SessionVariable -> SessionVariables -> Maybe SessionVariableValue

maybeRoleFromSessionVariables :: SessionVariables -> Maybe RoleName

mkClientHeadersForward :: [Header] -> [Header]