Hasura.Backends.BigQuery.Execute

Description

Execute a Select query against the BigQuery REST API.

Documentation

executeSelect :: Select -> Execute (Job, RecordSet)

runExecute :: MonadIO m => BigQuerySourceConfig -> Execute (Job, RecordSet) -> m (Either ExecuteProblem (Job, RecordSet))

streamBigQuery :: MonadIO m => BigQueryConnection -> BigQuery -> m (Either ExecuteProblem (Job, RecordSet))

TODO: WARNING: This function hasn't been tested on Big Data(tm), and therefore I was unable to get BigQuery to produce paginated results that would contain the pageToken field in the JSON response. Until that test has been done, we should consider this a preliminary implementation.

executeBigQuery :: MonadIO m => BigQueryConnection -> BigQuery -> m (Either ExecuteProblem ())

Execute a query without expecting any output (e.g. CREATE TABLE or INSERT)

executeProblemMessage :: ShowDetails -> ExecuteProblem -> Text

insertDataset :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Text -> m Dataset

Insert a new dataset

deleteDataset :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Text -> m ()

Delete a dataset

data BigQuery

Constructors

BigQuery

Fields

query :: Text

parameters :: InsOrdHashMap ParameterName Parameter

data Execute a

Execute monad; as queries are performed, the record sets are stored in the map.

data ExecuteProblem

Constructors

GetJobDecodeProblem String

CreateQueryJobDecodeProblem String

InsertDatasetDecodeProblem String

ExecuteRunBigQueryProblem BigQueryProblem

RESTRequestNonOK Status Value

newtype FieldNameText

As opposed to BigQuery.FieldName which is a qualified name, this is just the unqualified text name itself.

Constructors

FieldNameText Text

data Job

Constructors

Job

Fields

state :: Text

jobId :: Text

location :: Text

data OutputValue

Constructors

DecimalOutputValue Decimal

BigDecimalOutputValue BigDecimal

IntegerOutputValue Int64

FloatOutputValue Float64

GeographyOutputValue Geography

TextOutputValue Text

TimestampOutputValue Timestamp

DateOutputValue Date

TimeOutputValue Time

DatetimeOutputValue Datetime

BytesOutputValue Base64

BoolOutputValue Bool

ArrayOutputValue (Vector OutputValue)

RecordOutputValue (InsOrdHashMap FieldNameText OutputValue)

JsonOutputValue Value

NullOutputValue

data RecordSet

A set of records produced by the database. These are joined together. There are all sorts of optimizations possible here, from using a matrix/flat vector, unboxed sums for Value, etc. Presently we choose a naive implementation in the interest of getting other work done.

Constructors

RecordSet

Fields

rows :: Vector (InsOrdHashMap FieldNameText OutputValue)

wantedFields :: Maybe [Text]

data ShowDetails

We use this to hide certain details from the front-end, while allowing them in tests. We have not actually decided whether showing the details is insecure, but until we decide otherwise, it's probably best to err on the side of caution.

Constructors

HideDetails

InsecurelyShowDetails

data Value

Source for this represenation type:

https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.Value

BigQuery results come in via the REST API as one of these simply types.

TODO: This omits StructValue -- do we need it?

Constructors

NullValue

IntegerValue Int64
64-bit https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#integer_type

DecimalValue Decimal
Fixed precision https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#decimal_types

BigDecimalValue BigDecimal
Fixed precision https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#decimal_types

FloatValue Float64
Floating point https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#floating_point_types

GeographyValue Geography

StringValue Text

BytesValue Base64

BoolValue Bool

ArrayValue (Vector Value)

TimestampValue Timestamp

DateValue Date

TimeValue Time

JsonValue Value

DatetimeValue Datetime