Hasura.Backends.BigQuery.Execute

Description

Execute a Select query against the BigQuery REST API.

Documentation

data RecordSet

A set of records produced by the database. These are joined together. There are all sorts of optimizations possible here, from using a matrix/flat vector, unboxed sums for Value, etc. Presently we choose a naive implementation in the interest of getting other work done.

Constructors

RecordSet

Fields

rows :: Vector (InsOrdHashMap FieldNameText OutputValue)

wantedFields :: Maybe [Text]

newtype FieldNameText

As opposed to BigQuery.FieldName which is a qualified name, this is just the unqualified text name itself.

Constructors

FieldNameText Text

data OutputValue

Constructors

DecimalOutputValue Decimal

BigDecimalOutputValue BigDecimal

IntegerOutputValue Int64

FloatOutputValue Float64

GeographyOutputValue Geography

TextOutputValue Text

TimestampOutputValue Timestamp

DateOutputValue Date

TimeOutputValue Time

DatetimeOutputValue Datetime

BytesOutputValue Base64

BoolOutputValue Bool

ArrayOutputValue (Vector OutputValue)

RecordOutputValue (InsOrdHashMap FieldNameText OutputValue)

NullOutputValue

data ExecuteProblem

Constructors

GetJobDecodeProblem String

CreateQueryJobDecodeProblem String

InsertDatasetDecodeProblem String

ExecuteRunBigQueryProblem BigQueryProblem

RESTRequestNonOK Status Value

data ShowDetails

We use this to hide certain details from the front-end, while allowing them in tests. We have not actually decided whether showing the details is insecure, but until we decide otherwise, it's probably best to err on the side of caution.

Constructors

HideDetails

InsecurelyShowDetails

executeProblemMessage :: ShowDetails -> ExecuteProblem -> Text

data Execute a

Execute monad; as queries are performed, the record sets are stored in the map.

data BigQuery

Constructors

BigQuery

Fields

query :: Text

parameters :: InsOrdHashMap ParameterName Parameter

runExecute :: MonadIO m => BigQuerySourceConfig -> Execute RecordSet -> m (Either ExecuteProblem RecordSet)

executeSelect :: Select -> Execute RecordSet

streamBigQuery :: MonadIO m => BigQueryConnection -> BigQuery -> m (Either ExecuteProblem RecordSet)

TODO: WARNING: This function hasn't been tested on Big Data(tm), and therefore I was unable to get BigQuery to produce paginated results that would contain the JobResults field in the JSON response. Until that test has been done, we should consider this a preliminary implementation.

executeBigQuery :: MonadIO m => BigQueryConnection -> BigQuery -> m (Either ExecuteProblem ())

Execute a query without expecting any output (e.g. CREATE TABLE or INSERT)

deleteDataset :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Text -> m ()

Delete a dataset

insertDataset :: (MonadError ExecuteProblem m, MonadIO m) => BigQueryConnection -> Text -> m Dataset

Insert a new dataset