Hasura.Backends.BigQuery.FromIr

Description

Translate from the DML to the BigQuery dialect.

Documentation

mkSQLSelect :: JsonAggSelect -> AnnSelectG 'BigQuery (AnnFieldG 'BigQuery Void) Expression -> FromIr Select

Here is where we apply a top-level annotation to the select to indicate to the data loader that this select ought to produce a single object or an array.

fromRootField :: QueryDB 'BigQuery Void Expression -> FromIr Select

Convert from the IR database query into a select.

fromSelectAggregate :: Maybe (EntityAlias, HashMap ColumnName ColumnName) -> AnnSelectG 'BigQuery (TableAggregateFieldG 'BigQuery Void) Expression -> FromIr Select

data Error

Most of these errors should be checked for legitimacy.

Constructors

FromTypeUnsupported (SelectFromG 'BigQuery Expression)

NoOrderSpecifiedInOrderBy

MalformedAgg

FieldTypeUnsupportedForNow (AnnFieldG 'BigQuery Void Expression)

AggTypeUnsupportedForNow (TableAggregateFieldG 'BigQuery Void Expression)

NodesUnsupportedForNow (TableAggregateFieldG 'BigQuery Void Expression)

NoProjectionFields

NoAggregatesMustBeABug

UnsupportedArraySelect (ArraySelectG 'BigQuery Void Expression)

UnsupportedOpExpG (OpExpG 'BigQuery Expression)

UnsupportedSQLExp Expression

UnsupportedDistinctOn

UnexpectedEmptyList

InvalidIntegerishSql Expression

ConnectionsNotSupported

ActionsNotSupported

ComputedFieldsBooleanExpressionNotSupported
https://github.com/hasura/graphql-engine/issues/8526

ComputedFieldsOrderByNotSupported
https://github.com/hasura/graphql-engine/issues/8526

ScalarComputedFieldsNotSupported
https://github.com/hasura/graphql-engine/issues/8521

NoParentEntityInternalError

runFromIr :: FromIrConfig -> FromIr a -> Validate (NonEmpty Error) (a, FromIrWriter)

data FromIr a

The base monad used throughout this module for all conversion functions.

It's a Validate, so it'll continue going when it encounters errors to accumulate as many as possible.

It also contains a mapping from entity prefixes to counters. So if my prefix is "table" then there'll be a counter that lets me generate table1, table2, etc. Same for any other prefix needed (e.g. names for joins).

A ReaderT is used around this in most of the module too, for setting the current entity that a given field name refers to. See fromColumn.

newtype FromIrWriter

Collected from using a native query in a query. Each entry here because a CTE to be prepended to the query.

Constructors

FromIrWriter

Fields

fromIrWriterNativeQueries :: Map (Aliased NativeQueryName) (InterpolatedQuery Expression)

data FromIrConfig

Config values for the from-IR translator.

Constructors

FromIrConfig

Fields

globalSelectLimit :: Top

Applies globally to all selects, and may be reduced to something even smaller by permission/user args.

defaultFromIrConfig :: FromIrConfig

A default config.

bigQuerySourceConfigToFromIrConfig :: BigQuerySourceConfig -> FromIrConfig

data Top

Constructors

NoTop

Top Int64