Hasura.Backends.BigQuery.Meta

Documentation

data MetadataError

Constructors

RestProblem RestProblem

data RestProblem

Constructors

GetTablesProblem SomeException

GetTableProblem SomeException

GetRoutineProblem SomeException

GetMetaDecodeProblem String

GetTablesBigQueryProblem BigQueryProblem

GetRoutinesBigQueryProblem BigQueryProblem

RESTRequestNonOK Status

data RestTableList

Constructors

RestTableList

Fields

nextPageToken :: Maybe Text

tables :: [RestTableBrief]

data RestTableBrief

Constructors

RestTableBrief

Fields

tableReference :: RestTableReference

data RestTableReference

Constructors

RestTableReference

Fields

datasetId :: Text

projectId :: Text

tableId :: Text

data RestTable

Constructors

RestTable

Fields

tableReference :: RestTableReference

schema :: RestTableSchema

data RestTableSchema

Constructors

RestTableSchema

Fields

fields :: [RestFieldSchema]

data RestFieldSchema

Constructors

RestFieldSchema

Fields

name :: Text

type' :: RestType

The field data type. Possible values include STRING, BYTES, INTEGER, INT64 (same as INTEGER), FLOAT, FLOAT64 (same as FLOAT), BOOLEAN, BOOL (same as BOOLEAN), TIMESTAMP, DATE, TIME, DATETIME, GEOGRAPHY, NUMERIC, RECORD (where RECORD indicates that the field contains a nested schema) or STRUCT (same as RECORD).

mode :: Mode

data Mode

Constructors

Nullable

Required

Repeated

data RestType

Constructors

STRING

BYTES

INTEGER

FLOAT

BOOL

TIMESTAMP

DATE

TIME

DATETIME

GEOGRAPHY

DECIMAL

BIGDECIMAL

STRUCT

getTables :: MonadIO m => BigQuerySourceConfig -> m (Either RestProblem [RestTable])

Get all tables from all specified data sets.

getTablesForDataSet :: MonadIO m => BigQueryConnection -> Text -> m (Either RestProblem [RestTable])

Get tables in the dataset.

getTable :: MonadIO m => BigQueryConnection -> Text -> Text -> m (Either RestProblem RestTable)

Get tables in the schema.

encodeParams :: [(Text, Text)] -> Text

data RestRoutineType

The fine-grained type of the routine Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/routines#RoutineType

Constructors

ROUTINE_TYPE_UNSPECIFIED

SCALAR_FUNCTION

PROCEDURE

TABLE_VALUED_FUNCTION

data RestArgument

Input argument of a function/routine. Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/routines#Argument

Constructors

RestArgument

Fields

_raName :: Maybe Text

The name of this argument. Can be absent for function return argument.

data RestStandardSqlField

A field or a column. Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/StandardSqlField

Constructors

RestStandardSqlField

Fields

_rssfName :: Maybe Text

The field name is optional and is absent for fields with STRUCT type.

data RestStandardSqlTableType

A table type, which has only list of columns with names and types. Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/routines#StandardSqlTableType

Constructors

RestStandardSqlTableType

Fields

_rrttColumns :: [RestStandardSqlField]

data RestRoutineReference

Id path of a routine. Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/routines#RoutineReference

Constructors

RestRoutineReference

Fields

datasetId :: Text

projectId :: Text

routineId :: Text

routineReferenceToFunctionName :: RestRoutineReference -> FunctionName

data RestRoutine

A user-defined function. Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/routines#Routine

Constructors

RestRoutine

Fields

routineReference :: RestRoutineReference

Reference describing the ID of this routine

routineType :: RestRoutineType

The type of routine

arguments :: Maybe [RestArgument]

List of arguments defined

returnTableType :: Maybe RestStandardSqlTableType

Routines defined with 'RETURNS TABLE' clause has this information

data RestRoutineList

List of routines Ref: https://cloud.google.com/bigquery/docs/reference/rest/v2/routines/list

Constructors

RestRoutineList

Fields

_rrlRoutines :: [RestRoutine]

_rrlNextPageToken :: Maybe Text

getRoutines :: MonadIO m => BigQuerySourceConfig -> m (Either RestProblem [RestRoutine])

Get all routines from all specified data sets.

getRoutinesForDataSet :: MonadIO m => BigQueryConnection -> Text -> m (Either RestProblem [RestRoutine])

Get routines in the dataset.