Hasura.Backends.BigQuery.Types

Description

Types for Transact-SQL aka T-SQL; the language of SQL Server.

Documentation

data Select

Constructors

Select

Fields

selectTop :: !Top

selectProjections :: !(NonEmpty Projection)

selectFrom :: !From

selectJoins :: ![Join]

selectWhere :: !Where

selectOrderBy :: !(Maybe (NonEmpty OrderBy))

selectOffset :: !(Maybe Expression)

selectGroupBy :: [FieldName]

selectFinalWantedFields :: !(Maybe [Text])

selectCardinality :: !Cardinality

data PartitionableSelect

Helper type allowing addition of extra fields used in PARTITION BY.

The main purpose of this type is sumulation of DISTINCT ON implemented in Hasura.Backends.BigQuery.FromIr.simulateDistinctOn

Constructors

PartitionableSelect

Fields

pselectFinalize :: Maybe [FieldName] -> Select

pselectFrom :: !From

simpleSelect :: Select -> PartitionableSelect

noExtraPartitionFields :: PartitionableSelect -> Select

withExtraPartitionFields :: PartitionableSelect -> [FieldName] -> Select

data ArrayAgg

Constructors

ArrayAgg

Fields

arrayAggProjections :: !(NonEmpty Projection)

arrayAggOrderBy :: !(Maybe (NonEmpty OrderBy))

arrayAggTop :: !Top

data Reselect

Constructors

Reselect

Fields

reselectProjections :: !(NonEmpty Projection)

reselectWhere :: !Where

data OrderBy

Constructors

OrderBy

Fields

orderByFieldName :: FieldName

orderByOrder :: Order

orderByNullsOrder :: NullsOrder

data Order

Constructors

AscOrder

DescOrder

data NullsOrder

Constructors

NullsFirst

NullsLast

NullsAnyOrder

data FieldOrigin

Constructors

NoOrigin

AggregateOrigin [Aliased Aggregate]

aggregateProjectionsFieldOrigin :: Projection -> FieldOrigin

data Projection

Constructors

ExpressionProjection (Aliased Expression)

FieldNameProjection (Aliased FieldName)

AggregateProjections (Aliased (NonEmpty (Aliased Aggregate)))

AggregateProjection (Aliased Aggregate)

StarProjection

ArrayAggProjection (Aliased ArrayAgg)

EntityProjection (Aliased [(FieldName, FieldOrigin)])

ArrayEntityProjection EntityAlias (Aliased [FieldName])

WindowProjection (Aliased WindowFunction)

data WindowFunction

Constructors

RowNumberOverPartitionBy (NonEmpty FieldName) (Maybe (NonEmpty OrderBy))
ROW_NUMBER() OVER(PARTITION BY field)

data Join

Constructors

Join

Fields

joinSource :: !JoinSource

joinAlias :: !EntityAlias

joinOn :: [(FieldName, FieldName)]

joinProvenance :: !JoinProvenance

joinFieldName :: !Text

joinExtractPath :: !(Maybe Text)

joinRightTable :: !EntityAlias

data JoinProvenance

Constructors

OrderByJoinProvenance

ObjectJoinProvenance [Text]

ArrayAggregateJoinProvenance [(Text, FieldOrigin)]

ArrayJoinProvenance [Text]

MultiplexProvenance

data JoinSource

Constructors

JoinSelect Select

newtype Where

Constructors

Where [Expression]

data Cardinality

Constructors

Many

One

data AsStruct

Constructors

NoAsStruct

AsStruct

data Top

Constructors

NoTop

Top Int64

data Expression

Constructors

ValueExpression Value

InExpression Expression Value

AndExpression [Expression]

OrExpression [Expression]

NotExpression Expression

ExistsExpression Select

SelectExpression Select

IsNullExpression Expression

IsNotNullExpression Expression

ColumnExpression FieldName

EqualExpression Expression Expression

NotEqualExpression Expression Expression

JsonQueryExpression Expression
This one acts like a "cast to JSON" and makes SQL Server behave like it knows your field is JSON and not double-encode it.

ToStringExpression Expression

JsonValueExpression Expression JsonPath
This is for getting actual atomic values out of a JSON string.

OpExpression Op Expression Expression

ListExpression [Expression]

CastExpression Expression ScalarType

FunctionExpression !Text [Expression]

ConditionalProjection Expression FieldName

data JsonPath

Constructors

RootPath

FieldPath JsonPath Text

IndexPath JsonPath Integer

data Aggregate

Constructors

CountAggregate (Countable FieldName)

OpAggregates !Text (NonEmpty (Text, Expression))

OpAggregate !Text Expression

TextAggregate !Text

data Countable fieldname

Constructors

StarCountable

NonNullFieldCountable (NonEmpty fieldname)

DistinctCountable (NonEmpty fieldname)

data From

Constructors

FromQualifiedTable (Aliased TableName)

FromSelect (Aliased Select)

FromSelectJson (Aliased SelectJson)

data SelectJson

Constructors

SelectJson

Fields

selectJsonBody :: Expression

selectJsonFields :: [(ColumnName, ScalarType)]

data OpenJson

Constructors

OpenJson

Fields

openJsonExpression :: Expression

openJsonWith :: NonEmpty JsonFieldSpec

data JsonFieldSpec

Constructors

IntField Text

JsonField Text

data Aliased a

Constructors

Aliased

Fields

aliasedThing :: !a

aliasedAlias :: !Text

newtype SchemaName

Constructors

SchemaName

Fields

schemaNameParts :: [Text]

data TableName

Constructors

TableName

Fields

tableName :: Text

tableNameSchema :: Text

data FieldName

Constructors

FieldName

Fields

fieldName :: Text

fieldNameEntity :: !Text

newtype ColumnName

Constructors

ColumnName

Fields

columnName :: Text

data Comment

Constructors

DueToPermission

RequestedSingleObject

newtype EntityAlias

Constructors

EntityAlias

Fields

entityAliasText :: Text

data Op

Constructors

LessOp

LessOrEqualOp

MoreOp

MoreOrEqualOp

InOp

NotInOp

LikeOp

NotLikeOp

data Value

Source for this represenation type:

https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.Value

BigQuery results come in via the REST API as one of these simply types.

TODO: This omits StructValue -- do we need it?

Constructors

NullValue

IntegerValue !Int64
64-bit https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#integer_type

DecimalValue !Decimal
Fixed precision https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#decimal_types

BigDecimalValue !BigDecimal
Fixed precision https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#decimal_types

FloatValue !Float64
Floating point https://cloud.google.com/bigquery/docs/reference/standard-sql/data-types#floating_point_types

GeographyValue !Geography

StringValue !Text

BytesValue !Base64

BoolValue !Bool

ArrayValue !(Vector Value)

TimestampValue !Timestamp

DateValue !Date

TimeValue !Time

DatetimeValue !Datetime

newtype Timestamp

BigQuery's conception of a timestamp.

Constructors

Timestamp Text

textToUTCTime :: Text -> Parser UTCTime

utctimeToISO8601Text :: UTCTime -> Text

newtype Date

BigQuery's conception of a date.

Constructors

Date Text

newtype Time

BigQuery's conception of a time.

Constructors

Time Text

newtype Datetime

BigQuery's conception of a datetime.

Constructors

Datetime Text

newtype Int64

BigQuery's conception of an INTEGER/INT64 (they are the same).

Constructors

Int64 Text

intToInt64 :: Int64 -> Int64

int64Expr :: Int64 -> Expression

newtype Decimal

BigQuery's conception of a fixed precision decimal.

Constructors

Decimal Text

doubleToDecimal :: Double -> Decimal

newtype BigDecimal

BigQuery's conception of a "big" fixed precision decimal.

Constructors

BigDecimal Text

doubleToBigDecimal :: Double -> BigDecimal

newtype Float64

BigQuery's conception of a fixed precision decimal.

Constructors

Float64 Text

doubleToFloat64 :: Double -> Float64

newtype Base64

A base-64 encoded binary string.

Constructors

Base64

Fields

unBase64 :: ByteString

newtype Geography

Constructors

Geography

Fields

unGeography :: Text

data ScalarType

Constructors

StringScalarType

BytesScalarType

IntegerScalarType

FloatScalarType

BoolScalarType

TimestampScalarType

DateScalarType

TimeScalarType

DatetimeScalarType

GeographyScalarType

DecimalScalarType

BigDecimalScalarType

StructScalarType

data UnifiedMetadata

Constructors

UnifiedMetadata

Fields

tables :: ![UnifiedTableMetadata]

data UnifiedTableMetadata

Constructors

UnifiedTableMetadata

Fields

table :: !UnifiedTableName

object_relationships :: ![UnifiedObjectRelationship]

array_relationships :: ![UnifiedArrayRelationship]

columns :: ![UnifiedColumn]

data UnifiedColumn

Constructors

UnifiedColumn

Fields

name :: !Text

type' :: !ScalarType

data UnifiedTableName

Constructors

UnifiedTableName

Fields

schema :: !Text

name :: !Text

data UnifiedObjectRelationship

Constructors

UnifiedObjectRelationship

Fields

using :: !UnifiedUsing

name :: !Text

data UnifiedArrayRelationship

Constructors

UnifiedArrayRelationship

Fields

using :: !UnifiedUsing

name :: !Text

data UnifiedUsing

Constructors

UnifiedUsing

Fields

foreign_key_constraint_on :: !UnifiedOn

data UnifiedOn

Constructors

UnifiedOn

Fields

table :: !UnifiedTableName

column :: !Text

newtype FunctionName

Constructors

FunctionName Text

data BooleanOperators a

Constructors

ASTContains !a

ASTEquals !a

ASTTouches !a

ASTWithin !a

ASTIntersects !a

ASTDWithin !(DWithinGeogOp a)

scalarTypeGraphQLName :: ScalarType -> Either QErr Name

parseScalarValue :: ScalarType -> Value -> Either QErr Value

isComparableType :: ScalarType -> Bool

isNumType :: ScalarType -> Bool

getGQLTableName :: TableName -> Either QErr Name

liberalIntegralPrinter :: Coercible Text a => a -> Value

liberalDecimalPrinter :: Coercible a Text => a -> Value

liberalInt64Parser :: (Text -> a) -> Value -> Parser a

Parse from text by simply validating it contains digits; otherwise, require a JSON integer.

liberalDecimalParser :: (Text -> a) -> Value -> Parser a

Parse either a JSON native double number, or a text string containing something vaguely in scientific notation. In either case, producing a wrapped Text as the final result.

projectionAlias :: Projection -> Maybe Text