Hasura.Backends.DataConnector.Adapter.Types

Documentation

data ConnSourceConfig

Constructors

ConnSourceConfig

Fields

value :: Config

An arbitrary JSON payload to be passed to the agent in a header. HGE validates this against the OpenAPI Spec provided by the agent.

template :: Maybe Text

Kriti Template for transforming the supplied Config value.

timeout :: Maybe SourceTimeout

Timeout setting for HTTP requests to the agent. -- TODO: verify with lyndon

data SourceTimeout

Constructors

SourceTimeoutSeconds Int

SourceTimeoutMilliseconds Int

SourceTimeoutMicroseconds Int

sourceTimeoutMicroseconds :: SourceTimeout -> Int

data SourceConfig

Constructors

SourceConfig

Fields

_scEndpoint :: BaseUrl

_scConfig :: Config

_scTemplate :: Maybe Text

_scCapabilities :: Capabilities

_scSchema :: SchemaResponse

_scManager :: Manager

_scTimeoutMicroseconds :: Maybe Int

_scDataConnectorName :: DataConnectorName

newtype DataConnectorName

Constructors

DataConnectorName

Fields

unDataConnectorName :: NonEmptyText

data DataConnectorOptions

Constructors

DataConnectorOptions

Fields

_dcoUri :: BaseUrl

data DataConnectorInfo

Constructors

DataConnectorInfo

Fields

_dciOptions :: DataConnectorOptions

_dciCapabilities :: Capabilities

_dciConfigSchemaResponse :: ConfigSchemaResponse

newtype TableName

The fully qualified name of a table. The last element in the list is the table name and all other elements represent namespacing of the table name. For example, for a database that has schemas, the name would be '[schema,name]'

Constructors

TableName

Fields

unTableName :: NonEmpty Text

newtype ConstraintName

Constructors

ConstraintName

Fields

unConstraintName :: Text

newtype ColumnName

Constructors

ColumnName

Fields

unColumnName :: Text

newtype FunctionName

Constructors

FunctionName

Fields

unFunctionName :: NonEmpty Text

data CountAggregate

Constructors

StarCount

ColumnCount ColumnName

ColumnDistinctCount ColumnName

data Literal

Constructors

ValueLiteral Value

ArrayLiteral [Value]

data OrderDirection

Constructors

Ascending

Descending

data ScalarType

Constructors

StringTy

NumberTy

BoolTy

CustomTy Text

fromGQLType :: Name -> ScalarType

scTimeoutMicroseconds :: Lens' SourceConfig (Maybe Int)

scTemplate :: Lens' SourceConfig (Maybe Text)

scSchema :: Lens' SourceConfig SchemaResponse

scManager :: Lens' SourceConfig Manager

scEndpoint :: Lens' SourceConfig BaseUrl

scDataConnectorName :: Lens' SourceConfig DataConnectorName

scConfig :: Lens' SourceConfig Config

scCapabilities :: Lens' SourceConfig Capabilities