Hasura.Backends.DataConnector.Adapter.Types

Documentation

data ConnSourceConfig

Constructors

ConnSourceConfig

Fields

_cscValue :: Config

An arbitrary JSON payload to be passed to the agent in a header. HGE validates this against the OpenAPI Spec provided by the agent.

_cscTemplate :: Maybe Text

Kriti Template for transforming the supplied Config value.

_cscTemplateVariables :: Maybe (HashMap TemplateVariableName TemplateVariableSource)

Definitions of variables that can be accessed in the Kriti template via $vars

_cscTimeout :: Maybe SourceTimeout

Timeout setting for HTTP requests to the agent. -- TODO: verify with lyndon

newtype TemplateVariableName

Constructors

TemplateVariableName

Fields

unTemplateVariableName :: Text

data TemplateVariableSource

Constructors

TemplateVariableDynamicFromFile FilePath

data SourceTimeout

sourceTimeoutMicroseconds :: SourceTimeout -> Int

data SourceConfig

Constructors

SourceConfig

Fields

_scEndpoint :: BaseUrl

_scConfig :: Config

_scTemplate :: Maybe Text

_scTemplateVariables :: HashMap TemplateVariableName TemplateVariableSource

_scCapabilities :: Capabilities

_scManager :: Manager

_scTimeoutMicroseconds :: Maybe Int

_scDataConnectorName :: DataConnectorName

_scEnvironment :: Environment

scCapabilities :: Lens' SourceConfig Capabilities

scConfig :: Lens' SourceConfig Config

scDataConnectorName :: Lens' SourceConfig DataConnectorName

scEndpoint :: Lens' SourceConfig BaseUrl

scManager :: Lens' SourceConfig Manager

scTemplate :: Lens' SourceConfig (Maybe Text)

scTemplateVariables :: Lens' SourceConfig (HashMap TemplateVariableName TemplateVariableSource)

scTimeoutMicroseconds :: Lens' SourceConfig (Maybe Int)

scEnvironment :: Lens' SourceConfig Environment

data DataConnectorOptions

Constructors

DataConnectorOptions

Fields

_dcoUri :: BaseUrl

_dcoDisplayName :: Maybe Text

data DataConnectorInfo

Constructors

DataConnectorInfo

Fields

_dciOptions :: DataConnectorOptions

_dciCapabilities :: Capabilities

_dciConfigSchemaResponse :: ConfigSchemaResponse

_dciDisplayName :: Maybe Text

_dciReleaseName :: Maybe Text

newtype TableName

The fully qualified name of a table. The last element in the list is the table name and all other elements represent namespacing of the table name. For example, for a database that has schemas, the name would be '[schema,name]'

Constructors

TableName

Fields

unTableName :: NonEmpty Text

newtype ConstraintName

Constructors

ConstraintName

Fields

unConstraintName :: Text

newtype ColumnName

Constructors

ColumnName

Fields

unColumnName :: Text

data ColumnPath

Constructors

CPPath (NonEmpty ColumnName)

CPColumn ColumnName

newtype FunctionName

Constructors

FunctionName

Fields

unFunctionName :: NonEmpty Text

data FunctionReturnType

This represents what information can be known about the return type of a user-defined function. For now, either the return type will be the name of a table that exists in the schema, or Unknown - implying that this information can be derived from another source, or if there is no other source, then it is an error. In future, this type may be extended with additional constructors including scalar and row types from the Logical Models feature.

Note: This is very similar to ComputedFieldReturnType defined above. The two types may be unified in future.

Constructors

FunctionReturnsTable TableName

FunctionReturnsUnknown

data CountAggregate v

Constructors

StarCount

ColumnCount (ColumnName, AnnRedactionExp 'DataConnector v)

ColumnDistinctCount (ColumnName, AnnRedactionExp 'DataConnector v)

data Literal

Constructors

ValueLiteral ScalarType Value

ArrayLiteral ScalarType [Value]

data OrderDirection

Constructors

Ascending

Descending

data GraphQLType

Constructors

GraphQLInt

GraphQLFloat

GraphQLString

GraphQLBoolean

GraphQLID

newtype ScalarType

Constructors

ScalarType

Fields

unScalarType :: Text

data ArgumentExp a

Constructors

AEInput a
Table row accessor AETableRow | -- | Hardcoded reference to hdb_catalog.hdb_action_log.response_payload AEActionResponsePayload | -- | JSON/JSONB hasura session variable object AESession a |

fromGQLType :: Name -> ScalarType

data ExtraTableMetadata

This type captures backend-specific "extra" information about tables and is used on types like DBTableMetadata

Constructors

ExtraTableMetadata

Fields

_etmTableType :: TableType

_etmExtraColumnMetadata :: HashMap ColumnName ExtraColumnMetadata

data ExtraColumnMetadata

Constructors

ExtraColumnMetadata

Fields

_ecmValueGenerated :: Maybe ColumnValueGenerationStrategy

module Hasura.RQL.Types.DataConnector