Hasura.Backends.DataConnector.Adapter.Types

Documentation

data ConnSourceConfig

Constructors

ConnSourceConfig

Fields

value :: Config

An arbitrary JSON payload to be passed to the agent in a header. HGE validates this against the OpenAPI Spec provided by the agent.

template :: Maybe Text

Kriti Template for transforming the supplied Config value.

timeout :: Maybe SourceTimeout

Timeout setting for HTTP requests to the agent. -- TODO: verify with lyndon

data SourceTimeout

sourceTimeoutMicroseconds :: SourceTimeout -> Int

data SourceConfig

Constructors

SourceConfig

Fields

_scEndpoint :: BaseUrl

_scConfig :: Config

_scTemplate :: Maybe Text

_scCapabilities :: Capabilities

_scManager :: Manager

_scTimeoutMicroseconds :: Maybe Int

_scDataConnectorName :: DataConnectorName

_scEnvironment :: Environment

data DataConnectorOptions

Constructors

DataConnectorOptions

Fields

_dcoUri :: BaseUrl

_dcoDisplayName :: Maybe Text

data DataConnectorInfo

Constructors

DataConnectorInfo

Fields

_dciOptions :: DataConnectorOptions

_dciCapabilities :: Capabilities

_dciConfigSchemaResponse :: ConfigSchemaResponse

_dciDisplayName :: Maybe Text

_dciReleaseName :: Maybe Text

newtype TableName

The fully qualified name of a table. The last element in the list is the table name and all other elements represent namespacing of the table name. For example, for a database that has schemas, the name would be '[schema,name]'

Constructors

TableName

Fields

unTableName :: NonEmpty Text

newtype ConstraintName

Constructors

ConstraintName

Fields

unConstraintName :: Text

newtype ColumnName

Constructors

ColumnName

Fields

unColumnName :: Text

newtype FunctionName

Constructors

FunctionName

Fields

unFunctionName :: NonEmpty Text

data CountAggregate

Constructors

StarCount

ColumnCount ColumnName

ColumnDistinctCount ColumnName

data Literal

Constructors

ValueLiteral ScalarType Value

ArrayLiteral ScalarType [Value]

data OrderDirection

Constructors

Ascending

Descending

data ScalarType

Constructors

ScalarType Text (Maybe GraphQLType)

mkScalarType :: Capabilities -> ScalarType -> ScalarType

fromGQLType :: Name -> ScalarType

data ExtraTableMetadata

This type captures backend-specific "extra" information about tables and is used on types like DBTableMetadata

Constructors

ExtraTableMetadata

Fields

_etmTableType :: TableType

_etmExtraColumnMetadata :: HashMap ColumnName ExtraColumnMetadata

data ExtraColumnMetadata

Constructors

ExtraColumnMetadata

Fields

_ecmValueGenerated :: Maybe ColumnValueGenerationStrategy

scTimeoutMicroseconds :: Lens' SourceConfig (Maybe Int)

scTemplate :: Lens' SourceConfig (Maybe Text)

scManager :: Lens' SourceConfig Manager

scEnvironment :: Lens' SourceConfig Environment

scEndpoint :: Lens' SourceConfig BaseUrl

scDataConnectorName :: Lens' SourceConfig DataConnectorName

scConfig :: Lens' SourceConfig Config

scCapabilities :: Lens' SourceConfig Capabilities