Hasura.Backends.DataConnector.Plan.Common

Description

This module contains Data Connector request/response planning code and utility functions and types that are common across the different categories of requests (ie queries, mutations, etc). It contains code and concepts that are independent of these different categories.

Both QueryPlan and MutationPlan use the contents of this module, for example mkQueryPlan.

Documentation

data Plan request response

Represents a request to be sent to a data connector agent (_pRequest) and a function that is capable of reshaping the response to that request into the final JSON form expected to be returned by the GraphQL endpoint (_pResponseReshaper).

Constructors

Plan

Fields

_pRequest :: request

_pResponseReshaper :: forall m. MonadError QErr m => response -> m Encoding

newtype TableRelationships

A monoidal data structure used to record Table Relationships encountered during request translation. Used with recordTableRelationship.

Constructors

TableRelationships

Fields

unTableRelationships :: HashMap TableName (HashMap RelationshipName Relationship)

recordTableRelationship :: (Has TableRelationships writerOutput, Monoid writerOutput, MonadError QErr m) => TableName -> RelationshipName -> Relationship -> WriterT writerOutput m ()

Records a table relationship encountered during request translation into the output of the current WriterT

recordTableRelationshipFromRelInfo :: (Has TableRelationships writerOutput, Monoid writerOutput, MonadError QErr m) => TableName -> RelInfo 'DataConnector -> WriterT writerOutput m (RelationshipName, Relationship)

data FieldPrefix

Represents a potential prefix that can be applied to a field name, useful for namespacing field names that may be otherwise duplicated.

noPrefix :: FieldPrefix

prefixWith :: FieldName -> FieldPrefix

applyPrefix :: FieldPrefix -> FieldName -> FieldName

data Cardinality

Constructors

Single

Many

prepareLiteral :: MonadError QErr m => SessionVariables -> UnpreparedValue 'DataConnector -> m Literal

translateBoolExpToExpression :: (Has TableRelationships writerOutput, Monoid writerOutput, MonadError QErr m) => SessionVariables -> TableName -> AnnBoolExp 'DataConnector (UnpreparedValue 'DataConnector) -> WriterT writerOutput m (Maybe Expression)

mkRelationshipName :: RelName -> RelationshipName

mapFieldNameHashMap :: HashMap FieldName v -> HashMap FieldName v

encodeAssocListAsObject :: [(Text, Encoding)] -> Encoding