Hasura.Backends.DataConnector.Plan.Common

Description

This module contains Data Connector request/response planning code and utility functions and types that are common across the different categories of requests (ie queries, mutations, etc). It contains code and concepts that are independent of these different categories.

Both QueryPlan and MutationPlan use the contents of this module, for example mkQueryPlan.

Documentation

data Plan request response

Represents a request to be sent to a data connector agent (_pRequest) and a function that is capable of reshaping the response to that request into the final JSON form expected to be returned by the GraphQL endpoint (_pResponseReshaper).

Constructors

Plan

Fields

_pRequest :: request

_pResponseReshaper :: forall m. MonadError QErr m => response -> m Encoding

writeOutput :: (Semigroup output, MonadState state m, Has output state) => output -> m ()

Writes some output to state, like one might do if one was using a Writer monad. The output is combined with the existing output using <> from Semigroup

replaceOutput :: (MonadState state m, Has output state) => (output -> (output, a)) -> m a

Replaces some output in the state with a new version of the output. Also, a value can be returned from the replacement function.

This is useful if you need to inspect the existing state, make a decision, and update it based on that decision. The result of the decision can be returned from the transformation as your a value.

newtype TableRelationships

A monoidal data structure used to record Table Relationships encountered during request translation. Used with recordTableRelationship.

Constructors

TableRelationships

Fields

unTableRelationships :: HashMap TargetName (HashMap RelationshipName Relationship)

recordRelationship :: (MonadState state m, Has TableRelationships state) => TargetName -> RelationshipName -> Relationship -> m ()

Records a relationship encountered during request translation into the output of the current WriterT

recordTableRelationshipFromRelInfo :: (MonadState state m, Has TableRelationships state) => TargetName -> RelInfo 'DataConnector -> m (RelationshipName, Relationship)

recordNativeQuery :: (Has InterpolatedQueries state, Has ScalarTypesCapabilities r, MonadReader r m, MonadState state m, MonadError QErr m, Has SessionVariables r) => NativeQuery 'DataConnector (UnpreparedValue 'DataConnector) -> m InterpolatedQueryId

Records a Native Query encountered during request translation into the output of the current WriterT

newtype RedactionExpressionState

Collects encountered redaction expressions on a per table/function basis. Expressions are deduplicated and assigned a unique name (within that table/function) that is then used to reference the expression inside the query.

Constructors

RedactionExpressionState

Fields

unRedactionExpressionState :: HashMap TargetName (HashMap RedactionExpression RedactionExpressionName)

recordRedactionExpression :: (MonadState state m, Has TableRelationships state, Has RedactionExpressionState state, Has InterpolatedQueries state, MonadError QErr m, MonadReader r m, Has ScalarTypesCapabilities r, Has SessionVariables r, MonadIO m) => TargetName -> AnnRedactionExp 'DataConnector (UnpreparedValue 'DataConnector) -> m (Maybe RedactionExpressionName)

translateRedactionExpressions :: RedactionExpressionState -> Set TargetRedactionExpressions

data FieldPrefix

Represents a potential prefix that can be applied to a field name, useful for namespacing field names that may be otherwise duplicated.

noPrefix :: FieldPrefix

prefixWith :: FieldName -> FieldPrefix

applyPrefix :: FieldPrefix -> FieldName -> FieldName

data Cardinality

Constructors

Single

Many

prepareLiteral :: (MonadError QErr m, MonadReader r m, Has ScalarTypesCapabilities r, Has SessionVariables r) => UnpreparedValue 'DataConnector -> m Literal

data ColumnStack

emptyColumnStack :: ColumnStack

pushColumn :: ColumnStack -> ColumnName -> ColumnStack

toColumnSelector :: ColumnStack -> ColumnName -> ColumnSelector

translateBoolExpToExpression :: (MonadState state m, Has TableRelationships state, Has InterpolatedQueries state, Has RedactionExpressionState state, MonadError QErr m, MonadReader r m, Has ScalarTypesCapabilities r, Has SessionVariables r, MonadIO m) => TargetName -> AnnBoolExp 'DataConnector (UnpreparedValue 'DataConnector) -> m (Maybe Expression)

targetToTargetName :: Target -> TargetName

Helper function to convert targets into Keys

mkRelationshipName :: RelName -> RelationshipName

mapFieldNameHashMap :: HashMap FieldName v -> HashMap FieldName v

encodeAssocListAsObject :: [(Text, Encoding)] -> Encoding