Hasura.Backends.DataConnector.Plan.RemoteRelationshipPlan

Documentation

mkRemoteRelationshipPlan

Arguments

:: forall m r. (MonadError QErr m, MonadReader r m, Has ScalarTypesCapabilities r)

=> SessionVariables

-> SourceConfig

-> NonEmpty Object
List of join json objects, each of which contains IDs to be laterally-joined against as well as an argument ID that identifies the particular set of IDs (ie row to join against). This the response to the query constructed by this function is expected to contain a list of result objects, one per item in this list. Each object looks like: {"argument_id": 1, SomeIdColumn: 123}

-> HashMap FieldName (ColumnName, ScalarType)
Schema of the IDs in the join json objects, excluding the argument ID

-> FieldName
The name of the argument ID property in the join json object. This property name and associated value must be in the result object

-> FieldName
The field name the result of the join should be stored in the result object

-> SourceRelationshipSelection 'DataConnector Void UnpreparedValue

-> m (Plan QueryRequest QueryResponse)