Hasura.Backends.MSSQL.Connection

Description

MSSQL Connection

This module handles the connection against an MS SQL Server. It defines the connection string, connection pool, default settings, and conversion functions between MSSQL and graphql-engine.

Documentation

data MSSQLConnConfiguration

Constructors

MSSQLConnConfiguration MSSQLConnectionInfo (Maybe (NonEmpty MSSQLConnectionInfo))

data MSSQLSourceConfig

Constructors

MSSQLSourceConfig ConnectionString MSSQLExecCtx Int

data MSSQLConnectionInfo

Constructors

MSSQLConnectionInfo

Fields

mciConnectionString :: InputConnectionString

mciPoolSettings :: MSSQLPoolSettings

mciIsolationLevel :: TxIsolation

data MSSQLPoolSettings

Constructors

MSSQLPoolSettingsPool MSSQLPoolConnectionSettings

MSSQLPoolSettingsNoPool

data MSSQLPoolConnectionSettings

Constructors

MSSQLPoolConnectionSettings

Fields

mpsMaxConnections :: Maybe Int

mpsTotalMaxConnections :: Maybe Int

mpsIdleTimeout :: Int

data MSSQLExecCtx

Execution Context required to execute MSSQL transactions

Constructors

MSSQLExecCtx

Fields

mssqlRunReadOnly :: MSSQLRunTx

A function that runs read-only queries

mssqlRunReadWrite :: MSSQLRunTx

A function that runs read-write queries; run in a transaction

mssqlRunSerializableTx :: MSSQLRunTx

A function that runs a transaction in the SERIALIZABLE transaction isolation level. This is mainly intended to run source catalog migrations.

mssqlDestroyConn :: IO ()

Destroys connection pools

mssqlResizePools :: ServerReplicas -> IO SourceResizePoolSummary

Resize pools based on number of server instances

class MonadError QErr m => MonadMSSQLTx (m :: Type -> Type) where

Methods

liftMSSQLTx :: TxE QErr a -> m a

defaultMSSQLMaxConnections :: Int

createMSSQLPool :: (MonadIO m, QErrM m) => InputConnectionString -> ConnectionOptions -> Environment -> m (ConnectionString, MSSQLPool)

resizeMSSQLPool :: MSSQLPool -> Int -> ServerReplicas -> IO ()

Resize MSSQL pool by setting the number of connections equal to allowed maximum connections across all server instances divided by number of instances

getEnv :: QErrM m => Environment -> Text -> m Text

odbcValueToJValue :: Value -> Value

mkMSSQLExecCtx :: TxIsolation -> MSSQLPool -> ResizePoolStrategy -> MSSQLExecCtx

Creates a MSSQL execution context for a single primary pool

mkMSSQLAnyQueryTx :: Query -> TxET QErr IO ()

Run any query discarding its results

runMSSQLSourceReadTx :: (MonadIO m, MonadBaseControl IO m) => MSSQLSourceConfig -> TxET QErr m a -> m (Either QErr a)

runMSSQLSourceWriteTx :: (MonadIO m, MonadBaseControl IO m) => MSSQLSourceConfig -> TxET QErr m a -> m (Either QErr a)

Orphan instances

NFData ConnectionString

Instance details

Methods

rnf :: ConnectionString -> ()

Hashable ConnectionString

Instance details

Methods

hashWithSalt :: Int -> ConnectionString -> Int

hash :: ConnectionString -> Int