Hasura.Backends.MSSQL.FromIr

Description

The modules in the Hasura.Backends.MSSQL.FromIr namespace translates the RQL IR into TSQL, the SQL dialect of MSSQL, as defined in abstract syntax in Hasura.Backends.MSSQL.Types.

The translation happens in the FromIr monad, which manages identifier scoping and error collection.

The actual rendering of this AST into TSQL text happens in Hasura.Backends.MSSQL.ToQuery.

Documentation

data FromIr a

The central Monad used throughout for all conversion functions.

It has the following features:

It's a MonadValidate, so it'll continue going when it encounters Errors to accumulate as many as possible.

It has a facility for generating fresh, unique aliases, which lets the translation output retain a resemblance with source names without the translation process needing to be bothered about potential name shadowing. See generateAlias.

runFromIr :: MonadError QErr m => FromIr a -> m a

Run a FromIr action, throwing errors that have been collected using the supplied action.

data Error

Errors that may happen during translation.

Constructors

UnsupportedOpExpG (OpExpG 'MSSQL Expression)

FunctionNotSupported

data NameTemplate

Hints about the type of entity that generateAlias is producing an alias for.

Constructors

ArrayRelationTemplate Text

ArrayAggregateTemplate Text

ObjectRelationTemplate Text

TableTemplate Text

ForOrderAlias Text

generateAlias :: NameTemplate -> FromIr Text

Generate a fresh alias for a given entity to remove ambiguity and naming conflicts between scopes at the TSQL level.

Names are generated in the form type_name_occurrence, where:

type hints at the type of entity,

name refers to the source name being aliased, and

occurrence is an integer counter that distinguishes each occurrence of type_name.

Example outputs:

do
  "ar_articles_1" <- generateAlias (ArrayRelationTemplate "articles")
  "ar_articles_2" <- generateAlias (ArrayRelationTemplate "articles")
  "t_users_1"     <- generateAlias (TableTemplate "users")