Hasura.Backends.MSSQL.FromIr.Expression

Description

This module translates the IR of boolean expressions into TSQL boolean expressions.

Boolean expressions typically arise from permissions and where-clause filters.

Documentation

fromGBoolExp :: GBoolExp 'MSSQL (AnnBoolExpFld 'MSSQL Expression) -> ReaderT EntityAlias FromIr Expression

Translate boolean expressions into TSQL Expressions.

The AnnBoolExpFld references fields and columns. The entity (e.g. table) that binds these columns is supplied in the `ReaderT EntityAlias` environment, such that the columns can be referred to unambiguously.

fromAnnBoolExpFld :: AnnBoolExpFld 'MSSQL Expression -> ReaderT EntityAlias FromIr Expression

Translate boolean expressions into TSQL Expressions.

The AnnBoolExpFld references fields and columns. The entity (e.g. table) that binds these columns is supplied in the `ReaderT EntityAlias` environment, such that the columns can be referred to unambiguously.

scopedTo :: From -> ReaderT EntityAlias FromIr a -> ReaderT EntityAlias FromIr a

Scope a translation action to the table bound in a FROM clause.

fromColumnInfo :: ColumnInfo 'MSSQL -> ReaderT EntityAlias FromIr Expression

Translate a column reference occurring in a boolean expression into an equivalent Expression.

Different text types support different operators. Therefore we cast some text types to "varchar(max)", which supports the most operators.

fromColumn :: ColumnName -> ReaderT EntityAlias FromIr FieldName

Get FieldSource from a TAFExp type table aggregate field

fromOpExpG :: ColumnInfo 'MSSQL -> OpExpG 'MSSQL Expression -> ReaderT EntityAlias FromIr Expression

Translate a single OpExpG operation on a column into an expression.

nullableBoolEquality :: Expression -> Expression -> Expression

nullableBoolInequality :: Expression -> Expression -> Expression

aliasQualifiedTable :: TableName -> FromIr From