Hasura.Backends.MSSQL.FromIr.Query

Description

This module defines translation functions for queries which select data. Principally this includes translating the query root field, but parts are also reused for serving the responses for mutations.

Documentation

fromQueryRootField :: QueryDB 'MSSQL Void Expression -> FromIr Select

This is the top-level entry point for translation of Query root fields.

fromSelect :: JsonAggSelect -> AnnSelectG 'MSSQL (AnnFieldG 'MSSQL Void) Expression -> FromIr Select

fromSourceRelationship

Arguments

:: NonEmpty Object
List of json objects, each of which becomes a row of the table

-> HashMap FieldName (ColumnName, ScalarType)
The above objects have this schema

-> FieldName

-> (FieldName, SourceRelationshipSelection 'MSSQL Void (Const Expression))

-> FromIr Select

Used in planSourceRelationship, which is in turn used by to implement mkDBRemoteRelationship for BackendExecute. For more information, see the module/documentation of Source.

fromRemoteRelationFieldsG :: Map TableName EntityAlias -> HashMap ColumnName ColumnName -> (FieldName, SourceRelationshipSelection 'MSSQL Void (Const Expression)) -> ReaderT EntityAlias FromIr FieldSource

Build the FieldSource for the relation field, depending on whether it's an object, array, or aggregate relationship.

fromSelectRows :: AnnSelectG 'MSSQL (AnnFieldG 'MSSQL Void) Expression -> FromIr Select

Toproot-level Select. All descendentsub-translations are collected to produce a root TSQL.Select.

mkNodesSelect :: Args -> Where -> Expression -> Top -> From -> [(Int, (FieldName, [FieldSource]))] -> [(Int, Projection)]

mkAggregateSelect :: Args -> Where -> Expression -> From -> [(Int, (FieldName, [Projection]))] -> [(Int, Projection)]

fromSelectAggregate :: Maybe (EntityAlias, HashMap ColumnName ColumnName) -> AnnSelectG 'MSSQL (TableAggregateFieldG 'MSSQL Void) Expression -> FromIr Select

data Args

Constructors

Args

Fields

argsWhere :: Where

argsOrderBy :: Maybe (NonEmpty OrderBy)

argsJoins :: [Join]

argsTop :: Top

argsOffset :: Maybe Expression

argsDistinct :: Proxy (Maybe (NonEmpty FieldName))

argsExistingJoins :: Map TableName EntityAlias

fromSelectArgsG :: SelectArgsG 'MSSQL Expression -> ReaderT EntityAlias FromIr Args

fromQualifiedTable :: TableName -> FromIr From

fromAnnBoolExpFld :: AnnBoolExpFld 'MSSQL Expression -> ReaderT EntityAlias FromIr Expression

Translate an AnnBoolExpFld within an EntityAlias context referring to the table the AnnBoolExpFld field belongs to.

This is mutually recursive with fromGBoolExp, mirroring the mutually recursive structure between AnnBoolExpFld and 'AnnBoolExp b a' (alias of 'GBoolExp b (AnnBoolExpFld b a)').

fromColumnInfoForBoolExp :: ColumnInfo 'MSSQL -> ReaderT EntityAlias FromIr Expression

For boolean operators, various comparison operators used need special handling to ensure that SQL Server won't outright reject the comparison. See also shouldCastToVarcharMax.

shouldCastToVarcharMax :: ColumnType 'MSSQL -> Bool

There's a problem of comparing text fields with =, <, etc. that SQL Server completely refuses to do so. So one way to workaround this restriction is to automatically cast such text fields to varchar(max).

data FieldSource

Constructors

ExpressionFieldSource (Aliased Expression)

JoinFieldSource JsonCardinality (Aliased Join)

fromTableExpFieldG :: (Int, (FieldName, TableAggregateFieldG 'MSSQL Void Expression)) -> Maybe (ReaderT EntityAlias FromIr (Int, Projection))

Get FieldSource from a TAFExp type table aggregate field

fromTableAggFieldG :: (Int, (FieldName, TableAggregateFieldG 'MSSQL Void Expression)) -> Maybe (Int, (FieldName, [Projection]))

fromTableNodesFieldG :: Map TableName EntityAlias -> StringifyNumbers -> (Int, (FieldName, TableAggregateFieldG 'MSSQL Void Expression)) -> Maybe (ReaderT EntityAlias FromIr (Int, (FieldName, [FieldSource])))

fromAggregateField :: Text -> AggregateField 'MSSQL -> Projection

fromAnnFieldsG :: Map TableName EntityAlias -> StringifyNumbers -> (FieldName, AnnFieldG 'MSSQL Void Expression) -> ReaderT EntityAlias FromIr FieldSource

The main sources of fields, either constants, fields or via joins.

fromAnnColumnField :: StringifyNumbers -> AnnColumnField 'MSSQL Expression -> ReaderT EntityAlias FromIr Expression

Here is where we project a field as a column expression. If number stringification is on, then we wrap it in a ToStringExpression so that it's casted when being projected.

fromColumn :: ColumnName -> ReaderT EntityAlias FromIr FieldName

This is where a field name "foo" is resolved to a fully qualified field name [table].[foo]. The table name comes from EntityAlias in the ReaderT.

fieldSourceProjections :: FieldSource -> Projection

joinAliasToField :: JoinAlias -> FieldName

fieldSourceJoin :: FieldSource -> Maybe Join

fromObjectRelationSelectG :: Map TableName EntityAlias -> ObjectRelationSelectG 'MSSQL Void Expression -> ReaderT EntityAlias FromIr Join

lookupTableFrom :: Map TableName EntityAlias -> TableName -> FromIr (Either EntityAlias From)

fromArraySelectG :: ArraySelectG 'MSSQL Void Expression -> ReaderT EntityAlias FromIr Join

fromArrayAggregateSelectG :: AnnRelationSelectG 'MSSQL (AnnAggregateSelectG 'MSSQL Void Expression) -> ReaderT EntityAlias FromIr Join

fromArrayRelationSelectG :: ArrayRelationSelectG 'MSSQL Void Expression -> ReaderT EntityAlias FromIr Join

fromRelName :: RelName -> FromIr Text

fromMapping :: From -> HashMap ColumnName ColumnName -> ReaderT EntityAlias FromIr [Expression]

The context given by the reader is of the previous/parent "remote" table. The WHERE that we're generating goes in the child, "local" query. The From passed in as argument is the local table.

We should hope to see e.g. "post.category = category.id" for a local table of post and a remote table of category.

The left/right columns in HashMap ColumnName ColumnName corresponds to the left/right of select ... join .... Therefore left=remote, right=local in this context.

selectFromMapping :: Select -> HashMap ColumnName ColumnName -> ReaderT EntityAlias FromIr [Expression]

fromOpExpG :: Expression -> OpExpG 'MSSQL Expression -> FromIr Expression

nullableBoolEquality :: Expression -> Expression -> Expression

nullableBoolInequality :: Expression -> Expression -> Expression

fromGBoolExp :: GBoolExp 'MSSQL (AnnBoolExpFld 'MSSQL Expression) -> ReaderT EntityAlias FromIr Expression

Translate a GBoolExp of a AnnBoolExpFld, within an EntityAlias context.

It is mutually recursive with fromAnnBoolExpFld and fromGExists.

trueExpression :: Expression

safeJsonQueryExpression :: JsonCardinality -> Expression -> Expression

A version of JSON_QUERY(..) that returns a proper json literal, rather than SQL null, which does not compose properly with FOR JSON clauses.

nullExpression :: Expression

emptyArrayExpression :: Expression

jsonFieldName :: Text

aggSubselectName :: Text

existsFieldName :: Text

data UnfurledJoin

Constructors

UnfurledJoin

Fields

unfurledJoin :: Join

unfurledObjectTableAlias :: Maybe (TableName, EntityAlias)

Recorded if we joined onto an object relation.

fromAnnotatedOrderByItemG :: AnnotatedOrderByItemG 'MSSQL Expression -> WriterT (Seq UnfurledJoin) (ReaderT EntityAlias FromIr) OrderBy

unfurlAnnotatedOrderByElement :: AnnotatedOrderByElement 'MSSQL Expression -> WriterT (Seq UnfurledJoin) (ReaderT EntityAlias FromIr) (FieldName, Maybe ScalarType)

Unfurl the nested set of object relations (tell'd in the writer) that are terminated by field name (IR.AOCColumn and IR.AOCArrayAggregation).

tableNameText :: TableName -> Text

fromColumnInfo :: ColumnInfo 'MSSQL -> ReaderT EntityAlias FromIr FieldName

aggFieldName :: Text