Hasura.Backends.MSSQL.Plan

Description

MSSQL Plan

Planning T-SQL queries and subscription by translating IR to MSSQL-specific SQL query types.

Documentation

data PrepareState

Constructors

PrepareState

Fields

positionalArguments :: [ColumnValue 'MSSQL]

namedArguments :: HashMap Name (ColumnValue 'MSSQL)

sessionVariables :: HashSet SessionVariable

planQuery :: MonadError QErr m => SessionVariables -> QueryDB 'MSSQL Void (UnpreparedValue 'MSSQL) -> m (QueryWithDDL Select)

planSourceRelationship

Arguments

:: MonadError QErr m

=> SessionVariables

-> NonEmpty Object
List of json objects, each of which becomes a row of the table

-> HashMap FieldName (ColumnName, ScalarType)
The above objects have this schema

-> FieldName

-> (FieldName, SourceRelationshipSelection 'MSSQL Void UnpreparedValue)

-> m Select

For more information, see the module/documentation of Source.

planSubscription :: MonadError QErr m => InsOrdHashMap Name (QueryDB 'MSSQL Void (UnpreparedValue 'MSSQL)) -> SessionVariables -> m (Reselect, PrepareState)

prepareValueQuery :: MonadError QErr m => SessionVariables -> UnpreparedValue 'MSSQL -> m Expression

Prepare a value without any query planning; we just execute the query with the values embedded.

resultAlias :: Text

resultIdAlias :: Text

resultVarsAlias :: Text

rowAlias :: Text