Hasura.Backends.MySQL.Connection

Documentation

resolveSourceConfig :: MonadIO m => SourceName -> ConnSourceConfig -> BackendSourceKind 'MySQL -> BackendConfig 'MySQL -> Environment -> m (Either QErr SourceConfig)

resolveDatabaseMetadata :: MonadIO m => SourceConfig -> SourceTypeCustomization -> m (Either QErr (ResolvedSource 'MySQL))

postDropSourceHook :: MonadIO m => SourceConfig -> m ()

parseFieldResult :: Field -> Maybe ByteString -> Value

fieldsToAeson :: [Field] -> [[Maybe ByteString]] -> [Value]

runJSONPathQuery :: (MonadError QErr m, MonadIO m) => Pool Connection -> Query -> m Text

parseAndCollectRows :: [Field] -> [[Maybe ByteString]] -> Vector (InsOrdHashMap FieldName Value)

Used by the dataloader to produce rows of records. Those rows of records are then manipulated by the dataloader to do Haskell-side joins. Is a Vector of HashMaps the most efficient choice? A pandas-style data frame could also be more efficient, dependingly. However, this is a legible approach; efficiency improvements can be added later.

runQueryYieldingRows :: MonadIO m => Pool Connection -> Query -> m (Vector (InsOrdHashMap FieldName Value))

Run a query immediately and parse up the results into a vector.

fetchAllRows :: Result -> IO [[Maybe ByteString]]

parseTextRows :: [Field] -> [[Maybe ByteString]] -> [[Text]]

withMySQLPool :: MonadIO m => Pool Connection -> (Connection -> IO a) -> m a