Hasura.Backends.MySQL.DataLoader.Execute

Description

Execute the plan given from .Plan.

Documentation

data RecordSet

A set of records produced by the database. These are joined together. There are all sorts of optimizations possible here, from using a matrix/flat vector, unboxed sums for Value, etc. Presently we choose a naive implementation in the interest of getting other work done.

Constructors

RecordSet

Fields

origin :: !(Maybe PlannedAction)

rows :: !(Vector (InsOrdHashMap FieldName OutputValue))

wantedFields :: !(Maybe [Text])

data ExecuteReader

The read-only info. used by the Execute monad. Later, this IORef may become either atomically modified or in an STM or MVar so that jobs can be executed in parallel.

Constructors

ExecuteReader

Fields

recordSets :: IORef (InsOrdHashMap Ref RecordSet)

credentials :: !SourceConfig

data ExecuteProblem

Any problem encountered while executing the plan.

Constructors

GetJobDecodeProblem String

CreateQueryJobDecodeProblem String

JoinProblem ExecuteProblem

UnsupportedJoinBug JoinType

MissingRecordSetBug Ref

BrokenJoinInvariant [FieldName]

newtype Execute a

Execute monad; as queries are performed, the record sets are stored in the map.

Constructors

Execute

Fields

unExecute :: ReaderT ExecuteReader (ExceptT ExecuteProblem IO) a

data OutputValue

A value outputted by this execute module in a record set.

Constructors

ArrayOutputValue !(Vector OutputValue)

RecordOutputValue !(InsOrdHashMap FieldName OutputValue)

ScalarOutputValue !Value

NullOutputValue

runExecute :: MonadIO m => SourceConfig -> HeadAndTail -> Execute a -> m (Either ExecuteProblem RecordSet)

Using the config, run the execute action. Finally, resolve the head-and-tail to a record set.

execute :: Forest PlannedAction -> Execute ()

Execute the forest of actions.

executePlannedAction :: PlannedAction -> Execute ()

Execute an action, then store its result in the ref assigned to it.

fetchRecordSetForAction :: Action -> Execute RecordSet

Fetch the record set for the given action.

makeRecordSet :: Vector (InsOrdHashMap FieldName Value) -> RecordSet

Make a record set from a flat record from the DB.

saveRecordSet :: Ref -> RecordSet -> Execute ()

getRecordSet :: Ref -> Execute RecordSet

getFinalRecordSet :: HeadAndTail -> Execute RecordSet

See documentation for HeadAndTail.

leftObjectJoin :: Maybe [Text] -> Text -> [(FieldName, FieldName)] -> RecordSet -> RecordSet -> Either ExecuteProblem RecordSet

Inefficient but clean left object join.

leftArrayJoin :: Maybe [Text] -> Text -> [(FieldName, FieldName)] -> Top -> Maybe Int -> RecordSet -> RecordSet -> Either ExecuteProblem RecordSet

A naive, exponential reference implementation of a left join. It serves as a trivial sample implementation for correctness checking of more efficient ones.

joinArrayRows :: Maybe [Text] -> Text -> InsOrdHashMap FieldName OutputValue -> Vector (InsOrdHashMap FieldName OutputValue) -> InsOrdHashMap FieldName OutputValue

Join a row with another as an array join.

joinObjectRows :: Maybe [Text] -> Text -> InsOrdHashMap FieldName OutputValue -> Vector (InsOrdHashMap FieldName OutputValue) -> Either ExecuteProblem (InsOrdHashMap FieldName OutputValue)

Join a row with another as an object join.

If rightRow is not a single row, we throw BrokenJoinInvariant.