Hasura.Backends.MySQL.FromIr

Description

Translate from the DML to the MySQL dialect.

Documentation

data FieldSource

Constructors

ExpressionFieldSource (Aliased Expression)

JoinFieldSource (Aliased Join)

AggregateFieldSource [Aliased Aggregate]

data Error

Most of these errors should be checked for legitimacy.

Constructors

UnsupportedOpExpG (OpExpG 'MySQL Expression)

IdentifierNotSupported

FunctionNotSupported

NodesUnsupportedForNow

ConnectionsNotSupported

newtype FromIr a

The base monad used throughout this module for all conversion functions.

It's a Validate, so it'll continue going when it encounters errors to accumulate as many as possible.

It also contains a mapping from entity prefixes to counters. So if my prefix is "table" then there'll be a counter that lets me generate table1, table2, etc. Same for any other prefix needed (e.g. names for joins).

A ReaderT is used around this in most of the module too, for setting the current entity that a given field name refers to. See fromColumn.

Constructors

FromIr

Fields

unFromIr :: StateT (Map Text Int) (Validate (NonEmpty Error)) a

runFromIr :: FromIr a -> Validate (NonEmpty Error) a

data NameTemplate

Constructors

ArrayRelationTemplate Text

ArrayAggregateTemplate Text

ObjectRelationTemplate Text

TableTemplate Text

ForOrderAlias Text

IndexTemplate

generateEntityAlias :: NameTemplate -> FromIr Text

fromQualifiedTable :: TableName -> FromIr From

This is really the start where you query the base table, everything else is joins attached to it.

fromAlias :: From -> EntityAlias

trueExpression :: Expression

existsFieldName :: Text

fromGExists :: GExists 'MySQL Expression -> ReaderT EntityAlias FromIr Select

fromGBoolExp :: GBoolExp 'MySQL Expression -> ReaderT EntityAlias FromIr Expression

fromAnnBoolExp :: GBoolExp 'MySQL (AnnBoolExpFld 'MySQL Expression) -> ReaderT EntityAlias FromIr Expression

fromColumnInfoForBoolExp :: ColumnInfo 'MySQL -> ReaderT EntityAlias FromIr Expression

For boolean operators, various comparison operators used need special handling to ensure that SQL Server won't outright reject the comparison. See also shouldCastToVarcharMax.

fromAnnBoolExpFld :: AnnBoolExpFld 'MySQL Expression -> ReaderT EntityAlias FromIr Expression

fromMapping :: From -> HashMap Column Column -> ReaderT EntityAlias FromIr [Expression]

The context given by the reader is of the previous/parent "remote" table. The WHERE that we're generating goes in the child, "local" query. The From passed in as argument is the local table.

We should hope to see e.g. "post.category = category.id" for a local table of post and a remote table of category.

The left/right columns in HashMap Column Column corresponds to the left/right of select ... join .... Therefore left=remote, right=local in this context.

fromColumn :: Column -> ReaderT EntityAlias FromIr FieldName

columnNameToFieldName :: Column -> EntityAlias -> FieldName

fromOpExpG :: Expression -> OpExpG 'MySQL Expression -> FromIr Expression

data Args

Constructors

Args

Fields

argsWhere :: Where

argsOrderBy :: Maybe (NonEmpty OrderBy)

argsJoins :: [Join]

argsTop :: Top

argsOffset :: Maybe Int

argsDistinct :: Proxy (Maybe (NonEmpty FieldName))

argsExistingJoins :: Map TableName EntityAlias

data UnfurledJoin

Constructors

UnfurledJoin

Fields

unfurledJoin :: Join

unfurledObjectTableAlias :: Maybe (TableName, EntityAlias)

Recorded if we joined onto an object relation.

fromColumnInfo :: ColumnInfo 'MySQL -> ReaderT EntityAlias FromIr FieldName

tableNameText :: TableName -> Text

aggFieldName :: Text

unfurlAnnOrderByElement :: AnnotatedOrderByElement 'MySQL Expression -> WriterT (Seq UnfurledJoin) (ReaderT EntityAlias FromIr) (FieldName, Maybe ScalarType)

Unfurl the nested set of object relations (tell'd in the writer) that are terminated by field name (IR.AOCColumn and IR.AOCArrayAggregation).

fromAnnOrderByItemG :: AnnotatedOrderByItemG 'MySQL Expression -> WriterT (Seq UnfurledJoin) (ReaderT EntityAlias FromIr) OrderBy

Produce a valid ORDER BY construct, telling about any joins needed on the side.

fromSelectArgsG :: SelectArgsG 'MySQL Expression -> ReaderT EntityAlias FromIr Args

fromAnnColumnField :: StringifyNumbers -> AnnColumnField 'MySQL Expression -> ReaderT EntityAlias FromIr Expression

Here is where we project a field as a column expression. If number stringification is on, then we wrap it in a ToStringExpression so that it's casted when being projected.

fromRelName :: RelName -> FromIr Text

fromTableAggregateFieldG :: (FieldName, TableAggregateFieldG 'MySQL Void Expression) -> ReaderT EntityAlias FromIr FieldSource

fieldSourceProjections :: FieldSource -> [Projection]

fieldSourceJoin :: FieldSource -> Maybe Join

fromSelectAggregate :: Maybe (EntityAlias, HashMap Column Column) -> AnnSelectG 'MySQL (TableAggregateFieldG 'MySQL Void) Expression -> FromIr Select

fromArrayAggregateSelectG :: AnnRelationSelectG 'MySQL (AnnAggregateSelectG 'MySQL Void Expression) -> ReaderT EntityAlias FromIr Join

fromArraySelectG :: ArraySelectG 'MySQL Void Expression -> ReaderT EntityAlias FromIr Join

fromObjectRelationSelectG :: ObjectRelationSelectG 'MySQL Void Expression -> ReaderT EntityAlias FromIr Join

isEmptyExpression :: Expression -> Bool

fromSelectRows :: AnnSelectG 'MySQL (AnnFieldG 'MySQL Void) Expression -> FromIr Select

fromArrayRelationSelectG :: ArrayRelationSelectG 'MySQL Void Expression -> ReaderT EntityAlias FromIr Join

fromAnnFieldsG :: StringifyNumbers -> (FieldName, AnnFieldG 'MySQL Void Expression) -> ReaderT EntityAlias FromIr FieldSource

The main sources of fields, either constants, fields or via joins.

mkSQLSelect :: JsonAggSelect -> AnnSelectG 'MySQL (AnnFieldG 'MySQL Void) Expression -> FromIr Select

fromRootField :: QueryDB 'MySQL Void Expression -> FromIr Select

Convert from the IR database query into a select.

fromMappingFieldNames :: EntityAlias -> HashMap Column Column -> ReaderT EntityAlias FromIr [(FieldName, FieldName)]

fieldTextNames :: AnnFieldsG 'MySQL Void Expression -> [Text]