Hasura.Backends.Postgres.Connection.MonadTx

Description

Postgres Connection MonadTx

This module contains MonadTx and related combinators.

MonadTx, a class which abstracts the QErr in TxE via MonadError.

The combinators are used for running, tracing, or otherwise perform database related tasks. Please consult the individual documentation for more information.

Documentation

class MonadError QErr m => MonadTx (m :: Type -> Type) where

Methods

liftTx :: TxE QErr a -> m a

runTxWithCtx :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadTrace m, UserInfoM m) => PGExecCtx -> PGExecTxType -> PGExecFrom -> TxET QErr m a -> m a

runTxWithCtxAndUserInfo :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadTrace m) => UserInfo -> PGExecCtx -> PGExecTxType -> PGExecFrom -> TxET QErr m a -> m a

runQueryTx :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m) => PGExecCtx -> PGExecFrom -> TxET QErr m a -> m a

This runs the given set of statements (Tx) without wrapping them in BEGIN and COMMIT. This should only be used for running a single statement query!

withUserInfo :: forall (m :: Type -> Type) a. MonadIO m => UserInfo -> TxET QErr m a -> TxET QErr m a

withTraceContext :: forall (m :: Type -> Type) a. MonadIO m => Maybe TraceContext -> TxET QErr m a -> TxET QErr m a

Inject the trace context as a transaction-local variable, so that it can be picked up by any triggers (including event triggers).

setHeadersTx :: forall (m :: Type -> Type). MonadIO m => SessionVariables -> TxET QErr m ()

setTraceContextInTx :: forall (m :: Type -> Type). MonadIO m => Maybe TraceContext -> TxET QErr m ()

sessionInfoJsonExp :: SessionVariables -> SQLExp

checkDbConnection :: MonadTx m => m ()

doesSchemaExist :: MonadTx m => SchemaName -> m Bool

doesTableExist :: MonadTx m => SchemaName -> TableName -> m Bool

enablePgcryptoExtension :: MonadTx m => ExtensionsSchema -> m ()

dropHdbCatalogSchema :: MonadTx m => m ()

newtype ExtensionsSchema

The name of the schema in which the graphql-engine will install database extensions.

Constructors

ExtensionsSchema

Fields

_unExtensionsSchema :: Text

Orphan instances

MonadTrace m => MonadTrace (TxET e m)

Instance details

Methods

newTraceWith :: TraceContext -> SamplingPolicy -> Text -> TxET e m a -> TxET e m a

newSpanWith :: SpanId -> Text -> TxET e m a -> TxET e m a

attachMetadata :: TraceMetadata -> TxET e m ()

MonadTraceContext m => MonadTraceContext (TxET e m)

Instance details

Methods

currentContext :: TxET e m (Maybe TraceContext)