Hasura.Backends.Postgres.Connection.Settings

Description

Postgres Connection Settings

This module contains types and combinators related to postgres connection, pool, and replica related settings.

Documentation

data PostgresPoolSettings

Constructors

PostgresPoolSettings

Fields

ppsMaxConnections :: Maybe Int

ppsTotalMaxConnections :: Maybe Int

ppsIdleTimeout :: Maybe Int

ppsRetries :: Maybe Int

ppsPoolTimeout :: Maybe NominalDiffTime

ppsConnectionLifetime :: Maybe NominalDiffTime

data PostgresSourceConnInfo

Constructors

PostgresSourceConnInfo

Fields

psciDatabaseUrl :: UrlConf

psciPoolSettings :: Maybe PostgresPoolSettings

psciUsePreparedStatements :: Bool

psciIsolationLevel :: TxIsolation

psciSslConfiguration :: Maybe (PGClientCerts CertVar CertVar)

data PostgresConnConfiguration

Constructors

PostgresConnConfiguration

Fields

pccConnectionInfo :: PostgresSourceConnInfo

pccReadReplicas :: Maybe (NonEmpty PostgresSourceConnInfo)

pccExtensionsSchema :: ExtensionsSchema

pccConnectionTemplate :: Maybe ConnectionTemplate

pccConnectionSet :: Maybe PostgresConnectionSet

data PGClientCerts p a

Constructors

PGClientCerts

Fields

pgcSslCert :: Maybe a

pgcSslKey :: Maybe a

pgcSslRootCert :: Maybe a

pgcSslMode :: SSLMode

pgcSslPassword :: Maybe p

newtype CertVar

Constructors

CertVar String

newtype CertData

Constructors

CertData

Fields

unCert :: Text

data SSLMode

Constructors

Disable

Allow

Prefer

Require

VerifyCA

VerifyFull

data DefaultPostgresPoolSettings

Constructors

DefaultPostgresPoolSettings

Fields

dppsMaxConnections :: Int

dppsIdleTimeout :: Int

dppsRetries :: Int

dppsConnectionLifetime :: Maybe NominalDiffTime

data ConnectionTemplate

Connection template for the dynamic DB connection.

Constructors

ConnectionTemplate

Fields

ctVersion :: Int

Version for the connection template. Please read more about this in the dynamic DB connection RFC (Metadata API > Versioning).

ctTemplate :: KritiTemplate

`kriti-lang` template for the dynamic DB connection.

newtype PostgresConnectionSetMemberName

Name of the member of a connection set.

Constructors

PostgresConnectionSetMemberName

Fields

getPostgresConnectionSetMemberName :: NonEmptyText

newtype PostgresConnectionSet

HashMap of the connection set. This is used for the dynamic DB connection feature.

Constructors

PostgresConnectionSet

Fields

getPostgresConnectionSet :: NEHashMap PostgresConnectionSetMemberName PostgresConnectionSetMember

data PostgresConnectionSetMember

Constructors

PostgresConnectionSetMember

Fields

pscmName :: PostgresConnectionSetMemberName

pscmConnectionInfo :: PostgresSourceConnInfo

data KritiTemplate

`kriti-lang` template.

Constructors

KritiTemplate

Fields

ktSource :: Text

Raw kriti template

ktParsedAST :: ValueExt

Parsed kriti template

getDefaultPGPoolSettingIfNotExists :: Maybe PostgresPoolSettings -> DefaultPostgresPoolSettings -> (Int, Int, Int)

defaultPostgresPoolSettings :: DefaultPostgresPoolSettings

defaultPostgresExtensionsSchema :: ExtensionsSchema

setPostgresPoolSettings :: PostgresPoolSettings

Orphan instances

FromJSON TxIsolation

Instance details

Methods

parseJSON :: Value -> Parser TxIsolation

parseJSONList :: Value -> Parser [TxIsolation]

omittedField :: Maybe TxIsolation

ToJSON TxIsolation

Instance details

Methods

toJSON :: TxIsolation -> Value

toEncoding :: TxIsolation -> Encoding

toJSONList :: [TxIsolation] -> Value

toEncodingList :: [TxIsolation] -> Encoding

omitField :: TxIsolation -> Bool

HasCodec TxIsolation

Instance details

Methods

codec :: JSONCodec TxIsolation

listCodecForStringCompatibility :: JSONCodec [TxIsolation]

NFData TxIsolation

Instance details

Methods

rnf :: TxIsolation -> ()

Generic TxIsolation

Instance details

Associated Types

type Rep TxIsolation

Instance details

Defined in Hasura.Backends.Postgres.Connection.Settings

type Rep TxIsolation = D1 ('MetaData "TxIsolation" "Database.PG.Query.Transaction" "pg-client-0.1.0-inplace" 'False) (C1 ('MetaCons "ReadCommitted" 'PrefixI 'False) (U1 :: Type -> Type) :+: (C1 ('MetaCons "RepeatableRead" 'PrefixI 'False) (U1 :: Type -> Type) :+: C1 ('MetaCons "Serializable" 'PrefixI 'False) (U1 :: Type -> Type)))

Methods

from :: TxIsolation -> Rep TxIsolation x

to :: Rep TxIsolation x -> TxIsolation

Hashable TxIsolation

Instance details

Methods

hashWithSalt :: Int -> TxIsolation -> Int

hash :: TxIsolation -> Int