Hasura.Backends.Postgres.DDL

Description

Postgres DDL

Implements the DDL related methods of the BackendMetadata type class for the Postgres backend, which provides an interface for fetching information about the objects in the database, such as tables, relationships, etc.

Documentation

parseCollectableType :: forall pgKind m. (Backend ('Postgres pgKind), MonadError QErr m) => CollectableType (ColumnType ('Postgres pgKind)) -> Value -> m (PartialSQLExp ('Postgres pgKind))