Hasura.Backends.Postgres.DDL

Description

Postgres DDL

Implements the DDL related methods of the BackendMetadata type class for the Postgres backend, which provides an interface for fetching information about the objects in the database, such as tables, relationships, etc.

Documentation

parseCollectableType :: forall (pgKind :: PostgresKind) m. (Backend ('Postgres pgKind), MonadError QErr m) => CollectableType (ColumnType ('Postgres pgKind)) -> Value -> m (PartialSQLExp ('Postgres pgKind))

module Hasura.Backends.Postgres.DDL.BoolExp

module Hasura.Backends.Postgres.DDL.ComputedField

module Hasura.Backends.Postgres.DDL.EventTrigger

module Hasura.Backends.Postgres.DDL.Function

module Hasura.Backends.Postgres.DDL.Source

module Hasura.Backends.Postgres.DDL.Table