Hasura.Backends.Postgres.Execute.ConnectionTemplate

Documentation

data PrimaryTag

This connection tag represents primary database connection

Constructors

PrimaryTag

data DefaultTag

This connection tag represents default behaviour of database connections.

For example, if read replica is set, then it will redirect GQL queries to read replicas and mutations to the primary connection

Constructors

DefaultTag

data ReadReplicasTag

This connection tag represents read replica database connection

Constructors

ReadReplicasTag

data ConnectionSetMemberTemplateContext

Data type for single member in connection_set for connection template context

Constructors

ConnectionSetMemberTemplateContext

Fields

_cseType :: ConnectionSetTemplateContextType

_cseName :: PostgresConnectionSetMemberName

mkConnectionSetMemberTemplateContext :: PostgresConnectionSetMemberName -> ConnectionSetMemberTemplateContext

data PostgresResolvedConnectionTemplate

Outcome of the connection template resolution

Constructors

PCTODefault DefaultTag

PCTOPrimary PrimaryTag

PCTOReadReplicas ReadReplicasTag

PCTOConnectionSet PostgresConnectionSetMemberName

data RequestContextHeaders

Headers information for the connection template context

Constructors

RequestContextHeaders (HashMap Text Text)

data ConnectionSetTemplateContext

Data type for connection_set for connection template context

newtype QueryOperationType

Constructors

QueryOperationType OperationType

data QueryContext

Query information (operation name and operation type) for connection template context

Constructors

QueryContext

Fields

_qcOperationName :: Maybe Name

_qcOperationType :: QueryOperationType

data RequestContext

Request information for connection template context

Constructors

RequestContext

Fields

_rcHeaders :: RequestContextHeaders

_rcSession :: SessionVariables

_rcQuery :: Maybe QueryContext

data PostgresConnectionTemplateContext

The complete connection template context used for resolving connection template

Constructors

PostgresConnectionTemplateContext

Fields

_pctcRequest :: RequestContext

_pctcPrimary :: PrimaryTag

_pctcReadReplicas :: ReadReplicasTag

_pctcDefault :: DefaultTag

_pctcConnectionSet :: ConnectionSetTemplateContext

makeConnectionTemplateContext :: RequestContext -> [PostgresConnectionSetMemberName] -> PostgresConnectionTemplateContext

Construct template context

runKritiEval :: PostgresConnectionTemplateContext -> KritiTemplate -> Either EvalError Value

We should move this to Data.Aeson.Kriti.Functions

makeRequestContext :: Maybe QueryContext -> [Header] -> SessionVariables -> RequestContext