Hasura.Backends.Postgres.Execute.Prepare

Description

Postgres Execute Prepare

Deals with translating (session) variables to SQL expressions. Uses a state monad to keep track of things like variables and generating fresh variable names.

See Execute.

Documentation

type PlanVariables = HashMap PlanVariable Int

type PrepArgMap = IntMap (PrepArg, PGScalarValue)

The value is (PG.PrepArg, PGScalarValue) because we want to log the human-readable value of the prepared argument and not the binary encoding in PG format

data PlanningSt

Constructors

PlanningSt

Fields

_psArgNumber :: Int

_psVariables :: PlanVariables

_psPrepped :: PrepArgMap

type ExecutionPlan = RootFieldMap ExecutionStep

The series of steps that need to be executed for a given query. For now, those steps are all independent. In the future, when we implement a client-side dataloader and generalized joins, this will need to be changed into an annotated tree.

data ExecutionStep where

One execution step to processing a GraphQL query (e.g. one root field).

Constructors

ExecStepDB :: ResponseHeaders -> AnyBackend DBStepInfo -> Maybe RemoteJoins -> ExecutionStep
A query to execute against the database

ExecStepAction :: ActionExecutionPlan -> ActionsInfo -> Maybe RemoteJoins -> ExecutionStep
Execute an action

ExecStepRemote :: !RemoteSchemaInfo -> !ResultCustomizer -> !GQLReqOutgoing -> Maybe RemoteJoins -> ExecutionStep
A graphql query to execute against a remote schema

ExecStepRaw :: Value -> ExecutionStep
Output a plain JSON object

ExecStepMulti :: [ExecutionStep] -> ExecutionStep

initPlanningSt :: PlanningSt

prepareWithPlan :: forall m (pgKind :: PostgresKind). (MonadState PlanningSt m, MonadError QErr m) => UserInfo -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

If we're preparing a value with planning state, we favour referring to values by their prepared argument index. If the value refers to a session value, we look for it in prepared value (1) and access the particular keys using the JSONB ->> accessor.

prepareWithoutPlan :: forall m (pgKind :: PostgresKind). MonadError QErr m => UserInfo -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

If we're not using a prepared statement, substitution is pretty naïve: we resolve session variable names, ignore parameter names, and substitute into the SQLExp.

withUserVars :: SessionVariables -> PrepArgMap -> PrepArgMap

The map of user session variables is always given the number (1) as its variable argument number (see getVarArgNum). If we want to refer to a particular variable in this map, we use JSONB functions to interrogate variable (1).