Hasura.Backends.Postgres.Execute.Prepare

Description

Postgres Execute Prepare

Deals with translating (session) variables to SQL expressions. Uses a state monad to keep track of things like variables and generating fresh variable names.

See Execute.

Documentation

type PlanVariables = HashMap Name Int

type PrepArgMap = IntMap (PrepArg, PGScalarValue)

The value is (Q.PrepArg, PGScalarValue) because we want to log the human-readable value of the prepared argument and not the binary encoding in PG format

data PlanningSt

Constructors

PlanningSt

Fields

_psArgNumber :: !Int

_psVariables :: !PlanVariables

_psPrepped :: !PrepArgMap

initPlanningSt :: PlanningSt

prepareWithPlan :: (MonadState PlanningSt m, MonadError QErr m) => UserInfo -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

prepareWithoutPlan :: MonadError QErr m => UserInfo -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

resolveUnpreparedValue :: MonadError QErr m => UserInfo -> UnpreparedValue ('Postgres pgKind) -> m SQLExp

withUserVars :: SessionVariables -> PrepArgMap -> PrepArgMap

getVarArgNum :: MonadState PlanningSt m => Name -> m Int

addPrepArg :: MonadState PlanningSt m => Int -> (PrepArg, PGScalarValue) -> m ()

getNextArgNum :: MonadState PlanningSt m => m Int