Hasura.Backends.Postgres.Execute.Types

Description

Postgres Execute Types

Execution context and source configuration for Postgres databases. Provides support for things such as read-only transactions and read replicas.

Documentation

type RunTx = forall m a. (MonadIO m, MonadBaseControl IO m) => TxET QErr m a -> ExceptT QErr m a

data PGExecCtx

Constructors

PGExecCtx

Fields

_pecRunReadOnly :: RunTx

Run a Q.ReadOnly transaction

_pecRunReadNoTx :: RunTx

Run a read only statement without an explicit transaction block

_pecRunReadWrite :: RunTx

Run a Q.ReadWrite transaction

_pecDestroyConn :: IO ()

Destroys connection pools

mkPGExecCtx :: TxIsolation -> PGPool -> PGExecCtx

Creates a Postgres execution context for a single Postgres master pool

defaultTxErrorHandler :: PGTxErr -> QErr

dmlTxErrorHandler :: PGTxErr -> QErr

Constructs a transaction error handler tailored for the needs of RQL's DML.

mkTxErrorHandler :: (PGErrorType -> Bool) -> PGTxErr -> QErr

Constructs a transaction error handler given a predicate that determines which errors are expected and should be reported to the user. All other errors are considered internal errors.

data PGSourceConfig

Constructors

PGSourceConfig

Fields

_pscExecCtx :: PGExecCtx

_pscConnInfo :: ConnInfo

_pscReadReplicaConnInfos :: Maybe (NonEmpty ConnInfo)

_pscPostDropHook :: IO ()

_pscExtensionsSchema :: ExtensionsSchema

runPgSourceReadTx :: (MonadIO m, MonadBaseControl IO m) => PGSourceConfig -> TxET QErr m a -> m (Either QErr a)

runPgSourceWriteTx :: (MonadIO m, MonadBaseControl IO m) => PGSourceConfig -> TxET QErr m a -> m (Either QErr a)