Hasura.Backends.Postgres.Execute.Types

Description

Postgres Execute Types

Execution context and source configuration for Postgres databases. Provides support for things such as read-only transactions and read replicas.

Documentation

data PGExecCtx

Constructors

PGExecCtx

Fields

_pecRunTx :: PGExecCtxInfo -> RunTx

Run a PG transaction using the information provided by PGExecCtxInfo

_pecDestroyConnections :: IO ()

Destroy connection pools

_pecResizePools :: ServerReplicas -> IO SourceResizePoolSummary

Resize pools based on number of server instances and return the summary

data PGExecFrom

The level from where the transaction is being run

Constructors

GraphQLQuery (Maybe PostgresResolvedConnectionTemplate)
transaction initated via a GraphQLRequest

RunSQLQuery
transaction initiated during run_sql

InternalRawQuery
custom transaction Hasura runs on the database. This is usually used in event_trigger and actions

LegacyRQLQuery
transactions initiated via other API's other than run_sql in v1query or v2query

data PGExecCtxInfo

Holds the information required to exceute a PG transaction

Constructors

PGExecCtxInfo

Fields

_peciTxType :: PGExecTxType

The tranasction mode for executing the transaction

_peciFrom :: PGExecFrom

The level from where the PG transaction is being executed from

data PGExecTxType

The tranasction mode (isolation level, transaction access) for executing the transaction

Constructors

NoTxRead
a transaction without an explicit tranasction block

Tx TxAccess (Maybe TxIsolation)
a transaction block with custom transaction access and isolation level. Choose defaultIsolationLevel defined in SourceConnConfiguration if Nothing is provided for isolation level.

mkPGExecCtx :: TxIsolation -> PGPool -> ResizePoolStrategy -> PGExecCtx

Creates a Postgres execution context for a single Postgres master pool

mkTxErrorHandler :: (PGErrorType -> Bool) -> PGTxErr -> QErr

Constructs a transaction error handler given a predicate that determines which errors are expected and should be reported to the user. All other errors are considered internal errors.

defaultTxErrorHandler :: PGTxErr -> QErr

dmlTxErrorHandler :: PGTxErr -> QErr

Constructs a transaction error handler tailored for the needs of RQL's DML.

resizePostgresPool :: PGPool -> Int -> ServerReplicas -> IO ()

Resize Postgres pool by setting the number of connections equal to allowed maximum connections across all server instances divided by number of instances

data PostgresResolvedConnectionTemplate

Outcome of the connection template resolution

Constructors

PCTODefault DefaultTag

PCTOPrimary PrimaryTag

PCTOReadReplicas ReadReplicasTag

PCTOConnectionSet PostgresConnectionSetMemberName

Execution in a Postgres Source

data PGSourceConfig

Constructors

PGSourceConfig

Fields

_pscExecCtx :: PGExecCtx

_pscConnInfo :: ConnInfo

_pscReadReplicaConnInfos :: Maybe (NonEmpty ConnInfo)

_pscPostDropHook :: IO ()

_pscExtensionsSchema :: ExtensionsSchema

_pscConnectionSet :: HashMap PostgresConnectionSetMemberName ConnInfo

_pscConnectionTemplateConfig :: ConnectionTemplateConfig

data ConnectionTemplateConfig

Constructors

ConnTemplate_NotApplicable
Connection templates are disabled for Hasura CE

ConnTemplate_NotConfigured

ConnTemplate_Resolver ValueExt ConnectionTemplateResolver

connectionTemplateConfigResolver :: ConnectionTemplateConfig -> Maybe ConnectionTemplateResolver

newtype ConnectionTemplateResolver

A hook to resolve connection template

Constructors

ConnectionTemplateResolver

Fields

_runResolver :: forall (m :: Type -> Type). MonadError QErr m => SessionVariables -> [Header] -> Maybe QueryContext -> m PostgresResolvedConnectionTemplate

Runs the connection template resolver.

runPgSourceReadTx :: (MonadIO m, MonadBaseControl IO m) => PGSourceConfig -> TxET QErr m a -> m (Either QErr a)

runPgSourceWriteTx :: (MonadIO m, MonadBaseControl IO m) => PGSourceConfig -> PGExecFrom -> TxET QErr m a -> m (Either QErr a)

applyConnectionTemplateResolverNonAdmin :: MonadError QErr m => Maybe ConnectionTemplateResolver -> UserInfo -> [Header] -> Maybe QueryContext -> m (Maybe PostgresResolvedConnectionTemplate)

Resolve connection templates only for non-admin roles

pgResolveConnectionTemplate :: MonadError QErr m => PGSourceConfig -> RequestContext -> Maybe ConnectionTemplate -> m EncJSON

resolvePostgresConnectionTemplate :: MonadError QErr m => ConnectionTemplate -> [PostgresConnectionSetMemberName] -> SessionVariables -> [Header] -> Maybe QueryContext -> m PostgresResolvedConnectionTemplate

sourceConfigNumReadReplicas :: PGSourceConfig -> Int

sourceConfigConnectonTemplate :: PGSourceConfig -> Maybe ValueExt

getPGColValues :: (MonadIO m, MonadError QErr m) => SessionVariables -> SourceName -> PGSourceConfig -> QualifiedTable -> (PGScalarType, PGCol) -> (PGCol, [RemoteRelSupportedOp RemoteRelSessionVariableORLiteralValue]) -> m [Text]