Hasura.Backends.Postgres.Instances.Schema

Documentation

class PostgresSchema (pgKind :: PostgresKind) where

This class is an implementation detail of BackendSchema. Some functions of BackendSchema differ across different Postgres "kinds", or call to functions (such as those related to Relay) that have not been generalized to all kinds of Postgres and still explicitly work on Vanilla Postgres. This class allows each "kind" to specify its own specific implementation. All common code is directly part of BackendSchema.

Note: Users shouldn't ever put this as a constraint. Use `BackendSchema ('Postgres pgKind)` instead.

Methods

pgkBuildTableRelayQueryFields :: MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> NESeq (ColumnInfo ('Postgres pgKind)) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

pgkBuildFunctionRelayQueryFields :: MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> FunctionName ('Postgres pgKind) -> FunctionInfo ('Postgres pgKind) -> TableName ('Postgres pgKind) -> NESeq (ColumnInfo ('Postgres pgKind)) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

pgkRelayExtension :: Maybe (XRelay ('Postgres pgKind))

pgkNode :: MonadBuildSchema ('Postgres pgKind) r m n => m (Parser 'Output n (HashMap (TableName ('Postgres pgKind)) (SourceName, SourceConfig ('Postgres pgKind), SelPermInfo ('Postgres pgKind), PrimaryKeyColumns ('Postgres pgKind), AnnotatedFields ('Postgres pgKind))))

backendInsertParser :: forall pgKind m r n. MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableInfo ('Postgres pgKind) -> Maybe (SelPermInfo ('Postgres pgKind)) -> Maybe (UpdPermInfo ('Postgres pgKind)) -> m (InputFieldsParser n (BackendInsert pgKind (UnpreparedValue ('Postgres pgKind))))

buildTableRelayQueryFields :: forall pgKind m n r. MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> NESeq (ColumnInfo ('Postgres pgKind)) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildFunctionRelayQueryFields :: forall pgKind m n r. MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> FunctionName ('Postgres pgKind) -> FunctionInfo ('Postgres pgKind) -> TableName ('Postgres pgKind) -> NESeq (ColumnInfo ('Postgres pgKind)) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

columnParser :: (MonadSchema n m, MonadError QErr m, MonadReader r m, Has MkTypename r) => ColumnType ('Postgres pgKind) -> Nullability -> m (Parser 'Both n (ValueWithOrigin (ColumnValue ('Postgres pgKind))))

jsonPathArg :: MonadParse n => ColumnType ('Postgres pgKind) -> InputFieldsParser n (Maybe (ColumnOp ('Postgres pgKind)))

orderByOperators :: NonEmpty (Definition EnumValueInfo, (BasicOrderType ('Postgres pgKind), NullsOrderType ('Postgres pgKind)))

comparisonExps :: forall pgKind m n r. (BackendSchema ('Postgres pgKind), MonadSchema n m, MonadError QErr m, MonadReader r m, Has QueryContext r, Has MkTypename r) => ColumnType ('Postgres pgKind) -> m (Parser 'Input n [ComparisonExp ('Postgres pgKind)])

geographyWithinDistanceInput :: forall pgKind m n r. (MonadSchema n m, MonadError QErr m, MonadReader r m, Has MkTypename r) => m (Parser 'Input n (DWithinGeogOp (UnpreparedValue ('Postgres pgKind))))

geometryWithinDistanceInput :: forall pgKind m n r. (MonadSchema n m, MonadError QErr m, MonadReader r m, Has MkTypename r) => m (Parser 'Input n (DWithinGeomOp (UnpreparedValue ('Postgres pgKind))))

intersectsNbandGeomInput :: forall pgKind m n r. (MonadSchema n m, MonadError QErr m, MonadReader r m, Has MkTypename r) => m (Parser 'Input n (STIntersectsNbandGeommin (UnpreparedValue ('Postgres pgKind))))

intersectsGeomNbandInput :: forall pgKind m n r. (MonadSchema n m, MonadError QErr m, MonadReader r m, Has MkTypename r) => m (Parser 'Input n (STIntersectsGeomminNband (UnpreparedValue ('Postgres pgKind))))

countTypeInput :: MonadParse n => Maybe (Parser 'Both n (Column ('Postgres pgKind))) -> InputFieldsParser n (CountDistinct -> CountType ('Postgres pgKind))

prependOp :: forall pgKind m n r. (BackendSchema ('Postgres pgKind), MonadReader r m, MonadError QErr m, MonadSchema n m, Has MkTypename r) => UpdateOperator ('Postgres pgKind) m n (UnpreparedValue ('Postgres pgKind))

Update operator that prepends a value to a column containing jsonb arrays.

Note: Currently this is Postgres specific because json columns have not been ported to other backends yet.

appendOp :: forall pgKind m n r. (BackendSchema ('Postgres pgKind), MonadReader r m, MonadError QErr m, MonadSchema n m, Has MkTypename r) => UpdateOperator ('Postgres pgKind) m n (UnpreparedValue ('Postgres pgKind))

Update operator that appends a value to a column containing jsonb arrays.

Note: Currently this is Postgres specific because json columns have not been ported to other backends yet.

deleteKeyOp :: forall pgKind m n r. (BackendSchema ('Postgres pgKind), MonadReader r m, MonadError QErr m, MonadSchema n m, Has MkTypename r) => UpdateOperator ('Postgres pgKind) m n (UnpreparedValue ('Postgres pgKind))

Update operator that deletes a value at a specified key from a column containing jsonb objects.

Note: Currently this is Postgres specific because json columns have not been ported to other backends yet.

deleteElemOp :: forall pgKind m n r. (BackendSchema ('Postgres pgKind), MonadReader r m, MonadError QErr m, MonadSchema n m, Has MkTypename r) => UpdateOperator ('Postgres pgKind) m n (UnpreparedValue ('Postgres pgKind))

Update operator that deletes a value at a specific index from a column containing jsonb arrays.

Note: Currently this is Postgres specific because json columns have not been ported to other backends yet.

deleteAtPathOp :: forall pgKind m n r. (BackendSchema ('Postgres pgKind), MonadReader r m, MonadError QErr m, MonadSchema n m, Has MkTypename r) => UpdateOperator ('Postgres pgKind) m n [UnpreparedValue ('Postgres pgKind)]

Update operator that deletes a field at a certan path from a column containing jsonb objects.

Note: Currently this is Postgres specific because json columns have not been ported to other backends yet.

updateOperators :: forall pgKind m n r. (MonadParse n, MonadReader r m, Has MkTypename r, MonadError QErr m, MonadSchema n m, BackendSchema ('Postgres pgKind)) => TableInfo ('Postgres pgKind) -> UpdPermInfo ('Postgres pgKind) -> m (InputFieldsParser n (HashMap (Column ('Postgres pgKind)) (UpdateOpExpression (UnpreparedValue ('Postgres pgKind)))))

The update operators that we support on Postgres.

Orphan instances

(Backend ('Postgres pgKind), PostgresSchema pgKind) => BackendSchema ('Postgres pgKind)

Instance details

Methods

buildTableQueryFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildTableRelayQueryFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> NESeq (ColumnInfo ('Postgres pgKind)) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildTableInsertMutationFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> InsPermInfo ('Postgres pgKind) -> Maybe (SelPermInfo ('Postgres pgKind)) -> Maybe (UpdPermInfo ('Postgres pgKind)) -> m [FieldParser n (AnnInsert ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildTableUpdateMutationFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> UpdPermInfo ('Postgres pgKind) -> Maybe (SelPermInfo ('Postgres pgKind)) -> m [FieldParser n (AnnotatedUpdateG ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildTableDeleteMutationFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableName ('Postgres pgKind) -> TableInfo ('Postgres pgKind) -> Name -> DelPermInfo ('Postgres pgKind) -> Maybe (SelPermInfo ('Postgres pgKind)) -> m [FieldParser n (AnnDelG ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildFunctionQueryFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> FunctionName ('Postgres pgKind) -> FunctionInfo ('Postgres pgKind) -> TableName ('Postgres pgKind) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildFunctionRelayQueryFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> FunctionName ('Postgres pgKind) -> FunctionInfo ('Postgres pgKind) -> TableName ('Postgres pgKind) -> NESeq (ColumnInfo ('Postgres pgKind)) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (QueryDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

buildFunctionMutationFields :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> FunctionName ('Postgres pgKind) -> FunctionInfo ('Postgres pgKind) -> TableName ('Postgres pgKind) -> SelPermInfo ('Postgres pgKind) -> m [FieldParser n (MutationDB ('Postgres pgKind) (RemoteRelationshipField UnpreparedValue) (UnpreparedValue ('Postgres pgKind)))]

tableArguments :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> TableInfo ('Postgres pgKind) -> SelPermInfo ('Postgres pgKind) -> m (InputFieldsParser n (SelectArgsG ('Postgres pgKind) (UnpreparedValue ('Postgres pgKind))))

mkRelationshipParser :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> RelInfo ('Postgres pgKind) -> m (Maybe (InputFieldsParser n (Maybe (AnnotatedInsert ('Postgres pgKind) (UnpreparedValue ('Postgres pgKind))))))

relayExtension :: Maybe (XRelay ('Postgres pgKind))

nodesAggExtension :: Maybe (XNodesAgg ('Postgres pgKind))

columnParser :: forall (n :: Type -> Type) m r. (MonadSchema n m, MonadError QErr m, MonadReader r m, Has MkTypename r) => ColumnType ('Postgres pgKind) -> Nullability -> m (Parser 'Both n (ValueWithOrigin (ColumnValue ('Postgres pgKind))))

jsonPathArg :: forall (n :: Type -> Type). MonadParse n => ColumnType ('Postgres pgKind) -> InputFieldsParser n (Maybe (ColumnOp ('Postgres pgKind)))

orderByOperators :: NonEmpty (Definition EnumValueInfo, (BasicOrderType ('Postgres pgKind), NullsOrderType ('Postgres pgKind)))

comparisonExps :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => ColumnType ('Postgres pgKind) -> m (Parser 'Input n [ComparisonExp ('Postgres pgKind)])

countTypeInput :: forall (n :: Type -> Type). MonadParse n => Maybe (Parser 'Both n (Column ('Postgres pgKind))) -> InputFieldsParser n (CountDistinct -> CountType ('Postgres pgKind))

aggregateOrderByCountType :: ScalarType ('Postgres pgKind)

computedField :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => SourceName -> ComputedFieldInfo ('Postgres pgKind) -> TableName ('Postgres pgKind) -> SelPermInfo ('Postgres pgKind) -> m (Maybe (FieldParser n (AnnotatedField ('Postgres pgKind))))

node :: forall r m (n :: Type -> Type). MonadBuildSchema ('Postgres pgKind) r m n => m (Parser 'Output n (HashMap (TableName ('Postgres pgKind)) (SourceName, SourceConfig ('Postgres pgKind), SelPermInfo ('Postgres pgKind), PrimaryKeyColumns ('Postgres pgKind), AnnotatedFields ('Postgres pgKind))))

columnDefaultValue :: Column ('Postgres pgKind) -> SQLExpression ('Postgres pgKind)