Hasura.Backends.Postgres.Translate.BoolExp

Description

Postgres Translate BoolExp

Convert IR boolean expressions to Postgres-specific SQL expressions.

Documentation

equalsBoolExpBuilder :: SQLExpression ('Postgres pgKind) -> SQLExpression ('Postgres pgKind) -> BoolExp

notEqualsBoolExpBuilder :: SQLExpression ('Postgres pgKind) -> SQLExpression ('Postgres pgKind) -> BoolExp

annBoolExp :: (QErrM m, TableCoreInfoRM b m, BackendMetadata b) => BoolExpRHSParser b m v -> TableName b -> FieldInfoMap (FieldInfo b) -> GBoolExp b ColExp -> m (AnnBoolExp b v)

annColExp :: (QErrM m, TableCoreInfoRM b m, BackendMetadata b) => BoolExpRHSParser b m v -> TableName b -> FieldInfoMap (FieldInfo b) -> ColExp -> m (AnnBoolExpFld b v)

toSQLBoolExp

Arguments

:: forall pgKind. Backend ('Postgres pgKind)

=> Qual
The name of the tabular value in query scope that the boolean expression applies to

-> AnnBoolExpSQL ('Postgres pgKind)
The boolean expression to translate

-> BoolExp

Translate an IR boolean expression to an SQL boolean expression. References to columns etc are relative to the given rootReference.

data BoolExpCtx

The table context of boolean expression translation. This is used to resolve references to fields, as those may refer to the so-called 'root table' (identified by a $-sign in the expression input syntax) or the current table.

Constructors

BoolExpCtx

Fields

currTableReference :: Qual

Reference to the current tabular value.

rootReference :: Qual

Reference to the root tabular value.

newtype BoolExpM a

The monad that carries the translation of boolean expressions. This supports the generation of fresh names for aliasing sub-expressions and maintains the table context of the expressions being translated.

Constructors

BoolExpM

Fields

unBoolExpM :: ReaderT BoolExpCtx (State Word64) a

translateBoolExp :: forall pgKind. Backend ('Postgres pgKind) => AnnBoolExpSQL ('Postgres pgKind) -> BoolExpM BoolExp

Translate a GBoolExp with annotated SQLExpressions in the leaves into a bare SQL Boolean Expression.

withCurrentTable :: forall a. Qual -> BoolExpM a -> BoolExpM a

Call a given translation action recursively using the given identifier for the current table.

freshIdentifier :: forall a. ToTxt a => QualifiedObject a -> BoolExpM Identifier

Draw a fresh identifier intended to alias the given object.

identifierWithSuffix :: ToTxt a => QualifiedObject a -> Text -> Identifier

translateAVAggregationPredicates :: forall pgKind. Backend ('Postgres pgKind) => AggregationPredicatesImplementation ('Postgres pgKind) (SQLExpression ('Postgres pgKind)) -> BoolExpM BoolExp

Given a GraphQL aggregation filter of the form: > { where: {table_aggregate: {aggregate_field: {bool_op: value} } } }

Produces SQL of the form: > EXISTS ( > SELECT > 1 > FROM > ( > SELECT > aggregate_function AS aggregate_function_alias -- Note: we can have multiple aggregate expressions here > FROM > array_relationship_table AS array_relationship_table_alias > WHERE > relationship_table_key > AND row_permissions > AND filters > ) AS "_sub" > WHERE > "_sub".aggregate_function_alias bool_op value > )

translateAggPredBoolExp :: forall pgKind. TableName ('Postgres pgKind) -> Identifier -> AggregationPredicate ('Postgres pgKind) SQLExp -> BoolExpM BoolExp

translateAggPredsSubselect :: forall pgKind. Backend ('Postgres pgKind) => Identifier -> Identifier -> BoolExp -> AggregationPredicatesImplementation ('Postgres pgKind) SQLExp -> BoolExpM FromItem

translateAggPredExtractor :: forall pgKind field. Identifier -> TableName ('Postgres pgKind) -> AggregationPredicate ('Postgres pgKind) field -> Extractor

translateAggPredArguments :: forall pgKind. AggregationPredicateArguments ('Postgres pgKind) -> Identifier -> NonEmpty SQLExp

translateTableRelationship :: HashMap PGCol PGCol -> Identifier -> BoolExpM BoolExp

data LHSField b

Constructors

LColumn FieldName

LComputedField QualifiedFunction (FunctionArgsExp b (SQLExpression b))

mkComputedFieldFunctionExp :: Qual -> QualifiedFunction -> FunctionArgsExp ('Postgres pgKind) (SQLExpression ('Postgres pgKind)) -> Maybe TableAlias -> FunctionExp

sqlAnd :: [BoolExp] -> BoolExp

mkFieldCompExp :: Qual -> Qual -> LHSField ('Postgres pgKind) -> OpExpG ('Postgres pgKind) SQLExp -> BoolExp