Hasura.Backends.Postgres.Translate.BoolExp

Description

Postgres Translate BoolExp

Convert IR boolean expressions to Postgres-specific SQL expressions.

Documentation

toSQLBoolExp

Arguments

:: forall pgKind. Backend ('Postgres pgKind)

=> Qual
The name of the tabular value in query scope that the boolean expression applies to

-> AnnBoolExpSQL ('Postgres pgKind)
The boolean expression to translate

-> BoolExp

Translate an IR boolean expression to an SQL boolean expression. References to columns etc are relative to the given rootReference.

withRedactionExp :: Backend ('Postgres pgKind) => Qual -> AnnRedactionExp ('Postgres pgKind) SQLExp -> SQLExp -> SQLExp