Hasura.Backends.Postgres.Translate.Select

Description

This module is a translation layer between IR and postgres-specific select queries.

There are three main types of selects (as distinguished from the IR):

"simple" selects

aggregate selects

connection selects (used for relay)

Most exports from this module showcase this distinction. The "interesting" parts of the call tree of these functions is similar:

selectQuerySQL -> mkSQLSelect -> processAnnSimpleSelect -> processSelectParams/processAnnFields

selectAggregateQuerySQL -> mkAggregateSelect -> processAnnAggregateSelect -> processSelectParams/processAnnFields

connetionSelectQuerySQL -> mkConnectionSelect -> processConnectionSelection -> processSelectParams

Random thoughts that might help when diving deeper in this module:

Extractors are a pair of an SQL expression and an alias; they get translated like "[SELECT ...] expr as alias"

a SelectSource consists of a prefix, a source, a boolean conditional expression, and info on whether sorting or slicing is done (needed to handle the LIMIT optimisation)

For details on creating the selection tree for relationships via MonadWriter JoinTree, see withWriteJoinTree

Documentation

selectQuerySQL :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => JsonAggSelect -> AnnSimpleSelect ('Postgres pgKind) -> Query

Translates IR to Postgres queries for simple SELECTs (select queries that are not aggregations, including subscriptions).

See mkSQLSelect for the Postgres AST.

selectAggregateQuerySQL :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => AnnAggregateSelect ('Postgres pgKind) -> Query

Translates IR to Postgres queries for aggregated SELECTs.

See mkAggregateSelect for the Postgres AST.

connectionSelectQuerySQL :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => ConnectionSelect ('Postgres pgKind) Void SQLExp -> Query

Translates IR to Postgres queries for "connection" queries (used for Relay).

See mkConnectionSelect for the Postgres AST.

asSingleRowJsonResp :: Query -> [PrepArg] -> TxE QErr EncJSON

Helper function with no relation to anything else in the module.

This function is generally used on the result of selectQuerySQL, selectAggregateQuerySQL or connectionSelectSQL to run said query and get back the resulting JSON.

TODO: Perhaps this helper could find a new home.

functionToIdentifier :: QualifiedFunction -> Identifier

Converts a function name to an Identifier.

If the schema name is public, it will just use its name, otherwise it will prefix it by the schema name.

selectFromToFromItem :: Identifier -> SelectFrom ('Postgres pgKind) -> FromItem

selectFromToQual :: SelectFrom ('Postgres pgKind) -> Qual

aggregateFieldToExp :: AggregateFields ('Postgres pgKind) -> Bool -> SQLExp

asSingleRowExtr :: Alias -> SQLExp

withJsonAggExtr :: PermissionLimitSubQuery -> Maybe OrderByExp -> Alias -> SQLExp

asJsonAggExtr :: JsonAggSelect -> Alias -> PermissionLimitSubQuery -> Maybe OrderByExp -> Extractor

mkUniqArrayRelationAlias :: FieldName -> [FieldName] -> Identifier

mkArrayRelationTableAlias :: Identifier -> FieldName -> [FieldName] -> Identifier

mkObjectRelationTableAlias :: Identifier -> RelName -> Identifier

mkComputedFieldTableAlias :: Identifier -> FieldName -> Identifier

mkBaseTableAlias :: Identifier -> Identifier

mkBaseTableColumnAlias :: Identifier -> PGCol -> Identifier

mkOrderByFieldName :: ToTxt a => a -> FieldName

mkAggregateOrderByAlias :: AnnotatedAggregateOrderBy ('Postgres pgKind) -> Alias

mkArrayRelationSourcePrefix :: Identifier -> FieldName -> HashMap FieldName [FieldName] -> FieldName -> Identifier

mkArrayRelationAlias :: FieldName -> HashMap FieldName [FieldName] -> FieldName -> Alias

fromTableRowArgs :: Identifier -> FunctionArgsExpTableRow SQLExp -> FunctionArgs

withRowToJSON :: FieldName -> [Extractor] -> (Alias, SQLExp)

withJsonBuildObj :: FieldName -> [SQLExp] -> (Alias, SQLExp)

withForceAggregation :: TypeAnn -> SQLExp -> SQLExp

Forces aggregation

mkAggregateOrderByExtractorAndFields :: forall pgKind. Backend ('Postgres pgKind) => AnnotatedAggregateOrderBy ('Postgres pgKind) -> (Extractor, AggregateFields ('Postgres pgKind))

mkAnnOrderByAlias :: Identifier -> FieldName -> SimilarArrayFields -> AnnotatedOrderByElement ('Postgres pgKind) v -> Alias

Generate alias for order by extractors

applyDistinctOnAtBase :: NonEmpty PGCol -> DistinctExpr

applyDistinctOnAtNode :: Identifier -> NonEmpty PGCol -> (DistinctExpr, [(Alias, SQLExp)])

type SimilarArrayFields = HashMap FieldName [FieldName]

mkSimilarArrayFields :: forall pgKind v. (Backend ('Postgres pgKind), Eq v) => AnnFieldsG ('Postgres pgKind) Void v -> Maybe (NonEmpty (AnnotatedOrderByItemG ('Postgres pgKind) v)) -> SimilarArrayFields

getArrayRelNameAndSelectArgs :: ArraySelectG ('Postgres pgKind) r v -> (RelName, SelectArgsG ('Postgres pgKind) v)

getAnnArr :: (a, AnnFieldG ('Postgres pgKind) r v) -> Maybe (a, ArraySelectG ('Postgres pgKind) r v)

withWriteJoinTree :: MonadWriter JoinTree m => (JoinTree -> b -> JoinTree) -> m (a, b) -> m a

This is the lowest level function which deals with MonadWriter JoinTree, whose purpose is to essentially create the selection tree across relationships.

Each type of relationship uses a different kind of update function; see withWriteObjectRelation, withWriteArrayRelation, withWriteArrayConnection, and withWriteComputedFieldTableSet.

See the definition of JoinTree for details before diving further (particularly its components and Monoid instance).

withWriteObjectRelation :: MonadWriter JoinTree m => m (ObjectRelationSource, HashMap Alias SQLExp, a) -> m a

withWriteArrayRelation :: MonadWriter JoinTree m => m (ArrayRelationSource, Extractor, HashMap Alias SQLExp, a) -> m a

withWriteArrayConnection :: MonadWriter JoinTree m => m (ArrayConnectionSource, Extractor, HashMap Alias SQLExp, a) -> m a

withWriteComputedFieldTableSet :: MonadWriter JoinTree m => m (ComputedFieldTableSetSource, Extractor, HashMap Alias SQLExp, a) -> m a

processAnnSimpleSelect :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => SourcePrefixes -> FieldName -> PermissionLimitSubQuery -> AnnSimpleSelect ('Postgres pgKind) -> m (SelectSource, HashMap Alias SQLExp)

processAnnAggregateSelect :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => SourcePrefixes -> FieldName -> AnnAggregateSelect ('Postgres pgKind) -> m (SelectSource, HashMap Alias SQLExp, Extractor)

mkPermissionLimitSubQuery :: Maybe Int -> TableAggregateFields ('Postgres pgKind) -> Maybe (NonEmpty (AnnotatedOrderByItem ('Postgres pgKind))) -> PermissionLimitSubQuery

processArrayRelation :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => SourcePrefixes -> FieldName -> Alias -> ArraySelect ('Postgres pgKind) -> m ()

processSelectParams :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind)) => SourcePrefixes -> FieldName -> SimilarArrayFields -> SelectFrom ('Postgres pgKind) -> PermissionLimitSubQuery -> TablePerm ('Postgres pgKind) -> SelectArgs ('Postgres pgKind) -> m (SelectSource, [(Alias, SQLExp)], Maybe SQLExp)

processOrderByItems :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind)) => Identifier -> FieldName -> SimilarArrayFields -> Maybe (NonEmpty PGCol) -> Maybe (NonEmpty (AnnotatedOrderByItem ('Postgres pgKind))) -> m ([(Alias, SQLExp)], SelectSorting, Maybe SQLExp)

aggregateFieldsToExtractorExps :: Identifier -> AggregateFields ('Postgres pgKind) -> [(Alias, SQLExp)]

class PostgresAnnotatedFieldJSON (pgKind :: PostgresKind) where

Methods

annRowToJson :: FieldName -> [(FieldName, SQLExp)] -> (Alias, SQLExp)

processAnnFields :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => Identifier -> FieldName -> SimilarArrayFields -> AnnFields ('Postgres pgKind) -> m (Alias, SQLExp)

injectJoinCond

Arguments

:: BoolExp
Join condition

-> BoolExp
Where condition

-> WhereFrag
New where frag

mkJoinCond :: Alias -> HashMap PGCol PGCol -> BoolExp

generateSQLSelect

Arguments

:: BoolExp
Pre join condition

-> SelectSource

-> SelectNode

-> Select

generateSQLSelectFromArrayNode :: SelectSource -> MultiRowSelectNode -> BoolExp -> Select

mkAggregateSelect :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => AnnAggregateSelect ('Postgres pgKind) -> Select

mkSQLSelect :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => JsonAggSelect -> AnnSimpleSelect ('Postgres pgKind) -> Select

mkConnectionSelect :: forall pgKind. (Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => ConnectionSelect ('Postgres pgKind) Void SQLExp -> SelectWithG Select

mkFirstElementExp :: SQLExp -> SQLExp

First element extractor expression from given record set For example:- To get first "id" column from given row set, the function generates the SQL expression AS `(array_agg("id"))[1]`

mkLastElementExp :: SQLExp -> SQLExp

Last element extractor expression from given record set. For example:- To get first "id" column from given row set, the function generates the SQL expression AS `(array_agg("id"))[array_length(array_agg("id"), 1)]`

cursorIdentifier :: Identifier

startCursorIdentifier :: Identifier

endCursorIdentifier :: Identifier

hasPreviousPageIdentifier :: Identifier

hasNextPageIdentifier :: Identifier

pageInfoSelectAliasIdentifier :: Identifier

cursorsSelectAliasIdentifier :: Identifier

encodeBase64 :: SQLExp -> SQLExp

processConnectionSelect :: forall pgKind m. (MonadReader Bool m, MonadWriter JoinTree m, Backend ('Postgres pgKind), PostgresAnnotatedFieldJSON pgKind) => SourcePrefixes -> FieldName -> Alias -> HashMap PGCol PGCol -> ConnectionSelect ('Postgres pgKind) Void SQLExp -> m (ArrayConnectionSource, Extractor, HashMap Alias SQLExp)

connectionToSelectWith :: Alias -> ArrayConnectionSource -> MultiRowSelectNode -> SelectWithG Select