Hasura.Backends.Postgres.Translate.Select.Internal.JoinTree

Description

Stuff gutted from Translate.Select

Documentation

withWriteJoinTree :: MonadWriter JoinTree m => (JoinTree -> b -> JoinTree) -> m (a, b) -> m a

This is the lowest level function which deals with MonadWriter JoinTree, whose purpose is to essentially create the selection tree across relationships.

Each type of relationship uses a different kind of update function; see withWriteObjectRelation, withWriteArrayRelation, withWriteArrayConnection, and withWriteComputedFieldTableSet.

See the definition of JoinTree for details before diving further (particularly its components and Monoid instance).

withWriteObjectRelation :: MonadWriter JoinTree m => m (ObjectRelationSource, HashMap ColumnAlias SQLExp, a) -> m a

withWriteArrayRelation :: MonadWriter JoinTree m => m (ArrayRelationSource, Extractor, HashMap ColumnAlias SQLExp, a) -> m a

withWriteArrayConnection :: MonadWriter JoinTree m => m (ArrayConnectionSource, Extractor, HashMap ColumnAlias SQLExp, a) -> m a

withWriteComputedFieldTableSet :: MonadWriter JoinTree m => m (ComputedFieldTableSetSource, Extractor, HashMap ColumnAlias SQLExp, a) -> m a