Hasura.Backends.Postgres.Types.Update

Description

Postgres Types Update

This module defines the Update-related IR types specific to Postgres.

Documentation

data MultiRowUpdate pgKind v

Represents an entry in an update_table_many update.

Constructors

MultiRowUpdate

Fields

mruWhere :: AnnBoolExp ('Postgres pgKind) v

The where clause for each individual update.

Note that the single updates do not have a where clause, because it uses the one found in AnnotatedUpdateG. However, we have one for each update for update_many.

mruExpression :: HashMap PGCol (UpdateOpExpression v)

The update expression, e.g, "set", "inc", etc., for each column.

data BackendUpdate pgKind v

The PostgreSQL-specific data of an Update expression.

This is parameterised over v which enables different phases of IR transformation to maintain the overall structure while enriching/transforming the data at the leaves.

Constructors

BackendUpdate (HashMap PGCol (UpdateOpExpression v))
The update operations to perform on each colum.

BackendMultiRowUpdate [MultiRowUpdate pgKind v]
The update operations to perform, in sequence, for an update_table_many operation.

isEmpty :: BackendUpdate pgKind v -> Bool

Are we updating anything?

data UpdateOpExpression v

The various update operators supported by PostgreSQL, i.e. the _set, _inc operators that appear in the schema.

See Update Mutations User docs

Constructors

UpdateSet v

UpdateInc v

UpdateAppend v

UpdatePrepend v

UpdateDeleteKey v

UpdateDeleteElem v

UpdateDeleteAtPath [v]