Hasura.Backends.Postgres.Types.Update

Description

Postgres Types Update

This module defines the Update-related IR types specific to Postgres.

Documentation

data BackendUpdate v

The PostgreSQL-specific data of an Update expression.

This is parameterised over v which enables different phases of IR transformation to maintain the overall structure while enriching/transforming the data at the leaves.

Constructors

BackendUpdate

Fields

updateOperations :: !(HashMap PGCol (UpdateOpExpression v))

The update operations to perform on each colum.

data UpdateOpExpression v

The various update operators supported by PostgreSQL, i.e. the _set, _inc operators that appear in the schema.

See Update Mutations User docs

Constructors

UpdateSet !v

UpdateInc !v

UpdateAppend !v

UpdatePrepend !v

UpdateDeleteKey !v

UpdateDeleteElem !v

UpdateDeleteAtPath ![v]