Hasura.Eventing.HTTP

Description

Hasura.Eventing.HTTP

This module is an utility module providing HTTP utilities for Hasura.Eventing.EventTriggers and Hasura.Eventing.ScheduledTriggers.

The event triggers and scheduled triggers share the event delivery mechanism using the tryWebhook function defined in this module.

Documentation

data LogBehavior

Constructors

LogBehavior

Fields

_lbHeader :: !HeaderLogBehavior

_lbResponse :: !ResponseLogBehavior

data HeaderLogBehavior

Constructors

LogEnvValue

LogEnvVarname

data ResponseLogBehavior

Constructors

LogSanitisedResponse

LogEntireResponse

retryAfterHeader :: CI Text

data ExtraLogContext

Constructors

ExtraLogContext

Fields

elEventId :: !EventId

elEventName :: !(Maybe TriggerName)

data HTTPResp (a :: TriggerTypes)

Constructors

HTTPResp

Fields

hrsStatus :: !Int

hrsHeaders :: ![HeaderConf]

hrsBody :: !TByteString

hrsSize :: !Int64

data HTTPErr (a :: TriggerTypes)

Constructors

HClient !HttpException

HStatus !(HTTPResp a)

HOther !String

mkHTTPResp :: Response ByteString -> HTTPResp a

data RequestDetails

Constructors

RequestDetails

Fields

_rdOriginalRequest :: Request

_rdOriginalSize :: Int64

_rdTransformedRequest :: Maybe Request

_rdTransformedSize :: Maybe Int64

_rdReqTransformCtx :: Maybe RequestTransformCtx

_rdSessionVars :: Maybe SessionVariables

extractRequest :: RequestDetails -> Request

data HTTPRespExtra (a :: TriggerTypes)

Constructors

HTTPRespExtra

Fields

_hreResponse :: !(Either (HTTPErr a) (HTTPResp a))

_hreContext :: !ExtraLogContext

_hreRequest :: !RequestDetails

_hreLogResponse :: !ResponseLogBehavior

Whether to log the entire response, including the body and the headers, which may contain sensitive information.

isNetworkError :: HTTPErr a -> Bool

isNetworkErrorHC :: HttpException -> Bool

anyBodyParser :: Response ByteString -> Either (HTTPErr a) (HTTPResp a)

data HTTPReq

Constructors

HTTPReq

Fields

_hrqMethod :: !String

_hrqUrl :: !String

_hrqPayload :: !(Maybe Value)

_hrqTry :: !Int

_hrqDelay :: !(Maybe Int)

logHTTPForET :: (MonadReader r m, Has (Logger Hasura) r, MonadIO m) => Either (HTTPErr 'EventType) (HTTPResp 'EventType) -> ExtraLogContext -> RequestDetails -> LogBehavior -> m ()

logHTTPForST :: (MonadReader r m, Has (Logger Hasura) r, MonadIO m) => Either (HTTPErr 'ScheduledType) (HTTPResp 'ScheduledType) -> ExtraLogContext -> RequestDetails -> LogBehavior -> m ()

runHTTP :: MonadIO m => Manager -> Request -> m (Either (HTTPErr a) (HTTPResp a))

data TransformableRequestError a

Constructors

HTTPError Value (HTTPErr a)

TransformationError Value TransformErrorBundle

mkRequest

Arguments

:: MonadError (TransformableRequestError a) m

=> [Header]

-> ResponseTimeout

-> ByteString
the request body. It is passed as a Bytestring because we need to log the request size. As the logging happens outside the function, we pass it the final request body, instead of Value

-> Maybe RequestTransform

-> ResolvedWebhook

-> m RequestDetails

invokeRequest :: (MonadReader r m, MonadError (TransformableRequestError a) m, Has Manager r, Has (Logger Hasura) r, MonadIO m, MonadTrace m) => RequestDetails -> Maybe ResponseTransform -> Maybe SessionVariables -> (Either (HTTPErr a) (HTTPResp a) -> RequestDetails -> m ()) -> m (HTTPResp a)

mkResp :: Int -> TByteString -> [HeaderConf] -> Response a

mkClientErr :: TByteString -> Response a

mkWebhookReq :: Value -> [HeaderConf] -> InvocationVersion -> WebhookRequest

mkInvocationResp :: Maybe Int -> TByteString -> [HeaderConf] -> Response a

isClientError :: Int -> Bool

encodeHeader :: EventHeaderInfo -> Header

decodeHeader :: LogBehavior -> [EventHeaderInfo] -> (HeaderName, ByteString) -> HeaderConf

prepareHeaders :: LogBehavior -> [EventHeaderInfo] -> ([Header], [HeaderConf])

Encodes given request headers along with our defaultHeaders and returns them along with the re-decoded set of headers (for logging purposes).

getRetryAfterHeaderFromHTTPErr :: HTTPErr a -> Maybe Text

getRetryAfterHeaderFromResp :: HTTPResp a -> Maybe Text

parseRetryHeaderValue :: Text -> Maybe Int