Hasura.Function.Cache

Description

types and helpers for user-defined-functions after they have been resolved in the schema cache

Documentation

data FunctionVolatility

https://www.postgresql.org/docs/current/xfunc-volatility.html

Constructors

FTVOLATILE

FTIMMUTABLE

FTSTABLE

funcTypToTxt :: FunctionVolatility -> Text

newtype FunctionArgName

Constructors

FunctionArgName

Fields

getFuncArgNameTxt :: Text

data InputArgument a

Constructors

IAUserProvided a

IASessionVariables FunctionArgName

type FunctionInputArgument b = InputArgument (FunctionArgument b)

data FunctionExposedAs

Indicates whether the user requested the corresponding function to be tracked as a mutation or a query/subscription, in track_function.

Constructors

FEAQuery

FEAMutation

newtype FunctionPermissionInfo

Constructors

FunctionPermissionInfo

Fields

_fpmRole :: RoleName

type FunctionPermissionsMap = HashMap RoleName FunctionPermissionInfo

data FunctionCustomRootFields

Custom root fields for functions. When set, will be the names exposed to the user in the schema.

See rfcs/function-root-field-customisation.md for more information.

Constructors

FunctionCustomRootFields

Fields

_fcrfFunction :: Maybe Name

_fcrfFunctionAggregate :: Maybe Name

emptyFunctionCustomRootFields :: FunctionCustomRootFields

A function custom root fields without custom names set. This is the default.

data FunctionInfo (b :: BackendType)

Tracked SQL function metadata. See buildFunctionInfo.

Constructors

FunctionInfo

Fields

_fiSQLName :: FunctionName b

_fiGQLName :: Name

_fiGQLArgsName :: Name

_fiGQLAggregateName :: Name

_fiSystemDefined :: SystemDefined

_fiVolatility :: FunctionVolatility

_fiExposedAs :: FunctionExposedAs

In which part of the schema should this function be exposed?

See mkFunctionInfo and _fcExposedAs.

_fiInputArgs :: Seq (FunctionInputArgument b)

_fiReturnType :: TableName b

NOTE: when a table is created, a new composite type of the same name is automatically created; so strictly speaking this field means "the function returns the composite type corresponding to this table".

_fiDescription :: Maybe Text

this field represents the description of the function as present on the database

_fiPermissions :: FunctionPermissionsMap

Roles to which the function is accessible

_fiJsonAggSelect :: JsonAggSelect

_fiComment :: Maybe Text

type FunctionCache b = HashMap (FunctionName b) (FunctionInfo b)

data TrackableFunctionInfo b

Constructors

TrackableFunctionInfo

Fields

tfiFunctionName :: FunctionName b

tfiFunctionVolitility :: FunctionVolatility

newtype TrackableTableInfo b

Constructors

TrackableTableInfo

Fields

tfTableiName :: TableName b

data TrackableInfo b

Constructors

TrackableInfo

Fields

trackableFunctions :: [TrackableFunctionInfo b]

trackableTables :: [TrackableTableInfo b]

data FunctionConfig b

Tracked function configuration, and payload of the pg_track_function and pg_set_function_customization API calls.

Constructors

FunctionConfig

Fields

_fcSessionArgument :: Maybe FunctionArgName

_fcExposedAs :: Maybe FunctionExposedAs

In which top-level field should we expose this function?

The user might omit this, in which case we'll infer the location from the SQL functions volatility. See mkFunctionInfo or the track_function API docs for details of validation, etc.

_fcCustomRootFields :: FunctionCustomRootFields

_fcCustomName :: Maybe Name

_fcResponse :: Maybe (FunctionReturnType b)

emptyFunctionConfig :: FunctionConfig b

The default function config; v1 of the API implies this.

type DBFunctionsMetadata b = HashMap (FunctionName b) (FunctionOverloads b)

newtype FunctionOverloads b

Constructors

FunctionOverloads

Fields

getFunctionOverloads :: NonEmpty (RawFunctionInfo b)

data FunctionArgsExpG a

Constructors

FunctionArgsExp

Fields

_faePositional :: [a]

_faeNamed :: HashMap Text a

type FunctionArgsExp b v = FunctionArgsExpG (FunctionArgumentExp b v)

emptyFunctionArgsExp :: FunctionArgsExpG a