Hasura.GraphQL.Analyse

Description

Tools to analyze the structure of a GraphQL request.

Documentation

data Structure

Overall structure of a given query.

Constructors

Structure

Fields

_stSelection :: Selection

_stVariables :: HashMap Name VariableInfo

data Selection

Represents a selection of fields within a query.

Constructors

Selection

Fields

_seFields :: HashMap Name FieldInfo

data FieldInfo

Constructors

FieldInfo

Fields

_fiType :: GType

_fiTypeDefinition :: TypeDefinition [Name] InputValueDefinition

_fiSelection :: Maybe Selection

data VariableInfo

TODO: document this

Constructors

VariableInfo

Fields

_viType :: GType

_viDefaultValue :: Maybe (Value Void)

diagnoseGraphQLQuery :: SchemaIntrospection -> ExecutableDefinition Name -> Maybe [Text]

Given the schema's definition, and a query, validate that the query is consistent. We do this by running the analysis, but discarding the result: we do not care about the structure, only about the validity of the query.

Returns Nothing if the query is valid, or a list of messages otherwise.

analyzeGraphQLQuery :: SchemaIntrospection -> ExecutableDefinition Name -> (Structure, [Text])

Given the schema's definition, and a query, run the analysis.

We process all possible fields, and return a partially filled structure if necessary. Given the following query:

query {
  foo {
    bar
  }
  does_not_exist {
    ghsdflgh
  }
}

We would return a structure containing:

foo: {
  bar: {
  }
}

AND an error about "does_not_exist" not existing.

analyzeSelectionSet :: TypeDefinition [Name] InputValueDefinition -> SelectionSet FragmentSpread Name -> Analysis (Maybe Selection)

Check the consistency between the schema information about a type and a selection set (or lack thereof) on that type.

analyzeObjectSelectionSet :: ObjectTypeDefinition InputValueDefinition -> SelectionSet FragmentSpread Name -> Analysis Selection

Analyze the fields of an object selection set against its definition, and emit the corresponding Selection. We ignore the fields that fail, and we continue accumulating the others.

newtype Analysis a

The monad in which we run our analysis.

Has three capabilities: - reader carries the current path, and the full schema for lookups - writer logs all errors we have caught - except allows for short-circuiting errors

Constructors

Analysis (ExceptT AnalysisError (ReaderT (Path, SchemaIntrospection) (Writer [AnalysisError])) a)

runAnalysis :: Monoid a => SchemaIntrospection -> Analysis a -> (a, [Text])

lookupType :: Name -> Analysis (Maybe (TypeDefinition [Name] InputValueDefinition))

Look up a type in the schema.

withField :: Name -> Analysis a -> Analysis a

Add the current field to the error path.

throwDiagnosis :: Diagnosis -> Analysis a

Throws an AnalysisError by combining the given diagnosis with the current path. This interrupts the computation in the given branch, and must be caught for the analysis to resume.

withCatchAndRecord :: Analysis a -> Analysis (Maybe a)

Runs the given computation. if it fails, cacthes the error, records it in the monad, and return Nothing. This allows for a clean recovery.

data AnalysisError

Constructors

AnalysisError

Fields

_aePath :: Path

_aeDiagnosis :: Diagnosis

type Path = Seq Text

data Diagnosis

Constructors

TopLevelFragment

UntypedTopLevelOperation

NoTopLevelSelection

TypeNotFound Name

EnumSelectionSet Name

ScalarSelectionSet Name

InputObjectSelectionSet Name

ObjectFieldNotFound Name Name

render :: AnalysisError -> Text

queryRootName :: Name

mutationRootName :: Name

subscriptionRootName :: Name

typenameField :: FieldDefinition InputValueDefinition

schemaField :: FieldDefinition InputValueDefinition

typeField :: FieldDefinition InputValueDefinition

mkReservedField :: Name -> Name -> FieldDefinition InputValueDefinition