Hasura.GraphQL.Analyse

Documentation

data Analysis var

Analysis and FieldAnalysis are almost similar, except, Analysis will have a bool field (in future) to indicate whether the GQL query is valid or not

Constructors

Analysis

Fields

_aFields :: InsOrdHashMap FieldName (FieldDef, Maybe (FieldAnalysis var))

ordered hashmap from the fields to it's definition and analysis

_aVars :: HashMap VarName (GType, Maybe (Value Void))

map of all the variables, their type and default value (if there is any)

_aErrs :: [Text]

data FieldAnalysis var

Constructors

FieldAnalysis

Fields

_fFields :: InsOrdHashMap FieldName (FieldDef, Maybe (FieldAnalysis var))

_fVars :: HashMap VarName (GType, Maybe (Value var))

_fErrs :: [Text]

data FieldDef

FieldDef is analogous to GType from the Syntax module

Constructors

FieldInfo Nullability (TypeDefinition [Name] RemoteSchemaInputValueDefinition)

FieldList Nullability FieldDef

type FieldName = Name

type VarName = Name

safeInsertInFieldMap :: (Eq k, Hashable k, Semigroup a1) => InsOrdHashMap k (a2, Maybe a1) -> (k, (a2, Maybe a1)) -> InsOrdHashMap k (a2, Maybe a1)

inserts in field map, if there is already a key, then take a union of the fields inside i.e. for the following graphql query: query MyQuery { test { a b } test { b c } } The final field map of test will have a, b and c

analyzeGraphqlQuery :: ExecutableDefinition Name -> RemoteSchemaIntrospection -> Maybe (Analysis Name)

getFieldName :: Field frag var -> Name

getFieldsMap :: RemoteSchemaIntrospection -> [(OperationType, Selection frag var)] -> (InsOrdHashMap FieldName (FieldDef, Maybe (FieldAnalysis var)), [Text])

lookupRoot :: RemoteSchemaIntrospection -> (OperationType, Field frag var) -> Either (FieldDef, Maybe (FieldAnalysis var)) [Text]

lookupDefinition :: RemoteSchemaIntrospection -> GType -> Field frag var -> Either (FieldDef, Maybe (FieldAnalysis var)) [Text]

field :: Selection frag var -> Maybe (Field frag var)

(&&&&) :: Applicative f => (t -> f a1) -> (t -> f a2) -> t -> f (a1, a2) infixl 7

getRootFieldNameFromOpType :: OperationType -> Name

asObjectTypeDefinition :: TypeDefinition possibleTypes inputType -> Maybe (ObjectTypeDefinition inputType)

lookupRS :: RemoteSchemaIntrospection -> Name -> Maybe (TypeDefinition [Name] RemoteSchemaInputValueDefinition)

typeToSchemaM :: GType -> (Name -> Either (TypeDefinition [Name] RemoteSchemaInputValueDefinition, Maybe (FieldAnalysis var)) [Text]) -> Either (FieldDef, Maybe (FieldAnalysis var)) [Text]

getDefinition :: RemoteSchemaIntrospection -> TypeDefinition [Name] RemoteSchemaInputValueDefinition -> SelectionSet frag var -> (TypeDefinition [Name] RemoteSchemaInputValueDefinition, Maybe (FieldAnalysis var))

itrListWith :: GType -> (Name -> p) -> p

getFieldVars :: FieldDefinition RemoteSchemaInputValueDefinition -> HashMap Name (Value var) -> [(VarName, (GType, Maybe (Value var)))]

mkFieldAnalysis :: RemoteSchemaIntrospection -> Selection frag var -> FieldDefinition RemoteSchemaInputValueDefinition -> Maybe (FieldAnalysis var)

lookupFieldBySelection :: Selection frag0 var0 -> [FieldDefinition RemoteSchemaInputValueDefinition] -> Maybe (FieldDefinition RemoteSchemaInputValueDefinition)