Hasura.GraphQL.ApolloFederation

Description

Tools for generating fields for Apollo federation

Field Parser generators

apolloRootFields :: ApolloFederationStatus -> [(Name, Parser 'Output Parse (ApolloFederationParserFunction Parse))] -> [FieldParser Parse (SchemaIntrospection -> QueryRootField UnpreparedValue)]

data ApolloFederationParserFunction (n :: Type -> Type)

Internal parser function for entities field

Constructors

ApolloFederationParserFunction

Fields

aafuGetRootField :: ApolloFederationAnyType -> n (QueryRootField UnpreparedValue)

convertToApolloFedParserFunc :: forall (n :: Type -> Type) (b :: BackendType). (MonadParse n, Backend b) => SourceInfo b -> TableInfo b -> SelPermInfo b -> StringifyNumbers -> Maybe NamingCase -> NESeq (ColumnInfo b) -> Parser 'Output n (AnnotatedFields b) -> Parser 'Output n (ApolloFederationParserFunction n)

getApolloFederationStatus :: HashSet ExperimentalFeature -> Maybe ApolloFederationStatus -> ApolloFederationStatus

Check if the Apollo Federation feature is enabled or not. If the user has explicitly set the Apollo Federation status, then we use that else we fallback to the experimental feature flag

generateSDLWithAllTypes :: SchemaIntrospection -> Text

generateSDL :: SchemaIntrospection -> Text