Hasura.GraphQL.Context

Documentation

data RoleContext a

For storing both a normal GQLContext and one for the backend variant. Currently, this is to enable the backend variant to have certain insertupdatedelete permissions which the frontend variant does not.

Constructors

RoleContext

Fields

_rctxDefault :: !a

The default context for normal sessions

_rctxBackend :: !(Maybe a)

The context for sessions with backend privilege.

data GQLContext

Constructors

GQLContext

Fields

gqlQueryParser :: ParserFn (RootFieldMap (QueryRootField UnpreparedValue))

gqlMutationParser :: Maybe (ParserFn (RootFieldMap (MutationRootField UnpreparedValue)))

gqlSubscriptionParser :: Maybe (ParserFn (RootFieldMap (QueryRootField UnpreparedValue)))

type ParserFn a = SelectionSet NoFragments Variable -> Either QErr a