Hasura.GraphQL.Execute

Documentation

data ExecutionCtx

Execution context

Constructors

ExecutionCtx

Fields

_ecxLogger :: Logger Hasura

_ecxSqlGenCtx :: SQLGenCtx

_ecxSchemaCache :: SchemaCache

_ecxSchemaCacheVer :: SchemaCacheVer

_ecxHttpManager :: Manager

_ecxEnableAllowList :: Bool

_ecxReadOnlyMode :: ReadOnlyMode

getExecPlanPartial :: MonadError QErr m => UserInfo -> SchemaCache -> GraphQLQueryType -> GQLReqParsed -> m (GQLContext, SingleOperation)

data ResolvedExecutionPlan

Constructors

QueryExecutionPlan ExecutionPlan [QueryRootField UnpreparedValue] DirectiveMap
query execution; remote schemas and introspection possible

MutationExecutionPlan ExecutionPlan
mutation execution; only __typename introspection supported

SubscriptionExecutionPlan SubscriptionExecution
either action query or live query execution; remote schemas and introspection not supported

newtype MultiplexedSubscriptionQueryPlan (b :: BackendType)

Constructors

MultiplexedSubscriptionQueryPlan (SubscriptionQueryPlan b (MultiplexedQuery b))

newtype SubscriptionQueryPlan

Constructors

SubscriptionQueryPlan (AnyBackend MultiplexedSubscriptionQueryPlan)

data SourceSubscription

Constructors

SSLivequery !(HashSet ActionId) !(ActionLogResponseMap -> ExceptT QErr IO (SourceName, SubscriptionQueryPlan))

SSStreaming !RootFieldAlias !(SourceName, SubscriptionQueryPlan)

data SubscriptionExecution

The comprehensive subscription plan. We only support either 1. Fields with only async action queries with no associated relationships or 2. Source database query fields from same source and also can be mixed with async action query fields whose relationships are defined to tables in the source

Constructors

SEAsyncActionsWithNoRelationships !(RootFieldMap (ActionId, ActionLogResponse -> Either QErr EncJSON))

SEOnSourceDB !SourceSubscription

buildSubscriptionPlan :: forall m. (MonadError QErr m, MonadQueryTags m, MonadIO m, MonadBaseControl IO m) => UserInfo -> RootFieldMap (QueryRootField UnpreparedValue) -> ParameterizedQueryHash -> m SubscriptionExecution

checkQueryInAllowlist :: MonadError QErr m => Bool -> AllowlistMode -> UserInfo -> GQLReqParsed -> SchemaCache -> m ()

getResolvedExecPlan :: forall m. (MonadError QErr m, MonadMetadataStorage (MetadataStorageT m), MonadIO m, MonadBaseControl IO m, MonadTrace m, MonadGQLExecutionCheck m, MonadQueryTags m) => Environment -> Logger Hasura -> UserInfo -> SQLGenCtx -> ReadOnlyMode -> SchemaCache -> SchemaCacheVer -> GraphQLQueryType -> Manager -> [Header] -> (GQLReqUnparsed, GQLReqParsed) -> RequestId -> m (ParameterizedQueryHash, ResolvedExecutionPlan)

isSingleNamespace :: RootFieldMap a -> Bool

Even when directive _multiple_top_level_fields is given, we can't allow fields within differently-aliased namespaces. This is because the namespace is added to the result by wrapping the bytestring response we get back from the DB.