Hasura.GraphQL.Execute

Documentation

data ExecutionStep where

One execution step to processing a GraphQL query (e.g. one root field).

Constructors

ExecStepDB :: ResponseHeaders -> AnyBackend DBStepInfo -> Maybe RemoteJoins -> ExecutionStep
A query to execute against the database

ExecStepAction :: ActionExecutionPlan -> ActionsInfo -> Maybe RemoteJoins -> ExecutionStep
Execute an action

ExecStepRemote :: !RemoteSchemaInfo -> !ResultCustomizer -> !GQLReqOutgoing -> Maybe RemoteJoins -> ExecutionStep
A graphql query to execute against a remote schema

ExecStepRaw :: Value -> ExecutionStep
Output a plain JSON object

ExecStepMulti :: [ExecutionStep] -> ExecutionStep

data ResolvedExecutionPlan

Constructors

QueryExecutionPlan ExecutionPlan [QueryRootField UnpreparedValue] DirectiveMap
query execution; remote schemas and introspection possible

MutationExecutionPlan ExecutionPlan
mutation execution; only __typename introspection supported

SubscriptionExecutionPlan (SubscriptionExecution, Maybe (Endo Value))
either action query or live query execution; remote schemas and introspection not supported

data GraphQLQueryType

Constructors

QueryHasura

QueryRelay

getResolvedExecPlan :: (MonadError QErr m, MonadMetadataStorage m, MonadIO m, MonadBaseControl IO m, MonadTrace m, MonadGQLExecutionCheck m, MonadQueryTags m, ProvidesNetwork m, MonadGetPolicies m) => Environment -> Logger Hasura -> PrometheusMetrics -> UserInfo -> SQLGenCtx -> ReadOnlyMode -> SchemaCache -> GraphQLQueryType -> [Header] -> GQLReqUnparsed -> SingleOperation -> Maybe Name -> RequestId -> TraceQueryStatus -> m (ParameterizedQueryHash, ResolvedExecutionPlan, [ModelInfoPart])

Construct a ResolvedExecutionPlan from a GQLReqParsed and a bunch of metadata.

Labelling it as inlineable fixed a performance regression on GHC 8.10.7.

makeGQLContext :: UserInfo -> SchemaCache -> GraphQLQueryType -> GQLContext

Construct a single step of an execution plan.

execRemoteGQ

Arguments

:: (MonadIO m, MonadError QErr m, MonadTrace m, ProvidesNetwork m)

=> Environment

-> HttpPropagator

-> UserInfo

-> [Header]

-> ValidatedRemoteSchemaDef

-> GQLReqOutgoing

-> m (DiffTime, [Header], ByteString)
Returns the response body and headers, along with the time taken for the HTTP request to complete

Sends a GraphQL query to the given server.

data SubscriptionExecution

The comprehensive subscription plan. We only support either 1. Fields with only async action queries with no associated relationships or 2. Source database query fields from same source and also can be mixed with async action query fields whose relationships are defined to tables in the source

Constructors

SEAsyncActionsWithNoRelationships !(RootFieldMap (ActionId, ActionLogResponse -> Either QErr EncJSON))

SEOnSourceDB !SourceSubscription

buildSubscriptionPlan :: (MonadError QErr m, MonadQueryTags m, MonadIO m, MonadBaseControl IO m) => UserInfo -> RootFieldMap (QueryRootField UnpreparedValue) -> ParameterizedQueryHash -> [Header] -> Maybe Name -> m ((SubscriptionExecution, Maybe (Endo Value)), [ModelInfoPart])

class Monad m => MonadGQLExecutionCheck (m :: Type -> Type) where

Typeclass representing safety checks (if any) that need to be performed before a GraphQL query should be allowed to be executed. In OSS, the safety check is to check in the query is in the allow list.

the executeIntrospection function has different implementations in OSS and Pro. In Pro, the GraphQL schema introspection can be disabled for specified roles and in OSS there is no restrictions.

| TODO (from master): Limitation: This parses the query, which is not ideal if we already have the query cached. The parsing happens unnecessary. But getting this to either return a plan or parse was tricky and complicated.

Methods

checkGQLExecution

Arguments

:: UserInfo

-> ([Header], IpAddress)

-> AllowListStatus
allow list enabled?

-> SchemaCache
needs allow list

-> GQLReqUnparsed
the unparsed GraphQL query string (and related values)

-> RequestId

-> m (Either QErr GQLReqParsed)

executeIntrospection :: UserInfo -> Value -> SetGraphqlIntrospectionOptions -> m (Either QErr ExecutionStep)

checkGQLBatchedReqs :: UserInfo -> RequestId -> [GQLReq GQLQueryText] -> SchemaCache -> m (Either QErr ())

checkQueryInAllowlist :: MonadError QErr m => AllowListStatus -> AllowlistMode -> UserInfo -> GQLReqParsed -> SchemaCache -> m ()

newtype MultiplexedSubscriptionQueryPlan (b :: BackendType)

Constructors

MultiplexedSubscriptionQueryPlan (SubscriptionQueryPlan b (MultiplexedQuery b))

newtype SubscriptionQueryPlan

Constructors

SubscriptionQueryPlan (AnyBackend MultiplexedSubscriptionQueryPlan)

data SourceSubscription

Constructors

SSLivequery !(HashSet ActionId) !(ActionLogResponseMap -> ExceptT QErr IO ((SourceName, SubscriptionQueryPlan), [ModelInfoPart]))

SSStreaming !RootFieldAlias !((SourceName, SubscriptionQueryPlan), [ModelInfoPart])