Hasura.GraphQL.Execute.Action

Documentation

fetchActionLogResponses :: (MonadError QErr m, MonadMetadataStorage (MetadataStorageT m), Foldable t) => t ActionId -> m (ActionLogResponseMap, Bool)

runActionExecution :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadTrace m, MonadMetadataStorage (MetadataStorageT m)) => UserInfo -> ActionExecutionPlan -> m (DiffTime, (EncJSON, Maybe ResponseHeaders))

asSingleRowJsonResp :: Query -> [PrepArg] -> TxE QErr EncJSON

This function is generally used on the result of selectQuerySQL, selectAggregateQuerySQL or connectionSelectSQL to run said query and get back the resulting JSON.

resolveActionExecution :: Environment -> Logger Hasura -> UserInfo -> AnnActionExecution Void -> ActionExecContext -> Maybe GQLQueryText -> ActionExecution

Synchronously execute webhook handler and resolve response to action "output"

makeActionResponseNoRelations :: ActionFields -> GraphQLType -> ActionOutputFields -> Bool -> ActionWebhookResponse -> Value

Build action response from the Webhook JSON response when there are no relationships defined

gTypeContains :: (GType -> ActionOutputFields -> Bool) -> GType -> ActionOutputFields -> Bool

isCustomScalar :: GType -> ActionOutputFields -> Bool

resolveActionMutationAsync :: MonadMetadataStorage m => AnnActionMutationAsync -> [Header] -> SessionVariables -> m ActionId

Resolve asynchronous action mutation which returns only the action uuid

resolveAsyncActionQuery :: UserInfo -> AnnActionAsyncQuery ('Postgres 'Vanilla) Void -> AsyncActionQueryExecution (UnpreparedValue ('Postgres 'Vanilla))

See Note: [Resolving async action query]

asyncActionsProcessor :: forall m. (MonadIO m, MonadBaseControl IO m, Forall (Pure m), HasReporter m, MonadMetadataStorage (MetadataStorageT m)) => Environment -> Logger Hasura -> IO SchemaCache -> TVar (Set LockedActionEventId) -> Manager -> Milliseconds -> Maybe GQLQueryText -> m (Forever m)

Process async actions from hdb_catalog.hdb_action_log table. This functions is executed in a background thread. See Note [Async action architecture] above

callWebhook :: forall m r. (MonadIO m, MonadError QErr m, MonadTrace m, MonadReader r m, Has (Logger Hasura) r) => Environment -> Manager -> GraphQLType -> ActionOutputFields -> [Header] -> [HeaderConf] -> Bool -> EnvRecord ResolvedWebhook -> ActionWebhookPayload -> Timeout -> Maybe RequestTransform -> Maybe MetadataResponseTransform -> m (ActionWebhookResponse, ResponseHeaders)

processOutputSelectionSet :: ArgumentExp v -> GraphQLType -> [(PGCol, PGScalarType)] -> ActionFields -> StringifyNumbers -> AnnSimpleSelectG ('Postgres 'Vanilla) Void v

actionFieldToAnnField :: ActionFieldG Void -> AnnFieldG ('Postgres 'Vanilla) Void v

mkJsonAggSelect :: GraphQLType -> JsonAggSelect

insertActionTx :: ActionName -> SessionVariables -> [Header] -> Value -> TxE QErr ActionId

fetchUndeliveredActionEventsTx :: TxE QErr [ActionLogItem]

setActionStatusTx :: ActionId -> AsyncActionStatus -> TxE QErr ()

fetchActionResponseTx :: ActionId -> TxE QErr ActionLogResponse

clearActionDataTx :: ActionName -> TxE QErr ()

setProcessingActionLogsToPendingTx :: LockedActionIdArray -> TxE QErr ()