Hasura.GraphQL.Execute.RemoteJoin.Collect

Documentation

getRemoteJoinsQueryDB :: forall (b :: BackendType). Backend b => QueryDB b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (QueryDB b Void (UnpreparedValue b), Maybe RemoteJoins)

Collects remote joins from the a QueryDB if any, and transforms the selection to add new join fields where those occured.

Returns the transformed selection set, in which remote fields have been inserted, and for which the r type is now Void.

getRemoteJoinsMutationDB :: forall (b :: BackendType). Backend b => MutationDB b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b) -> (MutationDB b Void (UnpreparedValue b), Maybe RemoteJoins)

Collects remote joins from the a MutationDB if any, and transforms the selection to add new join fields where those occured.

Returns the transformed selection set, in which remote fields have been inserted, and for which the r type is now Void.

getRemoteJoinsActionQuery :: ActionQuery (RemoteRelationshipField UnpreparedValue) -> (ActionQuery Void, Maybe RemoteJoins)

getRemoteJoinsActionMutation :: ActionMutation (RemoteRelationshipField UnpreparedValue) -> (ActionMutation Void, Maybe RemoteJoins)

getRemoteJoinsGraphQLField :: GraphQLField (RemoteRelationshipField UnpreparedValue) var -> (GraphQLField Void var, Maybe RemoteJoins)