Hasura.GraphQL.Parser.Internal.Scalars

Description

This module defines all backend-agnostic scalars we use throughout the schema. This includes GraphQL scalars, and several other custom ones.

Documentation

boolean :: MonadParse m => Parser origin 'Both m Bool

int :: MonadParse m => Parser origin 'Both m Int32

float :: MonadParse m => Parser origin 'Both m Double

string :: MonadParse m => Parser origin 'Both m Text

identifier :: MonadParse m => Parser origin 'Both m Text

As an input type, any string or integer input value should be coerced to ID as Text https://spec.graphql.org/June2018/#sec-ID

uuid :: MonadParse m => Parser origin 'Both m UUID

json :: MonadParse m => Parser origin 'Both m Value

jsonb :: MonadParse m => Parser origin 'Both m Value

nonNegativeInt :: MonadParse m => Parser origin 'Both m Int32

Additional validation on integers. We do keep the same type name in the schema for backwards compatibility. TODO: when we can do a breaking change, we can rename the type to NonNegativeInt.

bigInt :: MonadParse m => Parser origin 'Both m Int64

GraphQL ints are 32-bit integers; but in some places we want to accept bigger ints. To do so, we declare a cusom scalar that can represent 64-bit ints, which accepts both int literals and string literals. We do keep the same type name in the schema for backwards compatibility. TODO: when we can do a breaking change, we can rename the type to BigInt.

scientific :: MonadParse m => Parser origin 'Both m Scientific

Parser for Scientific. Certain backends like BigQuery support Decimal/BigDecimal and need an arbitrary precision number.

jsonScalar :: MonadParse m => Name -> Maybe Description -> Parser origin 'Both m Value

Creates a parser that transforms its input into a JSON value. valueToJSON does properly unpack variables.

mkScalar :: MonadParse m => Name -> (InputValue Variable -> m a) -> Parser origin 'Both m a

typeNamed :: Name -> Maybe Description -> Type origin 'Both

scientificToInteger :: (MonadParse m, Integral i, Bounded i) => Scientific -> m i

scientificToFloat :: (MonadParse m, RealFloat f) => Scientific -> m f

parseJSON :: (MonadParse m, FromJSON b) => Value -> m b