Hasura.GraphQL.Parser.Schema

Description

Types for representing a GraphQL schema.

Documentation

data Kind

GraphQL types are divided into two classes: input types and output types. The GraphQL spec does not use the word “kind” to describe these classes, but it’s an apt term.

Some GraphQL types can be used at either kind, so we also include the Both kind, the superkind of both Input and Output. The <: class provides kind subsumption constraints.

For more details, see http://spec.graphql.org/June2018/#sec-Input-and-Output-Types.

Constructors

Both
see Note [The 'Both kind]

Input

Output

data k1 :<: k2 where

Evidence for <:.

Constructors

KRefl :: k :<: k

KBoth :: k :<: 'Both

class k1 <: k2 where

Kind subsumption. The GraphQL kind hierarchy is extremely simple:

    Both
    /  \
Input  Output

Various functions in this module use <: to allow Both to be used in places where Input or Output would otherwise be expected.

Methods

subKind :: k1 :<: k2

data Nullability

Constructors

Nullable

NonNullable

isNullable :: Nullability -> Bool

data Type origin k

Constructors

TNamed Nullability (Definition origin (TypeInfo origin k))

TList Nullability (Type origin k)

onTypeDef :: (forall a. Definition origin a -> Definition origin a) -> Type origin k -> Type origin k

Adjust the Definition underlying a Type

eqType :: Type origin k1 -> Type origin k2 -> Bool

Like ==, but can compare Types of different kinds.

typeNullability :: Type origin k -> Nullability

nullableType :: Type origin k -> Type origin k

nonNullableType :: Type origin k -> Type origin k

toGraphQLType :: Type origin k -> GType

data InputObjectInfo origin

Invariant: the list is sorted by dName

Constructors

InputObjectInfo__ ~[Definition origin (InputFieldInfo origin)]

pattern InputObjectInfo :: [Definition origin (InputFieldInfo origin)] -> InputObjectInfo origin

data ObjectInfo origin

Invariant: the lists are sorted by dName, maintained via pattern synonyms

Constructors

ObjectInfo__

Fields

_oiFields :: ~[Definition origin (FieldInfo origin)]

The fields that this object has. This consists of the fields of the interfaces that it implements, as well as any additional fields.

_oiImplements :: ~[Definition origin (InterfaceInfo origin)]

The interfaces that this object implements (inheriting all their fields). See Note [The interfaces story] for more details.

pattern ObjectInfo :: [Definition origin (FieldInfo origin)] -> [Definition origin (InterfaceInfo origin)] -> ObjectInfo origin

data InterfaceInfo origin

Type information for a GraphQL interface; see Note [The interfaces story] for more details.

Note: in the current working draft of the GraphQL specification (> June 2018), interfaces may implement other interfaces, but we currently don't support this.

Invariant: the lists are sorted by dName, maintained via pattern synonyms

Constructors

InterfaceInfo__

Fields

_iiFields :: ~[Definition origin (FieldInfo origin)]

Fields declared by this interface. Every object implementing this interface must include those fields.

_iiPossibleTypes :: ~[Definition origin (ObjectInfo origin)]

Objects that implement this interface. See Note [The interfaces story] for why we include that information here.

pattern InterfaceInfo :: [Definition origin (FieldInfo origin)] -> [Definition origin (ObjectInfo origin)] -> InterfaceInfo origin

data UnionInfo origin

Invariant: the list is sorted by dName

Constructors

UnionInfo__

Fields

_uiPossibleTypes :: ~[Definition origin (ObjectInfo origin)]

The member object types of this union.

pattern UnionInfo :: [Definition origin (ObjectInfo origin)] -> UnionInfo origin

data TypeInfo origin k where

Constructors

TIScalar :: TypeInfo origin 'Both

TIEnum__ :: NonEmpty (Definition origin EnumValueInfo) -> TypeInfo origin 'Both
Invariant: the NonEmpty is sorted by dName

TIInputObject :: InputObjectInfo origin -> TypeInfo origin 'Input

TIObject :: ObjectInfo origin -> TypeInfo origin 'Output

TIInterface :: InterfaceInfo origin -> TypeInfo origin 'Output

TIUnion :: UnionInfo origin -> TypeInfo origin 'Output

pattern TIEnum :: forall origin (k :: Kind). () => k ~ 'Both => NonEmpty (Definition origin EnumValueInfo) -> TypeInfo origin k

eqTypeInfo :: TypeInfo origin k1 -> TypeInfo origin k2 -> Bool

Like ==, but can compare TypeInfos of different kinds.

getTypeInfo :: Type origin k -> Definition origin (TypeInfo origin k)

getObjectInfo :: Type origin k -> Maybe (Definition origin (ObjectInfo origin))

getInterfaceInfo :: Type origin k -> Maybe (Definition origin (InterfaceInfo origin))

data SomeDefinitionTypeInfo origin

Constructors

forall k. SomeDefinitionTypeInfo (Definition origin (TypeInfo origin k))

data Definition origin a

Constructors

Definition

Fields

dName :: Name

dDescription :: Maybe Description

dOrigin :: Maybe origin

What piece of metadata was this fragment of GraphQL type information from? See also Parser.

Nothing can represent a couple of scenarios: 1. This is a native part of the GraphQL spec, e.g. the __Type introspection type 2. This is a native part of HGE, e.g. our scalar types and Relay-related types 3. We don't have a clear origin, because a. Semantically there is no clear origin because it arose from the combination of several things b. We generated this Definition in a context where origin information was no longer in scope

Maybe, at some point, it makes sense to represent the above options more accurately in the type of dOrigin.

dDirectives :: [Directive Void]

The directives for this object.

dInfo :: ~a

Lazy to allow mutually-recursive type definitions.

eqByName :: [Definition origin a] -> [Definition origin a] -> Bool

equivalent to, but faster than...

map dName x == map dName y

data EnumValueInfo

Enum values have no extra information except for the information common to all definitions, so this is just a placeholder for use as Definition EnumValueInfo.

Constructors

EnumValueInfo

data InputFieldInfo origin

Constructors

forall k.'Input <: k => InputFieldInfo (Type origin k) (Maybe (Value Void))
An input field with a type and possibly a default value. If a default value is provided, it should be a valid value for the type.

Note that a default value of VNull is subtly different from having no default value at all. If no default value is provided (i.e. Nothing), the GraphQL specification allows distinguishing provided null values from values left completely absent; see Note [The value of omitted fields] in Hasura.GraphQL.Parser.Internal.Parser.

data FieldInfo origin

Constructors

forall k.'Output <: k => FieldInfo

Fields

fArguments :: [Definition origin (InputFieldInfo origin)]

fType :: Type origin k

data DirectiveInfo origin

This type represents the directives information to be served over GraphQL introspection

Constructors

DirectiveInfo

Fields

diName :: !Name

diDescription :: !(Maybe Description)

diArguments :: ![Definition origin (InputFieldInfo origin)]

diLocations :: ![DirectiveLocation]

data Schema origin

This type contains all the information needed to efficiently serve GraphQL introspection queries. It corresponds to the GraphQL __Schema type defined in 4.5 Schema Introspection http://spec.graphql.org/June2018/#sec-Introspection. See also Note [Basics of introspection schema generation].

Constructors

Schema

Fields

sDescription :: Maybe Description

sTypes :: HashMap Name (SomeDefinitionTypeInfo origin)

sQueryType :: Type origin 'Output

sMutationType :: Maybe (Type origin 'Output)

sSubscriptionType :: Maybe (Type origin 'Output)

sDirectives :: [DirectiveInfo origin]

data TypeDefinitionsWrapper origin where

Constructors

TypeDefinitionsWrapper :: HasTypeDefinitions origin a => a -> TypeDefinitionsWrapper origin

collectTypeDefinitions :: HasTypeDefinitions origin a => a -> Either (ConflictingDefinitions origin) (HashMap Name (SomeDefinitionTypeInfo origin))

Recursively collects all type definitions accessible from the given value, attempting to detect any conflicting defintions that may have made it this far (See ConflictingDefinitions for details).

newtype TypeOriginStack

A path through Definition, accumulated in accumulateTypeDefinitions only to power ConflictingDefinitions in the error case.

Constructors

TypeOriginStack [Name]

typeOriginRecurse :: Name -> TypeOriginStack -> TypeOriginStack

typeRootRecurse :: Name -> TypeOriginStack -> TypeOriginStack

data ConflictingDefinitions origin

NOTE: it's not clear exactly where we'd get conflicting definitions at the point collectTypeDefinitions is called, but conflicting names from different data sources is apparently one place (TODO some tests that excercise this).

ALSO NOTE: it's difficult to see in isolation how or if this check is correct since Definition is cyclic and has no accomodations for observable sharing (formerly it had Uniques; see commit history and discussion in #3685). The check relies on dodgy Eq instances for the types that make up the Definition graph (see e.g. instance Eq ObjectInfo).

See Note [Collecting types from the GraphQL schema]

Constructors

ConflictingDefinitions (SomeDefinitionTypeInfo origin, TypeOriginStack) (SomeDefinitionTypeInfo origin, NonEmpty TypeOriginStack)
Type collection has found at least two types with the same name.

newtype TypeAccumulation origin a

Although the majority of graphql-engine is written in terms of abstract mtl-style effect monads, we figured out that this particular codepath is quite hot, and that mtl has a measurable negative effect for accumulating types from the schema, both in profiling and in benchmarking. Using an explicit transformers-style effect stack seems to overall memory usage by about 3-7%.

Constructors

TypeAccumulation

Fields

runTypeAccumulation :: ReaderT TypeOriginStack (StateT (HashMap Name (SomeDefinitionTypeInfo origin, NonEmpty TypeOriginStack)) (ExceptT (ConflictingDefinitions origin) Identity)) a

class HasTypeDefinitions origin a where

Methods

accumulateTypeDefinitions :: a -> TypeAccumulation origin ()

Recursively accumulates all type definitions accessible from the given value. This is done statefully to avoid infinite loops arising from recursive type definitions; see Note [Tying the knot] in Hasura.GraphQL.Parser.Class.