Hasura.GraphQL.RemoteServer

Documentation

introspectionQuery :: GQLReqOutgoing

validateSchemaCustomizations :: forall m. MonadError QErr m => RemoteSchemaCustomizer -> RemoteSchemaIntrospection -> m ()

validateSchemaCustomizationsConsistent :: forall m. MonadError QErr m => RemoteSchemaCustomizer -> RemoteSchemaIntrospection -> m ()

validateSchemaCustomizationsDistinct :: forall m. MonadError QErr m => RemoteSchemaCustomizer -> RemoteSchemaIntrospection -> m ()

parseIntrospectionResult :: Value -> Maybe IntrospectionResult

fetchRemoteSchema :: forall m. (MonadIO m, MonadError QErr m, MonadTrace m) => Environment -> Manager -> RemoteSchemaName -> ValidatedRemoteSchemaDef -> m RemoteSchemaCtx

Make an introspection query to the remote graphql server for the data we need to present and stitch the remote schema. This powers add_remote_schema, and also is called by schema cache rebuilding code in Hasura.RQL.DDL.Schema.Cache.

newtype FromIntrospection a

Parsing the introspection query result. We use this newtype wrapper to avoid orphan instances and parse JSON in the way that we need for GraphQL introspection results.

Constructors

FromIntrospection

Fields

fromIntrospection :: a

pErr :: MonadFail m => Text -> m a

kindErr :: MonadFail m => Text -> Text -> m a

objectWithoutNullValues :: [Pair] -> Value

toObjectTypeDefinition :: Name -> ObjectTypeDefinition InputValueDefinition

execRemoteGQ

Arguments

:: (MonadIO m, MonadError QErr m, MonadTrace m)

=> Environment

-> Manager

-> UserInfo

-> [Header]

-> ValidatedRemoteSchemaDef

-> GQLReqOutgoing

-> m (DiffTime, [Header], ByteString)
Returns the response body and headers, along with the time taken for the HTTP request to complete

identityCustomizer :: RemoteSchemaCustomizer

getCustomizer :: IntrospectionResult -> Maybe RemoteSchemaCustomization -> RemoteSchemaCustomizer

throwRemoteSchema :: QErrM m => Text -> m a

throwRemoteSchemaWithInternal :: (QErrM m, ToJSON a) => Text -> a -> m b

throwRemoteSchemaHttp :: QErrM m => URI -> HttpException -> m a

newtype RemoteSchemaIntrospectionErr

Constructors

RemoteSchemaIntrospectionErr

Fields

remoteSchemaIntrospectionErr :: String

errorE :: String -> c

getSchemaIntrospection :: HashMap RoleName (RoleContext GQLContext) -> Maybe RemoteSchemaIntrospection

nameToVariable :: Name -> Variable