Hasura.GraphQL.RemoteServer

Documentation

fetchRemoteSchema :: (MonadIO m, MonadError QErr m, MonadTrace m, ProvidesNetwork m) => Environment -> SchemaSampledFeatureFlags -> ValidatedRemoteSchemaDef -> m (IntrospectionResult, ByteString, RemoteSchemaInfo)

Make an introspection query to the remote graphql server for the data we need to present and stitch the remote schema. This powers add_remote_schema, and also is called by schema cache rebuilding code in Hasura.RQL.DDL.Schema.Cache.

stitchRemoteSchema :: (MonadIO m, MonadError QErr m) => SchemaSampledFeatureFlags -> ByteString -> ValidatedRemoteSchemaDef -> m (IntrospectionResult, RemoteSchemaInfo)

Parses the remote schema introspection result, and check whether it looks like it's a valid GraphQL endpoint even under the configured customization.

execRemoteGQ

Arguments

:: (MonadIO m, MonadError QErr m, MonadTrace m, ProvidesNetwork m)

=> Environment

-> HttpPropagator

-> UserInfo

-> [Header]

-> ValidatedRemoteSchemaDef

-> GQLReqOutgoing

-> m (DiffTime, [Header], ByteString)
Returns the response body and headers, along with the time taken for the HTTP request to complete

Sends a GraphQL query to the given server.

newtype FromIntrospection a

Parsing the introspection query result. We use this newtype wrapper to avoid orphan instances and parse JSON in the way that we need for GraphQL introspection results.

Constructors

FromIntrospection

Fields

fromIntrospection :: a