Hasura.GraphQL.RemoteServer

Documentation

fetchRemoteSchema :: forall m. (MonadIO m, MonadError QErr m, MonadTrace m) => Environment -> Manager -> RemoteSchemaName -> ValidatedRemoteSchemaDef -> m RemoteSchemaCtx

Make an introspection query to the remote graphql server for the data we need to present and stitch the remote schema. This powers add_remote_schema, and also is called by schema cache rebuilding code in Hasura.RQL.DDL.Schema.Cache.

execRemoteGQ

Arguments

:: (MonadIO m, MonadError QErr m, MonadTrace m)

=> Environment

-> Manager

-> UserInfo

-> [Header]

-> ValidatedRemoteSchemaDef

-> GQLReqOutgoing

-> m (DiffTime, [Header], ByteString)
Returns the response body and headers, along with the time taken for the HTTP request to complete

Sends a GraphQL query to the given server.