Hasura.GraphQL.Schema.BoolExp

Documentation

boolExp :: forall b r m n. MonadBuildSchema b r m n => SourceInfo b -> TableInfo b -> m (Parser 'Input n (AnnBoolExp b (UnpreparedValue b)))

input type_bool_exp {
  _or: [type_bool_exp!]
  _and: [type_bool_exp!]
  _not: type_bool_exp
  column: type_comparison_exp
  ...
}

mkBoolOperator

Arguments

:: (MonadParse n, 'Input <: k)

=> NamingCase
Naming convention for the field

-> Bool
shall this be collapsed to True when null is given?

-> GQLNameIdentifier
name of this operator

-> Maybe Description
optional description

-> Parser k n a
parser for the underlying value

-> InputFieldsParser n (Maybe a)

equalityOperators

Arguments

:: (MonadParse n, 'Input <: k)

=> NamingCase

-> Bool
shall this be collapsed to True when null is given?

-> Parser k n (UnpreparedValue b)
parser for one column value

-> Parser k n (UnpreparedValue b)
parser for a list of column values

-> [InputFieldsParser n (Maybe (OpExpG b (UnpreparedValue b)))]

comparisonOperators

Arguments

:: (MonadParse n, 'Input <: k)

=> NamingCase

-> Bool
shall this be collapsed to True when null is given?

-> Parser k n (UnpreparedValue b)
parser for one column value

-> [InputFieldsParser n (Maybe (OpExpG b (UnpreparedValue b)))]