Hasura.GraphQL.Schema.Common

Documentation

data SchemaOptions

Aggregation of options required to build the schema.

Constructors

SchemaOptions

Fields

soStringifyNum :: StringifyNumbers

how numbers should be represented in the ouput; this option does not influence the schema, but is bundled in the resulting IR

soDangerousBooleanCollapse :: Bool

should boolean fields be collapsed to True when null is given?

soQueryType :: GraphQLQueryType

what kind of schema is being built (Hasura or Relay)

soFunctionPermsContext :: FunctionPermissionsCtx

whether function permissions should be inferred

soRemoteSchemaPermsCtx :: RemoteSchemaPermsCtx

whether remote schema permissions are enabled

soOptimizePermissionFilters :: Bool

True when we should attempt to use experimental SQL optimization passes

data SchemaContext

Aggregation of contextual information required to build the schema.

Constructors

SchemaContext

Fields

scSourceCache :: SourceCache

the set of all sources (TODO: remove this?)

scRemoteRelationshipParserBuilder :: RemoteRelationshipParserBuilder

how to process remote relationships

type MonadBuildSchemaBase r m n = (MonadError QErr m, MonadReader r m, MonadSchema n m, Has SchemaOptions r, Has SchemaContext r, Has RoleName r, Has MkTypename r, Has MkRootFieldName r, Has CustomizeRemoteFieldName r, Has NamingCase r)

The set of common constraints required to build the schema.

newtype RemoteRelationshipParserBuilder

How a remote relationship field should be processed when building a schema. Injecting this function from the top level avoids having to know how to do top-level dispatch from deep within the schema code.

Note: the inner function type uses an existential qualifier: it is expected that the given function will work for _any_ monad m that has the relevant constraints. This prevents us from passing a function that is specfic to the monad in which the schema construction will run, but avoids having to propagate type annotations to each call site.

Constructors

RemoteRelationshipParserBuilder (forall lhsJoinField r n m. MonadBuildSchemaBase r m n => RemoteFieldInfo lhsJoinField -> m (Maybe [FieldParser n (RemoteRelationshipField UnpreparedValue)]))

ignoreRemoteRelationship :: RemoteRelationshipParserBuilder

A RemoteRelationshipParserBuilder that ignores the field altogether, that can be used in tests or to build a source or remote schema in isolation.

retrieve :: (MonadReader r m, Has a r) => (a -> b) -> m b

type SelectExp b = AnnSimpleSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type StreamSelectExp b = AnnSimpleStreamSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AggSelectExp b = AnnAggregateSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type ConnectionSelectExp b = ConnectionSelect b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type SelectArgs b = SelectArgsG b (UnpreparedValue b)

type SelectStreamArgs b = SelectStreamArgsG b (UnpreparedValue b)

type TablePerms b = TablePermG b (UnpreparedValue b)

type AnnotatedFields b = AnnFieldsG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AnnotatedField b = AnnFieldG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type ConnectionFields b = ConnectionFields b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type EdgeFields b = EdgeFields b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AnnotatedActionFields = ActionFieldsG (RemoteRelationshipField UnpreparedValue)

type AnnotatedActionField = ActionFieldG (RemoteRelationshipField UnpreparedValue)

data RemoteSchemaParser n

Constructors

RemoteSchemaParser

Fields

piQuery :: [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

piMutation :: Maybe [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

piSubscription :: Maybe [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

getTableRoles :: BackendSourceInfo -> [RoleName]

askTableInfo :: forall b m. (Backend b, MonadError QErr m) => SourceInfo b -> TableName b -> m (TableInfo b)

Looks up table information for the given table name. This function should never fail, since the schema cache construction process is supposed to ensure all dependencies are resolved. TODO: deduplicate this with CacheRM.

data Scenario

Whether the request is sent with `x-hasura-use-backend-only-permissions` set to true.

Constructors

Backend

Frontend

textToName :: MonadError QErr m => Text -> m Name

partialSQLExpToUnpreparedValue :: PartialSQLExp b -> UnpreparedValue b

mapField :: Functor m => InputFieldsParser m (Maybe a) -> (a -> b) -> InputFieldsParser m (Maybe b)

parsedSelectionsToFields

Arguments

:: (Text -> a)
how to handle __typename fields

-> InsOrdHashMap Name (ParsedSelection a)

-> Fields a

numericAggOperators :: [Name]

comparisonAggOperators :: [Name]

data NodeIdVersion

Constructors

NIVersion1

nodeIdVersionInt :: NodeIdVersion -> Int

currentNodeIdVersion :: NodeIdVersion

mkDescriptionWith :: Maybe PGDescription -> Text -> Description

takeValidTables :: forall b. Backend b => TableCache b -> TableCache b

takeValidFunctions :: forall b. FunctionCache b -> FunctionCache b

requiredFieldParser :: (Functor n, Functor m) => (a -> b) -> m (FieldParser n a) -> m (Maybe (FieldParser n b))

optionalFieldParser :: (Functor n, Functor m) => (a -> b) -> m (Maybe (FieldParser n a)) -> m (Maybe (FieldParser n b))

mkEnumTypeName :: forall b m r. (Backend b, MonadReader r m, Has MkTypename r, MonadError QErr m) => EnumReference b -> m Name

Builds the type name for referenced enum tables.

addEnumSuffix :: (MonadReader r m, Has MkTypename r) => Name -> Maybe Name -> m Name