Hasura.GraphQL.Schema.Common

Documentation

type MonadBuildSchemaBase r m n = (MonadError QErr m, MonadReader r m, MonadSchema n m, Has RoleName r, Has SourceCache r, Has QueryContext r, Has MkTypename r, Has MkRootFieldName r, Has CustomizeRemoteFieldName r, Has RemoteSchemaMap r)

the set of common constraints required to build the schema

type SelectExp b = AnnSimpleSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type StreamSelectExp b = AnnSimpleStreamSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AggSelectExp b = AnnAggregateSelectG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type ConnectionSelectExp b = ConnectionSelect b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type SelectArgs b = SelectArgsG b (UnpreparedValue b)

type SelectStreamArgs b = SelectStreamArgsG b (UnpreparedValue b)

type TablePerms b = TablePermG b (UnpreparedValue b)

type AnnotatedFields b = AnnFieldsG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AnnotatedField b = AnnFieldG b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type ConnectionFields b = ConnectionFields b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type EdgeFields b = EdgeFields b (RemoteRelationshipField UnpreparedValue) (UnpreparedValue b)

type AnnotatedActionFields = ActionFieldsG (RemoteRelationshipField UnpreparedValue)

type AnnotatedActionField = ActionFieldG (RemoteRelationshipField UnpreparedValue)

data RemoteSchemaParser n

Constructors

RemoteSchemaParser

Fields

piQuery :: [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

piMutation :: Maybe [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

piSubscription :: Maybe [FieldParser n (NamespacedField (RemoteSchemaRootField (RemoteRelationshipField UnpreparedValue) RemoteSchemaVariable))]

data QueryContext

Constructors

QueryContext

Fields

qcStringifyNum :: StringifyNumbers

qcDangerousBooleanCollapse :: Bool

should boolean fields be collapsed to True when null is given?

qcQueryType :: GraphQLQueryType

qcFunctionPermsContext :: FunctionPermissionsCtx

qcRemoteSchemaPermsCtx :: RemoteSchemaPermsCtx

qcOptimizePermissionFilters :: Bool

True when we should attempt to use experimental SQL optimization passes

getTableRoles :: BackendSourceInfo -> [RoleName]

askTableInfo :: forall b r m. (Backend b, MonadError QErr m, MonadReader r m, Has SourceCache r) => SourceName -> TableName b -> m (TableInfo b)

Looks up table information for the given table name. This function should never fail, since the schema cache construction process is supposed to ensure all dependencies are resolved. TODO: deduplicate this with CacheRM.

data Scenario

Whether the request is sent with `x-hasura-use-backend-only-permissions` set to true.

Constructors

Backend

Frontend

textToName :: MonadError QErr m => Text -> m Name

partialSQLExpToUnpreparedValue :: PartialSQLExp b -> UnpreparedValue b

mapField :: Functor m => InputFieldsParser m (Maybe a) -> (a -> b) -> InputFieldsParser m (Maybe b)

parsedSelectionsToFields

Arguments

:: (Text -> a)
how to handle __typename fields

-> InsOrdHashMap Name (ParsedSelection a)

-> Fields a

numericAggOperators :: [Name]

comparisonAggOperators :: [Name]

data NodeIdVersion

Constructors

NIVersion1

nodeIdVersionInt :: NodeIdVersion -> Int

currentNodeIdVersion :: NodeIdVersion

mkDescriptionWith :: Maybe PGDescription -> Text -> Description

takeValidTables :: forall b. Backend b => TableCache b -> TableCache b

takeValidFunctions :: forall b. FunctionCache b -> FunctionCache b

requiredFieldParser :: (Functor n, Functor m) => (a -> b) -> m (FieldParser n a) -> m (Maybe (FieldParser n b))

optionalFieldParser :: (Functor n, Functor m) => (a -> b) -> m (Maybe (FieldParser n a)) -> m (Maybe (FieldParser n b))

mkEnumTypeName :: forall b m r. (Backend b, MonadReader r m, Has MkTypename r, MonadError QErr m) => EnumReference b -> m Name

Builds the type name for referenced enum tables.

addEnumSuffix :: (MonadReader r m, Has MkTypename r) => Name -> Maybe Name -> m Name

getIndirectDependencies :: forall m. (QErrM m, CacheRM m) => SourceName -> m [SchemaObjId]

Return the indirect dependencies on a source. We return a [SchemaObjId] instead of [SourceObjId], because the latter has no reference to the source. Due to which we won't be able to drop the dependencies of a different source.

For eg: Say we have two sources: "source1" and "source2". "source1" has a remote relationship with "source2". So when "source2" is dropped, the remote relationship from "source1" also has to be dropped. If we were to only pass the SourceObjId, we wouldn't have know which source that pertains to. Hence we'll not be able to drop the remote relationship. That's why we return the [SchemaObjId] so that, we can drop the SourceObjId from the correct source.

purgeDependencies :: MonadError QErr m => [SchemaObjId] -> WriterT MetadataModifier m ()

dropRemoteSchemaRemoteRelationshipInMetadata :: RemoteSchemaName -> Name -> RelName -> MetadataModifier