Hasura.GraphQL.Schema.Parser

Description

In Parser, the Definition type has a dOrigin field that allows to track where a fragment of GraphQL type information comes from. This is useful for error reporting and internal repair mechanisms such as inconsistency tracking.

Morally, within the HGE codebase, this origin is always MetadataObjId. However, in order to avoid an import of RQL from Parser, the dOrigin has been defined through a type parameter of Definition. This type parameter then has to get threaded through throughout the Parser module hierarchy, so that it ends up in a lot of types. This is very noisy.

In order to avoid the noise of this type parameter, which really only has one value, and is really only used in one type, this module erases the type parameter by filling in the desired value, exporting type synonyms of the now-fixed notion of "origin". So most modules in the HGE codebase should import this module rather than Parser.

Documentation

type FieldParser = FieldParser MetadataObjId

pattern FieldParser :: Definition origin (FieldInfo origin) -> (Field NoFragments Variable -> m a) -> FieldParser origin m a

type InputFieldsParser = InputFieldsParser MetadataObjId

pattern InputFieldsParser :: [Definition origin (InputFieldInfo origin)] -> (HashMap Name (InputValue Variable) -> m a) -> InputFieldsParser origin m a

type InputObjectInfo = InputObjectInfo MetadataObjId

pattern InputObjectInfo :: [Definition origin (InputFieldInfo origin)] -> InputObjectInfo origin

type Parser = Parser MetadataObjId

pattern Parser :: ~(Type origin k) -> (ParserInput k -> m a) -> Parser origin k m a

type Schema = Schema MetadataObjId

pattern Schema :: Maybe Description -> HashMap Name (SomeDefinitionTypeInfo origin) -> Type origin 'Output -> Maybe (Type origin 'Output) -> Maybe (Type origin 'Output) -> [DirectiveInfo origin] -> Schema origin

type ConflictingDefinitions = ConflictingDefinitions MetadataObjId

pattern ConflictingDefinitions :: (SomeDefinitionTypeInfo origin, TypeOriginStack) -> (SomeDefinitionTypeInfo origin, NonEmpty TypeOriginStack) -> ConflictingDefinitions origin

Type collection has found at least two types with the same name.

type Definition = Definition MetadataObjId

pattern Definition :: Name -> Maybe Description -> Maybe origin -> [Directive Void] -> ~a -> Definition origin a

type Type = Type MetadataObjId

type TypeInfo = TypeInfo MetadataObjId

type Directive = Directive MetadataObjId

pattern Directive :: DirectiveInfo origin -> Bool -> (Directive Variable -> m a) -> Directive origin m

type DirectiveInfo = DirectiveInfo MetadataObjId

pattern DirectiveInfo :: !Name -> !(Maybe Description) -> ![Definition origin (InputFieldInfo origin)] -> ![DirectiveLocation] -> DirectiveInfo origin

type FieldInfo = FieldInfo MetadataObjId

pattern FieldInfo :: () => 'Output <: k => [Definition origin (InputFieldInfo origin)] -> Type origin k -> FieldInfo origin

type InputFieldInfo = InputFieldInfo MetadataObjId

pattern InputFieldInfo :: () => 'Input <: k => Type origin k -> Maybe (Value Void) -> InputFieldInfo origin

An input field with a type and possibly a default value. If a default value is provided, it should be a valid value for the type.

Note that a default value of VNull is subtly different from having no default value at all. If no default value is provided (i.e. Nothing), the GraphQL specification allows distinguishing provided null values from values left completely absent; see Note [The value of omitted fields] in Hasura.GraphQL.Parser.Internal.Parser.

type HasTypeDefinitions = HasTypeDefinitions MetadataObjId

type SomeDefinitionTypeInfo = SomeDefinitionTypeInfo MetadataObjId

pattern SomeDefinitionTypeInfo :: Definition origin (TypeInfo origin k) -> SomeDefinitionTypeInfo origin

type TypeDefinitionsWrapper = TypeDefinitionsWrapper MetadataObjId

pattern TypeDefinitionsWrapper :: () => HasTypeDefinitions a => a -> TypeDefinitionsWrapper

In order to aid type inference and type checking, we define this pattern synonym (an actual one) which restricts TypeDefinitionsWrapper to have MetadataObjId set for its origin type parameter.

data ParseErrorCode

Constructors

ValidationFailed

ParseFailed

ConflictingDefinitionsError

NotSupported

toQErr :: MonadError QErr m => Either ParseError a -> m a

uuid :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m UUID

scientific :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Scientific

Parser for Scientific. Certain backends like BigQuery support Decimal/BigDecimal and need an arbitrary precision number.

json :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Value

object :: forall (m :: Type -> Type) origin a. MonadParse m => Name -> Maybe Description -> InputFieldsParser origin m a -> Parser origin 'Input m a

string :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Text

list :: forall origin (k :: Kind) (m :: Type -> Type) a. (MonadParse m, 'Input <: k) => Parser origin k m a -> Parser origin k m [a]

int :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Int32

float :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Double

data Kind

GraphQL types are divided into two classes: input types and output types. The GraphQL spec does not use the word “kind” to describe these classes, but it’s an apt term.

Some GraphQL types can be used at either kind, so we also include the Both kind, the superkind of both Input and Output. The <: class provides kind subsumption constraints.

For more details, see http://spec.graphql.org/June2018/#sec-Input-and-Output-Types.

Constructors

Both
see Note [The 'Both kind]

Input

Output

runParser :: forall origin (k :: Kind) m a. Parser origin k m a -> ParserInput k -> m a

multiple :: forall (m :: Type -> Type) origin a. Parser origin 'Output m a -> Parser origin 'Output m a

data ParseError

Constructors

ParseError

Fields

pePath :: JSONPath

peMessage :: ErrorMessage

peCode :: ParseErrorCode

bind :: forall m origin (k :: Kind) a b. Monad m => Parser origin k m a -> (a -> m b) -> Parser origin k m b infixl 1

class HasName a where

Methods

getName :: a -> Name

eqType :: forall origin (k1 :: Kind) (k2 :: Kind). Type origin k1 -> Type origin k2 -> Bool

Like ==, but can compare Types of different kinds.

parseError :: MonadParse m => ErrorMessage -> m a

data Nullability

Constructors

Nullable

NonNullable

data Variable

Constructors

Variable

Fields

vInfo :: VariableInfo

vType :: GType

vValue :: Maybe (InputValue Void)

The following cases are distinguished:

A JSON value (including null) was provided: Just (JSONValue ...)

No JSON value was provided, but a default value exists: Just (GraphQLValue ...)

No JSON value was provided, and no default value exists: Nothing

field :: forall (m :: Type -> Type) (k :: Kind) origin a. (MonadParse m, 'Input <: k) => Name -> Maybe Description -> Parser origin k m a -> InputFieldsParser origin m a

Creates a parser for an input field. For nullable value parsers, if an input field is omitted, the provided parser will be called with a value of null.

selectionSet :: forall (m :: Type -> Type) origin a. MonadParse m => Name -> Maybe Description -> [FieldParser origin m a] -> Parser origin 'Output m (InsOrdHashMap Name (ParsedSelection a))

A variant of selectionSetObject which doesn't implement any interfaces

enum :: forall (m :: Type -> Type) origin a. MonadParse m => Name -> Maybe Description -> NonEmpty (Definition origin EnumValueInfo, a) -> Parser origin 'Both m a

nullable :: forall origin (k :: Kind) (m :: Type -> Type) a. (MonadParse m, 'Input <: k) => Parser origin k m a -> Parser origin k m (Maybe a)

identifier :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Text

jsonb :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Value

class (Monad m, Typeable m) => MonadParse (m :: Type -> Type) where

A class that provides functionality for parsing GraphQL queries, i.e. running a fully-constructed Parser.

Methods

withKey :: JSONPathElement -> m a -> m a

parseErrorWith :: ParseErrorCode -> ErrorMessage -> m a

Not the full power of MonadError because parse errors cannot be caught.

withPath :: MonadParse m => JSONPath -> m a -> m a

data Parse a

runParse :: MonadError ParseError m => Parse a -> m a

sDirectives :: Schema origin -> [DirectiveInfo origin]

sSubscriptionType :: Schema origin -> Maybe (Type origin 'Output)

sMutationType :: Schema origin -> Maybe (Type origin 'Output)

sQueryType :: Schema origin -> Type origin 'Output

sTypes :: Schema origin -> HashMap Name (SomeDefinitionTypeInfo origin)

sDescription :: Schema origin -> Maybe Description

diLocations :: DirectiveInfo origin -> [DirectiveLocation]

diArguments :: DirectiveInfo origin -> [Definition origin (InputFieldInfo origin)]

diDescription :: DirectiveInfo origin -> Maybe Description

diName :: DirectiveInfo origin -> Name

fType :: FieldInfo origin -> Type origin k

fArguments :: FieldInfo origin -> [Definition origin (InputFieldInfo origin)]

data EnumValueInfo

Enum values have no extra information except for the information common to all definitions, so this is just a placeholder for use as Definition EnumValueInfo.

Constructors

EnumValueInfo

dInfo :: Definition origin a -> a

Lazy to allow mutually-recursive type definitions.

dDirectives :: Definition origin a -> [Directive Void]

The directives for this object.

dOrigin :: Definition origin a -> Maybe origin

What piece of metadata was this fragment of GraphQL type information from? See also Parser.

Nothing can represent a couple of scenarios: 1. This is a native part of the GraphQL spec, e.g. the __Type introspection type 2. This is a native part of HGE, e.g. our scalar types and Relay-related types 3. We don't have a clear origin, because a. Semantically there is no clear origin because it arose from the combination of several things b. We generated this Definition in a context where origin information was no longer in scope

Maybe, at some point, it makes sense to represent the above options more accurately in the type of dOrigin.

dDescription :: Definition origin a -> Maybe Description

dName :: Definition origin a -> Name

pattern TIScalar :: TypeInfo origin 'Both

pattern TIInputObject :: InputObjectInfo origin -> TypeInfo origin 'Input

pattern TIObject :: ObjectInfo origin -> TypeInfo origin 'Output

pattern TIInterface :: InterfaceInfo origin -> TypeInfo origin 'Output

pattern TIUnion :: UnionInfo origin -> TypeInfo origin 'Output

pattern TIEnum :: () => k ~ 'Both => NonEmpty (Definition origin EnumValueInfo) -> TypeInfo origin k

data UnionInfo origin where

Invariant: the list is sorted by dName

Bundled Patterns

pattern UnionInfo :: [Definition origin (ObjectInfo origin)] -> UnionInfo origin

data InterfaceInfo origin where

Type information for a GraphQL interface; see Note [The interfaces story] for more details.

Note: in the current working draft of the GraphQL specification (> June 2018), interfaces may implement other interfaces, but we currently don't support this.

Invariant: the lists are sorted by dName, maintained via pattern synonyms

Bundled Patterns

pattern InterfaceInfo :: [Definition origin (FieldInfo origin)] -> [Definition origin (ObjectInfo origin)] -> InterfaceInfo origin

data ObjectInfo origin where

Invariant: the lists are sorted by dName, maintained via pattern synonyms

Bundled Patterns

pattern ObjectInfo :: [Definition origin (FieldInfo origin)] -> [Definition origin (InterfaceInfo origin)] -> ObjectInfo origin

pattern TNamed :: Nullability -> Definition origin (TypeInfo origin k) -> Type origin k

pattern TList :: Nullability -> Type origin k -> Type origin k

class (k1 :: Kind) <: (k2 :: Kind) where

Kind subsumption. The GraphQL kind hierarchy is extremely simple:

    Both
    /  \
Input  Output

Various functions in this module use <: to allow Both to be used in places where Input or Output would otherwise be expected.

Methods

subKind :: k1 :<: k2

data (k1 :: Kind) :<: (k2 :: Kind) where

Evidence for <:.

Constructors

KRefl :: forall (k1 :: Kind). k1 :<: k1

KBoth :: forall (k1 :: Kind). k1 :<: 'Both

onTypeDef :: forall origin (k :: Kind). (forall a. Definition origin a -> Definition origin a) -> Type origin k -> Type origin k

Adjust the Definition underlying a Type

typeNullability :: forall origin (k :: Kind). Type origin k -> Nullability

nullableType :: forall origin (k :: Kind). Type origin k -> Type origin k

nonNullableType :: forall origin (k :: Kind). Type origin k -> Type origin k

toGraphQLType :: forall origin (k :: Kind). Type origin k -> GType

eqTypeInfo :: forall origin (k1 :: Kind) (k2 :: Kind). TypeInfo origin k1 -> TypeInfo origin k2 -> Bool

Like ==, but can compare TypeInfos of different kinds.

getTypeInfo :: forall origin (k :: Kind). Type origin k -> Definition origin (TypeInfo origin k)

getObjectInfo :: forall origin (k :: Kind). Type origin k -> Maybe (Definition origin (ObjectInfo origin))

getInterfaceInfo :: forall origin (k :: Kind). Type origin k -> Maybe (Definition origin (InterfaceInfo origin))

accumulateTypeDefinitions :: HasTypeDefinitions origin a => a -> TypeAccumulation origin ()

Recursively accumulates all type definitions accessible from the given value. This is done statefully to avoid infinite loops arising from recursive type definitions; see Note [Tying the knot] in Hasura.GraphQL.Parser.Class.

collectTypeDefinitions :: HasTypeDefinitions origin a => a -> Either (ConflictingDefinitions origin) (HashMap Name (SomeDefinitionTypeInfo origin))

Recursively collects all type definitions accessible from the given value, attempting to detect any conflicting defintions that may have made it this far (See ConflictingDefinitions for details).

data VariableInfo

Constructors

VIRequired Name

VIOptional Name (Value Void)
Unlike fields (see InputFieldInfo), nullable variables with no default value are indistinguishable from variables with a default value of null, so we don’t distinguish those cases here.

data InputValue v

See Note [Parsing variable values].

Constructors

GraphQLValue (Value v)

JSONValue Value

pParser :: Parser origin k m a -> ParserInput k -> m a

pType :: Parser origin k m a -> Type origin k

Lazy for knot-tying reasons; see Note [Tying the knot] in Hasura.GraphQL.Parser.Class.

parserType :: forall origin (k :: Kind) (m :: Type -> Type) a. Parser origin k m a -> Type origin k

ifParser :: InputFieldsParser origin m a -> HashMap Name (InputValue Variable) -> m a

ifDefinitions :: InputFieldsParser origin m a -> [Definition origin (InputFieldInfo origin)]

fieldOptional :: forall (m :: Type -> Type) (k :: Kind) origin a. (MonadParse m, 'Input <: k) => Name -> Maybe Description -> Parser origin k m a -> InputFieldsParser origin m (Maybe a)

Creates a parser for a nullable field with no default value. If the field is omitted, the provided parser will not be called. This allows a field to distinguish an omitted field from a field supplied with null (which is permitted by the GraphQL specification); see Note [The value of omitted fields] for more details.

If you want a field with a default value of null, use field with a nullable value parser, instead.

fieldOptional' :: forall (m :: Type -> Type) (k :: Kind) origin a. (MonadParse m, 'Input <: k) => Name -> Maybe Description -> Parser origin k m a -> InputFieldsParser origin m (Maybe a)

fieldWithDefault

Arguments

:: forall (m :: Type -> Type) (k :: Kind) origin a. (MonadParse m, 'Input <: k)

=> Name

-> Maybe Description

-> Value Void
default value

-> Parser origin k m a

-> InputFieldsParser origin m a

Creates a parser for an input field with the given default value. The resulting field will always be optional, even if the underlying parser rejects null values. The underlying parser is always called.

fieldWithDefault'

Arguments

:: forall (m :: Type -> Type) (k :: Kind) origin a. (MonadParse m, 'Input <: k)

=> Name

-> Maybe Description

-> Value Void
default value

-> Parser origin k m a

-> InputFieldsParser origin m a

valueToJSON :: MonadParse m => GType -> InputValue Variable -> m Value

jsonToGraphQL :: Value -> Either ErrorMessage (Value Void)

namedBoolean :: forall (m :: Type -> Type) origin. MonadParse m => Name -> Parser origin 'Both m Bool

boolean :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Bool

namedInt :: forall (m :: Type -> Type) origin. MonadParse m => Name -> Parser origin 'Both m Int32

namedFloat :: forall (m :: Type -> Type) origin. MonadParse m => Name -> Parser origin 'Both m Double

namedString :: forall (m :: Type -> Type) origin. MonadParse m => Name -> Parser origin 'Both m Text

namedIdentifier :: forall (m :: Type -> Type) origin. MonadParse m => Name -> Parser origin 'Both m Text

As an input type, any string or integer input value should be coerced to ID as Text https://spec.graphql.org/June2018/#sec-ID

nonNegativeInt :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Int32

Additional validation on integers. We do keep the same type name in the schema for backwards compatibility. TODO: when we can do a breaking change, we can rename the type to NonNegativeInt.

bigInt :: forall (m :: Type -> Type) origin. MonadParse m => Parser origin 'Both m Int64

GraphQL ints are 32-bit integers; but in some places we want to accept bigger ints. To do so, we declare a cusom scalar that can represent 64-bit ints, which accepts both int literals and string literals. We do keep the same type name in the schema for backwards compatibility. TODO: when we can do a breaking change, we can rename the type to BigInt.

jsonScalar :: forall (m :: Type -> Type) origin. MonadParse m => Name -> Maybe Description -> Parser origin 'Both m Value

Creates a parser that transforms its input into a JSON value. valueToJSON does properly unpack variables.

dParser :: Directive origin m -> Directive Variable -> m a

dAdvertised :: Directive origin m -> Bool

dDefinition :: Directive origin m -> DirectiveInfo origin

data ParsedSelection a

A single parsed field in a selection set.

Constructors

SelectField a
An ordinary field.

SelectTypename Name
The magical __typename field, implicitly available on all objects part of GraphQL introspection http://spec.graphql.org/June2018/#sec-Type-Name-Introspection.

fParser :: FieldParser origin m a -> Field NoFragments Variable -> m a

fDefinition :: FieldParser origin m a -> Definition origin (FieldInfo origin)

bindFields :: Monad m => InputFieldsParser origin m a -> (a -> m b) -> InputFieldsParser origin m b infixl 1

bindField :: Monad m => FieldParser origin m a -> (a -> m b) -> FieldParser origin m b infixl 1

handleTypename :: (Name -> a) -> ParsedSelection a -> a

wrapFieldParser :: forall (m :: Type -> Type) origin a. GType -> FieldParser origin m a -> FieldParser origin m a

Decorate a schema field with reference to given @GType

nonNullableParser :: forall (m :: Type -> Type) origin (k :: Kind) a. Parser origin k m a -> Parser origin k m a

Decorate a schema type as NON_NULL. Note that this is unsafe for 'Output parsers, in the sense that nonNullableParser doesn't (and due to the polymorphic nature of the a type parameter, can't) ensure that the provided parser provides a semantically non-null value.

nullableParser :: forall (m :: Type -> Type) origin (k :: Kind) a. Parser origin k m a -> Parser origin k m a

Mark a schema type as nullable. Syntactically speaking, this is the default, because the GraphQL spec explicitly requires use of ! to mark types as non-nullable. But many constructions in our codebase are non-nullable by default, since this matches the Haskell type system better.

Note that this is unsafe for 'Input parsers, in the sense that nullableParser doesn't ensure that the provided parser actually deals with null input values.

setParserOrigin :: forall origin (k :: Kind) (m :: Type -> Type) a. origin -> Parser origin k m a -> Parser origin k m a

Set the metadata origin of a Parser

setFieldParserOrigin :: forall (m :: Type -> Type) origin a. origin -> FieldParser origin m a -> FieldParser origin m a

Set the metadata origin of a FieldParser

setInputFieldsParserOrigin :: forall (m :: Type -> Type) origin a. origin -> InputFieldsParser origin m a -> InputFieldsParser origin m a

Set the metadata origin of the arguments in a InputFieldsParser

safeSelectionSet :: forall n (m :: Type -> Type) origin a. (MonadError ErrorMessage n, MonadParse m, Hashable origin, ToErrorValue origin) => Name -> Maybe Description -> [FieldParser origin m a] -> n (Parser origin 'Output m (InsOrdHashMap Name (ParsedSelection a)))

selectionSetObject

Arguments

:: forall (m :: Type -> Type) origin a b. MonadParse m

=> Name

-> Maybe Description

-> [FieldParser origin m a]
Fields of this object, including any fields that are required from the interfaces that it implements. Note that we can't derive those fields from the list of interfaces (next argument), because the types of the fields of the object are only required to be *subtypes* of the types of the fields of the interfaces it implements.

-> [Parser origin 'Output m b]
Interfaces implemented by this object; see Note [The interfaces story] in Hasura.GraphQL.Parser.Schema.

-> Parser origin 'Output m (InsOrdHashMap Name (ParsedSelection a))

selectionSetInterface

Arguments

:: forall (n :: Type -> Type) t origin a b. (MonadParse n, Traversable t)

=> Name

-> Maybe Description

-> [FieldParser origin n a]
Fields defined in this interface

-> t (Parser origin 'Output n b)
Parsers for the object types that implement this interface; see Note [The interfaces story] in Hasura.GraphQL.Parser.Schema for details.

-> Parser origin 'Output n (t b)

selectionSetUnion

Arguments

:: forall (n :: Type -> Type) t origin b. (MonadParse n, Traversable t)

=> Name

-> Maybe Description

-> t (Parser origin 'Output n b)
The member object types.

-> Parser origin 'Output n (t b)

selection

Arguments

:: forall (m :: Type -> Type) origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Both m b
type of the result

-> FieldParser origin m a

Builds a FieldParser for a field that does not take a subselection set, i.e. a field that returns a scalar or enum. The field’s type is taken from the provided Parser, but the Parser is not otherwise used.

See also Note [The delicate balance of GraphQL kinds] in Hasura.GraphQL.Parser.Schema.

rawSelection

Arguments

:: forall (m :: Type -> Type) origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Both m b
type of the result

-> FieldParser origin m (Maybe Name, HashMap Name (Value Variable), a)
alias provided (if any), and the arguments

subselection

Arguments

:: forall (m :: Type -> Type) origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Output m b
parser for the subselection set

-> FieldParser origin m (a, b)

Builds a FieldParser for a field that takes a subselection set, i.e. a field that returns an object.

For example, subselection name _ args fields produces schema:

name (args) { fields }

See also Note [The delicate balance of GraphQL kinds] in Hasura.GraphQL.Parser.Schema.

rawSubselection

Arguments

:: forall (m :: Type -> Type) origin a b. MonadParse m

=> Name

-> Maybe Description

-> InputFieldsParser origin m a
parser for the input arguments

-> Parser origin 'Output m b
parser for the subselection set

-> FieldParser origin m (Maybe Name, HashMap Name (Value Variable), a, b)

selection_

Arguments

:: forall (m :: Type -> Type) origin a. MonadParse m

=> Name

-> Maybe Description

-> Parser origin 'Both m a
type of the result

-> FieldParser origin m ()

A shorthand for a selection that takes no arguments.

subselection_

Arguments

:: forall (m :: Type -> Type) origin a. MonadParse m

=> Name

-> Maybe Description

-> Parser origin 'Output m a
parser for the subselection set

-> FieldParser origin m a

A shorthand for a subselection that takes no arguments.

class Monad m => MonadMemoize (m :: Type -> Type)

Minimal complete definition

memoizeOn

memoizeOn

Arguments

:: forall m a p (n :: Type -> Type) b. (MonadMemoize m, Ord a, Typeable a, Typeable p, MonadParse n, Typeable b)

=> Name
A unique name used to identify the function being memoized. There isn’t really any metaprogramming going on here, we just use a Template Haskell Name as a convenient source for a static, unique identifier.

-> a
The value to use as the memoization key. It’s the caller’s responsibility to ensure multiple calls to the same function don’t use the same key.

-> m (p n b)
The value to be memoized. p is intended to be either 'Parser k' or FieldParser.

-> m (p n b)

memoize

Arguments

:: forall m a p (n :: Type -> Type) b. (MonadMemoize m, Ord a, Typeable a, Typeable p, MonadParse n, Typeable b)

=> Name

-> (a -> m (p n b))
A function generating something to be memoized. p is intended to be either 'Parser k' or FieldParser.

-> a

-> m (p n b)

A wrapper around memoizeOn that memoizes a function by using its argument as the key.