Hasura.GraphQL.Schema.Parser

Description

In Parser, the Definition type has a dOrigin field that allows to track where a fragment of GraphQL type information comes from. This is useful for error reporting and internal repair mechanisms such as inconsistency tracking.

Morally, within the HGE codebase, this origin is always MetadataObjId. However, in order to avoid an import of RQL from Parser, the dOrigin has been defined through a type parameter of Definition. This type parameter then has to get threaded through throughout the Parser module hierarchy, so that it ends up in a lot of types. This is very noisy.

In order to avoid the noise of this type parameter, which really only has one value, and is really only used in one type, this module erases the type parameter by filling in the desired value, exporting type synonyms of the now-fixed notion of "origin". So most modules in the HGE codebase should import this module rather than Parser.

Documentation

type FieldParser = FieldParser MetadataObjId

type Parser = Parser MetadataObjId

type Schema = Schema MetadataObjId

type Type = Type MetadataObjId

type InputFieldsParser = InputFieldsParser MetadataObjId

type Definition = Definition MetadataObjId

type Directive = Directive MetadataObjId

type DirectiveInfo = DirectiveInfo MetadataObjId

type FieldInfo = FieldInfo MetadataObjId

type InputFieldInfo = InputFieldInfo MetadataObjId

type HasTypeDefinitions = HasTypeDefinitions MetadataObjId

type SomeDefinitionTypeInfo = SomeDefinitionTypeInfo MetadataObjId

type TypeDefinitionsWrapper = TypeDefinitionsWrapper MetadataObjId

pattern TypeDefinitionsWrapper :: () => forall a. HasTypeDefinitions a => a -> TypeDefinitionsWrapper

In order to aid type inference and type checking, we define this pattern synonym (an actual one) which restricts TypeDefinitionsWrapper to have MetadataObjId set for its origin type parameter.