Hasura.GraphQL.Transport.HTTP

Description

Execution of GraphQL queries over HTTP transport

Documentation

data QueryCacheKey

Constructors

QueryCacheKey

Fields

qckQueryString :: !GQLReqParsed

qckUserRole :: !RoleName

qckSession :: !SessionVariables

type CacheStoreResponse = Either CacheStoreFailure CacheStoreSuccess

data CacheStoreSuccess

Constructors

CacheStoreSkipped

CacheStoreHit

data CacheStoreFailure

Constructors

CacheStoreLimitReached

CacheStoreNotEnoughCapacity

CacheStoreBackendError String

class Monad m => MonadExecuteQuery m where

Minimal complete definition

Nothing

Methods

cacheLookup

Arguments

:: [RemoteSchemaInfo]
Used to check if the elaborated query supports caching

-> [ActionsInfo]
Used to check if actions query supports caching (unsupported if forward_client_headers is set)

-> QueryCacheKey
Key that uniquely identifies the result of a query execution

-> Maybe CachedDirective
Cached Directive from GraphQL query AST

-> TraceT (ExceptT QErr m) (ResponseHeaders, Maybe EncJSON)
HTTP headers to be sent back to the caller for this GraphQL request, containing e.g. time-to-live information, and a cached value if found and within time-to-live. So a return value (non-empty-ttl-headers, Nothing) represents that we don't have a server-side cache of the query, but that the client should store it locally. The value ([], Just json) represents that the client should not store the response locally, but we do have a server-side cache value that can be used to avoid query execution.

This method does two things: it looks up a query result in the server-side cache, if a cache is used, and it additionally returns HTTP headers that can instruct a client how long a response can be cached locally (i.e. client-side).

default cacheLookup :: (m ~ t n, MonadTrans t, MonadExecuteQuery n) => [RemoteSchemaInfo] -> [ActionsInfo] -> QueryCacheKey -> Maybe CachedDirective -> TraceT (ExceptT QErr m) (ResponseHeaders, Maybe EncJSON)

cacheStore

Arguments

:: QueryCacheKey
Key under which to store the result of a query execution

-> Maybe CachedDirective
Cached Directive from GraphQL query AST

-> EncJSON
Result of a query execution

-> TraceT (ExceptT QErr m) CacheStoreResponse
Always succeeds

Store a json response for a query that we've executed in the cache. Note that, as part of this, cacheStore has to decide whether the response is cacheable. A very similar decision is also made in cacheLookup, since it has to construct corresponding cache-enabling headers that are sent to the client. But note that the HTTP headers influence client-side caching, whereas cacheStore changes the server-side cache.

default cacheStore :: (m ~ t n, MonadTrans t, MonadExecuteQuery n) => QueryCacheKey -> Maybe CachedDirective -> EncJSON -> TraceT (ExceptT QErr m) CacheStoreResponse

data AnnotatedResponsePart

A partial response, e.g. from a remote schema call or postgres postgres query, which we'll assemble into the final response for the client. It is annotated with timing metadata.

Constructors

AnnotatedResponsePart

Fields

arpTimeIO :: DiffTime

arpLocality :: Locality

arpResponse :: EncJSON

arpHeaders :: ResponseHeaders

data AnnotatedResponse

A full response, annotated with timing metadata.

Constructors

AnnotatedResponse

Fields

arQueryType :: QueryType

arTimeIO :: DiffTime

arLocality :: Locality

arResponse :: HttpResponse (Maybe GQResponse, EncJSON)

buildResponseFromParts :: MonadError QErr m => QueryType -> Either (Either GQExecError QErr) (RootFieldMap AnnotatedResponsePart) -> ResponseHeaders -> m AnnotatedResponse

Merge response parts into a full response.

buildResponse :: MonadError QErr m => QueryType -> Either (Either GQExecError QErr) a -> (a -> AnnotatedResponse) -> m AnnotatedResponse

newtype SessVarPred

A predicate on session variables. The Monoid instance makes it simple to combine several predicates disjunctively.

Constructors

SessVarPred

Fields

unSessVarPred :: SessionVariable -> SessionVariableValue -> Bool

keepAllSessionVariables :: SessVarPred

runSessVarPred :: SessVarPred -> SessionVariables -> SessionVariables

filterVariablesFromQuery :: Backend backend => [RootField (QueryDBRoot (RemoteRelationshipField UnpreparedValue) UnpreparedValue) RemoteField (ActionQuery backend (RemoteRelationshipField UnpreparedValue) (UnpreparedValue backend)) d] -> SessVarPred

Filter out only those session variables used by the query AST provided

runGQ :: forall m. (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadReader ExecutionCtx m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadTrace m, MonadExecuteQuery m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m, HasResourceLimits m) => Environment -> Logger Hasura -> RequestId -> UserInfo -> IpAddress -> [Header] -> GraphQLQueryType -> GQLReqUnparsed -> m (GQLQueryOperationSuccessLog, HttpResponse (Maybe GQResponse, EncJSON))

Run (execute) a single GraphQL query

coalescePostgresMutations :: ExecutionPlan -> Maybe (SourceConfig ('Postgres 'Vanilla), InsOrdHashMap RootFieldAlias (DBStepInfo ('Postgres 'Vanilla)))

data GraphQLResponse

Constructors

GraphQLResponseErrors [Value]

GraphQLResponseData Value

decodeGraphQLResponse :: ByteString -> Either Text GraphQLResponse

extractFieldFromResponse :: forall m. Monad m => RootFieldAlias -> ResultCustomizer -> ByteString -> ExceptT (Either GQExecError QErr) m Value

buildRaw :: Applicative m => Value -> m AnnotatedResponsePart

encodeAnnotatedResponseParts :: RootFieldMap AnnotatedResponsePart -> EncJSON

encodeEncJSONResults :: RootFieldMap EncJSON -> EncJSON

runGQBatched

Arguments

:: forall m. (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadReader ExecutionCtx m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadTrace m, MonadExecuteQuery m, HttpLog m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m, HasResourceLimits m)

=> Environment

-> Logger Hasura

-> RequestId

-> ResponseInternalErrorsConfig

-> UserInfo

-> IpAddress

-> [Header]

-> GraphQLQueryType

-> GQLBatchedReqs (GQLReq GQLQueryText)
the batched request with unparsed GraphQL query

-> m (HttpLogMetadata m, HttpResponse EncJSON)

Run (execute) a batched GraphQL query (see GQLBatchedReqs).