Hasura.GraphQL.Transport.WebSocket

Description

This file contains the handlers that are used within websocket server.

This module export three main handlers for the websocket server (onConn, onMessage, onClose), and two helpers for sending messages to the client (sendMsg, sendCloseWithMsg).

NOTE! The handler functions onClose, onMessage, etc. depend for correctness on two properties: - they run with async exceptions masked - they do not race on the same connection

Documentation

sendMsg :: MonadIO m => WSConn -> ServerMsg -> m ()

sendCloseWithMsg :: MonadIO m => Logger Hasura -> WSConn -> ServerErrorCode -> Maybe ServerMsg -> Maybe Word16 -> m ()

onConn :: (MonadIO m, MonadReader WSServerEnv m) => OnConnH m WSConnData

onMessage :: (MonadIO m, UserAuthentication (TraceT m), MonadGQLExecutionCheck m, MonadQueryLog m, HasReporter m, MonadExecuteQuery m, MonadBaseControl IO m, MonadMetadataStorage (MetadataStorageT m), MonadQueryTags m, HasResourceLimits m) => Environment -> HashSet (EngineLogType Hasura) -> AuthMode -> WSServerEnv -> WSConn -> ByteString -> WSActions WSConnData -> m ()

onClose :: MonadIO m => Logger Hasura -> ServerMetrics -> PrometheusMetrics -> SubscriptionsState -> WSConn -> m ()