Hasura.GraphQL.Transport.WebSocket.Server

Documentation

newtype WSId

Constructors

WSId

Fields

unWSId :: UUID

mkUnsafeWSId :: UUID -> WSId

data MessageDetails

Websocket message and other details

Constructors

MessageDetails

Fields

_mdMessage :: !SerializableBlob

_mdMessageSize :: !Int64

data WSEvent

Constructors

EConnectionRequest

EAccepted

ERejected

EMessageReceived !MessageDetails

EMessageSent !MessageDetails

EJwtExpired

ECloseReceived

ECloseSent !SerializableBlob

EClosed

data WSEventInfo

Constructors

WSEventInfo

Fields

_wseiEventType :: !(Maybe ServerMsgType)

_wseiOperationId :: !(Maybe OperationId)

_wseiOperationName :: !(Maybe OperationName)

_wseiQueryExecutionTime :: !(Maybe Double)

_wseiResponseSize :: !(Maybe Int64)

_wseiParameterizedQueryHash :: !(Maybe ParameterizedQueryHash)

data WSLog

Constructors

WSLog

Fields

_wslWebsocketId :: !WSId

_wslEvent :: !WSEvent

_wslMetadata :: !(Maybe WSEventInfo)

class Monad m => MonadWSLog m where

Methods

logWSLog :: Logger Hasura -> WSLog -> m ()

Takes WS server log data and logs it logWSServer

data WSQueueResponse

Constructors

WSQueueResponse

Fields

_wsqrMessage :: !ByteString

_wsqrEventInfo :: !(Maybe WSEventInfo)

extra metadata that we use for other actions, such as print log we don't want to inlcude them into websocket message payload

data WSConn a

Constructors

WSConn

Fields

_wcConnId :: !WSId

_wcLogger :: !(Logger Hasura)

_wcConnRaw :: !Connection

_wcSendQ :: !(TQueue WSQueueResponse)

_wcExtraData :: !a

getRawWebSocketConnection :: WSConn a -> Connection

getData :: WSConn a -> a

getWSId :: WSConn a -> WSId

closeConn :: WSConn a -> ByteString -> IO ()

forceConnReconnect :: MonadIO m => WSConn a -> ByteString -> m ()

Closes a connection with code 1012, which means "Server is restarting" good clients will implement a retry logic with a backoff of a few seconds

closeConnWithCode :: WSConn a -> Word16 -> ByteString -> IO ()

sendMsg :: WSConn a -> WSQueueResponse -> IO ()

type ConnMap a = Map WSId (WSConn a)

data ServerStatus a

Constructors

AcceptingConns !(ConnMap a)

ShuttingDown

data WSServer a

Constructors

WSServer

Fields

_wssLogger :: !(Logger Hasura)

_wssStatus :: !(TVar (ServerStatus a))

See e.g. createServerApp.onAccept for how we use STM to preserve consistency

createWSServer :: Logger Hasura -> STM (WSServer a)

closeAllWith :: (ByteString -> WSConn a -> IO ()) -> ByteString -> [(WSId, WSConn a)] -> IO ()

flushConnMap :: TVar (ServerStatus a) -> STM [(WSId, WSConn a)]

Resets the current connections map to an empty one if the server is running and returns the list of connections that were in the map before flushing it.

data AcceptWith a

Constructors

AcceptWith

Fields

_awData :: !a

_awReq :: !AcceptRequest

_awKeepAlive :: !(WSConn a -> IO ())

_awOnJwtExpiry :: !(WSConn a -> IO ())

type WSKeepAliveMessageAction a = WSConn a -> IO ()

These set of functions or message handlers is used by the server while communicating with the client. They are particularly useful for the case when the messages being sent to the client are different for each of the sub-protocol(s) supported by the server.

type WSPostExecErrMessageAction a = WSConn a -> OperationId -> GQExecError -> IO ()

type WSOnErrorMessageAction a = WSConn a -> ConnErrMsg -> Maybe String -> IO ()

type WSCloseConnAction a = WSConn a -> OperationId -> String -> IO ()

data WSActions a

Used for specific actions within the onConn and onMessage handlers

Constructors

WSActions

Fields

_wsaPostExecErrMessageAction :: !(WSPostExecErrMessageAction a)

_wsaOnErrorMessageAction :: !(WSOnErrorMessageAction a)

_wsaConnectionCloseAction :: !(WSCloseConnAction a)

_wsaKeepAliveAction :: !(WSKeepAliveMessageAction a)

NOTE: keep alive action was made redundant because we need to send this message after the connection has been successfully established after connection_init

_wsaGetDataMessageType :: !(DataMsg -> ServerMsg)

_wsaAcceptRequest :: !AcceptRequest

_wsaErrorMsgFormat :: !([Value] -> Value)

onClientMessageParseErrorText :: Maybe String

to be used with WSOnErrorMessageAction

onConnInitErrorText :: Maybe String

to be used with WSOnErrorMessageAction

type OnConnH m a = WSId -> RequestHead -> IpAddress -> WSActions a -> m (Either RejectRequest (AcceptWith a))

type OnCloseH m a = WSConn a -> m ()

type HasuraServerApp m = IpAddress -> PendingConnection -> m ()

aka generalized ServerApp over m, which takes an IPAddress

data WSHandlers m a

NOTE: The types of _hOnConn and _hOnMessage were updated from OnConnH and OnMessageH because we needed to pass the subprotcol here to these methods to eventually get to OnConnH and OnMessageH. Please see createServerApp to get a better understanding of how these handlers are used.

Constructors

WSHandlers

Fields

_hOnConn :: WSId -> RequestHead -> IpAddress -> WSSubProtocol -> m (Either RejectRequest (AcceptWith a))

_hOnMessage :: WSConn a -> ByteString -> WSSubProtocol -> m ()

_hOnClose :: OnCloseH m a

createServerApp

Arguments

:: (MonadIO m, MonadBaseControl IO m, Forall (Pure m), MonadWSLog m)

=> WSConnectionInitTimeout

-> WSServer a

-> WSHandlers m a
user provided handlers

-> HasuraServerApp m
aka WS.ServerApp

shutdown :: WSServer a -> IO ()