Hasura.GraphQL.Transport.WebSocket.Server

Documentation

data AcceptWith a

Constructors

AcceptWith !a !AcceptRequest !(WSConn a -> IO ()) !(WSConn a -> IO ())

type HasuraServerApp (m :: Type -> Type) = IpAddress -> PendingConnection -> m ()

aka generalized ServerApp over m, which takes an IPAddress

data MessageDetails

Websocket message and other details

Constructors

MessageDetails !SerializableBlob !Int64

class Monad m => MonadWSLog (m :: Type -> Type) where

Methods

logWSLog :: Logger Hasura -> WSLog -> m ()

Takes WS server log data and logs it logWSServer

type OnConnH (m :: Type -> Type) a = WSId -> RequestHead -> IpAddress -> WSActions a -> m (Either RejectRequest (AcceptWith a))

data WSActions a

Used for specific actions within the onConn and onMessage handlers

Constructors

WSActions

Fields

_wsaPostExecErrMessageAction :: !(WSPostExecErrMessageAction a)

_wsaOnErrorMessageAction :: !(WSOnErrorMessageAction a)

_wsaConnectionCloseAction :: !(WSCloseConnAction a)

_wsaKeepAliveAction :: !(WSKeepAliveMessageAction a)

NOTE: keep alive action was made redundant because we need to send this message after the connection has been successfully established after connection_init

_wsaGetDataMessageType :: !(DataMsg -> ServerMsg)

_wsaAcceptRequest :: !AcceptRequest

_wsaErrorMsgFormat :: !([Encoding] -> Encoding)

data WSConn a

data WSErrorMessage

Constructors

ClientMessageParseFailed

ConnInitFailed

data WSEvent

Constructors

EMessageSent !MessageDetails

data WSEventInfo

Constructors

WSEventInfo

Fields

_wseiEventType :: !(Maybe ServerMsgType)

_wseiOperationId :: !(Maybe OperationId)

_wseiOperationName :: !(Maybe OperationName)

_wseiQueryExecutionTime :: !(Maybe Double)

_wseiResponseSize :: !(Maybe Int64)

_wseiParameterizedQueryHash :: !(Maybe ParameterizedQueryHash)

data WSHandlers (m :: Type -> Type) a

NOTE: The types of _hOnConn and _hOnMessage were updated from OnConnH and OnMessageH because we needed to pass the subprotcol here to these methods to eventually get to OnConnH and OnMessageH. Please see createServerApp to get a better understanding of how these handlers are used.

Constructors

WSHandlers (WSId -> RequestHead -> IpAddress -> WSSubProtocol -> m (Either RejectRequest (AcceptWith a))) (WSConn a -> ByteString -> WSSubProtocol -> m ()) (OnCloseH m a)

data WSId

type WSKeepAliveMessageAction a = WSConn a -> IO ()

These set of functions or message handlers is used by the server while communicating with the client. They are particularly useful for the case when the messages being sent to the client are different for each of the sub-protocol(s) supported by the server.

data WSLog

Constructors

WSLog !WSId !WSEvent !(Maybe WSEventInfo)

type WSOnErrorMessageAction a = WSConn a -> ConnErrMsg -> WSErrorMessage -> IO ()

data WSQueueResponse

Constructors

WSQueueResponse ByteString (Maybe WSEventInfo) (IO DiffTime)

data WSServer a

Constructors

WSServer

Fields

_wssLogger :: Logger Hasura

_wssSecuritySensitiveUserConfig :: TVar SecuritySensitiveUserConfig

Keep track of the security sensitive user configuration to perform maintenance actions

_wssStatus :: TVar (ServerStatus a)

See e.g. createServerApp.onAccept for how we use STM to preserve consistency

websocketConnectionReaper :: IO (AuthMode, AllowListStatus, CorsPolicy, SQLGenCtx, HashSet ExperimentalFeature, NamingCase) -> IO SchemaCache -> WSServer a -> IO Void

The background thread responsible for closing all websocket connections when security sensitive user configuration changes. It checks for changes in the auth mode, allowlist, cors config, stringify num, dangerous boolean collapse, stringify big query numeric, experimental features and invalidates/closes all connections if there are any changes.

closeConn :: WSConn a -> ByteString -> IO ()

sendMsgAndCloseConn :: WSConn a -> Word16 -> ByteString -> ServerMsg -> IO ()

createServerApp

Arguments

:: (MonadIO m, MonadBaseControl IO m, Forall (Pure m), MonadWSLog m, MonadGetPolicies m)

=> IO MetricsConfig

-> WSConnectionInitTimeout

-> WSServer a

-> PrometheusMetrics

-> WSHandlers m a
user provided handlers

-> HasuraServerApp m
aka WS.ServerApp

createWSServer :: AuthMode -> AllowListStatus -> InlinedAllowlist -> CorsPolicy -> SQLGenCtx -> HashSet ExperimentalFeature -> NamingCase -> Logger Hasura -> STM (WSServer a)

closeAllConnectionsWithReason :: WSServer a -> String -> ByteString -> (SecuritySensitiveUserConfig -> SecuritySensitiveUserConfig) -> IO ()

getData :: WSConn a -> a

getRawWebSocketConnection :: WSConn a -> Connection

getWSId :: WSConn a -> WSId

mkWSServerErrorCode :: WSSubProtocol -> WSErrorMessage -> ConnErrMsg -> ServerErrorCode

sendMsg :: WSConn a -> WSQueueResponse -> IO ()

shutdown :: WSServer a -> IO ()

exported for testing

mkUnsafeWSId :: UUID -> WSId