Hasura.Logging

Documentation

newtype FormattedTime

Constructors

FormattedTime

Fields

_unFormattedTime :: Text

class (Eq (EngineLogType impl), Hashable (EngineLogType impl)) => EnabledLogTypes impl where

Typeclass representing any type which can be parsed into a list of enabled log types, and has a Set of default enabled log types, and can find out if a log type is enabled

Methods

parseEnabledLogTypes :: String -> Either String [EngineLogType impl]

defaultEnabledLogTypes :: HashSet (EngineLogType impl)

isLogTypeEnabled :: HashSet (EngineLogType impl) -> EngineLogType impl -> Bool

data family EngineLogType impl

A family of EngineLogType types

data Hasura

data InternalLogTypes

Constructors

ILTUnstructured
mostly for debug logs - see debugT, debugBS and debugLBS functions

ILTEventTrigger

ILTScheduledTrigger

ILTWsServer
internal logs for the websocket server

ILTPgClient

ILTMetadata
log type for logging metadata related actions; currently used in logging inconsistent metadata

ILTJwkRefreshLog

ILTTelemetry

ILTSchemaSyncThread

ILTSourceCatalogMigration

defaultEnabledEngineLogTypes :: HashSet (EngineLogType Hasura)

isEngineLogTypeEnabled :: HashSet (EngineLogType Hasura) -> EngineLogType Hasura -> Bool

readLogTypes :: String -> Either String [EngineLogType Hasura]

userAllowedLogTypes :: [EngineLogType Hasura]

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

data EngineLog impl

Constructors

EngineLog

Fields

_elTimestamp :: !FormattedTime

_elLevel :: !LogLevel

_elType :: !(EngineLogType impl)

_elDetail :: !Value

class EnabledLogTypes impl => ToEngineLog a impl where

Typeclass representing any data type that can be converted to EngineLog for the purpose of logging

Methods

toEngineLog :: a -> (LogLevel, EngineLogType impl, Value)

data UnstructuredLog

Constructors

UnstructuredLog

Fields

_ulLevel :: !LogLevel

_ulPayload :: !TByteString

debugT :: Text -> UnstructuredLog

debugBS :: ByteString -> UnstructuredLog

debugLBS :: ByteString -> UnstructuredLog

data LoggerCtx impl

Constructors

LoggerCtx

Fields

_lcLoggerSet :: !LoggerSet

_lcLogLevel :: !LogLevel

_lcTimeGetter :: !(IO FormattedTime)

_lcEnabledLogTypes :: !(HashSet (EngineLogType impl))

data LoggerSettings

Constructors

LoggerSettings

Fields

_lsCachedTimestamp :: !Bool

should current time be cached (refreshed every sec)

_lsTimeZone :: !(Maybe TimeZone)

_lsLevel :: !LogLevel

defaultLoggerSettings :: Bool -> LogLevel -> LoggerSettings

getFormattedTime :: Maybe TimeZone -> IO FormattedTime

mkLoggerCtx :: (MonadIO io, MonadBaseControl IO io) => LoggerSettings -> HashSet (EngineLogType impl) -> ManagedT io (LoggerCtx impl)

cleanLoggerCtx :: LoggerCtx a -> IO ()

newtype Logger impl

Constructors

Logger

Fields

unLogger :: forall a m. (ToEngineLog a impl, MonadIO m) => a -> m ()

mkLogger :: LoggerCtx Hasura -> Logger Hasura

eventTriggerLogType :: EngineLogType Hasura

scheduledTriggerLogType :: EngineLogType Hasura

sourceCatalogMigrationLogType :: EngineLogType Hasura