Hasura.LogicalModel.API

Description

Define and handle v1/metadata API operations to track, untrack, and get logical models.

Documentation

data GetLogicalModel (b :: BackendType)

API payload for the get_logical_model endpoint.

Constructors

GetLogicalModel

Fields

glmSource :: SourceName

data TrackLogicalModel (b :: BackendType)

Default implementation of the track_logical_model request payload.

Constructors

TrackLogicalModel

Fields

tlmSource :: SourceName

tlmName :: LogicalModelName

tlmDescription :: Maybe Text

tlmFields :: InsOrdHashMap (Column b) (LogicalModelField b)

data UntrackLogicalModel (b :: BackendType)

API payload for the untrack_logical_model endpoint.

Constructors

UntrackLogicalModel

Fields

utlmSource :: SourceName

utlmName :: LogicalModelName

runGetLogicalModel :: forall (b :: BackendType) m. (BackendMetadata b, MonadError QErr m, MetadataM m) => GetLogicalModel b -> m EncJSON

Handler for the get_logical_model endpoint.

execTrackLogicalModel :: forall (b :: BackendType) m. (BackendMetadata b, MonadError QErr m) => TrackLogicalModel b -> Metadata -> m (MetadataObjId, MetadataModifier)

Handler for the track_logical_model endpoint. The type 'TrackLogicalModel b' (appearing here in wrapped as 'BackendTrackLogicalModel b' for AnyBackend compatibility) is defined in 'class LogicalModelMetadata'.

execUntrackLogicalModel :: forall (b :: BackendType) m. (BackendMetadata b, MonadError QErr m) => UntrackLogicalModel b -> Metadata -> m (MetadataObjId, MetadataModifier)

Handler for the untrack_logical_model endpoint.

dropLogicalModelInMetadata :: forall (b :: BackendType). BackendMetadata b => SourceName -> LogicalModelName -> MetadataModifier

TODO: should this cascade and also delete associated permissions?

data CreateLogicalModelPermission (a :: BackendType -> Type) (b :: BackendType)

A permission for logical models is tied to a specific name and source. This wrapper adds both of those things to the JSON object that describes the permission.

Constructors

CreateLogicalModelPermission

Fields

clmpSource :: SourceName

clmpLocation :: LogicalModelLocation

clmpInfo :: PermDef b a

data DropLogicalModelPermission (b :: BackendType)

To drop a permission, we need to know the source and name of the logical model, as well as the role whose permission we want to drop.

Constructors

DropLogicalModelPermission

Fields

dlmpSource :: SourceName

dlmpLocation :: LogicalModelLocation

dlmpRole :: RoleName

runCreateSelectLogicalModelPermission :: forall (b :: BackendType) m. (Backend b, CacheRWM m, MetadataM m, MonadError QErr m) => CreateLogicalModelPermission SelPerm b -> m EncJSON

runDropSelectLogicalModelPermission :: forall (b :: BackendType) m. (Backend b, CacheRWM m, MetadataM m, MonadError QErr m) => DropLogicalModelPermission b -> m EncJSON

data LogicalModelField (b :: BackendType)

a single field in a Logical Model

data LogicalModelLocation

Constructors

LMLLogicalModel LogicalModelName

LMLNativeQuery NativeQueryName

data LogicalModelName

logicalModelFieldMapCodec :: forall (b :: BackendType). Backend b => Codec Value (LogicalModelFields b) (LogicalModelFields b)