Hasura.LogicalModel.Metadata

Documentation

data LogicalModelMetadata (b :: BackendType)

Description of a logical model for use in metadata (before schema cache)

Constructors

LogicalModelMetadata

Fields

_lmmName :: LogicalModelName

_lmmFields :: InsOrdHashMap (Column b) (LogicalModelField b)

_lmmDescription :: Maybe Text

_lmmSelectPermissions :: InsOrdHashMap RoleName (SelPermDef b)

data WithLogicalModel a

A wrapper to tie something to a particular native query. Specifically, it assumes the underlying _wlmInfo is represented as an object, and adds two keys to that object: source and root_field_name.

Constructors

WithLogicalModel

Fields

_wlmSource :: SourceName

_wlmName :: LogicalModelName

_wlmInfo :: a