Hasura.LogicalModel.Metadata

Description

Metadata representation of a logical model in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

data InterpolatedItem variable

A component of an interpolated query

Constructors

IIText Text
normal text

IIVariable variable
a captured variable

newtype InterpolatedQuery variable

A list of logical model components representing a single logical model, separating the variables from the text.

Constructors

InterpolatedQuery

Fields

getInterpolatedQuery :: [InterpolatedItem variable]

newtype LogicalModelArgumentName

Constructors

LogicalModelArgumentName

Fields

getLogicalModelArgumentName :: Text

data LogicalModelInfo (b :: BackendType)

A representation of a logical model metadata info object.

Constructors

LogicalModelInfo

Fields

lmiRootFieldName :: LogicalModelName

lmiCode :: InterpolatedQuery LogicalModelArgumentName

lmiReturns :: CustomReturnType b

lmiArguments :: HashMap LogicalModelArgumentName (ScalarType b)

lmiDescription :: Maybe Text

parseInterpolatedQuery :: Text -> Either Text (InterpolatedQuery LogicalModelArgumentName)

extract all of the `{{ variable }}` inside our query string