Hasura.Metadata.Class

Description

This module has type class and types which implements the Metadata Storage Abstraction

Documentation

data SchemaSyncEventProcessResult

Constructors

SchemaSyncEventProcessResult

Fields

_sseprShouldReload :: !Bool

_sseprCacheInvalidations :: !CacheInvalidations

class MonadError QErr m => MonadMetadataStorage m where

Metadata storage abstraction via a type class.

This type class enables storing and managing Hasura metadata in an isolated database which will not interfere with user's database where tables/functions are defined. Hence, it'll enable support for databases of multiple backends like MySQL, MSSQL etc.

This class has functions broadly related to:

Metadata Management ---------------------- Basic metadata management functions such as retrieving metadata from storage database and replacing the given metadata. TODO: Console specific operations

Scheduled Triggers --------------------- Eventing sub-system for scheduled triggers is implemented via metadata storage. For more details, refer description in ScheduledTrigger module.

TODO: Functions need to be added to the type class - Retrieving invocation logs from storage (console requirement) - Deleting an scheduled event - Creating an one-off scheduled event

Async Actions ---------------- Operations to implement async actions sub-system. This includes recording an async action event and retreiving the details of action delivery to the webhook. For more details see Note [Async action architecture] in Action module.

It is believed that all the above three are implemented in a single storage system (ex: a Postgres database). We can split the functions into appropriate and specific type classes in future iterations if required.

Methods

fetchMetadataResourceVersion :: m MetadataResourceVersion

fetchMetadata :: m (Metadata, MetadataResourceVersion)

fetchMetadataNotifications :: MetadataResourceVersion -> InstanceId -> m [(MetadataResourceVersion, CacheInvalidations)]

setMetadata :: MetadataResourceVersion -> Metadata -> m MetadataResourceVersion

notifySchemaCacheSync :: MetadataResourceVersion -> InstanceId -> CacheInvalidations -> m ()

getCatalogState :: m CatalogState

setCatalogState :: CatalogStateType -> Value -> m ()

getDatabaseUid :: m Text

checkMetadataStorageHealth :: m Bool

getDeprivedCronTriggerStats :: [TriggerName] -> m [CronTriggerStats]

getScheduledEventsForDelivery :: m ([CronEvent], [OneOffScheduledEvent])

insertCronEvents :: [CronEventSeed] -> m ()

insertOneOffScheduledEvent :: OneOffEvent -> m EventId

insertScheduledEventInvocation :: Invocation 'ScheduledType -> ScheduledEventType -> m ()

setScheduledEventOp :: ScheduledEventId -> ScheduledEventOp -> ScheduledEventType -> m ()

unlockScheduledEvents :: ScheduledEventType -> [ScheduledEventId] -> m Int

unlockAllLockedScheduledEvents :: m ()

clearFutureCronEvents :: ClearCronEvents -> m ()

getOneOffScheduledEvents :: ScheduledEventPagination -> [ScheduledEventStatus] -> m (WithTotalCount [OneOffScheduledEvent])

getCronEvents :: TriggerName -> ScheduledEventPagination -> [ScheduledEventStatus] -> m (WithTotalCount [CronEvent])

getInvocations :: GetInvocationsBy -> ScheduledEventPagination -> m (WithTotalCount [ScheduledEventInvocation])

deleteScheduledEvent :: ScheduledEventId -> ScheduledEventType -> m ()

insertAction :: ActionName -> SessionVariables -> [Header] -> Value -> m ActionId

fetchUndeliveredActionEvents :: m [ActionLogItem]

setActionStatus :: ActionId -> AsyncActionStatus -> m ()

fetchActionResponse :: ActionId -> m ActionLogResponse

clearActionData :: ActionName -> m ()

setProcessingActionLogsToPending :: LockedActionIdArray -> m ()

newtype MetadataStorageT m a

The MetadataStorageT transformer adds ability to throw exceptions for monads deriving @MonadMetadataStorage instance. For more details see Note [Generic MetadataStorageT transformer]

Constructors

MetadataStorageT

Fields

unMetadataStorageT :: ExceptT QErr m a

runMetadataStorageT :: MetadataStorageT m a -> m (Either QErr a)

class MonadMetadataStorage m => MonadMetadataStorageQueryAPI m where

Operations from @MonadMetadataStorage used in 'v1query' and 'v1metadata' APIs

Minimal complete definition

Nothing

Methods

createOneOffScheduledEvent :: OneOffEvent -> m EventId

Record a one-off event

createCronEvents :: [CronEventSeed] -> m ()

Record a cron event

dropFutureCronEvents :: ClearCronEvents -> m ()

Clear cron events

deleteActionData :: ActionName -> m ()

Delete async action logs

fetchInvocations :: GetInvocationsBy -> ScheduledEventPagination -> m (WithTotalCount [ScheduledEventInvocation])

Fetch cron/oneoff scheduled event invocations

fetchScheduledEvents :: GetScheduledEvents -> m Value

Fetch cron/oneoff scheduled events

dropEvent :: ScheduledEventId -> ScheduledEventType -> m ()

Drop a cron/oneoff scheduled event

fetchCatalogState :: m CatalogState

Retrieve the state from metadata storage catalog

updateCatalogState :: CatalogStateType -> Value -> m ()

Update the state from metadata storage catalog