Hasura.NativeQuery.Metadata

Description

This module houses the types and functions associated with the default implementation of the metadata of native queries.

Documentation

data InterpolatedItem variable

A component of an interpolated query

Constructors

IIText Text
normal text

IIVariable variable
a captured variable

newtype InterpolatedQuery variable

A list of native query components representing a single native query, separating the variables from the text.

Constructors

InterpolatedQuery

Fields

getInterpolatedQuery :: [InterpolatedItem variable]

newtype NativeQueryArgumentName

Constructors

NativeQueryArgumentName

Fields

getNativeQueryArgumentName :: Text

data NativeQueryInfo (b :: BackendType)

Default implementation of the Native Query metadata info object.

Constructors

NativeQueryInfo

Fields

nqiRootFieldName :: NativeQueryName

nqiCode :: InterpolatedQuery NativeQueryArgumentName

nqiReturns :: TableName b

nqiArguments :: HashMap NativeQueryArgumentName (ScalarType b)

nqiDescription :: Maybe Text

parseInterpolatedQuery :: Text -> Either Text (InterpolatedQuery NativeQueryArgumentName)

extract all of the `{{ variable }}` inside our query string