Hasura.NativeQuery.Metadata

Description

Metadata representation of a native query in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

data NativeQueryMetadata (b :: BackendType)

The representation of native queries within the metadata structure.

Constructors

NativeQueryMetadata

Fields

_nqmRootFieldName :: NativeQueryName

_nqmCode :: InterpolatedQuery NativeQueryArgumentName

_nqmReturns :: LogicalModelName

_nqmArguments :: HashMap NativeQueryArgumentName (NullableScalarType b)

_nqmArrayRelationships :: Relationships (RelDef (RelManualConfig b))

_nqmDescription :: Maybe Text

nqmRootFieldName :: forall b. Lens' (NativeQueryMetadata b) NativeQueryName

nqmReturns :: forall b. Lens' (NativeQueryMetadata b) LogicalModelName

nqmDescription :: forall b. Lens' (NativeQueryMetadata b) (Maybe Text)

nqmCode :: forall b. Lens' (NativeQueryMetadata b) (InterpolatedQuery NativeQueryArgumentName)

nqmArrayRelationships :: forall b. Lens' (NativeQueryMetadata b) (Relationships (RelDef (RelManualConfig b)))

nqmArguments :: forall b. Lens' (NativeQueryMetadata b) (HashMap NativeQueryArgumentName (NullableScalarType b))