Hasura.NativeQuery.Metadata

Description

Metadata representation of a native query in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

data NativeQueryMetadata (b :: BackendType)

The representation of native queries within the metadata structure.

Constructors

NativeQueryMetadata

Fields

_nqmRootFieldName :: NativeQueryName

_nqmCode :: InterpolatedQuery ArgumentName

_nqmReturns :: LogicalModelIdentifier b

_nqmArguments :: HashMap ArgumentName (NullableScalarType b)

_nqmArrayRelationships :: Relationships (RelDef (RelManualNativeQueryConfig b))

_nqmObjectRelationships :: Relationships (RelDef (RelManualConfig b))

_nqmDescription :: Maybe Text

data WithNativeQuery a

A wrapper to tie something to a particular native query. Specifically, it assumes the underlying _wlmInfo is represented as an object, and adds two keys to that object: source and root_field_name.

Constructors

WithNativeQuery

Fields

_wnqSource :: SourceName

_wnqName :: NativeQueryName

_wnqInfo :: a