Hasura.NativeQuery.Metadata

Description

Metadata representation of a native query in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

newtype NativeQueryName

Constructors

NativeQueryName

Fields

getNativeQueryName :: Name

data NativeQueryMetadata (b :: BackendType)

The representation of native queries within the metadata structure.

Constructors

NativeQueryMetadata

Fields

_nqmRootFieldName :: NativeQueryName

_nqmCode :: InterpolatedQuery ArgumentName

_nqmReturns :: LogicalModelIdentifier b

_nqmArguments :: HashMap ArgumentName (NullableScalarType b)

_nqmArrayRelationships :: Relationships (RelDef (RelManualConfig b))

_nqmObjectRelationships :: Relationships (RelDef (RelManualConfig b))

_nqmDescription :: Maybe Text

newtype ArgumentName

A name of an argument to a native query or stored procedure.

Constructors

ArgumentName

Fields

getArgumentName :: Text

data InterpolatedItem variable

A component of an interpolated query

Constructors

IIText Text
normal text

IIVariable variable
a captured variable

newtype InterpolatedQuery variable

A list of stored procedure components representing a single stored procedure, separating the variables from the text.

Constructors

InterpolatedQuery

Fields

getInterpolatedQuery :: [InterpolatedItem variable]

parseInterpolatedQuery :: Text -> Either Text (InterpolatedQuery ArgumentName)

extract all of the `{{ variable }}` inside our query string

module Hasura.NativeQuery.Types

data WithNativeQuery a

A wrapper to tie something to a particular native query. Specifically, it assumes the underlying _wlmInfo is represented as an object, and adds two keys to that object: source and root_field_name.

Constructors

WithNativeQuery

Fields

_wnqSource :: SourceName

_wnqName :: NativeQueryName

_wnqInfo :: a