Hasura.NativeQuery.Types

Description

This module houses the types that are necessary to even talk about native queries abstract of a concrete implementation.

The default implementation is given in modules Metadata, and API, but backends are free to provide their own as needed.

Documentation

class (APIType (Voidable (NativeQueryName b)), APIType (Voidable (TrackNativeQuery b)), FromJSON (Voidable [NativeQueryInfo b]), Ord (NativeQueryName b), Representable (NativeQueryInfo b), Representable (NativeQueryName b), Eq (Voidable [NativeQueryInfo b]), HasCodec (Voidable [NativeQueryInfo b]), ToJSON (Voidable [NativeQueryInfo b]), ToTxt (NativeQueryName b)) => NativeQueryMetadata (b :: BackendType) where

This type class models the types and functions necessary to talk about Native Queries.

Uninstantiable defaults are given for types and methods. In order to work gracefully with this approach, types require instances wrapped in the Voidable newtype, see module Voidable for more on this.

Minimal complete definition

Nothing

Associated Types

type NativeQueryInfo b :: Type

The type of persisted metadata.

type NativeQueryInfo b = Void

type NativeQueryName b :: Type

The types of names of native queries.

type NativeQueryName b = Void

type TrackNativeQuery b :: Type

The API payload of the track_native_query api endpoint.

type TrackNativeQuery b = Void

Methods

trackNativeQuerySource :: TrackNativeQuery b -> SourceName

Projection function identifying the name of the source a track_native_query request concerns.

default trackNativeQuerySource :: TrackNativeQuery b ~ Void => TrackNativeQuery b -> SourceName

nativeQueryInfoName :: NativeQueryInfo b -> NativeQueryName b

Projection function giving the name of a native query.

default nativeQueryInfoName :: NativeQueryInfo b ~ Void => NativeQueryInfo b -> NativeQueryName b

nativeQueryTrackToInfo :: TrackNativeQuery b -> Either NativeQueryParseError (NativeQueryInfo b)

Projection function, producing a 'NativeQueryInfo b' from a 'TrackNativeQuery b'.

default nativeQueryTrackToInfo :: TrackNativeQuery b ~ Void => TrackNativeQuery b -> Either NativeQueryParseError (NativeQueryInfo b)

newtype BackendTrackNativeQuery b

Our API endpoint solution wraps all request payload types in AnyBackend for its multi-backend support, but type families must be fully applied to all their arguments.

So in order to be usable as an API request payload data type, 'TrackNativeQuery b' needs to be wrapped in a newtype.

Constructors

BackendTrackNativeQuery

Fields

getBackendTrackNativeQuery :: Voidable (TrackNativeQuery b)

newtype NativeQueryParseError

Constructors

NativeQueryParseError Text