Hasura.NativeQuery.Types

Description

A name for a native query as it is recognized by the graphql schema.

Documentation

newtype NativeQueryName

Constructors

NativeQueryName

Fields

getNativeQueryName :: Name

data NullableScalarType b

A ScalarType that can be nullable with an optional description

Constructors

NullableScalarType

Fields

nstType :: ScalarType b

nstNullable :: Bool

nstDescription :: Maybe Text

nullableScalarTypeMapCodec :: forall b. Backend b => Codec Value (InsOrdHashMap (Column b) (NullableScalarType b)) (InsOrdHashMap (Column b) (NullableScalarType b))

nativeQueryArrayRelationshipsCodec :: forall b. Backend b => Codec Value (InsOrdHashMap RelName (RelDef (RelManualConfig b))) (InsOrdHashMap RelName (RelDef (RelManualConfig b)))