Hasura.NativeQuery.Types

Description

A name for a native query as it is recognized by the graphql schema.

Documentation

newtype NativeQueryName

Constructors

NativeQueryName

Fields

getNativeQueryName :: Name

data NullableScalarType (b :: BackendType)

A ScalarType that can be nullable with an optional description

Constructors

NullableScalarType

Fields

nstType :: ScalarType b

nstNullable :: Bool

nstDescription :: Maybe Text