Hasura.QueryTags

Description

This module has the various metadata we want to attach to the generated/executed query

Documentation

data QueryTags

Query Tags are SQL comments which are made up of (key=value) pairs.

These are appended to the SQL statements generated by Hasura for GraphQL operations. This enables the ability to get some application context in the database logs and also use native database monitoring tools (e.g. pganalyze) for better performance analysis.

The application context(query tags) can be used to detect slow GQL operation and relate them back to the SQL that was generated.

For eg: SELECT name FROM child * request_id=487c2ed5-08a4-429a-b0e0-4666a82e3cc6, field_name=child, operation_name=GetChild *

For more usage information, refer Query Tags Docs

Constructors

QTQuery !QueryMetadata

QTMutation !MutationMetadata

QTLiveQuery !LivequeryMetadata

newtype QueryTagsComment

query-tags as SQL comment which is appended to the prepared SQL statement

Constructors

QueryTagsComment

Fields

_unQueryTagsComment :: Text

type Attribute = (Text, Text)

newtype QueryTagsAttributes

Constructors

QueryTagsAttributes

Fields

_unQueryTagsAttributes :: [Attribute]

emptyQueryTagsAttributes :: QueryTagsAttributes

emptyQueryTagsComment :: QueryTagsComment

data QueryMetadata

Constructors

QueryMetadata

Fields

qmRequestId :: !RequestId

qmOperationName :: !(Maybe Name)

qmFieldName :: !RootFieldAlias

qmParameterizedQueryHash :: !ParameterizedQueryHash

data MutationMetadata

Constructors

MutationMetadata

Fields

mmRequestId :: !RequestId

mmOperationName :: !(Maybe Name)

mmFieldName :: !RootFieldAlias

mmParameterizedQueryHash :: !ParameterizedQueryHash

data LivequeryMetadata

Constructors

LivequeryMetadata

Fields

lqmFieldName :: !RootFieldAlias

lqmParameterizedQueryHash :: !ParameterizedQueryHash

encodeQueryTags :: QueryTags -> QueryTagsAttributes

operationNameAttributes :: Maybe Name -> [(Text, Text)]