Hasura.RQL.DDL.CustomTypes

Documentation

validateCustomTypeDefinitions

Arguments

:: forall m. MonadValidate [CustomTypeValidationError] m

=> SourceCache

-> CustomTypes

-> DMap BackendTag ScalarSet
all Postgres base types. See Note [Postgres scalars in custom types]

-> m AnnotatedCustomTypes

Validate the custom types and return any reused Postgres base types (as scalars).

lookupPGScalar :: DMap BackendTag ScalarSet -> Name -> (PGScalarType -> a) -> Maybe a

data CustomTypeValidationError

Constructors

DuplicateTypeNames !(HashSet Name)
type names have to be unique across all types

InputObjectFieldTypeDoesNotExist !InputObjectTypeName !InputObjectFieldName !Name
field name and the field's base type

InputObjectDuplicateFields !InputObjectTypeName !(HashSet InputObjectFieldName)
duplicate field declaration in input objects

ObjectFieldTypeDoesNotExist !ObjectTypeName !ObjectFieldName !Name
field name and the field's base type

ObjectDuplicateFields !ObjectTypeName !(HashSet Name)
duplicate field declaration in objects

ObjectFieldArgumentsNotAllowed !ObjectTypeName !ObjectFieldName
object fields can't have arguments

ObjectFieldObjectBaseType !ObjectTypeName !ObjectFieldName !Name
object fields can't have object types as base types

ObjectRelationshipTableDoesNotExist !ObjectTypeName !RelationshipName !QualifiedTable
The table specified in the relationship does not exist

ObjectRelationshipFieldDoesNotExist !ObjectTypeName !RelationshipName !ObjectFieldName
The field specified in the relationship mapping does not exist

ObjectRelationshipFieldListType !ObjectTypeName !RelationshipName !ObjectFieldName
The field specified in the relationship mapping is a list type

ObjectRelationshipColumnDoesNotExist !ObjectTypeName !RelationshipName !QualifiedTable !PGCol
The column specified in the relationship mapping does not exist

ObjectRelationshipMultiSources !ObjectTypeName
Object relationship refers to table in multiple sources

DuplicateEnumValues !EnumTypeName !(HashSet EnumValue)
duplicate enum values

showCustomTypeValidationError :: CustomTypeValidationError -> Text

runSetCustomTypes :: (MonadError QErr m, CacheRWM m, MetadataM m) => CustomTypes -> m EncJSON

clearCustomTypesInMetadata :: MetadataModifier

resolveCustomTypes :: MonadError QErr m => SourceCache -> CustomTypes -> DMap BackendTag ScalarSet -> m AnnotatedCustomTypes