Hasura.RQL.DDL.InheritedRoles

Documentation

runAddInheritedRole :: (MonadError QErr m, CacheRWM m, MetadataM m) => InheritedRole -> m EncJSON

runDropInheritedRole :: (MonadError QErr m, CacheRWM m, MetadataM m) => DropInheritedRole -> m EncJSON

dropInheritedRoleInMetadata :: RoleName -> MetadataModifier

resolveInheritedRole :: MonadError QErr m => HashSet RoleName -> InheritedRole -> m (Role, Seq SchemaDependency)

resolveInheritedRole resolves an inherited role by checking if all the parent roles of an inherited role exists and report the dependencies of the inherited role which will be the list of the parent roles