Hasura.RQL.DDL.Metadata.Types

Description

Types related to metadata management API

Documentation

data ClearMetadata

ClearMetadata can be used to reset the state of Hasura -- clean the current state by forgetting the tables tracked, relationships, permissions, event triggers etc.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#metadata-clear-metadata

Constructors

ClearMetadata

data ExportMetadata

ExportMetadata is used to export the current metadata from the server as a JSON file.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#metadata-export-metadata

Constructors

ExportMetadata

data ReloadSpec a

Constructors

RSReloadAll

RSReloadList (HashSet a)

data ReloadMetadata

ReloadMetadata should be used when there is a change in underlying Postgres database that Hasura should be aware of. Example: a new column is added to a table using psql and this column should now be added to the GraphQL schema.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#metadata-reload-metadata

Constructors

ReloadMetadata

Fields

_rmReloadRemoteSchemas :: ReloadRemoteSchemas

_rmReloadSources :: ReloadSources

_rmRecreateEventTriggers :: ReloadSources

Provides a way for the user to allow to explicitly recreate event triggers for some or all the sources. This is useful when a user may have fiddled with the SQL trigger in the source and they'd simply want the event trigger to be recreated without deleting and creating the event trigger. By default, no source's event triggers will be recreated.

_rmReloadDataConnectors :: ReloadDataConnectors

data DumpInternalState

Undocumented Metadata API action which serializes the entire SchemaCache.

Constructors

DumpInternalState

data GetInconsistentMetadata

GetInconsistentMetadata can be used to fetch all inconsistent metadata objects.

https://hasura.io/docs/latest/api-reference/schema-metadata-api/manage-metadata/#schema-metadata-get-inconsistent-metadata

Constructors

GetInconsistentMetadata

data DropInconsistentMetadata

DropInconsistentMetadata can be used to purge all inconsistent objects from the metadata.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#metadata-drop-inconsistent-metadata

Constructors

DropInconsistentMetadata

data AllowInconsistentMetadata

Constructors

AllowInconsistentMetadata

NoAllowInconsistentMetadata

data ReplaceMetadataV1

Replace metadata either with or without metadata sources.

Constructors

RMWithSources Metadata

RMWithoutSources MetadataNoSources

data ReplaceMetadataV2

Replace metadata while allowing for inconsitency in the metadata object.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#metadata-replace-metadata-syntax

Constructors

ReplaceMetadataV2

Fields

_rmv2AllowInconsistentMetadata :: AllowInconsistentMetadata

_rmv2Metadata :: ReplaceMetadataV1

data ReplaceMetadata

ReplaceMetadata is used to replace/import metadata into Hasura. Existing metadata will be replaced with the new one.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#metadata-replace-metadata TODO: If additional API versions are supported in future it would be ideal to include a version field Rather than differentiating on the "metadata" field.

Constructors

RMReplaceMetadataV1 ReplaceMetadataV1

RMReplaceMetadataV2 ReplaceMetadataV2

data WebHookUrl

Constructors

EnvVar String

URL Text

data TestWebhookTransform

TestWebhookTransform can be used to test out request transformations using mock data.

https://hasura.io/docs/latest/api-reference/metadata-api/manage-metadata/#test-webhook-transform

Constructors

TestWebhookTransform

Fields

_twtEnv :: Environment

_twtHeaders :: [Header]

_twtWebhookUrl :: WebHookUrl

_twtPayload :: Value

_twtTransformer :: RequestTransform

_twtResponseTransformer :: Maybe MetadataResponseTransform

_twtSessionVariables :: Maybe SessionVariables

twtTransformer :: Lens' TestWebhookTransform RequestTransform