Hasura.RQL.DDL.Permission.Internal

Documentation

newtype CreatePerm (a :: BackendType -> Type) (b :: BackendType)

Constructors

CreatePerm (WithTable b (PermDef b a))

data DropPerm (b :: BackendType)

Constructors

DropPerm

Fields

dipSource :: SourceName

dipTable :: TableName b

dipRole :: RoleName

permissionIsDefined :: forall (backend :: BackendType). PermType -> RolePermInfo backend -> Bool

assertPermDefined :: forall (backend :: BackendType) m. (Backend backend, MonadError QErr m) => RoleName -> PermType -> TableInfo backend -> m ()

interpColSpec :: forall (b :: BackendType). [Column b] -> PermColSpec b -> [Column b]

Intrepet a PermColSpec column specification, which can either refer to a list of named columns or all columns.

getDepHeadersFromVal :: Value -> [Text]

getDependentHeaders :: forall (b :: BackendType). BoolExp b -> HashSet Text

annBoolExp :: forall m (b :: BackendType) v. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b) => BoolExpRHSParser b m v -> FieldInfoMap (FieldInfo b) -> FieldInfoMap (FieldInfo b) -> GBoolExp b ColExp -> m (AnnBoolExp b v)

procBoolExp :: forall m (b :: BackendType) r. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> BoolExp b -> m (AnnBoolExpPartialSQL b, Seq SchemaDependency)

procLogicalModelBoolExp :: forall (b :: BackendType) m r. (QErrM m, TableCoreInfoRM b m, LogicalModelFieldsRM b m, BackendMetadata b, GetAggregationPredicatesDeps b, MonadReader r m, Has (ScalarTypeParsingContext b) r) => SourceName -> LogicalModelLocation -> FieldInfoMap (FieldInfo b) -> BoolExp b -> m (AnnBoolExpPartialSQL b, Seq SchemaDependency)

Interpret a BoolExp into an AnnBoolExp, collecting any dependencies as we go. At the moment, the only dependencies we're likely to encounter are independent dependencies on other tables. For example, "this user can only select from this logical model if their ID is in the allowed_users table".