Hasura.RQL.DDL.Schema.Cache

Description

Top-level functions concerned specifically with operations on the schema cache, such as rebuilding it from the catalog and incorporating schema changes. See the module documentation for Hasura.RQL.DDL.Schema for more details.

Note: this module is mutually recursive with other Hasura.RQL.DDL.Schema.* modules, which both define pieces of the implementation of building the schema cache and define handlers that trigger schema cache rebuilds.

Documentation

data RebuildableSchemaCache

lastBuiltSchemaCache :: RebuildableSchemaCache -> SchemaCache

buildRebuildableSchemaCache :: Logger Hasura -> Environment -> MetadataWithResourceVersion -> CacheDynamicConfig -> Maybe SchemaRegistryContext -> CacheBuild RebuildableSchemaCache

data CacheRWT (m :: Type -> Type) a

runCacheRWT :: Monad m => CacheDynamicConfig -> RebuildableSchemaCache -> CacheRWT m a -> m (a, RebuildableSchemaCache, CacheInvalidations, SourcesIntrospectionStatus, SchemaRegistryAction)

mkBooleanPermissionMap :: (RoleName -> a) -> HashMap RoleName a -> OrderedRoles -> HashMap RoleName a

saveSourcesIntrospection :: (MonadTraceContext m, MonadIO m, MonadMetadataStorage m) => Logger Hasura -> SourcesIntrospectionStatus -> MetadataResourceVersion -> m ()