Hasura.RQL.DDL.Schema.Cache.Common

Description

Types/functions shared between modules that implement Hasura.RQL.DDL.Schema.Cache. Other modules should not import this module directly.

Documentation

data InvalidationKeys

InvalidationKeys used to apply requested CacheInvalidations.

Constructors

InvalidationKeys

Fields

_ikMetadata :: InvalidationKey

_ikRemoteSchemas :: HashMap RemoteSchemaName InvalidationKey

_ikSources :: HashMap SourceName InvalidationKey

ikSources :: Lens' InvalidationKeys (HashMap SourceName InvalidationKey)

ikRemoteSchemas :: Lens' InvalidationKeys (HashMap RemoteSchemaName InvalidationKey)

ikMetadata :: Lens' InvalidationKeys InvalidationKey

initialInvalidationKeys :: InvalidationKeys

invalidateKeys :: CacheInvalidations -> InvalidationKeys -> InvalidationKeys

data TableBuildInput b

Constructors

TableBuildInput

Fields

_tbiName :: TableName b

_tbiIsEnum :: Bool

_tbiConfiguration :: TableConfig b

_tbiApolloFederationConfig :: Maybe ApolloFederationConfig

data NonColumnTableInputs b

Constructors

NonColumnTableInputs

Fields

_nctiTable :: TableName b

_nctiObjectRelationships :: [ObjRelDef b]

_nctiArrayRelationships :: [ArrRelDef b]

_nctiComputedFields :: [ComputedFieldMetadata b]

_nctiRemoteRelationships :: [RemoteRelationship]

data TablePermissionInputs b

Constructors

TablePermissionInputs

Fields

_tpiTable :: TableName b

_tpiInsert :: [InsPermDef b]

_tpiSelect :: [SelPermDef b]

_tpiUpdate :: [UpdPermDef b]

_tpiDelete :: [DelPermDef b]

mkTableInputs :: TableMetadata b -> (TableBuildInput b, NonColumnTableInputs b, TablePermissionInputs b)

data BuildOutputs

The direct output of buildSchemaCacheRule. Contains most of the things necessary to build a schema cache, but dependencies and inconsistent metadata objects are collected via a separate MonadWriter side channel.

Constructors

BuildOutputs

Fields

_boSources :: SourceCache

_boActions :: ActionCache

_boRemoteSchemas :: HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject)

We preserve the MetadataObject from the original catalog metadata in the output so we can reuse it later if we need to mark the remote schema inconsistent during GraphQL schema generation (because of field conflicts).

_boAllowlist :: InlinedAllowlist

_boCustomTypes :: AnnotatedCustomTypes

_boCronTriggers :: HashMap TriggerName CronTriggerInfo

_boEndpoints :: HashMap EndpointName (EndpointMetadata GQLQueryWithText)

_boApiLimits :: ApiLimit

_boMetricsConfig :: MetricsConfig

_boRoles :: HashMap RoleName Role

_boTlsAllowlist :: [TlsAllow]

_boQueryCollections :: QueryCollections

_boBackendCache :: BackendCache

boTlsAllowlist :: Lens' BuildOutputs [TlsAllow]

boSources :: Lens' BuildOutputs SourceCache

boRoles :: Lens' BuildOutputs (HashMap RoleName Role)

boRemoteSchemas :: Lens' BuildOutputs (HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject))

boQueryCollections :: Lens' BuildOutputs QueryCollections

boMetricsConfig :: Lens' BuildOutputs MetricsConfig

boEndpoints :: Lens' BuildOutputs (HashMap EndpointName (EndpointMetadata GQLQueryWithText))

boCustomTypes :: Lens' BuildOutputs AnnotatedCustomTypes

boCronTriggers :: Lens' BuildOutputs (HashMap TriggerName CronTriggerInfo)

boBackendCache :: Lens' BuildOutputs BackendCache

boApiLimits :: Lens' BuildOutputs ApiLimit

boAllowlist :: Lens' BuildOutputs InlinedAllowlist

boActions :: Lens' BuildOutputs ActionCache

data CacheBuildParams

Parameters required for schema cache build

Constructors

CacheBuildParams

Fields

_cbpManager :: Manager

_cbpPGSourceResolver :: SourceResolver ('Postgres 'Vanilla)

_cbpMSSQLSourceResolver :: SourceResolver 'MSSQL

_cbpServerConfigCtx :: ServerConfigCtx

newtype CacheBuild a

The monad in which @RebuildableSchemaCache is being run

Constructors

CacheBuild (ReaderT CacheBuildParams (ExceptT QErr IO) a)

runCacheBuild :: (MonadIO m, MonadError QErr m) => CacheBuildParams -> CacheBuild a -> m a

runCacheBuildM :: (MonadIO m, MonadError QErr m, HasHttpManagerM m, HasServerConfigCtx m, MonadResolveSource m) => CacheBuild a -> m a

data RebuildableSchemaCache

Constructors

RebuildableSchemaCache

Fields

lastBuiltSchemaCache :: SchemaCache

_rscInvalidationMap :: InvalidationKeys

_rscRebuild :: Rule (ReaderT BuildReason CacheBuild) (Metadata, InvalidationKeys) SchemaCache

bindErrorA :: (ArrowChoice arr, ArrowKleisli m arr, ArrowError e arr, MonadError e m) => arr (m a) a

withRecordDependencies :: ArrowWriter (Seq CollectedInfo) arr => WriterA (Seq SchemaDependency) arr (e, s) a -> arr (e, (MetadataObject, (SchemaObjId, s))) a

noDuplicates :: MonadWriter (Seq CollectedInfo) m => (a -> MetadataObject) -> [a] -> m (Maybe a)

buildInfoMap :: (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectedInfo) arr, Eq k, Hashable k) => (a -> k) -> (a -> MetadataObject) -> ((e, a) `arr` Maybe b) -> (e, [a]) `arr` HashMap k b

Processes a list of catalog metadata into a map of processed information, marking any duplicate entries inconsistent.

buildInfoMapPreservingMetadata :: (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectedInfo) arr, Eq k, Hashable k) => (a -> k) -> (a -> MetadataObject) -> ((e, a) `arr` Maybe b) -> (e, [a]) `arr` HashMap k (b, MetadataObject)

Like buildInfo, but includes each processed info’s associated MetadataObject in the result. This is useful if the results will be further processed, and the MetadataObject is still needed to mark the object inconsistent.

addTableContext :: Backend b => TableName b -> Text -> Text