Hasura.RQL.DDL.Schema.Cache.Common

Description

Types/functions shared between modules that implement Hasura.RQL.DDL.Schema.Cache. Other modules should not import this module directly.

Documentation

data ApolloFederationConfig

Constructors

ApolloFederationConfig

Fields

enable :: ApolloFederationVersion

data ApolloFederationVersion

Constructors

V1

newtype BackendInvalidationKeysWrapper (b :: BackendType)

Constructors

BackendInvalidationKeysWrapper

Fields

unBackendInvalidationKeysWrapper :: BackendInvalidationKeys b

data BuildOutputs

The direct output of buildSchemaCacheRule. Contains most of the things necessary to build a schema cache, but dependencies and inconsistent metadata objects are collected via a separate MonadWriter side channel.

See also Note [Avoiding GraphQL schema rebuilds when changing irrelevant Metadata]

Constructors

BuildOutputs

Fields

_boSources :: SourceCache

_boActions :: ActionCache

_boRemoteSchemas :: HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject)

We preserve the MetadataObject from the original catalog metadata in the output so we can reuse it later if we need to mark the remote schema inconsistent during GraphQL schema generation (because of field conflicts).

_boCustomTypes :: AnnotatedCustomTypes

_boRoles :: HashMap RoleName Role

_boBackendCache :: BackendCache

data CacheBuild a

The monad in which @RebuildableSchemaCache is being run

data CacheBuildParams

Parameters required for schema cache build

Constructors

CacheBuildParams Manager (SourceResolver ('Postgres 'Vanilla)) (SourceResolver 'MSSQL) CacheStaticConfig

data InvalidationKeys

InvalidationKeys used to apply requested CacheInvalidations.

Constructors

InvalidationKeys

Fields

_ikMetadata :: InvalidationKey

_ikRemoteSchemas :: HashMap RemoteSchemaName InvalidationKey

_ikSources :: HashMap SourceName InvalidationKey

_ikBackends :: BackendMap BackendInvalidationKeysWrapper

ikMetadata :: Lens' InvalidationKeys InvalidationKey

ikRemoteSchemas :: Lens' InvalidationKeys (HashMap RemoteSchemaName InvalidationKey)

ikSources :: Lens' InvalidationKeys (HashMap SourceName InvalidationKey)

ikBackends :: Lens' InvalidationKeys (BackendMap BackendInvalidationKeysWrapper)

data NonColumnTableInputs (b :: BackendType)

Constructors

NonColumnTableInputs

Fields

_nctiTable :: TableName b

_nctiObjectRelationships :: [ObjRelDef b]

_nctiArrayRelationships :: [ArrRelDef b]

_nctiComputedFields :: [ComputedFieldMetadata b]

_nctiRemoteRelationships :: [RemoteRelationship]

data RebuildableSchemaCache

Constructors

RebuildableSchemaCache SchemaCache InvalidationKeys (Rule (ReaderT BuildReason CacheBuild) (MetadataWithResourceVersion, CacheDynamicConfig, InvalidationKeys, Maybe StoredIntrospection) (SchemaCache, (SourcesIntrospectionStatus, SchemaRegistryAction)))

data TableBuildInput (b :: BackendType)

Constructors

TableBuildInput (TableName b) Bool (TableConfig b) (Maybe ApolloFederationConfig) (Maybe LogicalModelName)

data TablePermissionInputs (b :: BackendType)

Constructors

TablePermissionInputs

Fields

_tpiTable :: TableName b

_tpiInsert :: [InsPermDef b]

_tpiSelect :: [SelPermDef b]

_tpiUpdate :: [UpdPermDef b]

_tpiDelete :: [DelPermDef b]

addTableContext :: forall (b :: BackendType). Backend b => TableName b -> Text -> Text

addLogicalModelContext :: LogicalModelLocation -> Text -> Text

boActions :: Lens' BuildOutputs ActionCache

boCustomTypes :: Lens' BuildOutputs AnnotatedCustomTypes

boBackendCache :: Lens' BuildOutputs BackendCache

boRemoteSchemas :: Lens' BuildOutputs (HashMap RemoteSchemaName (RemoteSchemaCtx, MetadataObject))

boRoles :: Lens' BuildOutputs (HashMap RoleName Role)

boSources :: Lens' BuildOutputs SourceCache

buildInfoMap :: (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectItem) arr, Hashable k) => (a -> k) -> (a -> MetadataObject) -> arr (e, a) (Maybe b) -> arr (e, [a]) (HashMap k b)

Processes a list of catalog metadata into a map of processed information, marking any duplicate entries inconsistent.

buildInfoMapM :: (MonadWriter (Seq CollectItem) m, Hashable k) => (a -> k) -> (a -> MetadataObject) -> (a -> m (Maybe b)) -> [a] -> m (HashMap k b)

buildInfoMapPreservingMetadata :: (ArrowChoice arr, ArrowDistribute arr, ArrowWriter (Seq CollectItem) arr, Hashable k) => (a -> k) -> (a -> MetadataObject) -> arr (e, a) (Maybe b) -> arr (e, [a]) (HashMap k (b, MetadataObject))

Like buildInfoMap, but includes each processed info’s associated MetadataObject in the result. This is useful if the results will be further processed, and the MetadataObject is still needed to mark the object inconsistent.

buildInfoMapPreservingMetadataM :: (MonadWriter (Seq CollectItem) m, Hashable k) => (a -> k) -> (a -> MetadataObject) -> (a -> m (Maybe b)) -> [a] -> m (HashMap k (b, MetadataObject))

initialInvalidationKeys :: InvalidationKeys

invalidateKeys :: CacheInvalidations -> InvalidationKeys -> InvalidationKeys

mkTableInputs :: forall (b :: BackendType). TableMetadata b -> (TableBuildInput b, NonColumnTableInputs b, TablePermissionInputs b)

runCacheBuild :: (MonadIO m, MonadError QErr m) => CacheBuildParams -> CacheBuild a -> m a

runCacheBuildM :: (MonadIO m, MonadError QErr m, MonadResolveSource m, ProvidesNetwork m, HasCacheStaticConfig m) => CacheBuild a -> m a

withRecordDependencies :: ArrowWriter (Seq CollectItem) arr => WriterA (Seq SchemaDependency) arr (e, s) a -> arr (e, (MetadataObject, (SchemaObjId, s))) a

data SourcesIntrospectionStatus

The status of collection of stored introspections of remote schemas and data sources.

Constructors

SourcesIntrospectionChangedFull StoredIntrospection
A full introspection collection of all available remote schemas and data sources.

SourcesIntrospectionChangedPartial StoredIntrospection
A partial introspection collection. Does not include all configured remote schemas and data sources, because they were not available.

SourcesIntrospectionUnchanged
None of remote schemas or data sources introspection is refetched.