Hasura.RQL.DDL.Schema.Cache.Permission

Documentation

mkBooleanPermissionMap :: (RoleName -> a) -> HashMap RoleName a -> OrderedRoles -> HashMap RoleName a

newtype OrderedRoles

OrderedRoles is a data type to hold topologically sorted roles according to each role's parent roles, see orderRoles for more details.

Constructors

OrderedRoles

Fields

_unOrderedRoles :: [Role]

orderRoles :: MonadError QErr m => [Role] -> m OrderedRoles

orderRoles is used to order the roles, in such a way that given a role R with n parent roles - PR1, PR2 .. PRn, then the orderRoles function will order the roles in such a way that all the parent roles precede the role R. Note that the order of the parent roles itself doesn't matter as long as they precede the roles on which they are dependent on.

For example, the orderRoles may return `[PR1, PR3, PR2, ... PRn, R]` or `[PR5, PR3, PR1 ... R]`, both of them are correct because all the parent roles precede the inherited role R, assuming the parent roles themselves don't have any parents for the sake of this example.

resolveCheckPermission :: forall arr m p. (ArrowChoice arr, ArrowCache m arr, ArrowWriter (Seq CollectedInfo) arr) => (CheckPermission p, RoleName, InconsistentRoleEntity) `arr` Maybe p

resolveCheckPermission is a helper arrow function which will convert the indermediate type CheckPermission to its original type. It will record any metadata inconsistencies, if exists.

resolveCheckTablePermission :: forall b a arr m. (ArrowChoice arr, ArrowCache m arr, ArrowWriter (Seq CollectedInfo) arr, BackendMetadata b) => (CheckPermission (PermInfo a b), Maybe (RolePermInfo b), RolePermInfo b -> Maybe (PermInfo a b), RoleName, SourceName, TableName b, PermType) `arr` Maybe (PermInfo a b)

buildTablePermissions :: forall b arr m. (ArrowChoice arr, ArrowDistribute arr, ArrowCache m arr, MonadError QErr m, ArrowWriter (Seq CollectedInfo) arr, BackendMetadata b, Cacheable (Proxy b)) => (Proxy b, SourceName, Dependency (TableCoreCache b), FieldInfoMap (FieldInfo b), TablePermissionInputs b, OrderedRoles) `arr` RolePermInfoMap b

mkPermissionMetadataObject :: forall b a. (ToJSON (a b), BackendMetadata b, IsPerm a) => SourceName -> TableName b -> PermDef (a b) -> MetadataObject

mkRemoteSchemaPermissionMetadataObject :: AddRemoteSchemaPermission -> MetadataObject

withPermission :: forall bknd a b c s arr. (ArrowChoice arr, ArrowWriter (Seq CollectedInfo) arr, BackendMetadata bknd, ToJSON (c bknd), IsPerm c) => WriterA (Seq SchemaDependency) (ErrorA QErr arr) (a, s) b -> (a, ((SourceName, TableName bknd, PermDef (c bknd), Proxy bknd), s)) `arr` Maybe b

buildPermission :: forall b a arr m. (ArrowChoice arr, ArrowWriter (Seq CollectedInfo) arr, ArrowCache m arr, Cacheable (a b), Cacheable (Proxy b), MonadError QErr m, BackendMetadata b, ToJSON (a b), IsPerm a) => (Proxy b, Dependency (TableCoreCache b), SourceName, TableName b, FieldInfoMap (FieldInfo b), Maybe (PermDef (a b))) `arr` Maybe (PermInfo a b)