Hasura.RQL.DDL.Schema.Enum

Description

Types and functions for interacting with and manipulating SQL enums represented by single-column tables, not native Postgres enum types. Native enum types in Postgres are difficult to change, so we discourage their use, but we might add support for native enum types in the future.

Re-exports from Hasura.RQL.Types.Column

data EnumReference (b :: BackendType)

Represents a reference to an “enum table,” a single-column Postgres table that is referenced via foreign key.

Constructors

EnumReference

Fields

erTable :: TableName b

erValues :: EnumValues

erTableCustomName :: Maybe Name

type EnumValues = HashMap EnumValue EnumValueInfo

newtype EnumValueInfo

Constructors

EnumValueInfo

Fields

evComment :: Maybe Text

newtype EnumValue

Constructors

EnumValue

Fields

getEnumValue :: Name

Loading table info

resolveEnumReferences :: forall (b :: BackendType). Backend b => HashMap (TableName b) (PrimaryKey b (Column b), TableConfig b, EnumValues) -> HashSet (ForeignKey b) -> HashMap (Column b) (NonEmpty (EnumReference b))

Given a map of enum tables, computes all enum references implied by the given set of foreign keys. A foreign key constitutes an enum reference iff the following conditions hold:

The key only includes a single column.

The referenced column is the table’s primary key.

The referenced table is, in fact, an enum table.