Hasura.RQL.DDL.Schema.Rename

Description

Functions for updating the metadata (with integrity checking) to incorporate schema changes discovered after applying a user-supplied SQL query. None of these functions modify the schema cache, so it must be reloaded after the metadata is updated.

Documentation

data RenameItem (b :: BackendType) a

Constructors

RenameItem

Fields

_riTable :: !(TableName b)

_riOld :: !a

_riNew :: !a

type RenameCol (b :: BackendType) = RenameItem b (Column b)

data RenameField b

Constructors

RFCol !(RenameCol b)

RFRel !(RenameItem b RelName)

type RenameTable b = (TableName b, TableName b)

data Rename b

Constructors

RTable !(RenameTable b)

RField !(RenameField b)

otherDeps :: QErrM m => Text -> SchemaObjId -> m ()

renameTableInMetadata :: forall b m. (MonadError QErr m, CacheRM m, MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> TableName b -> TableName b -> m ()

Replace all references to a given table name by its new name across the entire metadata.

This function will make use of the metadata dependency graph (see getDependentObjs) to identify all places that refer to the old table name, and replace it accordingly. Most operations will occur within the same source, such as table references in relationships and permissions. Dependencies across sources can happen in the case of cross-source relationships.

This function will fail if it encounters a nonsensical dependency; for instance, if there's a dependency from that table to a source.

For more information about the dependency graph, see SchemaObjId.

renameColumnInMetadata :: forall b m. (MonadError QErr m, CacheRM m, MonadWriter MetadataModifier m, BackendMetadata b) => Column b -> Column b -> SourceName -> TableName b -> FieldInfoMap (FieldInfo b) -> m ()

Replace all references to a given column name by its new name across the entire metadata.

This function will make use of the metadata dependency graph (see getDependentObjs) to identify all places that refer to the old column name, and replace it accordingly. Most operations will occur within the same source, such as column references in relationships and permissions. Dependencies across sources can happen in the case of cross-source relationships.

This function will fail if it encounters a nonsensical dependency; for instance, if there's a dependency from that table to a source.

For more information about the dependency graph, see SchemaObjId.

renameRelationshipInMetadata :: forall b m. (MonadError QErr m, CacheRM m, MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> TableName b -> RelName -> RelType -> RelName -> m ()

updateRelDefs :: forall b m. (MonadError QErr m, CacheRM m, MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> TableName b -> RelName -> RenameTable b -> m ()

updatePermFlds :: forall b m. (MonadError QErr m, CacheRM m, MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> TableName b -> RoleName -> PermType -> Rename b -> m ()

update fields in permissions

updateInsPermFlds :: (MonadReader (TableCache b) m, Backend b) => TableName b -> Rename b -> InsPerm b -> m (InsPerm b)

updateSelPermFlds :: (MonadReader (TableCache b) m, Backend b) => TableName b -> Rename b -> SelPerm b -> m (SelPerm b)

updateUpdPermFlds :: (MonadReader (TableCache b) m, Backend b) => TableName b -> Rename b -> UpdPerm b -> m (UpdPerm b)

updateDelPermFlds :: (MonadReader (TableCache b) m, Backend b) => TableName b -> Rename b -> DelPerm b -> m (DelPerm b)

updatePreset :: Backend b => TableName b -> RenameField b -> ColumnValues b Value -> ColumnValues b Value

updateCols :: Backend b => TableName b -> RenameField b -> PermColSpec b -> PermColSpec b

updateTableInBoolExp :: Backend b => RenameTable b -> BoolExp b -> BoolExp b

updateFieldInBoolExp :: (MonadReader (TableCache b) m, Backend b) => TableName b -> RenameField b -> BoolExp b -> m (BoolExp b)

updateColExp :: forall b m. (MonadReader (TableCache b) m, Backend b) => TableName b -> RenameField b -> ColExp -> m ColExp

updateColInRel :: forall b m. (CacheRM m, MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> TableName b -> RelName -> RenameCol b -> m ()

updateColInRemoteRelationshipLHS :: forall b m. (MonadError QErr m, MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> RelName -> RenameCol b -> m ()

Local helper: update a column's name in the left-hand side of a remote relationship.

There are two kinds or remote relationships: remote source relationships, across sources, and remote schema relationships, on remote schemas. In both cases, we maintain a mapping from the source table's colunns to what they should be joined against in the target; when a column is renamed, those references must be renamed as well. This function handles both cases.

See renameColumnInMetadata.

updateColInRemoteRelationshipRHS :: forall source target m. (MonadWriter MetadataModifier m, Backend source, Backend target) => SourceName -> TableName source -> RelName -> RenameCol target -> m ()

Local helper: update a column's name in the right-hand side of a remote relationship.

In the case of remote _source_ relationships, the mapping from column to column needs to be updated if one of the rhs columns has been renamed. A dependency is tracked from the rhs source's column to the lhs source's relationship: when a rhs source's column has been renamed, this function performs the corresponding update in the lhs source's relationship definition.

See renameColumnInMetadata.

updateTableInRemoteRelationshipRHS :: forall source target m. (MonadWriter MetadataModifier m, Backend source, Backend target) => SourceName -> TableName source -> RelName -> RenameTable target -> m ()

Local helper: update a table's name in the right-hand side of a remote relationship.

In the case of remote _source_ relationships, the relationship definition targets a specific table in the rhs source, and that reference needs to be updated if the targeted table has been renamed. A dependency is tracked from the rhs source's table to the lhs source's relationship: when a rhs table has been renamed, this function performs the corresponding update in the lhs source's relationship definition.

See renameTableInMetadata.

updateColInObjRel :: Backend b => TableName b -> TableName b -> RenameCol b -> ObjRelUsing b -> ObjRelUsing b

updateRelChoice :: Backend b => TableName b -> TableName b -> RenameCol b -> ObjRelUsingChoice b -> ObjRelUsingChoice b

updateColInArrRel :: Backend b => TableName b -> TableName b -> RenameCol b -> ArrRelUsing b -> ArrRelUsing b

type ColMap b = HashMap (Column b) (Column b)

getNewCol :: forall b f. Backend b => Functor f => RenameCol b -> TableName b -> f (Column b) -> f (Column b)

updateRelManualConfig :: forall b. Backend b => TableName b -> TableName b -> RenameCol b -> RelManualConfig b -> RelManualConfig b

updateColMap :: forall b. Backend b => TableName b -> TableName b -> RenameCol b -> ColMap b -> ColMap b

possiblyUpdateCustomColumnNames :: forall b m. (MonadWriter MetadataModifier m, BackendMetadata b) => SourceName -> TableName b -> Column b -> Column b -> m ()