Hasura.RQL.DDL.Webhook.Transform

Description

Webhook Transformations are data transformations used to modify HTTP Requests/Responses before requests are executed and after responses are received.

Transformations are supplied by users as part of the Metadata for a particular Action or EventTrigger as a RequestTransform record. Per-field Transformations are stored as data (defunctionalized), often in the form of a Kriti template, and then converted into actual functions (reified) at runtime by the Transform typeclass.

We take a Higher Kinded Data (HKD) approach to representing the transformations. RequestFields is an HKD which can represent the actual request data as 'RequestFields Identity' or the defunctionalized transforms as 'RequestFields (WithOptional TransformFn)'.

We can then traverse over the entire RequestFields HKD to reify all the fields at once and apply them to our actual request data.

NOTE: We don't literally use traverse or the HKD equivalent btraverse, but you can think of this operation morally as a traversal. See applyRequestTransform for implementation details.

Documentation

data RequestTransform

RequestTransform is the metadata representation of a request transformation. It consists of a record of higher kinded data (HKD) along with some regular data. We seperate the HKD data into its own record field called requestFields which we nest inside our non-HKD record. The actual transformation operations are contained in the HKD.

Constructors

RequestTransform

Fields

version :: Version

requestFields :: RequestFields (WithOptional TransformFn)

templateEngine :: TemplatingEngine

type RequestTransformFn = RequestFields (WithOptional TransformFn)

Defunctionalized Webhook Transformation

We represent a defunctionalized request transformation by parameterizing our HKD with WithOptional TransformFn, which marks each of the fields as optional and supplies the appropriate transformation function to them if if they are provided. 'WithOptional _' is equivalent to 'Compose Maybe _'.

type RequestData = RequestFields Identity

Actual Request Data

We represent the actual request data by parameterizing our HKD with Identity, which allows us to trivially unwrap the fields (which should exist after any transformations have been applied).

data RequestFields f

This is our HKD type. It is a record with fields for each component of an Request we wish to transform.

Constructors

RequestFields

Fields

method :: f Method

url :: f Url

body :: f Body

queryParams :: f QueryParams

requestHeaders :: f Headers

request :: Lens' Request RequestData

A lens for mapping from an actual Request term to our HKD.

applyRequestTransform :: MonadError TransformErrorBundle m => (Request -> RequestTransformCtx) -> RequestTransformFn -> Request -> m Request

Transform an Request with a RequestTransform.

Note: we pass in the request url explicitly for use in the ReqTransformCtx. We do this so that we can ensure that the url is syntactically identical to what the use submits. If we use the parsed request from the Request term then it is possible that the url is semantically equivalent but syntactically different. An example of this is the presence or lack of a trailing slash on the URL path. This important when performing string interpolation on the request url.

(.=?) :: ToJSON v => k -> Maybe v -> Maybe (k, Value) infixr 8

newtype WithOptional fn result

WithOptional has Eq and Show instances with difficult to satisfy constraints. WithOptional is isomorphic to 'Compose Maybe f'.

TODO(jkachmar): Add better documentation now that we've specialized this type.

Constructors

WithOptional

Fields

getOptional :: Maybe (fn result)

data ResponseTransform

A set of data transformation functions generated from a MetadataResponseTransform. Nothing means use the original response value.

Constructors

ResponseTransform

Fields

respTransformBody :: Maybe (ResponseTransformCtx -> Either TransformErrorBundle Value)

respTransformTemplateEngine :: TemplatingEngine

data MetadataResponseTransform

Constructors

MetadataResponseTransform

Fields

mrtVersion :: Version

mrtBodyTransform :: Maybe BodyTransformAction

mrtTemplatingEngine :: TemplatingEngine

buildRespTransformCtx :: Maybe RequestTransformCtx -> Maybe SessionVariables -> TemplatingEngine -> ByteString -> ResponseTransformCtx

A helper function for constructing the RespTransformCtx

mkRespTemplateTransform :: TemplatingEngine -> BodyTransformAction -> ResponseTransformCtx -> Either TransformErrorBundle Value

Construct a Template Transformation function for Responses

mkResponseTransform :: MetadataResponseTransform -> ResponseTransform

applyResponseTransform :: ResponseTransform -> ResponseTransformCtx -> Either TransformErrorBundle ByteString

At the moment we only transform the body of Responses. 'http-client' does not export the constructors for Response. If we want to transform then we will need additional apply functions.