Hasura.RQL.DDL.Webhook.Transform.Body

Documentation

validateBodyTransformFn :: TemplatingEngine -> BodyTransformFn -> Validation TransformErrorBundle ()

Validate that the provided BodyTransformFn is correct in the context of a particular TemplatingEngine.

This is a product of the fact that the correctness of a given transformation may be dependent on zero, one, or more of the templated transformations encoded within the given BodyTransformFn.

foldFormEncoded :: HashMap Text ByteString -> ByteString

Fold a HashMap of header key/value pairs into an x-www-form-urlencoded message body.

Orphan instances

Transform Body

Instance details

Methods

transform :: MonadError TransformErrorBundle m => TransformFn Body -> TransformCtx Body -> Body -> m Body

validate :: TemplatingEngine -> TransformFn Body -> Validation TransformErrorBundle ()