Hasura.RQL.DDL.Webhook.Transform.Body

Documentation

data Body

HTTP message body being transformed.

Constructors

JSONBody (Maybe Value)

RawBody ByteString

data BodyTransformFn

The transformations which can be applied to an HTTP message body.

Constructors

Remove
Remove the HTTP message body.

ModifyAsJSON Template
Modify the JSON message body by applying a Template transformation.

ModifyAsFormURLEncoded (HashMap Text UnescapedTemplate)
Modify the JSON message body by applying UnescapedTemplate transformations to each field with a matching Text key.

foldFormEncoded :: HashMap Text ByteString -> ByteString

Fold a HashMap of header key/value pairs into an x-www-form-urlencoded message body.