Hasura.RQL.DDL.Webhook.Transform.Class

Description

The Transform typeclass with various types and helper functions for evaluating transformations.

Documentation

class Transform a where

Transform describes how to reify a defunctionalized transformation for a particular request field.

Methods

transform :: MonadError TransformErrorBundle m => TransformFn a -> TransformCtx a -> a -> m a

transform is a function which takes TransformFn of a and reifies it into a function of the form:

  ReqTransformCtx -> a -> m a

validate :: TemplatingEngine -> TransformFn a -> Validation TransformErrorBundle ()

Validate a TransformFn of a.

throwErrorBundle :: MonadError TransformErrorBundle m => Text -> Maybe Value -> m a

A helper function for serializing transformation errors to JSON.

wrapUnescapedTemplate :: UnescapedTemplate -> Template

Wrap an UnescapedTemplate with escaped double quotes.

encodeScalar :: MonadError TransformErrorBundle m => Value -> m ByteString

Encode a JSON Scalar Value as a ByteString. If a non-Scalar value is provided, will return a TrnasformErrorBundle