Hasura.RQL.DDL.Webhook.Transform.Class

Description

The Transform typeclass with various types and helper functions for evaluating transformations.

Documentation

class Transform a where

Transform describes how to reify a defunctionalized transformation for a particular request field.

Associated Types

data TransformFn a :: Type

The associated type 'TransformFn a' is the defunctionalized version of some transformation that should be applied to a given request field.

In most cases it is some variation on a piece of template text describing the transformation.

Methods

transform :: MonadError TransformErrorBundle m => TransformFn a -> RequestTransformCtx -> a -> m a

transform is a function which takes TransformFn of a and reifies it into a function of the form:

  ReqTransformCtx -> a -> m a

validate :: TemplatingEngine -> TransformFn a -> Validation TransformErrorBundle ()

Validate a TransformFn of a.

newtype TransformErrorBundle

We use collect all transformation failures as a '[J.Value]'.

Constructors

TransformErrorBundle

Fields

tebMessages :: [Value]

throwErrorBundle :: MonadError TransformErrorBundle m => Text -> Maybe Value -> m a

A helper function for serializing transformation errors to JSON.

data RequestTransformCtx

Common context that is made available to all request transformations.

Constructors

RequestTransformCtx

Fields

rtcBaseUrl :: Maybe Value

rtcBody :: Value

rtcSessionVariables :: Value

rtcQueryParams :: Maybe Value

rtcEngine :: TemplatingEngine

rtcFunctions :: HashMap Text (Value -> Either CustomFunctionError Value)

mkReqTransformCtx :: Text -> Maybe SessionVariables -> TemplatingEngine -> Request -> RequestTransformCtx

A smart constructor for constructing the RequestTransformCtx

XXX: This function makes internal usage of decodeUtf8, which throws an impure exception when the supplied ByteString cannot be decoded into valid UTF8 text!

data ResponseTransformCtx

Common context that is made available to all response transformations.

Constructors

ResponseTransformCtx

Fields

responseTransformBody :: Value

responseTransformReqCtx :: Value

responseTransformFunctions :: HashMap Text (Value -> Either CustomFunctionError Value)

responseTransformEngine :: TemplatingEngine

data TemplatingEngine

Available templating engines.

Constructors

Kriti

newtype Template

Textual transformation template.

Constructors

Template

Fields

unTemplate :: Text

runRequestTemplateTransform :: Template -> RequestTransformCtx -> Either TransformErrorBundle Value

A helper function for executing transformations from a Template and a RequestTransformCtx.

NOTE: This and all related funtions are hard-coded to Kriti at the moment. When we add additional template engines this function will need to take a TemplatingEngine parameter.

validateRequestTemplateTransform :: TemplatingEngine -> Template -> Either TransformErrorBundle ()

validateRequestTemplateTransform' :: TemplatingEngine -> Template -> Validation TransformErrorBundle ()

runResponseTemplateTransform :: Template -> ResponseTransformCtx -> Either TransformErrorBundle Value

A helper function for executing transformations from a Template and a ResponseTransformCtx.

NOTE: This and all related funtions are hard-coded to Kriti at the moment. When we add additional template engines this function will need to take a TemplatingEngine parameter.

data Version

RequestTransform Versioning

Constructors

V1

V2

newtype UnescapedTemplate

Validated textual transformation template /for string interpolation only/.

This is necessary due to Kriti not distinguishing between string literals and string templates.

Constructors

UnescapedTemplate

Fields

getUnescapedTemplate :: Text

wrapUnescapedTemplate :: UnescapedTemplate -> Template

Wrap an UnescapedTemplate with escaped double quotes.

runUnescapedRequestTemplateTransform :: RequestTransformCtx -> UnescapedTemplate -> Either TransformErrorBundle ByteString

A helper function for executing Kriti transformations from a UnescapedTemplate and a RequestTrasformCtx.

The difference from runRequestTemplateTransform is that this function will wrap the template text in double quotes before running Kriti.

runUnescapedRequestTemplateTransform' :: RequestTransformCtx -> UnescapedTemplate -> Validation TransformErrorBundle ByteString

Run a Kriti transformation with an unescaped template in Validation instead of Either.

validateRequestUnescapedTemplateTransform :: TemplatingEngine -> UnescapedTemplate -> Either TransformErrorBundle ()

validateRequestUnescapedTemplateTransform' :: TemplatingEngine -> UnescapedTemplate -> Validation TransformErrorBundle ()

runUnescapedResponseTemplateTransform :: ResponseTransformCtx -> UnescapedTemplate -> Either TransformErrorBundle ByteString

Run an UnescapedTemplate with a ResponseTransformCtx.

runUnescapedResponseTemplateTransform' :: ResponseTransformCtx -> UnescapedTemplate -> Validation TransformErrorBundle ByteString

Run an UnescapedTemplate with a ResponseTransformCtx in Validation.

encodeScalar :: MonadError TransformErrorBundle m => Value -> m ByteString

Encode a JSON Scalar Value as a ByteString. If a non-Scalar value is provided, will return a TrnasformErrorBundle