Hasura.RQL.DDL.Webhook.Transform.Headers

Documentation

newtype Headers

The actual header data we are transforming..

This newtype is necessary because otherwise we end up with an orphan instance.

Constructors

Headers [Header]

newtype HeadersTransformFn

The defunctionalized transformation on Headers

Constructors

AddReplaceOrRemove AddReplaceOrRemoveFields
Add or replace matching Headers.

data AddReplaceOrRemoveFields

The user can supply a set of header keys to be filtered from the request and a set of headers to be added to the request.

Constructors

AddReplaceOrRemoveFields

Fields

addOrReplaceHeaders :: [(CI Text, UnescapedTemplate)]

A list of key-value pairs for Headers which should be added (if they don't exist) or replaced (if they do) within the HTTP message.

removeHeaders :: [CI Text]

A list of Header keys which should be removed from the HTTP message.