Hasura.RQL.DDL.Webhook.Transform.Validation

Description

We validate TransformFn terms inside RequestTransform before dispatching Metadata actions in runMetadataQueryV1M. Validation follows the same HKD pattern from applyRequestTransform but using btraverseC to call validate from the Transform class on all the HKD fields.

Documentation

validateRequestTransform :: MonadError TransformErrorBundle m => RequestTransform -> m RequestTransform

Validate all 'TransformFn a' fields in the RequestTransform.

newtype Unvalidated a

Used to annotate that a RequestTransform, or some record containing a RequestTransform has not yet been validated.

Constructors

Unvalidated

Fields

_unUnvalidate :: a

unUnvalidate :: Lens' (Unvalidated a) a

A lens for focusing through Unvalidated in validateTransforms.

newtype Unvalidated1 (f :: k -> Type) (a :: k)

Used to annotate that a higher kinded type containing a RequestTransform has not yet been validated.

This is needed specifically for CreateEventTriggerQuery and any other type that is paramterized by a BackendType.

Constructors

Unvalidated1

Fields

_unUnvalidate1 :: f a

unUnvalidate1 :: Lens' (Unvalidated1 f a) (f a)

A lens for focusing through Unvalidated1 in validateTransforms.

validateTransforms :: MonadError QErr m => LensLike (Either TransformErrorBundle) api api RequestTransform RequestTransform -> (api -> m EncJSON) -> api -> m EncJSON

Used to focus into a records in RQLMetadataV1 and validate any RequestTransform terms present.