Hasura.RQL.IR.Action

Documentation

data ActionFieldG (r :: Type)

Constructors

ACFScalar Name

ACFRemote (ActionRemoteRelationshipSelect r)

ACFExpression Text

ACFNestedObject Name !(ActionFieldsG r)

type ActionFieldsG r = Fields (ActionFieldG r)

type ActionFields = ActionFieldsG Void

data ActionRemoteRelationshipSelect r

Constructors

ActionRemoteRelationshipSelect

Fields

_arrsLHSJoinFields :: HashMap FieldName Name

The fields on the table that are required for the join condition of the remote relationship

_arrsRelationship :: r

The field that captures the relationship r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

_ACFNestedObject :: forall r. Prism' (ActionFieldG r) (Name, ActionFieldsG r)

_ACFExpression :: forall r. Prism' (ActionFieldG r) Text

_ACFRemote :: forall r. Prism' (ActionFieldG r) (ActionRemoteRelationshipSelect r)

_ACFScalar :: forall r. Prism' (ActionFieldG r) Name