Hasura.RQL.IR.Action

Documentation

data ActionFieldG r

Internal representation for a selection of fields on the result of an action. Type parameter r will be either r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

Constructors

ACFScalar Name
Scalar value. G.Name is the original field name from the object type.

ACFRemote (ActionRemoteRelationshipSelect r)
Remote relationship

ACFExpression Text
Constant text value (used for __typename fields)

ACFNestedObject Name (ActionFieldsG r)
Nested object. G.Name is the original field name from the object type.

type ActionFieldsG r = Fields (ActionFieldG r)

type ActionFields = ActionFieldsG Void

data ActionRemoteRelationshipSelect r

Constructors

ActionRemoteRelationshipSelect

Fields

_arrsLHSJoinFields :: HashMap FieldName Name

The fields on the table that are required for the join condition of the remote relationship

_arrsRelationship :: r

The field that captures the relationship r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

_ACFExpression :: forall r p f. (Choice p, Applicative f) => p Text (f Text) -> p (ActionFieldG r) (f (ActionFieldG r))

_ACFNestedObject :: forall r p f. (Choice p, Applicative f) => p (Name, ActionFieldsG r) (f (Name, ActionFieldsG r)) -> p (ActionFieldG r) (f (ActionFieldG r))

_ACFRemote :: forall r p f. (Choice p, Applicative f) => p (ActionRemoteRelationshipSelect r) (f (ActionRemoteRelationshipSelect r)) -> p (ActionFieldG r) (f (ActionFieldG r))

_ACFScalar :: forall r p f. (Choice p, Applicative f) => p Name (f Name) -> p (ActionFieldG r) (f (ActionFieldG r))

data AnnActionExecution r

Constructors

AnnActionExecution

Fields

_aaeName :: ActionName

_aaeOutputType :: GraphQLType

output type

_aaeFields :: ActionFieldsG r

output selection

_aaePayload :: Value

jsonified input arguments

_aaeOutputFields :: ActionOutputFields

to validate the response fields from webhook

_aaeWebhook :: EnvRecord ResolvedWebhook

_aaeHeaders :: [HeaderConf]

_aaeForwardClientHeaders :: Bool

_aaeTimeOut :: Timeout

_aaeRequestTransform :: Maybe RequestTransform

_aaeResponseTransform :: Maybe MetadataResponseTransform

aaeName :: forall r f. Functor f => (ActionName -> f ActionName) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeOutputType :: forall r f. Functor f => (GraphQLType -> f GraphQLType) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeFields :: forall r1 r2 f. Functor f => (ActionFieldsG r1 -> f (ActionFieldsG r2)) -> AnnActionExecution r1 -> f (AnnActionExecution r2)

aaePayload :: forall r f. Functor f => (Value -> f Value) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeOutputFields :: forall r f. Functor f => (ActionOutputFields -> f ActionOutputFields) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeWebhook :: forall r f. Functor f => (EnvRecord ResolvedWebhook -> f (EnvRecord ResolvedWebhook)) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeHeaders :: forall r f. Functor f => ([HeaderConf] -> f [HeaderConf]) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeForwardClientHeaders :: forall r f. Functor f => (Bool -> f Bool) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeTimeOut :: forall r f. Functor f => (Timeout -> f Timeout) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeRequestTransform :: forall r f. Functor f => (Maybe RequestTransform -> f (Maybe RequestTransform)) -> AnnActionExecution r -> f (AnnActionExecution r)

aaeResponseTransform :: forall r f. Functor f => (Maybe MetadataResponseTransform -> f (Maybe MetadataResponseTransform)) -> AnnActionExecution r -> f (AnnActionExecution r)

data AnnActionMutationAsync

Constructors

AnnActionMutationAsync

Fields

_aamaName :: ActionName

_aamaForwardClientHeaders :: Bool

_aamaPayload :: Value

jsonified input arguments

data AsyncActionQueryFieldG r

Constructors

AsyncTypename Text

AsyncOutput (ActionFieldsG r)

AsyncId

AsyncCreatedAt

AsyncErrors

_AsyncTypename :: forall r p f. (Choice p, Applicative f) => p Text (f Text) -> p (AsyncActionQueryFieldG r) (f (AsyncActionQueryFieldG r))

_AsyncOutput :: forall r1 r2 p f. (Choice p, Applicative f) => p (ActionFieldsG r1) (f (ActionFieldsG r2)) -> p (AsyncActionQueryFieldG r1) (f (AsyncActionQueryFieldG r2))

_AsyncId :: forall r p f. (Choice p, Applicative f) => p () (f ()) -> p (AsyncActionQueryFieldG r) (f (AsyncActionQueryFieldG r))

_AsyncCreatedAt :: forall r p f. (Choice p, Applicative f) => p () (f ()) -> p (AsyncActionQueryFieldG r) (f (AsyncActionQueryFieldG r))

_AsyncErrors :: forall r p f. (Choice p, Applicative f) => p () (f ()) -> p (AsyncActionQueryFieldG r) (f (AsyncActionQueryFieldG r))

data AnnActionAsyncQuery (b :: BackendType) r

Constructors

AnnActionAsyncQuery

Fields

_aaaqName :: ActionName

_aaaqActionId :: ActionId

_aaaqOutputType :: GraphQLType

_aaaqFields :: AsyncActionQueryFieldsG r

_aaaqDefinitionList :: [(Column b, ScalarType b)]

_aaaqStringifyNum :: StringifyNumbers

_aaaqForwardClientHeaders :: Bool

_aaaqSource :: ActionSourceInfo b

aaaqName :: forall (b :: BackendType) r f. Functor f => (ActionName -> f ActionName) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

aaaqActionId :: forall (b :: BackendType) r f. Functor f => (ActionId -> f ActionId) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

aaaqOutputType :: forall (b :: BackendType) r f. Functor f => (GraphQLType -> f GraphQLType) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

aaaqFields :: forall (b :: BackendType) r1 r2 f. Functor f => (AsyncActionQueryFieldsG r1 -> f (AsyncActionQueryFieldsG r2)) -> AnnActionAsyncQuery b r1 -> f (AnnActionAsyncQuery b r2)

aaaqDefinitionList :: forall (b :: BackendType) r f. Functor f => ([(Column b, ScalarType b)] -> f [(Column b, ScalarType b)]) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

aaaqStringifyNum :: forall (b :: BackendType) r f. Functor f => (StringifyNumbers -> f StringifyNumbers) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

aaaqForwardClientHeaders :: forall (b :: BackendType) r f. Functor f => (Bool -> f Bool) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

aaaqSource :: forall (b :: BackendType) r f. Functor f => (ActionSourceInfo b -> f (ActionSourceInfo b)) -> AnnActionAsyncQuery b r -> f (AnnActionAsyncQuery b r)

data ActionSourceInfo (b :: BackendType)

Constructors

ASINoSource
No relationships defined on the action output object

ASISource SourceName (SourceConfig b)
All relationships refer to tables in one source

type ActionOutputFields = HashMap Name GType

getActionOutputFields :: AnnotatedOutputType -> ActionOutputFields