Hasura.RQL.IR.BoolExp

Documentation

data GBoolExp (b :: BackendType) a

This type represents a hierarchical boolean expression. It is parametric over the actual implementation of the actual boolean term values. It nonetheless leaks some information: "exists" is only used in permissions, to add conditions based on another table.

Constructors

BoolAnd ![GBoolExp b a]

BoolOr ![GBoolExp b a]

BoolNot !(GBoolExp b a)

BoolExists !(GExists b a)

BoolFld !a

gBoolExpTrue :: GBoolExp b a

data GExists (b :: BackendType) a

Represents a condition on an aribtrary table. Used as part of our permissions boolean expressions. See our documentation for more information: https://hasura.io/docs/latest/graphql/core/auth/authorization/permission-rules.html#using-unrelated-tables-views

Constructors

GExists

Fields

_geTable :: !(TableName b)

_geWhere :: !(GBoolExp b a)

geWhere :: forall b a a. Lens (GExists b a) (GExists b a) (GBoolExp b a) (GBoolExp b a)

geTable :: forall b a. Lens' (GExists b a) (TableName b)

data ColExp

We don't allow conditions across relationships in permissions: the type we use as the terms in GBoolExp is this one, ColExp, which only contains a FieldName and a JSON Value.

Constructors

ColExp

Fields

ceCol :: !FieldName

ceVal :: !Value

newtype BoolExp (b :: BackendType)

This BoolExp type is a simple alias for the boolean expressions used in permissions, that uses ColExp as the term in GBoolExp.

Constructors

BoolExp

Fields

unBoolExp :: GBoolExp b ColExp

_BoolFld :: forall b a. Prism' (GBoolExp b a) a

_BoolExists :: forall b a. Prism' (GBoolExp b a) (GExists b a)

_BoolNot :: forall b a. Prism' (GBoolExp b a) (GBoolExp b a)

_BoolOr :: forall b a. Prism' (GBoolExp b a) [GBoolExp b a]

_BoolAnd :: forall b a. Prism' (GBoolExp b a) [GBoolExp b a]

data PartialSQLExp (b :: BackendType)

Permissions get translated into boolean expressions that are threaded throuhgout the parsers. For the leaf values of those permissions, we use this type, which references but doesn't inline the session variables.

Constructors

PSESessVar !(SessionVarType b) !SessionVariable

PSESession

PSESQLExp !(SQLExpression b)

isStaticValue :: PartialSQLExp backend -> Bool

hasStaticExp :: Backend b => OpExpG b (PartialSQLExp b) -> Bool

type CastExp b a = HashMap (ScalarType b) [OpExpG b a]

Operand for cast operator

data OpExpG (b :: BackendType) a

This type represents the boolean operators that can be applied on values of a column. This type only contains the common core, that we expect to be ultimately entirely supported in most if not all backends. Backends can extend this with the BooleanOperators type in Backend.

Constructors

ACast !(CastExp b a)

AEQ !Bool !a

ANE !Bool !a

AIN !a

ANIN !a

AGT !a

ALT !a

AGTE !a

ALTE !a

ALIKE !a

ANLIKE !a

CEQ (RootOrCurrentColumn b)

CNE (RootOrCurrentColumn b)

CGT (RootOrCurrentColumn b)

CLT (RootOrCurrentColumn b)

CGTE (RootOrCurrentColumn b)

CLTE (RootOrCurrentColumn b)

ANISNULL

ANISNOTNULL

ABackendSpecific !(BooleanOperators b a)

data RootOrCurrentColumn b

Constructors

RootOrCurrentColumn RootOrCurrent (Column b)

data RootOrCurrent

The arguments of column-operators may refer to either the so-called 'root tabular value' or 'current tabular value'.

Constructors

IsRoot

IsCurrent

opExpDepCol :: OpExpG backend a -> Maybe (RootOrCurrentColumn backend)

data SessionArgumentPresence a

The presence of session argument in the SQL function of a computed field. Since we only support computed fields with SQL functions having maximum of 2 arguments in boolean expression, the position (if present) is either first or second. The other mandatory argument is table row input.

Constructors

SAPNotPresent

SAPFirst a

SAPSecond a

mkSessionArgumentPresence :: forall v a. v -> Maybe a -> FunctionTableArgument -> SessionArgumentPresence v

Determine the position of session argument

data ComputedFieldBoolExp (b :: BackendType) a

This type is used to represent the kinds of boolean expression used for compouted fields based on the return type of the SQL function

Constructors

CFBEScalar ![OpExpG b a]

CFBETable !(TableName b) !(AnnBoolExp b a)

data AnnComputedFieldBoolExp (b :: BackendType) a

Using a computed field in boolean expression. Example: A computed field "full_name" ("first_name" || "last_name") is defined to the "user" table. Boolean expression to filter whose "full_name" is LIKE "%bob%" query { user(where: {full_name: {_like: "%bob%"}}){ id first_name last_name full_name } } Limitation: Computed field whose function with no input arguments are allowed in boolean expression. It is complex to generate schema for `where` clause with functions having input arguments.

Constructors

AnnComputedFieldBoolExp

Fields

_acfbXFieldInfo :: !(XComputedField b)

_acfbName :: !ComputedFieldName

_acfbFunction :: !(FunctionName b)

_acfbSessionArgumentPresence :: !(SessionArgumentPresence a)

_acfbBoolExp :: !(ComputedFieldBoolExp b a)

data AnnBoolExpFld (b :: BackendType) a

This type is used for boolean terms in GBoolExp in the schema; there are two kinds boolean terms: - operators on a column of the current table, using the OpExpG kind of operators - arbitrary expressions on columns of tables in relationships (in the same source)

This type is parametric over the type of leaf values, the values on which we operate.

Constructors

AVColumn !(ColumnInfo b) ![OpExpG b a]

AVRelationship !(RelInfo b) !(AnnBoolExp b a)

AVComputedField !(AnnComputedFieldBoolExp b a)

type AnnBoolExp b a = GBoolExp b (AnnBoolExpFld b a)

A simple alias for the kind of boolean expressions used in the schema, that ties together GBoolExp, OpExpG, and AnnBoolExpFld.

type AnnBoolExpFldSQL b = AnnBoolExpFld b (SQLExpression b)

type AnnBoolExpSQL b = AnnBoolExp b (SQLExpression b)

type AnnBoolExpPartialSQL b = AnnBoolExp b (PartialSQLExp b)

annBoolExpTrue :: AnnBoolExp backend a

andAnnBoolExps :: AnnBoolExp backend a -> AnnBoolExp backend a -> AnnBoolExp backend a

data DWithinGeomOp a

Operand for STDWithin opoerator

Constructors

DWithinGeomOp

Fields

dwgeomDistance :: !a

dwgeomFrom :: !a

data DWithinGeogOp a

Operand for STDWithin opoerator

Constructors

DWithinGeogOp

Fields

dwgeogDistance :: !a

dwgeogFrom :: !a

dwgeogUseSpheroid :: !a

data STIntersectsNbandGeommin a

Operand for STIntersect

Constructors

STIntersectsNbandGeommin

Fields

singNband :: !a

singGeommin :: !a

data STIntersectsGeomminNband a

Operand for STIntersect

Constructors

STIntersectsGeomminNband

Fields

signGeommin :: !a

signNband :: !(Maybe a)

newtype AnnColumnCaseBoolExpField (b :: BackendType) a

This is a simple newtype over AnnBoolExpFld. At time of writing, I do not know why we want this, and why it exists. It might be a relic of a needed differentiation, now lost? TODO: can this be removed?

Constructors

AnnColumnCaseBoolExpField

Fields

_accColCaseBoolExpField :: AnnBoolExpFld b a

type AnnColumnCaseBoolExp b a = GBoolExp b (AnnColumnCaseBoolExpField b a)

Similar to AnnBoolExp, this type alias ties together GBoolExp, OpExpG, and AnnColumnCaseBoolExpFld.

type AnnColumnCaseBoolExpPartialSQL b = AnnColumnCaseBoolExp b (PartialSQLExp b)

type PreSetColsG b v = HashMap (Column b) v

type PreSetColsPartial b = HashMap (Column b) (PartialSQLExp b)