Hasura.RQL.IR.BoolExp

Description

Boolean Expressions

This module defines the IR representation of boolean expressions used in _where clauses in GraphQL queries, permissions, and so on.

The types in this module define a generic structure with "holes" to be filled by each backend. Specifically, holes will include things like types for table names, and backend field types.

Documentation

newtype BoolExp (b :: BackendType)

This BoolExp type is a simple alias for the boolean expressions used in permissions, that uses ColExp as the term in GBoolExp.

Constructors

BoolExp

Fields

unBoolExp :: GBoolExp b ColExp

data ColExp

We don't allow conditions across relationships in permissions: the type we use as the terms in GBoolExp is this one, ColExp, which only contains a FieldName and a JSON Value.

Constructors

ColExp

Fields

ceCol :: FieldName

ceVal :: Value

data GBoolExp (backend :: BackendType) field

This type represents a boolean expression tree. It is parametric over the actual implementation of the actual boolean term values. It nonetheless leaks some information: "exists" is only used in permissions, to add conditions based on another table.

The backend parameter is used to find the backend-specific type for table names in the BoolExists constructor.

The field type represent the type of database-specific field types.

Constructors

BoolAnd [GBoolExp backend field]

BoolOr [GBoolExp backend field]

BoolNot (GBoolExp backend field)

BoolExists (GExists backend field)
Represents a condition on an aribtrary table. since the backend and field are the same, the table must be of the same database type.

BoolField field
A column field

gBoolExpTrue :: forall (backend :: BackendType) field. GBoolExp backend field

A default representation for a true boolean value.

data GExists (backend :: BackendType) field

Represents a condition on an aribtrary table. Used as part of our permissions boolean expressions. See our documentation for more information: https://hasura.io/docs/latest/graphql/core/auth/authorization/permission-rules.html#using-unrelated-tables-views

Constructors

GExists

Fields

_geTable :: TableName backend

_geWhere :: GBoolExp backend field

data DWithinGeomOp field

Operand for STDWithin opoerator

Constructors

DWithinGeomOp

Fields

dwgeomDistance :: field

dwgeomFrom :: field

data DWithinGeogOp field

Operand for STDWithin opoerator

Constructors

DWithinGeogOp

Fields

dwgeogDistance :: field

dwgeogFrom :: field

dwgeogUseSpheroid :: field

type CastExp (backend :: BackendType) field = HashMap (ScalarType backend) [OpExpG backend field]

Operand for cast operator

data OpExpG (backend :: BackendType) field

This type represents the boolean operators that can be applied on values of a column. This type only contains the common core, that we expect to be ultimately entirely supported in most if not all backends. Backends can extend this with the BooleanOperators type in Backend.

Constructors

ACast (CastExp backend field)

AEQ ComparisonNullability field

ANE ComparisonNullability field

AIN field

ANIN field

AGT field

ALT field

AGTE field

ALTE field

ALIKE field

ANLIKE field

CEQ (RootOrCurrentColumn backend)

CNE (RootOrCurrentColumn backend)

CGT (RootOrCurrentColumn backend)

CLT (RootOrCurrentColumn backend)

CGTE (RootOrCurrentColumn backend)

CLTE (RootOrCurrentColumn backend)

ANISNULL

ANISNOTNULL

ABackendSpecific (BooleanOperators backend field)

opExpDepCol :: forall (backend :: BackendType) field. OpExpG backend field -> Maybe (RootOrCurrentColumn backend)

data ComparisonNullability

Constructors

NonNullableComparison

NullableComparison

data STIntersectsNbandGeommin field

Operand for STIntersect

Constructors

STIntersectsNbandGeommin

Fields

singNband :: field

singGeommin :: field

data STIntersectsGeomminNband field

Operand for STIntersect

Constructors

STIntersectsGeomminNband

Fields

signGeommin :: field

signNband :: Maybe field

data ComputedFieldBoolExp (backend :: BackendType) scalar

This type is used to represent the kinds of boolean expression used for computed fields based on the return type of the SQL function.

Constructors

CFBEScalar (AnnRedactionExp backend scalar) [OpExpG backend scalar]
SQL function returning a scalar

CFBETable (TableName backend) (AnnBoolExp backend scalar)
SQL function returning SET OF table

data AnnComputedFieldBoolExp (backend :: BackendType) scalar

Using a computed field in boolean expression. Example: A computed field "full_name" ("first_name" || "last_name") is defined to the "user" table. Boolean expression to filter whose "full_name" is LIKE "%bob%" query { user(where: {full_name: {_like: "%bob%"}}){ id first_name last_name full_name } } Limitation: We only support computed fields in boolean expressions when they are functions with no input arguments, because it is complex to generate schema for where clauses for functions that have input arguments.

Constructors

AnnComputedFieldBoolExp

Fields

_acfbXFieldInfo :: XComputedField backend

_acfbName :: ComputedFieldName

_acfbFunction :: FunctionName backend

_acfbFunctionArgsExp :: FunctionArgsExp backend scalar

_acfbBoolExp :: ComputedFieldBoolExp backend scalar

data AnnBoolExpFld (backend :: BackendType) leaf

This type is used for boolean terms in GBoolExp in the schema; there are four kinds boolean terms: - operators on a column of the current table, using the OpExpG kind of operators - arbitrary expressions on columns of tables in relationships (in the same source) - A computed field of the current table - aggregation operations on array relationships on the current tables.

This type is parameterized over the type of leaf values, the values on which we operate.

Constructors

AVColumn (ColumnInfo backend) (AnnRedactionExp backend leaf) [OpExpG backend leaf]

AVNestedObject (NestedObjectInfo backend) (AnnBoolExp backend leaf)

AVRelationship (RelInfo backend) (RelationshipFilters backend leaf)

AVComputedField (AnnComputedFieldBoolExp backend leaf)

AVAggregationPredicates (AggregationPredicates backend leaf)

AVRemoteRelationship (RemoteRelPermBoolExp backend leaf)

data RelationshipFilters (backend :: BackendType) leaf

This type represents a boolean expression over a relationship. In addition to the actual user-specified predicate, we need to also consider the permissions of the target table.

Because the permissions may include column-comparison-operators, they need to be translated in the context of the table they apply to. Thus we keep the permissions and filters separate.

Constructors

RelationshipFilters

Fields

rfTargetTablePermissions :: AnnBoolExp backend leaf

rfFilter :: AnnBoolExp backend leaf

type AnnBoolExp (backend :: BackendType) scalar = GBoolExp backend (AnnBoolExpFld backend scalar)

A simple alias for the kind of boolean expressions used in the schema, that ties together GBoolExp, OpExpG, and AnnBoolExpFld.

type AnnRedactionExpPartialSQL (b :: BackendType) = AnnRedactionExp b (PartialSQLExp b)

type AnnRedactionExpUnpreparedValue (b :: BackendType) = AnnRedactionExp b (UnpreparedValue b)

data AnnRedactionExp (b :: BackendType) v

This captures a boolean expression where, if it is false, some associated data needs to be redacted (in practice, nulled out) because the user doesn't have access to it. Alternatively, "no redaction" is explicitly defined, which is used as an optimization to avoid evaluating a boolexp if unnecessary (as opposed to defining a boolean exp which always evaluates to true).

Constructors

NoRedaction

RedactIfFalse (GBoolExp b (AnnBoolExpFld b v))

annBoolExpTrue :: forall (backend :: BackendType) scalar. AnnBoolExp backend scalar

andAnnBoolExps :: forall (backend :: BackendType) scalar. AnnBoolExp backend scalar -> AnnBoolExp backend scalar -> AnnBoolExp backend scalar

type AnnBoolExpFldSQL (backend :: BackendType) = AnnBoolExpFld backend (SQLExpression backend)

type AnnBoolExpSQL (backend :: BackendType) = AnnBoolExp backend (SQLExpression backend)

data PartialSQLExp (backend :: BackendType)

Permissions get translated into boolean expressions that are threaded throuhgout the parsers. For the leaf values of those permissions, we use this type, which references but doesn't inline the session variables.

Constructors

PSESessVar (SessionVarType backend) SessionVariable

PSESession

PSESQLExp (SQLExpression backend)

isStaticValue :: forall (backend :: BackendType). PartialSQLExp backend -> Bool

hasStaticExp :: forall (b :: BackendType). Backend b => OpExpG b (PartialSQLExp b) -> Bool

type AnnBoolExpPartialSQL (backend :: BackendType) = AnnBoolExp backend (PartialSQLExp backend)

type PreSetColsG (b :: BackendType) v = HashMap (Column b) v

type PreSetColsPartial (b :: BackendType) = HashMap (Column b) (PartialSQLExp b)

data RootOrCurrentColumn (b :: BackendType)

Constructors

RootOrCurrentColumn RootOrCurrent (Column b)

data RootOrCurrent

The arguments of column-operators may refer to either the so-called 'root tabular value' or 'current tabular value'.

Constructors

IsRoot

IsCurrent

data RemoteRelPermBoolExp (b :: BackendType) field

Constructors

RemoteRelPermBoolExp

Fields

rawBoolExp :: (RelName, Value)

lhsCol :: (Column b, ColumnType b)

rhsFetchInfo :: AnyBackend (RemoteRelRHSFetchInfo field)

data RemoteRelRHSFetchInfo field (b :: BackendType)

Constructors

RemoteRelRHSFetchInfo

Fields

rrrfiColumn :: (ScalarType b, Column b)

rrrfiTable :: TableName b

rrrfiWhere :: RemoteRelRHSFetchWhereExp (Column b)

rrrfiSource :: SourceName

rrrfiSourceConfig :: SourceConfig b