Hasura.RQL.IR.BoolExp.AggregationPredicates

Description

This module contains the default types and function that model aggregation predicates.

Documentation

data AggregationPredicatesImplementation (b :: BackendType) field

This type the default non-empty implementation of the AggregationPredicates type family of 'class Backend'.

This represents an _applied_ aggregation predicate, i.e. _not_ an aggegation function in isolation.

In the default schema implementation, this type results from parsing graphql such as:

table(_where( relation_aggregate: {functionname: { arguments: arguments, predicate: predicate, distinct: bool } } )) { ... }

Note that we make no attempt at modelling window functions or so-called analytical functions such as percentile_cont.

Constructors

AggregationPredicatesImplementation

Fields

aggRelation :: RelInfo b

aggRowPermission :: AnnBoolExp b field

aggPredicate :: AggregationPredicate b field

data AggregationPredicate (b :: BackendType) field

Constructors

AggregationPredicate

Fields

aggPredFunctionName :: Text

aggPredDistinct :: Bool

aggPredFilter :: Maybe (AnnBoolExp b field)

aggPredArguments :: AggregationPredicateArguments b field

aggPredPredicate :: [OpExpG b field]

data AggregationPredicateArguments (b :: BackendType) field

Constructors

AggregationPredicateArgumentsStar

AggregationPredicateArguments (NonEmpty (Column b, AnnRedactionExp b field))