Hasura.RQL.IR.Root

Documentation

data SourceConfigWith (db :: BackendType -> Type) (b :: BackendType)

Constructors

SourceConfigWith (SourceConfig b) (Maybe QueryTagsConfig) (db b)

data RootField (db :: BackendType -> Type) remote action raw where

Constructors

RFDB :: forall (db :: BackendType -> Type) remote action raw. SourceName -> AnyBackend (SourceConfigWith db) -> RootField db remote action raw

RFRemote :: forall remote (db :: BackendType -> Type) action raw. RemoteSchemaName -> remote -> RootField db remote action raw

RFAction :: forall action (db :: BackendType -> Type) remote raw. action -> RootField db remote action raw

RFRaw :: forall raw (db :: BackendType -> Type) remote action. raw -> RootField db remote action raw

RFMulti :: forall (db :: BackendType -> Type) remote action raw. [RootField db remote action raw] -> RootField db remote action raw

data MutationDB (b :: BackendType) r v

Constructors

MDBInsert (AnnotatedInsert b r v)

MDBUpdate (AnnotatedUpdateG b r v)

MDBDelete (AnnDelG b r v)

MDBFunction JsonAggSelect (AnnSimpleSelectG b r v)
This represents a VOLATILE function, and is AnnSimpleSelG for easy re-use of non-VOLATILE function tracking code.

data ActionQuery r

Constructors

AQQuery (AnnActionExecution r)

AQAsync (AnnActionAsyncQuery ('Postgres 'Vanilla) r)

data ActionMutation r

Constructors

AMSync (AnnActionExecution r)

AMAsync AnnActionMutationAsync

type QueryRootField (v :: BackendType -> Type) = RootField (QueryDBRoot (RemoteRelationshipField v) v) (RemoteSchemaRootField (RemoteRelationshipField v) RemoteSchemaVariable) (QueryActionRoot v) Value

type MutationRootField (v :: BackendType -> Type) = RootField (MutationDBRoot (RemoteRelationshipField v) v) (RemoteSchemaRootField (RemoteRelationshipField v) RemoteSchemaVariable) (MutationActionRoot v) Value

type SubscriptionRootField (v :: BackendType -> Type) = RootField (QueryDBRoot (RemoteRelationshipField v) v) (RemoteSchemaRootField (RemoteRelationshipField v) RemoteSchemaVariable) (QueryActionRoot v) Value

newtype QueryDBRoot r (v :: BackendType -> Type) (b :: BackendType)

Constructors

QDBR (QueryDB b r (v b))

newtype MutationDBRoot r (v :: BackendType -> Type) (b :: BackendType)

Constructors

MDBR (MutationDB b r (v b))

data RemoteRelationshipField (vf :: BackendType -> Type)

IR of a remote relationship. A remote relationship currently can be to either a remote schema or a database's table. See RemoteSourceSelect for explanation on vf.

Constructors

RemoteSchemaField (RemoteSchemaSelect (RemoteRelationshipField vf))

RemoteSourceField (AnyBackend (RemoteSourceSelect (RemoteRelationshipField vf) vf))
AnyBackend is used here to capture a relationship to an arbitrary target