Hasura.RQL.IR.Select

Description

This modules defines the tree of Select types: how we represent a query internally, from its top level QueryDB down to each individual field. Most of those types have three type arguments:

b: BackendType The backend that is targeted by that specific select (Postgres Vanilla, MSSQL...); we use the type families in the Backend class to decide how different parts of the IR are represented in different backends.

v: Type The type of the leaf values in our AST; used almost exclusively for column values, over which queries can be parameterized. The output of the parser phase will use UnpreparedValue b for the leaves, and most backends will then transform the AST to interpret those values and consequently change v to be SQLExpression b

r: BackendType -> Type Joins across backends mean that the aforementioned b parameter won't be the same throughout the entire tree; at some point we will have an AnyBackend used to encapsulate a branch that uses a different b. We still want, however, to be able to parameterize the values of the leaves in that separate branch, and that's what the r parameter is for. We also use UnpreparedValue here during the parsing phase, meaning all leaf values will be UnpreparedValue b for their respective backend b, and most backends will then transform their AST, cutting all such remote branches, and therefore using Const Void for r.

Documentation

data QueryDB (b :: BackendType) (r :: Type) v

Constructors

QDBMultipleRows (AnnSimpleSelectG b r v)

QDBSingleRow (AnnSimpleSelectG b r v)

QDBAggregation (AnnAggregateSelectG b r v)

QDBConnection (ConnectionSelect b r v)

QDBStreamMultipleRows (AnnSimpleStreamSelectG b r v)

data AnnSelectG (b :: BackendType) (f :: Type -> Type) (v :: Type)

Constructors

AnnSelectG

Fields

_asnFields :: Fields (f v)

_asnFrom :: SelectFromG b v

_asnPerm :: TablePermG b v

_asnArgs :: SelectArgsG b v

_asnStrfyNum :: StringifyNumbers

data AnnSelectStreamG (b :: BackendType) (f :: Type -> Type) (v :: Type)

IR type representing nodes for streaming subscriptions

Constructors

AnnSelectStreamG

Fields

_assnXStreamingSubscription :: XStreamingSubscription b

type to indicate if streaming subscription has been enabled in the BackendType. This type helps avoiding missing case match patterns for backends where it's disabled.

_assnFields :: Fields (f v)

output selection fields

_assnFrom :: SelectFromG b v

table information to select from

_assnPerm :: TablePermG b v

select permissions

_assnArgs :: SelectStreamArgsG b v

streaming arguments

_assnStrfyNum :: StringifyNumbers

type AnnSimpleSelectG b r v = AnnSelectG b (AnnFieldG b r) v

type AnnAggregateSelectG b r v = AnnSelectG b (TableAggregateFieldG b r) v

type AnnSimpleStreamSelectG b r v = AnnSelectStreamG b (AnnFieldG b r) v

type AnnSimpleSelect b = AnnSimpleSelectG b Void (SQLExpression b)

type AnnAggregateSelect b = AnnAggregateSelectG b Void (SQLExpression b)

type AnnSimpleStreamSelect b = AnnSimpleStreamSelectG b Void (SQLExpression b)

bifoldMapAnnSelectG :: (Backend b, Bifoldable (f b), Monoid m) => (r -> m) -> (v -> m) -> AnnSelectG b (f b r) v -> m

We can't write a Bifoldable instance for AnnSelectG because the types don't line up. Instead, we provide this function which can be used to help define Bifoldable instances of other types containing AnnSelectG values.

bifoldMapAnnSelectStreamG :: (Backend b, Bifoldable (f b), Monoid m) => (r -> m) -> (v -> m) -> AnnSelectStreamG b (f b r) v -> m

data ConnectionSelect (b :: BackendType) (r :: Type) v

Constructors

ConnectionSelect

Fields

_csXRelay :: XRelay b

_csPrimaryKeyColumns :: PrimaryKeyColumns b

_csSplit :: Maybe (NonEmpty (ConnectionSplit b v))

_csSlice :: Maybe ConnectionSlice

_csSelect :: AnnSelectG b (ConnectionField b r) v

data ConnectionSplit (b :: BackendType) v

Constructors

ConnectionSplit

Fields

_csKind :: ConnectionSplitKind

_csValue :: v

_csOrderBy :: OrderByItemG b (AnnotatedOrderByElement b v)

data ConnectionSlice

Constructors

SliceFirst Int

SliceLast Int

data ConnectionSplitKind

Constructors

CSKBefore

CSKAfter

newtype FIIdentifier

Identifier used exclusively as the argument to FromIdentifier

Constructors

FIIdentifier

Fields

unFIIdentifier :: Text

data SelectFromG (b :: BackendType) v

Constructors

FromTable (TableName b)

FromIdentifier FIIdentifier

FromFunction (FunctionName b) (FunctionArgsExp b v) (Maybe [(Column b, ScalarType b)])

type SelectFrom b = SelectFromG b (SQLExpression b)

data SelectStreamArgsG (b :: BackendType) v

Constructors

SelectStreamArgsG

Fields

_ssaWhere :: !(Maybe (AnnBoolExp b v))

optional filter to filter the stream results

_ssaBatchSize :: !Int

maximum number of rows to be returned in a single fetch

_ssaCursorArg :: !(StreamCursorItem b)

info related to the cursor column, a single item data type currently because only single column cursors are supported

type SelectStreamArgs b = SelectStreamArgsG b (SQLExpression b)

data SelectArgsG (b :: BackendType) v

Constructors

SelectArgs

Fields

_saWhere :: Maybe (AnnBoolExp b v)

_saOrderBy :: Maybe (NonEmpty (AnnotatedOrderByItemG b v))

_saLimit :: Maybe Int

_saOffset :: Maybe Int64

_saDistinct :: Maybe (NonEmpty (Column b))

type SelectArgs b = SelectArgsG b (SQLExpression b)

noSelectArgs :: SelectArgsG backend v

data ComputedFieldOrderByElement (b :: BackendType) v

The order by element for a computed field based on its return type

Constructors

CFOBEScalar (ScalarType b)
Sort by the scalar computed field

CFOBETableAggregation

Fields

(TableName b)

(AnnBoolExp b v)

Permission filter of the retuning table

(AnnotatedAggregateOrderBy b)

Sort by aggregation fields of table rows returned by computed field

data ComputedFieldOrderBy (b :: BackendType) v

Constructors

ComputedFieldOrderBy

Fields

_cfobXField :: XComputedField b

_cfobName :: ComputedFieldName

_cfobFunction :: FunctionName b

_cfobFunctionArgsExp :: FunctionArgsExp b v

_cfobOrderByElement :: ComputedFieldOrderByElement b v

data AnnotatedOrderByElement (b :: BackendType) v

Constructors

AOCColumn (ColumnInfo b)

AOCObjectRelation

Fields

(RelInfo b)

(AnnBoolExp b v)

Permission filter of the remote table to which the relationship is defined

(AnnotatedOrderByElement b v)

AOCArrayAggregation

Fields

(RelInfo b)

(AnnBoolExp b v)

Permission filter of the remote table to which the relationship is defined

(AnnotatedAggregateOrderBy b)

AOCComputedField (ComputedFieldOrderBy b v)

data AnnotatedAggregateOrderBy (b :: BackendType)

Constructors

AAOCount

AAOOp Text !(ColumnInfo b)

type AnnotatedOrderByItemG b v = OrderByItemG b (AnnotatedOrderByElement b v)

type AnnotatedOrderByItem b = AnnotatedOrderByItemG b (SQLExpression b)

data StreamCursorItem (b :: BackendType)

Cursor for streaming subscription

Constructors

StreamCursorItem

Fields

_sciOrdering :: !CursorOrdering

Specifies how the cursor item should be ordered

_sciColInfo :: !(ColumnInfo b)

Column info of the cursor item

_sciInitialValue :: !(ColumnValue b)

Initial value of the cursor item from where the streaming should start

data RemoteRelationshipSelect b r

captures a remote relationship's selection and the necessary context

Constructors

RemoteRelationshipSelect

Fields

_rrsLHSJoinFields :: HashMap FieldName (DBJoinField b)

The fields on the table that are required for the join condition of the remote relationship

_rrsRelationship :: r

The field that captures the relationship r ~ (RemoteRelationshipField UnpreparedValue) when the AST is emitted by the parser. r ~ Void when an execution tree is constructed so that a backend is absolved of dealing with remote relationships.

data AnnFieldG (b :: BackendType) (r :: Type) v

Constructors

AFColumn (AnnColumnField b v)

AFObjectRelation (ObjectRelationSelectG b r v)

AFArrayRelation (ArraySelectG b r v)

AFComputedField (XComputedField b) ComputedFieldName (ComputedFieldSelect b r v)

AFRemote (RemoteRelationshipSelect b r)
A remote relationship field

AFNodeId (XRelay b) SourceName (TableName b) (PrimaryKeyColumns b)

AFExpression Text

type AnnField b = AnnFieldG b Void (SQLExpression b)

type AnnFields b = AnnFieldsG b Void (SQLExpression b)

mkAnnColumnField :: Column backend -> ColumnType backend -> Maybe (AnnColumnCaseBoolExp backend v) -> Maybe (ScalarSelectionArguments backend) -> AnnFieldG backend r v

mkAnnColumnFieldAsText :: ColumnInfo backend -> AnnFieldG backend r v

traverseSourceRelationshipSelection :: (Applicative f, Backend backend) => (vf backend -> f (vg backend)) -> SourceRelationshipSelection backend r vf -> f (SourceRelationshipSelection backend r vg)

data TableAggregateFieldG (b :: BackendType) (r :: Type) v

Constructors

TAFAgg (AggregateFields b)

TAFNodes (XNodesAgg b) (AnnFieldsG b r v)

TAFExp Text

data AggregateField (b :: BackendType)

Constructors

AFCount (CountType b)

AFOp (AggregateOp b)

AFExp Text

data AggregateOp (b :: BackendType)

Constructors

AggregateOp

Fields

_aoOp :: Text

_aoFields :: ColumnFields b

data ColFld (b :: BackendType)

Constructors

CFCol (Column b) !(ColumnType b)

CFExp Text

type TableAggregateField b = TableAggregateFieldG b Void (SQLExpression b)

type TableAggregateFields b = TableAggregateFieldsG b Void (SQLExpression b)

type TableAggregateFieldsG b r v = Fields (TableAggregateFieldG b r v)

type ColumnFields b = Fields (ColFld b)

type AggregateFields b = Fields (AggregateField b)

type AnnFieldsG b r v = Fields (AnnFieldG b r v)

data ConnectionField (b :: BackendType) (r :: Type) v

Constructors

ConnectionTypename Text

ConnectionPageInfo PageInfoFields

ConnectionEdges (EdgeFields b r v)

data PageInfoField

Constructors

PageInfoTypename Text

PageInfoHasNextPage

PageInfoHasPreviousPage

PageInfoStartCursor

PageInfoEndCursor

data EdgeField (b :: BackendType) (r :: Type) v

Constructors

EdgeTypename Text

EdgeCursor

EdgeNode (AnnFieldsG b r v)

type ConnectionFields b r v = Fields (ConnectionField b r v)

type PageInfoFields = Fields PageInfoField

type EdgeFields b r v = Fields (EdgeField b r v)

data AnnColumnField (b :: BackendType) v

Constructors

AnnColumnField

Fields

_acfColumn :: Column b

_acfType :: ColumnType b

_acfAsText :: Bool

If this field is True, columns are explicitly casted to text when fetched, which avoids an issue that occurs because we don’t currently have proper support for array types. See https://github.com/hasura/graphql-engine/pull/3198 for more details.

_acfArguments :: Maybe (ScalarSelectionArguments b)

Arguments of this column's selection. See ScalarSelectionArguments

_acfCaseBoolExpression :: Maybe (AnnColumnCaseBoolExp b v)

This type is used to determine whether the column should be nullified. When the value is Nothing, the column value will be outputted as computed and when the value is `Just c`, the column will be outputted when c evaluates to true and null when c evaluates to false.

data ComputedFieldScalarSelect (b :: BackendType) v

Constructors

ComputedFieldScalarSelect

Fields

_cfssFunction :: FunctionName b

_cfssArguments :: FunctionArgsExp b v

_cfssType :: ScalarType b

_cfssScalarArguments :: Maybe (ScalarSelectionArguments b)

data ComputedFieldSelect (b :: BackendType) (r :: Type) v

Constructors

CFSScalar

Fields

(ComputedFieldScalarSelect b v)

Type containing info about the computed field

(Maybe (AnnColumnCaseBoolExp b v))

This type is used to determine if whether the scalar computed field should be nullified. When the value is Nothing, the scalar computed value will be outputted as computed and when the value is `Just c`, the scalar computed field will be outputted when c evaluates to true and null when c evaluates to false

CFSTable JsonAggSelect !(AnnSimpleSelectG b r v)

data AnnRelationSelectG (b :: BackendType) a

Constructors

AnnRelationSelectG

Fields

_aarRelationshipName :: RelName

_aarColumnMapping :: HashMap (Column b) (Column b)

_aarAnnSelect :: a

type ArrayRelationSelectG b r v = AnnRelationSelectG b (AnnSimpleSelectG b r v)

type ArrayAggregateSelectG b r v = AnnRelationSelectG b (AnnAggregateSelectG b r v)

type ArrayConnectionSelect b r v = AnnRelationSelectG b (ConnectionSelect b r v)

type ArrayAggregateSelect b = ArrayAggregateSelectG b Void (SQLExpression b)

data AnnObjectSelectG (b :: BackendType) (r :: Type) v

Constructors

AnnObjectSelectG

Fields

_aosFields :: AnnFieldsG b r v

_aosTableFrom :: TableName b

_aosTableFilter :: AnnBoolExp b v

type AnnObjectSelect b r = AnnObjectSelectG b r (SQLExpression b)

type ObjectRelationSelectG b r v = AnnRelationSelectG b (AnnObjectSelectG b r v)

type ObjectRelationSelect b = ObjectRelationSelectG b Void (SQLExpression b)

data ArraySelectG (b :: BackendType) (r :: Type) v

Constructors

ASSimple (ArrayRelationSelectG b r v)

ASAggregate (ArrayAggregateSelectG b r v)

ASConnection (ArrayConnectionSelect b r v)

type ArraySelect b = ArraySelectG b Void (SQLExpression b)

type ArraySelectFieldsG b r v = Fields (ArraySelectG b r v)

data SourceRelationshipSelection (b :: BackendType) (r :: Type) (vf :: BackendType -> Type)

Captures the selection set of a remote source relationship.

Constructors

SourceRelationshipObject (AnnObjectSelectG b r (vf b))

SourceRelationshipArray (AnnSimpleSelectG b r (vf b))

SourceRelationshipArrayAggregate (AnnAggregateSelectG b r (vf b))

data RemoteSourceSelect (r :: Type) (vf :: BackendType -> Type) (tgt :: BackendType)

A relationship to a remote source. vf (could use a better name) is analogous to v in other IR types such as AnnFieldG. vf's kind is (BackendType -> Type) instead of v's Type so that v of AnnFieldG can be specific to the backend that it captures (b of an AnnFieldG changes as we walk down the IR branches which capture relationships to other databases)

Constructors

RemoteSourceSelect

Fields

_rssName :: SourceName

_rssConfig :: SourceConfig tgt

_rssSelection :: SourceRelationshipSelection tgt r vf

_rssJoinMapping :: HashMap FieldName (ScalarType tgt, Column tgt)

Additional information about the source's join columns: (ScalarType tgt) so that the remote can interpret the join values coming from src (Column tgt) so that an appropriate join condition / IN clause can be built by the remote

data TablePermG (b :: BackendType) v

Constructors

TablePerm

Fields

_tpFilter :: AnnBoolExp b v

_tpLimit :: Maybe Int

type TablePerm b = TablePermG b (SQLExpression b)

noTablePermissions :: TablePermG backend v

insertFunctionArg :: FunctionArgName -> Int -> a -> FunctionArgsExpG a -> FunctionArgsExpG a

If argument positional index is less than or equal to length of positional arguments then insert the value in positional arguments else insert the value with argument name in named arguments

data CountDistinct

The "distinct" input field inside "count" aggregate field

count ( distinct: Boolean ): Int

Constructors

SelectCountDistinct

SelectCountNonDistinct

asnStrfyNum :: forall b f v. Lens' (AnnSelectG b f v) StringifyNumbers

asnPerm :: forall b f v. Lens' (AnnSelectG b f v) (TablePermG b v)

asnFrom :: forall b f v. Lens' (AnnSelectG b f v) (SelectFromG b v)

asnFields :: forall b f v f. Lens (AnnSelectG b f v) (AnnSelectG b f v) (Fields (f v)) (Fields (f v))

asnArgs :: forall b f v. Lens' (AnnSelectG b f v) (SelectArgsG b v)

aosTableFrom :: forall b r v. Lens' (AnnObjectSelectG b r v) (TableName b)

aosTableFilter :: forall b r v. Lens' (AnnObjectSelectG b r v) (AnnBoolExp b v)

aosFields :: forall b r v r. Lens (AnnObjectSelectG b r v) (AnnObjectSelectG b r v) (AnnFieldsG b r v) (AnnFieldsG b r v)

aarRelationshipName :: forall b a. Lens' (AnnRelationSelectG b a) RelName

aarColumnMapping :: forall b a b. Lens (AnnRelationSelectG b a) (AnnRelationSelectG b a) (HashMap (Column b) (Column b)) (HashMap (Column b) (Column b))

aarAnnSelect :: forall b a a. Lens (AnnRelationSelectG b a) (AnnRelationSelectG b a) a a

csXRelay :: forall b r v. Lens' (ConnectionSelect b r v) (XRelay b)

csSplit :: forall b r v. Lens' (ConnectionSelect b r v) (Maybe (NonEmpty (ConnectionSplit b v)))

csSlice :: forall b r v. Lens' (ConnectionSelect b r v) (Maybe ConnectionSlice)

csSelect :: forall b r v r. Lens (ConnectionSelect b r v) (ConnectionSelect b r v) (AnnSelectG b (ConnectionField b r) v) (AnnSelectG b (ConnectionField b r) v)

csPrimaryKeyColumns :: forall b r v. Lens' (ConnectionSelect b r v) (PrimaryKeyColumns b)

saWhere :: forall b v. Lens' (SelectArgsG b v) (Maybe (AnnBoolExp b v))

saOrderBy :: forall b v. Lens' (SelectArgsG b v) (Maybe (NonEmpty (AnnotatedOrderByItemG b v)))

saOffset :: forall b v. Lens' (SelectArgsG b v) (Maybe Int64)

saLimit :: forall b v. Lens' (SelectArgsG b v) (Maybe Int)

saDistinct :: forall b v. Lens' (SelectArgsG b v) (Maybe (NonEmpty (Column b)))

_AFExpression :: forall b r v. Prism' (AnnFieldG b r v) Text

_AFNodeId :: forall b r v. Prism' (AnnFieldG b r v) (XRelay b, SourceName, TableName b, PrimaryKeyColumns b)

_AFRemote :: forall b r v. Prism' (AnnFieldG b r v) (RemoteRelationshipSelect b r)

_AFComputedField :: forall b r v. Prism' (AnnFieldG b r v) (XComputedField b, ComputedFieldName, ComputedFieldSelect b r v)

_AFArrayRelation :: forall b r v. Prism' (AnnFieldG b r v) (ArraySelectG b r v)

_AFObjectRelation :: forall b r v. Prism' (AnnFieldG b r v) (ObjectRelationSelectG b r v)

_AFColumn :: forall b r v. Prism' (AnnFieldG b r v) (AnnColumnField b v)

_AOCComputedField :: forall b v. Prism' (AnnotatedOrderByElement b v) (ComputedFieldOrderBy b v)

_AOCArrayAggregation :: forall b v. Prism' (AnnotatedOrderByElement b v) (RelInfo b, AnnBoolExp b v, AnnotatedAggregateOrderBy b)

_AOCObjectRelation :: forall b v. Prism' (AnnotatedOrderByElement b v) (RelInfo b, AnnBoolExp b v, AnnotatedOrderByElement b v)

_AOCColumn :: forall b v. Prism' (AnnotatedOrderByElement b v) (ColumnInfo b)

_TAFExp :: forall b r v. Prism' (TableAggregateFieldG b r v) Text

_TAFNodes :: forall b r v r v. Prism (TableAggregateFieldG b r v) (TableAggregateFieldG b r v) (XNodesAgg b, AnnFieldsG b r v) (XNodesAgg b, AnnFieldsG b r v)

_TAFAgg :: forall b r v. Prism' (TableAggregateFieldG b r v) (AggregateFields b)

_ConnectionEdges :: forall b r v b r v. Prism (ConnectionField b r v) (ConnectionField b r v) (EdgeFields b r v) (EdgeFields b r v)

_ConnectionPageInfo :: forall b r v. Prism' (ConnectionField b r v) PageInfoFields

_ConnectionTypename :: forall b r v. Prism' (ConnectionField b r v) Text

_EdgeNode :: forall b r v b r v. Prism (EdgeField b r v) (EdgeField b r v) (AnnFieldsG b r v) (AnnFieldsG b r v)

_EdgeCursor :: forall b r v. Prism' (EdgeField b r v) ()

_EdgeTypename :: forall b r v. Prism' (EdgeField b r v) Text