Hasura.RQL.IR.Select.Args

Description

Removed from Select to speed up compilation

Documentation

type SelectArgs (b :: BackendType) = SelectArgsG b (SQLExpression b)

data SelectArgsG (b :: BackendType) v

Constructors

SelectArgs

Fields

_saWhere :: Maybe (AnnBoolExp b v)

_saOrderBy :: Maybe (NonEmpty (AnnotatedOrderByItemG b v))

_saLimit :: Maybe Int

_saOffset :: Maybe Int64

_saDistinct :: Maybe (NonEmpty (AnnDistinctColumn b v))

data AnnDistinctColumn (b :: BackendType) v

Constructors

AnnDistinctColumn

Fields

_adcColumn :: Column b

_adcRedactionExpression :: AnnRedactionExp b v

This type is used to determine whether the column should be redacted before being distincted-upon

data SelectStreamArgsG (b :: BackendType) v

Constructors

SelectStreamArgsG

Fields

_ssaWhere :: Maybe (AnnBoolExp b v)

optional filter to filter the stream results

_ssaBatchSize :: Int

maximum number of rows to be returned in a single fetch

_ssaCursorArg :: StreamCursorItem b v

info related to the cursor column, a single item data type currently because only single column cursors are supported

type SelectStreamArgs (b :: BackendType) = SelectStreamArgsG b (SQLExpression b)

data StreamCursorItem (b :: BackendType) v

Cursor for streaming subscription

Constructors

StreamCursorItem

Fields

_sciOrdering :: CursorOrdering

Specifies how the cursor item should be ordered

_sciColInfo :: ColumnInfo b

Column info of the cursor item

_sciRedactionExpression :: AnnRedactionExp b v

This type is used to determine whether the column should be redacted before it is sorted upon

_sciInitialValue :: ColumnValue b

Initial value of the cursor item from where the streaming should start

noSelectArgs :: forall (backend :: BackendType) v. SelectArgsG backend v