Hasura.RQL.IR.Value

Documentation

data Provenance

Where did this variable come from?

Constructors

FromGraphQL VariableInfo

FromInternal Text
An internal source

FreshVar
A unique, fresh occurrence of a variable. E.g. a native query argument, or generated values that benefit from being prepared rather than inlined.

data UnpreparedValue (b :: BackendType)

Constructors

UVParameter Provenance (ColumnValue b)
A SQL value that can be parameterized over.

UVLiteral (SQLExpression b)
A literal SQL expression that cannot be parameterized over.

UVSession
The entire session variables JSON object.

UVSessionVar (SessionVarType b) SessionVariable
A single session variable.

data ValueWithOrigin a

This indicates whether a variable value came from a certain GraphQL variable

Constructors

ValueWithOrigin VariableInfo a

ValueNoOrigin a

openValueOrigin :: ValueWithOrigin a -> a

mkParameter :: ValueWithOrigin (ColumnValue b) -> UnpreparedValue b