Hasura.RQL.Types

Documentation

askSourceInfo :: forall b m. (CacheRM m, MonadError QErr m, Backend b, MetadataM m) => SourceName -> m (SourceInfo b)

askSourceConfig :: forall b m. (CacheRM m, MonadError QErr m, Backend b, MetadataM m) => SourceName -> m (SourceConfig b)

askSourceTables :: forall b m. (Backend b, CacheRM m) => SourceName -> m (TableCache b)

askTabInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (TableInfo b)

askFunInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> FunctionName b -> m (FunctionInfo b)

Tries to extract the function information from the metadata.

If the function does not exist, will throw a 'not-exists' error.

askTabInfoSource :: forall b m. (QErrM m, TableInfoRM b m, Backend b) => TableName b -> m (TableInfo b)

askTableMetadata :: forall b m. (QErrM m, MetadataM m, Backend b, BackendMetadata b) => SourceName -> TableName b -> m (TableMetadata b)

Retrieve TableMetadata from the stateful MetadataM environment that is associated with the given SourceName and TableName for a particular Backend.

class Monad m => HasServerConfigCtx m where

Methods

askServerConfigCtx :: m ServerConfigCtx

class Monad m => HasSystemDefined m where

Methods

askSystemDefined :: m SystemDefined

newtype HasSystemDefinedT m a

Constructors

HasSystemDefinedT

Fields

unHasSystemDefinedT :: ReaderT SystemDefined m a

runHasSystemDefinedT :: SystemDefined -> HasSystemDefinedT m a -> m a

throwTableDoesNotExist :: forall b m a. (QErrM m, Backend b) => TableName b -> m a

findTable :: forall b m a. (QErrM m, Backend b) => TableName b -> HashMap (TableName b) a -> m a

askTableCache :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> m (TableCache b)

askTableCoreInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (TableCoreInfo b)

askTableCoreInfoSource :: forall b m. (QErrM m, Backend b, TableCoreInfoRM b m) => TableName b -> m (TableCoreInfo b)

Asking for a table core info without explicit SourceName argument. The source name is implicitly inferred from SourceM via @TableCoreInfoRM. This is useful in RQL DML queries which are executed in a particular source database.

askFieldInfoMap :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (FieldInfoMap (FieldInfo b))

askFieldInfoMapSource :: (QErrM m, Backend b, TableCoreInfoRM b m) => TableName b -> m (FieldInfoMap (FieldInfo b))

Asking for a table's fields info without explicit SourceName argument. The source name is implicitly inferred from SourceM via @TableCoreInfoRM. This is useful in RQL DML queries which are executed in a particular source database.

askComputedFieldInfo :: MonadError QErr m => FieldInfoMap (FieldInfo backend) -> ComputedFieldName -> m (ComputedFieldInfo backend)

assertColumnExists :: forall backend m. (MonadError QErr m, Backend backend) => FieldInfoMap (FieldInfo backend) -> Text -> Column backend -> m ()

askRelType :: MonadError QErr m => FieldInfoMap (FieldInfo backend) -> RelName -> Text -> m (RelInfo backend)

askRemoteRel :: MonadError QErr m => FieldInfoMap (FieldInfo backend) -> RelName -> m (RemoteFieldInfo (DBJoinField backend))