Hasura.RQL.Types.Backend

Documentation

class (HasSourceConfiguration b, Representable (BasicOrderType b), Representable (Column b), Representable (ColumnPath b), Representable (ComputedFieldDefinition b), Representable (ComputedFieldImplicitArguments b), Representable (ComputedFieldReturn b), Representable (ConstraintName b), Representable (ExtraTableMetadata b), Representable (FunctionArgument b), Representable (FunctionName b), Representable (FunctionReturnType b), Representable (HealthCheckTest b), Representable (NullsOrderType b), Representable (SQLExpression b), Representable (ScalarSelectionArguments b), Representable (ScalarType b), Representable (XComputedField b), Representable (XGroupBy b), Representable (TableName b), Eq (RawFunctionInfo b), Show (RawFunctionInfo b), Representable (ResolvedConnectionTemplate b), Ord (TableName b), Ord (FunctionName b), Ord (ScalarType b), Ord (Column b), Ord (ColumnPath b), Ord (ComputedFieldReturn b), Ord (ComputedFieldImplicitArguments b), Ord (ConstraintName b), Ord (FunctionArgument b), Ord (XComputedField b), Data (TableName b), From (Column b) (ColumnPath b), FromJSON (BackendConfig b), FromJSON (Column b), FromJSON (ColumnPath b), FromJSON (ColumnPath b), FromJSON (ComputedFieldDefinition b), FromJSON (ConnectionTemplateRequestContext b), FromJSON (ConstraintName b), FromJSON (ExtraTableMetadata b), FromJSON (FunctionName b), FromJSON (FunctionReturnType b), FromJSON (HealthCheckTest b), FromJSON (RawFunctionInfo b), FromJSON (ScalarType b), FromJSON (TableName b), FromJSONKey (Column b), FromJSONKey (ColumnPath b), FromJSONKey (ConstraintName b), HasCodec (BackendConfig b), HasCodec (BackendSourceKind b), HasCodec (Column b), HasCodec (ColumnPath b), HasCodec (ComputedFieldDefinition b), HasCodec (FunctionName b), HasCodec (FunctionReturnType b), HasCodec (ScalarType b), HasCodec (TableName b), Hashable (Column b), Hashable (ColumnPath b), ToJSON (BackendConfig b), ToJSON (Column b), ToJSON (ColumnPath b), ToJSON (ConstraintName b), ToJSON (ExecutionStatistics b), ToJSON (FunctionArgument b), ToJSON (FunctionName b), ToJSON (FunctionReturnType b), ToJSON (RawFunctionInfo b), ToJSON (ScalarType b), ToJSON (TableName b), ToJSON (ExtraTableMetadata b), ToJSON (SQLExpression b), ToJSON (ComputedFieldDefinition b), ToJSON (ComputedFieldImplicitArguments b), ToJSON (ComputedFieldReturn b), ToJSON (HealthCheckTest b), ToJSON (ResolvedConnectionTemplate b), ToJSONKey (Column b), ToJSONKey (ColumnPath b), ToJSONKey (ConstraintName b), ToJSONKey (ScalarType b), ToTxt (Column b), ToTxt (FunctionName b), ToTxt (ScalarType b), ToTxt (TableName b), ToTxt (ConstraintName b), ToErrorValue (Column b), ToErrorValue (TableName b), Typeable (Column b), Typeable (ColumnPath b), Typeable b, HasTag b, Traversable (CountType b), Traversable (FunctionArgumentExp b), Eq (BackendConfig b), Show (BackendConfig b), Eq (BackendInfo b), Show (BackendInfo b), Monoid (BackendInfo b), Eq (ScalarValue b), Show (ScalarValue b), Eq (XNodesAgg b), Show (XNodesAgg b), Eq (XRelay b), Show (XRelay b), Eq (XStreamingSubscription b), Show (XStreamingSubscription b), Eq (XNestedObjects b), Ord (XNestedObjects b), Show (XNestedObjects b), NFData (XNestedObjects b), Hashable (XNestedObjects b), ToJSON (XNestedObjects b), FromJSON (XNestedObjects b), ToTxt (XNestedObjects b), Traversable (BooleanOperators b), Traversable (UpdateVariant b), Traversable (BackendInsert b), Traversable (AggregationPredicates b), LiftedConstraint Eq (FunctionArgumentExp b), LiftedConstraint Eq (AggregationPredicates b), LiftedConstraint Eq (BooleanOperators b), LiftedConstraint Eq (CountType b), LiftedConstraint Show (CountType b), LiftedConstraint Show (BooleanOperators b), LiftedConstraint NFData (BooleanOperators b), LiftedConstraint Hashable (BooleanOperators b), ComposeConstraint ToJSONKeyValue ToJSON (BooleanOperators b), LiftedConstraint Show (FunctionArgumentExp b), LiftedConstraint NFData (FunctionArgumentExp b), LiftedConstraint Hashable (FunctionArgumentExp b), LiftedConstraint Show (AggregationPredicates b), LiftedConstraint NFData (AggregationPredicates b), LiftedConstraint Hashable (AggregationPredicates b), ComposeConstraint ToJSONKeyValue ToJSON (AggregationPredicates b), LiftedConstraint Eq (UpdateVariant b), LiftedConstraint Show (UpdateVariant b)) => Backend (b :: BackendType) where

Mapping from abstract types to concrete backend representation

The RQL IR, used as the output of GraphQL parsers and of the RQL parsers, is backend-agnostic: it uses an abstract representation of the structure of a query, and delegates to the backends the task of choosing an appropriate concrete representation.

Additionally, grouping all those types under one typeclass rather than having dedicated type families allows to explicitly list all typeclass requirements, which simplifies the instance declarations of all IR types.

There are no injectivity requirements on those type families: it's okay for two different backends to use the same types. That means, however, that functions cannot identify to what backend b a given TableName b refers to; most generic functions will need either a type application or a Proxy parameter to disambiguate between different backends at the call site.

Minimal complete definition

isComparableType, isNumType, textToScalarValue, parseScalarValue, scalarValueToJSON, functionToTable, tableToFunction, computedFieldFunction, computedFieldReturnType, fromComputedFieldImplicitArguments, tableGraphQLName, functionGraphQLName, getTableIdentifier, snakeCaseTableName, namingConventionSupport, resizeSourcePools, defaultTriggerOnReplication, getColVals, getColumnPathColumn, tryColumnPathToColumn

Associated Types

type BackendConfig (b :: BackendType)

Backend configuration stored in metadata

type BackendInfo (b :: BackendType)

Runtime backend info derived from (possibly enriched) BackendConfig and stored in SchemaCache

type TableName (b :: BackendType)

type FunctionName (b :: BackendType)

type FunctionReturnType (b :: BackendType)

type FunctionReturnType (b :: BackendType) = XDisable

type RawFunctionInfo (b :: BackendType)

type ConstraintName (b :: BackendType)

type BasicOrderType (b :: BackendType)

type NullsOrderType (b :: BackendType)

type CountType (b :: BackendType) :: Type -> Type

The type that captures how count aggregations are modelled

It is parameterised over the type of fields, which changes during the IR translation phases.

type Column (b :: BackendType)

type ColumnPath (b :: BackendType)

type ScalarValue (b :: BackendType)

type ScalarType (b :: BackendType)

type SQLExpression (b :: BackendType)

type ComputedFieldDefinition (b :: BackendType)

type ScalarSelectionArguments (b :: BackendType)

Arguments of a scalar field's selection { query { some_table { # a scalar field column(ScalarSelectionArguments) } } }

type ExtraTableMetadata (b :: BackendType)

type FunctionArgument (b :: BackendType)

FunctionArgument

type FunctionArgumentExp (b :: BackendType) :: Type -> Type

Function input argument expression

It is parameterised over the type of fields, which changes during the IR translation phases.

type ComputedFieldImplicitArguments (b :: BackendType)

Computed field function argument values which are being implicitly inferred from table and/or session information

type ComputedFieldReturn (b :: BackendType)

Computed field return information

type HealthCheckTest (b :: BackendType)

A config type for health check tests

type BooleanOperators (b :: BackendType) :: Type -> Type

Intermediate Representation of extensions to the shared set of boolean operators on table fields.

It is parameterised over the type of fields, which changes during the IR translation phases.

type AggregationPredicates (b :: BackendType) :: Type -> Type

Intermediate Representation of aggregation predicates. The default implementation makes aggregation predicates uninstantiable.

It is parameterised over the type of fields, which changes during the IR translation phases.

type AggregationPredicates (b :: BackendType) = Const Void :: Type -> Type

type UpdateVariant (b :: BackendType) :: Type -> Type

The different variants of update supported by a backend for their intermediate representation. For example, a backend could use a sum type encapsulating either a single batch update or multiple batch updates.

The default implementation makes update expressions uninstantiable.

It is parameterised over the type of fields, which changes during the IR translation phases.

type UpdateVariant (b :: BackendType) = Const Void :: Type -> Type

type BackendInsert (b :: BackendType) :: Type -> Type

Intermediate Representation of Insert Mutations. The default implementation makes insert expressions uninstantiable.

It is parameterised over the type of fields, which changes during the IR translation phases.

type BackendInsert (b :: BackendType) = Const Void :: Type -> Type

type XComputedField (b :: BackendType)

type XRelay (b :: BackendType)

type XNodesAgg (b :: BackendType)

type XEventTriggers (b :: BackendType)

Flag the availability of event triggers.

type XNestedInserts (b :: BackendType)

Extension to flag the availability of object and array relationships in inserts (aka nested inserts).

type XStreamingSubscription (b :: BackendType)

type XNestedObjects (b :: BackendType)

type XNestedObjects (b :: BackendType) = XDisable

type XGroupBy (b :: BackendType)

type XGroupBy (b :: BackendType) = XDisable

type ResolvedConnectionTemplate (b :: BackendType)

type ResolvedConnectionTemplate (b :: BackendType) = ()

type ConnectionTemplateRequestContext (b :: BackendType)

type ConnectionTemplateRequestContext (b :: BackendType) = ()

type ExecutionStatistics (b :: BackendType)

Information about the query execution that may be useful for debugging or reporting.

type ExecutionStatistics (b :: BackendType) = ()

Methods

healthCheckImplementation :: Maybe (HealthCheckImplementation (HealthCheckTest b))

A backend type can opt into supporting health checks by providing an implementation that includes a default health check test, and a health check test codec.

versionCheckImplementation :: Environment -> SourceName -> SourceConnConfiguration b -> IO (Either QErr ())

An Implementation for version checking when adding a source.

runPingSource :: Environment -> (String -> IO ()) -> SourceName -> SourceConnConfiguration b -> IO ()

A backend type can opt into providing an implementation for fingerprinted pings to the source, useful for attribution that the user is using Hasura

resolveConnectionTemplate :: SourceConfig b -> ConnectionTemplateRequestContext b -> Maybe ConnectionTemplate -> Either QErr EncJSON

isComparableType :: ScalarType b -> Bool

isNumType :: ScalarType b -> Bool

getCustomAggregateOperators :: SourceConfig b -> HashMap Name (HashMap (ScalarType b) (ScalarType b))

Custom aggregate operators supported by the backend. Backends that support custom aggregate operators should return a HashMap from operator name to a scalar type mapping. In the scalar type mapping the key represents the input type for the operator and the value represents the result type. Backends that do not support custom aggregate operators can use the default implementation which returns an empty map.

textToScalarValue :: Maybe Text -> ScalarValue b

parseScalarValue :: ScalarTypeParsingContext b -> ScalarType b -> Value -> Either QErr (ScalarValue b)

scalarValueToJSON :: ScalarValue b -> Value

functionToTable :: FunctionName b -> TableName b

tableToFunction :: TableName b -> FunctionName b

computedFieldFunction :: ComputedFieldDefinition b -> FunctionName b

computedFieldReturnType :: ComputedFieldReturn b -> ComputedFieldReturnType b

supportsAggregateComputedFields :: Bool

Backends that don't support aggregate computed fields will never encounter an SelectionField. However, backends are expected to provide a total transformation from SelectionField to the backend's query language.

Rather than implement error handling for every backend that doesn't support aggregate computed fields, and then remove that error handling for each backend when we add support - honestly, adding error handling would probably take longer than adding aggregate computed field support - we instead have a flag.

If a backend declares this flag as False, computed fields will not be added to the GraphQL schema. This means that backends can safely handle SFComputedField with a runtime exception as long as this flag is False.

Once all backends support all aggregate computed field operations, this flag can be deleted.

fromComputedFieldImplicitArguments :: v -> ComputedFieldImplicitArguments b -> [FunctionArgumentExp b v]

Build function arguments expression from computed field implicit arguments

tableGraphQLName :: TableName b -> Either QErr Name

functionGraphQLName :: FunctionName b -> Either QErr Name

getTableIdentifier :: TableName b -> Either QErr GQLNameIdentifier

snakeCaseTableName :: TableName b -> Text

namingConventionSupport :: SupportedNamingCase

resizeSourcePools :: SourceConfig b -> ServerReplicas -> IO SourceResizePoolSummary

defaultTriggerOnReplication :: Maybe (XEventTriggers b, TriggerOnReplication)

Default behaviour of SQL triggers on logically replicated database. Setting this to Nothing will disable event trigger configuration in the metadata.

getColVals :: (MonadIO m, MonadError QErr m) => SessionVariables -> SourceName -> SourceConfig b -> TableName b -> (ScalarType b, Column b) -> (Column b, [RemoteRelSupportedOp RemoteRelSessionVariableORLiteralValue]) -> m [Text]

Get values from a column in a table with some filters. This function is used in evaluating remote relationship predicate in permissions

TODO (paritosh): This function should return a JSON array of column values. We shouldn't have to parse the column values as Text. The database's JSON serialize/deserialize can take care of correct casting of values (GS-642).

getColumnPathColumn :: ColumnPath b -> Column b

Get the top-level column from a ColumnPath For backends that don't support nested objects (i.e. where ColumnPath b = Column b) this will be id.

tryColumnPathToColumn :: ColumnPath b -> Maybe (Column b)

Convert a singleton ColumnPath to a Column Should return Nothing for paths to nested fields

backendSupportsNestedObjects :: Either QErr (XNestedObjects b)

default backendSupportsNestedObjects :: XNestedObjects b ~ XDisable => Either QErr (XNestedObjects b)

sourceSupportsSchemalessTables :: SourceConfig b -> Bool

getAggregationPredicatesModels :: MonadState [ModelNameInfo] m => SourceName -> ModelSourceType -> AggregationPredicates b a -> m ()

type SessionVarType (b :: BackendType) = CollectableType (ScalarType b)

type XDisable = Void

type XEnable = ()

data ComputedFieldReturnType (b :: BackendType)

Constructors

ReturnsScalar (ScalarType b)

ReturnsTable (TableName b)

ReturnsOthers

_ReturnsTable :: forall (b :: BackendType) p f. (Choice p, Applicative f) => p (TableName b) (f (TableName b)) -> p (ComputedFieldReturnType b) (f (ComputedFieldReturnType b))

data SupportedNamingCase

Used for keeping track of the extent of support of naming convention across different backends.

AllConventions implies a full support whereas OnlyHasuraCase implies a partial support of only HasuraCase

Constructors

OnlyHasuraCase

AllConventions

class (Representable (SourceConnConfiguration b), HasCodec (SourceConnConfiguration b), FromJSON (SourceConnConfiguration b), Show (SourceConfig b), ToJSON (SourceConfig b), ToJSON (SourceConnConfiguration b), Eq (SourceConfig b), Has (ScalarTypeParsingContext b) (SourceConfig b), Eq (ScalarTypeParsingContext b), Ord (ScalarTypeParsingContext b), HasTag b) => HasSourceConfiguration (b :: BackendType) where

Associated Types

type ScalarTypeParsingContext (b :: BackendType)

type ScalarTypeParsingContext (b :: BackendType) = ()

type SourceConnConfiguration (b :: BackendType)

User facing connection configuration for a database.

type SourceConfig (b :: BackendType)

Internal connection configuration for a database - connection string, connection pool etc

Methods

sourceConfigNumReadReplicas :: SourceConfig b -> Int

The number of read replicas specified in the source configuration

sourceConfigConnectonTemplate :: SourceConfig b -> Maybe ValueExt

The connection template specified in the source configuration, if any

sourceSupportsColumnRedaction :: SourceConfig b -> Bool

Whether or not the source supports performing column redaction. See note [SQL generation for inherited roles] for more information

sourceConfigBackendSourceKind :: SourceConfig b -> BackendSourceKind b

type Representable a = (Show a, Eq a, Hashable a, NFData a)