Hasura.RQL.Types.BackendType

Documentation

data PostgresKind

Argument to Postgres; we represent backends which are variations on Postgres as sub-types of Postgres. This value indicates which "flavour" of Postgres a backend is.

Constructors

Vanilla

Citus

Cockroach

data BackendType

An enum that represents each backend we support.

Constructors

Postgres PostgresKind

MSSQL

BigQuery

DataConnector

data BackendSourceKind (b :: BackendType) where

Similar to BackendType, however, in the case of DataConnectorKind we need to be able capture the name of the data connector that should be used by the DataConnector backend. This type correlates to the kind property of SourceMetadata, which is usually just postgres, mssql, etc for static backends, but can be a configurable value for DataConnector hence requiring DataConnectorName for DataConnectorKind

This type cannot entirely replace BackendType because BackendType has a fixed number of possible values which can be enumerated over at compile time, but BackendSourceKind does not because DataConnector fundamentally is configured at runtime with DataConnectorName.

Constructors

PostgresVanillaKind :: BackendSourceKind ('Postgres 'Vanilla)

PostgresCitusKind :: BackendSourceKind ('Postgres 'Citus)

PostgresCockroachKind :: BackendSourceKind ('Postgres 'Cockroach)

MSSQLKind :: BackendSourceKind 'MSSQL

BigQueryKind :: BackendSourceKind 'BigQuery

DataConnectorKind :: DataConnectorName -> BackendSourceKind 'DataConnector

backendShortName :: BackendType -> Maybe Text

Some generated APIs use a shortened version of the backend's name rather than its full name. This function returns the "short form" of a backend, if any.

supportedBackends :: [BackendType]

backendTextNames :: BackendType -> [Text]

backendTypeFromText :: Text -> Maybe BackendType

This uses this lookup mechanism to avoid having to duplicate and hardcode the backend string. We accept both the short form and the long form of the backend's name.

parseBackendTypeFromText :: Text -> Parser BackendType

backendTypeFromBackendSourceKind :: forall (b :: BackendType). BackendSourceKind b -> BackendType