Hasura.RQL.Types.Column

Documentation

newtype EnumValue

Constructors

EnumValue

Fields

getEnumValue :: Name

newtype EnumValueInfo

Constructors

EnumValueInfo

Fields

evComment :: Maybe Text

type EnumValues = HashMap EnumValue EnumValueInfo

data EnumReference (b :: BackendType)

Represents a reference to an “enum table,” a single-column Postgres table that is referenced via foreign key.

Constructors

EnumReference

Fields

erTable :: !(TableName b)

erValues :: !EnumValues

erTableCustomName :: !(Maybe Name)

data ColumnType (b :: BackendType)

The type we use for columns, which are currently always “scalars” (though see the note about CollectableType). Unlike ScalarType, which represents a type that a backend knows about, this type characterizes distinctions we make but the backend doesn’t.

Constructors

ColumnScalar !(ScalarType b)
Ordinary Postgres columns.

ColumnEnumReference !(EnumReference b)
Columns that reference enum tables (see Hasura.RQL.Schema.Enum). This is not actually a distinct type from the perspective of Postgres (at the time of this writing, we ensure they always have type text), but we really want to distinguish this case, since we treat it completely differently in the GraphQL schema.

_ColumnEnumReference :: forall b. Prism' (ColumnType b) (EnumReference b)

_ColumnScalar :: forall b. Prism' (ColumnType b) (ScalarType b)

type ValueParser b m v = CollectableType (ColumnType b) -> Value -> m v

A parser to parse a json value with enforcing column type

data ColumnValue (b :: BackendType)

Constructors

ColumnValue

Fields

cvType :: ColumnType b

cvValue :: ScalarValue b

isScalarColumnWhere :: (ScalarType b -> Bool) -> ColumnType b -> Bool

parseScalarValueColumnType :: forall m b. (MonadError QErr m, Backend b) => ColumnType b -> Value -> m (ScalarValue b)

Note: Unconditionally accepts null values and returns PGNull.

parseScalarValuesColumnType :: (MonadError QErr m, Backend b) => ColumnType b -> [Value] -> m [ScalarValue b]

data RawColumnInfo (b :: BackendType)

“Raw” column info, as stored in the catalog (but not in the schema cache). Instead of containing a PGColumnType, it only contains a PGScalarType, which is combined with the pcirReferences field and other table data to eventually resolve the type to a PGColumnType.

Constructors

RawColumnInfo

Fields

rciName :: !(Column b)

rciPosition :: !Int

The “ordinal position” of the column according to Postgres. Numbering starts at 1 and increases. Dropping a column does not cause the columns to be renumbered, so a column can be consistently identified by its position.

rciType :: !(ScalarType b)

rciIsNullable :: !Bool

rciDescription :: !(Maybe Description)

rciMutability :: ColumnMutability

data ColumnMutability

Indicates whether a column may participate in certain mutations.

For example, identity columns may sometimes be insertable but rarely updatable, depending on the backend and how they're declared.

This guides the schema parsers such that they only generate fields for columns where they're valid without having to model the exact circumstances which cause a column to appear or not.

See https://github.com/hasura/graphql-engine/blob/master/rfcs/column-mutability.md.

Constructors

ColumnMutability

Fields

_cmIsInsertable :: Bool

_cmIsUpdatable :: Bool

data ColumnInfo (b :: BackendType)

“Resolved” column info, produced from a RawColumnInfo value that has been combined with other schema information to produce a PGColumnType.

Constructors

ColumnInfo

Fields

ciColumn :: !(Column b)

ciName :: !Name

field name exposed in GraphQL interface

ciPosition :: !Int

ciType :: !(ColumnType b)

ciIsNullable :: !Bool

ciDescription :: !(Maybe Description)

ciMutability :: ColumnMutability

type PrimaryKeyColumns b = NESeq (ColumnInfo b)

onlyNumCols :: forall b. Backend b => [ColumnInfo b] -> [ColumnInfo b]

isNumCol :: forall b. Backend b => ColumnInfo b -> Bool

onlyComparableCols :: forall b. Backend b => [ColumnInfo b] -> [ColumnInfo b]

getColInfos :: Backend b => [Column b] -> [ColumnInfo b] -> [ColumnInfo b]

fromCol :: Backend b => Column b -> FieldName

type ColumnValues b a = HashMap (Column b) a

data ColumnReference (b :: BackendType)

Represents a reference to a source column, possibly casted an arbitrary number of times. Used within parseBoolExpOperations for bookkeeping.

Constructors

ColumnReferenceColumn !(ColumnInfo b)

ColumnReferenceComputedField !ComputedFieldName !(ScalarType b)

ColumnReferenceCast !(ColumnReference b) !(ColumnType b)

columnReferenceType :: ColumnReference backend -> ColumnType backend