Hasura.RQL.Types.ComputedField

Description

Documentation

newtype ComputedFieldName

Constructors

ComputedFieldName

Fields

unComputedFieldName :: NonEmptyText

computedFieldNameToText :: ComputedFieldName -> Text

fromComputedField :: ComputedFieldName -> FieldName

data ComputedFieldDefinition b

Constructors

ComputedFieldDefinition

Fields

_cfdFunction :: !(FunctionName b)

_cfdTableArgument :: !(Maybe FunctionArgName)

_cfdSessionArgument :: !(Maybe FunctionArgName)

data FunctionTableArgument

The function table argument is either the very first argument or the named argument with an index. The index is 0 if the named argument is the first.

Constructors

FTAFirst

FTANamed

Fields

!FunctionArgName

argument name

!Int

argument index

data FunctionSessionArgument

The session argument, which passes Hasura session variables to a SQL function as a JSON object.

Constructors

FunctionSessionArgument

Fields

!FunctionArgName

The argument name

!Int

The ordinal position in the function input parameters

data FunctionTrackedAs (b :: BackendType)

Constructors

FTAComputedField ComputedFieldName SourceName (TableName b)

FTACustomFunction CustomFunctionNames

data CustomFunctionNames

The function name and input arguments name for the "args" field parser.

function_name(args: args_name)

Constructors

CustomFunctionNames

Fields

cfnFunctionName :: Name

cfnArgsName :: Name

data ComputedFieldReturn (b :: BackendType)

Constructors

CFRScalar !(ScalarType b)

CFRSetofTable !(TableName b)

_CFRSetofTable :: forall b. Prism' (ComputedFieldReturn b) (TableName b)

_CFRScalar :: forall b. Prism' (ComputedFieldReturn b) (ScalarType b)

data ComputedFieldFunction (b :: BackendType)

Constructors

ComputedFieldFunction

Fields

_cffName :: !(FunctionName b)

_cffInputArgs :: !(Seq (FunctionArg b))

_cffTableArgument :: !FunctionTableArgument

_cffSessionArgument :: !(Maybe FunctionSessionArgument)

_cffDescription :: !(Maybe PGDescription)

data ComputedFieldInfo (b :: BackendType)

Constructors

ComputedFieldInfo

Fields

_cfiXComputedFieldInfo :: !(XComputedField b)

_cfiName :: !ComputedFieldName

_cfiFunction :: !(ComputedFieldFunction b)

_cfiReturnType :: !(ComputedFieldReturn b)

_cfiDescription :: !(Maybe Text)

cfiXComputedFieldInfo :: forall b. Lens' (ComputedFieldInfo b) (XComputedField b)

cfiReturnType :: forall b. Lens' (ComputedFieldInfo b) (ComputedFieldReturn b)

cfiName :: forall b. Lens' (ComputedFieldInfo b) ComputedFieldName

cfiFunction :: forall b. Lens' (ComputedFieldInfo b) (ComputedFieldFunction b)

cfiDescription :: forall b. Lens' (ComputedFieldInfo b) (Maybe Text)

onlyScalarComputedFields :: [ComputedFieldInfo backend] -> [ComputedFieldInfo backend]