Hasura.RQL.Types.ComputedField

Description

Documentation

data ComputedFieldFunction (b :: BackendType)

Constructors

ComputedFieldFunction

Fields

_cffName :: FunctionName b

_cffInputArgs :: Seq (FunctionArgument b)

_cffComputedFieldImplicitArgs :: ComputedFieldImplicitArguments b

_cffDescription :: Maybe PGDescription

data ComputedFieldInfo (b :: BackendType)

Constructors

ComputedFieldInfo

Fields

_cfiXComputedFieldInfo :: XComputedField b

_cfiName :: ComputedFieldName

_cfiFunction :: ComputedFieldFunction b

_cfiReturnType :: ComputedFieldReturn b

_cfiDescription :: Maybe Text

newtype ComputedFieldName

Constructors

ComputedFieldName

Fields

unComputedFieldName :: NonEmptyText

data CustomFunctionNames

The function name and input arguments name for the "args" field parser.

function_name(args: args_name)

Constructors

CustomFunctionNames

Fields

cfnFunctionName :: Name

cfnArgsName :: Name

data FunctionTrackedAs (b :: BackendType)

Constructors

FTAComputedField ComputedFieldName SourceName (TableName b)

FTACustomFunction CustomFunctionNames

cfiDescription :: forall (b :: BackendType) f. Functor f => (Maybe Text -> f (Maybe Text)) -> ComputedFieldInfo b -> f (ComputedFieldInfo b)

cfiFunction :: forall (b :: BackendType) f. Functor f => (ComputedFieldFunction b -> f (ComputedFieldFunction b)) -> ComputedFieldInfo b -> f (ComputedFieldInfo b)

cfiName :: forall (b :: BackendType) f. Functor f => (ComputedFieldName -> f ComputedFieldName) -> ComputedFieldInfo b -> f (ComputedFieldInfo b)

cfiReturnType :: forall (b :: BackendType) f. Functor f => (ComputedFieldReturn b -> f (ComputedFieldReturn b)) -> ComputedFieldInfo b -> f (ComputedFieldInfo b)

cfiXComputedFieldInfo :: forall (b :: BackendType) f. Functor f => (XComputedField b -> f (XComputedField b)) -> ComputedFieldInfo b -> f (ComputedFieldInfo b)

computedFieldNameToText :: ComputedFieldName -> Text

fromComputedField :: ComputedFieldName -> FieldName

onlyNumComputedFields :: forall (b :: BackendType). Backend b => [ComputedFieldInfo b] -> [ComputedFieldInfo b]

Return all the computed fields in the given list that have numeric types.

isNumComputedField :: forall (b :: BackendType). Backend b => ComputedFieldInfo b -> Bool

Check whether a computed field has a numeric type.

onlyComparableComputedFields :: forall (b :: BackendType). Backend b => [ComputedFieldInfo b] -> [ComputedFieldInfo b]

Return all the computed fields in the given list that have numeric types.

isComparableComputedField :: forall (b :: BackendType). Backend b => ComputedFieldInfo b -> Bool

Check whether a computed field has a numeric type.

removeComputedFieldsReturningExistingTable :: forall (backend :: BackendType). Backend backend => [ComputedFieldInfo backend] -> [ComputedFieldInfo backend]

Filter computed fields not returning rows of existing table