Hasura.RQL.Types.DataConnector

Documentation

data DataConnectorName

Note: Currently you should not use underscores in this name. This should be enforced in instances, and the mkDataConnectorName smart constructor is available to assist.

unDataConnectorName :: DataConnectorName -> Name

mkDataConnectorName :: Name -> Either String DataConnectorName