Hasura.RQL.Types.Endpoint.Trie

Types

type MultiMapPathTrie a k v = Trie (PathComponent a) (MultiMap k v)

Trie from PathComponents to MultiMaps

data MatchResult a k v

Result of matching a path [PathComponent] a and key k in a MultiMapPathTrie.

MatchResult is a lattice where MatchNotFound is the bottom element and MatchAmbiguous is the top element:

MatchAmbiguous / MatchFound v0 as0 MatchFound v1 as1 / MatchMissingKey (ks0 <> ks1) / MatchMissingKey ks0 MatchMissingKey ks1 / MatchNotFound

Constructors

MatchAmbiguous
Multiple results.

MatchFound v [a]
A single unambiguous result. Returns the value found and a list of parameter bindings.

MatchMissingKey (NonEmpty k)
A path was found, but not a key. Returns a list of keys found.

MatchNotFound
Path was not found in the MultiMapPathTrie.

data PathComponent a

A component in a URL path: either a literal or a wildcard parameter

Constructors

PathLiteral a

PathParam

Path matching

matchPath :: (Hashable k, Hashable a) => k -> [a] -> MultiMapPathTrie a k v -> MatchResult a k v

Match a key k and path [a] against a MultiMapPathTrie a k v

ambiguousPaths :: (Hashable a, Hashable k, Ord v) => MultiMapPathTrie a k v -> [([PathComponent a], Set v)]

Detect and return all ambiguous paths in the MultiMapPathTrie A path p is ambiguous if matchPath k p can return MatchAmbiguous for some k.

ambiguousPathsGrouped :: (Hashable a, Hashable k, Ord v, Ord a) => MultiMapPathTrie a k v -> [(Set [PathComponent a], Set v)]

A version of ambiguousPaths that attempts to group all ambiguous paths that have overlapping endpoints