Hasura.RQL.Types.EventTrigger

Documentation

newtype TriggerName

Unique name for event trigger.

Constructors

TriggerName

Fields

unTriggerName :: NonEmptyText

triggerNameToTxt :: TriggerName -> Text

data Ops

Constructors

INSERT

UPDATE

DELETE

MANUAL

data SubscribeColumns (b :: BackendType)

Constructors

SubCStar

SubCArray [Column b]

data SubscribeOpSpec (b :: BackendType)

Constructors

SubscribeOpSpec

Fields

sosColumns :: SubscribeColumns b

Columns of the table that user can subscribe to listen for changes.

sosPayload :: Maybe (SubscribeColumns b)

Columns that the event trigger payload should consists. If set, only those columns will be visible in the payload. By default, the payload consists of all the columns of the table.

defaultNumRetries :: Int

defaultRetryInterval :: Int

defaultTimeoutSeconds :: Int

defaultRetryConf :: RetryConf

data RetryConf

Constructors

RetryConf

Fields

rcNumRetries :: Int

rcIntervalSec :: Int

rcTimeoutSec :: Maybe Int

data EventHeaderInfo

Constructors

EventHeaderInfo

Fields

ehiHeaderConf :: HeaderConf

ehiCachedValue :: Text

data WebhookConf

Constructors

WCValue InputWebhook

WCEnv Text

data WebhookConfInfo

Constructors

WebhookConfInfo

Fields

wciWebhookConf :: WebhookConf

wciCachedValue :: EnvRecord ResolvedWebhook

data TriggerOpsDef (b :: BackendType)

The table operations on which the event trigger will be invoked.

Constructors

TriggerOpsDef

Fields

tdInsert :: Maybe (SubscribeOpSpec b)

tdUpdate :: Maybe (SubscribeOpSpec b)

tdDelete :: Maybe (SubscribeOpSpec b)

tdEnableManual :: Maybe Bool

data EventTriggerCleanupStatus

Constructors

ETCSPaused

ETCSUnpaused

data AutoTriggerLogCleanupConfig

Automatic event trigger log cleanup configuration

Constructors

AutoTriggerLogCleanupConfig

Fields

_atlccSchedule :: CronSchedule

cron schedule for the automatic cleanup

_atlccBatchSize :: Int

maximum number of events to be deleted in a single cleanup action

_atlccRetentionPeriod :: Int

retention period (in hours) for the event trigger logs

_atlccQueryTimeout :: Int

SQL query timeout (in seconds)

_atlccCleanInvocationLogs :: Bool

should we clean the invocation logs as well

_atlccPaused :: EventTriggerCleanupStatus

is the cleanup action paused

data TriggerLogCleanupConfig

Manual event trigger log cleanup configuration

Constructors

TriggerLogCleanupConfig

Fields

tlccEventTriggerName :: TriggerName

name of the event trigger

tlccSourceName :: SourceName

source of the event trigger

tlccBatchSize :: Int

batch size of for the cleanup action

tlccRetentionPeriod :: Int

retention period (in hours) for the event trigger logs

tlccQueryTimeout :: Int

SQL query timeout (in seconds)

tlccCleanInvocationLogs :: Bool

should we clean the invocation logs as well

data EventTriggerQualifier

Constructors

EventTriggerQualifier

Fields

_etqSourceName :: SourceName

_etqEventTriggers :: NonEmpty TriggerName

data TriggerLogCleanupSources

Constructors

TriggerAllSource

TriggerSource (NonEmpty SourceName)

data TriggerLogCleanupToggleConfig

Constructors

TriggerLogCleanupSources TriggerLogCleanupSources

TriggerQualifier (NonEmpty EventTriggerQualifier)

data EventTriggerConf (b :: BackendType)

Constructors

EventTriggerConf

Fields

etcName :: TriggerName

etcDefinition :: TriggerOpsDef b

etcWebhook :: Maybe InputWebhook

etcWebhookFromEnv :: Maybe Text

etcRetryConf :: RetryConf

etcHeaders :: Maybe [HeaderConf]

etcRequestTransform :: Maybe RequestTransform

etcResponseTransform :: Maybe MetadataResponseTransform

etcCleanupConfig :: Maybe AutoTriggerLogCleanupConfig

updateCleanupConfig :: Maybe AutoTriggerLogCleanupConfig -> EventTriggerConf b -> EventTriggerConf b

data RecreateEventTriggers

Constructors

RETRecreate

RETDoNothing

data TriggerMetadata

Constructors

TriggerMetadata

Fields

tmName :: TriggerName

data Event (b :: BackendType)

Change data for a particular row

https://docs.hasura.io/1.0/graphql/manual/event-triggers/payload.html

Constructors

Event

Fields

eId :: EventId

eSource :: SourceName

eTable :: TableName b

eTrigger :: TriggerMetadata

eEvent :: Value

eTries :: Int

eCreatedAt :: UTCTime

data EventWithSource (b :: BackendType)

The event payload processed by processEvent

Constructors

EventWithSource

Fields

_ewsEvent :: Event b

_ewsSourceConfig :: SourceConfig b

_ewsSourceName :: SourceName

_ewsFetchTime :: UTCTime

The UTCTime represents the time when the event was fetched from DB. ^ Used to calculate Event Lock time

data ProcessEventError

Constructors

PESetRetry UTCTime

PESetError

data EventTriggerInfo (b :: BackendType)

Constructors

EventTriggerInfo

Fields

etiName :: TriggerName

etiOpsDef :: TriggerOpsDef b

etiRetryConf :: RetryConf

etiWebhookInfo :: WebhookConfInfo

The HTTP(s) URL which will be called with the event payload on configured operation. Must be a POST handler. This URL can be entered manually or can be picked up from an environment variable (the environment variable needs to be set before using it for this configuration).

etiHeaders :: [EventHeaderInfo]

Custom headers can be added to an event trigger. Each webhook request will have these headers added.

etiRequestTransform :: Maybe RequestTransform

etiResponseTransform :: Maybe MetadataResponseTransform

etiCleanupConfig :: Maybe AutoTriggerLogCleanupConfig

type EventTriggerInfoMap b = HashMap TriggerName (EventTriggerInfo b)

newtype FetchBatchSize

Constructors

FetchBatchSize

Fields

_unFetchBatchSize :: Int

data DeletedEventLogStats

Statistics of deleted event logs and invocation logs

Constructors

DeletedEventLogStats

Fields

deletedEventLogs :: Int

deletedInvocationLogs :: Int