Hasura.RQL.Types.EventTrigger

Documentation

newtype TriggerName

Unique name for event trigger.

Constructors

TriggerName

Fields

unTriggerName :: NonEmptyText

triggerNameToTxt :: TriggerName -> Text

data Ops

Constructors

INSERT

UPDATE

DELETE

MANUAL

data SubscribeColumns (b :: BackendType)

Constructors

SubCStar

SubCArray [Column b]

data SubscribeOpSpec (b :: BackendType)

Constructors

SubscribeOpSpec

Fields

sosColumns :: !(SubscribeColumns b)

Columns of the table that user can subscribe to listen for changes.

sosPayload :: !(Maybe (SubscribeColumns b))

Columns that the event trigger payload should consists. If set, only those columns will be visible in the payload. By default, the payload consists of all the columns of the table.

defaultNumRetries :: Int

defaultRetryInterval :: Int

defaultTimeoutSeconds :: Int

defaultRetryConf :: RetryConf

data RetryConf

Constructors

RetryConf

Fields

rcNumRetries :: !Int

rcIntervalSec :: !Int

rcTimeoutSec :: !(Maybe Int)

data EventHeaderInfo

Constructors

EventHeaderInfo

Fields

ehiHeaderConf :: !HeaderConf

ehiCachedValue :: !Text

data WebhookConf

Constructors

WCValue InputWebhook

WCEnv Text

data WebhookConfInfo

Constructors

WebhookConfInfo

Fields

wciWebhookConf :: !WebhookConf

wciCachedValue :: !ResolvedWebhook

data TriggerOpsDef (b :: BackendType)

The table operations on which the event trigger will be invoked.

Constructors

TriggerOpsDef

Fields

tdInsert :: !(Maybe (SubscribeOpSpec b))

tdUpdate :: !(Maybe (SubscribeOpSpec b))

tdDelete :: !(Maybe (SubscribeOpSpec b))

tdEnableManual :: !(Maybe Bool)

data EventTriggerConf (b :: BackendType)

Constructors

EventTriggerConf

Fields

etcName :: !TriggerName

etcDefinition :: !(TriggerOpsDef b)

etcWebhook :: !(Maybe InputWebhook)

etcWebhookFromEnv :: !(Maybe Text)

etcRetryConf :: !RetryConf

etcHeaders :: !(Maybe [HeaderConf])

etcRequestTransform :: !(Maybe MetadataTransform)

data RecreateEventTriggers

Constructors

RETRecreate

RETDoNothing

data TriggerMetadata

Constructors

TriggerMetadata

Fields

tmName :: TriggerName

data Event (b :: BackendType)

Change data for a particular row

https://docs.hasura.io/1.0/graphql/manual/event-triggers/payload.html

Constructors

Event

Fields

eId :: !EventId

eSource :: !SourceName

eTable :: !(TableName b)

eTrigger :: !TriggerMetadata

eEvent :: !Value

eTries :: !Int

eCreatedAt :: !UTCTime

data EventWithSource (b :: BackendType)

The event payload processed by processEvent

Constructors

EventWithSource

Fields

_ewsEvent :: !(Event b)

_ewsSourceConfig :: !(SourceConfig b)

_ewsSourceName :: !SourceName

_ewsFetchTime :: !UTCTime

The UTCTime represents the time when the event was fetched from DB. ^ Used to calculate Event Lock time

data ProcessEventError

Constructors

PESetRetry !UTCTime

PESetError

data EventTriggerInfo (b :: BackendType)

Constructors

EventTriggerInfo

Fields

etiName :: !TriggerName

etiOpsDef :: !(TriggerOpsDef b)

etiRetryConf :: !RetryConf

etiWebhookInfo :: !WebhookConfInfo

The HTTP(s) URL which will be called with the event payload on configured operation. Must be a POST handler. This URL can be entered manually or can be picked up from an environment variable (the environment variable needs to be set before using it for this configuration).

etiHeaders :: ![EventHeaderInfo]

Custom headers can be added to an event trigger. Each webhook request will have these headers added.

etiRequestTransform :: !(Maybe MetadataTransform)

type EventTriggerInfoMap b = HashMap TriggerName (EventTriggerInfo b)

newtype FetchBatchSize

Constructors

FetchBatchSize

Fields

_unFetchBatchSize :: Int