Hasura.RQL.Types.Metadata

Documentation

parseListAsMap :: (Hashable k, Eq k, ToTxt k) => Text -> (a -> k) -> Parser [a] -> Parser (InsOrdHashMap k a)

Parse a list of objects into a map from a derived key, failing if the list has duplicates.

data MetadataVersion

Versioning the Metadata JSON structure to track backwards incompatible changes. This value is included in the metadata JSON object at top level version key. Always metadata is emitted in the latest version via export metadata API (runExportMetadata handler). Adding a new value constructor to @MetadataVersion type bumps the metadata version.

NOTE: When metadata version is bumped: 1. The Hasura CLI and Console actively use export metadata API to read metadata. Hence, it is necessary to update CLI and Console to read latest metadata. All changes SHOULD be released hand in hand (preferebly in one pull request) 2. There might be other third party services (developed by Hasura users) which use the export metadata API. Apart from changelog, we need to establish the metadata version update by bumping up the minor version of the GraphQL Engine.

Constructors

MVVersion1

MVVersion2

MVVersion3

currentMetadataVersion :: MetadataVersion

data ComputedFieldMetadata b

Constructors

ComputedFieldMetadata

Fields

_cfmName :: !ComputedFieldName

_cfmDefinition :: !(ComputedFieldDefinition b)

_cfmComment :: !Comment

data RemoteSchemaPermissionMetadata

Constructors

RemoteSchemaPermissionMetadata

Fields

_rspmRole :: !RoleName

_rspmDefinition :: !RemoteSchemaPermissionDefinition

_rspmComment :: !(Maybe Text)

rspmRole :: Lens' RemoteSchemaPermissionMetadata RoleName

rspmDefinition :: Lens' RemoteSchemaPermissionMetadata RemoteSchemaPermissionDefinition

rspmComment :: Lens' RemoteSchemaPermissionMetadata (Maybe Text)

type Relationships a = InsOrdHashMap RelName a

type ComputedFields b = InsOrdHashMap ComputedFieldName (ComputedFieldMetadata b)

type RemoteRelationships = InsOrdHashMap RelName RemoteRelationship

type SchemaRemoteRelationships = InsOrdHashMap Name RemoteSchemaTypeRelationships

type Permissions a = InsOrdHashMap RoleName a

type EventTriggers b = InsOrdHashMap TriggerName (EventTriggerConf b)

data RemoteSchemaTypeRelationships

Constructors

RemoteSchemaTypeRelationships

Fields

_rstrsName :: Name

_rstrsRelationships :: RemoteRelationships

data RemoteSchemaMetadata

Constructors

RemoteSchemaMetadata

Fields

_rsmName :: RemoteSchemaName

_rsmDefinition :: RemoteSchemaDef

_rsmComment :: Maybe Text

_rsmPermissions :: [RemoteSchemaPermissionMetadata]

_rsmRemoteRelationships :: SchemaRemoteRelationships

rstrsRelationships :: Lens' RemoteSchemaTypeRelationships RemoteRelationships

rstrsName :: Lens' RemoteSchemaTypeRelationships Name

rsmRemoteRelationships :: Lens' RemoteSchemaMetadata SchemaRemoteRelationships

rsmPermissions :: Lens' RemoteSchemaMetadata [RemoteSchemaPermissionMetadata]

rsmName :: Lens' RemoteSchemaMetadata RemoteSchemaName

rsmDefinition :: Lens' RemoteSchemaMetadata RemoteSchemaDef

rsmComment :: Lens' RemoteSchemaMetadata (Maybe Text)

data TableMetadata b

Constructors

TableMetadata

Fields

_tmTable :: !(TableName b)

_tmIsEnum :: !Bool

_tmConfiguration :: !(TableConfig b)

_tmObjectRelationships :: !(Relationships (ObjRelDef b))

_tmArrayRelationships :: !(Relationships (ArrRelDef b))

_tmComputedFields :: !(ComputedFields b)

_tmRemoteRelationships :: !RemoteRelationships

_tmInsertPermissions :: !(Permissions (InsPermDef b))

_tmSelectPermissions :: !(Permissions (SelPermDef b))

_tmUpdatePermissions :: !(Permissions (UpdPermDef b))

_tmDeletePermissions :: !(Permissions (DelPermDef b))

_tmEventTriggers :: !(EventTriggers b)

tmUpdatePermissions :: forall b. Lens' (TableMetadata b) (Permissions (UpdPermDef b))

tmTable :: forall b. Lens' (TableMetadata b) (TableName b)

tmSelectPermissions :: forall b. Lens' (TableMetadata b) (Permissions (SelPermDef b))

tmRemoteRelationships :: forall b. Lens' (TableMetadata b) RemoteRelationships

tmObjectRelationships :: forall b. Lens' (TableMetadata b) (Relationships (ObjRelDef b))

tmIsEnum :: forall b. Lens' (TableMetadata b) Bool

tmInsertPermissions :: forall b. Lens' (TableMetadata b) (Permissions (InsPermDef b))

tmEventTriggers :: forall b. Lens' (TableMetadata b) (EventTriggers b)

tmDeletePermissions :: forall b. Lens' (TableMetadata b) (Permissions (DelPermDef b))

tmConfiguration :: forall b. Lens' (TableMetadata b) (TableConfig b)

tmComputedFields :: forall b. Lens' (TableMetadata b) (ComputedFields b)

tmArrayRelationships :: forall b. Lens' (TableMetadata b) (Relationships (ArrRelDef b))

mkTableMeta :: TableName b -> Bool -> TableConfig b -> TableMetadata b

data FunctionMetadata b

Constructors

FunctionMetadata

Fields

_fmFunction :: !(FunctionName b)

_fmConfiguration :: !FunctionConfig

_fmPermissions :: ![FunctionPermissionInfo]

_fmComment :: !(Maybe Text)

fmPermissions :: forall b. Lens' (FunctionMetadata b) [FunctionPermissionInfo]

fmFunction :: forall b b. Lens (FunctionMetadata b) (FunctionMetadata b) (FunctionName b) (FunctionName b)

fmConfiguration :: forall b. Lens' (FunctionMetadata b) FunctionConfig

fmComment :: forall b. Lens' (FunctionMetadata b) (Maybe Text)

type Tables b = InsOrdHashMap (TableName b) (TableMetadata b)

type Functions b = InsOrdHashMap (FunctionName b) (FunctionMetadata b)

type RemoteSchemas = InsOrdHashMap RemoteSchemaName RemoteSchemaMetadata

type Endpoints = InsOrdHashMap EndpointName CreateEndpoint

type Actions = InsOrdHashMap ActionName ActionMetadata

type CronTriggers = InsOrdHashMap TriggerName CronTriggerMetadata

type InheritedRoles = InsOrdHashMap RoleName InheritedRole

data SourceMetadata b

Constructors

SourceMetadata

Fields

_smName :: !SourceName

_smKind :: !(BackendSourceKind b)

_smTables :: !(Tables b)

_smFunctions :: !(Functions b)

_smConfiguration :: !(SourceConnConfiguration b)

_smQueryTags :: !(Maybe QueryTagsConfig)

_smCustomization :: !SourceCustomization

smTables :: forall b. Lens' (SourceMetadata b) (Tables b)

smQueryTags :: forall b. Lens' (SourceMetadata b) (Maybe QueryTagsConfig)

smName :: forall b. Lens' (SourceMetadata b) SourceName

smKind :: forall b. Lens' (SourceMetadata b) (BackendSourceKind b)

smFunctions :: forall b. Lens' (SourceMetadata b) (Functions b)

smCustomization :: forall b. Lens' (SourceMetadata b) SourceCustomization

smConfiguration :: forall b. Lens' (SourceMetadata b) (SourceConnConfiguration b)

mkSourceMetadata :: forall (b :: BackendType). Backend b => SourceName -> BackendSourceKind b -> SourceConnConfiguration b -> SourceCustomization -> BackendSourceMetadata

type BackendSourceMetadata = AnyBackend SourceMetadata

toSourceMetadata :: forall b. Backend b => Prism' BackendSourceMetadata (SourceMetadata b)

getSourceName :: BackendSourceMetadata -> SourceName

type Sources = InsOrdHashMap SourceName BackendSourceMetadata

parseNonSourcesMetadata :: Object -> Parser (RemoteSchemas, QueryCollections, MetadataAllowlist, CustomTypes, Actions, CronTriggers, ApiLimit, MetricsConfig, InheritedRoles, SetGraphqlIntrospectionOptions)

newtype BackendConfigWrapper b

This newtype simply wraps the BackendConfig type family so that it can be used with BackendMap in the Metadata type. GHC will not allow the type family to be used directly. :(

Constructors

BackendConfigWrapper

Fields

unBackendConfigWrapper :: BackendConfig b

data Metadata

A complete GraphQL Engine metadata representation to be stored, exported/replaced via metadata queries.

Constructors

Metadata

Fields

_metaSources :: !Sources

_metaRemoteSchemas :: !RemoteSchemas

_metaQueryCollections :: !QueryCollections

_metaAllowlist :: !MetadataAllowlist

_metaCustomTypes :: !CustomTypes

_metaActions :: !Actions

_metaCronTriggers :: !CronTriggers

_metaRestEndpoints :: !Endpoints

_metaApiLimits :: !ApiLimit

_metaMetricsConfig :: !MetricsConfig

_metaInheritedRoles :: !InheritedRoles

_metaSetGraphqlIntrospectionOptions :: !SetGraphqlIntrospectionOptions

_metaNetwork :: !Network

_metaBackendConfigs :: BackendMap BackendConfigWrapper

metaSources :: Lens' Metadata Sources

metaSetGraphqlIntrospectionOptions :: Lens' Metadata SetGraphqlIntrospectionOptions

metaRestEndpoints :: Lens' Metadata Endpoints

metaRemoteSchemas :: Lens' Metadata RemoteSchemas

metaQueryCollections :: Lens' Metadata QueryCollections

metaNetwork :: Lens' Metadata Network

metaMetricsConfig :: Lens' Metadata MetricsConfig

metaInheritedRoles :: Lens' Metadata InheritedRoles

metaCustomTypes :: Lens' Metadata CustomTypes

metaCronTriggers :: Lens' Metadata CronTriggers

metaBackendConfigs :: Lens' Metadata (BackendMap BackendConfigWrapper)

metaApiLimits :: Lens' Metadata ApiLimit

metaAllowlist :: Lens' Metadata MetadataAllowlist

metaActions :: Lens' Metadata Actions

emptyMetadata :: Metadata

tableMetadataSetter :: Backend b => SourceName -> TableName b -> ASetter' Metadata (TableMetadata b)

functionMetadataSetter :: Backend b => SourceName -> FunctionName b -> ASetter' Metadata (FunctionMetadata b)

A lens setter for the metadata of a specific function as identified by the source name and function name.

class Monad m => MetadataM m where

A simple monad class which enables fetching and setting @Metadata in the state.

Methods

getMetadata :: m Metadata

putMetadata :: Metadata -> m ()

data MetadataNoSources

Constructors

MetadataNoSources

Fields

_mnsTables :: !(Tables ('Postgres 'Vanilla))

_mnsFunctions :: !(Functions ('Postgres 'Vanilla))

_mnsRemoteSchemas :: !RemoteSchemas

_mnsQueryCollections :: !QueryCollections

_mnsAllowlist :: !MetadataAllowlist

_mnsCustomTypes :: !CustomTypes

_mnsActions :: !Actions

_mnsCronTriggers :: !CronTriggers

newtype MetadataModifier

Constructors

MetadataModifier

Fields

runMetadataModifier :: Metadata -> Metadata

dropTableInMetadata :: forall b. Backend b => SourceName -> TableName b -> MetadataModifier

dropRelationshipInMetadata :: RelName -> TableMetadata b -> TableMetadata b

dropPermissionInMetadata :: RoleName -> PermType -> TableMetadata b -> TableMetadata b

dropComputedFieldInMetadata :: ComputedFieldName -> TableMetadata b -> TableMetadata b

dropEventTriggerInMetadata :: TriggerName -> TableMetadata b -> TableMetadata b

dropRemoteRelationshipInMetadata :: RelName -> TableMetadata b -> TableMetadata b

dropFunctionInMetadata :: forall b. Backend b => SourceName -> FunctionName b -> MetadataModifier

metadataToOrdJSON :: Metadata -> Value

Encode Metadata to JSON with deterministic ordering (e.g. "version" being at the top). The CLI system stores metadata in files and has option to show changes in git diff style. The diff shouldn't appear different for no metadata changes. So, the ordering of object keys and array elements should remain consistent across versions of graphql-engine if possible.

Note: While modifying any part of the code below, make sure the encoded JSON of each type is parsable via its FromJSON instance.

TODO: we can use 'aeson-pretty' to serialize in a consistent way, and to specify a (partial) order of keys, while getting the benefits of auto-generated To/FromJSON instances. `FromJSON TableMetadata` complicates this though...

See: https://github.com/hasura/graphql-engine/issues/6348

data CatalogStateType

Constructors

CSTCli

CSTConsole

data SetCatalogState

Constructors

SetCatalogState

Fields

_scsType :: !CatalogStateType

_scsState :: !Value

data CatalogState

Constructors

CatalogState

Fields

_csId :: !Text

_csCliState :: !Value

_csConsoleState :: !Value

data GetCatalogState

Constructors

GetCatalogState