Hasura.RQL.Types.Metadata.Common

Description

In order to avoid circular dependencies while splitting Metadata into multiple modules, some definitions must be moved out of that module. This module is the bucket for definitions that have not been specifically moved elsewhere.

Documentation

type EventTriggers b = InsOrdHashMap TriggerName (EventTriggerConf b)

type RemoteSchemaMetadata = RemoteSchemaMetadataG RemoteRelationshipDefinition

type RemoteSchemas = InsOrdHashMap RemoteSchemaName RemoteSchemaMetadata

type Tables b = InsOrdHashMap (TableName b) (TableMetadata b)

type Functions b = InsOrdHashMap (FunctionName b) (FunctionMetadata b)

type NativeQueries b = InsOrdHashMap NativeQueryName (NativeQueryMetadata b)

type StoredProcedures b = InsOrdHashMap (FunctionName b) (StoredProcedureMetadata b)

type LogicalModels b = InsOrdHashMap LogicalModelName (LogicalModelMetadata b)

type Endpoints = InsOrdHashMap EndpointName CreateEndpoint

type Actions = InsOrdHashMap ActionName ActionMetadata

type CronTriggers = InsOrdHashMap TriggerName CronTriggerMetadata

type InheritedRoles = InsOrdHashMap RoleName InheritedRole

data SourceMetadata b

Source configuration for a source of backend type b as stored in the Metadata DB.

Constructors

SourceMetadata

Fields

_smName :: SourceName

_smKind :: BackendSourceKind b

_smTables :: Tables b

_smFunctions :: Functions b

_smNativeQueries :: NativeQueries b

_smStoredProcedures :: StoredProcedures b

_smLogicalModels :: LogicalModels b

_smConfiguration :: SourceConnConfiguration b

_smQueryTags :: Maybe QueryTagsConfig

_smCustomization :: SourceCustomization

_smHealthCheckConfig :: Maybe (HealthCheckConfig b)

https://hasura.io/docs/latest/deployment/health-checks/source-health-check/

smTables :: forall b. Lens' (SourceMetadata b) (Tables b)

smStoredProcedures :: forall b. Lens' (SourceMetadata b) (StoredProcedures b)

smQueryTags :: forall b. Lens' (SourceMetadata b) (Maybe QueryTagsConfig)

smNativeQueries :: forall b. Lens' (SourceMetadata b) (NativeQueries b)

smName :: forall b. Lens' (SourceMetadata b) SourceName

smLogicalModels :: forall b. Lens' (SourceMetadata b) (LogicalModels b)

smKind :: forall b. Lens' (SourceMetadata b) (BackendSourceKind b)

smHealthCheckConfig :: forall b. Lens' (SourceMetadata b) (Maybe (HealthCheckConfig b))

smFunctions :: forall b. Lens' (SourceMetadata b) (Functions b)

smCustomization :: forall b. Lens' (SourceMetadata b) SourceCustomization

smConfiguration :: forall b. Lens' (SourceMetadata b) (SourceConnConfiguration b)

backendSourceMetadataCodec :: JSONCodec BackendSourceMetadata

mkSourceMetadata :: forall (b :: BackendType). Backend b => SourceName -> BackendSourceKind b -> SourceConnConfiguration b -> SourceCustomization -> Maybe (HealthCheckConfig b) -> BackendSourceMetadata

newtype BackendSourceMetadata

Source configuration as stored in the Metadata DB for some existentialized backend.

Constructors

BackendSourceMetadata

Fields

unBackendSourceMetadata :: AnyBackend SourceMetadata

toSourceMetadata :: forall b. Backend b => Prism' BackendSourceMetadata (SourceMetadata b)

getSourceName :: BackendSourceMetadata -> SourceName

type Sources = InsOrdHashMap SourceName BackendSourceMetadata

sourcesCodec :: JSONCodec Sources

parseNonSourcesMetadata :: Object -> Parser (RemoteSchemas, QueryCollections, MetadataAllowlist, CustomTypes, Actions, CronTriggers, ApiLimit, MetricsConfig, InheritedRoles, SetGraphqlIntrospectionOptions)

newtype BackendConfigWrapper b

This newtype simply wraps the BackendConfig type family so that it can be used with BackendMap in the Metadata type. GHC will not allow the type family to be used directly. :(

Constructors

BackendConfigWrapper

Fields

unBackendConfigWrapper :: BackendConfig b

data CatalogStateType

Constructors

CSTCli

CSTConsole

data SetCatalogState

Constructors

SetCatalogState

Fields

_scsType :: CatalogStateType

_scsState :: Value

data CatalogState

Constructors

CatalogState

Fields

_csId :: Text

_csCliState :: Value

_csConsoleState :: Value

data GetCatalogState

Constructors

GetCatalogState