Hasura.RQL.Types.Metadata.Common

Description

In order to avoid circular dependencies while splitting Metadata into multiple modules, some definitions must be moved out of that module. This module is the bucket for definitions that have not been specifically moved elsewhere.

Documentation

parseListAsMap :: (Hashable k, Eq k, ToTxt k) => Text -> (a -> k) -> Parser [a] -> Parser (InsOrdHashMap k a)

Parse a list of objects into a map from a derived key, failing if the list has duplicates.

data ComputedFieldMetadata b

Constructors

ComputedFieldMetadata

Fields

_cfmName :: ComputedFieldName

_cfmDefinition :: ComputedFieldDefinition b

_cfmComment :: Comment

data RemoteSchemaPermissionMetadata

Constructors

RemoteSchemaPermissionMetadata

Fields

_rspmRole :: RoleName

_rspmDefinition :: RemoteSchemaPermissionDefinition

_rspmComment :: Maybe Text

rspmRole :: Lens' RemoteSchemaPermissionMetadata RoleName

rspmDefinition :: Lens' RemoteSchemaPermissionMetadata RemoteSchemaPermissionDefinition

rspmComment :: Lens' RemoteSchemaPermissionMetadata (Maybe Text)

type Relationships a = InsOrdHashMap RelName a

type ComputedFields b = InsOrdHashMap ComputedFieldName (ComputedFieldMetadata b)

type RemoteRelationships = InsOrdHashMap RelName RemoteRelationship

type SchemaRemoteRelationships = InsOrdHashMap Name RemoteSchemaTypeRelationships

type Permissions a = InsOrdHashMap RoleName a

type EventTriggers b = InsOrdHashMap TriggerName (EventTriggerConf b)

data RemoteSchemaTypeRelationships

Constructors

RemoteSchemaTypeRelationships

Fields

_rstrsName :: Name

_rstrsRelationships :: RemoteRelationships

data RemoteSchemaMetadata

Constructors

RemoteSchemaMetadata

Fields

_rsmName :: RemoteSchemaName

_rsmDefinition :: RemoteSchemaDef

_rsmComment :: Maybe Text

_rsmPermissions :: [RemoteSchemaPermissionMetadata]

_rsmRemoteRelationships :: SchemaRemoteRelationships

rstrsRelationships :: Lens' RemoteSchemaTypeRelationships RemoteRelationships

rstrsName :: Lens' RemoteSchemaTypeRelationships Name

rsmRemoteRelationships :: Lens' RemoteSchemaMetadata SchemaRemoteRelationships

rsmPermissions :: Lens' RemoteSchemaMetadata [RemoteSchemaPermissionMetadata]

rsmName :: Lens' RemoteSchemaMetadata RemoteSchemaName

rsmDefinition :: Lens' RemoteSchemaMetadata RemoteSchemaDef

rsmComment :: Lens' RemoteSchemaMetadata (Maybe Text)

data TableMetadata b

Constructors

TableMetadata

Fields

_tmTable :: TableName b

_tmIsEnum :: Bool

_tmConfiguration :: TableConfig b

_tmObjectRelationships :: Relationships (ObjRelDef b)

_tmArrayRelationships :: Relationships (ArrRelDef b)

_tmComputedFields :: ComputedFields b

_tmRemoteRelationships :: RemoteRelationships

_tmInsertPermissions :: Permissions (InsPermDef b)

_tmSelectPermissions :: Permissions (SelPermDef b)

_tmUpdatePermissions :: Permissions (UpdPermDef b)

_tmDeletePermissions :: Permissions (DelPermDef b)

_tmEventTriggers :: EventTriggers b

_tmApolloFederationConfig :: Maybe ApolloFederationConfig

tmUpdatePermissions :: forall b. Lens' (TableMetadata b) (Permissions (UpdPermDef b))

tmTable :: forall b. Lens' (TableMetadata b) (TableName b)

tmSelectPermissions :: forall b. Lens' (TableMetadata b) (Permissions (SelPermDef b))

tmRemoteRelationships :: forall b. Lens' (TableMetadata b) RemoteRelationships

tmObjectRelationships :: forall b. Lens' (TableMetadata b) (Relationships (ObjRelDef b))

tmIsEnum :: forall b. Lens' (TableMetadata b) Bool

tmInsertPermissions :: forall b. Lens' (TableMetadata b) (Permissions (InsPermDef b))

tmEventTriggers :: forall b. Lens' (TableMetadata b) (EventTriggers b)

tmDeletePermissions :: forall b. Lens' (TableMetadata b) (Permissions (DelPermDef b))

tmConfiguration :: forall b. Lens' (TableMetadata b) (TableConfig b)

tmComputedFields :: forall b. Lens' (TableMetadata b) (ComputedFields b)

tmArrayRelationships :: forall b. Lens' (TableMetadata b) (Relationships (ArrRelDef b))

tmApolloFederationConfig :: forall b. Lens' (TableMetadata b) (Maybe ApolloFederationConfig)

mkTableMeta :: TableName b -> Bool -> TableConfig b -> TableMetadata b

data FunctionMetadata b

Constructors

FunctionMetadata

Fields

_fmFunction :: FunctionName b

_fmConfiguration :: FunctionConfig

_fmPermissions :: [FunctionPermissionInfo]

_fmComment :: Maybe Text

fmPermissions :: forall b. Lens' (FunctionMetadata b) [FunctionPermissionInfo]

fmFunction :: forall b b. Lens (FunctionMetadata b) (FunctionMetadata b) (FunctionName b) (FunctionName b)

fmConfiguration :: forall b. Lens' (FunctionMetadata b) FunctionConfig

fmComment :: forall b. Lens' (FunctionMetadata b) (Maybe Text)

type Tables b = InsOrdHashMap (TableName b) (TableMetadata b)

type Functions b = InsOrdHashMap (FunctionName b) (FunctionMetadata b)

type RemoteSchemas = InsOrdHashMap RemoteSchemaName RemoteSchemaMetadata

type Endpoints = InsOrdHashMap EndpointName CreateEndpoint

type Actions = InsOrdHashMap ActionName ActionMetadata

type CronTriggers = InsOrdHashMap TriggerName CronTriggerMetadata

type InheritedRoles = InsOrdHashMap RoleName InheritedRole

data SourceMetadata b

Source configuration for a source of backend type b as stored in the Metadata DB.

Constructors

SourceMetadata

Fields

_smName :: SourceName

_smKind :: BackendSourceKind b

_smTables :: Tables b

_smFunctions :: Functions b

_smConfiguration :: SourceConnConfiguration b

_smQueryTags :: Maybe QueryTagsConfig

_smCustomization :: SourceCustomization

_smHealthCheckConfig :: Maybe (HealthCheckConfig b)

smTables :: forall b. Lens' (SourceMetadata b) (Tables b)

smQueryTags :: forall b. Lens' (SourceMetadata b) (Maybe QueryTagsConfig)

smName :: forall b. Lens' (SourceMetadata b) SourceName

smKind :: forall b. Lens' (SourceMetadata b) (BackendSourceKind b)

smHealthCheckConfig :: forall b. Lens' (SourceMetadata b) (Maybe (HealthCheckConfig b))

smFunctions :: forall b. Lens' (SourceMetadata b) (Functions b)

smCustomization :: forall b. Lens' (SourceMetadata b) SourceCustomization

smConfiguration :: forall b. Lens' (SourceMetadata b) (SourceConnConfiguration b)

backendSourceMetadataCodec :: JSONCodec BackendSourceMetadata

anySourceMetadataCodec :: HasTag b => JSONCodec (SourceMetadata b) -> JSONCodec BackendSourceMetadata

mkSourceMetadata :: forall (b :: BackendType). Backend b => SourceName -> BackendSourceKind b -> SourceConnConfiguration b -> SourceCustomization -> Maybe (HealthCheckConfig b) -> BackendSourceMetadata

newtype BackendSourceMetadata

Source configuration as stored in the Metadata DB for some existentialized backend.

Constructors

BackendSourceMetadata

Fields

unBackendSourceMetadata :: AnyBackend SourceMetadata

toSourceMetadata :: forall b. Backend b => Prism' BackendSourceMetadata (SourceMetadata b)

getSourceName :: BackendSourceMetadata -> SourceName

type Sources = InsOrdHashMap SourceName BackendSourceMetadata

sourcesCodec :: JSONCodec Sources

parseNonSourcesMetadata :: Object -> Parser (RemoteSchemas, QueryCollections, MetadataAllowlist, CustomTypes, Actions, CronTriggers, ApiLimit, MetricsConfig, InheritedRoles, SetGraphqlIntrospectionOptions)

newtype BackendConfigWrapper b

This newtype simply wraps the BackendConfig type family so that it can be used with BackendMap in the Metadata type. GHC will not allow the type family to be used directly. :(

Constructors

BackendConfigWrapper

Fields

unBackendConfigWrapper :: BackendConfig b

data CatalogStateType

Constructors

CSTCli

CSTConsole

data SetCatalogState

Constructors

SetCatalogState

Fields

_scsType :: CatalogStateType

_scsState :: Value

data CatalogState

Constructors

CatalogState

Fields

_csId :: Text

_csCliState :: Value

_csConsoleState :: Value

data GetCatalogState

Constructors

GetCatalogState