Hasura.RQL.Types.Metadata.Object

Documentation

data TableMetadataObjId

Constructors

MTORel !RelName !RelType

MTOComputedField !ComputedFieldName

MTOPerm !RoleName !PermType

MTOTrigger !TriggerName

MTORemoteRelationship !RelName

data SourceMetadataObjId b

Constructors

SMOTable !(TableName b)

SMOFunction !(FunctionName b)

SMOFunctionPermission !(FunctionName b) !RoleName

SMOTableObj !(TableName b) !TableMetadataObjId

data MetadataObjId

Constructors

MOSource !SourceName

MOSourceObjId !SourceName !(AnyBackend SourceMetadataObjId)

MORemoteSchema !RemoteSchemaName
Originates from user-defined _arsqName

MORemoteSchemaPermissions !RemoteSchemaName !RoleName

MORemoteSchemaRemoteRelationship !RemoteSchemaName !Name !RelName
A remote relationship on a remote schema type, identified by 1. remote schema name 2. remote schema type on which the relationship is defined 3. name of the relationship

MOCustomTypes

MOAction !ActionName

MOActionPermission !ActionName !RoleName

MOCronTrigger !TriggerName

MOInheritedRole !RoleName

MOEndpoint !EndpointName

MOHostTlsAllowlist !String

MOQueryCollectionsQuery !CollectionName !ListedQuery

_MOQueryCollectionsQuery :: Prism' MetadataObjId (CollectionName, ListedQuery)

_MOHostTlsAllowlist :: Prism' MetadataObjId String

_MOEndpoint :: Prism' MetadataObjId EndpointName

_MOInheritedRole :: Prism' MetadataObjId RoleName

_MOCronTrigger :: Prism' MetadataObjId TriggerName

_MOActionPermission :: Prism' MetadataObjId (ActionName, RoleName)

_MOAction :: Prism' MetadataObjId ActionName

_MOCustomTypes :: Prism' MetadataObjId ()

_MORemoteSchemaRemoteRelationship :: Prism' MetadataObjId (RemoteSchemaName, Name, RelName)

_MORemoteSchemaPermissions :: Prism' MetadataObjId (RemoteSchemaName, RoleName)

_MORemoteSchema :: Prism' MetadataObjId RemoteSchemaName

_MOSourceObjId :: Prism' MetadataObjId (SourceName, AnyBackend SourceMetadataObjId)

_MOSource :: Prism' MetadataObjId SourceName

moiTypeName :: MetadataObjId -> Text

moiName :: MetadataObjId -> Text

data MetadataObject

Constructors

MetadataObject

Fields

_moId :: !MetadataObjId

_moDefinition :: !Value

moId :: Lens' MetadataObject MetadataObjId

moDefinition :: Lens' MetadataObject Value

data InconsistentRoleEntity

Constructors

InconsistentTablePermission

Fields

!SourceName

!Text

Table name -- using Text here instead of `TableName b` for simplification, Otherwise, we'll have to create a newtype wrapper around `TableName b` and then use it with AnyBackend

!PermType

InconsistentRemoteSchemaPermission !RemoteSchemaName

data InconsistentMetadata

Constructors

InconsistentObject !Text !(Maybe Value) !MetadataObject

ConflictingObjects !Text ![MetadataObject]

DuplicateObjects !MetadataObjId ![Value]

DuplicateRestVariables !Text !MetadataObject

InvalidRestSegments !Text !MetadataObject

AmbiguousRestEndpoints !Text ![MetadataObject]

ConflictingInheritedPermission !RoleName !InconsistentRoleEntity

_ConflictingInheritedPermission :: Prism' InconsistentMetadata (RoleName, InconsistentRoleEntity)

_AmbiguousRestEndpoints :: Prism' InconsistentMetadata (Text, [MetadataObject])

_InvalidRestSegments :: Prism' InconsistentMetadata (Text, MetadataObject)

_DuplicateRestVariables :: Prism' InconsistentMetadata (Text, MetadataObject)

_DuplicateObjects :: Prism' InconsistentMetadata (MetadataObjId, [Value])

_ConflictingObjects :: Prism' InconsistentMetadata (Text, [MetadataObject])

_InconsistentObject :: Prism' InconsistentMetadata (Text, Maybe Value, MetadataObject)

droppableInconsistentMetadata :: InconsistentMetadata -> Bool

Helper function to differentiate which type of inconsistent metadata can be dropped, if an inconsistency cannot be resolved by dropping any part of the metadata then this function should return False, otherwise it should return True

getInconsistentRemoteSchemas :: [InconsistentMetadata] -> [RemoteSchemaName]

imObjectIds :: InconsistentMetadata -> [MetadataObjId]

imReason :: InconsistentMetadata -> Text

groupInconsistentMetadataById :: [InconsistentMetadata] -> HashMap MetadataObjId (NonEmpty InconsistentMetadata)

Builds a map from each unique metadata object id to the inconsistencies associated with it. Note that a single inconsistency can involve multiple metadata objects, so the same inconsistency may appear in the resulting map multiple times!