Hasura.RQL.Types.OpenTelemetry

User-facing configuration (metadata)

data OpenTelemetryConfig

Metadata configuration for all OpenTelemetry-related features

Constructors

OpenTelemetryConfig

Fields

_ocStatus :: OtelStatus

_ocEnabledDataTypes :: Set OtelDataType

_ocExporterOtlp :: OtelExporterConfig

_ocBatchSpanProcessor :: OtelBatchSpanProcessorConfig

emptyOpenTelemetryConfig :: OpenTelemetryConfig

data OpenTelemetryConfigSubobject

Subsets of the fields of OpenTelemetryConfig, serving as metadata object names for MetadataObjId.

Constructors

OtelSubobjectAll
The entire OpenTelemetry configuration

OtelSubobjectExporterOtlp

OtelSubobjectBatchSpanProcessor

data OtelStatus

Should the OpenTelemetry exporter be enabled?

Constructors

OtelEnabled

OtelDisabled

data OtelDataType

Constructors

OtelTraces

OtelMetrics

data OtelExporterConfig

https://opentelemetry.io/docs/reference/specification/protocol/exporter/

Constructors

OtelExporterConfig

Fields

_oecTracesEndpoint :: Maybe Text

Target URL to which the exporter is going to send traces. No default. Used as-is without modification (e.g. appending v1traces).

_oecMetricsEndpoint :: Maybe Text

Target URL to which the exporter is going to send metrics. No default. Used as-is without modification (e.g. appending v1metrics).

_oecProtocol :: OtlpProtocol

The transport protocol, for all telemetry types.

_oecHeaders :: [HeaderConf]

Key-value pairs to be used as headers to send with an export request, for all telemetry types. We currently only support string-valued attributes. Like OTEL_EXPORTER_OTLP_HEADERS.

_oecResourceAttributes :: [NameValue]

Attributes to send as the resource attributes of an export request, for all telemetry types.

_oecTracesPropagators :: [TracePropagator]

Trace propagator to be used to extract and inject trace headers

defaultOtelExporterConfig :: OtelExporterConfig

data OtlpProtocol

Possible protocol to use with OTLP. Currently, only http/protobuf is supported.

Constructors

OtlpProtocolHttpProtobuf

data NameValue

Constructors

NameValue

Fields

nv_name :: Text

nv_value :: Text

data TracePropagator

Constructors

B3

TraceContext

defaultOtelExporterTracesPropagators :: [TracePropagator]

newtype OtelBatchSpanProcessorConfig

Constructors

OtelBatchSpanProcessorConfig

Fields

_obspcMaxExportBatchSize :: Int

The maximum batch size of every export. It must be smaller or equal to maxQueueSize (not yet configurable). Default 512.

defaultOtelBatchSpanProcessorConfig :: OtelBatchSpanProcessorConfig

ocStatus :: Lens' OpenTelemetryConfig OtelStatus

ocExporterOtlp :: Lens' OpenTelemetryConfig OtelExporterConfig

ocEnabledDataTypes :: Lens' OpenTelemetryConfig (Set OtelDataType)

ocBatchSpanProcessor :: Lens' OpenTelemetryConfig OtelBatchSpanProcessorConfig

Parsed configuration (schema cache)

data OpenTelemetryInfo

Schema cache configuration for all OpenTelemetry-related features

Constructors

OpenTelemetryInfo

Fields

_otiExporterOtlp :: OtelExporterInfo

_otiBatchSpanProcessorInfo :: OtelBatchSpanProcessorInfo

This configuration will be used for traces and logs (when implemented)

data OtelExporterInfo

Constructors

OtelExporterInfo

Fields

_oteleiTracesBaseRequest :: Maybe Request

HTTP Request containing (1) the target URL to which the exporter is going to send spans, and (2) the user-specified request headers. A value of Nothing indicates that the export of trace data is disabled.

_oteleiMetricsBaseRequest :: Maybe Request

HTTP Request containing (1) the target URL to which the exporter is going to send metrics, and (2) the user-specified request headers. A value of Nothing indicates that the export of trace data is disabled.

_oteleiResourceAttributes :: Map Text Text

Attributes to send as the resource attributes of an export request. We currently only support string-valued attributes.

Using Data.Map.Strict over Data.Hashmap.Strict because currently the only operations on data are (1) folding and (2) union with a small map of default attributes, and Map should be is faster than HashMap for the latter.

_oteleiTracesPropagator :: Propagator RequestHeaders ResponseHeaders

Trace propagator to be used to extract and inject trace headers

emptyOtelExporterInfo :: OtelExporterInfo

data OtelBatchSpanProcessorInfo

Batch processor configuration for trace export.

https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/sdk.md

NOTE: we could share this configuration with logs when implemented, but that change is invasive; I recommend just adding a new separate configuration. See: https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/logs/sdk.md#batching-processor

Constructors

OtelBatchSpanProcessorInfo

Fields

_obspiMaxExportBatchSize :: Int

The maximum batch size of every export. It must be smaller or equal to maxQueueSize. Default 512.

_obspiMaxQueueSize :: Int

The maximum span queue size. After the size is reached spans are dropped. Default 2048.

getMaxExportBatchSize :: OtelBatchSpanProcessorInfo -> Int

getMaxQueueSize :: OtelBatchSpanProcessorInfo -> Int

mkOtelTracesPropagator :: [TracePropagator] -> HttpPropagator

getOtelTracesPropagator :: OpenTelemetryInfo -> HttpPropagator

defaultOtelBatchSpanProcessorInfo :: OtelBatchSpanProcessorInfo

Defaults taken from https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/sdk.md#batching-processor

otiExporterOtlp :: Lens' OpenTelemetryInfo OtelExporterInfo

otiBatchSpanProcessorInfo :: Lens' OpenTelemetryInfo OtelBatchSpanProcessorInfo