Hasura.RQL.Types.Relationships.Remote

Documentation

type RemoteRelationship = RemoteRelationshipG RemoteRelationshipDefinition

data RemoteRelationshipDefinition

Metadata representation of the internal definition of a remote relationship.

Constructors

RelationshipToSource ToSourceRelationshipDef
Remote relationship targetting a source.

RelationshipToSchema RRFormat ToSchemaRelationshipDef
Remote relationship targetting a remote schema.

data RemoteSourceRelationshipBuilder

Specify whether remote schema <> source relationships should be built

Constructors

IncludeRemoteSourceRelationship

ExcludeRemoteSourceRelationship

parseRemoteRelationshipDefinition :: RRParseMode -> Value -> Parser RemoteRelationshipDefinition

Parse RemoteRelationshipDefinition letting the caller decide how lenient to be.

This is necessary because RemoteRelationshipDefinition is parsed in different contexts. In RemoteRelationship, the RemoteRelationshipDefinition is always parsed out from a top-level @"definition" field. Thus, a legacy payload looks like this:

{
  "name": "thing",
  "definition": {
    "remote_schema": "stuff",
    "hasura_fields": ...
    "remote_field": ...
  }
}

and a new payload looks like this:

{
  "name": "thing",
  "definition": {
    "to_remote_schema": {
      "schema": "stuff",
      "lhs_fields": ...
      "remote_field": ...
    }
  }
}

In contrast, CreateFromSourceRelationship does not have a top- level "definition" in its legacy format. Instead, the legacy fields themselves are top-level:

{
  "remote_schema": "stuff",
  "hasura_fields": ...
  "remote_field": ...
}

Furthermore, the presence of a "definition" field is used to detect that the new payload is being used:

{
  "definition": {
    "to_remote_schema": {
      "schema": "stuff",
      "lhs_fields": ...
      "remote_field": ...
    }
  }
}

In this latter case, we should not allow "remote_schema" to appear under "definition".

data RRFormat

Represents the format of the metadata a remote relationship was read from and must be written back as. We don't have a good way of doing metadata versioning yet, and we therefore use this to keep track of the format used.

Constructors

RRFOldDBToRemoteSchema
The remote relationship was parsed from the old format, that was only used only for db-to-rs schemas.

RRFUnifiedFormat
The remote relationship was parsed from the new unified format.

data RRParseMode

Whether to accept legacy fields when parsing RemoteRelationshipDefinition

Constructors

RRPLegacy
Only allow legacy fields when parsing RemoteRelationshipDefinition

RRPLenient
Allow legacy fields when parsing RemoteRelationshipDefinition

RRPStrict
Reject legacy fields when parsing RemoteRelationshipDefinition

_RelationshipToSource :: Prism' RemoteRelationshipDefinition ToSourceRelationshipDef

_RelationshipToSchema :: Prism' RemoteRelationshipDefinition (RRFormat, ToSchemaRelationshipDef)

rrName :: forall def f. Functor f => (RelName -> f RelName) -> RemoteRelationshipG def -> f (RemoteRelationshipG def)

rrDefinition :: forall def def' f. Functor f => (def -> f def') -> RemoteRelationshipG def -> f (RemoteRelationshipG def')

data RemoteSchemaFieldInfo

Schema cache information for a table field targeting a remote schema.

Constructors

RemoteSchemaFieldInfo

Fields

_rrfiName :: RelName

Field name to which we'll map the remote in hasura; this becomes part of the hasura schema.

_rrfiParamMap :: HashMap Name RemoteSchemaInputValueDefinition

Input arguments to the remote field info; The _rfiParamMap will only include the arguments to the remote field that is being joined. The names of the arguments here are modified, it will be in the format of Field Name_remote_rel_table schema_table namerelationship name

_rrfiRemoteFields :: RemoteFields

_rrfiRemoteSchema :: RemoteSchemaInfo

_rrfiInputValueDefinitions :: [TypeDefinition [Name] RemoteSchemaInputValueDefinition]

The new input value definitions created for this remote field

_rrfiRemoteSchemaName :: RemoteSchemaName

Name of the remote schema, that's used for joining

_rrfiLHSIdentifier :: LHSIdentifier

TODO: this one should be gone when validateRemoteRelationship function is cleaned up

data RemoteSourceFieldInfo (tgt :: BackendType)

Schema cache information for a table field targeting a remote source.

Constructors

RemoteSourceFieldInfo

Fields

_rsfiName :: RelName

_rsfiType :: RelType

_rsfiSource :: SourceName

_rsfiSourceConfig :: SourceConfig tgt

_rsfiTable :: TableName tgt

this is parsed from Value

_rsfiMapping :: HashMap FieldName (ScalarType tgt, Column tgt)

LHS field name -> RHS Column, RHS Column type

data RemoteFieldInfoRHS

Resolved remote relationship's RHS

Constructors

RFISchema RemoteSchemaFieldInfo

RFISource (AnyBackend RemoteSourceFieldInfo)

data RemoteFieldInfo lhsJoinField

Resolved remote relationship, as stored in the schema cache.

Constructors

RemoteFieldInfo

Fields

_rfiLHS :: HashMap FieldName lhsJoinField

_rfiRHS :: RemoteFieldInfoRHS

data DBJoinField (b :: BackendType)

Information about the field on the LHS of a join against a remote schema.

Constructors

JoinColumn (Column b) (ColumnType b)

JoinComputedField (ScalarComputedField b)

data ScalarComputedField (b :: BackendType)

Information about a computed field appearing on the LHS of a remote join. FIXME: why do we need all of this?

Constructors

ScalarComputedField

Fields

_scfXField :: XComputedField b

_scfName :: ComputedFieldName

_scfFunction :: FunctionName b

_scfComputedFieldImplicitArgs :: ComputedFieldImplicitArguments b

_scfType :: ScalarType b

graphQLValueToJSON :: Value Void -> Value

newtype LHSIdentifier

Constructors

LHSIdentifier

Fields

getLHSIdentifier :: Text

tableNameToLHSIdentifier :: forall (b :: BackendType). Backend b => TableName b -> LHSIdentifier