Hasura.RQL.Types.SchemaCache

Documentation

newtype MetadataResourceVersion

Constructors

MetadataResourceVersion

Fields

getMetadataResourceVersion :: Int64

initialResourceVersion :: MetadataResourceVersion

mkParentDep :: forall b. Backend b => SourceName -> TableName b -> SchemaDependency

mkColDep :: forall b. Backend b => DependencyReason -> SourceName -> TableName b -> Column b -> SchemaDependency

mkComputedFieldDep :: forall b. Backend b => DependencyReason -> SourceName -> TableName b -> ComputedFieldName -> SchemaDependency

type WithDeps a = (a, [SchemaDependency])

data IntrospectionResult

Constructors

IntrospectionResult

Fields

irDoc :: RemoteSchemaIntrospection

irQueryRoot :: Name

irMutationRoot :: Maybe Name

irSubscriptionRoot :: Maybe Name

type RemoteSchemaRelationships = InsOrdHashMap Name (InsOrdHashMap RelName (RemoteFieldInfo Name))

data RemoteSchemaCtx

See fetchRemoteSchema.

Constructors

RemoteSchemaCtx

Fields

_rscName :: RemoteSchemaName

_rscIntroOriginal :: IntrospectionResult

Original remote schema without customizations

_rscInfo :: RemoteSchemaInfo

_rscRawIntrospectionResult :: ByteString

The raw response from the introspection query against the remote server. We store this so we can efficiently service introspect_remote_schema.

_rscPermissions :: HashMap RoleName IntrospectionResult

_rscRemoteRelationships :: RemoteSchemaRelationships

getIntrospectionResult :: RemoteSchemaPermissions -> RoleName -> RemoteSchemaCtx -> Maybe IntrospectionResult

rscRemoteRelationships :: Lens' RemoteSchemaCtx RemoteSchemaRelationships

rscRawIntrospectionResult :: Lens' RemoteSchemaCtx ByteString

rscPermissions :: Lens' RemoteSchemaCtx (HashMap RoleName IntrospectionResult)

rscName :: Lens' RemoteSchemaCtx RemoteSchemaName

rscIntroOriginal :: Lens' RemoteSchemaCtx IntrospectionResult

rscInfo :: Lens' RemoteSchemaCtx RemoteSchemaInfo

type RemoteSchemaMap = HashMap RemoteSchemaName RemoteSchemaCtx

type DepMap = HashMap SchemaObjId (HashSet SchemaDependency)

data CronTriggerInfo

Constructors

CronTriggerInfo

Fields

ctiName :: TriggerName

ctiSchedule :: CronSchedule

ctiPayload :: Maybe Value

ctiRetryConf :: STRetryConf

ctiWebhookInfo :: EnvRecord ResolvedWebhook

ctiHeaders :: [EventHeaderInfo]

ctiComment :: Maybe Text

ctiRequestTransform :: Maybe RequestTransform

ctiResponseTransform :: Maybe MetadataResponseTransform

newtype SchemaCacheVer

Constructors

SchemaCacheVer

Fields

unSchemaCacheVer :: Word64

initSchemaCacheVer :: SchemaCacheVer

incSchemaCacheVer :: SchemaCacheVer -> SchemaCacheVer

type ActionCache = HashMap ActionName ActionInfo

type InheritedRolesCache = HashMap RoleName (HashSet RoleName)

askSourceInfo :: forall b m. (CacheRM m, MetadataM m, MonadError QErr m, Backend b) => SourceName -> m (SourceInfo b)

Retrieves the source info for a given source name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. This function must be used with a _type annotation_, such as `askSourceInfo @('Postgres 'Vanilla)`. It throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

askSourceConfig :: forall b m. (CacheRM m, MonadError QErr m, Backend b, MetadataM m) => SourceName -> m (SourceConfig b)

Retrieves the source config for a given source name.

This function relies on askSourceInfo and similarly throws an error if the source isn't found.

unsafeTableCache :: forall b. Backend b => SourceName -> SourceCache -> Maybe (TableCache b)

Retrieves the table cache for a given source cache and source name.

This function must be used with a _type annotation_, such as `unsafeTableCache ('Postgres 'Vanilla)`. It returns Nothing@ if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

askTableCache :: forall b m. (Backend b, CacheRM m) => SourceName -> m (Maybe (TableCache b))

Retrieves the table cache for a given source name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. It must be used with a _type annotation_, such as `unsafeTableCache ('Postgres 'Vanilla)`. It returns Nothing@ if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

unsafeTableInfo :: forall b. Backend b => SourceName -> TableName b -> SourceCache -> Maybe (TableInfo b)

Retrieves the information about a table from the source cache, the source name, and the table name.

This function returns Nothing if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

askTableInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (TableInfo b)

Retrieves the information about a table for a given source name and table name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. it throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

askTableCoreInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (TableCoreInfo b)

Similar to askTableInfo, but drills further down to extract the underlying core info.

askTableFieldInfoMap :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (FieldInfoMap (FieldInfo b))

Similar to askTableCoreInfo, but drills further down to extract the underlying field info map.

askTableMetadata :: forall b m. (QErrM m, MetadataM m, Backend b) => SourceName -> TableName b -> m (TableMetadata b)

Retrieves the metadata information about a table for a given source name and table name.

Unlike most other ask functions in this module, this function does not drill through the schema cache, and instead inspects the metadata. Like most others, it throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

unsafeFunctionCache :: forall b. Backend b => SourceName -> SourceCache -> Maybe (FunctionCache b)

Retrieves the function cache for a given source cache and source name.

This function must be used with a _type annotation_, such as `unsafeFunctionCache ('Postgres 'Vanilla)`. It returns Nothing@ if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

unsafeFunctionInfo :: forall b. Backend b => SourceName -> FunctionName b -> SourceCache -> Maybe (FunctionInfo b)

Retrieves the information about a function from the source cache, the source name, and the function name.

This function returns Nothing if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

askFunctionInfo :: forall b m. (QErrM m, CacheRM m, Backend b) => SourceName -> FunctionName b -> m (FunctionInfo b)

Retrieves the information about a function cache for a given source name and function name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. It throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

newtype BackendInfoWrapper (b :: BackendType)

Constructors

BackendInfoWrapper

Fields

unBackendInfoWrapper :: BackendInfo b

type BackendCache = BackendMap BackendInfoWrapper

data SchemaCache

Constructors

SchemaCache

Fields

scSources :: SourceCache

scActions :: ActionCache

scRemoteSchemas :: RemoteSchemaMap

scAllowlist :: InlinedAllowlist

scAdminIntrospection :: SchemaIntrospection

scGQLContext :: HashMap RoleName (RoleContext GQLContext)

scUnauthenticatedGQLContext :: GQLContext

scRelayContext :: HashMap RoleName (RoleContext GQLContext)

scUnauthenticatedRelayContext :: GQLContext

scDepMap :: DepMap

scInconsistentObjs :: [InconsistentMetadata]

scCronTriggers :: HashMap TriggerName CronTriggerInfo

scEndpoints :: EndpointTrie GQLQueryWithText

scApiLimits :: ApiLimit

scMetricsConfig :: MetricsConfig

scMetadataResourceVersion :: Maybe MetadataResourceVersion

scSetGraphqlIntrospectionOptions :: SetGraphqlIntrospectionOptions

scTlsAllowlist :: [TlsAllow]

scQueryCollections :: QueryCollections

scBackendCache :: BackendCache

scSourceHealthChecks :: SourceHealthCheckCache

getAllRemoteSchemas :: SchemaCache -> [RemoteSchemaName]

class Monad m => SourceM m where

Methods

askCurrentSource :: m SourceName

newtype SourceT m a

Constructors

SourceT

Fields

runSourceT :: SourceName -> m a

class SourceM m => TableCoreInfoRM b m where

A more limited version of CacheRM that is used when building the schema cache, since the entire schema cache has not been built yet.

Methods

lookupTableCoreInfo :: TableName b -> m (Maybe (TableCoreInfo b))

newtype TableCoreCacheRT b m a

Constructors

TableCoreCacheRT

Fields

runTableCoreCacheRT :: (SourceName, Dependency (TableCoreCache b)) -> m a

class TableCoreInfoRM b m => TableInfoRM b m where

All our RQL DML queries operate over a single source. This typeclass facilitates that.

Methods

lookupTableInfo :: TableName b -> m (Maybe (TableInfo b))

newtype TableCacheRT b m a

Constructors

TableCacheRT

Fields

runTableCacheRT :: (SourceName, TableCache b) -> m a

class Monad m => CacheRM m where

Methods

askSchemaCache :: m SchemaCache

getDependentObjs :: SchemaCache -> SchemaObjId -> [SchemaObjId]

getDependentObjsWith :: (DependencyReason -> Bool) -> SchemaCache -> SchemaObjId -> [SchemaObjId]

getRemoteDependencies :: SchemaCache -> SourceName -> [SchemaObjId]

Compute all remote dependencies on a source.

Given a source name, this function computes all of its dependencies, direct or indirect, and returns all of the dependencies that are not "local" to the source, i.e. that belong to another source or to a remote schema, here dubbed "remote dependencies".

This functions returns a SchemaObjId for each such dependency, but makes no attempt at extracting the underlying SourceObjId (if any), for two reasons: 1. a SourceObjId no longer contains the source name, which most callers need to identify where the corresponding dependency is 2. this would prevent us from returning remote schema dependencies, which by definition do not have a corresponding SourceObjId

getBoolExpDeps :: forall b. GetAggregationPredicatesDeps b => SourceName -> TableName b -> AnnBoolExpPartialSQL b -> [SchemaDependency]

Discover the schema dependencies of an AnnBoolExpPartialSQL.

getBoolExpDeps' :: forall b. (Backend b, GetAggregationPredicatesDeps b) => AnnBoolExpPartialSQL b -> BoolExpM b [SchemaDependency]

getColExpDeps :: forall b. (Backend b, GetAggregationPredicatesDeps b) => AnnBoolExpFld b (PartialSQLExp b) -> BoolExpM b [SchemaDependency]

getOpExpDeps :: forall b. Backend b => [OpExpG b (PartialSQLExp b)] -> BoolExpM b [SchemaDependency]

askFieldInfoMapSource :: (QErrM m, Backend b, TableCoreInfoRM b m) => TableName b -> m (FieldInfoMap (FieldInfo b))

Asking for a table's fields info without explicit SourceName argument. The source name is implicitly inferred from SourceM via @TableCoreInfoRM.