Hasura.RQL.Types.SchemaCache

Documentation

newtype MetadataResourceVersion

Constructors

MetadataResourceVersion

Fields

getMetadataResourceVersion :: Int64

initialResourceVersion :: MetadataResourceVersion

mkParentDep :: forall b. Backend b => SourceName -> TableName b -> SchemaDependency

mkColDep :: forall b. Backend b => DependencyReason -> SourceName -> TableName b -> Column b -> SchemaDependency

mkComputedFieldDep :: forall b. Backend b => DependencyReason -> SourceName -> TableName b -> ComputedFieldName -> SchemaDependency

type WithDeps a = (a, [SchemaDependency])

data IntrospectionResult

Constructors

IntrospectionResult

Fields

irDoc :: !RemoteSchemaIntrospection

irQueryRoot :: !Name

irMutationRoot :: !(Maybe Name)

irSubscriptionRoot :: !(Maybe Name)

type RemoteSchemaRelationships = InsOrdHashMap Name (InsOrdHashMap RelName (RemoteFieldInfo Name))

data RemoteSchemaCtx

See fetchRemoteSchema.

Constructors

RemoteSchemaCtx

Fields

_rscName :: !RemoteSchemaName

_rscIntroOriginal :: !IntrospectionResult

Original remote schema without customizations

_rscInfo :: !RemoteSchemaInfo

_rscRawIntrospectionResult :: !ByteString

The raw response from the introspection query against the remote server. We store this so we can efficiently service introspect_remote_schema.

_rscPermissions :: !(HashMap RoleName IntrospectionResult)

_rscRemoteRelationships :: RemoteSchemaRelationships

getIntrospectionResult :: RemoteSchemaPermsCtx -> RoleName -> RemoteSchemaCtx -> Maybe IntrospectionResult

rscRemoteRelationships :: Lens' RemoteSchemaCtx RemoteSchemaRelationships

rscRawIntrospectionResult :: Lens' RemoteSchemaCtx ByteString

rscPermissions :: Lens' RemoteSchemaCtx (HashMap RoleName IntrospectionResult)

rscName :: Lens' RemoteSchemaCtx RemoteSchemaName

rscIntroOriginal :: Lens' RemoteSchemaCtx IntrospectionResult

rscInfo :: Lens' RemoteSchemaCtx RemoteSchemaInfo

type RemoteSchemaMap = HashMap RemoteSchemaName RemoteSchemaCtx

type DepMap = HashMap SchemaObjId (HashSet SchemaDependency)

data CronTriggerInfo

Constructors

CronTriggerInfo

Fields

ctiName :: !TriggerName

ctiSchedule :: !CronSchedule

ctiPayload :: !(Maybe Value)

ctiRetryConf :: !STRetryConf

ctiWebhookInfo :: !ResolvedWebhook

ctiHeaders :: ![EventHeaderInfo]

ctiComment :: !(Maybe Text)

ctiRequestTransform :: !(Maybe RequestTransform)

ctiResponseTransform :: !(Maybe MetadataResponseTransform)

newtype SchemaCacheVer

Constructors

SchemaCacheVer

Fields

unSchemaCacheVer :: Word64

initSchemaCacheVer :: SchemaCacheVer

incSchemaCacheVer :: SchemaCacheVer -> SchemaCacheVer

type ActionCache = HashMap ActionName ActionInfo

unsafeFunctionCache :: forall b. Backend b => SourceName -> SourceCache -> Maybe (FunctionCache b)

unsafeFunctionInfo :: forall b. Backend b => SourceName -> FunctionName b -> SourceCache -> Maybe (FunctionInfo b)

type InheritedRolesCache = HashMap RoleName (HashSet RoleName)

unsafeTableCache :: forall b. Backend b => SourceName -> SourceCache -> Maybe (TableCache b)

unsafeTableInfo :: forall b. Backend b => SourceName -> TableName b -> SourceCache -> Maybe (TableInfo b)

data SchemaCache

Constructors

SchemaCache

Fields

scSources :: !SourceCache

scActions :: !ActionCache

scRemoteSchemas :: !RemoteSchemaMap

scAllowlist :: !InlinedAllowlist

scAdminIntrospection :: !SchemaIntrospection

scGQLContext :: !(HashMap RoleName (RoleContext GQLContext))

scUnauthenticatedGQLContext :: !GQLContext

scRelayContext :: !(HashMap RoleName (RoleContext GQLContext))

scUnauthenticatedRelayContext :: !GQLContext

scDepMap :: !DepMap

scInconsistentObjs :: ![InconsistentMetadata]

scCronTriggers :: !(HashMap TriggerName CronTriggerInfo)

scEndpoints :: !(EndpointTrie GQLQueryWithText)

scApiLimits :: !ApiLimit

scMetricsConfig :: !MetricsConfig

scMetadataResourceVersion :: !(Maybe MetadataResourceVersion)

scSetGraphqlIntrospectionOptions :: !SetGraphqlIntrospectionOptions

scTlsAllowlist :: ![TlsAllow]

scQueryCollections :: !QueryCollections

getAllRemoteSchemas :: SchemaCache -> [RemoteSchemaName]

class Monad m => SourceM m where

Methods

askCurrentSource :: m SourceName

newtype SourceT m a

Constructors

SourceT

Fields

runSourceT :: SourceName -> m a

class SourceM m => TableCoreInfoRM b m where

A more limited version of CacheRM that is used when building the schema cache, since the entire schema cache has not been built yet.

Methods

lookupTableCoreInfo :: TableName b -> m (Maybe (TableCoreInfo b))

newtype TableCoreCacheRT b m a

Constructors

TableCoreCacheRT

Fields

runTableCoreCacheRT :: (SourceName, Dependency (TableCoreCache b)) -> m a

class TableCoreInfoRM b m => TableInfoRM b m where

All our RQL DML queries operate over a single source. This typeclass facilitates that.

Methods

lookupTableInfo :: TableName b -> m (Maybe (TableInfo b))

newtype TableCacheRT b m a

Constructors

TableCacheRT

Fields

runTableCacheRT :: (SourceName, TableCache b) -> m a

class Monad m => CacheRM m where

Methods

askSchemaCache :: m SchemaCache

getDependentObjs :: SchemaCache -> SchemaObjId -> [SchemaObjId]

getDependentObjsWith :: (DependencyReason -> Bool) -> SchemaCache -> SchemaObjId -> [SchemaObjId]

data BoolExpCtx b

The table type context of schema dependency discovery. Boolean expressions may refer to a so-called 'root table' (identified by a $-sign in the expression input syntax) or the current table.

Constructors

BoolExpCtx

Fields

source :: SourceName

currTable :: TableName b

Reference to the current table type.

rootTable :: TableName b

Reference to the root table type.

getBoolExpDeps :: forall b. Backend b => SourceName -> TableName b -> AnnBoolExpPartialSQL b -> [SchemaDependency]

Discover the schema dependencies of an AnnBoolExpPartialSQL.

newtype BoolExpM b a

The monad for doing schema dependency discovery for boolean expressions. maintains the table context of the expressions being translated.

Constructors

BoolExpM

Fields

unBoolExpM :: Reader (BoolExpCtx b) a

runBoolExpM :: BoolExpCtx b -> BoolExpM b a -> a

getBoolExpDeps' :: forall b. Backend b => AnnBoolExpPartialSQL b -> BoolExpM b [SchemaDependency]

getColExpDeps :: forall b. Backend b => AnnBoolExpFld b (PartialSQLExp b) -> BoolExpM b [SchemaDependency]