Hasura.RQL.Types.SchemaCache

Documentation

data SchemaCache

Constructors

SchemaCache

Fields

scSources :: SourceCache

scActions :: ActionCache

scRemoteSchemas :: RemoteSchemaMap

scAllowlist :: InlinedAllowlist

scAdminIntrospection :: SchemaIntrospection

scGQLContext :: HashMap RoleName (RoleContext GQLContext)

scUnauthenticatedGQLContext :: GQLContext

scRelayContext :: HashMap RoleName (RoleContext GQLContext)

scUnauthenticatedRelayContext :: GQLContext

scDepMap :: DepMap

scInconsistentObjs :: [InconsistentMetadata]

scCronTriggers :: HashMap TriggerName CronTriggerInfo

scEndpoints :: EndpointTrie GQLQueryWithText

scApiLimits :: ApiLimit

scMetricsConfig :: MetricsConfig

scMetadataResourceVersion :: MetadataResourceVersion

scSetGraphqlIntrospectionOptions :: SetGraphqlIntrospectionOptions

scTlsAllowlist :: [TlsAllow]

scQueryCollections :: QueryCollections

scBackendCache :: BackendCache

scSourceHealthChecks :: SourceHealthCheckCache

scSourcePingConfig :: SourcePingCache

scOpenTelemetryConfig :: OpenTelemetryInfo

data TableConfig (b :: BackendType)

Constructors

TableConfig

Fields

_tcCustomRootFields :: TableCustomRootFields

_tcColumnConfig :: HashMap (Column b) ColumnConfig

_tcCustomName :: Maybe Name

_tcComment :: Comment

emptyTableConfig :: forall (b :: BackendType). TableConfig b

getAllRemoteSchemas :: SchemaCache -> [RemoteSchemaName]

unsafeFunctionCache :: forall (b :: BackendType). Backend b => SourceName -> SourceCache -> Maybe (FunctionCache b)

Retrieves the function cache for a given source cache and source name.

This function must be used with a _type annotation_, such as `unsafeFunctionCache ('Postgres 'Vanilla)`. It returns Nothing@ if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

unsafeFunctionInfo :: forall (b :: BackendType). Backend b => SourceName -> FunctionName b -> SourceCache -> Maybe (FunctionInfo b)

Retrieves the information about a function from the source cache, the source name, and the function name.

This function returns Nothing if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

unsafeTableCache :: forall (b :: BackendType). Backend b => SourceName -> SourceCache -> Maybe (TableCache b)

Retrieves the table cache for a given source cache and source name.

This function must be used with a _type annotation_, such as `unsafeTableCache ('Postgres 'Vanilla)`. It returns Nothing@ if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

unsafeTableInfo :: forall (b :: BackendType). Backend b => SourceName -> TableName b -> SourceCache -> Maybe (TableInfo b)

Retrieves the information about a table from the source cache, the source name, and the table name.

This function returns Nothing if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

askSourceInfo :: forall (b :: BackendType) m. (CacheRM m, MetadataM m, MonadError QErr m, Backend b) => SourceName -> m (SourceInfo b)

Retrieves the source info for a given source name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. This function must be used with a _type annotation_, such as `askSourceInfo @('Postgres 'Vanilla)`. It throws an error if: 1. The function fails to find the named source at all 2. The named source exists but does not match the expected type 3. The named source exists, and is of the expected type, but is inconsistent

askSourceInfoMaybe :: forall (b :: BackendType) m. (CacheRM m, Backend b) => SourceName -> m (Maybe (SourceInfo b))

askSourceConfig :: forall (b :: BackendType) m. (CacheRM m, MonadError QErr m, Backend b, MetadataM m) => SourceName -> m (SourceConfig b)

Retrieves the source config for a given source name.

This function relies on askSourceInfo and similarly throws an error if the source isn't found.

askSourceConfigMaybe :: forall (b :: BackendType) m. (CacheRM m, Backend b) => SourceName -> m (Maybe (SourceConfig b))

askTableCache :: forall (b :: BackendType) m. (Backend b, CacheRM m) => SourceName -> m (Maybe (TableCache b))

Retrieves the table cache for a given source name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. It must be used with a _type annotation_, such as `unsafeTableCache ('Postgres 'Vanilla)`. It returns Nothing@ if it fails to find that source or if the kind of the source does not match the type annotation, and does not distinguish between the two cases.

askTableInfo :: forall (b :: BackendType) m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (TableInfo b)

Retrieves the information about a table for a given source name and table name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. it throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

askTableCoreInfo :: forall (b :: BackendType) m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (TableCoreInfo b)

Similar to askTableInfo, but drills further down to extract the underlying core info.

askTableFieldInfoMap :: forall (b :: BackendType) m. (QErrM m, CacheRM m, Backend b) => SourceName -> TableName b -> m (FieldInfoMap (FieldInfo b))

Similar to askTableCoreInfo, but drills further down to extract the underlying field info map.

askLogicalModelCache :: forall (b :: BackendType) m. (Backend b, CacheRM m) => SourceName -> m (Maybe (LogicalModelCache b))

askTableMetadata :: forall (b :: BackendType) m. (QErrM m, MetadataM m, Backend b) => SourceName -> TableName b -> m (TableMetadata b)

Retrieves the metadata information about a table for a given source name and table name.

Unlike most other ask functions in this module, this function does not drill through the schema cache, and instead inspects the metadata. Like most others, it throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

askFunctionInfo :: forall (b :: BackendType) m. (QErrM m, CacheRM m, Backend b) => SourceName -> FunctionName b -> m (FunctionInfo b)

Retrieves the information about a function cache for a given source name and function name.

This function retrieves the schema cache from the monadic context, and attempts to look the corresponding source up in the source cache. It throws an error if it fails to find that source, in which case it looks that source up in the metadata, to differentiate between the source not existing or the type of the source not matching.

askFieldInfoMapSource :: forall m (b :: BackendType). (QErrM m, Backend b, TableCoreInfoRM b m) => TableName b -> m (FieldInfoMap (FieldInfo b))

Asking for a table's fields info without explicit SourceName argument. The source name is implicitly inferred from SourceM via @TableCoreInfoRM.

type TableCoreCache (b :: BackendType) = HashMap (TableName b) (TableCoreInfo b)

type TableCache (b :: BackendType) = HashMap (TableName b) (TableInfo b)

type ActionCache = HashMap ActionName ActionInfo

type InheritedRolesCache = HashMap RoleName (HashSet RoleName)

data TableCoreInfoG (b :: BackendType) field primaryKeyColumn

The field and primaryKeyColumn type parameters vary as the schema cache is built and more information is accumulated. See also TableCoreInfo.

Constructors

TableCoreInfo

Fields

_tciName :: TableName b

_tciDescription :: Maybe PGDescription

_tciFieldInfoMap :: FieldInfoMap field

_tciPrimaryKey :: Maybe (PrimaryKey b primaryKeyColumn)

_tciUniqueConstraints :: HashSet (UniqueConstraint b)

Does not include the primary key; use tciUniqueOrPrimaryKeyConstraints if you need both.

_tciForeignKeys :: HashSet (ForeignKey b)

_tciViewInfo :: Maybe ViewInfo

_tciEnumValues :: Maybe EnumValues

_tciCustomConfig :: TableConfig b

_tciExtraTableMetadata :: ExtraTableMetadata b

_tciApolloFederationConfig :: Maybe ApolloFederationConfig

_tciRawColumns :: [RawColumnInfo b]

type TableCoreInfo (b :: BackendType) = TableCoreInfoG b (FieldInfo b) (ColumnInfo b)

Fully-processed table info that includes non-column fields.

tciName :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (TableName b -> f (TableName b)) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciDescription :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (Maybe PGDescription -> f (Maybe PGDescription)) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciFieldInfoMap :: forall (b :: BackendType) field1 primaryKeyColumn field2 f. Functor f => (FieldInfoMap field1 -> f (FieldInfoMap field2)) -> TableCoreInfoG b field1 primaryKeyColumn -> f (TableCoreInfoG b field2 primaryKeyColumn)

tciPrimaryKey :: forall (b :: BackendType) field primaryKeyColumn1 primaryKeyColumn2 f. Functor f => (Maybe (PrimaryKey b primaryKeyColumn1) -> f (Maybe (PrimaryKey b primaryKeyColumn2))) -> TableCoreInfoG b field primaryKeyColumn1 -> f (TableCoreInfoG b field primaryKeyColumn2)

tciUniqueConstraints :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (HashSet (UniqueConstraint b) -> f (HashSet (UniqueConstraint b))) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciForeignKeys :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (HashSet (ForeignKey b) -> f (HashSet (ForeignKey b))) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciViewInfo :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (Maybe ViewInfo -> f (Maybe ViewInfo)) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciEnumValues :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (Maybe EnumValues -> f (Maybe EnumValues)) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciCustomConfig :: forall (b :: BackendType) field primaryKeyColumn f. Functor f => (TableConfig b -> f (TableConfig b)) -> TableCoreInfoG b field primaryKeyColumn -> f (TableCoreInfoG b field primaryKeyColumn)

tciUniqueOrPrimaryKeyConstraints :: forall (b :: BackendType) f. Hashable (Column b) => TableCoreInfoG b f (ColumnInfo b) -> Maybe (NonEmpty (UniqueConstraint b))

data TableInfo (b :: BackendType)

Constructors

TableInfo

Fields

_tiCoreInfo :: TableCoreInfo b

_tiRolePermInfoMap :: RolePermInfoMap b

_tiEventTriggerInfoMap :: EventTriggerInfoMap b

_tiAdminRolePermInfo :: RolePermInfo b

tiCoreInfo :: forall (b :: BackendType) f. Functor f => (TableCoreInfo b -> f (TableCoreInfo b)) -> TableInfo b -> f (TableInfo b)

tiRolePermInfoMap :: forall (b :: BackendType) f. Functor f => (RolePermInfoMap b -> f (RolePermInfoMap b)) -> TableInfo b -> f (TableInfo b)

tiEventTriggerInfoMap :: forall (b :: BackendType) f. Functor f => (EventTriggerInfoMap b -> f (EventTriggerInfoMap b)) -> TableInfo b -> f (TableInfo b)

data ViewInfo

Constructors

ViewInfo

Fields

viIsUpdatable :: Bool

viIsDeletable :: Bool

viIsInsertable :: Bool

isMutable :: (ViewInfo -> Bool) -> Maybe ViewInfo -> Bool

data IntrospectionResult

Constructors

IntrospectionResult

Fields

irDoc :: RemoteSchemaIntrospection

irQueryRoot :: Name

irMutationRoot :: Maybe Name

irSubscriptionRoot :: Maybe Name

data RemoteSchemaCustomizer

Constructors

RemoteSchemaCustomizer

Fields

_rscNamespaceFieldName :: Maybe Name

_rscCustomizeTypeName :: HashMap Name Name

type name -> type name

_rscCustomizeFieldName :: HashMap Name (HashMap Name Name)

type name -> field name -> field name

remoteSchemaCustomizeTypeName :: RemoteSchemaCustomizer -> MkTypename

remoteSchemaCustomizeFieldName :: RemoteSchemaCustomizer -> CustomizeRemoteFieldName

type RemoteSchemaRelationships = RemoteSchemaRelationshipsG (RemoteFieldInfo Name)

data RemoteSchemaCtxG remoteFieldInfo

The resolved information of a remote schema. It is parameterized by remoteFieldInfo so as to work on an arbitrary 'remote relationship' TODO: Get rid of this G suffix using pattern synonyms or qualified usage

Constructors

RemoteSchemaCtx

Fields

_rscName :: RemoteSchemaName

_rscIntroOriginal :: IntrospectionResult

Original remote schema without customizations

_rscInfo :: RemoteSchemaInfo

_rscRawIntrospectionResult :: ByteString

The raw response from the introspection query against the remote server. We store this so we can efficiently service introspect_remote_schema.

_rscPermissions :: HashMap RoleName IntrospectionResult

_rscRemoteRelationships :: RemoteSchemaRelationshipsG remoteFieldInfo

type PartiallyResolvedRemoteSchemaCtx = RemoteSchemaCtxG (PartiallyResolvedRemoteRelationship RemoteRelationshipDefinition)

type RemoteSchemaCtx = RemoteSchemaCtxG (RemoteFieldInfo Name)

type PartiallyResolvedRemoteSchemaMap = HashMap RemoteSchemaName PartiallyResolvedRemoteSchemaCtx

type RemoteSchemaMap = HashMap RemoteSchemaName RemoteSchemaCtx

type DepMap = HashMap SchemaObjId (HashSet SchemaDependency)

type WithDeps a = (a, Seq SchemaDependency)

class Monad m => TableCoreInfoRM (b :: BackendType) (m :: Type -> Type) where

A more limited version of CacheRM that is used when building the schema cache, since the entire schema cache has not been built yet.

Methods

lookupTableCoreInfo :: TableName b -> m (Maybe (TableCoreInfo b))

newtype TableCoreCacheRT (b :: BackendType) (m :: Type -> Type) a

Constructors

TableCoreCacheRT

Fields

runTableCoreCacheRT :: TableCoreCache b -> m a

class TableCoreInfoRM b m => TableInfoRM (b :: BackendType) (m :: Type -> Type) where

All our RQL DML queries operate over a single source. This typeclass facilitates that.

Methods

lookupTableInfo :: TableName b -> m (Maybe (TableInfo b))

newtype TableCacheRT (b :: BackendType) (m :: Type -> Type) a

Constructors

TableCacheRT

Fields

runTableCacheRT :: TableCache b -> m a

class Monad m => CacheRM (m :: Type -> Type) where

Methods

askSchemaCache :: m SchemaCache

type FieldInfoMap = HashMap FieldName

data FieldInfo (b :: BackendType)

Constructors

FIColumn (StructuredColumnInfo b)

FIRelationship (RelInfo b)

FIComputedField (ComputedFieldInfo b)

FIRemoteRelationship (RemoteFieldInfo (DBJoinField b))

_FIColumn :: forall (b :: BackendType) p f. (Choice p, Applicative f) => p (StructuredColumnInfo b) (f (StructuredColumnInfo b)) -> p (FieldInfo b) (f (FieldInfo b))

_FIRelationship :: forall (b :: BackendType) p f. (Choice p, Applicative f) => p (RelInfo b) (f (RelInfo b)) -> p (FieldInfo b) (f (FieldInfo b))

_FIComputedField :: forall (b :: BackendType) p f. (Choice p, Applicative f) => p (ComputedFieldInfo b) (f (ComputedFieldInfo b)) -> p (FieldInfo b) (f (FieldInfo b))

fieldInfoName :: forall (b :: BackendType). Backend b => FieldInfo b -> FieldName

fieldInfoGraphQLNames :: forall (b :: BackendType). FieldInfo b -> [Name]

Returns all the field names created for the given field. Columns, object relationships, and computed fields only ever produce a single field, but array relationships also contain an _aggregate field.

getCols :: forall (backend :: BackendType). FieldInfoMap (FieldInfo backend) -> [StructuredColumnInfo backend]

getRels :: forall (backend :: BackendType). FieldInfoMap (FieldInfo backend) -> [RelInfo backend]

getComputedFieldInfos :: forall (backend :: BackendType). FieldInfoMap (FieldInfo backend) -> [ComputedFieldInfo backend]

type RolePermInfoMap (b :: BackendType) = HashMap RoleName (RolePermInfo b)

data InsPermInfo (b :: BackendType)

Constructors

InsPermInfo

Fields

ipiCols :: HashSet (Column b)

ipiCheck :: AnnBoolExpPartialSQL b

ipiSet :: PreSetColsPartial b

ipiBackendOnly :: Bool

ipiRequiredHeaders :: HashSet Text

ipiValidateInput :: Maybe (ValidateInput ResolvedWebhook)

data UpdPermInfo (b :: BackendType)

Constructors

UpdPermInfo

Fields

upiCols :: HashSet (Column b)

upiTable :: TableName b

upiFilter :: AnnBoolExpPartialSQL b

upiCheck :: Maybe (AnnBoolExpPartialSQL b)

upiSet :: PreSetColsPartial b

upiBackendOnly :: Bool

upiRequiredHeaders :: HashSet Text

upiValidateInput :: Maybe (ValidateInput ResolvedWebhook)

data DelPermInfo (b :: BackendType)

Constructors

DelPermInfo

Fields

dpiTable :: TableName b

dpiFilter :: AnnBoolExpPartialSQL b

dpiBackendOnly :: !Bool

dpiRequiredHeaders :: HashSet Text

dpiValidateInput :: Maybe (ValidateInput ResolvedWebhook)

type PreSetColsPartial (b :: BackendType) = HashMap (Column b) (PartialSQLExp b)

data EventTriggerInfo (b :: BackendType)

Constructors

EventTriggerInfo

Fields

etiName :: TriggerName

etiOpsDef :: TriggerOpsDef b

etiRetryConf :: RetryConf

etiWebhookInfo :: WebhookConfInfo

The HTTP(s) URL which will be called with the event payload on configured operation. Must be a POST handler. This URL can be entered manually or can be picked up from an environment variable (the environment variable needs to be set before using it for this configuration).

etiHeaders :: [EventHeaderInfo]

Custom headers can be added to an event trigger. Each webhook request will have these headers added.

etiRequestTransform :: Maybe RequestTransform

etiResponseTransform :: Maybe MetadataResponseTransform

etiCleanupConfig :: Maybe AutoTriggerLogCleanupConfig

etiTriggerOnReplication :: TriggerOnReplication

type EventTriggerInfoMap (b :: BackendType) = HashMap TriggerName (EventTriggerInfo b)

data TableObjId (b :: BackendType)

Constructors

TOCol (Column b)

TORel RelName

TOComputedField ComputedFieldName

TORemoteRel RelName

TOForeignKey (ConstraintName b)

TOPerm RoleName PermType

TOTrigger TriggerName

data SchemaObjId

Constructors

SOSource SourceName

SOSourceObj SourceName (AnyBackend SourceObjId)

SORemoteSchema RemoteSchemaName

SORemoteSchemaPermission RemoteSchemaName RoleName

SORemoteSchemaRemoteRelationship RemoteSchemaName Name RelName
A remote relationship on a remote schema type, identified by 1. remote schema name 2. remote schema type on which the relationship is defined 3. name of the relationship

SORole RoleName

reportSchemaObj :: SchemaObjId -> Text

reportSchemaObjs :: [SchemaObjId] -> Text

data DependencyReason

Constructors

DRTable

DRColumn

DRRemoteTable

DRLeftColumn

DRRightColumn

DRUsingColumn

DRFkey

DRRemoteFkey

DRUntyped

DROnType

DRSessionVariable

DRPayload

DRParent

DRRemoteSchema

DRRemoteRelationship

DRParentRole

DRLogicalModel

DRReferencedLogicalModel

data SchemaDependency

Constructors

SchemaDependency

Fields

sdObjId :: SchemaObjId

sdReason :: DependencyReason

mkParentDep :: forall (b :: BackendType). Backend b => SourceName -> TableName b -> SchemaDependency

mkLogicalModelParentDep :: forall (b :: BackendType). Backend b => SourceName -> LogicalModelLocation -> SchemaDependency

When we depend on anything to do with logical models, we also declare that we depend on the logical model as a whole. This is the "parent" dependency in the dependency tree for a given logical model.

mkColDep :: forall (b :: BackendType). Backend b => DependencyReason -> SourceName -> TableName b -> Column b -> SchemaDependency

mkLogicalModelColDep :: forall (b :: BackendType). Backend b => DependencyReason -> SourceName -> LogicalModelLocation -> Column b -> SchemaDependency

Declare a dependency on a particular column of a logical model

mkComputedFieldDep :: forall (b :: BackendType). Backend b => DependencyReason -> SourceName -> TableName b -> ComputedFieldName -> SchemaDependency

getDependentObjs :: SchemaCache -> SchemaObjId -> [SchemaObjId]

getDependentObjsWith :: (DependencyReason -> Bool) -> SchemaCache -> SchemaObjId -> [SchemaObjId]

getRemoteDependencies :: SchemaCache -> SourceName -> [SchemaObjId]

Compute all remote dependencies on a source.

Given a source name, this function computes all of its dependencies, direct or indirect, and returns all of the dependencies that are not "local" to the source, i.e. that belong to another source or to a remote schema, here dubbed "remote dependencies".

This functions returns a SchemaObjId for each such dependency, but makes no attempt at extracting the underlying SourceObjId (if any), for two reasons: 1. a SourceObjId no longer contains the source name, which most callers need to identify where the corresponding dependency is 2. this would prevent us from returning remote schema dependencies, which by definition do not have a corresponding SourceObjId

data FunctionVolatility

https://www.postgresql.org/docs/current/xfunc-volatility.html

Constructors

FTVOLATILE

FTIMMUTABLE

FTSTABLE

newtype FunctionArgName

Constructors

FunctionArgName

Fields

getFuncArgNameTxt :: Text

data FunctionInfo (b :: BackendType)

Tracked SQL function metadata. See buildFunctionInfo.

Constructors

FunctionInfo

Fields

_fiSQLName :: FunctionName b

_fiGQLName :: Name

_fiGQLArgsName :: Name

_fiGQLAggregateName :: Name

_fiSystemDefined :: SystemDefined

_fiVolatility :: FunctionVolatility

_fiExposedAs :: FunctionExposedAs

In which part of the schema should this function be exposed?

See mkFunctionInfo and _fcExposedAs.

_fiInputArgs :: Seq (FunctionInputArgument b)

_fiReturnType :: TableName b

NOTE: when a table is created, a new composite type of the same name is automatically created; so strictly speaking this field means "the function returns the composite type corresponding to this table".

_fiDescription :: Maybe Text

this field represents the description of the function as present on the database

_fiPermissions :: FunctionPermissionsMap

Roles to which the function is accessible

_fiJsonAggSelect :: JsonAggSelect

_fiComment :: Maybe Text

type FunctionCache (b :: BackendType) = HashMap (FunctionName b) (FunctionInfo b)

data CronTriggerInfo

Constructors

CronTriggerInfo

Fields

ctiName :: TriggerName

ctiSchedule :: CronSchedule

ctiPayload :: Maybe Value

ctiRetryConf :: STRetryConf

ctiWebhookInfo :: EnvRecord ResolvedWebhook

ctiHeaders :: [EventHeaderInfo]

ctiComment :: Maybe Text

ctiRequestTransform :: Maybe RequestTransform

ctiResponseTransform :: Maybe MetadataResponseTransform

newtype MetadataResourceVersion

Constructors

MetadataResourceVersion

Fields

getMetadataResourceVersion :: Int64

showMetadataResourceVersion :: MetadataResourceVersion -> Text

initialResourceVersion :: MetadataResourceVersion

data MetadataWithResourceVersion

Constructors

MetadataWithResourceVersion

Fields

_mwrvMetadata :: Metadata

_mwrvResourceVersion :: MetadataResourceVersion

getLogicalModelBoolExpDeps :: forall (b :: BackendType). GetAggregationPredicatesDeps b => SourceName -> LogicalModelLocation -> AnnBoolExpPartialSQL b -> [SchemaDependency]

What schema dependencies does a given row permission for a logical model have? This will almost certainly involve some number of dependencies on logical models, but may also involve dependencies on tables. Although we can't relate tables and logical models yet, we can still declare permissions like, "you can only see this logical model if your user ID exists in this table".

getBoolExpDeps :: forall (b :: BackendType). GetAggregationPredicatesDeps b => SourceName -> TableName b -> AnnBoolExpPartialSQL b -> [SchemaDependency]

Discover the schema dependencies of an AnnBoolExpPartialSQL.

data InlinedAllowlist

InlinedAllowlist is the data type with which the allowlist is represented in the schema cache, it contains a global and a per role allowlist and when allowlist is enabled in the graphql-engine, the incoming query for a non-admin role should either be in the global allowlist or in the given role's role based allowlist.

Essentially, it's a memoization of allowlistAllowsQuery implemented in terms of MetadataAllowlist.

newtype BoolExpM (b :: BackendType) a

The monad for doing schema dependency discovery for boolean expressions. maintains the table context of the expressions being translated.

Constructors

BoolExpM

Fields

unBoolExpM :: Reader (BoolExpCtx b) a

data BoolExpCtx (b :: BackendType)

The table type context of schema dependency discovery. Boolean expressions may refer to a so-called 'root table' (identified by a $-sign in the expression input syntax) or the current table.

Constructors

BoolExpCtx

Fields

source :: SourceName

currTable :: TableName b

Reference to the current table type.

rootTable :: TableName b

Reference to the root table type.

getOpExpDeps :: forall (b :: BackendType). Backend b => [OpExpG b (PartialSQLExp b)] -> BoolExpM b [SchemaDependency]

newtype BackendInfoWrapper (b :: BackendType)

Constructors

BackendInfoWrapper

Fields

unBackendInfoWrapper :: BackendInfo b

type BackendCache = BackendMap BackendInfoWrapper

getBackendInfo :: forall (b :: BackendType) m. (CacheRM m, HasTag b) => m (Maybe (BackendInfo b))