Hasura.RQL.Types.SchemaCache.Build

Description

Types and functions used in the process of building the schema cache from metadata information stored in the hdb_catalog schema in Postgres.

Documentation

data CollectedInfo

Constructors

CIInconsistency !InconsistentMetadata

CIDependency

Fields

!MetadataObject

for error reporting on missing dependencies

!SchemaObjId

!SchemaDependency

_CIDependency :: Prism' CollectedInfo (MetadataObject, SchemaObjId, SchemaDependency)

_CIInconsistency :: Prism' CollectedInfo InconsistentMetadata

class AsInconsistentMetadata s where

Methods

_InconsistentMetadata :: Prism' s InconsistentMetadata

partitionCollectedInfo :: Seq CollectedInfo -> ([InconsistentMetadata], [(MetadataObject, SchemaObjId, SchemaDependency)])

recordInconsistency :: (ArrowWriter (Seq w) arr, AsInconsistentMetadata w) => ((Maybe Value, MetadataObject), Text) `arr` ()

recordInconsistencies :: (ArrowWriter (Seq w) arr, AsInconsistentMetadata w) => ([MetadataObject], Text) `arr` ()

recordInconsistencies' :: (ArrowWriter (Seq w) arr, AsInconsistentMetadata w) => ([(Maybe Value, MetadataObject)], Text) `arr` ()

recordDependencies :: ArrowWriter (Seq CollectedInfo) arr => (MetadataObject, SchemaObjId, [SchemaDependency]) `arr` ()

withRecordInconsistency :: (ArrowChoice arr, ArrowWriter (Seq w) arr, AsInconsistentMetadata w) => ErrorA QErr arr (e, s) a -> arr (e, (MetadataObject, s)) (Maybe a)

class CacheRM m => CacheRWM m where

Methods

buildSchemaCacheWithOptions :: BuildReason -> CacheInvalidations -> Metadata -> m ()

setMetadataResourceVersionInSchemaCache :: MetadataResourceVersion -> m ()

data BuildReason

Constructors

CatalogUpdate (Maybe (HashSet SourceName))
The build was triggered by an update this instance made to the catalog (in the currently-active transaction), so information in Postgres that needs to be kept in sync with the catalog (i.e. table event triggers in hdb_catalog schema) should be updated.

CatalogSync
The build was triggered by a notification that some other currently-running Hasura instance updated the catalog. Since that instance already updated table event triggers in hdb_catalog, this build should be read-only.

data CacheInvalidations

Constructors

CacheInvalidations

Fields

ciMetadata :: !Bool

Force reloading of all database information, including information not technically stored in metadata (currently just enum values). Set by the reload_metadata API.

ciRemoteSchemas :: !(HashSet RemoteSchemaName)

Force refetching of the given remote schemas, even if their definition has not changed. Set by the reload_remote_schema API.

ciSources :: !(HashSet SourceName)

Force re-establishing connections of the given data sources, even if their configuration has not changed. Set by the pg_reload_source API.

class Monad m => MetadataM m where

A simple monad class which enables fetching and setting @Metadata in the state.

Methods

getMetadata :: m Metadata

putMetadata :: Metadata -> m ()

newtype MetadataT m a

Constructors

MetadataT

Fields

unMetadataT :: StateT Metadata m a

runMetadataT :: Metadata -> MetadataT m a -> m (a, Metadata)

buildSchemaCacheWithInvalidations :: (MetadataM m, CacheRWM m) => CacheInvalidations -> MetadataModifier -> m ()

buildSchemaCache :: (MetadataM m, CacheRWM m) => MetadataModifier -> m ()

buildSchemaCacheFor :: (QErrM m, CacheRWM m, MetadataM m) => MetadataObjId -> MetadataModifier -> m ()

Rebuilds the schema cache after modifying metadata. If an object with the given object id became newly inconsistent, raises an error about it specifically. Otherwise, raises a generic metadata inconsistency error.

buildSchemaCacheStrict :: (QErrM m, CacheRWM m, MetadataM m) => m ()

Like buildSchemaCache, but fails if there is any inconsistent metadata.

withNewInconsistentObjsCheck :: (QErrM m, CacheRM m) => m a -> m a

Executes the given action, and if any new InconsistentMetadatas are added to the schema cache as a result of its execution, raises an error.