Hasura.RQL.Types.SchemaCache.Build

Description

Types and functions used in the process of building the schema cache from metadata information stored in the hdb_catalog schema in Postgres.

Inconsistencies

recordInconsistencies :: (ArrowWriter (Seq CollectItem) arr, Functor f, Foldable f) => ((Maybe Value, f MetadataObject), Text) `arr` ()

recordInconsistencyM :: MonadWriter (Seq CollectItem) m => Maybe Value -> MetadataObject -> Text -> m ()

recordInconsistenciesM :: MonadWriter (Seq CollectItem) m => [MetadataObject] -> Text -> m ()

Dependencies

data MetadataDependency

Constructors

MetadataDependency

Fields

MetadataObject

for error reporting on missing dependencies

SchemaObjId

SchemaDependency

recordDependencies :: ArrowWriter (Seq CollectItem) arr => (MetadataObject, SchemaObjId, Seq SchemaDependency) `arr` ()

recordDependenciesM :: MonadWriter (Seq CollectItem) m => MetadataObject -> SchemaObjId -> Seq SchemaDependency -> m ()

Helpers

withRecordInconsistencyM :: MonadWriter (Seq CollectItem) m => MetadataObject -> ExceptT QErr m a -> m (Maybe a)

Monadic version of withRecordInconsistency

withRecordInconsistency :: (ArrowChoice arr, ArrowWriter (Seq CollectItem) arr) => ErrorA QErr arr (e, s) a -> arr (e, (MetadataObject, s)) (Maybe a)

Record any errors resulting from a computation as inconsistencies

withRecordInconsistencies :: (ArrowChoice arr, ArrowWriter (Seq CollectItem) arr) => ErrorA QErr arr (e, s) a -> arr (e, ([MetadataObject], s)) (Maybe a)

class CacheRM m => CacheRWM m where

Methods

tryBuildSchemaCacheWithOptions :: BuildReason -> CacheInvalidations -> Metadata -> Maybe MetadataResourceVersion -> ValidateNewSchemaCache a -> m a

setMetadataResourceVersionInSchemaCache :: MetadataResourceVersion -> m ()

buildSchemaCacheWithOptions :: CacheRWM m => BuildReason -> CacheInvalidations -> Metadata -> Maybe MetadataResourceVersion -> m ()

data BuildReason

Constructors

CatalogUpdate (Maybe (HashSet SourceName))
The build was triggered by an update this instance made to the catalog (in the currently-active transaction), so information in Postgres that needs to be kept in sync with the catalog (i.e. table event triggers in hdb_catalog schema) should be updated.

CatalogSync
The build was triggered by a notification that some other currently-running Hasura instance updated the catalog. Since that instance already updated table event triggers in hdb_catalog, this build should be read-only.

data CacheInvalidations

Constructors

CacheInvalidations

Fields

ciMetadata :: Bool

Force reloading of all database information, including information not technically stored in metadata (currently just enum values). Set by the reload_metadata API.

ciRemoteSchemas :: HashSet RemoteSchemaName

Force refetching of the given remote schemas, even if their definition has not changed. Set by the reload_remote_schema API.

ciSources :: HashSet SourceName

Force re-establishing connections of the given data sources, even if their configuration has not changed. Set by the pg_reload_source API.

ciDataConnectors :: HashSet DataConnectorName

Force re-fetching of DataConnectorInfo from the named data connectors.

type ValidateNewSchemaCache a = SchemaCache -> SchemaCache -> (ValidateNewSchemaCacheResult, a)

Function that validates the new schema cache (usually involves checking for any metadata inconsistencies) and can decide whether or not to keep or discard the new schema cache (ValidateNewSchemaCacheResult). It can also return some arbitrary extra information that will be returned from tryBuildSchemaCacheWithOptions.

First parameter is the old schema cache, the second is the new schema cache.

data ValidateNewSchemaCacheResult

Constructors

KeepNewSchemaCache

DiscardNewSchemaCache

newtype MetadataT m a

Constructors

MetadataT

Fields

unMetadataT :: StateT Metadata m a

runMetadataT :: Metadata -> MetadataDefaults -> MetadataT m a -> m (a, Metadata)

runMetadataT puts a stateful metadata in scope. MetadataDefaults is provided so that it can be considered from the --metadataDefaults arguments.

buildSchemaCacheWithInvalidations :: (MetadataM m, CacheRWM m) => CacheInvalidations -> MetadataModifier -> m ()

buildSchemaCache :: (MetadataM m, CacheRWM m) => MetadataModifier -> m ()

tryBuildSchemaCache :: (CacheRWM m, MetadataM m) => MetadataModifier -> m (HashMap MetadataObjId (NonEmpty InconsistentMetadata))

Rebuilds the schema cache after modifying metadata and returns any _new_ metadata inconsistencies. If there are any new inconsistencies, the changes to the metadata and the schema cache are abandoned.

tryBuildSchemaCacheWithModifiers :: (CacheRWM m, MetadataM m) => [Metadata -> m Metadata] -> m (HashMap MetadataObjId (NonEmpty InconsistentMetadata))

Rebuilds the schema cache after modifying metadata sequentially and returns any _new_ metadata inconsistencies. If there are any new inconsistencies, the changes to the metadata and the schema cache are abandoned. If the metadata modifiers run into validation issues (e.g. a native query is already tracked in the metadata), we throw these errors back without changing the metadata and schema cache.

tryBuildSchemaCacheAndWarnOnFailingObjects

Arguments

:: forall m a. (CacheRWM m, MonadWarnings m, QErrM m, MetadataM m)

=> (a -> m MetadataModifier)
Function makes a metadata modifier for a metadata object

-> WarningCode
Warning code to use for failed metadata objects

-> HashMap MetadataObjId a
Map of metadata objects to apply to metadata using the mkMetadataModifier function

-> m (HashMap MetadataObjId a)
Successfully applied metadata objects

Tries to modify the metadata for all the specified metadata objects. If the modification fails, any objects that directly caused a new metadata inconsistency are removed and the modification is attempted again without those failing objects. The failing objects are raised as warnings in MonadWarnings and the successful objects are returned. If there are metadata inconsistencies that are not directly related to the specified metadata objects, an error is thrown.

buildSchemaCacheFor :: (QErrM m, CacheRWM m, MetadataM m) => MetadataObjId -> MetadataModifier -> m ()

Rebuilds the schema cache after modifying metadata. If an object with the given object id became newly inconsistent, raises an error about it specifically. Otherwise, raises a generic metadata inconsistency error.

throwOnInconsistencies :: (QErrM m, CacheRWM m) => m ()

Requests the schema cache, and fails if there is any inconsistent metadata.

withNewInconsistentObjsCheck :: (QErrM m, CacheRM m) => m a -> m a

Executes the given action, and if any new InconsistentMetadatas are added to the schema cache as a result of its execution, raises an error.

getInconsistentQueryCollections :: SchemaIntrospection -> QueryCollections -> ((CollectionName, ListedQuery) -> MetadataObject) -> EndpointTrie GQLQueryWithText -> [NormalizedQuery] -> [InconsistentMetadata]

getInconsistentQueryCollections is a helper function that runs the static analysis over the saved queries and reports any inconsistenties with the current schema.

data StoredIntrospection

Constructors

StoredIntrospection

Fields

siBackendIntrospection :: HashMap SourceName EncJSON

siRemotes :: HashMap RemoteSchemaName EncJSON

data StoredIntrospectionItem

Represents remote schema or source introspection data to be persisted in a storage (database).

Constructors

SourceIntrospectionItem SourceName EncJSON

RemoteSchemaIntrospectionItem RemoteSchemaName EncJSON

data CollectItem

Items to be collected while building schema cache See @buildSchemaCacheRule for more details.

Constructors

CollectInconsistentMetadata InconsistentMetadata

CollectMetadataDependency MetadataDependency

CollectStoredIntrospection StoredIntrospectionItem