Hasura.RQL.Types.Source

Documentation

data SourceInfo b

Constructors

SourceInfo

Fields

_siName :: !SourceName

_siTables :: !(TableCache b)

_siFunctions :: !(FunctionCache b)

_siConfiguration :: !(SourceConfig b)

_siQueryTagsConfig :: !(Maybe QueryTagsConfig)

_siCustomization :: !SourceCustomization

siTables :: forall b. Lens' (SourceInfo b) (TableCache b)

siQueryTagsConfig :: forall b. Lens' (SourceInfo b) (Maybe QueryTagsConfig)

siName :: forall b. Lens' (SourceInfo b) SourceName

siFunctions :: forall b. Lens' (SourceInfo b) (FunctionCache b)

siCustomization :: forall b. Lens' (SourceInfo b) SourceCustomization

siConfiguration :: forall b. Lens' (SourceInfo b) (SourceConfig b)

type BackendSourceInfo = AnyBackend SourceInfo

type SourceCache = HashMap SourceName BackendSourceInfo

unsafeSourceInfo :: forall b. HasTag b => BackendSourceInfo -> Maybe (SourceInfo b)

unsafeSourceName :: BackendSourceInfo -> SourceName

unsafeSourceTables :: forall b. HasTag b => BackendSourceInfo -> Maybe (TableCache b)

unsafeSourceFunctions :: forall b. HasTag b => BackendSourceInfo -> Maybe (FunctionCache b)

unsafeSourceConfiguration :: forall b. HasTag b => BackendSourceInfo -> Maybe (SourceConfig b)

getTableRoles :: BackendSourceInfo -> [RoleName]

data ResolvedSource b

Contains Postgres connection configuration and essential metadata from the database to build schema cache for tables and function.

Constructors

ResolvedSource

Fields

_rsConfig :: !(SourceConfig b)

_rsCustomization :: !SourceTypeCustomization

_rsTables :: !(DBTablesMetadata b)

_rsFunctions :: !(DBFunctionsMetadata b)

_rsPgScalars :: !(HashSet (ScalarType b))

type SourceTables b = HashMap SourceName (TableCache b)

type SourceResolver b = SourceName -> SourceConnConfiguration b -> IO (Either QErr (SourceConfig b))

class Monad m => MonadResolveSource m where

Methods

getPGSourceResolver :: m (SourceResolver ('Postgres 'Vanilla))

getMSSQLSourceResolver :: m (SourceResolver 'MSSQL)

data MaintenanceModeVersion

Constructors

PreviousMMVersion
should correspond to the source catalog version from which the user is migrating from

CurrentMMVersion
should correspond to the latest source catalog version