Hasura.RQL.Types.SourceConfiguration

Documentation

class (Representable (SourceConnConfiguration b), HasCodec (SourceConnConfiguration b), FromJSON (SourceConnConfiguration b), Show (SourceConfig b), ToJSON (SourceConfig b), ToJSON (SourceConnConfiguration b), Eq (SourceConfig b), Has (ScalarTypeParsingContext b) (SourceConfig b), Eq (ScalarTypeParsingContext b), Ord (ScalarTypeParsingContext b), HasTag b) => HasSourceConfiguration (b :: BackendType) where

Associated Types

type ScalarTypeParsingContext (b :: BackendType)

type ScalarTypeParsingContext (b :: BackendType) = ()

type SourceConnConfiguration (b :: BackendType)

User facing connection configuration for a database.

type SourceConfig (b :: BackendType)

Internal connection configuration for a database - connection string, connection pool etc

Methods

sourceConfigNumReadReplicas :: SourceConfig b -> Int

The number of read replicas specified in the source configuration

sourceConfigConnectonTemplate :: SourceConfig b -> Maybe ValueExt

The connection template specified in the source configuration, if any

sourceSupportsColumnRedaction :: SourceConfig b -> Bool

Whether or not the source supports performing column redaction. See note [SQL generation for inherited roles] for more information

sourceConfigBackendSourceKind :: SourceConfig b -> BackendSourceKind b

type Representable a = (Show a, Eq a, Hashable a, NFData a)