Hasura.RQL.Types.SourceConfiguration

Documentation

type Representable a = (Show a, Eq a, Hashable a, NFData a)

class (Representable (SourceConnConfiguration b), HasCodec (SourceConnConfiguration b), FromJSON (SourceConnConfiguration b), ToJSON (SourceConfig b), ToJSON (SourceConnConfiguration b), Eq (SourceConfig b), Has (ScalarTypeParsingContext b) (SourceConfig b), Eq (ScalarTypeParsingContext b), Ord (ScalarTypeParsingContext b), HasTag b) => HasSourceConfiguration (b :: BackendType) where

Associated Types

type ScalarTypeParsingContext b

type ScalarTypeParsingContext b = ()

type SourceConnConfiguration b :: Type

User facing connection configuration for a database.

type SourceConfig b :: Type

Internal connection configuration for a database - connection string, connection pool etc

Methods

sourceConfigNumReadReplicas :: SourceConfig b -> Int

The number of read replicas specified in the source configuration

sourceConfigConnectonTemplateEnabled :: SourceConfig b -> Bool

Whether the source configuration specifies the use of a connection template

sourceConfigBackendSourceKind :: SourceConfig b -> BackendSourceKind b