Hasura.RQL.Types.SourceCustomization

Documentation

data RootFieldsCustomization

Constructors

RootFieldsCustomization

Fields

_rootfcNamespace :: !(Maybe Name)

_rootfcPrefix :: !(Maybe Name)

_rootfcSuffix :: !(Maybe Name)

emptyRootFieldsCustomization :: RootFieldsCustomization

data SourceTypeCustomization

Constructors

SourceTypeCustomization

Fields

_stcPrefix :: !(Maybe Name)

_stcSuffix :: !(Maybe Name)

emptySourceTypeCustomization :: SourceTypeCustomization

data NamingCase

Represents the different possible type cases for fields and types, i.e. HasuraCase and GraphqlCase (CamelCase fields and PascalCase types).

Constructors

HasuraCase

GraphqlCase

parseNamingConventionFromText :: Text -> Either String NamingCase

parses naming convention from Text

mkCustomizedTypename :: Maybe SourceTypeCustomization -> NamingCase -> MkTypename

mkCustomizedFieldName :: Maybe RootFieldsCustomization -> MkRootFieldName

applyTypeCust :: Maybe SourceTypeCustomization -> Name -> Name

apply prefix and suffix to type name according to the source customization

applyFieldCust :: Maybe RootFieldsCustomization -> Name -> Name

apply prefix and suffix to field name according to the source customization

applyTypeNameCaseCust :: NamingCase -> Name -> Name

apply naming convention to type name

applyFieldNameCaseCust :: NamingCase -> Name -> Name

apply naming convention to field name

getTextFieldName :: NamingCase -> GQLNameIdentifier -> Text

returns field name according to the naming conventions as Text

applyTypeNameCaseIdentifier :: NamingCase -> GQLNameIdentifier -> Name

applies naming convention and returns field name

Note: This can't possibly fail as GQLNameIdentifier contains already validated identifiers

getTextTypeName :: NamingCase -> GQLNameIdentifier -> Text

returns type name according to the naming conventions as Text

applyFieldNameCaseIdentifier :: NamingCase -> GQLNameIdentifier -> Name

applies naming convention and returns type name

Note: This can't possibly fail as GQLNameIdentifier contains already validated identifiers

applyEnumValueCase :: NamingCase -> Name -> Name

applyPrefixSuffix :: Maybe Name -> Maybe Name -> Name -> Name

appendprepend the suffixprefix in the graphql name

data SourceCustomization

Constructors

SourceCustomization

Fields

_scRootFields :: !(Maybe RootFieldsCustomization)

_scTypeNames :: !(Maybe SourceTypeCustomization)

_scNamingConvention :: !(Maybe NamingCase)

emptySourceCustomization :: SourceCustomization

getRootFieldsCustomization :: SourceCustomization -> RootFieldsCustomization

getSourceTypeCustomization :: SourceCustomization -> SourceTypeCustomization

getNamingConvention :: SourceCustomization -> Maybe NamingCase -> NamingCase

newtype MkRootFieldName

Function to apply root field name customizations.

Constructors

MkRootFieldName

Fields

runMkRootFieldName :: Name -> Name

withRootFieldNameCustomization :: forall m r a. (MonadReader r m, Has MkRootFieldName r) => MkRootFieldName -> m a -> m a

Inject a new root field name customization function into the environment. This can be used by schema-building code (with MonadBuildSchema constraint) to ensure the correct root field name customizations are applied.

mkRootFieldName :: (MonadReader r m, Has MkRootFieldName r) => Name -> m Name

Apply the root field name customization function from the current environment.

withSourceCustomization :: forall m r a. (MonadReader r m, Has MkTypename r, Has MkRootFieldName r, Has NamingCase r, MonadError QErr m) => SourceCustomization -> SupportedNamingCase -> Maybe NamingCase -> m a -> m a

Inject NamingCase, typename and root field name customizations from SourceCustomization into the environment.

withNamingCaseCustomization :: forall m r a. (MonadReader r m, Has NamingCase r) => NamingCase -> m a -> m a

newtype CustomizeRemoteFieldName

Constructors

CustomizeRemoteFieldName

Fields

runCustomizeRemoteFieldName :: Name -> Name -> Name

withRemoteFieldNameCustomization :: forall m r a. (MonadReader r m, Has CustomizeRemoteFieldName r) => CustomizeRemoteFieldName -> m a -> m a

mkSelectField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectAggregateField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkInsertField :: GQLNameIdentifier -> GQLNameIdentifier

mkInsertOneField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkDeleteField :: GQLNameIdentifier -> GQLNameIdentifier

mkDeleteByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkRelayConnectionField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectStreamField :: GQLNameIdentifier -> GQLNameIdentifier