Hasura.RQL.Types.SourceCustomization

Documentation

data RootFieldsCustomization

Constructors

RootFieldsCustomization

Fields

_rootfcNamespace :: !(Maybe Name)

_rootfcPrefix :: !(Maybe Name)

_rootfcSuffix :: !(Maybe Name)

emptyRootFieldsCustomization :: RootFieldsCustomization

data SourceTypeCustomization

Constructors

SourceTypeCustomization

Fields

_stcPrefix :: !(Maybe Name)

_stcSuffix :: !(Maybe Name)

emptySourceTypeCustomization :: SourceTypeCustomization

mkCustomizedTypename :: Maybe SourceTypeCustomization -> MkTypename

mkCustomizedFieldName :: Maybe RootFieldsCustomization -> MkRootFieldName

applyPrefixSuffix :: Maybe Name -> Maybe Name -> Name -> Name

data SourceCustomization

Constructors

SourceCustomization

Fields

_scRootFields :: !(Maybe RootFieldsCustomization)

_scTypeNames :: !(Maybe SourceTypeCustomization)

emptySourceCustomization :: SourceCustomization

getRootFieldsCustomization :: SourceCustomization -> RootFieldsCustomization

getSourceTypeCustomization :: SourceCustomization -> SourceTypeCustomization

newtype MkRootFieldName

Function to apply root field name customizations.

Constructors

MkRootFieldName

Fields

runMkRootFieldName :: Name -> Name

withRootFieldNameCustomization :: forall m r a. (MonadReader r m, Has MkRootFieldName r) => MkRootFieldName -> m a -> m a

Inject a new root field name customization function into the environment. This can be used by schema-building code (with MonadBuildSchema constraint) to ensure the correct root field name customizations are applied.

mkRootFieldName :: (MonadReader r m, Has MkRootFieldName r) => Name -> m Name

Apply the root field name customization function from the current environment.

withSourceCustomization :: forall m r a. (MonadReader r m, Has MkTypename r, Has MkRootFieldName r) => SourceCustomization -> m a -> m a

Inject typename and root field name customizations from SourceCustomization into the environment.

newtype CustomizeRemoteFieldName

Constructors

CustomizeRemoteFieldName

Fields

runCustomizeRemoteFieldName :: Name -> Name -> Name

withRemoteFieldNameCustomization :: forall m r a. (MonadReader r m, Has CustomizeRemoteFieldName r) => CustomizeRemoteFieldName -> m a -> m a