Hasura.RQL.Types.SourceCustomization

Documentation

data RootFieldsCustomization

Constructors

RootFieldsCustomization

Fields

_rootfcNamespace :: Maybe Name

_rootfcPrefix :: Maybe Name

_rootfcSuffix :: Maybe Name

data SourceTypeCustomization

emptySourceTypeCustomization :: SourceTypeCustomization

mkCustomizedTypename :: Maybe SourceTypeCustomization -> NamingCase -> MkTypename

applyTypeNameCaseCust :: NamingCase -> Name -> Name

apply naming convention to type name

applyFieldNameCaseCust :: NamingCase -> Name -> Name

apply naming convention to field name

getTextFieldName :: NamingCase -> GQLNameIdentifier -> Text

returns field name according to the naming conventions as Text

applyTypeNameCaseIdentifier :: NamingCase -> GQLNameIdentifier -> Name

applies naming convention and returns field name

Note: This can't possibly fail as GQLNameIdentifier contains already validated identifiers

getTextTypeName :: NamingCase -> GQLNameIdentifier -> Text

returns type name according to the naming conventions as Text

applyFieldNameCaseIdentifier :: NamingCase -> GQLNameIdentifier -> Name

applies naming convention and returns type name

Note: This can't possibly fail as GQLNameIdentifier contains already validated identifiers

applyEnumValueCase :: NamingCase -> Name -> Name

setFieldNameCase :: NamingCase -> TableInfo b -> CustomRootField -> (GQLNameIdentifier -> GQLNameIdentifier) -> GQLNameIdentifier -> Name

Builds field name with proper case. Please note that this is a pure function as all the validation has already been done while preparing GQLNameIdentifier.

data SourceCustomization

Source customization information as it appears in the metadata.

Constructors

SourceCustomization

Fields

_scRootFields :: Maybe RootFieldsCustomization

_scTypeNames :: Maybe SourceTypeCustomization

_scNamingConvention :: Maybe NamingCase

emptySourceCustomization :: SourceCustomization

getSourceTypeCustomization :: SourceCustomization -> SourceTypeCustomization

data ResolvedSourceCustomization

Source customization as it appears in the SchemaCache.

Constructors

ResolvedSourceCustomization

Fields

_rscRootFields :: MkRootFieldName

_rscTypeNames :: MkTypename

_rscNamingConvention :: NamingCase

_rscRootNamespace :: Maybe Name

mkResolvedSourceCustomization :: SourceCustomization -> NamingCase -> ResolvedSourceCustomization

newtype MkRootFieldName

Function to apply root field name customizations.

Constructors

MkRootFieldName

Fields

runMkRootFieldName :: Name -> Name

getNamingCase :: forall m. MonadError QErr m => SourceCustomization -> SupportedNamingCase -> NamingCase -> m NamingCase

mkSelectField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectAggregateField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkInsertField :: GQLNameIdentifier -> GQLNameIdentifier

mkInsertOneField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkUpdateManyField :: GQLNameIdentifier -> GQLNameIdentifier

mkDeleteField :: GQLNameIdentifier -> GQLNameIdentifier

mkDeleteByPkField :: GQLNameIdentifier -> GQLNameIdentifier

mkRelayConnectionField :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectStreamField :: GQLNameIdentifier -> GQLNameIdentifier

mkMultiRowUpdateTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkOnConflictTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableConstraintTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateFieldTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkFunctionArgsTypeName :: Name -> GQLNameIdentifier -> GQLNameIdentifier

mkTableBoolExpTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableInsertInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableObjRelInsertInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableArrRelInsertInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableMutationResponseTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableOrderByTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateOrderByTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateOrderByOpTypeName :: GQLNameIdentifier -> GQLNameIdentifier -> GQLNameIdentifier

mkTableAggOperatorTypeName :: GQLNameIdentifier -> GQLNameIdentifier -> GQLNameIdentifier

mkTableSelectColumnTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableUpdateColumnTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkTableOperatorInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier -> GQLNameIdentifier

mkTablePkColumnsInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkEnumTableTypeName :: GQLNameIdentifier -> Maybe Name -> GQLNameIdentifier

mkStreamCursorInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkStreamCursorValueInputTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkSelectColumnPredTypeName :: GQLNameIdentifier -> GQLNameIdentifier -> GQLNameIdentifier

mkTableAggregateBoolExpTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkGroupByTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkGroupByKeyTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkGroupByKeyFieldsTypeName :: GQLNameIdentifier -> GQLNameIdentifier

mkRelationFunctionIdentifier :: GQLNameIdentifier -> Name -> GQLNameIdentifier

mkRelationFunctionArgumentsFieldName :: GQLNameIdentifier -> Name -> GQLNameIdentifier

updateColumnsFieldName :: GQLNameIdentifier

affectedRowsFieldName :: GQLNameIdentifier

pkColumnsFieldName :: GQLNameIdentifier