Hasura.RemoteSchema.Metadata.RemoteRelationship

Documentation

data ToSchemaRelationshipDef

Metadata representation of a relationship to a remote schema.

Constructors

ToSchemaRelationshipDef

Fields

_trrdRemoteSchema :: RemoteSchemaName

Identifier for this mapping.

_trrdLhsFields :: HashSet FieldName

The lhs fields that must be forwarded to the remote schema.

_trrdRemoteField :: RemoteFields

trrdRemoteField :: Lens' ToSchemaRelationshipDef RemoteFields

trrdLhsFields :: Lens' ToSchemaRelationshipDef (HashSet FieldName)

trrdRemoteSchema :: Lens' ToSchemaRelationshipDef RemoteSchemaName

data FieldCall

Associates a field name with the arguments it will be passed in the query.

https://graphql.github.io/graphql-spec/June2018/#sec-Language.Arguments

Constructors

FieldCall

Fields

fcName :: Name

fcArguments :: RemoteArguments

newtype RemoteArguments

Arguments to a remote GraphQL fields, represented as a mapping from name to GraphQL Value. Said values can be variable names, in which case they'll be referring to values we're closed over. TODO: expand on this

Constructors

RemoteArguments

Fields

getRemoteArguments :: HashMap Name (Value Name)

newtype RemoteFields

Targeted field in a remote schema relationship. TODO: explain about subfields and why this is a container

Constructors

RemoteFields

Fields

unRemoteFields :: NonEmpty FieldCall

type SchemaRemoteRelationships r = InsOrdHashMap Name (RemoteSchemaTypeRelationships r)

data RemoteSchemaTypeRelationships r

Constructors

RemoteSchemaTypeRelationships

Fields

_rstrsName :: Name

_rstrsRelationships :: RemoteRelationships r

rstrsName :: forall r f. Functor f => (Name -> f Name) -> RemoteSchemaTypeRelationships r -> f (RemoteSchemaTypeRelationships r)

rstrsRelationships :: forall r1 r2 f. Functor f => (RemoteRelationships r1 -> f (RemoteRelationships r2)) -> RemoteSchemaTypeRelationships r1 -> f (RemoteSchemaTypeRelationships r2)