Hasura.SQL.AnyBackend

Documentation

data AnyBackend (i :: BackendType -> Type)

This type is essentially an unlabeled box for types indexed by BackendType. Given some type defined as 'data T (b :: BackendType) = ...', we can define 'AnyBackend T' without mentioning any BackendType.

This is useful for having generic containers of potentially different types of T. For instance, SourceCache is defined as a HashMap SourceName (AnyBackend SourceInfo).

This type is generated with Template Haskell to have one constructor per backend. This declaration generates the following type:

data AnyBackend (i :: BackendType -> Type) = PostgresValue (i 'Postgres) | MSSQLValue (i 'MSSQL) | ...

Constructors

PostgresVanillaValue (i ('Postgres 'Vanilla))

PostgresCitusValue (i ('Postgres 'Citus))

MSSQLValue (i 'MSSQL)

BigQueryValue (i 'BigQuery)

MySQLValue (i 'MySQL)

DataWrapperValue (i 'DataWrapper)

type AllBackendsSatisfy (c :: BackendType -> Constraint) = (c ('Postgres 'Vanilla), c ('Postgres 'Citus), c 'MSSQL, c 'BigQuery, c 'MySQL, c 'DataWrapper)

Generates a constraint for all backends. This Template Haskell expression generates the following constraint type:

type AllBackendsSatisfy (c :: BackendType -> Constraint) = ( c 'Postgres , c 'MSSQL , ... )

That is, given a class C, this creates the constraint that dictates that all backend must satisfy C.

type SatisfiesForAllBackends (i :: BackendType -> Type) (c :: Type -> Constraint) = (c (i ('Postgres 'Vanilla)), c (i ('Postgres 'Citus)), c (i 'MSSQL), c (i 'BigQuery), c (i 'MySQL), c (i 'DataWrapper))

Generates a constraint for a generic type over all backends. This Template Haskell expression generates the following constraint type:

type SatisfiesForAllBackends (i :: BackendType -> Type) (c :: Type -> Constraint) = ( c (i 'Postgres) , c (i 'MSSQL) , ... )

That is, given a type I and a class C, this creates the constraint that dictates that for all backends b, I b must satisfy C.

liftTag :: BackendType -> AnyBackend BackendTag

How to obtain a tag from a runtime value. This function is generated with Template Haskell for each Backend. The case switch looks like this:

Postgres -> PostgresValue PostgresTag MSSQL -> MSSQLValue MSSQLTag ...

mapBackend :: forall (i :: BackendType -> Type) (j :: BackendType -> Type). AnyBackend i -> (forall b. i b -> j b) -> AnyBackend j

Transforms an `AnyBackend i` into an `AnyBackend j`.

traverseBackend :: forall (c :: BackendType -> Constraint) (i :: BackendType -> Type) (j :: BackendType -> Type) f. (AllBackendsSatisfy c, Applicative f) => AnyBackend i -> (forall b. c b => i b -> f (j b)) -> f (AnyBackend j)

Traverse an `AnyBackend i` into an `f (AnyBackend j)`.

mkAnyBackend :: forall (b :: BackendType) (i :: BackendType -> Type). HasTag b => i b -> AnyBackend i

Creates a new AnyBackend i for a given backend b by wrapping the given i b.

runBackend :: forall (i :: BackendType -> Type) (r :: Type). AnyBackend i -> (forall (b :: BackendType). i b -> r) -> r

Dispatch a function to the value inside the AnyBackend, that does not require bringing into scope a new class constraint.

dispatchAnyBackend :: forall (c :: BackendType -> Constraint) (i :: BackendType -> Type) (r :: Type). AllBackendsSatisfy c => AnyBackend i -> (forall (b :: BackendType). c b => i b -> r) -> r

Dispatch an existential using an universally quantified function while also resolving a different constraint. Use this to dispatch Backend* instances. This is essentially a wrapper around 'runAnyBackend f . repackAnyBackend @c'.

dispatchAnyBackendWithTwoConstraints :: forall (c1 :: BackendType -> Constraint) (c2 :: BackendType -> Constraint) (i :: BackendType -> Type) (r :: Type). AllBackendsSatisfy c1 => AllBackendsSatisfy c2 => AnyBackend i -> (forall (b :: BackendType). c1 b => c2 b => i b -> r) -> r

dispatchAnyBackend' :: forall (c :: Type -> Constraint) (i :: BackendType -> Type) (r :: Type). i `SatisfiesForAllBackends` c => AnyBackend i -> (forall (b :: BackendType). c (i b) => i b -> r) -> r

Unlike dispatchAnyBackend, the expected constraint has a different kind. Use for classes like Show, ToJSON, etc.

composeAnyBackend :: forall (c :: BackendType -> Constraint) (i :: BackendType -> Type) (r :: Type). AllBackendsSatisfy c => (forall (b :: BackendType). c b => i b -> i b -> r) -> AnyBackend i -> AnyBackend i -> r -> r

Sometimes we need to run operations on two backends of the same type. If the backends don't contain the same type, the given r value is returned. Otherwise, the function is called with the two wrapped values.

unpackAnyBackend :: forall (b :: BackendType) (i :: BackendType -> Type). HasTag b => AnyBackend i -> Maybe (i b)

Try to unpack the type of an existential. Returns Just x upon a succesful match, Nothing otherwise.

type BackendChoice (i :: BackendType -> Type) = Either (i ('Postgres 'Vanilla)) (Either (i ('Postgres 'Citus)) (Either (i 'MSSQL) (Either (i 'BigQuery) (Either (i 'MySQL) (Either (i 'DataWrapper) Void)))))

First, we create a type to represent our complicated Either type. We use Void as a terminating case for our recursion. This declaration creates the following type:

type BackendChoice (i :: BackendType -> Type) = Either (i 'Postgres) ( Either (i 'MSSQL) ( Either ... Void

spreadChoice :: forall (i :: BackendType -> Type) (arr :: Type -> Type -> Type). ArrowChoice arr => arr (AnyBackend i) (BackendChoice i)

Spread a AnyBackend into a BackendChoice.

Given backends Foo, Bar, Baz, the type of `BackendChoice c` will be: ( Either (c 'Foo) ( Either (c 'Bar) ( Either (c 'Baz) Void )))

Accordingly, the following Template Haskell splice generates the following code:

case e of FooValue x -> Left x BarValue x -> Right (Left x) BazValue x -> Right (Right (Left x))

coalesceChoice :: forall (c1 :: BackendType -> Constraint) (c2 :: BackendType -> Constraint) (i :: BackendType -> Type) (r :: Type) (arr :: Type -> Type -> Type). (ArrowChoice arr, AllBackendsSatisfy c1, AllBackendsSatisfy c2) => (forall b. c1 b => c2 b => arr (i b) r) -> arr (BackendChoice i) r

Coalesce a BackendChoice into a result, given an arrow from each possibilty to a common result.

Given backends Foo, Bar, Baz, the type of `BackendChoice c` will be: ( Either (c 'Foo) ( Either (c 'Bar) ( Either (c 'Baz) Void )))

Accordingly, the following Template Haskell splice generates the following code:

( arrow ||| ( arrow ||| ( arrow ||| absurd )))

dispatchAnyBackendArrow' :: forall (c1 :: BackendType -> Constraint) (c2 :: BackendType -> Constraint) (i :: BackendType -> Type) (r :: Type) (arr :: Type -> Type -> Type). (ArrowChoice arr, AllBackendsSatisfy c1, AllBackendsSatisfy c2) => (forall b. c1 b => c2 b => arr (i b) r) -> arr (AnyBackend i) r

Dispatch variant for use with arrow syntax. The universally quantified dispatch function is an arrow instead. Since we can't express this using Template Haskell, we instead generate the arrow by combining spreadChoice and coalesceChoice.

newtype BackendArrowTuple x i (b :: BackendType)

While dispatchAnyBackendArrow' is expressed over an AnyBackend, in practice we need slightly more complex types. Specifically: the only call site for dispatchAnyBackendArrow uses a four element tuple containing an AnyBackend.

Constructors

BackendArrowTuple

Fields

unTuple :: (i b, x)

dispatchAnyBackendArrow :: forall (c1 :: BackendType -> Constraint) (c2 :: BackendType -> Constraint) (i :: BackendType -> Type) (r :: Type) (arr :: Type -> Type -> Type) x. (ArrowChoice arr, AllBackendsSatisfy c1, AllBackendsSatisfy c2) => (forall b. c1 b => c2 b => arr (i b, x) r) -> arr (AnyBackend i, x) r

Finally, we can do the dispatch on the four-elements tuple. Here's what happens, step by step:

┌─────────────────────────┐ │ (x, y, AnyBackend i, z) │ └─┬───────────────────────┘ │ │ cons ▼ ┌────────────────────────────────────────┐ ┌─────────────────────────────┐ │ AnyBackend (BackendArrowTuple x y z i) │ ┌───► │ BackendArrowTuple x y z i b │ └─┬──────────────────────────────────────┘ │ └─┬───────────────────────────┘ │ │ │ │ spreadChoice │ │ arr unTuple ▼ │ ▼ ┌───────────────────────────────────────────┐ │ ┌────────────────┐ │ BackendChoice (BackendArrowTuple x y z i) │ │ │ (x, y, i b, z) │ └─┬─────────────────────────────────────────┘ │ └─┬──────────────┘ │ │ │ │ coalesceChoice (arr unTuple >>> arrow) ◄─────┘ │ arrow ▼ ▼ ┌───┐ ┌───┐ │ r │ │ r │ └───┘ └───┘

NOTE: The below function accepts two constraints, if the arrow you want to dispatch only has one constraint then repeat the constraint twice. For example: ```AB.dispatchAnyBackendArrow BackendMetadata BackendMetadata (proc (sourceMetadata, invalidationKeys)```

parseAnyBackendFromJSON :: i `SatisfiesForAllBackends` FromJSON => BackendType -> Value -> Parser (AnyBackend i)

Attempts to parse an AnyBackend from a JSON value, using the provided backend information.

debugAnyBackendToJSON :: i `SatisfiesForAllBackends` ToJSON => AnyBackend i -> Value

Outputs a debug JSON value from an AnyBackend. This function must only be used for debug purposes, as it has no way of inserting the backend kind in the output, since there's no guarantee that the output will be an object.