Hasura.Server.Auth

Documentation

class Monad m => UserAuthentication m where

Typeclass representing the UserInfo authorization and resolving effect

Methods

resolveUserInfo

Arguments

:: Logger Hasura

-> Manager

-> [Header]
request headers

-> AuthMode

-> Maybe ReqsText

-> m (Either QErr (UserInfo, Maybe UTCTime, [Header]))

newtype AdminSecretHash

The hashed admin password. hashAdminSecret is our public interface for constructing the secret.

To prevent misuse and leaking we keep this opaque and don't provide instances that could leak information. Likewise for AuthMode.

Although this exists only in memory we store only a hash of the admin secret primarily in order to:

prevent theoretical timing attacks from a naive == check

prevent misuse or inadvertent leaking of the secret

Constructors

AdminSecretHash (Digest SHA512)

unsafeMkAdminSecretHash :: Digest SHA512 -> AdminSecretHash

hashAdminSecret :: Text -> AdminSecretHash

data AuthMode

The methods we'll use to derive roles for authenticating requests.

Maybe RoleName below is the optionally-defined role for the unauthenticated (anonymous) user.

See: https://hasura.io/docs/latest/graphql/core/auth/authentication/unauthenticated-access.html

Constructors

AMNoAuth

AMAdminSecret !(HashSet AdminSecretHash) !(Maybe RoleName)

AMAdminSecretAndHook !(HashSet AdminSecretHash) !AuthHook

AMAdminSecretAndJWT !(HashSet AdminSecretHash) ![JWTCtx] !(Maybe RoleName)

setupAuthMode :: (ForkableMonadIO m, HasReporter m) => HashSet AdminSecretHash -> Maybe AuthHook -> [JWTConfig] -> Maybe RoleName -> Manager -> Logger Hasura -> ExceptT Text (ManagedT m) AuthMode

Validate the user's requested authentication configuration, launching any required maintenance threads for JWT etc.

This must only be run once, on launch.

getUserInfoWithExpTime :: forall m. (MonadIO m, MonadBaseControl IO m, MonadError QErr m, MonadTrace m) => Logger Hasura -> Manager -> [Header] -> AuthMode -> Maybe ReqsText -> m (UserInfo, Maybe UTCTime, [Header])

Authenticate the request using the headers and the configured AuthMode.

getUserInfoWithExpTime_

Arguments

:: forall m mgr logger. (MonadIO m, MonadError QErr m)

=> (logger -> mgr -> AuthHook -> [Header] -> Maybe ReqsText -> m (UserInfo, Maybe UTCTime, [Header]))
mock userInfoFromAuthHook

-> ([JWTCtx] -> [Header] -> Maybe RoleName -> m (UserInfo, Maybe UTCTime, [Header]))
mock processJwt

-> logger

-> mgr

-> [Header]

-> AuthMode

-> Maybe ReqsText

-> m (UserInfo, Maybe UTCTime, [Header])