Hasura.Server.Auth

Documentation

getUserInfoWithExpTime :: (MonadIO m, MonadBaseControl IO m, MonadError QErr m) => Logger Hasura -> Manager -> [Header] -> AuthMode -> Maybe ReqsText -> m (UserInfo, Maybe UTCTime, [Header])

Authenticate the request using the headers and the configured AuthMode.

data AuthMode

The methods we'll use to derive roles for authenticating requests.

Maybe RoleName below is the optionally-defined role for the unauthenticated (anonymous) user.

See: https://hasura.io/docs/latest/graphql/core/auth/authentication/unauthenticated-access.html

Constructors

AMNoAuth

AMAdminSecret !(HashSet AdminSecretHash) !(Maybe RoleName)

AMAdminSecretAndHook !(HashSet AdminSecretHash) !AuthHook

AMAdminSecretAndJWT !(HashSet AdminSecretHash) ![JWTCtx] !(Maybe RoleName)

compareAuthMode :: AuthMode -> AuthMode -> IO Bool

In case JWT is used as an authentication mode, the JWKs are stored inside JWTCtx as an IORef. IORef has pointer equality, so we need to compare the values inside the IORef to check if the JWTCtx is same.

setupAuthMode :: (MonadError Text m, MonadIO m, MonadBaseControl IO m) => HashSet AdminSecretHash -> Maybe AuthHook -> [JWTConfig] -> Maybe RoleName -> Logger Hasura -> Manager -> m AuthMode

Validate the user's requested authentication configuration, launching any required maintenance threads for JWT etc.

This must only be run once, on launch.

data AdminSecretHash

The hashed admin password. hashAdminSecret is our public interface for constructing the secret.

To prevent misuse and leaking we keep this opaque and don't provide instances that could leak information. Likewise for AuthMode.

Although this exists only in memory we store only a hash of the admin secret primarily in order to:

prevent theoretical timing attacks from a naive == check

prevent misuse or inadvertent leaking of the secret

unsafeMkAdminSecretHash :: Digest SHA512 -> AdminSecretHash

hashAdminSecret :: Text -> AdminSecretHash

updateJwkCtx :: (MonadIO m, MonadBaseControl IO m) => AuthMode -> Manager -> Logger Hasura -> m ()

Update the JWK based on the expiry time specified in Expires header or Cache-Control header

WebHook related

data AuthHookType

Constructors

AHTGet

AHTPost

data AuthHook

Constructors

AuthHook

Fields

ahUrl :: Text

ahType :: AuthHookType

ahSendRequestBody :: Bool

Whether to send the request body to the auth hook

JWT related

data RawJWT

data JWTConfig

The JWT configuration we got from the user.

Constructors

JWTConfig

Fields

jcKeyOrUrl :: !(Either JWK URI)

jcAudience :: !(Maybe Audience)

jcIssuer :: !(Maybe StringOrURI)

jcClaims :: !JWTClaims

jcAllowedSkew :: !(Maybe NominalDiffTime)

jcHeader :: !(Maybe JWTHeader)

data JWTCtx

The validated runtime JWT configuration returned by mkJwtCtx in setupAuthMode.

Constructors

JWTCtx

Fields

jcxUrl :: !(Maybe URI)

jcxKeyConfig :: !(IORef (JWKSet, Maybe UTCTime))

This needs to be a mutable variable for fetchJwk. | We add the expiry time of the JWK to the IORef, to determine | if the JWK has expired and needs to be refreshed.

jcxAudience :: !(Maybe Audience)

jcxIssuer :: !(Maybe StringOrURI)

jcxClaims :: !JWTClaims

jcxAllowedSkew :: !(Maybe NominalDiffTime)

jcxHeader :: !JWTHeader

newtype JWKSet

RFC 7517 5. JWK Set Format

Constructors

JWKSet [JWK]

processJwt :: (MonadIO m, MonadError QErr m) => [JWTCtx] -> RequestHeaders -> Maybe RoleName -> m (UserInfo, Maybe UTCTime, [Header], Maybe JWTCtx)

Process the request headers to verify the JWT and extract UserInfo from it From the JWT config, we check which header to expect, it can be the Authorization or Cookie header

If no Authorization/Cookie header was passed, we will fall back to the unauthenticated user role [1], if one was configured at server start.

When no 'x-hasura-user-role' is specified in the request, the mandatory 'x-hasura-default-role' [2] from the JWT claims will be used.

class Monad m => UserAuthentication (m :: Type -> Type) where

Typeclass representing the UserInfo authorization and resolving effect

Methods

resolveUserInfo

Arguments

:: Logger Hasura

-> Manager

-> [Header]
request headers

-> AuthMode

-> Maybe ReqsText

-> m (Either QErr (UserInfo, Maybe UTCTime, [Header], ExtraUserInfo))

mkJwtCtx :: (MonadIO m, MonadBaseControl IO m, MonadError Text m) => JWTConfig -> Logger Hasura -> Manager -> m JWTCtx

updateJwkFromUrl :: (MonadIO m, MonadBaseControl IO m) => JWTCtx -> Manager -> Logger Hasura -> m ()

Exposed for testing

getUserInfoWithExpTime_

Arguments

:: forall m mgr logger. (MonadIO m, MonadError QErr m)

=> (logger -> mgr -> AuthHook -> [Header] -> Maybe ReqsText -> m (UserInfo, Maybe UTCTime, [Header]))
mock userInfoFromAuthHook

-> ([JWTCtx] -> [Header] -> Maybe RoleName -> m (UserInfo, Maybe UTCTime, [Header], Maybe JWTCtx))
mock processJwt

-> logger

-> mgr

-> [Header]

-> AuthMode

-> Maybe ReqsText

-> m (UserInfo, Maybe UTCTime, [Header])