Hasura.Server.Auth.JWT

Description

This module implements the bulk of Hasura's JWT capabilities and interactions. Its main point of non-testing invocation is Auth.

It exports both processJwt and processJwt_ with processJwt_ being the majority of the implementation with the JWT Token processing function passed in as an argument in order to enable mocking in test-code.

In processJwt_, prior to validation of the token, first the token locations and issuers are reconciled. Locations are either specified as auth or cookie (with cookie name) or assumed to be auth. Issuers can be omitted or specified, where an omitted configured issuer can match any issuer specified by a request.

If none match, then this is considered an no-auth request, if one matches, then normal token auth is performed, and if multiple match, then this is considered an ambiguity error.

Documentation

newtype RawJWT

Constructors

RawJWT ByteString

data JWTClaimsFormat

Constructors

JCFJson

JCFStringifiedJson

data JWTHeader

Constructors

JHAuthorization

JHCookie Text

defaultClaimsFormat :: JWTClaimsFormat

allowedRolesClaim :: SessionVariable

defaultRoleClaim :: SessionVariable

defaultClaimsNamespace :: Text

data JWTCustomClaimsMapValueG v

JWTCustomClaimsMapValueG is used to represent a single value of the JWTCustomClaimsMap. A JWTCustomClaimsMapValueG can either be an JSON object or the literal value of the claim. If the value is an JSON object, then it should contain a key path, which is the JSON path to the claim value in the JWT token. There's also an option to specify a default value in the map via the 'default' key, which will be used when a peek at the JWT token using the JSON path fails (key does not exist).

Constructors

JWTCustomClaimsMapJSONPath !JSONPath !(Maybe v)
JSONPath to the key in the claims map, in case the key doesn't exist in the claims map then the default value will be used (if provided)

JWTCustomClaimsMapStatic !v

type JWTCustomClaimsMapDefaultRole = JWTCustomClaimsMapValueG RoleName

type JWTCustomClaimsMapAllowedRoles = JWTCustomClaimsMapValueG [RoleName]

type JWTCustomClaimsMapValue = JWTCustomClaimsMapValueG SessionVariableValue

type CustomClaimsMap = HashMap SessionVariable JWTCustomClaimsMapValue

data JWTCustomClaimsMap

JWTClaimsMap is an option to provide a custom JWT claims map. The JWTClaimsMap should be specified in the HASURA_GRAPHQL_JWT_SECRET in the claims_map. The JWTClaimsMap, if specified, requires two mandatory fields, namely, `x-hasura-allowed-roles` and the `x-hasura-default-role`, other claims may also be provided in the claims map.

Constructors

JWTCustomClaimsMap

Fields

jcmDefaultRole :: !JWTCustomClaimsMapDefaultRole

jcmAllowedRoles :: !JWTCustomClaimsMapAllowedRoles

jcmCustomClaims :: !CustomClaimsMap

data JWTNamespace

JWTNamespace is used to locate the claims map within the JWT token. The location can be either provided via a JSON path or the name of the key in the JWT token.

Constructors

ClaimNsPath JSONPath

ClaimNs Text

data JWTClaims

Constructors

JCNamespace !JWTNamespace !JWTClaimsFormat

JCMap !JWTCustomClaimsMap

newtype StringOrURI

Hashable Wrapper for constructing a HashMap of JWTConfigs

Constructors

StringOrURI

Fields

unStringOrURI :: StringOrURI

data JWTConfig

The JWT configuration we got from the user.

Constructors

JWTConfig

Fields

jcKeyOrUrl :: !(Either JWK URI)

jcAudience :: !(Maybe Audience)

jcIssuer :: !(Maybe StringOrURI)

jcClaims :: !JWTClaims

jcAllowedSkew :: !(Maybe NominalDiffTime)

jcHeader :: !(Maybe JWTHeader)

data JWTCtx

The validated runtime JWT configuration returned by mkJwtCtx in setupAuthMode.

This is also evidence that the jwkRefreshCtrl thread is running, if an expiration schedule could be determined.

Constructors

JWTCtx

Fields

jcxKey :: !(IORef JWKSet)

This needs to be a mutable variable for updateJwkRef.

jcxAudience :: !(Maybe Audience)

jcxIssuer :: !(Maybe StringOrURI)

jcxClaims :: !JWTClaims

jcxAllowedSkew :: !(Maybe NominalDiffTime)

jcxHeader :: !JWTHeader

data HasuraClaims

Constructors

HasuraClaims

Fields

_cmAllowedRoles :: ![RoleName]

_cmDefaultRole :: !RoleName

jwkRefreshCtrl :: (MonadIO m, MonadBaseControl IO m, HasReporter m) => Logger Hasura -> Manager -> URI -> IORef JWKSet -> DiffTime -> m void

An action that refreshes the JWK at intervals in an infinite loop.

updateJwkRef :: (MonadIO m, MonadBaseControl IO m, MonadError JwkFetchError m, MonadTrace m) => Logger Hasura -> Manager -> URI -> IORef JWKSet -> m (Maybe NominalDiffTime)

Given a JWK url, fetch JWK from it and update the IORef

determineJwkExpiryLifetime :: forall m. (MonadIO m, MonadError JwkFetchError m) => m UTCTime -> Logger Hasura -> ResponseHeaders -> m (Maybe NominalDiffTime)

First check for Cache-Control header, if not found, look for Expires header

type ClaimsMap = HashMap SessionVariable Value

decodeClaimsSet :: RawJWT -> Maybe ClaimsSet

Decode a Jose ClaimsSet without verifying the signature

tokenIssuer :: RawJWT -> Maybe StringOrURI

Extract the issuer from a bearer tokena _without_ verifying it.

processJwt :: (MonadIO m, MonadError QErr m) => [JWTCtx] -> RequestHeaders -> Maybe RoleName -> m (UserInfo, Maybe UTCTime, [Header])

Process the request headers to verify the JWT and extract UserInfo from it From the JWT config, we check which header to expect, it can be the Authorization or Cookie header

Iff no Authorization/Cookie header was passed, we will fall back to the unauthenticated user role [1], if one was configured at server start.

When no 'x-hasura-user-role' is specified in the request, the mandatory 'x-hasura-default-role' [2] from the JWT claims will be used.

type AuthTokenLocation = JWTHeader

processJwt_

Arguments

:: MonadError QErr m

=> (JWTCtx -> ByteString -> m (ClaimsMap, Maybe UTCTime))
mock processAuthZOrCookieHeader

-> (RawJWT -> Maybe StringOrURI)

-> (JWTCtx -> JWTHeader)

-> [JWTCtx]

-> RequestHeaders

-> Maybe RoleName

-> m (UserInfo, Maybe UTCTime, [Header])

processHeaderSimple :: (MonadIO m, MonadError QErr m) => JWTCtx -> ByteString -> m (ClaimsMap, Maybe UTCTime)

Processes a token payload (excluding the `Bearer ` prefix in the context of a JWTCtx)

parseClaimsMap

Arguments

:: MonadError QErr m

=> ClaimsSet
Unregistered JWT claims

-> JWTClaims
Claims config

-> m ClaimsMap
Hasura claims and other claims

parse the claims map from the JWT token or custom claims from the JWT config

verifyJwt :: (MonadError JWTError m, MonadIO m) => JWTCtx -> RawJWT -> m ClaimsSet

Verify the JWT against given JWK

parseHasuraClaims :: forall m. MonadError QErr m => ClaimsMap -> m HasuraClaims

parseJwtClaim :: (FromJSON a, MonadError QErr m) => Value -> Text -> m a