Hasura.Server.Auth.JWT

Description

This module implements the bulk of Hasura's JWT capabilities and interactions. Its main point of non-testing invocation is Auth.

It exports both processJwt and processJwt_ with processJwt_ being the majority of the implementation with the JWT Token processing function passed in as an argument in order to enable mocking in test-code.

In processJwt_, prior to validation of the token, first the token locations and issuers are reconciled. Locations are either specified as auth or cookie (with cookie name) or assumed to be auth. Issuers can be omitted or specified, where an omitted configured issuer can match any issuer specified by a request.

If none match, then this is considered an no-auth request, if one matches, then normal token auth is performed, and if multiple match, then this is considered an ambiguity error.

Documentation

processJwt :: (MonadIO m, MonadError QErr m) => [JWTCtx] -> RequestHeaders -> Maybe RoleName -> m (UserInfo, Maybe UTCTime, [Header], Maybe JWTCtx)

Process the request headers to verify the JWT and extract UserInfo from it From the JWT config, we check which header to expect, it can be the Authorization or Cookie header

If no Authorization/Cookie header was passed, we will fall back to the unauthenticated user role [1], if one was configured at server start.

When no 'x-hasura-user-role' is specified in the request, the mandatory 'x-hasura-default-role' [2] from the JWT claims will be used.

data RawJWT

newtype StringOrURI

Hashable Wrapper for constructing a HashMap of JWTConfigs

Constructors

StringOrURI

Fields

unStringOrURI :: StringOrURI

data JWTConfig

The JWT configuration we got from the user.

Constructors

JWTConfig

Fields

jcKeyOrUrl :: !(Either JWK URI)

jcAudience :: !(Maybe Audience)

jcIssuer :: !(Maybe StringOrURI)

jcClaims :: !JWTClaims

jcAllowedSkew :: !(Maybe NominalDiffTime)

jcHeader :: !(Maybe JWTHeader)

data JWTCtx

The validated runtime JWT configuration returned by mkJwtCtx in setupAuthMode.

Constructors

JWTCtx

Fields

jcxUrl :: !(Maybe URI)

jcxKeyConfig :: !(IORef (JWKSet, Maybe UTCTime))

This needs to be a mutable variable for fetchJwk. | We add the expiry time of the JWK to the IORef, to determine | if the JWK has expired and needs to be refreshed.

jcxAudience :: !(Maybe Audience)

jcxIssuer :: !(Maybe StringOrURI)

jcxClaims :: !JWTClaims

jcxAllowedSkew :: !(Maybe NominalDiffTime)

jcxHeader :: !JWTHeader

newtype JWKSet

RFC 7517 5. JWK Set Format

Constructors

JWKSet [JWK]

data JWTClaimsFormat

Constructors

JCFJson

JCFStringifiedJson

data JWTClaims

Constructors

JCNamespace !JWTNamespace !JWTClaimsFormat

JCMap !JWTCustomClaimsMap

data JwkFetchError

Possible errors during fetching and parsing JWK (the Text type at the end is a friendly error message)

Constructors

JFEHttpException !HttpException !Text
Exception while making the HTTP request

JFEHttpError !URI !Status !ByteString !Text
Non-2xx HTTP errors from the upstream server

JFEJwkParseError !Text !Text
Error parsing the JWK response itself

JFEExpiryParseError !(Maybe Text) Text
Error parsing the expiry of the JWK

data JWTHeader

Constructors

JHAuthorization

JHCookie Text

JHCustomHeader Text

data JWTNamespace

JWTNamespace is used to locate the claims map within the JWT token. The location can be either provided via a JSON path or the name of the key in the JWT token.

Constructors

ClaimNsPath JSONPath

ClaimNs Text

type JWTCustomClaimsMapDefaultRole = JWTCustomClaimsMapValueG RoleName

type JWTCustomClaimsMapAllowedRoles = JWTCustomClaimsMapValueG [RoleName]

type JWTCustomClaimsMapValue = JWTCustomClaimsMapValueG SessionVariableValue

type ClaimsMap = HashMap SessionVariable Value

fetchAndUpdateJWKs :: (MonadIO m, MonadBaseControl IO m) => Logger Hasura -> Manager -> URI -> IORef (JWKSet, Maybe UTCTime) -> m ()

An action that fetches the JWKs and updates the expiry time and JWKs in the IORef

fetchJwk :: (MonadIO m, MonadBaseControl IO m, MonadError JwkFetchError m) => Logger Hasura -> Manager -> URI -> m (JWKSet, ResponseHeaders)

Given a JWK url, fetch JWK from it

defaultClaimsFormat :: JWTClaimsFormat

defaultClaimsNamespace :: Text

Exposed for testing

processJwt_

Arguments

:: MonadError QErr m

=> (JWTCtx -> ByteString -> m (ClaimsMap, Maybe UTCTime))
mock processAuthZOrCookieHeader

-> (RawJWT -> Maybe StringOrURI)

-> (JWTCtx -> JWTHeader)

-> [JWTCtx]

-> RequestHeaders

-> Maybe RoleName

-> m (UserInfo, Maybe UTCTime, [Header], Maybe JWTCtx)

tokenIssuer :: RawJWT -> Maybe StringOrURI

Extract the issuer from a bearer tokena _without_ verifying it.

allowedRolesClaim :: SessionVariable

defaultRoleClaim :: SessionVariable

parseClaimsMap

Arguments

:: MonadError QErr m

=> ClaimsSet
Unregistered JWT claims

-> JWTClaims
Claims config

-> m ClaimsMap
Hasura claims and other claims

parse the claims map from the JWT token or custom claims from the JWT config

parseObjectAsClaims :: KeyMap v -> HashMap SessionVariable v

Parses a JSON key-value object, including only x-hasura claims

data JWTCustomClaimsMapValueG v

JWTCustomClaimsMapValueG is used to represent a single value of the JWTCustomClaimsMap. A JWTCustomClaimsMapValueG can either be an JSON object or the literal value of the claim. If the value is an JSON object, then it should contain a key path, which is the JSON path to the claim value in the JWT token. There's also an option to specify a default value in the map via the 'default' key, which will be used when a peek at the JWT token using the JSON path fails (key does not exist).

Constructors

JWTCustomClaimsMapJSONPath !JSONPath !(Maybe v)
JSONPath to the key in the claims map, in case the key doesn't exist in the claims map then the default value will be used (if provided)

JWTCustomClaimsMapStatic !v

data JWTCustomClaimsMap

JWTClaimsMap is an option to provide a custom JWT claims map. The JWTClaimsMap should be specified in the HASURA_GRAPHQL_JWT_SECRET in the claims_map. The JWTClaimsMap, if specified, requires two mandatory fields, namely, `x-hasura-allowed-roles` and the `x-hasura-default-role`, other claims may also be provided in the claims map.

Constructors

JWTCustomClaimsMap

Fields

jcmDefaultRole :: !JWTCustomClaimsMapDefaultRole

jcmAllowedRoles :: !JWTCustomClaimsMapAllowedRoles

jcmCustomClaims :: !CustomClaimsMap

determineJwkExpiryLifetime :: (MonadIO m, MonadError JwkFetchError m) => m UTCTime -> Logger Hasura -> ResponseHeaders -> m (Maybe UTCTime)

First check for Cache-Control header, if not found, look for Expires header