Hasura.Server.Init

Description

Types and functions related to the server initialisation

Documentation

getDbId :: TxE QErr Text

getPgVersion :: TxE QErr PGVersion

generateInstanceId :: IO InstanceId

data StartupTimeInfo

Constructors

StartupTimeInfo

Fields

_stiMessage :: !Text

_stiTimeTaken :: !Double

returnJust :: Monad m => a -> m (Maybe a)

considerEnv :: FromEnv a => String -> WithEnv (Maybe a)

Lookup a key in the application environment then parse the value with a FromEnv instance'

considerEnvs :: FromEnv a => [String] -> WithEnv (Maybe a)

Lookup a list of keys with considerEnv and return the first value to parse successfully.

withEnv :: FromEnv a => Maybe a -> String -> WithEnv (Maybe a)

withEnvs :: FromEnv a => Maybe a -> [String] -> WithEnv (Maybe a)

Return a if Just or else call considerEnv with a list of env keys k

withEnvBool :: Bool -> String -> WithEnv Bool

withEnvJwtConf :: Maybe JWTConfig -> String -> WithEnv (Maybe JWTConfig)

mkHGEOptions :: EnabledLogTypes impl => RawHGEOptions impl -> WithEnv (HGEOptions impl)

processPostgresConnInfo :: PostgresConnInfo (Maybe PostgresRawConnInfo) -> WithEnv (PostgresConnInfo (Maybe UrlConf))

rawConnInfoToUrlConf :: Maybe PostgresRawConnInfo -> WithEnv (Maybe UrlConf)

mkServeOptions :: EnabledLogTypes impl => RawServeOptions impl -> WithEnv (ServeOptions impl)

mkExamplesDoc :: [[String]] -> Doc

mkEnvVarDoc :: [(String, String)] -> Doc

mainCmdFooter :: Doc

databaseUrlEnv :: (String, String)

metadataDbUrlEnv :: (String, String)

serveCmdFooter :: Doc

eventsHttpPoolSizeEnv :: (String, String)

eventsFetchIntervalEnv :: (String, String)

eventsFetchBatchSizeEnv :: (String, String)

asyncActionsFetchIntervalEnv :: (String, String)

logHeadersFromEnvEnv :: (String, String)

retriesNumEnv :: (String, String)

servePortEnv :: (String, String)

serveHostEnv :: (String, String)

pgConnsEnv :: (String, String)

pgStripesEnv :: (String, String)

pgTimeoutEnv :: (String, String)

pgConnLifetimeEnv :: (String, String)

pgPoolTimeoutEnv :: (String, String)

pgUsePrepareEnv :: (String, String)

txIsoEnv :: (String, String)

accessKeyEnv :: (String, String)

adminSecretEnv :: (String, String)

authHookEnv :: (String, String)

authHookModeEnv :: (String, String)

jwtSecretEnv :: (String, String)

unAuthRoleEnv :: (String, String)

corsDisableEnv :: (String, String)

corsDomainEnv :: (String, String)

enableConsoleEnv :: (String, String)

enableTelemetryEnv :: (String, String)

wsReadCookieEnv :: (String, String)

stringifyNumEnv :: (String, String)

dangerousBooleanCollapseEnv :: (String, String)

defaultEnabledAPIs :: [API]

enabledAPIsEnv :: (String, String)

experimentalFeaturesEnv :: (String, String)

gracefulShutdownEnv :: (String, String)

consoleAssetsDirEnv :: (String, String)

enabledLogsEnv :: (String, String)

logLevelEnv :: (String, String)

devModeEnv :: (String, String)

enableRemoteSchemaPermsEnv :: (String, String)

inferFunctionPermsEnv :: (String, String)

maintenanceModeEnv :: (String, String)

schemaPollIntervalEnv :: (String, String)

adminInternalErrorsEnv :: (String, String)

parsePostgresConnInfo :: Parser (PostgresConnInfo (Maybe PostgresRawConnInfo))

parseDatabaseUrl :: Parser (Maybe URLTemplate)

parseRawConnDetails :: Parser (Maybe PostgresRawConnDetails)

parseMetadataDbUrl :: Parser (Maybe String)

parseTxIsolation :: Parser (Maybe TxIsolation)

parseConnParams :: Parser RawConnParams

parseServerPort :: Parser (Maybe Int)

parseServerHost :: Parser (Maybe HostPreference)

parseAccessKey :: Parser (Maybe AdminSecretHash)

parseAdminSecret :: Parser (Maybe AdminSecretHash)

parseWebHook :: Parser RawAuthHook

parseJwtSecret :: Parser (Maybe JWTConfig)

jwtSecretHelp :: String

parseUnAuthRole :: Parser (Maybe RoleName)

parseCorsConfig :: Parser (Maybe CorsConfig)

parseEnableConsole :: Parser Bool

parseConsoleAssetsDir :: Parser (Maybe Text)

parseEnableTelemetry :: Parser (Maybe Bool)

parseWsReadCookie :: Parser Bool

parseStringifyNum :: Parser Bool

parseDangerousBooleanCollapse :: Parser (Maybe Bool)

parseEnabledAPIs :: Parser (Maybe [API])

parseExperimentalFeatures :: Parser (Maybe [ExperimentalFeature])

parseGracefulShutdownTimeout :: Parser (Maybe Seconds)

parseMxRefetchInt :: Parser (Maybe RefetchInterval)

parseMxBatchSize :: Parser (Maybe BatchSize)

parseEnableAllowlist :: Parser Bool

parseGraphqlDevMode :: Parser Bool

parseGraphqlAdminInternalErrors :: Parser (Maybe Bool)

parseGraphqlEventsHttpPoolSize :: Parser (Maybe Int)

parseGraphqlEventsFetchInterval :: Parser (Maybe Milliseconds)

parseEventsFetchBatchSize :: Parser (Maybe NonNegativeInt)

parseGraphqlAsyncActionsFetchInterval :: Parser (Maybe Milliseconds)

parseLogHeadersFromEnv :: Parser Bool

parseEnableRemoteSchemaPerms :: Parser Bool

parseInferFunctionPerms :: Parser (Maybe Bool)

parseEnableMaintenanceMode :: Parser Bool

parseSchemaPollInterval :: Parser (Maybe Milliseconds)

mxRefetchDelayEnv :: (String, String)

mxBatchSizeEnv :: (String, String)

enableAllowlistEnv :: (String, String)

parsePlanCacheSize :: Parser (Maybe CacheSize)

parseEnabledLogs :: EnabledLogTypes impl => Parser (Maybe [EngineLogType impl])

parseLogLevel :: Parser (Maybe LogLevel)

censorQueryItem :: Text -> QueryItem -> QueryItem

censorQuery :: Text -> Query -> Query

updateQuery :: (Query -> Query) -> URI -> URI

censorURI :: Text -> URI -> URI

connInfoToLog :: ConnInfo -> StartupLog

serveOptsToLog :: ToJSON (EngineLogType impl) => ServeOptions impl -> StartupLog

mkGenericStrLog :: LogLevel -> Text -> String -> StartupLog

mkGenericLog :: ToJSON a => LogLevel -> Text -> a -> StartupLog

inconsistentMetadataLog :: SchemaCache -> StartupLog

serveOptionsParser :: EnabledLogTypes impl => Parser (RawServeOptions impl)

downgradeShortcuts :: [(String, String)]

This implements the mapping between application versions and catalog schema versions.

downgradeOptionsParser :: Parser DowngradeOptions

webSocketCompressionEnv :: (String, String)

parseWebSocketCompression :: Parser Bool

webSocketKeepAliveEnv :: (String, String)

parseWebSocketKeepAlive :: Parser (Maybe Int)

webSocketConnectionInitTimeoutEnv :: (String, String)

parseWebSocketConnectionInitTimeout :: Parser (Maybe Int)