Hasura.Server.Init.Config

Description

Types and classes related to configuration when the server is initialised

Documentation

data RawConnParams

Constructors

RawConnParams

Fields

rcpStripes :: !(Maybe Int)

rcpConns :: !(Maybe Int)

rcpIdleTime :: !(Maybe Int)

rcpConnLifetime :: !(Maybe NominalDiffTime)

Time from connection creation after which to destroy a connection and choose a different/new one.

rcpAllowPrepare :: !(Maybe Bool)

rcpPoolTimeout :: !(Maybe NominalDiffTime)

See HASURA_GRAPHQL_PG_POOL_TIMEOUT

type RawAuthHook = AuthHookG (Maybe Text) (Maybe AuthHookType)

data OptionalInterval

Sleep time interval for recurring activities such as (asyncActionsProcessor) Presently msToOptionalInterval interprets `0` as Skip.

Constructors

Skip
No polling

Interval !Milliseconds
Interval time

msToOptionalInterval :: Milliseconds -> OptionalInterval

data API

Constructors

METADATA

GRAPHQL

PGDUMP

DEVELOPER

CONFIG

data RawServeOptions impl

Constructors

RawServeOptions

Fields

rsoPort :: Maybe Int

rsoHost :: Maybe HostPreference

rsoConnParams :: RawConnParams

rsoTxIso :: Maybe TxIsolation

rsoAdminSecret :: Maybe AdminSecretHash

rsoAuthHook :: RawAuthHook

rsoJwtSecret :: Maybe JWTConfig

rsoUnAuthRole :: Maybe RoleName

rsoCorsConfig :: Maybe CorsConfig

rsoEnableConsole :: Bool

rsoConsoleAssetsDir :: Maybe Text

rsoEnableTelemetry :: Maybe Bool

rsoWsReadCookie :: Bool

rsoStringifyNum :: Bool

rsoDangerousBooleanCollapse :: Maybe Bool

rsoEnabledAPIs :: Maybe [API]

rsoMxRefetchInt :: Maybe RefetchInterval

rsoMxBatchSize :: Maybe BatchSize

rsoStreamingMxRefetchInt :: Maybe RefetchInterval

rsoStreamingMxBatchSize :: Maybe BatchSize

rsoEnableAllowlist :: Bool

rsoEnabledLogTypes :: Maybe [EngineLogType impl]

rsoLogLevel :: Maybe LogLevel

rsoDevMode :: Bool

rsoAdminInternalErrors :: Maybe Bool

rsoEventsHttpPoolSize :: Maybe Int

rsoEventsFetchInterval :: Maybe Milliseconds

rsoAsyncActionsFetchInterval :: Maybe Milliseconds

rsoLogHeadersFromEnv :: Bool

rsoEnableRemoteSchemaPermissions :: Bool

rsoWebSocketCompression :: Bool

rsoWebSocketKeepAlive :: Maybe Int

rsoInferFunctionPermissions :: Maybe Bool

rsoEnableMaintenanceMode :: Bool

rsoSchemaPollInterval :: Maybe Milliseconds

rsoExperimentalFeatures :: Maybe [ExperimentalFeature]

rsoEventsFetchBatchSize :: Maybe NonNegativeInt

rsoGracefulShutdownTimeout :: Maybe Seconds

rsoWebSocketConnectionInitTimeout :: Maybe Int

rsoEnableMetadataQueryLoggingEnv :: Bool

data ResponseInternalErrorsConfig

@ResponseInternalErrorsConfig represents the encoding of the internal errors in the response to the client. See the github comment https://github.com/hasura/graphql-engine/issues/4031#issuecomment-609747705 for more details.

Constructors

InternalErrorsAllRequests

InternalErrorsAdminOnly

InternalErrorsDisabled

shouldIncludeInternal :: RoleName -> ResponseInternalErrorsConfig -> Bool

newtype KeepAliveDelay

Constructors

KeepAliveDelay

Fields

unKeepAliveDelay :: Seconds

defaultKeepAliveDelay :: KeepAliveDelay

newtype WSConnectionInitTimeout

Constructors

WSConnectionInitTimeout

Fields

unWSConnectionInitTimeout :: Seconds

defaultWSConnectionInitTimeout :: WSConnectionInitTimeout

data ServeOptions impl

Constructors

ServeOptions

Fields

soPort :: Int

soHost :: HostPreference

soConnParams :: ConnParams

soTxIso :: TxIsolation

soAdminSecret :: HashSet AdminSecretHash

soAuthHook :: Maybe AuthHook

soJwtSecret :: [JWTConfig]

soUnAuthRole :: Maybe RoleName

soCorsConfig :: CorsConfig

soEnableConsole :: Bool

soConsoleAssetsDir :: Maybe Text

soEnableTelemetry :: Bool

soStringifyNum :: StringifyNumbers

soDangerousBooleanCollapse :: Bool

soEnabledAPIs :: HashSet API

soLiveQueryOpts :: LiveQueriesOptions

soStreamingQueryOpts :: StreamQueriesOptions

soEnableAllowlist :: Bool

soEnabledLogTypes :: HashSet (EngineLogType impl)

soLogLevel :: LogLevel

soResponseInternalErrorsConfig :: ResponseInternalErrorsConfig

soEventsHttpPoolSize :: Maybe Int

soEventsFetchInterval :: Maybe Milliseconds

soAsyncActionsFetchInterval :: OptionalInterval

soLogHeadersFromEnv :: Bool

soEnableRemoteSchemaPermissions :: RemoteSchemaPermsCtx

soConnectionOptions :: ConnectionOptions

soWebsocketKeepAlive :: KeepAliveDelay

soInferFunctionPermissions :: FunctionPermissionsCtx

soEnableMaintenanceMode :: MaintenanceMode

soSchemaPollInterval :: OptionalInterval

soExperimentalFeatures :: HashSet ExperimentalFeature

soEventsFetchBatchSize :: NonNegativeInt

soDevMode :: Bool

soGracefulShutdownTimeout :: Seconds

soWebsocketConnectionInitTimeout :: WSConnectionInitTimeout

soEventingMode :: EventingMode

soReadOnlyMode :: ReadOnlyMode

soEnableMetadataQueryLogging :: MetadataQueryLoggingMode

data DowngradeOptions

Constructors

DowngradeOptions

Fields

dgoTargetVersion :: !Text

dgoDryRun :: !Bool

data PostgresConnInfo a

Constructors

PostgresConnInfo

Fields

_pciDatabaseConn :: !a

_pciRetries :: !(Maybe Int)

data PostgresRawConnDetails

Constructors

PostgresRawConnDetails

Fields

connHost :: !String

connPort :: !Int

connUser :: !String

connPassword :: !String

connDatabase :: !String

connOptions :: !(Maybe String)

data PostgresRawConnInfo

Constructors

PGConnDatabaseUrl !URLTemplate

PGConnDetails !PostgresRawConnDetails

rawConnDetailsToUrl :: PostgresRawConnDetails -> URLTemplate

rawConnDetailsToUrlText :: PostgresRawConnDetails -> Text

data HGECommandG a

Constructors

HCServe !a

HCExport

HCClean

HCVersion

HCDowngrade !DowngradeOptions

type HGECommand impl = HGECommandG (ServeOptions impl)

type RawHGECommand impl = HGECommandG (RawServeOptions impl)

data HGEOptionsG a b

Constructors

HGEOptionsG

Fields

hoDatabaseUrl :: !(PostgresConnInfo a)

hoMetadataDbUrl :: !(Maybe String)

hoCommand :: !(HGECommandG b)

type RawHGEOptions impl = HGEOptionsG (Maybe PostgresRawConnInfo) (RawServeOptions impl)

type HGEOptions impl = HGEOptionsG (Maybe UrlConf) (ServeOptions impl)

type Env = [(String, String)]

readHookType :: String -> Either String AuthHookType

parseStrAsBool :: String -> Either String Bool

readNonNegativeInt :: String -> Either String NonNegativeInt

readAPIs :: String -> Either String [API]

readExperimentalFeatures :: String -> Either String [ExperimentalFeature]

readLogLevel :: String -> Either String LogLevel

class FromEnv a where

Methods

fromEnv :: String -> Either String a

type WithEnv a = ReaderT Env (ExceptT String Identity) a

runWithEnv :: Env -> WithEnv a -> Either String a