Hasura.Server.Init.Config

Description

Types and classes related to configuration when the server is initialised

Documentation

data Option def

The collected default value, env var, and help message for an option. If there should be no default value then use 'Option ()'.

Constructors

Option

Fields

_default :: def

_envVar :: String

_helpMessage :: String

optionPP :: Option a -> (String, String)

Helper function for pretty printing Option a.

data HGEOptionsRaw impl

Raw HGE Options from the arg parser and the env.

Constructors

HGEOptionsRaw

Fields

_horDatabaseUrl :: PostgresConnInfo (Maybe PostgresConnInfoRaw)

_horMetadataDbUrl :: Maybe String

_horCommand :: HGECommand impl

horDatabaseUrl :: Lens' (HGEOptionsRaw impl) (PostgresConnInfo (Maybe PostgresConnInfoRaw))

horMetadataDbUrl :: Lens' (HGEOptionsRaw impl) (Maybe String)

horCommand :: Lens' (HGEOptionsRaw impl) (HGECommand impl)

data HGEOptions impl

The final processed HGE options.

Constructors

HGEOptions

Fields

_hoDatabaseUrl :: PostgresConnInfo (Maybe UrlConf)

_hoMetadataDbUrl :: Maybe String

_hoCommand :: HGECommand impl

hoCommand :: Lens' (HGEOptions impl) (HGECommand impl)

data PostgresConnInfo a

Postgres connection info tupled with a retry count.

In practice, the a here is one of the following: 1. 'Maybe PostgresConnInfoRaw' 2. 'Maybe UrlConf' 3. 'Maybe Text' 4. 'Maybe DatabaseUrl' where DatabaseUrl is an alias for Text

If it contains a 'Maybe PostgresConnInfoRaw' then you have not yet processed your arg parser results.

Constructors

PostgresConnInfo

Fields

_pciDatabaseConn :: a

_pciRetries :: Maybe Int

pciDatabaseConn :: Lens' (PostgresConnInfo a) a

pciRetries :: Lens' (PostgresConnInfo a) (Maybe Int)

data PostgresConnInfoRaw

Postgres Connection info in the form of a templated URI string or structured data.

Constructors

PGConnDatabaseUrl URLTemplate

PGConnDetails PostgresConnDetailsRaw

mkUrlConnInfo :: String -> PostgresConnInfoRaw

_PGConnDatabaseUrl :: Prism' PostgresConnInfoRaw URLTemplate

_PGConnDetails :: Prism' PostgresConnInfoRaw PostgresConnDetailsRaw

rawConnDetailsToUrl :: PostgresConnDetailsRaw -> URLTemplate

data PostgresConnDetailsRaw

Structured Postgres connection information as provided by the arg parser or env vars.

Constructors

PostgresConnDetailsRaw

Fields

connHost :: String

connPort :: Int

connUser :: String

connPassword :: String

connDatabase :: String

connOptions :: Maybe String

rawConnDetailsToUrlText :: PostgresConnDetailsRaw -> Text

data HGECommand a

The HGE Arg parser Command choices.

This is polymorphic so that we can pack either ServeOptionsRaw or ProServeOptionsRaw in it.

Constructors

HCServe a

HCExport

HCClean

HCVersion

HCDowngrade !DowngradeOptions

_HCServe :: Prism' (HGECommand a) a

data ServeOptionsRaw impl

The Serve Command options accumulated from the Arg and Env parsers.

NOTE: A Nothing value indicates the absence of a particular flag. Hence types such as 'Maybe (HashSet X)' or 'Maybe Bool'.

Constructors

ServeOptionsRaw

Fields

rsoPort :: Maybe Port

rsoHost :: Maybe HostPreference

rsoConnParams :: ConnParamsRaw

rsoTxIso :: Maybe TxIsolation

rsoAdminSecret :: Maybe AdminSecretHash

rsoAuthHook :: AuthHookRaw

rsoJwtSecret :: Maybe JWTConfig

rsoUnAuthRole :: Maybe RoleName

rsoCorsConfig :: Maybe CorsConfig

rsoEnableConsole :: Bool

rsoConsoleAssetsDir :: Maybe Text

rsoEnableTelemetry :: Maybe Bool

rsoWsReadCookie :: Bool

rsoStringifyNum :: StringifyNumbers

rsoDangerousBooleanCollapse :: Maybe DangerouslyCollapseBooleans

rsoEnabledAPIs :: Maybe (HashSet API)

rsoMxRefetchInt :: Maybe RefetchInterval

rsoMxBatchSize :: Maybe BatchSize

rsoStreamingMxRefetchInt :: Maybe RefetchInterval

rsoStreamingMxBatchSize :: Maybe BatchSize

rsoEnableAllowlist :: Bool

rsoEnabledLogTypes :: Maybe (HashSet (EngineLogType impl))

rsoLogLevel :: Maybe LogLevel

rsoDevMode :: Bool

rsoAdminInternalErrors :: Maybe Bool

rsoEventsHttpPoolSize :: Maybe PositiveInt

rsoEventsFetchInterval :: Maybe (NonNegative Milliseconds)

rsoAsyncActionsFetchInterval :: Maybe OptionalInterval

rsoEnableRemoteSchemaPermissions :: RemoteSchemaPermissions

rsoWebSocketCompression :: Bool

rsoWebSocketKeepAlive :: Maybe KeepAliveDelay

rsoInferFunctionPermissions :: Maybe InferFunctionPermissions

rsoEnableMaintenanceMode :: MaintenanceMode ()

rsoSchemaPollInterval :: Maybe OptionalInterval

rsoExperimentalFeatures :: Maybe (HashSet ExperimentalFeature)

See Note '$experimentalFeatures' at bottom of module

rsoEventsFetchBatchSize :: Maybe NonNegativeInt

rsoGracefulShutdownTimeout :: Maybe (NonNegative Seconds)

rsoWebSocketConnectionInitTimeout :: Maybe WSConnectionInitTimeout

rsoEnableMetadataQueryLoggingEnv :: MetadataQueryLoggingMode

rsoDefaultNamingConvention :: Maybe NamingCase

stores global default naming convention

rsoExtensionsSchema :: Maybe ExtensionsSchema

newtype Port

An Int representing a Port number in the range 0 to 65536.

Constructors

Port

Fields

_getPort :: Int

mkPort :: Int -> Maybe Port

unsafePort :: Int -> Port

data API

Constructors

METADATA

GRAPHQL

PGDUMP

DEVELOPER

CONFIG

METRICS

data AuthHookRaw

Constructors

AuthHookRaw

Fields

ahrUrl :: Maybe Text

ahrType :: Maybe AuthHookType

data OptionalInterval

Sleep time interval for recurring activities such as (@asyncActionsProcessor) Presently msToOptionalInterval interprets `0` as Skip.

Constructors

Skip
No polling

Interval (NonNegative Milliseconds)
Interval time

msToOptionalInterval :: NonNegative Milliseconds -> OptionalInterval

data ConnParamsRaw

The Raw configuration data from the Arg and Env parsers needed to construct a ConnParams

Constructors

ConnParamsRaw

Fields

rcpStripes :: Maybe NonNegativeInt

rcpConns :: Maybe NonNegativeInt

rcpIdleTime :: Maybe NonNegativeInt

rcpConnLifetime :: Maybe (NonNegative NominalDiffTime)

Time from connection creation after which to destroy a connection and choose a different/new one.

rcpAllowPrepare :: Maybe Bool

rcpPoolTimeout :: Maybe (NonNegative NominalDiffTime)

See HASURA_GRAPHQL_PG_POOL_TIMEOUT

newtype KeepAliveDelay

Constructors

KeepAliveDelay

Fields

unKeepAliveDelay :: NonNegative Seconds

newtype WSConnectionInitTimeout

The timeout duration in Seconds for a WebSocket connection.

Constructors

WSConnectionInitTimeout

Fields

unWSConnectionInitTimeout :: NonNegative Seconds

data ServeOptions impl

The final Serve Command options accummulated from the Arg Parser and the Environment, fully processed and ready to apply when running the server.

Constructors

ServeOptions

Fields

soPort :: Port

soHost :: HostPreference

soConnParams :: ConnParams

soTxIso :: TxIsolation

soAdminSecret :: HashSet AdminSecretHash

soAuthHook :: Maybe AuthHook

soJwtSecret :: [JWTConfig]

soUnAuthRole :: Maybe RoleName

soCorsConfig :: CorsConfig

soEnableConsole :: Bool

soConsoleAssetsDir :: Maybe Text

soEnableTelemetry :: Bool

soStringifyNum :: StringifyNumbers

soDangerousBooleanCollapse :: DangerouslyCollapseBooleans

soEnabledAPIs :: HashSet API

soLiveQueryOpts :: LiveQueriesOptions

soStreamingQueryOpts :: StreamQueriesOptions

soEnableAllowlist :: Bool

soEnabledLogTypes :: HashSet (EngineLogType impl)

soLogLevel :: LogLevel

soResponseInternalErrorsConfig :: ResponseInternalErrorsConfig

soEventsHttpPoolSize :: PositiveInt

soEventsFetchInterval :: NonNegative Milliseconds

soAsyncActionsFetchInterval :: OptionalInterval

soEnableRemoteSchemaPermissions :: RemoteSchemaPermissions

soConnectionOptions :: ConnectionOptions

soWebSocketKeepAlive :: KeepAliveDelay

soInferFunctionPermissions :: InferFunctionPermissions

soEnableMaintenanceMode :: MaintenanceMode ()

soSchemaPollInterval :: OptionalInterval

soExperimentalFeatures :: HashSet ExperimentalFeature

See note '$experimentalFeatures'

soEventsFetchBatchSize :: NonNegativeInt

soDevMode :: Bool

soGracefulShutdownTimeout :: NonNegative Seconds

soWebSocketConnectionInitTimeout :: WSConnectionInitTimeout

soEventingMode :: EventingMode

soReadOnlyMode :: ReadOnlyMode

See note '$readOnlyMode'

soEnableMetadataQueryLogging :: MetadataQueryLoggingMode

soDefaultNamingConvention :: Maybe NamingCase

soExtensionsSchema :: ExtensionsSchema

data ResponseInternalErrorsConfig

ResponseInternalErrorsConfig represents the encoding of the internal errors in the response to the client.

For more details, see this github comment: https://github.com/hasura/graphql-engine/issues/4031#issuecomment-609747705

Constructors

InternalErrorsAllRequests

InternalErrorsAdminOnly

InternalErrorsDisabled

shouldIncludeInternal :: RoleName -> ResponseInternalErrorsConfig -> Bool

data DowngradeOptions

The Downgrade Command options. These are only sourced from the Arg Parser and are used directly in Migrate.

Constructors

DowngradeOptions

Fields

dgoTargetVersion :: Text

dgoDryRun :: Bool

Note Experimental features:

The graphql-engine accepts a list of experimental features that can be enabled at the startup. Experimental features are a way to introduce new, but not stable features to our users in a manner in which they have the choice to enable or disable a certain feature(s).

The objective of an experimental feature should be that when the feature is disabled, the graphql-engine should work the same way as it worked before adding the said feature.

The experimental feature's flag is `--experimental-features` and the corresponding environment variable is HASURA_GRAPHQL_EXPERIMENTAL_FEATURES which expects a comma-seperated value.

When an experimental feature is stable enough i.e. it's stable through multiple non-beta releases then we make the feature not experimental i.e. it will always be enabled. Note that when we do this we still have to support parsing of the experimental feature because users of the previous version will have it enabled and when they upgrade an error should not be thrown at the startup. For example:

The inherited roles was an experimental feature when introduced and it was enabled by setting `--experimental-features` to inherited_roles and then it was decided to make the inherited roles a stable feature, so it was removed as an experimental feature but the code was modified such that `--experimental-features inherited_roles` to not throw an error.

Note ReadOnly Mode:

This mode starts the server in a (database) read-only mode. That is, only read-only queries are allowed on users' database sources, and write queries throw a runtime error. The use-case is for failsafe operations. Metadata APIs are also disabled.

Following is the precise behaviour - 1. For any GraphQL API (relayhasura; httpwebsocket) - disable execution of mutations 2. Metadata API is disabled 3. v2query API - insert, delete, update, run_sql are disabled 4. v1query API - insert, delete, update, run_sql are disabled 5. No source catalog migrations are run 6. During build schema cache phase, building event triggers are disabled (as they create corresponding database triggers)