Hasura.Server.Init.Env

Documentation

considerEnv :: FromEnv a => String -> WithEnv (Maybe a)

Lookup a key in the application environment then parse the value with a FromEnv instance'

considerEnvs :: FromEnv a => [String] -> WithEnv (Maybe a)

Lookup a list of keys with considerEnv and return the first value to parse successfully.

withEnv :: FromEnv a => Maybe a -> String -> WithEnv (Maybe a)

withEnvs :: FromEnv a => Maybe a -> [String] -> WithEnv (Maybe a)

Return a if Just or else call considerEnv with a list of env keys k

withEnvBool :: Bool -> String -> WithEnv Bool

If bVal is True then return it, else lookup the envVar in the environment and return False if absent.

withEnvList :: (FromEnv b, Coercible b [a]) => Proxy [a] -> b -> String -> WithEnv b

withEnv for types isomorphic to an array. Proxy is generally used to pass on the type info. Here Proxy helps us to specify what the underlying array type should be.

class FromEnv a where

A Read style parser used for consuming envvars and building ReadM parsers for 'optparse-applicative'.

Methods

fromEnv :: String -> Either String a

type WithEnv a = ReaderT [(String, String)] (ExceptT String Identity) a

runWithEnv :: [(String, String)] -> WithEnv a -> Either String a

Given an environment run a WithEnv action producing either a parse error or an a.