Hasura.Server.Limits

Documentation

class Monad m => HasResourceLimits (m :: Type -> Type) where

Monads which support resource (memory, CPU time, etc.) limiting

Minimal complete definition

Nothing

Methods

askHTTPHandlerLimit :: m ResourceLimits

default askHTTPHandlerLimit :: forall (t :: (Type -> Type) -> Type -> Type) (n :: Type -> Type). (m ~ t n, MonadTrans t, HasResourceLimits n) => m ResourceLimits

askGraphqlOperationLimit :: RequestId -> UserInfo -> ApiLimit -> m ResourceLimits

default askGraphqlOperationLimit :: forall (t :: (Type -> Type) -> Type -> Type) (n :: Type -> Type). (m ~ t n, MonadTrans t, HasResourceLimits n) => RequestId -> UserInfo -> ApiLimit -> m ResourceLimits

data ResourceLimits

Resource limits, represented by a function which modifies IO actions to enforce those limits by throwing errors using MonadError in the case where they are exceeded.

Constructors

ResourceLimits

Fields

runResourceLimits :: forall (m :: Type -> Type) a. (MonadBaseControl IO m, MonadError QErr m) => m a -> m a