Hasura.Server.Logging

Description

This is taken from wai-logger and customised for our use

Documentation

data StartupLog

Constructors

StartupLog

Fields

slLogLevel :: !LogLevel

slKind :: !Text

slInfo :: !Value

data PGLog

Constructors

PGLog

Fields

plLogLevel :: !LogLevel

plMessage :: !Value

data RequestMode

whether a request is executed in batched mode or not

Constructors

RequestModeBatched
this request is batched

RequestModeSingle
this is a single request

RequestModeNonBatchable
this request is of a kind for which batching is not done or does not make sense

RequestModeError
the execution of this request failed

mkInconsMetadataLog :: [InconsistentMetadata] -> MetadataLog

mkHttpAccessLogContext

Arguments

:: Maybe UserInfo
Maybe because it may not have been resolved

-> LoggingSettings

-> RequestId

-> Request

-> (ByteString, Maybe Value)

-> Int64
Size of response body, before compression

-> ByteString

-> Maybe (DiffTime, DiffTime)

-> Maybe CompressionType

-> [Header]

-> RequestMode

-> Maybe (GQLBatchedReqs GQLBatchQueryOperationLog)

-> HttpLogContext

mkHttpErrorLogContext

Arguments

:: Maybe UserInfo
Maybe because it may not have been resolved

-> LoggingSettings

-> RequestId

-> Request

-> (ByteString, Maybe Value)

-> QErr

-> Maybe (DiffTime, DiffTime)

-> Maybe CompressionType

-> [Header]

-> HttpLogContext

mkHttpLog :: HttpLogContext -> HttpLogLine

data HttpInfoLog

Log information about the HTTP request

Constructors

HttpInfoLog

Fields

hlStatus :: !Status

hlMethod :: !Text

hlSource :: !IpAddress

hlPath :: !Text

hlHttpVersion :: !HttpVersion

hlCompression :: !(Maybe CompressionType)

hlHeaders :: ![Header]

all the request headers

data OperationLog

Information about a GraphQL/Hasura metadata operation over HTTP

Constructors

OperationLog

Fields

olRequestId :: !RequestId

olUserVars :: !(Maybe SessionVariables)

olResponseSize :: !(Maybe Int64)

olUncompressedResponseSize :: !Int64

Response size before compression

olRequestReadTime :: !(Maybe Seconds)

Request IO wait time, i.e. time spent reading the full request from the socket.

olQueryExecutionTime :: !(Maybe Seconds)

Service time, not including request IO wait time.

olQuery :: !(Maybe Value)

olRawQuery :: !(Maybe Text)

olError :: !(Maybe QErr)

olRequestMode :: !RequestMode

data HttpLogContext

Constructors

HttpLogContext

Fields

hlcHttpInfo :: !HttpInfoLog

hlcOperation :: !OperationLog

hlcRequestId :: !RequestId

hlcBatchedOperations :: !(Maybe (NonEmpty BatchOperationLog))

data WebHookLog

Constructors

WebHookLog

Fields

whlLogLevel :: !LogLevel

whlStatusCode :: !(Maybe Status)

whlUrl :: !Text

whlMethod :: !StdMethod

whlError :: !(Maybe HttpException)

whlResponse :: !(Maybe Text)

whlMessage :: !(Maybe Text)

data HttpException

class Monad m => HttpLog (m :: Type -> Type) where

Associated Types

type ExtraHttpLogMetadata (m :: Type -> Type)

Extra http-log metadata that we attach when operating in m.

Methods

emptyExtraHttpLogMetadata :: ExtraHttpLogMetadata m

buildExtraHttpLogMetadata :: ParameterizedQueryHashList -> ExtraUserInfo -> ExtraHttpLogMetadata m

logHttpError

Arguments

:: Logger Hasura
the logger

-> LoggingSettings
setting used to control the information in logs

-> Maybe UserInfo
user info may or may not be present (error can happen during user resolution)

-> RequestId
request id of the request

-> Request
the Wai.Request object

-> (ByteString, Maybe Value)
the request body and parsed request

-> QErr
the error

-> [Header]
list of request headers

-> HttpLogMetadata m

-> Bool
flag to indicate if the request/response size should be added to the Prometheus Counter

-> m ()

logHttpSuccess

Arguments

:: Logger Hasura
the logger

-> LoggingSettings
setting used to control the information in logs

-> Maybe UserInfo
user info may or may not be present (error can happen during user resolution)

-> RequestId
request id of the request

-> Request
the Wai.Request object

-> (ByteString, Maybe Value)
the request body and parsed request

-> ByteString
the response bytes

-> ByteString
the compressed response bytes ^ TODO (from master): make the above two type represented

-> Maybe (DiffTime, DiffTime)
IO/network wait time and service time (respectively) for this request, if available.

-> Maybe CompressionType
possible compression type

-> [Header]
list of request headers

-> HttpLogMetadata m

-> Bool
flag to indicate if the request/response size should be added to the Prometheus Counter

-> m ()

data GQLBatchQueryOperationLog

Constructors

GQLQueryOperationSuccess !GQLQueryOperationSuccessLog

GQLQueryOperationError !GQLQueryOperationErrorLog

data GQLQueryOperationSuccessLog

GQLQueryOperationSuccessLog captures all the data required to construct an HTTP success log.

Constructors

GQLQueryOperationSuccessLog

Fields

gqolQuery :: !GQLReqUnparsed

gqolQueryExecutionTime :: !DiffTime

gqolResponseSize :: !Int64

gqolRequestSize :: !Int64

gqolParameterizedQueryHash :: !ParameterizedQueryHash

data GQLQueryOperationErrorLog

GQLQueryOperationErrorLog captures the request along with the error message

Constructors

GQLQueryOperationErrorLog

Fields

gqelQuery :: !GQLReqUnparsed

gqelError :: !QErr

data MetadataLog

Constructors

MetadataLog

Fields

mlLogLevel :: !LogLevel

mlMessage :: !Text

mlInfo :: !Value

newtype EnvVarsMovedToMetadata

The environment variables that were moved to metadata. These environment variables are available if a v1 hasura project is run an v2 hasura server. These environment variables are marked as deprecated only when the v1 hasura project is migrated to v2 project.

Constructors

EnvVarsMovedToMetadata

Fields

unEnvVarsMovedToMetadata :: [String]

newtype DeprecatedEnvVars

These env vars are completely deprecated

Constructors

DeprecatedEnvVars

Fields

unDeprecatedEnvVars :: [String]

logDeprecatedEnvVars :: Logger Hasura -> Environment -> SourceCache -> IO ()

Log warning messages for deprecated environment variables

data CommonHttpLogMetadata

Constructors

CommonHttpLogMetadata

Fields

_chlmRequestMode :: !RequestMode

_chlmBatchOperationLog :: !(Maybe (GQLBatchedReqs GQLBatchQueryOperationLog))

type HttpLogMetadata (m :: Type -> Type) = (CommonHttpLogMetadata, ExtraHttpLogMetadata m)

The http-log metadata attached to HTTP requests running in the monad m, split into a common portion that is present regardless of m, and a monad-specific one defined in the HttpLog instance.

This allows us to not have to duplicate the code that generates the common part of the metadata across OSS and Pro, so that instances only have to implement the part of it unique to them.

buildHttpLogMetadata :: forall (m :: Type -> Type). HttpLog m => HttpLogGraphQLInfo -> ExtraUserInfo -> HttpLogMetadata m

emptyHttpLogMetadata :: forall (m :: Type -> Type). HttpLog m => HttpLogMetadata m

synonym for clarity, writing `emptyHttpLogMetadata m` instead of `def (HttpLogMetadata m)`

data MetadataQueryLoggingMode

Constructors

MetadataQueryLoggingEnabled

MetadataQueryLoggingDisabled

data HttpLogQueryOnlyOnError

Constructors

HttpLogQueryOnlyOnErrorEnabled

HttpLogQueryOnlyOnErrorDisabled

data LoggingSettings

Setting used to control the information in logs

Constructors

LoggingSettings

Fields

_lsEnabledLogTypes :: HashSet (EngineLogType Hasura)

this is only required for the short-term fix in https://github.com/hasura/graphql-engine-mono/issues/1770 See Note [Disable query printing when query-log is disabled]

_lsMetadataQueryLoggingMode :: MetadataQueryLoggingMode

_lsHttpLogQueryOnlyOnError :: HttpLogQueryOnlyOnError

data SchemaSyncThreadType

Constructors

TTListener

TTProcessor

TTMetadataApi

data SchemaSyncLog

Constructors

SchemaSyncLog

Fields

sslLogLevel :: !LogLevel

sslThreadType :: !SchemaSyncThreadType

sslInfo :: !Value

type HttpLogGraphQLInfo = (CommonHttpLogMetadata, ParameterizedQueryHashList)

emptyHttpLogGraphQLInfo :: HttpLogGraphQLInfo

data ModelInfo

Constructors

ModelInfo

Fields

miModelName :: !Text

miModelType :: !Text

miSourceName :: !(Maybe Text)

miSourceType :: !(Maybe Text)

miQueryType :: !Text

miIsCached :: !Bool

data ModelInfoLog

Constructors

ModelInfoLog

Fields

milLogType :: !LogLevel

milModelInfo :: !ModelInfo