Hasura.Server.Migrate

Description

Migrations for the Hasura catalog.

To add a new migration:

Bump the catalog version number in src-rsr/catalog_version.txt.

Add a migration script in the src-rsrmigrations directory with the name version_to_version.sql.

Create a downgrade script in the src-rsrmigrations directory with the name version_to_version.sql.

If making a new release, add the mapping from application version to catalog schema version in src-rsr/catalog_versions.txt.

If appropriate, add the change to serversrc-rsrinitialise.sql for fresh installations of hasura.

The Template Haskell code in this module will automatically compile the new migration script into the graphql-engine executable.

NOTE: Please have a look at the `serverdocumentationmigration-guidelines.md` before adding any new migration if you haven't already looked at it

Documentation

data MigrationResult

Constructors

MRNothingToDo

MRInitialized

MRMigrated Text
old catalog version

MRMaintanenceMode

migrateCatalog :: (MonadTx m, MonadIO m, MonadBaseControl IO m) => Maybe (SourceConnConfiguration ('Postgres 'Vanilla)) -> ExtensionsSchema -> MaintenanceMode () -> UTCTime -> m (MigrationResult, MetadataWithResourceVersion)

latestCatalogVersion :: MetadataCatalogVersion

The current catalog schema version. We store this in a file because we want to append the current verson to the catalog_versions file when tagging a new release, in tag-release.sh.

downgradeCatalog :: (MonadIO m, MonadTx m) => Maybe (SourceConnConfiguration ('Postgres 'Vanilla)) -> DowngradeOptions -> UTCTime -> m MigrationResult