Hasura.Server.OpenAPI

Description

The implementation currently iterates over the endpoints building up EndpointData for each then exposes this as an OpenAPI Schema.

Most functions are in the Declare monad so that they can add new component definitions on the fly that can be referenced. This is especially useful for the params and request body documentation.

The response body recurses over the SelectionSet Fields associated with an endpoint and looks up types by name in a RemoteSchemaIntrospection result generated from the SchemaCache.

Response bodies are mostly delcared inline, since the associated query will likely be unique and determine the fields contained in the response.

Documentation

data EndpointData

Constructors

EndpointData

Fields

_edUrl :: String

_edMethod :: [Text]

_edVarList :: [Referenced Param]

_edProperties :: InsOrdHashMap Text (Referenced Schema)

_edResponse :: Maybe Response

_edDescription :: Text

_edName :: Text

Response Body Related Functions.

mdDefinitions :: EndpointMetadata GQLQueryWithText -> [ExecutableDefinition Name]

mkResponse :: [Text] -> String -> Maybe RemoteSchemaIntrospection -> Analysis Name -> Declare (Definitions Schema) (Maybe Response)

getSelectionSchema :: RemoteSchemaIntrospection -> [(Name, (FieldDef, Maybe (FieldAnalysis var)))] -> Declare (Definitions Schema) Schema

setPattern :: Maybe Pattern -> Schema -> Schema

A helper function to set the pattern field in Schema Why not lens pattern? hlint doesn't like the name pattern https://github.com/ndmitchell/hlint/issues/607

getDefinitionSchema :: RemoteSchemaIntrospection -> Name -> FieldDef -> Maybe (FieldAnalysis var) -> Declare (Definitions Schema) Schema

typeToSchemaM :: Monad m => FieldDef -> (TypeDefinition [Name] RemoteSchemaInputValueDefinition -> m Schema) -> m Schema

(&&&&) :: Applicative f => (t -> f a1) -> (t -> f a2) -> t -> f (a1, a2) infixl 7

URL / Query Params and Request Body Functions

getParams :: Analysis Name -> EndpointUrl -> [Referenced Param]

getType :: GType -> Either GType (OpenApiType, Maybe Pattern)

mkProperties :: Maybe RemoteSchemaIntrospection -> Analysis Name -> Declare (Definitions Schema) (InsOrdHashMap Text (Referenced Schema))

mkProperty :: InsOrdHashMap Text (TypeDefinition [Name] RemoteSchemaInputValueDefinition) -> (Name, (GType, Maybe (Value Void))) -> Declare (Definitions Schema) (Text, Referenced Schema)

handleRefType :: InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> GType -> Declare (Definitions Schema) (Referenced Schema)

declareReference :: Nullability -> Text -> InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> Declare (Definitions Schema) ()

referenceAvailable :: Text -> DeclareT (Definitions Schema) Identity Bool

getPropertyReferences :: InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> Maybe [RemoteSchemaInputValueDefinition] -> InsOrdHashMap Text (Referenced Schema)

processProperty' :: InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> RemoteSchemaInputValueDefinition -> (Text, Referenced Schema)

handleRefType' :: InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> GType -> Referenced Schema

typeDescription :: TypeDefinition possibleTypes inputType -> Maybe Text

typeProperties :: TypeDefinition possibleTypes RemoteSchemaInputValueDefinition -> Maybe [RemoteSchemaInputValueDefinition]

referenceType :: Bool -> Text -> (Maybe OpenApiType, Maybe Pattern)

processProperties :: InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> Maybe [RemoteSchemaInputValueDefinition] -> DeclareT (Definitions Schema) Identity (InsOrdHashMap Text (Referenced Schema))

processProperty :: InsOrdHashMap Text (TypeDefinition a RemoteSchemaInputValueDefinition) -> RemoteSchemaInputValueDefinition -> DeclareT (Definitions Schema) Identity (Text, Referenced Schema)

getGQLQueryFromTrie :: EndpointMetadata GQLQueryWithText -> Text

mkParam :: Text -> Maybe Text -> Maybe Bool -> ParamLocation -> Maybe Bool -> Maybe Value -> Maybe OpenApiType -> Maybe Pattern -> Param

gqlToJsonValue :: Value Void -> Value

Top level schema construction

getComment :: EndpointMetadata GQLQueryWithText -> Text

getURL :: EndpointMetadata GQLQueryWithText -> Text

extractEndpointInfo :: Maybe RemoteSchemaIntrospection -> EndpointMethod -> EndpointMetadata GQLQueryWithText -> Declare (Definitions Schema) EndpointData

getEndpointsData :: Maybe RemoteSchemaIntrospection -> SchemaCache -> Declare (Definitions Schema) [EndpointData]

squashEndpointGroup :: NonEmpty EndpointData -> EndpointData

serveJSON :: SchemaCache -> OpenApi

isRequestBodyRequired :: EndpointData -> Bool

If all variables are scalar or optional then the entire request body can be marked as optional

Entry point

declareOpenApiSpec :: SchemaCache -> Declare (Definitions Schema) OpenApi