Hasura.Server.Prometheus

Description

Mutable references for Prometheus metrics.

These metrics are independent from the metrics in Hasura.Server.Metrics.

Documentation

data PrometheusMetrics

Mutable references for Prometheus metrics.

Constructors

PrometheusMetrics

Fields

pmConnections :: ConnectionsGauge

pmGraphQLRequestMetrics :: GraphQLRequestMetrics

pmEventTriggerMetrics :: EventTriggerMetrics

pmWebSocketBytesReceived :: Counter

pmWebSocketBytesSent :: CounterVector DynamicGraphqlOperationLabel

pmActionBytesReceived :: Counter

pmActionBytesSent :: Counter

pmScheduledTriggerMetrics :: ScheduledTriggerMetrics

pmSubscriptionMetrics :: SubscriptionMetrics

pmWebsocketMsgQueueTimeSeconds :: Histogram

pmWebsocketMsgWriteTimeSeconds :: Histogram

pmCacheRequestMetrics :: CacheRequestMetrics

pmOpenTelemetryMetrics :: OpenTelemetryMetrics

data GraphQLRequestMetrics

Constructors

GraphQLRequestMetrics

Fields

gqlRequests :: CounterVector GraphQLRequestsLabels

gqlExecutionTimeSecondsQuery :: Histogram

gqlExecutionTimeSecondsMutation :: Histogram

data EventTriggerMetrics

Constructors

EventTriggerMetrics

Fields

eventTriggerHTTPWorkers :: Gauge

eventsFetchedPerBatch :: Gauge

eventQueueTimeSeconds :: HistogramVector (Maybe DynamicEventTriggerLabel)

eventsFetchTimePerBatch :: Histogram

eventWebhookProcessingTime :: HistogramVector (Maybe DynamicEventTriggerLabel)

eventProcessingTime :: HistogramVector (Maybe DynamicEventTriggerLabel)

eventTriggerBytesReceived :: Counter

eventTriggerBytesSent :: Counter

eventProcessedTotal :: CounterVector EventStatusWithTriggerLabel

eventInvocationTotal :: CounterVector EventStatusWithTriggerLabel

data CacheRequestMetrics

Constructors

CacheRequestMetrics

Fields

crmCacheHits :: Counter

crmCacheMisses :: Counter

data OpenTelemetryMetrics

Metrics related to OTel telemetry export; for now the volume of logs and trace spans shipped, and counts of log lines and spans dropped due to high volume.

Constructors

OpenTelemetryMetrics

Fields

otmSentSpans :: Counter

otmDroppedSpansInBuffer :: Counter

Dropped due to the send buffer being full

otmDroppedSpansInSend :: Counter

Dropped due to some error (after retrying) when sending to collector

otmSentLogs :: Counter

otmDroppedLogsInBuffer :: Counter

otmDroppedLogsInSend :: Counter

makeDummyPrometheusMetrics :: IO PrometheusMetrics

Create dummy mutable references without associating them to a metrics store.

data ConnectionsGauge

A mutable reference for atomically sampling the number of websocket connections and number of threads forked by the warp webserver.

Because we derive the number of (non-websocket) HTTP connections by the difference of these two metrics, we must sample them simultaneously, otherwise we might report a negative number of HTTP connections.

data Connections

Constructors

Connections

Fields

connWarpThreads :: Int64

connWebsockets :: Int64

newConnectionsGauge :: IO ConnectionsGauge

readConnectionsGauge :: ConnectionsGauge -> IO Connections

incWarpThreads :: ConnectionsGauge -> IO ()

decWarpThreads :: ConnectionsGauge -> IO ()

incWebsocketConnections :: ConnectionsGauge -> IO ()

decWebsocketConnections :: ConnectionsGauge -> IO ()

data ScheduledTriggerMetrics

Constructors

ScheduledTriggerMetrics

Fields

stmScheduledTriggerBytesReceived :: Counter

stmScheduledTriggerBytesSent :: Counter

stmCronEventsInvocationTotalSuccess :: Counter

stmCronEventsInvocationTotalFailure :: Counter

stmOneOffEventsInvocationTotalSuccess :: Counter

stmOneOffEventsInvocationTotalFailure :: Counter

stmCronEventsProcessedTotalSuccess :: Counter

stmCronEventsProcessedTotalFailure :: Counter

stmOneOffEventsProcessedTotalSuccess :: Counter

stmOneOffEventsProcessedTotalFailure :: Counter

data SubscriptionMetrics

Constructors

SubscriptionMetrics

Fields

submActiveLiveQueryPollers :: Gauge

submActiveStreamingPollers :: Gauge

submActiveLiveQueryPollersInError :: Gauge

submActiveStreamingPollersInError :: Gauge

submTotalTime :: HistogramVector SubscriptionLabel

submDBExecTotalTime :: HistogramVector SubscriptionLabel

submActiveSubscriptions :: GaugeVector SubscriptionLabel

data DynamicEventTriggerLabel

Constructors

DynamicEventTriggerLabel

Fields

_detlTriggerName :: TriggerName

_detlSourceName :: SourceName

data ResponseStatus

Constructors

Success

Failed

responseStatusToLabelValue :: ResponseStatus -> Text

newtype EventStatusLabel

Constructors

EventStatusLabel

Fields

status :: Text

eventSuccessLabel :: EventStatusLabel

eventFailedLabel :: EventStatusLabel

data EventStatusWithTriggerLabel

Constructors

EventStatusWithTriggerLabel

Fields

_eswtlStatus :: EventStatusLabel

_eswtlDynamicLabels :: Maybe DynamicEventTriggerLabel

data GranularPrometheusMetricsState

Whether or not to enable granular metrics for Prometheus.

GranularMetricsOn will enable the dynamic labels for the metrics. GranularMetricsOff will disable the dynamic labels for the metrics.

*Warning**: Enabling dynamic labels for Prometheus metrics can cause cardinality issues and can cause memory usage to increase.

Constructors

GranularMetricsOff

GranularMetricsOn

observeHistogramWithLabel :: (Ord l, MonadIO m) => IO GranularPrometheusMetricsState -> Bool -> HistogramVector (Maybe l) -> l -> Double -> m ()

Observe a histogram metric with a label.

If the granularity is set to GranularMetricsOn, the label will be included in the metric. Otherwise, the label will be set to Nothing

data SubscriptionKindLabel

Constructors

SubscriptionKindLabel

Fields

subscription_kind :: Text

data SubscriptionLabel

Constructors

SubscriptionLabel

Fields

_slKind :: SubscriptionKindLabel

_slDynamicLabels :: Maybe DynamicGraphqlOperationLabel

data DynamicGraphqlOperationLabel

Constructors

DynamicGraphqlOperationLabel

Fields

_dslParamQueryHash :: Maybe ParameterizedQueryHash

_dslOperationName :: Maybe OperationName

streamingSubscriptionLabel :: SubscriptionKindLabel

liveQuerySubscriptionLabel :: SubscriptionKindLabel

recordMetricWithLabel :: MonadIO m => IO GranularPrometheusMetricsState -> Bool -> IO () -> IO () -> m ()

Record metrics with dynamic label

recordSubscriptionMetric :: MonadIO m => IO GranularPrometheusMetricsState -> Bool -> HashMap (Maybe OperationName) Int -> ParameterizedQueryHash -> SubscriptionKindLabel -> (SubscriptionLabel -> IO ()) -> m ()

Record a subscription metric for all the operation names present in the subscription. Use this when you want to update the same value of the metric for all the operation names.

data GraphQLRequestsLabels

recordGraphqlOperationMetric :: MonadIO m => IO GranularPrometheusMetricsState -> Maybe OperationType -> ResponseStatus -> Maybe OperationName -> Maybe ParameterizedQueryHash -> (GraphQLRequestsLabels -> IO ()) -> m ()