Hasura.Server.Prometheus

Description

Mutable references for Prometheus metrics.

These metrics are independent from the metrics in Hasura.Server.Metrics.

Documentation

data PrometheusMetrics

Mutable references for Prometheus metrics.

Constructors

PrometheusMetrics

Fields

pmConnections :: ConnectionsGauge

pmGraphQLRequestMetrics :: GraphQLRequestMetrics

pmEventTriggerMetrics :: EventTriggerMetrics

pmWebSocketBytesReceived :: Counter

pmWebSocketBytesSent :: Counter

pmActionBytesReceived :: Counter

pmActionBytesSent :: Counter

pmScheduledTriggerMetrics :: ScheduledTriggerMetrics

pmSubscriptionMetrics :: SubscriptionMetrics

pmWebsocketMsgQueueTimeSeconds :: Histogram

pmCacheRequestMetrics :: CacheRequestMetrics

data GraphQLRequestMetrics

Constructors

GraphQLRequestMetrics

Fields

gqlRequestsQuerySuccess :: Counter

gqlRequestsQueryFailure :: Counter

gqlRequestsMutationSuccess :: Counter

gqlRequestsMutationFailure :: Counter

gqlRequestsSubscriptionSuccess :: Counter

gqlRequestsSubscriptionFailure :: Counter

gqlRequestsUnknownFailure :: Counter

gqlExecutionTimeSecondsQuery :: Histogram

gqlExecutionTimeSecondsMutation :: Histogram

data EventTriggerMetrics

Constructors

EventTriggerMetrics

Fields

eventTriggerHTTPWorkers :: Gauge

eventsFetchedPerBatch :: Gauge

eventQueueTimeSeconds :: Histogram

eventsFetchTimePerBatch :: Histogram

eventWebhookProcessingTime :: Histogram

eventProcessingTime :: HistogramVector (Maybe TriggerNameLabel)

eventTriggerBytesReceived :: Counter

eventTriggerBytesSent :: Counter

eventProcessedTotalSuccess :: Counter

eventProcessedTotalFailure :: Counter

eventInvocationTotalSuccess :: Counter

eventInvocationTotalFailure :: Counter

data ScheduledTriggerMetrics

Constructors

ScheduledTriggerMetrics

Fields

stmScheduledTriggerBytesReceived :: Counter

stmScheduledTriggerBytesSent :: Counter

stmCronEventsInvocationTotalSuccess :: Counter

stmCronEventsInvocationTotalFailure :: Counter

stmOneOffEventsInvocationTotalSuccess :: Counter

stmOneOffEventsInvocationTotalFailure :: Counter

stmCronEventsProcessedTotalSuccess :: Counter

stmCronEventsProcessedTotalFailure :: Counter

stmOneOffEventsProcessedTotalSuccess :: Counter

stmOneOffEventsProcessedTotalFailure :: Counter

data SubscriptionMetrics

Constructors

SubscriptionMetrics

Fields

submActiveLiveQueryPollers :: Gauge

submActiveStreamingPollers :: Gauge

submActiveLiveQueryPollersInError :: Gauge

submActiveStreamingPollersInError :: Gauge

submTotalTime :: HistogramVector SubscriptionLabel

submDBExecTotalTime :: HistogramVector SubscriptionLabel

submActiveSubscriptions :: GaugeVector SubscriptionLabel

data CacheRequestMetrics

Constructors

CacheRequestMetrics

Fields

crmCacheHits :: Counter

crmCacheMisses :: Counter

makeDummyPrometheusMetrics :: IO PrometheusMetrics

Create dummy mutable references without associating them to a metrics store.

data ConnectionsGauge

A mutable reference for atomically sampling the number of websocket connections and number of threads forked by the warp webserver.

Because we derive the number of (non-websocket) HTTP connections by the difference of these two metrics, we must sample them simultaneously, otherwise we might report a negative number of HTTP connections.

data Connections

Constructors

Connections

Fields

connWarpThreads :: Int64

connWebsockets :: Int64

newConnectionsGauge :: IO ConnectionsGauge

readConnectionsGauge :: ConnectionsGauge -> IO Connections

incWarpThreads :: ConnectionsGauge -> IO ()

decWarpThreads :: ConnectionsGauge -> IO ()

incWebsocketConnections :: ConnectionsGauge -> IO ()

decWebsocketConnections :: ConnectionsGauge -> IO ()

newtype TriggerNameLabel

Constructors

TriggerNameLabel TriggerName

data SubscriptionKindLabel

Constructors

SubscriptionKindLabel

Fields

subscription_kind :: Text

streamingSubscriptionLabel :: SubscriptionKindLabel

liveQuerySubscriptionLabel :: SubscriptionKindLabel

data DynamicSubscriptionLabel

Constructors

DynamicSubscriptionLabel

Fields

_dslParamQueryHash :: ParameterizedQueryHash

_dslOperationName :: Maybe OperationName

data SubscriptionLabel

Constructors

SubscriptionLabel

Fields

_slKind :: SubscriptionKindLabel

_slDynamicLabels :: Maybe DynamicSubscriptionLabel

recordMetricWithLabel :: MonadIO m => IO GranularPrometheusMetricsState -> Bool -> IO () -> IO () -> m ()

Record metrics with dynamic label

observeHistogramWithLabel :: (Ord l, MonadIO m) => IO GranularPrometheusMetricsState -> Bool -> HistogramVector (Maybe l) -> l -> Double -> m ()

Observe a histogram metric with a label.

If the granularity is set to GranularMetricsOn, the label will be included in the metric. Otherwise, the label will be set to Nothing

recordSubcriptionMetric :: MonadIO m => IO GranularPrometheusMetricsState -> Bool -> HashMap (Maybe OperationName) Int -> ParameterizedQueryHash -> SubscriptionKindLabel -> (SubscriptionLabel -> IO ()) -> m ()

Record a subscription metric for all the operation names present in the subscription. Use this when you want to update the same value of the metric for all the operation names.