Hasura.Server.Prometheus

Description

Mutable references for Prometheus metrics.

These metrics are independent from the metrics in Hasura.Server.Metrics.

Documentation

data PrometheusMetrics

Mutable references for Prometheus metrics.

Constructors

PrometheusMetrics

Fields

pmConnections :: ConnectionsGauge

pmActiveSubscriptions :: Gauge

pmGraphQLRequestMetrics :: GraphQLRequestMetrics

pmEventTriggerMetrics :: EventTriggerMetrics

data GraphQLRequestMetrics

Constructors

GraphQLRequestMetrics

Fields

gqlRequestsQuerySuccess :: Counter

gqlRequestsQueryFailure :: Counter

gqlRequestsMutationSuccess :: Counter

gqlRequestsMutationFailure :: Counter

gqlRequestsUnknownFailure :: Counter

gqlExecutionTimeSecondsQuery :: Histogram

gqlExecutionTimeSecondsMutation :: Histogram

data EventTriggerMetrics

Constructors

EventTriggerMetrics

Fields

eventTriggerHTTPWorkers :: Gauge

eventQueueTimeSeconds :: Histogram

makeDummyPrometheusMetrics :: IO PrometheusMetrics

Create dummy mutable references without associating them to a metrics store.

data ConnectionsGauge

A mutable reference for atomically sampling the number of websocket connections and number of threads forked by the warp webserver.

Because we derive the number of (non-websocket) HTTP connections by the difference of these two metrics, we must sample them simultaneously, otherwise we might report a negative number of HTTP connections.

data Connections

Constructors

Connections

Fields

connWarpThreads :: Int64

connWebsockets :: Int64

newConnectionsGauge :: IO ConnectionsGauge

readConnectionsGauge :: ConnectionsGauge -> IO Connections

incWarpThreads :: ConnectionsGauge -> IO ()

decWarpThreads :: ConnectionsGauge -> IO ()

incWebsocketConnections :: ConnectionsGauge -> IO ()

decWebsocketConnections :: ConnectionsGauge -> IO ()