Hasura.Server.Rest

Documentation

runCustomEndpoint :: (MonadIO m, MonadError QErr m, MonadTrace m, MonadBaseControl IO m, MonadGQLExecutionCheck m, MonadQueryLog m, MonadExecutionLog m, MonadExecuteQuery m, MonadMetadataStorage m, MonadQueryTags m, HasResourceLimits m, ProvidesNetwork m, MonadGetPolicies m) => Environment -> SQLGenCtx -> SchemaCache -> AllowListStatus -> ReadOnlyMode -> TraceQueryStatus -> PrometheusMetrics -> Logger Hasura -> Maybe (CredentialCache AgentLicenseKey) -> RequestId -> UserInfo -> [Header] -> IpAddress -> RestRequest EndpointMethod -> EndpointTrie GQLQueryWithText -> m (HttpLogGraphQLInfo, HttpResponse EncJSON)

Implements all the custom endpoints by looking up the path/methods in the endpoint trie and delegating to the graphql handler.

data RestRequest method

Constructors

RestRequest

Fields

reqPath :: Text

Remainder of the url path after `api/rest`

reqMethod :: method

reqArgs :: [(Text, Either Text Value)]

URL Query/Request Body Arguments