Hasura.Server.SchemaUpdate

Documentation

data ThreadType

Constructors

TTListener

TTProcessor

data ThreadError

Constructors

TEPayloadParse !Text

TEQueryError !QErr

startSchemaSyncListenerThread :: ForkableMonadIO m => Logger Hasura -> PGPool -> InstanceId -> Refined NonNegative Milliseconds -> TMVar MetadataResourceVersion -> ManagedT m Thread

An async thread which listen to Postgres notify to enable schema syncing See Note [Schema Cache Sync]

startSchemaSyncProcessorThread :: (ForkableMonadIO m, MonadMetadataStorage (MetadataStorageT m), MonadResolveSource m) => Logger Hasura -> Manager -> TMVar MetadataResourceVersion -> SchemaCacheRef -> InstanceId -> ServerConfigCtx -> TVar Bool -> ManagedT m Thread

An async thread which processes the schema sync events See Note [Schema Cache Sync]