Hasura.Server.SchemaUpdate

Documentation

data ThreadType

Constructors

TTListener

TTProcessor

data SchemaSyncThreadLog

Constructors

SchemaSyncThreadLog

Fields

suelLogLevel :: !LogLevel

suelThreadType :: !ThreadType

suelInfo :: !Value

data ThreadError

Constructors

TEPayloadParse !Text

TEQueryError !QErr

logThreadStarted :: MonadIO m => Logger Hasura -> InstanceId -> ThreadType -> Thread -> m ()

startSchemaSyncListenerThread :: ForkableMonadIO m => Logger Hasura -> PGPool -> InstanceId -> Milliseconds -> TMVar MetadataResourceVersion -> ManagedT m Thread

An async thread which listen to Postgres notify to enable schema syncing See Note [Schema Cache Sync]

startSchemaSyncProcessorThread :: (ForkableMonadIO m, MonadMetadataStorage (MetadataStorageT m), MonadResolveSource m) => Logger Hasura -> Manager -> TMVar MetadataResourceVersion -> SchemaCacheRef -> InstanceId -> ServerConfigCtx -> TVar Bool -> ManagedT m Thread

An async thread which processes the schema sync events See Note [Schema Cache Sync]

forcePut :: TMVar a -> a -> IO ()

schemaVersionCheckHandler :: PGPool -> TMVar MetadataResourceVersion -> IO (Either QErr ())

data ErrorState

Constructors

ErrorState

Fields

_esLastErrorSeen :: !(Maybe QErr)

_esLastMetadataVersion :: !(Maybe MetadataResourceVersion)

defaultErrorState :: ErrorState

updateErrorInState :: ErrorState -> QErr -> MetadataResourceVersion -> ErrorState

NOTE: this can be updated to use lenses

isInErrorState :: ErrorState -> Bool

toLogError :: ErrorState -> QErr -> MetadataResourceVersion -> Bool

listener :: MonadIO m => Logger Hasura -> PGPool -> TMVar MetadataResourceVersion -> Milliseconds -> m void

An IO action that listens to postgres for events and pushes them to a Queue, in a loop forever.

processor :: forall m void. (ForkableMonadIO m, MonadMetadataStorage (MetadataStorageT m), MonadResolveSource m) => Logger Hasura -> Manager -> TMVar MetadataResourceVersion -> SchemaCacheRef -> InstanceId -> ServerConfigCtx -> TVar Bool -> m void

An IO action that processes events from Queue, in a loop forever.

refreshSchemaCache :: (MonadIO m, MonadBaseControl IO m, MonadMetadataStorage m, MonadResolveSource m) => MetadataResourceVersion -> InstanceId -> Logger Hasura -> Manager -> SchemaCacheRef -> ThreadType -> ServerConfigCtx -> TVar Bool -> m ()

logInfo :: MonadIO m => Logger Hasura -> ThreadType -> Value -> m ()

logError :: (MonadIO m, ToJSON a) => Logger Hasura -> ThreadType -> a -> m ()

logDebug :: MonadIO m => Logger Hasura -> ThreadType -> String -> m ()