Hasura.Server.Telemetry

Description

Send anonymized metrics to the telemetry server regarding usage of various features of Hasura.

Documentation

data RelationshipMetric

Constructors

RelationshipMetric

Fields

_rmManual :: !Int

_rmAuto :: !Int

data PermissionMetric

Constructors

PermissionMetric

Fields

_pmSelect :: !Int

_pmInsert :: !Int

_pmUpdate :: !Int

_pmDelete :: !Int

_pmRoles :: !Int

data ActionMetric

Constructors

ActionMetric

Fields

_amSynchronous :: !Int

_amAsynchronous :: !Int

_amQueryActions :: !Int

_amTypeRelationships :: !Int

_amCustomTypes :: !Int

data Metrics

Constructors

Metrics

Fields

_mtTables :: !Int

_mtViews :: !Int

_mtEnumTables :: !Int

_mtRelationships :: !RelationshipMetric

_mtPermissions :: !PermissionMetric

_mtEventTriggers :: !Int

_mtRemoteSchemas :: !Int

_mtFunctions :: !Int

_mtServiceTimings :: !ServiceTimingMetrics

_mtPgVersion :: !PGVersion

_mtActions :: !ActionMetric

data HasuraTelemetry

Constructors

HasuraTelemetry

Fields

_htDbUid :: !Text

_htInstanceUid :: !InstanceId

_htVersion :: !Version

_htCi :: !(Maybe CI)

_htMetrics :: !Metrics

data TelemetryPayload

Constructors

TelemetryPayload

Fields

_tpTopic :: !Text

_tpData :: !HasuraTelemetry

telemetryUrl :: Text

mkPayload :: Text -> InstanceId -> Version -> Metrics -> IO TelemetryPayload

runTelemetry

Arguments

:: Logger Hasura

-> Manager

-> IO SchemaCache
an action that always returns the latest schema cache

-> Text

-> InstanceId

-> PGVersion

-> IO void

An infinite loop that sends updated telemetry data (Metrics) every 24 hours. The send time depends on when the server was started and will naturally drift.

computeMetrics :: SchemaCache -> ServiceTimingMetrics -> PGVersion -> Metrics

computeActionsMetrics :: ActionCache -> ActionMetric

data TelemetryLog

Logging related

Constructors

TelemetryLog

Fields

_tlLogLevel :: !LogLevel

_tlType :: !Text

_tlMessage :: !Text

_tlHttpError :: !(Maybe TelemetryHttpError)

data TelemetryHttpError

Constructors

TelemetryHttpError

Fields

tlheStatus :: !(Maybe Status)

tlheUrl :: !Text

tlheHttpException :: !(Maybe HttpException)

tlheResponse :: !(Maybe Text)

mkHttpError :: Text -> Maybe (Response ByteString) -> Maybe HttpException -> TelemetryHttpError

mkTelemetryLog :: Text -> Text -> Maybe TelemetryHttpError -> TelemetryLog