Hasura.Server.Types

Documentation

newtype RequestId

Constructors

RequestId

Fields

unRequestId :: Text

getRequestId :: MonadIO m => [Header] -> m (RequestId, [Header])

newtype DbUid

A uuid of a source database.

Constructors

DbUid

Fields

getDbUid :: Text

newtype DbVersion

Constructors

DbVersion

Fields

unDbVersion :: Text

newtype PGVersion

Constructors

PGVersion

Fields

unPGVersion :: Int

pgToDbVersion :: PGVersion -> DbVersion

newtype MetadataDbId

A uuid of the postgres metadata db.

Constructors

MetadataDbId

Fields

getMetadataDbId :: Text

mdDbIdToDbUid :: MetadataDbId -> DbUid

newtype InstanceId

Constructors

InstanceId

Fields

getInstanceId :: Text

generateInstanceId :: IO InstanceId

Generate an InstanceId from a UUID

data ExperimentalFeature

Constructors

EFInheritedRoles

EFOptimizePermissionFilters

EFNamingConventions

EFStreamingSubscriptions

EFApolloFederation

EFHideUpdateManyFields

data MaintenanceMode a

Constructors

MaintenanceModeEnabled a

MaintenanceModeDisabled

data ReadOnlyMode

See Note [ReadOnly Mode]

Constructors

ReadOnlyModeEnabled

ReadOnlyModeDisabled

data EventingMode

EventingMode decides whether the eventing subsystem should be enabled or disabled. EventDisabled mode disables Event Triggers, Async Actions, Scheduled Events and source catalaog migrations. This is an internal feature and will not be exposed to users.

Constructors

EventingEnabled

EventingDisabled

data ServerConfigCtx

Constructors

ServerConfigCtx

Fields

_sccFunctionPermsCtx :: InferFunctionPermissions

_sccRemoteSchemaPermsCtx :: RemoteSchemaPermissions

_sccSQLGenCtx :: SQLGenCtx

_sccMaintenanceMode :: MaintenanceMode ()

_sccExperimentalFeatures :: HashSet ExperimentalFeature

_sccEventingMode :: EventingMode

_sccReadOnlyMode :: ReadOnlyMode

_sccDefaultNamingConvention :: Maybe NamingCase

stores global default naming convention

class Monad m => HasServerConfigCtx m where

Methods

askServerConfigCtx :: m ServerConfigCtx