Hasura.Server.Types

Documentation

newtype RequestId

Constructors

RequestId

Fields

unRequestId :: Text

getRequestId :: MonadIO m => [Header] -> m (RequestId, [Header])

newtype DbUid

Constructors

DbUid

Fields

getDbUid :: Text

newtype PGVersion

Constructors

PGVersion

Fields

unPGVersion :: Int

newtype InstanceId

Constructors

InstanceId

Fields

getInstanceId :: Text

data ExperimentalFeature

Constructors

EFInheritedRoles

data MaintenanceMode

Constructors

MaintenanceModeEnabled

MaintenanceModeDisabled

data ReadOnlyMode

See Note [ReadOnly Mode]

Constructors

ReadOnlyModeEnabled

ReadOnlyModeDisabled

data EventingMode

EventingMode decides whether the eventing subsystem should be enabled or disabled. EventDisabled mode disables Event Triggers, Async Actions, Scheduled Events and source catalaog migrations. This is an internal feature and will not be exposed to users.

Constructors

EventingEnabled

EventingDisabled

data ServerConfigCtx

Constructors

ServerConfigCtx

Fields

_sccFunctionPermsCtx :: FunctionPermissionsCtx

_sccRemoteSchemaPermsCtx :: RemoteSchemaPermsCtx

_sccSQLGenCtx :: SQLGenCtx

_sccMaintenanceMode :: MaintenanceMode

_sccExperimentalFeatures :: HashSet ExperimentalFeature

_sccEventingMode :: EventingMode

_sccReadOnlyMode :: ReadOnlyMode