Hasura.Server.Types

Documentation

newtype RequestId

Constructors

RequestId

Fields

unRequestId :: Text

getRequestId :: MonadIO m => [Header] -> m (RequestId, [Header])

newtype DbUid

A uuid of a source database.

Constructors

DbUid

Fields

getDbUid :: Text

newtype DbVersion

Constructors

DbVersion Text

newtype PGVersion

Constructors

PGVersion Int

pgToDbVersion :: PGVersion -> DbVersion

newtype MetadataDbId

A uuid of the postgres metadata db.

Constructors

MetadataDbId

Fields

getMetadataDbId :: Text

mdDbIdToDbUid :: MetadataDbId -> DbUid

newtype InstanceId

A UUID for each running instance of graphql-engine, generated fresh each time graphql-engine starts up

Constructors

InstanceId

Fields

getInstanceId :: Text

generateInstanceId :: IO InstanceId

Generate an InstanceId from a UUID

data ExperimentalFeature

Constructors

EFInheritedRoles

EFOptimizePermissionFilters

EFNamingConventions

EFStreamingSubscriptions

EFApolloFederation

EFHideUpdateManyFields

EFBigQueryStringNumericInput

EFHideAggregationPredicates

EFHideStreamFields

EFGroupByAggregations

EFDisablePostgresArrays

experimentalFeatureKey :: ExperimentalFeature -> Text

data MaintenanceMode a

Constructors

MaintenanceModeEnabled a

MaintenanceModeDisabled

data ReadOnlyMode

See Note [ReadOnly Mode]

Constructors

ReadOnlyModeEnabled

ReadOnlyModeDisabled

data EventingMode

EventingMode decides whether the eventing subsystem should be enabled or disabled. EventDisabled mode disables Event Triggers, Async Actions, Scheduled Events and source catalaog migrations. This is an internal feature and will not be exposed to users.

Constructors

EventingEnabled

EventingDisabled

data ApolloFederationStatus

Whether or not to enable apollo federation fields.

Constructors

ApolloFederationEnabled

ApolloFederationDisabled

isApolloFederationEnabled :: ApolloFederationStatus -> Bool

data GranularPrometheusMetricsState

Whether or not to enable granular metrics for Prometheus.

GranularMetricsOn will enable the dynamic labels for the metrics. GranularMetricsOff will disable the dynamic labels for the metrics.

*Warning**: Enabling dynamic labels for Prometheus metrics can cause cardinality issues and can cause memory usage to increase.

Constructors

GranularMetricsOff

GranularMetricsOn

data OpenTelemetryExporterState

Whether or not to enable OpenTelemetry Exporter.

OpenTelemetryExporterOn will enable exporting of traces & metrics via the OTel Exporter. OpenTelemetryExporterOff will disable exporting of traces & metrics via the OTel Exporter.

Constructors

OpenTelemetryExporterOff

OpenTelemetryExporterOn

data CloseWebsocketsOnMetadataChangeStatus

Whether or not to close websocket connections on metadata change.

Constructors

CWMCEnabled

CWMCDisabled

isCloseWebsocketsOnMetadataChangeStatusEnabled :: CloseWebsocketsOnMetadataChangeStatus -> Bool

class Monad m => MonadGetPolicies m where

Methods

runGetApiTimeLimit :: m (Maybe MaxTime)

runGetPrometheusMetricsGranularity :: m (IO GranularPrometheusMetricsState)