Hasura.Session

Documentation

sessionVariableToGraphQLName :: SessionVariable -> Maybe Name

Converts a SessionVariable value to a GraphQL name. This will fail if the session variable contains characters that are not valid for a graphql names. It is the caller's responsibility to decide what to do in such a case.

filterSessionVariables :: (SessionVariable -> SessionVariableValue -> Bool) -> SessionVariables -> SessionVariables

mkSessionVariablesHeaders :: [Header] -> SessionVariables

sessionVariablesToHeaders :: SessionVariables -> [Header]

getSessionVariables :: SessionVariables -> [Text]

getSessionVariablesSet :: SessionVariables -> HashSet SessionVariable

getSessionVariableValue :: SessionVariable -> SessionVariables -> Maybe SessionVariableValue

data UserAdminSecret

Represent the admin secret state; whether the secret is sent in the request or if actually authorization is not configured.

Constructors

UAdminSecretSent

UAdminSecretNotSent

UAuthNotSet

askCurRole :: UserInfoM m => m RoleName

mkUserInfo :: forall m. MonadError QErr m => UserRoleBuild -> UserAdminSecret -> SessionVariables -> m UserInfo

Build UserInfo from SessionVariables

maybeRoleFromSessionVariables :: SessionVariables -> Maybe RoleName

adminUserInfo :: UserInfo