Hasura.StoredProcedure.Metadata

Description

Metadata representation of a stored procedure in the metadata, as well as a parser and prettyprinter for the query code.

Documentation

data StoredProcedureMetadata (b :: BackendType)

The representation of stored procedures within the metadata structure.

Constructors

StoredProcedureMetadata

Fields

_spmStoredProcedure :: FunctionName b

_spmConfig :: StoredProcedureConfig

_spmReturns :: LogicalModelName

_spmArguments :: HashMap ArgumentName (NullableScalarType b)

_spmDescription :: Maybe Text

spmStoredProcedure :: forall b. Lens' (StoredProcedureMetadata b) (FunctionName b)

spmReturns :: forall b. Lens' (StoredProcedureMetadata b) LogicalModelName

spmDescription :: forall b. Lens' (StoredProcedureMetadata b) (Maybe Text)

spmConfig :: forall b. Lens' (StoredProcedureMetadata b) StoredProcedureConfig

spmArguments :: forall b. Lens' (StoredProcedureMetadata b) (HashMap ArgumentName (NullableScalarType b))