Hasura.Table.Metadata

Description

In order to avoid circular dependencies while splitting Metadata into multiple modules, some definitions must be moved out of that module. This module is the bucket for definitions that have not been specifically moved elsewhere.

Documentation

data ComputedFieldMetadata b

Constructors

ComputedFieldMetadata

Fields

_cfmName :: ComputedFieldName

_cfmDefinition :: ComputedFieldDefinition b

_cfmComment :: Comment

type Relationships a = InsOrdHashMap RelName a

type ComputedFields b = InsOrdHashMap ComputedFieldName (ComputedFieldMetadata b)

type Permissions a = InsOrdHashMap RoleName a

data TableMetadata b

Constructors

TableMetadata

Fields

_tmTable :: TableName b

_tmIsEnum :: Bool

_tmConfiguration :: TableConfig b

_tmObjectRelationships :: Relationships (ObjRelDef b)

_tmArrayRelationships :: Relationships (ArrRelDef b)

_tmComputedFields :: ComputedFields b

_tmRemoteRelationships :: RemoteRelationships

_tmInsertPermissions :: Permissions (InsPermDef b)

_tmSelectPermissions :: Permissions (SelPermDef b)

_tmUpdatePermissions :: Permissions (UpdPermDef b)

_tmDeletePermissions :: Permissions (DelPermDef b)

_tmEventTriggers :: EventTriggers b

_tmApolloFederationConfig :: Maybe ApolloFederationConfig

tmUpdatePermissions :: forall b. Lens' (TableMetadata b) (Permissions (UpdPermDef b))

tmTable :: forall b. Lens' (TableMetadata b) (TableName b)

tmSelectPermissions :: forall b. Lens' (TableMetadata b) (Permissions (SelPermDef b))

tmRemoteRelationships :: forall b. Lens' (TableMetadata b) RemoteRelationships

tmObjectRelationships :: forall b. Lens' (TableMetadata b) (Relationships (ObjRelDef b))

tmIsEnum :: forall b. Lens' (TableMetadata b) Bool

tmInsertPermissions :: forall b. Lens' (TableMetadata b) (Permissions (InsPermDef b))

tmEventTriggers :: forall b. Lens' (TableMetadata b) (EventTriggers b)

tmDeletePermissions :: forall b. Lens' (TableMetadata b) (Permissions (DelPermDef b))

tmConfiguration :: forall b. Lens' (TableMetadata b) (TableConfig b)

tmComputedFields :: forall b. Lens' (TableMetadata b) (ComputedFields b)

tmArrayRelationships :: forall b. Lens' (TableMetadata b) (Relationships (ArrRelDef b))

tmApolloFederationConfig :: forall b. Lens' (TableMetadata b) (Maybe ApolloFederationConfig)

mkTableMeta :: TableName b -> Bool -> TableConfig b -> TableMetadata b