Hasura.Tracing.Class

Description

Defines the Tracing API.

The MonadTrace class defines the "public API" of this component.

Documentation

class MonadTraceContext m => MonadTrace (m :: Type -> Type) where

Methods

newTraceWith :: TraceContext -> SamplingPolicy -> Text -> m a -> m a

Trace the execution of a block of code, attaching a human-readable name. This starts a new trace and its corresponding root span, to which subsequent spans will be attached.

newSpanWith :: SpanId -> Text -> m a -> m a

Starts a new span within the current trace. No-op if there's no current trace.

TODO: we could rewrite this to start a new trace if there isn't one, using the default reporter and policy? This would guarantee that no span is ever lost, but would also risk reporting undesired spans.

attachMetadata :: TraceMetadata -> m ()

Log some arbitrary metadata to be attached to the current span, if any.

class Monad m => MonadTraceContext (m :: Type -> Type) where

Access to the current tracing context, factored out of MonadTrace so we can use it separately and dispatch the constraint in cases outside of a MonadTrace context.

Methods

currentContext :: m (Maybe TraceContext)

Ask for the current tracing context, so that we can provide it to any downstream services, e.g. in HTTP headers. Returns Nothing if we're not currently tracing anything.

runNoMonadTraceContext :: NoMonadTraceContext f a -> f a

Provide an empty TraceContext where we aren't in a MonadTrace context.

newTrace :: (MonadIO m, MonadTrace m) => SamplingPolicy -> Text -> m a -> m a

Create a new trace using a randomly-generated context.

newSpan :: (MonadIO m, MonadTrace m) => Text -> m a -> m a

Create a new span with a randomly-generated id.