Hasura.Tracing.TraceId

TraceId

data TraceId

128-bit trace identifiers.

TraceIds are guaranteed to have at least one non-zero bit.

randomTraceId :: IO TraceId

traceIdFromBytes :: ByteString -> Maybe TraceId

Create a TraceId from a ByteString.

Fails if the ByteString is not exactly 16 bytes long, or if it contains only zero bytes.

traceIdToBytes :: TraceId -> ByteString

Convert a TraceId to a ByteString of 16 bytes.

traceIdFromHex :: ByteString -> Maybe TraceId

Create a TraceId from a ByteString of hex characters.

Fails if the ByteString is not exactly 32 characters long, or if it contains only zero characters.

traceIdToHex :: TraceId -> ByteString

Convert a TraceId to a ByteString of 32 lowercase hex characters.

data SpanId

64-bit span identifiers

SpanIds are guaranteed to have at least one non-zero bit.

randomSpanId :: IO SpanId

spanIdFromBytes :: ByteString -> Maybe SpanId

Create a SpanId from a ByteString.

Fails if the ByteString is not exactly 8 bytes long, or if it contains only zero bytes.

spanIdToBytes :: SpanId -> ByteString

Convert a SpanId to a ByteString of 8 bytes.

spanIdFromHex :: ByteString -> Maybe SpanId

Create a SpanId from a ByteString of hex characters.

Fails if the ByteString is not exactly 16 characters long, or if it contains only zero characters.

spanIdToHex :: SpanId -> ByteString

Convert a SpanId to a ByteString of 16 lowercase hex characters.